R6 Class Representing PosteriorBVARGROUPPANEL

The class PosteriorBVARGROUPPANEL contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model with country grouping. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Super class

PosteriorBVARPANEL -> PosteriorBVARGROUPPANEL

Public fields

last_draw

an object of class BVARGROUPPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

PosteriorBVARGROUPPANEL$new()

PosteriorBVARGROUPPANEL$clone()

Inherited methods

PosteriorBVARPANEL$get_last_draw()

PosteriorBVARPANEL$get_posterior()

PosteriorBVARGROUPPANEL$new()

Create a new posterior output PosteriorBVARGROUPPANEL.

Usage

PosteriorBVARGROUPPANEL$new(specification, posterior)

Arguments

specification

an object of class BVARGROUPPANEL with the last draw of the current MCMC run as the starting value.

posterior

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARGROUPPANEL.

PosteriorBVARGROUPPANEL$clone()

The objects of this class are cloneable with this method.

Usage

PosteriorBVARGROUPPANEL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

spec = specify_bvarGroupPANEL$new(
   data = ilo_dynamic_panel[1:5],
   G = 2
)
#> Country groupings will be estimated.
#posterior       = estimate(specification, 5)
#class(posterior)