R6 Class Representing PosteriorBVARPANEL

The class PosteriorBVARPANEL contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvarPANEL$new()

specify_posterior_bvarPANEL$get_posterior()

specify_posterior_bvarPANEL$get_last_draw()

specify_posterior_bvarPANEL$clone()

Method new()

Create a new posterior output PosteriorBVARPANEL.

Usage

specify_posterior_bvarPANEL$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARPANEL.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvarPANEL$get_posterior()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvarPANEL$get_last_draw()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvarPANEL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARPANEL" "R6"                

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_posterior`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $A
#> , , 1
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.9086899  0.5827200 -0.02068583  0.3045406
#> [2,] -0.3649608  1.1156961 -0.25225521 -0.3177456
#> [3,] -0.2773026 -1.2514920  0.72214210  0.2133890
#> [4,] -0.8041688 -0.5682655 -0.03936709  0.3541840
#> [5,]  1.0601679  2.5662142  0.37587442  0.8258970
#> 
#> , , 2
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.18452580  0.2990168  0.3485201  0.9478213
#> [2,]  0.04981037  0.8855061 -0.5638026 -0.4193130
#> [3,] -0.13164040  0.1542738 -0.1114723 -0.4641296
#> [4,]  0.30334256 -0.4587238  0.3631515  0.5745584
#> [5,] -0.01914217  3.9565975  2.2686231  3.9371139
#> 
#> , , 3
#> 
#>             [,1]       [,2]        [,3]         [,4]
#> [1,]  0.96065340 -0.2768343  0.30891590  0.003421184
#> [2,] -0.16204895  0.9388404  0.17454745  0.485851296
#> [3,] -0.17429179  0.4570013 -0.08564776 -0.294083759
#> [4,]  0.08616396 -0.3737322  0.36011111  0.419210875
#> [5,]  0.34528767  3.1035371  4.23345302  5.772935818
#> 
#> , , 4
#> 
#>               [,1]        [,2]        [,3]       [,4]
#> [1,]  0.9560219516  0.01289984  0.10844615 0.06164947
#> [2,]  0.0080975340  1.42947636 -0.07745594 0.77849189
#> [3,] -0.0040626409  1.09887566 -0.35678042 0.10361630
#> [4,]  0.0263885772 -0.80905463  0.61967623 0.39806186
#> [5,] -0.0002704809  3.54366186  7.08131066 8.99605950
#> 
#> , , 5
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.94863833  0.10585470  0.3525482  0.3664758
#> [2,]  0.03270860  0.69882219  1.2526726  1.6971305
#> [3,]  0.13987411 -0.07973851  2.1212171  2.0024928
#> [4,] -0.03885925  0.03729357 -0.6369875 -0.4573504
#> [5,]  0.05610116  1.12178485  4.8249127  5.0416099
#> 
#> 
#> $V
#> , , 1
#> 
#>             [,1]      [,2]       [,3]        [,4]       [,5]
#> [1,]  1.81772340 0.1659609 0.01714014 -0.04783207  2.1572760
#> [2,]  0.16596093 0.9281721 0.18089058  0.53599578  0.4507438
#> [3,]  0.01714014 0.1808906 1.41999482  0.24622244  0.8820763
#> [4,] -0.04783207 0.5359958 0.24622244  3.70454252 -2.4087344
#> [5,]  2.15727597 0.4507438 0.88207628 -2.40873441 23.5066559
#> 
#> , , 2
#> 
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,]  1.8887732 -0.3694837 -0.4275662 -0.2399615  0.1083812
#> [2,] -0.3694837  2.0807643  0.4489534 -0.1687459 -1.0786276
#> [3,] -0.4275662  0.4489534  1.7959940  0.3995261 -0.4846039
#> [4,] -0.2399615 -0.1687459  0.3995261  1.3845515  0.4387445
#> [5,]  0.1083812 -1.0786276 -0.4846039  0.4387445 22.8428439
#> 
#> , , 3
#> 
#>             [,1]        [,2]       [,3]        [,4]        [,5]
#> [1,]  2.48076726 -0.25826609 -0.6258476 -0.01598454 -0.55900659
#> [2,] -0.25826609  2.34208182  0.4322407 -0.83979262 -0.08862941
#> [3,] -0.62584760  0.43224068  3.1413885 -1.08787927 -1.34994024
#> [4,] -0.01598454 -0.83979262 -1.0878793  3.38410599  3.36454727
#> [5,] -0.55900659 -0.08862941 -1.3499402  3.36454727 28.18683290
#> 
#> , , 4
#> 
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,]  1.2128992 -0.4534195 -0.7646554  0.6802681 -2.474545
#> [2,] -0.4534195  3.3233588  3.4254436 -3.0899527 -1.536585
#> [3,] -0.7646554  3.4254436  6.2703020 -4.0259274 -3.993598
#> [4,]  0.6802681 -3.0899527 -4.0259274  4.7493430 -1.326804
#> [5,] -2.4745447 -1.5365852 -3.9935983 -1.3268038 54.591790
#> 
#> , , 5
#> 
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,]  1.6388871 -0.2687321 -0.789413  0.4347459 -3.249478
#> [2,] -0.2687321  3.8909067  2.242666 -3.0612433 -2.435822
#> [3,] -0.7894130  2.2426659  5.512898 -3.8359104 -1.689059
#> [4,]  0.4347459 -3.0612433 -3.835910  6.7171010 -3.106353
#> [5,] -3.2494776 -2.4358220 -1.689059 -3.1063526 68.796308
#> 
#> 
#> $Sigma
#> , , 1
#> 
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  8.2164005 -3.644756 -0.1089406 0.2447229
#> [2,] -3.6447564  9.620172  3.6430839 1.0719004
#> [3,] -0.1089406  3.643084 10.0826840 2.8453192
#> [4,]  0.2447229  1.071900  2.8453192 8.1163641
#> 
#> , , 2
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  2.0376899  -0.8176513  -0.1035995 -0.1605059
#> [2,] -0.8176513  15.6378574 -10.6615132 -1.4580289
#> [3,] -0.1035995 -10.6615132  15.3123102  9.2500502
#> [4,] -0.1605059  -1.4580289   9.2500502 10.5989596
#> 
#> , , 3
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  0.5986883 -0.2368056  0.330923  0.619249
#> [2,] -0.2368056  7.1897027 -6.061808 -1.144409
#> [3,]  0.3309230 -6.0618080 10.709705  7.033190
#> [4,]  0.6192490 -1.1444089  7.033190  7.639681
#> 
#> , , 4
#> 
#>             [,1]        [,2]       [,3]      [,4]
#> [1,]  0.23490568 -0.03711554  0.2727292 0.2433105
#> [2,] -0.03711554  6.08776948 -2.9606317 1.0651893
#> [3,]  0.27272921 -2.96063173  8.2396070 6.3896580
#> [4,]  0.24331053  1.06518933  6.3896580 7.4540653
#> 
#> , , 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.1260828 -0.3598900  0.1731689 -0.1018938
#> [2,] -0.3598900 12.4794702 -8.9744753 -0.2621721
#> [3,]  0.1731689 -8.9744753 13.1955189  7.7032966
#> [4,] -0.1018938 -0.2621721  7.7032966  8.5911055
#> 
#> 
#> $nu
#>          [,1]
#> [1,] 5.099974
#> [2,] 5.099863
#> [3,] 5.099978
#> [4,] 5.100197
#> [5,] 5.100080
#> 
#> $m
#>            [,1]
#> [1,] 1.17150253
#> [2,] 2.00302119
#> [3,] 0.07236432
#> [4,] 0.96247861
#> [5,] 3.47005931
#> 
#> $w
#>          [,1]
#> [1,] 36.55243
#> [2,] 32.34737
#> [3,] 46.72427
#> [4,] 19.89430
#> [5,] 40.45522
#> 
#> $s
#>           [,1]
#> [1,] 0.3605776
#> [2,] 0.8098917
#> [3,] 1.0460671
#> [4,] 0.9373831
#> [5,] 1.5632431
#> 
#> $scale
#>            [,1]
#> [1,] 0.01427928
#> [2,] 0.01427928
#> [3,] 0.01641726
#> [4,] 0.01838007
#> [5,] 0.02049814
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.03746886 -0.0624251  0.05238723  0.03759366
#> [2,] -0.06242510  0.5298027 -0.40165160 -0.21558972
#> [3,]  0.05238723 -0.4016516  0.61948323  0.41947700
#> [4,]  0.03759366 -0.2155897  0.41947700  0.35521183
#> 
#> , , 2, 1
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.062922098 -0.080086376  0.07599655 0.003521851
#> [2,] -0.080086376  0.415855307 -0.26169069 0.004686959
#> [3,]  0.075996554 -0.261690691  0.53646336 0.345042057
#> [4,]  0.003521851  0.004686959  0.34504206 0.423742200
#> 
#> , , 3, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.1036735  0.1341840 -0.2813493 -0.2195710
#> [2,]  0.1341840  5.9660346 -4.4180582 -0.3698526
#> [3,] -0.2813493 -4.4180582  5.2974635  2.6391287
#> [4,] -0.2195710 -0.3698526  2.6391287  2.8799303
#> 
#> , , 4, 1
#> 
#>              [,1]         [,2]         [,3]        [,4]
#> [1,]  0.047027312 -0.006129287  0.005816185  0.01323058
#> [2,] -0.006129287  0.276385641 -0.163451170 -0.04689123
#> [3,]  0.005816185 -0.163451170  0.357304499  0.23140510
#> [4,]  0.013230581 -0.046891231  0.231405101  0.21468441
#> 
#> , , 5, 1
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.02470429 -0.01138853  0.0456765  0.05156836
#> [2,] -0.01138853  3.09099086 -2.8539216 -1.09268939
#> [3,]  0.04567650 -2.85392159  3.5494614  1.99249947
#> [4,]  0.05156836 -1.09268939  1.9924995  1.51619667
#> 
#> , , 1, 2
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.016832837 -0.008747094  0.01171873  0.01265563
#> [2,] -0.008747094  0.278114679 -0.28063464 -0.11949213
#> [3,]  0.011718729 -0.280634639  0.43078094  0.29069742
#> [4,]  0.012655633 -0.119492131  0.29069742  0.26818453
#> 
#> , , 2, 2
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.014071736 -0.02824039  0.02837288 0.003460435
#> [2,] -0.028240386  0.30937216 -0.24227102 0.021397661
#> [3,]  0.028372875 -0.24227102  0.42508455 0.269029887
#> [4,]  0.003460435  0.02139766  0.26902989 0.363650048
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  0.1764930 -0.2539984 -1.128272 -1.536306
#> [2,] -0.2539984 11.8793692 -5.569693  2.860782
#> [3,] -1.1282725 -5.5696933 12.876212 10.828108
#> [4,] -1.5363057  2.8607817 10.828108 15.140846
#> 
#> , , 4, 2
#> 
#>              [,1]        [,2]       [,3]         [,4]
#> [1,]  0.016871481 -0.02679462  0.0065599 -0.002282126
#> [2,] -0.026794622  0.29460133 -0.2203703 -0.018068021
#> [3,]  0.006559900 -0.22037028  0.3302669  0.166199783
#> [4,] -0.002282126 -0.01806802  0.1661998  0.154237712
#> 
#> , , 5, 2
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.01060140 -0.0820689  0.1179417  0.07982028
#> [2,] -0.08206890  2.8528963 -2.7511295 -1.14754306
#> [3,]  0.11794175 -2.7511295  3.7599123  2.33234612
#> [4,]  0.07982028 -1.1475431  2.3323461  1.83290531
#> 
#> , , 1, 3
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.007746535 -0.01413453  0.01682095  0.009301721
#> [2,] -0.014134532  0.22522306 -0.19348534 -0.074148239
#> [3,]  0.016820953 -0.19348534  0.42117346  0.342203955
#> [4,]  0.009301721 -0.07414824  0.34220395  0.342305337
#> 
#> , , 2, 3
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.01108852 -0.03081633  0.04526663  0.02633728
#> [2,] -0.03081633  0.19663879 -0.25555599 -0.10778878
#> [3,]  0.04526663 -0.25555599  0.46830515  0.29848998
#> [4,]  0.02633728 -0.10778878  0.29848998  0.25368048
#> 
#> , , 3, 3
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.1277513 -0.4729194 -0.7875319 -1.306141
#> [2,] -0.4729194  9.5411212 -2.7704177  4.178612
#> [3,] -0.7875319 -2.7704177 11.2690270 11.275007
#> [4,] -1.3061407  4.1786120 11.2750074 16.755169
#> 
#> , , 4, 3
#> 
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.009670540 -0.01643394 -0.006287566 -0.01373974
#> [2,] -0.016433937  0.23069183 -0.078102053  0.10507963
#> [3,] -0.006287566 -0.07810205  0.270511580  0.19921870
#> [4,] -0.013739736  0.10507963  0.199218704  0.27932757
#> 
#> , , 5, 3
#> 
#>              [,1]       [,2]        [,3]        [,4]
#> [1,]  0.007783509 -0.1042222  0.08748927  0.02670442
#> [2,] -0.104222153  3.9823565 -2.80208780 -0.41933834
#> [3,]  0.087489268 -2.8020878  3.22959395  1.72740633
#> [4,]  0.026704419 -0.4193383  1.72740633  1.67808728
#> 
#> , , 1, 4
#> 
#>              [,1]       [,2]        [,3]        [,4]
#> [1,]  0.006928296 -0.0197445  0.03022421  0.01883585
#> [2,] -0.019744497  0.2668504 -0.26302177 -0.14017049
#> [3,]  0.030224210 -0.2630218  0.52200884  0.43016680
#> [4,]  0.018835847 -0.1401705  0.43016680  0.41060613
#> 
#> , , 2, 4
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.01732864 -0.06337269  0.07668484  0.02706565
#> [2,] -0.06337269  0.38665389 -0.40642580 -0.10901978
#> [3,]  0.07668484 -0.40642580  0.67631093  0.40010115
#> [4,]  0.02706565 -0.10901978  0.40010115  0.36179291
#> 
#> , , 3, 4
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.04536379 -0.09537895 -0.1833882 -0.2857444
#> [2,] -0.09537895  6.92329303 -3.9261537  0.7678108
#> [3,] -0.18338816 -3.92615370  4.7266912  2.4723580
#> [4,] -0.28574443  0.76781081  2.4723580  3.4861887
#> 
#> , , 4, 4
#> 
#>              [,1]         [,2]        [,3]         [,4]
#> [1,]  0.001924002 -0.005039681  0.00149233 -0.002641346
#> [2,] -0.005039681  0.161958929 -0.15602098 -0.035111511
#> [3,]  0.001492330 -0.156020976  0.33619807  0.225061081
#> [4,] -0.002641346 -0.035111511  0.22506108  0.205532927
#> 
#> , , 5, 4
#> 
#>              [,1]        [,2]        [,3]       [,4]
#> [1,]  0.003683876 -0.08465103  0.08746826  0.0401423
#> [2,] -0.084651034  3.35016686 -2.69631260 -0.7394827
#> [3,]  0.087468262 -2.69631260  3.15925767  1.6948169
#> [4,]  0.040142297 -0.73948274  1.69481694  1.3875455
#> 
#> , , 1, 5
#> 
#>              [,1]        [,2]       [,3]        [,4]
#> [1,]  0.005405793 -0.02355524  0.0298605  0.01788398
#> [2,] -0.023555243  0.32362130 -0.3474980 -0.19625542
#> [3,]  0.029860499 -0.34749798  0.5608852  0.41205057
#> [4,]  0.017883976 -0.19625542  0.4120506  0.34788167
#> 
#> , , 2, 5
#> 
#>               [,1]        [,2]        [,3]          [,4]
#> [1,]  0.0055060147 -0.01028597  0.00676690 -0.0003543028
#> [2,] -0.0102859699  0.14082802 -0.06037669  0.0551408007
#> [3,]  0.0067668996 -0.06037669  0.36868071  0.3821264023
#> [4,] -0.0003543028  0.05514080  0.38212640  0.5051140199
#> 
#> , , 3, 5
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.01452449 -0.2548305  0.05716717 -0.1348496
#> [2,] -0.25483050  9.5507192 -3.66568957  3.1878813
#> [3,]  0.05716717 -3.6656896  3.29615966  1.0349470
#> [4,] -0.13484957  3.1878813  1.03494703  3.7771235
#> 
#> , , 4, 5
#> 
#>               [,1]         [,2]         [,3]          [,4]
#> [1,]  0.0012470684 -0.004925084  0.003794871  0.0001405886
#> [2,] -0.0049250836  0.331566649 -0.331656451 -0.0789577751
#> [3,]  0.0037948715 -0.331656451  0.528513279  0.2832951929
#> [4,]  0.0001405886 -0.078957775  0.283295193  0.2321498081
#> 
#> , , 5, 5
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.00310631 -0.05725965  0.05548931  0.02261522
#> [2,] -0.05725965  2.53903313 -1.85568319 -0.34182127
#> [3,]  0.05548931 -1.85568319  2.04695755  1.03401852
#> [4,]  0.02261522 -0.34182127  1.03401852  0.93924054
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  1.0097408  0.2139403 -0.3625144 -0.18445459
#> [2,] -0.1365917  1.0611217 -0.1576020 -0.01379712
#> [3,] -0.2222604  0.8193727  0.2023718  0.21064863
#> [4,]  0.1999007 -0.8748673  0.8748398  0.82335058
#> [5,]  1.1263502  0.5946367  2.5192194  3.30934568
#> 
#> , , 2, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9323086  0.1525013 -0.3093775 -0.2967337
#> [2,] -0.3969594  1.5659401 -0.3732342  0.5252613
#> [3,] -0.4801359  1.1652272  0.2057319  0.8526685
#> [4,]  0.5172159 -1.1581672  0.8640366  0.2891325
#> [5,] -0.6981721  0.9653101 -1.5865826 -1.5328866
#> 
#> , , 3, 1
#> 
#>             [,1]      [,2]       [,3]       [,4]
#> [1,]  1.01105648 -3.102430   3.041585  1.0347282
#> [2,] -0.08106446  3.400740  -2.600533 -0.7524385
#> [3,] -0.14378332  4.791890  -3.640962 -1.0386305
#> [4,]  0.11309956 -3.838035   3.715619  1.7877282
#> [5,]  1.49904129 21.258935 -17.061735 -6.7087277
#> 
#> , , 4, 1
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,] 0.779880734  0.1446003  0.7457684  0.7886330
#> [2,] 0.182361915  0.3076411 -0.3054587 -0.4070326
#> [3,] 0.009890258 -0.5603260  0.5243332 -0.2730709
#> [4,] 0.049503578  0.4794425  0.1934544  0.9801177
#> [5,] 0.801958805  3.0489623  2.1035999  2.3652827
#> 
#> , , 5, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  1.05094374 -0.7510936  1.1681680  0.78023729
#> [2,] -0.08457141  0.4789946  0.0563066  0.04339331
#> [3,] -0.14287928 -0.6347240  0.9451557  0.05015726
#> [4,]  0.13298496  0.9171788 -0.4453612  0.62623448
#> [5,] -0.77340958  4.6557940 -2.1035386 -1.55536327
#> 
#> , , 1, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.0262048  0.2337941 -0.2417386 -0.1320361
#> [2,] -0.1268157  1.2573757 -0.6387464 -0.2678235
#> [3,] -0.2288136  1.1799741 -0.6347610 -0.3017978
#> [4,]  0.2040984 -1.1930575  1.5448749  1.2520116
#> [5,]  0.7590057 -2.1633344  8.4686334  6.1889420
#> 
#> , , 2, 2
#> 
#>            [,1]       [,2]       [,3]          [,4]
#> [1,]  1.0750597 -0.3610986  0.2554424 -0.1796524331
#> [2,] -0.2600011  1.6312782 -0.8739207 -0.0611132786
#> [3,] -0.4301613  1.5090113 -0.8062735  0.0007253843
#> [4,]  0.3980494 -1.3618305  1.5395038  0.9293223916
#> [5,] -0.4837422  6.2738137  5.6966335 10.8032278820
#> 
#> , , 3, 2
#> 
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  0.7897491 -2.629293  4.0728960  2.559218
#> [2,]  0.1718691  3.288711 -4.0116246 -2.588503
#> [3,]  0.2766878  4.192962 -5.7809503 -4.139803
#> [4,] -0.1929750 -3.276269  5.2050009  4.079636
#> [5,] -0.3320907  9.549169  0.3294754  7.204742
#> 
#> , , 4, 2
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.71478832  0.6992252  0.5309484  0.9611935
#> [2,]  0.24130855  0.7381718 -0.8762673 -0.4538120
#> [3,]  0.15811628  0.3363708 -0.3087018 -0.3036720
#> [4,] -0.06794246 -0.5579024  1.0656770  0.9387095
#> [5,]  0.38083713 -0.1607337  4.4016036  3.3012570
#> 
#> , , 5, 2
#> 
#>              [,1]       [,2]      [,3]       [,4]
#> [1,]  1.049031685 -1.2183439  1.098599  0.3300597
#> [2,] -0.025728245 -0.0268649  2.111416  2.1041847
#> [3,] -0.052673264 -1.2440059  4.013212  3.2438705
#> [4,]  0.030320001  1.7026446 -3.562510 -2.4933170
#> [5,] -0.002750975  7.5163249  1.411807  5.3189098
#> 
#> , , 1, 3
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.98369636  0.03275652  0.1052875  0.04534187
#> [2,] -0.09635022  1.47520821 -1.1101064 -0.56147059
#> [3,] -0.24505137  1.18932696 -0.7854124 -0.39125205
#> [4,]  0.23137844 -1.17329772  1.5224732  1.20317019
#> [5,]  0.91999303 -0.68761879 11.8982828 10.64783689
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.0730156 -0.6382535  0.3250569 -0.3188674
#> [2,] -0.3122899  2.1325395 -0.6930042  0.8185578
#> [3,] -0.4641658  1.9655907 -0.6176461  0.7892118
#> [4,]  0.4293032 -1.7240686  1.2958179  0.1345952
#> [5,]  0.1931082  3.2864757  7.7042599 10.1992557
#> 
#> , , 3, 3
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  0.6017607  -0.426252  3.653072  3.787119
#> [2,]  0.2941174   2.361813 -3.876379 -3.087237
#> [3,]  0.5087474   2.871610 -5.820613 -5.129329
#> [4,] -0.3905718  -2.262122  5.258465  4.855915
#> [5,]  2.1753104 -20.513360  6.774353 -9.538928
#> 
#> , , 4, 3
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.80638919 -0.03622430  0.7471989  0.5833113
#> [2,]  0.33874972 -0.04522015 -0.7155789 -0.9230140
#> [3,]  0.34016683 -1.05307399  0.1046462 -0.9879804
#> [4,] -0.27546658  1.02470102  0.6518337  1.7803220
#> [5,] -0.01537344  3.81198892 -0.8180123  1.1625502
#> 
#> , , 5, 3
#> 
#>             [,1]       [,2]      [,3]        [,4]
#> [1,]  0.95969700 -1.8727452  1.244224  0.01493242
#> [2,]  0.09404215  0.9275623  1.625165  2.25400491
#> [3,]  0.13813747  0.5625334  3.020132  3.44271374
#> [4,] -0.12382547  0.1563828 -2.692455 -2.64087524
#> [5,]  0.12743298 10.5464089  3.892818 10.37436149
#> 
#> , , 1, 4
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.94853108  0.1574927 -0.01995598  0.02331588
#> [2,] -0.09360231  1.3999032 -1.32422285 -0.86942105
#> [3,] -0.23756792  1.2007151 -1.29029039 -0.92967059
#> [4,]  0.23609493 -1.1944620  2.05529682  1.74259278
#> [5,]  1.17701845 -2.4803472 13.46937181 11.73630589
#> 
#> , , 2, 4
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  1.0356636 -0.3587834  0.01967666 -0.45000000
#> [2,] -0.2229469  1.9057562 -0.32182360  1.05798603
#> [3,] -0.3799686  1.7886846 -0.29670463  1.04344187
#> [4,]  0.3565794 -1.5874636  1.01908804 -0.08963499
#> [5,] -0.1870141  0.9876464  9.85141042 10.43684094
#> 
#> , , 3, 4
#> 
#>            [,1]      [,2]      [,3]      [,4]
#> [1,]  0.7652743 -2.637696  3.926117  2.419223
#> [2,]  0.3290291  2.413033 -4.313086 -3.619208
#> [3,]  0.5420387  3.137863 -6.436493 -5.727117
#> [4,] -0.4448446 -2.552437  6.048857  5.643390
#> [5,] -0.4620628 33.873050 -9.773597 12.579870
#> 
#> , , 4, 4
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.97220237 -0.02691389  0.6569678  0.5935486
#> [2,]  0.15850704  1.21357807  0.1099988  0.9324779
#> [3,]  0.19855698  0.54486604  1.3197368  1.5888132
#> [4,] -0.18762264 -0.56609706 -0.5970800 -0.8571961
#> [5,] -0.09208732  2.83881483  4.3660531  5.7180754
#> 
#> , , 5, 4
#> 
#>             [,1]       [,2]       [,3]      [,4]
#> [1,]  0.95894142 -0.7294687  0.6550412  0.105581
#> [2,]  0.13867333 -1.1595334  3.2940190  2.730802
#> [3,]  0.22986358 -3.2444106  6.1466384  4.399158
#> [4,] -0.20971935  3.4491448 -5.4797923 -3.559214
#> [5,] -0.08641376  9.2590161  0.8865257  6.367421
#> 
#> , , 1, 5
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9820024  0.08105482 -0.1139498 -0.1048811
#> [2,] -0.0763632  1.60689895 -0.9369520 -0.3933297
#> [3,] -0.1949379  1.28410081 -0.6007777 -0.2313444
#> [4,]  0.1955764 -1.17455642  1.4035433  1.1269482
#> [5,]  0.3221588 -6.96154290 13.2908062  9.5439647
#> 
#> , , 2, 5
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.99760352 -0.15737130  0.05492005 -0.1077409
#> [2,] -0.04422862  0.91300466  2.16439234  2.8146361
#> [3,] -0.03555682  0.09815350  3.26534384  3.2013934
#> [4,]  0.04085484 -0.07176483 -2.32574738 -2.2163561
#> [5,] -0.03403162  4.54752343 -3.83108182 -2.1816379
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  0.8324291 -0.9807164  1.978059  1.444834
#> [2,]  0.2140335  2.5616178 -2.819169 -1.707050
#> [3,]  0.3638153  2.6391757 -3.648448 -2.780831
#> [4,] -0.3010124 -2.2903706  3.795778  3.141449
#> [5,]  0.1732541  2.1571422  5.942535  5.701537
#> 
#> , , 4, 5
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.96025814  0.1414327  0.8020721  0.8707983
#> [2,]  0.10813113  1.1647104 -0.0900561  0.7706785
#> [3,]  0.13891133  0.5666024  1.0513361  1.4323740
#> [4,] -0.12456251 -0.5982241 -0.3759279 -0.7516534
#> [5,] -0.01548822 -0.5691667  4.1049189  2.4259686
#> 
#> , , 5, 5
#> 
#>            [,1]        [,2]          [,3]      [,4]
#> [1,]  0.9135271 -1.48463888  0.0001606106 -1.101524
#> [2,]  0.1098563  1.04457935  3.0365968528  3.929210
#> [3,]  0.1815091  0.46862677  5.6479674961  6.338280
#> [4,] -0.1531751  0.01976895 -4.7258302959 -5.024687
#> [5,]  0.6265354 13.04113715  2.4844388872  9.936992
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_last_draw`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|