R6 Class Representing PosteriorBVARs

The class PosteriorBVARs contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvars$new()

specify_posterior_bvars$get_posterior()

specify_posterior_bvars$get_last_draw()

specify_posterior_bvars$clone()

Method new()

Create a new posterior output PosteriorBVARs.

Usage

specify_posterior_bvars$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARs with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARs.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvars$get_posterior()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvars$get_last_draw()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvars$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARs" "R6"            

## ------------------------------------------------
## Method `specify_posterior_bvars$get_posterior`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $nu
#>      [,1] [,2]     [,3]     [,4]     [,5]
#> [1,]  5.1  5.1 5.100000 5.100000 5.100000
#> [2,]  5.1  5.1 5.100000 5.100000 5.140012
#> [3,]  5.1  5.1 5.100000 5.100000 5.013836
#> [4,]  5.1  5.1 5.116075 5.116075 5.116075
#> [5,]  5.1  5.1 5.100000 5.100000 5.100000
#> 
#> $m
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 1.4868103 0.8755315 0.8645049 0.9005044 0.5662365
#> [2,] 0.6288214 0.7696653 0.8814639 0.7293005 0.6005604
#> [3,] 0.8963327 0.7548340 0.5104913 0.2296726 0.5063411
#> [4,] 0.3334429 0.5314702 0.6579005 0.7739592 0.9697870
#> [5,] 0.6370646 1.1348699 0.8042455 0.5058683 0.8319645
#> 
#> $w
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.17318053 0.17695628 0.13989653 0.10863056 0.08411075
#> [2,] 0.07733595 0.09781292 0.19060010 0.18970959 0.24146743
#> [3,] 0.17743648 0.08799122 0.08407696 0.06157735 0.12671760
#> [4,] 0.03626684 0.06472561 0.18563891 0.20126832 0.14252859
#> [5,] 0.18605247 0.25490475 0.28331495 0.20695872 0.16110155
#> 
#> $s
#>           [,1]     [,2]     [,3]      [,4]     [,5]
#> [1,]  89.79368 55.42911 68.66958  79.07351 76.82960
#> [2,]  68.15605 59.86642 66.72588  84.94093 42.65014
#> [3,]  78.33562 33.74934 85.02910  46.72481 50.84069
#> [4,]  67.34748 60.38448 33.45911 137.79456 58.02779
#> [5,] 186.79696 34.30409 87.53810  39.69232 82.11393
#> 
#> $scale
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
#> [4,]    0
#> [5,]    0
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  4.0614812 -0.6100346 0.2980856 -0.4087125
#> [2,] -0.6100346  3.2421449 0.3143606  0.2179333
#> [3,]  0.2980856  0.3143606 2.4149544  0.6453966
#> [4,] -0.4087125  0.2179333 0.6453966  1.6384151
#> 
#> , , 2, 1
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 2.5628297  0.2305140  0.6035791 1.1015397
#> [2,] 0.2305140  2.3360649 -0.2780292 0.8945180
#> [3,] 0.6035791 -0.2780292  3.1766137 0.9472493
#> [4,] 1.1015397  0.8945180  0.9472493 3.2807402
#> 
#> , , 3, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.6861833 -0.7743625  0.1900833 -0.2571557
#> [2,] -0.7743625 14.6935985 -8.9881493  0.0821440
#> [3,]  0.1900833 -8.9881493 10.8381995  2.0724237
#> [4,] -0.2571557  0.0821440  2.0724237  4.2833420
#> 
#> , , 4, 1
#> 
#>           [,1]      [,2]     [,3]      [,4]
#> [1,] 2.4849288 0.5523043 1.229243 0.9003981
#> [2,] 0.5523043 2.2270613 1.295412 0.5753739
#> [3,] 1.2292433 1.2954123 4.039617 1.5287303
#> [4,] 0.9003981 0.5753739 1.528730 2.9543302
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  7.1539211 -0.3168971  1.320254 -1.417133
#> [2,] -0.3168971  7.3184119 -2.299974 -1.497381
#> [3,]  1.3202544 -2.2999738  7.830198  1.326479
#> [4,] -1.4171329 -1.4973807  1.326479  9.641894
#> 
#> , , 1, 2
#> 
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 1.7663746 0.4521881 0.5360532 0.0480621
#> [2,] 0.4521881 2.5042842 0.6128352 0.3933795
#> [3,] 0.5360532 0.6128352 2.4372940 0.3221243
#> [4,] 0.0480621 0.3933795 0.3221243 1.5937278
#> 
#> , , 2, 2
#> 
#>           [,1]        [,2]        [,3]      [,4]
#> [1,] 2.0931299 0.122470851 0.286977411 0.6976282
#> [2,] 0.1224709 1.849236592 0.007249194 0.5615633
#> [3,] 0.2869774 0.007249194 2.621052922 2.0404515
#> [4,] 0.6976282 0.561563274 2.040451534 5.1952055
#> 
#> , , 3, 2
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  1.04269570 -0.1700297  0.4783807  0.03306296
#> [2,] -0.17002965  6.6302841 -4.3649297 -0.27745928
#> [3,]  0.47838066 -4.3649297  5.3170473  1.50763011
#> [4,]  0.03306296 -0.2774593  1.5076301  2.26318237
#> 
#> , , 4, 2
#> 
#>           [,1]         [,2]       [,3]         [,4]
#> [1,] 1.7025925  0.338258013 0.59425485  0.216804272
#> [2,] 0.3382580  1.851468168 0.09749793 -0.003521768
#> [3,] 0.5942549  0.097497932 2.80768667  0.278096381
#> [4,] 0.2168043 -0.003521768 0.27809638  2.257762549
#> 
#> , , 5, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.4673309  0.7208407 -0.1653304  0.2906957
#> [2,]  0.7208407  4.6280265 -2.1313380 -0.7239093
#> [3,] -0.1653304 -2.1313380  3.4634440  1.5799651
#> [4,]  0.2906957 -0.7239093  1.5799651  3.0564115
#> 
#> , , 1, 3
#> 
#>              [,1]       [,2]        [,3]         [,4]
#> [1,]  1.734783568  0.5847667  0.06066383 -0.001395869
#> [2,]  0.584766676  2.6109223 -0.17891004 -0.046853300
#> [3,]  0.060663825 -0.1789100  1.95485598  0.375569711
#> [4,] -0.001395869 -0.0468533  0.37556971  1.777356470
#> 
#> , , 2, 3
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  2.03477114 -0.07850037  0.1236051 0.64091605
#> [2,] -0.07850037  2.58107361 -0.4796254 0.03223496
#> [3,]  0.12360508 -0.47962542  2.6606832 0.77518664
#> [4,]  0.64091605  0.03223496  0.7751866 2.31257998
#> 
#> , , 3, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  4.9463121  0.3728788  0.2536934 -0.2136696
#> [2,]  0.3728788 14.9079047 -8.1163217  0.3523093
#> [3,]  0.2536934 -8.1163217  8.9074414  0.8476781
#> [4,] -0.2136696  0.3523093  0.8476781  2.9743859
#> 
#> , , 4, 3
#> 
#>             [,1]         [,2]       [,3]         [,4]
#> [1,]  0.91417153  0.181090038 -0.1084598  0.026709678
#> [2,]  0.18109004  1.055591152 -0.2186628 -0.008759854
#> [3,] -0.10845984 -0.218662820  1.8756638  1.026149141
#> [4,]  0.02670968 -0.008759854  1.0261491  1.601445705
#> 
#> , , 5, 3
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 2.5377748  0.7618514  0.7386182 0.3689907
#> [2,] 0.7618514  4.6557058 -0.1989100 0.5836920
#> [3,] 0.7386182 -0.1989100  5.1702215 2.0849554
#> [4,] 0.3689907  0.5836920  2.0849554 4.7174080
#> 
#> , , 1, 4
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  3.13196530 0.09934177 -0.2960583 -0.2320545
#> [2,]  0.09934177 2.89480054  0.5237719  0.1807424
#> [3,] -0.29605833 0.52377193  1.9955102  0.2325977
#> [4,] -0.23205446 0.18074242  0.2325977  2.2622514
#> 
#> , , 2, 4
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  1.8160529 -0.55991281  0.7184202  0.51249651
#> [2,] -0.5599128  2.85477933 -0.4491301 -0.02551121
#> [3,]  0.7184202 -0.44913011  3.4907131  1.15808007
#> [4,]  0.5124965 -0.02551121  1.1580801  3.40937595
#> 
#> , , 3, 4
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 1.6387718  0.2818442  0.4313474 0.3574445
#> [2,] 0.2818442 10.4985795 -5.0563279 1.2107754
#> [3,] 0.4313474 -5.0563279  7.5505847 2.5163081
#> [4,] 0.3574445  1.2107754  2.5163081 4.3280441
#> 
#> , , 4, 4
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  3.4353491  0.73967895  1.31821461 -0.3992913
#> [2,]  0.7396789  7.26638812 -0.07232883  1.1912206
#> [3,]  1.3182146 -0.07232883  4.80379377  0.4325855
#> [4,] -0.3992913  1.19122064  0.43258555  3.8060378
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,] 1.48120541  0.3302571  0.3328488  0.09579203
#> [2,] 0.33025712  7.1561078 -3.7116245 -1.25163312
#> [3,] 0.33284877 -3.7116245  4.8246084  2.54188276
#> [4,] 0.09579203 -1.2516331  2.5418828  3.35219730
#> 
#> , , 1, 5
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  2.61007041 -0.54412607 -0.02070207 0.44528908
#> [2,] -0.54412607  1.80128469 -0.13920485 0.05377003
#> [3,] -0.02070207 -0.13920485  3.21194056 0.68285363
#> [4,]  0.44528908  0.05377003  0.68285363 2.17755652
#> 
#> , , 2, 5
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  1.15556335 -0.2562665 0.034357530 0.382394269
#> [2,] -0.25626653  2.1585745 0.111121735 0.281143426
#> [3,]  0.03435753  0.1111217 1.491951628 0.008764346
#> [4,]  0.38239427  0.2811434 0.008764346 1.845701602
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  1.9742962  0.5425906 -0.883472 -0.3665213
#> [2,]  0.5425906 10.5365376 -5.674034  0.1724546
#> [3,] -0.8834720 -5.6740344  6.999135  1.3773906
#> [4,] -0.3665213  0.1724546  1.377391  3.2282200
#> 
#> , , 4, 5
#> 
#>            [,1]        [,2]       [,3]      [,4]
#> [1,]  2.8375954 -0.14309260 0.45410468 0.9065576
#> [2,] -0.1430926  1.82421067 0.06386532 0.2442673
#> [3,]  0.4541047  0.06386532 2.90019243 0.5521724
#> [4,]  0.9065576  0.24426730 0.55217241 2.6460202
#> 
#> , , 5, 5
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  3.5658399 -1.0793197 -1.701393 -0.6721104
#> [2,] -1.0793197  5.3886320 -1.330381  0.2319736
#> [3,] -1.7013931 -1.3303813  8.028399  3.7737827
#> [4,] -0.6721104  0.2319736  3.773783  4.8603113
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,] -0.1646743  0.3420376 -0.53117289 -0.08971627
#> [2,]  0.5031697  0.5971230 -0.04689761 -0.20509645
#> [3,] -0.2379120 -0.8107409  0.26612436 -0.54101641
#> [4,]  0.7220958  0.5566051  0.87164547  1.50368331
#> [5,] -1.0285066  3.8554748  3.18483207  3.34285067
#> 
#> , , 2, 1
#> 
#>              [,1]       [,2]       [,3]        [,4]
#> [1,]  0.079757605 -0.2599432 -0.6142349 -1.18823183
#> [2,] -0.008282407  1.1921429 -0.1092303  0.02712052
#> [3,]  0.781035447  0.5308754  1.5803116  1.68233030
#> [4,] -0.354382923 -0.4083385 -0.2770998 -0.04979600
#> [5,] -0.175781014  0.3596512  0.5665151  1.26663762
#> 
#> , , 3, 1
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  1.0412268 -0.06713373  0.7609376  0.31252322
#> [2,] -0.2274730  0.63503574 -0.3326022 -0.05118748
#> [3,] -0.3223820  0.16383206 -0.1066793  0.15139012
#> [4,]  0.2998315 -0.06036012  0.7017834  0.77738451
#> [5,]  0.8683787  2.66697958  1.1939068 -0.99177103
#> 
#> , , 4, 1
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,] -0.38887142 -0.1343553 -1.0981205 -1.0649475
#> [2,] -0.21929419  0.4465230  0.1542411  0.1041658
#> [3,]  0.46446746 -0.0752963  0.5841382  0.5022530
#> [4,]  0.03157512  0.1301605  0.7850541  0.8673856
#> [5,] -0.09533969  0.4976634 -0.4318791  0.6980959
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9437392 -1.0630368  1.7674547  0.5152274
#> [2,]  0.1804671  1.3941719 -0.0492724  0.9535243
#> [3,]  0.3661438  0.6111136  0.9306897  1.6574242
#> [4,] -0.3302360 -0.3298129 -0.6737885 -0.9719213
#> [5,] -0.1670921 11.6865196 -2.3451210  5.6354838
#> 
#> , , 1, 2
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.4574505 -0.16445170 -0.4311631  0.01906595
#> [2,]  0.1451562  1.29063512 -0.2095285 -0.09564515
#> [3,] -0.6703655 -0.02355196  0.7763286 -0.04129404
#> [4,]  0.8844780  0.16166606  0.4654238  1.04894603
#> [5,] -1.2629772 -5.08308312 -0.1522606 -0.01324625
#> 
#> , , 2, 2
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.3957159 -0.2019118 -0.22982778 -0.9796903
#> [2,] -0.4953472  1.0473376 -0.02443033  0.6744281
#> [3,]  0.5956645  0.5521272  1.51826320  1.8899892
#> [4,] -0.2155249 -0.3988412 -0.37752755 -0.4348154
#> [5,]  1.2028328 -1.0492028  1.11440773  0.8129129
#> 
#> , , 3, 2
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  1.0622211 -0.59258901  0.73132219  0.04056612
#> [2,] -0.2302702  0.34274060 -0.09032473  0.08873081
#> [3,] -0.3388294 -0.01330027  0.24635122  0.27197182
#> [4,]  0.3047239  0.40415181  0.35650238  0.75786747
#> [5,]  1.0895617 -0.74020753  2.06990908 -1.43090501
#> 
#> , , 4, 2
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,] -0.19061964 -0.04895557 -0.9166149 -1.0246536
#> [2,] -0.04514396  0.67831603 -0.2210838  0.1031589
#> [3,] -0.31065021 -0.49908268  0.4294405  0.3125650
#> [4,]  0.70633569  0.47068719  0.8762973  1.0273997
#> [5,]  0.50487357  3.38204895  0.3882256  1.1944108
#> 
#> , , 5, 2
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.95630738 -0.5657958  0.8450613  0.3919945
#> [2,]  0.16117979  1.1461813  0.4593385  1.2961779
#> [3,]  0.38352306  0.7485877  1.3195997  1.9632295
#> [4,] -0.34991167 -0.5203460 -0.7615463 -1.1810260
#> [5,]  0.03686398  4.7770961  1.4520429  1.8608497
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.4098798 -0.2822880 -0.31398697  0.2229213
#> [2,]  0.0384688  1.0479628  0.08773409 -0.3127667
#> [3,] -0.9693758  0.3118988  0.80888401  0.2526158
#> [4,]  1.1707286 -0.1588717  0.29802685  0.6718893
#> [5,]  0.1165800  0.9107091 -0.09091385  1.3176867
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.4451205 -0.2024004 -0.2789489 -0.7515051
#> [2,]  0.1673597  0.9477791  0.1991396  0.4421698
#> [3,]  0.7980574  0.4265872  1.5404207  1.7305394
#> [4,] -0.4960054 -0.2681891 -0.4698900 -0.3551103
#> [5,] -2.3165482 -1.1911777  4.6876208  3.7933536
#> 
#> , , 3, 3
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.62219382  0.02591212  0.2476382 0.06874454
#> [2,] -0.09728599  0.67884619 -0.2453448 0.17627589
#> [3,] -0.11139855 -0.01808840  0.1948448 0.29255967
#> [4,]  0.29294833  0.08146840  0.6599688 0.66129643
#> [5,] -1.93553702  1.41200680 -1.0514541 1.10560246
#> 
#> , , 4, 3
#> 
#>            [,1]         [,2]       [,3]       [,4]
#> [1,] -0.3617169 -0.231977996 -0.8637519 -0.7043934
#> [2,] -0.1573312  0.766501946 -0.3697051 -0.2329524
#> [3,]  0.1416438 -0.205828804  0.3944621 -0.1693529
#> [4,]  0.3165909  0.292290069  0.8725885  1.3948504
#> [5,]  1.3626612 -0.009518724  2.0732918  1.2226661
#> 
#> , , 5, 3
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.01120411 -1.3098698  1.3534642  0.7552736
#> [2,]  0.19970804  1.3394452  0.7132136  1.2417520
#> [3,]  0.71190476  1.0724227  1.8866029  1.9211508
#> [4,] -0.22504876 -0.4501945 -1.4133642 -1.2577971
#> [5,] -0.52123565  0.9437585 -5.9968499 -0.1622370
#> 
#> , , 1, 4
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.5017176  0.4712920 -0.36348550  0.05746039
#> [2,]  0.2566088  0.6358343  0.06853472  0.02196432
#> [3,] -0.0361106 -0.4062302  1.12808168  0.08359248
#> [4,]  0.2838279  0.2604515  0.02513327  0.89066982
#> [5,] -2.1725238 -2.5283939  0.65885148 -0.64376642
#> 
#> , , 2, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.7944801 -0.1020380  0.4146800 -1.2120832
#> [2,]  0.1801236  0.7026159  0.7210596  0.2745169
#> [3,]  0.6018509  0.2828594  1.7108029  1.4187805
#> [4,] -0.4709960 -0.1404377 -0.8051731  0.1446892
#> [5,] -0.0846959 -0.1607728 -5.7757594 -0.1807108
#> 
#> , , 3, 4
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.9174481 -0.41588609  0.9270658  0.66154709
#> [2,] -0.1137070  0.56686227 -0.2048782  0.09723027
#> [3,] -0.1448914  0.02170848  0.1249546  0.12071365
#> [4,]  0.1931092  0.17198425  0.4471974  0.63700892
#> [5,] -0.5429951  3.81846366 -0.4882073 -1.06362410
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]        [,3]         [,4]
#> [1,] -0.4212855 -0.5356531 -1.30419794 -0.762585244
#> [2,] -0.1294754  0.3788896  0.20077968  0.069972554
#> [3,]  0.2554778  1.0670176  1.54467279  0.008016871
#> [4,]  0.2164032 -0.8624913 -0.08957470  1.318479650
#> [5,]  1.7959241  1.9488472 -0.01979125 -5.122434422
#> 
#> , , 5, 4
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.48687027 -0.6012652  1.6410637  1.1140862
#> [2,] -0.16253175  1.0049580  0.1740567  0.7219372
#> [3,] -0.28131045  0.4473477  0.7414649  0.8259650
#> [4,]  0.05837855 -0.1620266 -0.5496376 -0.4138365
#> [5,]  0.34630730  0.9142114  1.8212356  3.8534746
#> 
#> , , 1, 5
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.85790246 -0.09518379  0.4102741  0.5419987
#> [2,]  0.30132980  1.04409037 -0.3520548 -0.1997036
#> [3,] -0.01863075  0.08201965  0.6956537  0.3539919
#> [4,]  0.02584368 -0.07521715  0.1405609  0.4358527
#> [5,]  0.74031240  2.01633697 -0.2087838  0.3433123
#> 
#> , , 2, 5
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.73237138 -0.1616442  0.04370268 -0.3806302
#> [2,] -0.06508932  1.3947923  0.61588788  1.0691778
#> [3,]  0.38447731  0.4712272  1.41691805  1.5890142
#> [4,] -0.26114171 -0.4954236 -0.52481031 -0.4925649
#> [5,]  2.84030884  5.2996805  2.52834545  4.5073869
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8938487 -0.7069599  0.9514613  0.1706677
#> [2,] -0.2188253  0.7888469 -0.4438760  0.1011326
#> [3,] -0.1322621  0.1285649  0.0336617  0.1973100
#> [4,]  0.1969676  0.2411017  0.5675382  0.7766781
#> [5,]  0.7953254 -0.8949619 -0.6585687 -2.2739604
#> 
#> , , 4, 5
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,] -0.26781681 -0.24371323 -0.9974421 -0.8318256
#> [2,] -0.02630308  0.72530525 -0.2011509  0.3226842
#> [3,] -0.33023629  0.09494476  0.9930083 -0.1521672
#> [4,]  0.74988900  0.04170197  0.3583715  1.4234204
#> [5,]  0.36108208 -3.40645709 -0.1676175  0.3166464
#> 
#> , , 5, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8599212 -0.7676640  1.7052908  0.9475954
#> [2,]  0.1302119  1.0132635  0.1767969  0.8605666
#> [3,]  0.2743667 -0.0458264  0.9693669  1.2764959
#> [4,] -0.2165269  0.3617802 -0.8406916 -0.8394537
#> [5,]  1.4401477  0.6718085  4.1006804  6.9943199
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvars$get_last_draw`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|