R6 Class Representing StartingValuesBVARGROUPPANEL

The class StartingValuesBVARGROUPPANEL presents starting values for the Bayesian hierarchical panel VAR model with country grouping

Super class

bpvars::StartingValuesBVARPANEL -> StartingValuesBVARGROUPPANEL

Public fields

group_allocation

a numeric vector with integer numbers denoting group allocations

A_c

an KxNxC array of starting values for the local parameter \(\mathbf{A}_c\).

Sigma_c

an NxNxC array of starting values for the local parameter \(\mathbf{\Sigma}_c\).

A_g

an KxNxG array of starting values for the group parameter \(\mathbf{A}_g\).

Sigma_g

an NxNxG array of starting values for the group parameter \(\mathbf{\Sigma}_g\).

A

an KxN matrix of starting values for the global parameter \(\mathbf{A}\).

V

an KxK matrix of starting values for the global parameter \(\mathbf{V}\).

Sigma

an NxN matrix of starting values for the global parameter \(\mathbf{\Sigma}\).

nu

a positive scalar with starting values for the global parameter \(\nu\).

m

a positive scalar with starting values for the global hyper-parameter \(m\).

w

a positive scalar with starting values for the global hyper-parameter \(w\).

s

a positive scalar with starting values for the global hyper-parameter \(s\).

Methods

Public methods

specify_starting_values_bvarGroupPANEL$new()

specify_starting_values_bvarGroupPANEL$get_starting_values()

specify_starting_values_bvarGroupPANEL$set_starting_values()

specify_starting_values_bvarGroupPANEL$clone()

Method new()

Create new starting values StartingValuesBVARGROUPPANEL

Usage

specify_starting_values_bvarGroupPANEL$new(
  group_allocation = 1:C,
  C,
  G = C,
  N,
  p,
  d = 0
)

Arguments

group_allocation

a numeric vector with integer numbers denoting group allocations

C

a positive integer - the number of countries in the data.

G

a positive integer specifying the number of country groups.

N

a positive integer - the number of dependent variables in the model.

p

a positive integer - the autoregressive lag order of the SVAR model.

d

a positive integer - the number of exogenous variables in the model.

Returns

Starting values StartingValuesBVARGROUPPANEL

Examples

# starting values for Bayesian Panel VAR 2-country model with 4 lags for a 3-variable system.
sv = specify_starting_values_bvarGroupPANEL$new(C = 2, N = 3, p = 1)

Method get_starting_values()

Returns the elements of the starting values StartingValuesBVARGROUPPANEL as a list.

Usage

specify_starting_values_bvarGroupPANEL$get_starting_values()

Examples

# starting values for a homoskedastic bsvar with 1 lag for a 3-variable system
sv = specify_starting_values_bvarGroupPANEL$new(rep(1,2), C = 2, N = 3, p = 1)
sv$get_starting_values()   # show starting values as list

Method set_starting_values()

Returns the elements of the starting values StartingValuesBVARGROUPPANEL as a list.

Usage

specify_starting_values_bvarGroupPANEL$set_starting_values(last_draw)

Arguments

last_draw

a list containing the same elements as object StartingValuesBVARGROUPPANEL

Returns

An object of class StartingValuesBVARGROUPPANEL including the last draw of the current MCMC as the starting value to be passed to the continuation of the MCMC estimation.

Examples

sv = specify_starting_values_bvarGroupPANEL$new(rep(1,2), C = 2, N = 3, p = 1)

# Modify the starting values by:
sv_list = sv$get_starting_values()   # getting them as list
sv_list$A <- matrix(rnorm(12), 3, 4) # modifying the entry
sv$set_starting_values(sv_list)      # providing to the class object

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_starting_values_bvarGroupPANEL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# starting values for a Bayesian Panel VAR
sv = specify_starting_values_bvarGroupPANEL$new(rep(1,2), C = 2, G = 1, N = 3, p = 1)

## ------------------------------------------------
## Method `specify_starting_values_bvarGroupPANEL$new`
## ------------------------------------------------

# starting values for Bayesian Panel VAR 2-country model with 4 lags for a 3-variable system.
sv = specify_starting_values_bvarGroupPANEL$new(C = 2, N = 3, p = 1)

## ------------------------------------------------
## Method `specify_starting_values_bvarGroupPANEL$get_starting_values`
## ------------------------------------------------

# starting values for a homoskedastic bsvar with 1 lag for a 3-variable system
sv = specify_starting_values_bvarGroupPANEL$new(rep(1,2), C = 2, N = 3, p = 1)
sv$get_starting_values()   # show starting values as list
#> $group_allocation
#> [1] 1 1
#> 
#> $A_c
#> , , 1
#> 
#>               [,1]          [,2]          [,3]
#> [1,] -0.0015409396 -0.0005372022 -0.0019094408
#> [2,]  0.0007129495 -0.0012855869  0.0013784741
#> [3,] -0.0014160534 -0.0003337530  0.0008538241
#> [4,] -0.0002007286  0.0010795019  0.0007465007
#> 
#> , , 2
#> 
#>               [,1]          [,2]          [,3]
#> [1,] -0.0015409396 -0.0005372022 -0.0019094408
#> [2,]  0.0007129495 -0.0012855869  0.0013784741
#> [3,] -0.0014160534 -0.0003337530  0.0008538241
#> [4,] -0.0002007286  0.0010795019  0.0007465007
#> 
#> 
#> $Sigma_c
#> , , 1
#> 
#>             [,1]       [,2]        [,3]
#> [1,]  1.73194256 0.08286903 -0.03017792
#> [2,]  0.08286903 3.48489997  1.73578114
#> [3,] -0.03017792 1.73578114  2.57768388
#> 
#> , , 2
#> 
#>             [,1]       [,2]        [,3]
#> [1,]  1.73194256 0.08286903 -0.03017792
#> [2,]  0.08286903 3.48489997  1.73578114
#> [3,] -0.03017792 1.73578114  2.57768388
#> 
#> 
#> $A_g
#> , , 1
#> 
#>               [,1]          [,2]          [,3]
#> [1,] -0.0015409396 -0.0005372022 -0.0019094408
#> [2,]  0.0007129495 -0.0012855869  0.0013784741
#> [3,] -0.0014160534 -0.0003337530  0.0008538241
#> [4,] -0.0002007286  0.0010795019  0.0007465007
#> 
#> , , 2
#> 
#>               [,1]          [,2]          [,3]
#> [1,]  0.0007635323  9.303441e-04 -2.131131e-04
#> [2,] -0.0005668458 -6.262695e-05  7.818445e-04
#> [3,] -0.0001816927  1.617785e-03  2.419852e-04
#> [4,] -0.0013807560 -1.250350e-03 -7.192453e-05
#> 
#> 
#> $Sigma_g
#> , , 1
#> 
#>             [,1]       [,2]        [,3]
#> [1,]  1.73194256 0.08286903 -0.03017792
#> [2,]  0.08286903 3.48489997  1.73578114
#> [3,] -0.03017792 1.73578114  2.57768388
#> 
#> , , 2
#> 
#>           [,1]       [,2]       [,3]
#> [1,] 0.3965792  0.4508118  0.3028781
#> [2,] 0.4508118  8.2418577 -8.9290791
#> [3,] 0.3028781 -8.9290791 13.0557789
#> 
#> 
#> $A
#>               [,1]          [,2]          [,3]
#> [1,]  0.9997134700 -0.0002411063 -0.0012610662
#> [2,] -0.0014597641  1.0020769458 -0.0006352611
#> [3,] -0.0004252436  0.0010970582  0.9996156789
#> [4,]  0.0009898912 -0.0006540818 -0.0009220008
#> 
#> $V
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.6671702  0.7124846 -0.4991816  0.2553879
#> [2,]  0.7124846  2.5271199 -0.5073277 -0.2193948
#> [3,] -0.4991816 -0.5073277  2.1018716 -1.3151244
#> [4,]  0.2553879 -0.2193948 -1.3151244  3.5336773
#> 
#> $Sigma
#>           [,1]      [,2]      [,3]
#> [1,]  4.079553 -2.051535  2.544386
#> [2,] -2.051535  1.364992 -1.078083
#> [3,]  2.544386 -1.078083  2.499324
#> 
#> $nu
#> [1] 4.1
#> 
#> $m
#> [1] 4.194018e-05
#> 
#> $w
#> [1] 0.1077431
#> 
#> $s
#> [1] 0.2912586
#> 

## ------------------------------------------------
## Method `specify_starting_values_bvarGroupPANEL$set_starting_values`
## ------------------------------------------------

sv = specify_starting_values_bvarGroupPANEL$new(rep(1,2), C = 2, N = 3, p = 1)

# Modify the starting values by:
sv_list = sv$get_starting_values()   # getting them as list
sv_list$A <- matrix(rnorm(12), 3, 4) # modifying the entry
sv$set_starting_values(sv_list)      # providing to the class object