Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARGROUPPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARGROUPPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarGroupPANEL$new(
      data = ilo_dynamic_panel[1:5],
      exogenous = ilo_exogenous_variables[1:5],
      group_allocation = country_grouping_region[1:5]
)
#> Country groupings have been pre-specified and will not be estimated.
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004747848 0.0004493699 0.004141264  0.005084028
#> 
#> $AFG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01475498 0.004270623 -0.01972869 -0.009880682
#> Sigma[2,2]  0.22152997 0.037072859  0.18235823  0.260582232
#> 
#> $AFG$Sigma$equation3
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.01967969 0.004083247  0.01503024   0.02452986
#> Sigma[3,2] -0.21291646 0.053073936 -0.28099863  -0.16659121
#> Sigma[3,3]  0.35029710 0.077759079  0.28563154   0.45306928
#> 
#> $AFG$Sigma$equation4
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[4,1]  0.01499869 0.005283967  0.01037151   0.02196077
#> Sigma[4,2] -0.13335131 0.047134701 -0.19467119  -0.09459729
#> Sigma[4,3]  0.29708913 0.068985244  0.24143957   0.38981819
#> Sigma[4,4]  0.27672511 0.061572771  0.23121676   0.35860774
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.99145439 0.03295184  0.96070143   1.03432948
#> lag1_var2 -0.06330932 0.13951377 -0.21937859   0.07020405
#> lag1_var3 -0.10288437 0.28517128 -0.41562071   0.19173439
#> lag1_var4  0.08643451 0.24326940 -0.16955160   0.35108937
#> const      1.15165636 2.22309155 -1.46707747   3.48124340
#> exo1      -0.03884395 0.06437183 -0.09252035   0.04709254
#> exo2      -0.03078427 0.01247129 -0.04329334  -0.01508454
#> exo3      -0.16996736 0.06818938 -0.23987292  -0.08364285
#> 
#> $AFG$A$equation2
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.386688440  0.2274781   0.1934632    0.6912645
#> lag1_var2 -0.084572981  0.5972852  -0.6372705    0.6914842
#> lag1_var3 -2.279534860  1.3741437  -3.3998228   -0.4165842
#> lag1_var4  1.975906414  1.2036162   0.3418453    2.9594813
#> const      5.596341669 11.8478786 -10.4559401   14.7671787
#> exo1       0.009429942  0.3437059  -0.4079272    0.3889412
#> exo2      -0.031225935  0.2714889  -0.3538072    0.2787550
#> exo3      -0.478640675  0.5805724  -1.0964989    0.1651303
#> 
#> $AFG$A$equation3
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  -0.2718312  0.1555594  -0.42557308  -0.07852477
#> lag1_var2   3.4451248  1.0403731   2.09795725   4.45087804
#> lag1_var3   9.0962953  2.2368764   6.12477522  11.02148901
#> lag1_var4  -7.1752990  1.9300670  -8.76516853  -4.58906082
#> const     -34.2069470 16.4466087 -49.65417131 -13.39982871
#> exo1       -0.1008959  0.7910853  -0.95412192   0.80580692
#> exo2       -0.4069909  0.2483299  -0.66888123  -0.09927514
#> exo3        0.5589052  0.6005147   0.02978077   1.33566915
#> 
#> $AFG$A$equation4
#>                   mean          sd 5% quantile  95% quantile
#> lag1_var1  -0.12438018  0.09831868  -0.2133199   0.001138708
#> lag1_var2   3.85700615  0.99580044   2.6442769   4.904688182
#> lag1_var3   9.08182239  2.08396947   6.4820536  11.145259592
#> lag1_var4  -7.04144214  1.78709468  -8.7633346  -4.785283883
#> const     -42.72936883 14.71692613 -58.5383690 -26.046664612
#> exo1       -0.09628826  0.75953230  -0.9698077   0.743484919
#> exo2       -0.39184849  0.29749428  -0.7183425  -0.054757614
#> exo3        0.49652018  0.55268526   0.1870822   1.257157419
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                  mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00765144 0.004739351  0.00425303   0.01412054
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01262057 0.008366145 -0.02293090 -0.004021584
#> Sigma[2,2]  0.11676124 0.031843031  0.07696252  0.147784259
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.008626688 0.004121839  0.004244737   0.01332385
#> Sigma[3,2] -0.050137862 0.017953257 -0.070345852  -0.02869507
#> Sigma[3,3]  0.047934380 0.013287950  0.035409315   0.06503162
#> 
#> $AGO$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1] -0.001716404 0.009844907 -0.0151498381  0.004196894
#> Sigma[4,2]  0.046611415 0.019158626  0.0347976110  0.072676946
#> Sigma[4,3]  0.009598597 0.006824826  0.0009118129  0.016083134
#> Sigma[4,4]  0.052325142 0.012929256  0.0424615230  0.069645094
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.01150930 0.02906328  0.98839961   1.05098623
#> lag1_var2 -0.34518526 0.19403741 -0.56423641  -0.11230261
#> lag1_var3 -0.41435303 0.24610497 -0.72791879  -0.16343110
#> lag1_var4  0.39170752 0.22422705  0.15620496   0.67350207
#> const      1.93484874 1.24432357  0.30821710   3.12180218
#> exo1      -0.02537156 0.06240343 -0.08846077   0.05037035
#> exo2      -0.02639566 0.04948782 -0.06578256   0.03985860
#> exo3       0.12202226 0.06705635  0.05100097   0.20219119
#> 
#> $AGO$A$equation2
#>                 mean         sd  5% quantile 95% quantile
#> lag1_var1 -0.1099610  0.2320959  -0.40310012    0.1094408
#> lag1_var2  1.2561251  0.8599006   0.29813228    2.2424330
#> lag1_var3  0.4132686  1.0669226  -0.71583409    1.6623040
#> lag1_var4 -0.3576548  0.9275448  -1.48661528    0.5937147
#> const     -0.6381636 15.5889901 -12.44201548   20.4266312
#> exo1       0.2884669  0.2464211   0.11485993    0.6246347
#> exo2       0.3019637  0.1788114   0.08388915    0.4840690
#> exo3      -0.6094828  0.5809916  -1.13975628    0.1610182
#> 
#> $AGO$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.16981631 0.1531217 -0.27694233   0.04001016
#> lag1_var2 -0.18852010 0.1726667 -0.33574389   0.03967012
#> lag1_var3  0.63291228 0.2664792  0.36249205   0.93089520
#> lag1_var4  0.24903279 0.2221175 -0.02340476   0.47448409
#> const     11.71454525 5.3663126  6.60579848  18.68511037
#> exo1      -0.02167807 0.1595476 -0.23029417   0.10437059
#> exo2      -1.51540491 0.2221051 -1.77506016  -1.28692326
#> exo3       3.01456564 0.2070010  2.81842458   3.26532430
#> 
#> $AGO$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.2997332  0.0749547  -0.3855629   -0.2199058
#> lag1_var2  0.1930106  0.8089066  -0.7574982    1.1258965
#> lag1_var3  0.1899311  1.0432311  -1.0059989    1.4169911
#> lag1_var4  0.7687687  0.9166167  -0.3237594    1.7934326
#> const      9.7027522 11.0186390  -2.2147976   22.9824958
#> exo1       0.2328238  0.2036338  -0.0217319    0.4415223
#> exo2      -1.5208656  0.2533775  -1.7628135   -1.2001351
#> exo3       3.0863447  0.4622567   2.6607700    3.6668337
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003039629 0.001059463 0.001839471  0.003996013
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05111084 0.01081939 -0.06182851  -0.03747855
#> Sigma[2,2]  7.16026190 1.57793343  5.83832345   9.28177283
#> 
#> $ALB$Sigma$equation3
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.0333467 0.02408995  0.01704197   0.06653422
#> Sigma[3,2] -5.1463114 1.18084441 -6.43970666  -3.85992093
#> Sigma[3,3]  7.0971076 1.47358278  5.07479430   8.18212125
#> 
#> $ALB$Sigma$equation4
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.002750457 0.02266159 -0.01896871   0.02835593
#> Sigma[4,2] -0.384315842 0.67524715 -1.16287109   0.37820802
#> Sigma[4,3]  4.327485385 1.59969131  2.52490664   6.23404056
#> Sigma[4,4]  4.867416328 2.06993817  2.70826387   7.44965582
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.97091917 0.05294181  0.90299212   1.01954748
#> lag1_var2 -0.03420566 0.02409376 -0.06257235  -0.01204728
#> lag1_var3 -0.05543415 0.03077092 -0.09058985  -0.02516046
#> lag1_var4  0.04556275 0.02479893  0.02218302   0.07466563
#> const      1.31172838 1.60765981 -0.09224158   3.33059140
#> exo1       0.02359124 0.05739124 -0.03064430   0.08813124
#> exo2      -0.10048154 0.06597572 -0.15607369  -0.01482966
#> exo3       0.05634907 0.03721712  0.01486973   0.09705116
#> 
#> $ALB$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -2.9078605  2.212324   -5.892181  -1.04405661
#> lag1_var2  2.1291961  3.377730   -1.375718   6.38058671
#> lag1_var3  2.7636190  5.007202   -2.186890   9.22561795
#> lag1_var4 -2.2464105  4.189476   -7.702755   1.80366465
#> const     44.0438895 90.349189  -17.335480 167.55538066
#> exo1      -3.0584880  2.414242   -5.326584  -0.30150547
#> exo2      -0.5864603  4.002155   -5.208270   3.41371243
#> exo3      -0.9257012  1.134081   -2.481344  -0.09672819
#> 
#> $ALB$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.32706191  2.5380013  -2.2785672    3.5204150
#> lag1_var2 -1.10641310  3.2517395  -5.2147428    1.9896494
#> lag1_var3 -0.93417695  4.7072887  -6.7639354    3.9164673
#> lag1_var4  1.47407276  4.0035712  -2.7506203    6.3849145
#> const     19.01021734 98.8183367 -90.0627133  136.1069608
#> exo1      -0.52237758  3.3412071  -4.3754854    3.0731425
#> exo2      -0.44928551  1.1443372  -1.6700011    0.9760711
#> exo3      -0.07882348  0.7580192  -0.9928911    0.7028114
#> 
#> $ALB$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -1.83495968  2.2526135   -4.849711  -0.03652074
#> lag1_var2 -0.04197494  1.8987359   -1.756628   2.43752261
#> lag1_var3  0.47266799  2.8560938   -1.894089   4.28922878
#> lag1_var4  0.49355155  2.4909258   -2.823999   2.61191028
#> const     49.01304615 65.7362919   13.010315 138.87431144
#> exo1      -2.53244213  3.7899966   -7.510826   0.66522727
#> exo2      -0.87564111  3.2053489   -4.601097   2.61333547
#> exo3      -0.84594093  0.7479812   -1.557303   0.09023177
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                  mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.01369105 0.008696326 0.006251705     0.025202
#> 
#> $ARE$Sigma$equation2
#>                    mean         sd  5% quantile 95% quantile
#> Sigma[2,1] -0.003755451 0.00344399 -0.006917773 0.0002686858
#> Sigma[2,2]  0.055465422 0.00714921  0.048676542 0.0643655105
#> 
#> $ARE$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04077568 0.01960973  0.02026747   0.06549975
#> Sigma[3,2] -0.05809951 0.01999569 -0.08297365  -0.03745303
#> Sigma[3,3]  0.25916109 0.04888235  0.20960228   0.32156079
#> 
#> $ARE$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1]  0.03636327 0.01756921  0.01874521  0.058713143
#> Sigma[4,2] -0.01514239 0.01849554 -0.03583805  0.006880967
#> Sigma[4,3]  0.21236200 0.04968174  0.16212839  0.273935489
#> Sigma[4,4]  0.19954737 0.05298600  0.14461010  0.257967602
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.64672947 0.1907557   0.5061125   0.90814149
#> lag1_var2 -1.47568627 0.1646288  -1.6222039  -1.26398653
#> lag1_var3 -2.06627343 0.3163745  -2.2873536  -1.63706481
#> lag1_var4  2.07671913 0.3403769   1.6107324   2.32171875
#> const      8.28531502 3.2049984   4.5001945  11.70179287
#> exo1      -0.02233886 0.1238904  -0.1305979   0.14121433
#> exo2      -0.16930748 0.1954882  -0.4063049   0.02469001
#> exo3      -0.33710131 0.2857008  -0.6298774   0.02658863
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.02794550 0.2473766  -0.3127855    0.1996702
#> lag1_var2  1.58953134 0.6833964   0.9752164    2.4707849
#> lag1_var3  1.13705870 0.8854033   0.3190221    2.2610195
#> lag1_var4 -1.13741766 0.8058885  -2.1288725   -0.3776071
#> const     -0.07582523 2.4560855  -2.7539592    2.5013836
#> exo1       0.11584893 0.3530438  -0.3429879    0.4456350
#> exo2       1.95506030 0.3501532   1.4801745    2.1962906
#> exo3      -0.64452635 0.4212617  -1.1665927   -0.2225164
#> 
#> $ARE$A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.8482879  0.7213603   -1.702848  -0.10428202
#> lag1_var2 -2.1479601  1.5310638   -4.215715  -0.93084740
#> lag1_var3 -2.5999223  1.9774288   -5.171280  -0.82709782
#> lag1_var4  3.6880002  1.8737270    1.843414   6.02132876
#> const     14.4065592 11.6368141    2.012891  27.13375941
#> exo1      -0.6323318  0.5687529   -1.222432   0.01612848
#> exo2      -2.6121311  0.9340927   -3.757811  -1.60698708
#> exo3      -2.3555555  0.7843355   -3.078952  -1.34928715
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.7858106 0.6513195   -1.510842  -0.02009189
#> lag1_var2 -1.2855485 1.0947181   -2.731765  -0.42307422
#> lag1_var3 -2.1819578 1.3739299   -3.963645  -1.10304282
#> lag1_var4  3.2779279 1.3253782    2.086259   4.92669855
#> const     12.6658049 9.9100283    1.843401  23.57597062
#> exo1      -0.5478404 0.4266389   -1.042526  -0.05377461
#> exo2      -1.0792687 0.7421157   -2.055609  -0.41938874
#> exo3      -2.8799692 0.6210749   -3.411448  -2.07048125
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003650703 0.0006198133 0.003037829  0.004435629
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05781621 0.03277961 -0.09900644  -0.02475701
#> Sigma[2,2]  3.67433356 1.26648123  2.13789107   5.07653661
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.05092514 0.01804343  0.03552437   0.07494256
#> Sigma[3,2] -2.60037725 1.06878193 -3.88900221  -1.41304786
#> Sigma[3,3]  2.16743658 0.76227154  1.50029430   3.18697926
#> 
#> $ARG$Sigma$equation4
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01769357 0.01027607  0.005181429   0.02885156
#> Sigma[4,2] -0.41047263 0.38103197 -0.914321623  -0.11137542
#> Sigma[4,3]  0.65067521 0.23415521  0.412831344   0.94683825
#> Sigma[4,4]  0.44363039 0.19430089  0.264455914   0.69185936
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                    mean         sd 5% quantile 95% quantile
#> lag1_var1  0.9156395412 0.05505155  0.84839716   0.97513394
#> lag1_var2 -0.2335692210 0.09967239 -0.35044147  -0.13474190
#> lag1_var3 -0.3778511188 0.16197558 -0.57153593  -0.22002314
#> lag1_var4  0.3347151911 0.14709239  0.19205399   0.50521203
#> const      4.9431830126 1.64651701  3.09238034   6.50824210
#> exo1      -0.0004412172 0.09780332 -0.09638981   0.11452734
#> exo2      -0.1304888300 0.04181580 -0.17200776  -0.08397161
#> exo3       0.0740631578 0.05910393  0.01538962   0.13955890
#> 
#> $ARG$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.48228473  1.540673  -2.2183887     1.085108
#> lag1_var2  0.82443137  3.650975  -2.2876983     5.673746
#> lag1_var3  0.03977008  5.900852  -5.0676177     7.857785
#> lag1_var4  0.34552303  5.366071  -6.7308251     5.118098
#> const     -8.24970329 65.997031 -88.1723784    63.604795
#> exo1      -0.89762765  2.387152  -3.3290205     2.019714
#> exo2       2.47279895  1.966889   0.2475602     4.615086
#> exo3      -0.21936958  4.062812  -4.8606597     4.265227
#> 
#> $ARG$A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.6968203  0.9515938  -1.5370663     0.534049
#> lag1_var2  0.8569030  3.3716357  -2.5975049     4.955488
#> lag1_var3  2.2228901  5.4795585  -3.3242490     8.934142
#> lag1_var4 -1.5611872  4.9488262  -7.6496493     3.460224
#> const     38.1469442 44.2235653 -15.1907718    87.199619
#> exo1      -0.1606458  2.1045045  -2.7764928     1.813127
#> exo2      -5.3983893  1.6513288  -7.2822565    -3.631718
#> exo3       2.3717149  2.6855235  -0.3961661     5.735830
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -1.1188962  0.3038696  -1.4839611  -0.79171424
#> lag1_var2  1.1988489  2.1296560  -1.2986552   3.60544316
#> lag1_var3  1.9853635  3.4843122  -2.0906244   5.91665856
#> lag1_var4 -1.0331315  3.0647625  -4.5136259   2.54294204
#> const     33.1937405 32.7020139  -0.5966133  74.39877815
#> exo1      -0.7986993  0.7557194  -1.7316195  -0.02140263
#> exo2      -4.4164888  0.6238319  -5.0792711  -3.63748993
#> exo3       2.5887679  0.5248872   2.1175753   3.27672418
#> 
#> 
#>