Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003345663 0.0005145532 0.002928516  0.004043494
#> 
#> $AFG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01203394 0.003370767 -0.01661382 -0.009927392
#> Sigma[2,2]  0.19446311 0.024612958  0.17262434  0.223674705
#> 
#> $AFG$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01169155 0.007045267  0.004498448   0.01969644
#> Sigma[3,2] -0.17355463 0.054821916 -0.237652805  -0.10985820
#> Sigma[3,3]  0.30316476 0.089418368  0.189834520   0.39521298
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.007313399 0.006878802 -8.210429e-05   0.01436736
#> Sigma[4,2] -0.101881303 0.051970243 -1.612930e-01  -0.04084082
#> Sigma[4,3]  0.259150467 0.076450932  1.611477e-01   0.33623009
#> Sigma[4,4]  0.244909836 0.065012664  1.618625e-01   0.30630458
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                    mean         sd 5% quantile 95% quantile
#> lag1_var1  0.9985778161 0.01649534  0.98719085  1.021078503
#> lag1_var2 -0.0239631236 0.02634927 -0.04689331  0.007204474
#> lag1_var3 -0.0059181155 0.04151642 -0.03856481  0.041820414
#> lag1_var4 -0.0005271563 0.03029355 -0.03393144  0.023135282
#> const      0.5727217904 0.70788819 -0.22795860  1.358726226
#> exo1      -0.0867287454 0.06119244 -0.15214626 -0.017002270
#> exo2      -0.0354577563 0.06164998 -0.08653931  0.043180079
#> exo3      -0.1692678076 0.04556548 -0.20788373 -0.108599071
#> 
#> $AFG$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.3486962 0.1980872   0.1457631    0.5927404
#> lag1_var2  0.8125076 0.1166209   0.6617815    0.9168461
#> lag1_var3 -0.3129364 0.1744279  -0.5409640   -0.1603587
#> lag1_var4  0.2524491 0.1431609   0.1088617    0.4240197
#> const     -4.6967289 2.0243941  -6.7103982   -2.1543854
#> exo1      -0.1028266 0.6731728  -0.8076954    0.6459376
#> exo2       0.2290483 0.5688484  -0.5044143    0.7843783
#> exo3      -0.1008671 0.2712727  -0.3604647    0.2162007
#> 
#> $AFG$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.15925794 0.2135574 -0.35997539    0.1117671
#> lag1_var2  0.19988823 0.2520933 -0.08534498    0.4965598
#> lag1_var3  2.10675051 0.4629557  1.66908652    2.6842934
#> lag1_var4 -1.09337721 0.3650652 -1.54985618   -0.7666351
#> const      5.41420363 5.1497622 -1.06967540   10.2940395
#> exo1       0.30315877 0.5091107 -0.14589205    0.9040768
#> exo2      -1.12145193 0.8421020 -2.00943816   -0.2567281
#> exo3       0.03534498 0.6352737 -0.66996222    0.7887294
#> 
#> $AFG$A$equation4
#>                   mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.002830895 0.1710319  -0.1659539    0.2198327
#> lag1_var2  0.693199905 0.2459674   0.3933905    0.9624698
#> lag1_var3  2.286969211 0.4607673   1.8205732    2.8466844
#> lag1_var4 -1.131943388 0.3634626  -1.5776280   -0.7899071
#> const     -4.273750431 5.6710544 -10.7438618    1.4788514
#> exo1       0.305353104 0.3748363  -0.1439433    0.6744012
#> exo2      -1.075837797 0.7338610  -1.9165711   -0.4353683
#> exo3       0.093236888 0.6639244  -0.5914395    0.8683432
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.006929253 0.000393589 0.006529791  0.007379355
#> 
#> $AGO$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01036123 0.01630943 -0.02920878   0.00593206
#> Sigma[2,2]  0.19166461 0.07269097  0.11539926   0.27389520
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[3,1]  0.008533547 0.006867097  0.0006183718   0.01530830
#> Sigma[3,2] -0.082042612 0.035478071 -0.1199741034  -0.04111826
#> Sigma[3,3]  0.047382341 0.015831350  0.0285736872   0.06295752
#> 
#> $AGO$Sigma$equation4
#>                     mean          sd 5% quantile 95% quantile
#> Sigma[4,1]  0.0009919334 0.006729443 -0.00771327  0.007007182
#> Sigma[4,2]  0.0721971091 0.024139560  0.04906565  0.098923413
#> Sigma[4,3] -0.0162673202 0.013240912 -0.03118675 -0.002504867
#> Sigma[4,4]  0.0451451144 0.008416644  0.03353889  0.050871491
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.99980181 0.03618093  0.95873288   1.04177513
#> lag1_var2 -0.29788501 0.20004766 -0.53365495  -0.11713742
#> lag1_var3 -0.35949896 0.20150299 -0.59004513  -0.15650789
#> lag1_var4  0.32465528 0.18854267  0.13216065   0.54105874
#> const      3.07314188 2.42131869  0.09030288   5.41964690
#> exo1       0.02777815 0.12128506 -0.08095408   0.18518459
#> exo2      -0.03660928 0.11354263 -0.16779869   0.07840439
#> exo3       0.04433988 0.11020940 -0.10002567   0.14862223
#> 
#> $AGO$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.07488528 0.2181926 -0.27856349    0.2091990
#> lag1_var2 -0.11910704 1.3749874 -1.90558454    1.1207136
#> lag1_var3 -1.10417139 2.0891386 -3.81415531    0.9039473
#> lag1_var4  1.14204380 1.8850666 -0.62764131    3.6068890
#> const      3.25898451 7.8434614 -6.98506874   10.1884832
#> exo1       0.20747628 0.3871939 -0.28449732    0.5870355
#> exo2       0.61313572 0.6991765  0.04690725    1.5463199
#> exo3       0.22095534 0.6059057 -0.26441096    1.0425193
#> 
#> $AGO$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.14262666 0.1154900  -0.3003248  -0.05771927
#> lag1_var2 -0.42441728 0.8575850  -1.1153307   0.73912078
#> lag1_var3  0.29931696 1.2822296  -0.8000999   2.01908795
#> lag1_var4  0.52516251 1.1415977  -1.0211322   1.45861177
#> const     15.08661347 7.3969655   7.8517188  24.09074384
#> exo1      -0.03161534 0.2080174  -0.2939826   0.13648789
#> exo2      -1.49567759 0.4080996  -2.0224596  -1.16505643
#> exo3       3.03608951 0.3825710   2.6067698   3.47098566
#> 
#> $AGO$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.2588154 0.0673849 -0.34635226   -0.1997168
#> lag1_var2 -1.0265552 0.4311280 -1.55270385   -0.5682031
#> lag1_var3 -1.2055868 0.5621916 -1.81861315   -0.5860779
#> lag1_var4  2.1028753 0.5142699  1.54331841    2.6476095
#> const     15.7623676 4.7338457 11.31461743   21.8719322
#> exo1       0.1787720 0.1711950 -0.02143048    0.3733657
#> exo2      -1.2920753 0.1731396 -1.43539643   -1.0578571
#> exo3       3.7124211 0.3611362  3.36678278    4.1115463
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001976959 0.0005756137 0.001319462  0.002643628
#> 
#> $ALB$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0743515 0.03837756  -0.1165691  -0.03358203
#> Sigma[2,2]  8.3483716 3.98059665   4.2577794  13.35958212
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[3,1]  0.03368402 0.02589469  0.008564432   0.06580807
#> Sigma[3,2] -4.86466080 2.91425582 -8.828290749  -2.64188029
#> Sigma[3,3]  4.65474711 1.91880970  3.152070188   7.26461320
#> 
#> $ALB$Sigma$equation4
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.0193655 0.03232668 -0.06361285  0.001087876
#> Sigma[4,2]  0.8553964 1.73216744 -0.13027308  3.234216428
#> Sigma[4,3]  1.6775160 0.60900115  1.10743627  2.476045694
#> Sigma[4,4]  2.7040704 1.98929162  1.44129028  5.385858264
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.99457506 0.04268640  0.943148404  1.039344386
#> lag1_var2 -0.01424055 0.02262312 -0.038561956  0.013663988
#> lag1_var3 -0.03306875 0.03635777 -0.074639762  0.009448043
#> lag1_var4  0.02720680 0.03098677 -0.008261695  0.063257294
#> const      0.43163051 1.13575006 -0.707314332  1.857808907
#> exo1       0.01940854 0.03754986 -0.024834950  0.057629719
#> exo2      -0.06996750 0.04441517 -0.101890342 -0.009527762
#> exo3       0.09081788 0.04422292  0.032175663  0.128054025
#> 
#> $ALB$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  -5.5719155  2.857064  -8.6428500   -2.9714217
#> lag1_var2   1.7255539  1.567409  -0.4386390    2.7559759
#> lag1_var3   2.3505904  2.383063  -0.9274945    4.0701415
#> lag1_var4  -2.0733757  1.950676  -3.6967608    0.5523897
#> const     121.2886333 96.158101  33.0077089  238.5961455
#> exo1       -2.2364411  3.188713  -5.2915129    1.8762533
#> exo2       -0.4425358  2.257161  -2.6649246    2.2532713
#> exo3       -1.6497727  3.278295  -4.8429073    2.5543383
#> 
#> $ALB$A$equation3
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1   3.7874269  2.056029    1.8511945 6.0196577849
#> lag1_var2  -1.4486678  1.526848   -3.4495259 0.0006091368
#> lag1_var3  -1.8272105  2.231750   -4.7804585 0.2117381730
#> lag1_var4   2.3212215  1.857171    0.6949469 4.8077835375
#> const     -60.2403888 73.414790 -143.5522433 9.0285485808
#> exo1        0.7548715  2.113473   -1.9097160 2.7743981496
#> exo2       -1.4687815  1.840268   -2.5284988 1.0786366116
#> exo3        1.4428727  2.354329   -0.7307470 4.5087697924
#> 
#> $ALB$A$equation4
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.07597276  0.6024797  -0.64832076   0.74931922
#> lag1_var2 -0.79263303  1.4174747  -2.64722779   0.55767644
#> lag1_var3 -0.93394459  2.1424215  -3.71708066   1.14551493
#> lag1_var4  1.65393105  1.7260026  -0.09356771   3.86288656
#> const     18.42859491 38.4116133 -15.03252364  68.96404143
#> exo1      -0.38355424  0.3660738  -0.69339962   0.05802926
#> exo2      -2.04277936  1.9805056  -4.54954451  -0.15569950
#> exo3       0.41298050  1.4829321  -1.09620353   2.01032020
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001899643 0.0002327134 0.001625878  0.002162903
#> 
#> $ARE$Sigma$equation2
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.0005830021 0.001305726 -0.001621049  0.001123371
#> Sigma[2,2]  0.0583586327 0.017536143  0.041021588  0.079996615
#> 
#> $ARE$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.00770528 0.003939926  0.003990655   0.01203696
#> Sigma[3,2] -0.02933816 0.010980644 -0.042305923  -0.01724916
#> Sigma[3,3]  0.15596509 0.035306158  0.117539894   0.19525577
#> 
#> $ARE$Sigma$equation4
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[4,1] 0.006929943 0.003341586 0.003529866    0.0104685
#> Sigma[4,2] 0.015463967 0.019751301 0.001629080    0.0425871
#> Sigma[4,3] 0.132547358 0.031238906 0.096207028    0.1656067
#> Sigma[4,4] 0.144253785 0.028031024 0.109077159    0.1722808
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  1.008204402 0.01987269  0.99090190  1.033469937
#> lag1_var2 -0.290615802 0.06207795 -0.36906118 -0.231455041
#> lag1_var3 -0.389490883 0.06897412 -0.46561890 -0.313154147
#> lag1_var4  0.367192830 0.07313612  0.29130767  0.448607649
#> const      1.514914267 0.45757951  0.93841394  1.972862993
#> exo1      -0.024201038 0.04134405 -0.06761018  0.018657169
#> exo2      -0.056554808 0.04503280 -0.10285553 -0.005935168
#> exo3       0.001158729 0.10023446 -0.11798728  0.092054560
#> 
#> $ARE$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.2105330 0.4061464  -0.1692726    0.7120089
#> lag1_var2 -0.3605487 0.6802842  -1.2869958    0.1349534
#> lag1_var3 -1.3637048 0.9902648  -2.6986647   -0.8727756
#> lag1_var4  1.2365410 1.0058963   0.7177060    2.5990193
#> const      4.9019830 5.7400343  -0.7440931   11.8407630
#> exo1       0.1102384 0.2532465  -0.1389710    0.4368133
#> exo2       2.1164819 0.3118745   1.7301344    2.4188080
#> exo3      -0.2493589 0.4596102  -0.7245943    0.2495064
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.6847675 0.4809668  0.06170715   1.09872825
#> lag1_var2 -2.5430147 0.6580002 -3.24592811  -1.80267259
#> lag1_var3 -2.3946713 0.9524160 -3.33135277  -1.20379179
#> lag1_var4  3.1649276 1.0454265  1.93028654   4.29404771
#> const     -0.3305462 2.9320980 -3.77630950   3.04928899
#> exo1      -0.1979098 0.2522541 -0.46725781   0.08815491
#> exo2      -2.1724524 0.8245385 -2.85766596  -1.05730373
#> exo3      -1.1468155 0.5898490 -1.64453984  -0.36118298
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.8299440 0.3822594   0.3579437    1.1692982
#> lag1_var2 -2.9558298 0.7696822  -3.6604244   -1.9977352
#> lag1_var3 -3.6212196 1.1961656  -4.7151348   -2.0398726
#> lag1_var4  4.3249067 1.2399506   2.6788746    5.3931694
#> const      1.4831562 4.4774698  -2.3822747    7.3273737
#> exo1      -0.1393701 0.2750244  -0.4923456    0.1238360
#> exo2      -0.5572339 0.6131071  -1.1536473    0.2355733
#> exo3      -1.4002011 0.8524966  -1.9507186   -0.2233729
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003575866 0.000964048 0.002606744  0.004679289
#> 
#> $ARG$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0372069 0.01995582 -0.05878211   -0.0146606
#> Sigma[2,2]  2.9094052 0.44365666  2.41818874    3.4309221
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.03564447 0.01180347  0.02234517   0.04983123
#> Sigma[3,2] -1.92391755 0.26231186 -2.12437109  -1.56638913
#> Sigma[3,3]  1.55166290 0.23254213  1.27221646   1.81314013
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01418815 0.009635601  0.007431916   0.02734204
#> Sigma[4,2] -0.17582643 0.159727322 -0.389733492  -0.04373148
#> Sigma[4,3]  0.42686633 0.135901107  0.334204499   0.61247644
#> Sigma[4,4]  0.35710345 0.071334805  0.277830740   0.43400893
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.89478598 0.04428846  0.84726512   0.94948926
#> lag1_var2 -0.18966014 0.17605554 -0.41995535  -0.02827328
#> lag1_var3 -0.30370568 0.29270945 -0.68961519  -0.04160766
#> lag1_var4  0.26495201 0.25357965  0.03847371   0.59959837
#> const      5.25401285 2.04987534  2.79478329   7.29523060
#> exo1       0.04005825 0.05092833 -0.01429753   0.10309839
#> exo2      -0.06495713 0.04744094 -0.12518075  -0.01829626
#> exo3       0.07424270 0.09978207 -0.05037951   0.17799166
#> 
#> $ARG$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  -1.7095465  1.534877  -2.9987278   0.34140255
#> lag1_var2  -5.7170706  4.601704 -10.5055754   0.08882724
#> lag1_var3 -10.3849512  7.553865 -18.2378206  -0.85811153
#> lag1_var4  10.0477624  6.773294   1.4539121  17.02000542
#> const      70.4870278 18.465101  47.3988751  89.88903022
#> exo1       -0.8652224  1.683753  -3.0529364   0.73704686
#> exo2        1.7815887  2.323532  -0.5446522   4.57761222
#> exo3        2.1869156  1.401283   0.3361818   3.43015762
#> 
#> $ARG$A$equation3
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1   0.4667111  1.422864  -1.29130317    1.8282903
#> lag1_var2   4.7918075  3.975461   0.01748350    9.1908681
#> lag1_var3   8.4533793  6.565646   0.55072803   15.7048915
#> lag1_var4  -7.3518429  5.874128 -13.79921558   -0.2389283
#> const     -21.2549360 23.217998 -43.81094872    6.9247752
#> exo1        0.4442849  1.197408  -0.92725470    1.6769855
#> exo2       -4.8028751  1.296983  -6.34863062   -3.3462246
#> exo3        1.5953243  1.216342   0.08777459    2.8816613
#> 
#> $ARG$A$equation4
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1 -0.62284992  0.8457014 -1.741357277   0.03333532
#> lag1_var2  1.37549154  1.3295288  0.003907196   3.03166356
#> lag1_var3  2.22674121  2.2089924 -0.101412813   4.94564553
#> lag1_var4 -1.24066060  1.9697245 -3.655062560   0.84058563
#> const     17.46733440 22.0855615  3.771540176  48.03368476
#> exo1      -0.08922122  0.6113044 -0.684839047   0.70011248
#> exo2      -4.22170869  0.5745383 -5.004100513  -3.76450170
#> exo3       3.27026973  0.6694759  2.445296703   3.98332428
#> 
#> 
#>