Provides posterior estimation summary for Bayesian Vector Autoregressions for dynamic panel data

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARs'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARs obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005009 0.002029632 0.003005247  0.007522037
#> 
#> $AFG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.00810137 0.00530327 -0.01445214 -0.003158853
#> Sigma[2,2]  0.17184041 0.03009305  0.14986416  0.210785543
#> 
#> $AFG$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.002068684 0.002831819 -0.001663514  0.004453378
#> Sigma[3,2] -0.134471661 0.035288898 -0.171437950 -0.096590848
#> Sigma[3,3]  0.255000515 0.057038605  0.189569798  0.319763903
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1] -0.001666804 0.002543722 -0.004103378  0.001363855
#> Sigma[4,2] -0.067508489 0.040324143 -0.113221844 -0.019519831
#> Sigma[4,3]  0.223365222 0.047724492  0.167549694  0.277452520
#> Sigma[4,4]  0.221417317 0.038054495  0.180814305  0.268681102
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1  0.949184171 0.03226214  0.907072209   0.97810713
#> lag1_var2  0.146688437 0.14744886 -0.042271640   0.29037771
#> lag1_var3  0.351131329 0.28571056 -0.005722749   0.62145164
#> lag1_var4 -0.311107023 0.24483263 -0.540715438  -0.01131584
#> const     -0.495548971 1.71952914 -1.751197109   1.85442365
#> exo1      -0.057295222 0.05552914 -0.097558527   0.01896708
#> exo2       0.007926761 0.05765619 -0.034154517   0.08390960
#> exo3      -0.103845382 0.05910063 -0.179031990  -0.05339428
#> 
#> $AFG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.4702087  0.1079681   0.3888498   0.61794316
#> lag1_var2  0.1819457  0.9659150  -0.9192229   1.04191709
#> lag1_var3 -1.6667106  1.8041868  -3.8339827  -0.16148116
#> lag1_var4  1.4264038  1.4878944   0.1869120   3.24880801
#> const      0.9416078 15.4468186 -14.3310666  18.57530140
#> exo1      -0.3316402  0.5428431  -0.9169582   0.30131461
#> exo2      -0.1422337  0.2640691  -0.4919010   0.09344816
#> exo3      -0.4746254  0.5330020  -1.0387441   0.17060874
#> 
#> $AFG$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.25659205 0.0787199  -0.3527209  -0.17009603
#> lag1_var2  0.03293506 0.5173817  -0.4929010   0.61187340
#> lag1_var3  1.69289048 0.8471442   0.9166496   2.68985863
#> lag1_var4 -0.72394222 0.6574043  -1.5090530  -0.09909322
#> const      9.48501963 9.5734483  -1.1147971  20.03775444
#> exo1       0.40858549 0.7810634  -0.4876036   1.23554053
#> exo2      -1.10330323 0.3295937  -1.5114044  -0.75045760
#> exo3       0.05879083 0.8111896  -0.7005119   1.06477562
#> 
#> $AFG$A$equation4
#>                  mean         sd 5% quantile  95% quantile
#> lag1_var1 -0.05309393 0.03985182 -0.08753431  0.0002577016
#> lag1_var2  0.09244599 0.66504565 -0.71811846  0.8267171699
#> lag1_var3  0.93467856 1.38410013 -0.81172569  2.3491479387
#> lag1_var4  0.05071101 1.18142675 -1.12558108  1.5659462978
#> const      4.54240659 9.67711693 -7.05894285 15.3133338325
#> exo1       0.32487173 0.62464348 -0.37427848  0.9814655551
#> exo2      -1.24232810 0.31842925 -1.65606205 -0.9973002369
#> exo3      -0.08955832 0.73248045 -0.90835718  0.7881090750
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005311222 0.0005752992 0.004570892  0.005854487
#> 
#> $AGO$Sigma$equation2
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.007998423 0.004073726 -0.01270462 -0.004001049
#> Sigma[2,2]  0.108199740 0.021859711  0.08301777  0.133792068
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.007195347 0.001883901  0.00562162  0.009392736
#> Sigma[3,2] -0.039844596 0.012932601 -0.05209679 -0.024010975
#> Sigma[3,3]  0.029784913 0.003647678  0.02570537  0.034185881
#> 
#> $AGO$Sigma$equation4
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.0016522963 0.001759611 -0.000124185  0.003886490
#> Sigma[4,2]  0.0487646461 0.008816885  0.039030197  0.059064106
#> Sigma[4,3] -0.0003940637 0.007815377 -0.009216352  0.008480446
#> Sigma[4,4]  0.0436994886 0.008809906  0.033165071  0.053545425
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.969374874 0.02897367  0.94663072   1.00862788
#> lag1_var2  0.063018408 0.03821914  0.01431482   0.10219982
#> lag1_var3  0.113087564 0.03616979  0.06509379   0.14278032
#> lag1_var4 -0.091197924 0.03827971 -0.13933127  -0.05736168
#> const     -0.509554183 2.40809062 -3.33064647   1.89732906
#> exo1       0.013639519 0.08757609 -0.08384414   0.11943015
#> exo2      -0.006330537 0.05501558 -0.06911330   0.05754271
#> exo3       0.031332736 0.15083477 -0.09957411   0.23084744
#> 
#> $AGO$A$equation2
#>                   mean        sd  5% quantile 95% quantile
#> lag1_var1 -0.003274876 0.1609315 -0.189603837   0.17556092
#> lag1_var2  0.716735933 0.1321711  0.581407337   0.88427538
#> lag1_var3 -0.323943037 0.2211292 -0.470393747  -0.02363609
#> lag1_var4  0.307082829 0.1870478  0.057093860   0.45626551
#> const      1.869169530 7.6946934 -7.172124398  10.08570748
#> exo1       0.257315171 0.3357000 -0.059754480   0.66455932
#> exo2       0.252192296 0.2241463  0.009212193   0.52268869
#> exo3      -0.788519177 0.3611400 -1.221731300  -0.38237529
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.22177123 0.05918059  -0.2784783   -0.1466101
#> lag1_var2 -1.03739828 0.09667148  -1.1490419   -0.9363632
#> lag1_var3 -0.39735412 0.13771900  -0.5835308   -0.2897644
#> lag1_var4  1.12086429 0.12948984   0.9944211    1.2886321
#> const     26.05510701 3.80070368  21.4947111   29.3886088
#> exo1       0.05853217 0.22163613  -0.1869519    0.3254540
#> exo2      -1.44167457 0.21861836  -1.7168138   -1.2127507
#> exo3       3.22995680 0.30856422   2.9290963    3.6268738
#> 
#> $AGO$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.3150435 0.08994883 -0.40249092   -0.2031561
#> lag1_var2 -0.5821886 0.07420452 -0.64515173   -0.4841977
#> lag1_var3 -0.7454425 0.04799913 -0.78366056   -0.6900735
#> lag1_var4  1.5602215 0.08343006  1.44765152    1.6260286
#> const     22.0170863 4.17869999 17.82771258   27.1710630
#> exo1       0.2726740 0.29002015 -0.01664147    0.6120461
#> exo2      -1.5717226 0.03716619 -1.60830917   -1.5268827
#> exo3       3.0426347 0.23318866  2.73185558    3.2491558
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.008576554 0.008991128 0.002011574    0.0205441
#> 
#> $ALB$Sigma$equation2
#>                  mean       sd 5% quantile 95% quantile
#> Sigma[2,1] -0.1019155 0.074522  -0.1972132  -0.04330152
#> Sigma[2,2]  7.7203446 3.925840   4.5447528  12.96618648
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[3,1]  0.07191687 0.08411435  0.009411038    0.1873706
#> Sigma[3,2] -4.63489087 2.44368538 -7.357057499   -2.4891281
#> Sigma[3,3]  4.51443286 2.16696728  2.740033973    7.3433372
#> 
#> $ALB$Sigma$equation4
#>                     mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.0006283232 0.05009071 -0.03002838   0.06872301
#> Sigma[4,2]  0.6936952822 1.37505146 -0.99964805   2.17365476
#> Sigma[4,3]  1.6373852830 1.12881128  0.56991042   3.10917532
#> Sigma[4,4]  2.5168242940 0.66386160  1.75603782   3.01346615
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.82393195 0.18452612  0.62210399   0.99024157
#> lag1_var2 -0.15432090 0.18481005 -0.40743004  -0.04954559
#> lag1_var3 -0.23251710 0.26637824 -0.59562432  -0.08077263
#> lag1_var4  0.18482290 0.21653291  0.06477498   0.47847772
#> const      7.13953186 7.18395641  1.50297350  16.46708394
#> exo1       0.01467844 0.07028908 -0.05687531   0.08860218
#> exo2       0.03451355 0.07186042 -0.01585094   0.13301557
#> exo3       0.14473866 0.10332148  0.05267448   0.28152132
#> 
#> $ALB$A$equation2
#>                mean        sd 5% quantile 95% quantile
#> lag1_var1 -4.336096  3.597988   -7.978920  -0.07372285
#> lag1_var2  3.029504  1.832259    1.336578   5.30727013
#> lag1_var3  4.375033  3.026254    1.667896   8.23185642
#> lag1_var4 -3.646516  2.743925   -7.135648  -1.24788822
#> const     65.098205 73.206762  -30.936617 126.77800019
#> exo1      -3.535070  3.584227   -8.220096  -0.09921056
#> exo2      -2.112194  2.354867   -4.272538   0.95933071
#> exo3      -1.244908  2.924814   -3.936910   2.44579449
#> 
#> $ALB$A$equation3
#>                  mean          sd  5% quantile 95% quantile
#> lag1_var1   2.9909531   4.2201489   -2.4699063     7.066591
#> lag1_var2  -2.6670567   0.8366868   -3.6600613    -1.930504
#> lag1_var3  -3.8285327   1.5034516   -5.4748663    -2.652995
#> lag1_var4   3.9248272   1.4618494    2.6936459     5.592138
#> const     -18.1805768 102.3965029 -117.5852975   114.881717
#> exo1        1.5396547   1.4858214   -0.1583312     3.053865
#> exo2       -0.3978508   1.8188717   -1.8587767     2.038083
#> exo3        1.3802151   3.0864871   -1.3290120     5.428307
#> 
#> $ALB$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.3434357  2.3507712   -2.755337   2.42296977
#> lag1_var2 -0.9088740  0.6940638   -1.533579   0.01122137
#> lag1_var3 -1.2647798  0.9716604   -2.233944  -0.02095453
#> lag1_var4  1.9504575  0.7340284    1.050837   2.74701376
#> const     13.1903164 66.7507404  -57.049183  95.85306270
#> exo1      -0.5728791  2.9174258   -3.081908   3.26497266
#> exo2      -2.1002526  2.6840352   -5.148733   0.86368467
#> exo3       0.5884722  2.4846758   -1.835019   3.64903948
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                  mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00215776 0.0007730841 0.001506202   0.00309378
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.002181491 0.002516935 -0.005297607 0.0003816122
#> Sigma[2,2]  0.048231309 0.007928530  0.037977132 0.0556581982
#> 
#> $ARE$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.00721038 0.003896627  0.003045553   0.01157209
#> Sigma[3,2] -0.05850317 0.010944157 -0.071081536  -0.04935232
#> Sigma[3,3]  0.18225517 0.025252457  0.157267535   0.21276804
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.005409451 0.001911400  0.003322428  0.006986265
#> Sigma[4,2] -0.021282481 0.009837139 -0.031747413 -0.009806251
#> Sigma[4,3]  0.139273262 0.019544213  0.119321264  0.163376958
#> Sigma[4,4]  0.125524820 0.015511406  0.110801405  0.145643683
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.95611374 0.06041886  0.91089988   1.03807847
#> lag1_var2 -0.46999505 0.02963439 -0.50683488  -0.44064637
#> lag1_var3 -0.64505770 0.02840228 -0.67712895  -0.61431539
#> lag1_var4  0.62841341 0.03557687  0.60051896   0.67695908
#> const      2.41295307 0.35319112  1.98302509   2.78874460
#> exo1      -0.02053340 0.03940200 -0.07227870   0.01564006
#> exo2      -0.06402767 0.04306418 -0.11195700  -0.01762250
#> exo3      -0.02959032 0.05292485 -0.07819006   0.04055459
#> 
#> $ARE$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.01361381 0.12524398 -0.12230545   0.15147056
#> lag1_var2  0.83791872 0.09613664  0.75341809   0.96662105
#> lag1_var3  0.10234223 0.04266686  0.06208245   0.15817712
#> lag1_var4 -0.12765094 0.03889996 -0.17431626  -0.08441158
#> const      2.08816664 1.22732893  1.02918455   3.66815046
#> exo1       0.10066870 0.26768551 -0.20691667   0.37761126
#> exo2       2.11630172 0.17663137  1.88438155   2.27950205
#> exo3      -0.51613900 0.23903434 -0.81633845  -0.32818102
#> 
#> $ARE$A$equation3
#>                 mean        sd  5% quantile 95% quantile
#> lag1_var1  0.2593862 0.2008147 -0.006694314   0.42944210
#> lag1_var2 -1.6004488 0.3590846 -2.011154729  -1.19450080
#> lag1_var3 -1.5741727 0.3645519 -1.958653092  -1.15362982
#> lag1_var4  2.4944291 0.3230309  2.099408962   2.82467515
#> const     -1.2535206 1.0181070 -2.352461768  -0.01801147
#> exo1      -0.7894391 0.6293669 -1.371724251   0.01461717
#> exo2      -2.4967017 0.3279345 -2.791722115  -2.09527992
#> exo3      -1.5289519 0.3733239 -1.811017356  -1.01817020
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.2667800 0.2064403   0.0136675    0.4832945
#> lag1_var2 -0.9870925 0.3797789  -1.4120017   -0.5994958
#> lag1_var3 -1.5321758 0.3830774  -1.9294445   -1.0758696
#> lag1_var4  2.4584905 0.3493607   2.0309524    2.8259769
#> const     -1.5169006 1.5330607  -3.0336720    0.2215789
#> exo1      -0.7243274 0.4699105  -1.1175019   -0.1151174
#> exo2      -0.8642441 0.2947174  -1.1050795   -0.4658572
#> exo3      -1.9384783 0.2421505  -2.1312499   -1.6085900
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002474568 0.0004666056 0.001962763  0.003013415
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05780406 0.01595344  -0.0779425  -0.04125971
#> Sigma[2,2]  2.94780587 0.64288919   2.2167634   3.59640728
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.03875964 0.01091143  0.02667116   0.05196164
#> Sigma[3,2] -2.08962556 0.36917958 -2.47171850  -1.65853375
#> Sigma[3,3]  1.76552888 0.27140954  1.45223793   2.05948999
#> 
#> $ARG$Sigma$equation4
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[4,1]  0.004483981 0.002371486  0.00178402  0.006688998
#> Sigma[4,2] -0.328732547 0.065476855 -0.39710789 -0.247923456
#> Sigma[4,3]  0.547950405 0.075334789  0.46309742  0.621012008
#> Sigma[4,4]  0.387233811 0.084096396  0.29473630  0.481398943
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.884478950 0.03863588  0.86117181   0.93698703
#> lag1_var2  0.133998313 0.07706776  0.04184402   0.19608448
#> lag1_var3  0.221534107 0.12965666  0.06504983   0.32847580
#> lag1_var4 -0.210898156 0.12235157 -0.30971212  -0.06022769
#> const      2.490086013 1.08584546  1.07292789   3.47281679
#> exo1       0.006302232 0.06337606 -0.06258505   0.07161919
#> exo2      -0.112788840 0.05556672 -0.17180450  -0.04325009
#> exo3       0.008339061 0.08240527 -0.07372927   0.11364538
#> 
#> $ARG$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  -2.4634765  2.350855  -5.6983046   -0.8011747
#> lag1_var2  -5.1469968  2.400919  -7.8535375   -3.0819234
#> lag1_var3  -9.6277560  4.080391 -14.1399508   -6.0221711
#> lag1_var4   9.0811100  3.545250   5.7269625   13.0944495
#> const     102.3069512 76.851743  21.3510675  195.3757403
#> exo1       -0.1604251  2.253392  -2.2269836    2.7808807
#> exo2        2.1125125  1.241983   0.8187693    3.4050817
#> exo3        0.2821029  2.632386  -3.0883537    2.8848940
#> 
#> $ARG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1   0.1886368  1.7953306  -1.2580202    2.6188058
#> lag1_var2   3.9005585  2.8032195   1.1275816    7.1328153
#> lag1_var3   7.1880520  4.6691162   2.5052649   12.4677189
#> lag1_var4  -6.1163662  4.2169857 -10.9911815   -1.7524331
#> const     -10.6098836 58.1649488 -76.3341429   55.6408391
#> exo1       -0.4671523  1.0857654  -1.6182925    0.8645534
#> exo2       -5.1431012  0.6834066  -5.7136384   -4.2935414
#> exo3        2.2989439  1.9521408   0.4459747    4.7978113
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -1.4718391  0.4269969  -1.9304270   -0.9442158
#> lag1_var2  0.7347133  1.4771743  -0.9301075    2.4162702
#> lag1_var3  1.2762175  2.3970798  -1.4361047    3.9620805
#> lag1_var4 -0.4751932  2.2517925  -3.0665643    2.0686074
#> const     52.2491766 16.2699132  39.7305863   73.1283554
#> exo1      -0.5937805  0.7021790  -1.2562021    0.3287145
#> exo2      -4.3746106  0.6523404  -5.2527469   -3.8833353
#> exo3       2.7535994  0.9143739   1.5217639    3.5173608
#> 
#> 
#>