R6 Class Representing PosteriorBVARPANEL

The class PosteriorBVARPANEL contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvarPANEL$new()

specify_posterior_bvarPANEL$get_posterior()

specify_posterior_bvarPANEL$get_last_draw()

specify_posterior_bvarPANEL$clone()

Method new()

Create a new posterior output PosteriorBVARPANEL.

Usage

specify_posterior_bvarPANEL$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARPANEL.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvarPANEL$get_posterior()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvarPANEL$get_last_draw()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvarPANEL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARPANEL" "R6"                

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_posterior`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $A
#> , , 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.2091296 -0.8198144  0.6501130  0.4160633
#> [2,]  0.5230142  2.4006682 -2.4834941  2.9411319
#> [3,] -0.8794125 -0.7137465 -0.0213957  0.2272853
#> [4,] -0.2975666 -2.4844152  3.6157879 -1.7282743
#> [5,] -2.6534396  3.4168048 -5.1724992  6.6680135
#> 
#> , , 2
#> 
#>            [,1]        [,2]       [,3]      [,4]
#> [1,]  0.9937338 -0.03337961  0.4641589 0.3136671
#> [2,]  0.4428724  1.00630619 -0.3222763 0.1335242
#> [3,]  0.3282941  0.01262473  0.8080393 0.3136764
#> [4,] -0.2264444  0.26397737 -0.3195209 0.3801622
#> [5,]  1.0130728 -1.18540124  6.4149276 5.2922619
#> 
#> , , 3
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.9031180  0.5353928 -0.1690155 0.1153060
#> [2,]  0.2291984  1.0637616 -0.2373135 0.3332979
#> [3,]  0.4444479  0.9489389 -0.2124847 0.1445806
#> [4,] -0.3184537 -0.6819335  1.0696053 1.0001549
#> [5,]  0.5347412 -4.2620349  4.5050276 2.2407337
#> 
#> , , 4
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  0.8993963  0.4810570 0.1564374 0.3076635
#> [2,]  0.3029692  0.7486698 0.1316854 0.6394991
#> [3,]  0.4604273 -0.4069879 1.3123633 0.8869595
#> [4,] -0.3740486 -0.3636832 0.6515530 0.5334991
#> [5,]  0.2134411 -4.3503949 4.5322582 1.3465794
#> 
#> , , 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8406205  0.2341408  0.3757571  0.5310461
#> [2,]  0.4633123 -0.1780016  1.1618080  0.7796010
#> [3,]  0.6748111 -0.7328647  2.4073256  1.6238938
#> [4,] -0.6738714  1.0546757 -1.5319567 -0.5170690
#> [5,]  0.3841150 -7.6137043  9.4240049  3.0770362
#> 
#> 
#> $V
#> , , 1
#> 
#>           [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 10.927013  4.1645568 -1.3961236  -6.505179   3.254408
#> [2,]  4.164557 14.5968118  0.7322427  -9.623045   6.410101
#> [3,] -1.396124  0.7322427  8.2952622  -2.413582   3.218898
#> [4,] -6.505179 -9.6230450 -2.4135821  25.293908 -23.691230
#> [5,]  3.254408  6.4101006  3.2188983 -23.691230 112.768134
#> 
#> , , 2
#> 
#>            [,1]       [,2]       [,3]       [,4]        [,5]
#> [1,]  0.6811842 -0.3310249 -0.6008633  0.5681044  -1.0497276
#> [2,] -0.3310249  2.1066624  1.3705077 -1.1850504   3.8155044
#> [3,] -0.6008633  1.3705077  3.1166512 -2.6088012   1.2719538
#> [4,]  0.5681044 -1.1850504 -2.6088012  3.0070789  -0.9402219
#> [5,] -1.0497276  3.8155044  1.2719538 -0.9402219 101.8002257
#> 
#> , , 3
#> 
#>            [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,]  2.7243544 -0.4724752 -0.7733218  1.305449 -3.7472594
#> [2,] -0.4724752  5.7466027  3.3802663 -3.368423 -0.8013077
#> [3,] -0.7733218  3.3802663  6.6853612 -5.659279  4.4427863
#> [4,]  1.3054491 -3.3684234 -5.6592794  7.349743 -5.8171268
#> [5,] -3.7472594 -0.8013077  4.4427863 -5.817127 64.3974004
#> 
#> , , 4
#> 
#>           [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,]  1.719726 -1.403372 -1.644965   1.228083  -4.113452
#> [2,] -1.403372  7.832757  6.626837  -6.672307  12.470342
#> [3,] -1.644965  6.626837 10.589702  -9.144812  13.723073
#> [4,]  1.228083 -6.672307 -9.144812  11.867781 -17.218524
#> [5,] -4.113452 12.470342 13.723073 -17.218524 115.019657
#> 
#> , , 5
#> 
#>           [,1]      [,2]       [,3]       [,4]       [,5]
#> [1,]  1.723010 -2.248603  -2.822203   2.373362  -5.523142
#> [2,] -2.248603  8.575288   9.472246  -9.093354   9.491263
#> [3,] -2.822203  9.472246  12.790269 -12.114997  11.852436
#> [4,]  2.373362 -9.093354 -12.114997  13.286561 -12.819864
#> [5,] -5.523142  9.491263  11.852436 -12.819864 101.097740
#> 
#> 
#> $Sigma
#> , , 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] 11.3447077  0.1151519 -2.0306284 -3.5531300
#> [2,]  0.1151519 10.5267408 -7.2659768  0.7175921
#> [3,] -2.0306284 -7.2659768 12.4266910  0.9130984
#> [4,] -3.5531300  0.7175921  0.9130984 16.9958154
#> 
#> , , 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.2314817 -0.1845811  0.8451171 -0.3571719
#> [2,] -0.1845811 10.9189639 -6.5277001  2.3269382
#> [3,]  0.8451171 -6.5277001 11.2199178  3.2440342
#> [4,] -0.3571719  2.3269382  3.2440342  7.9794439
#> 
#> , , 3
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  0.6188665  0.4968919 -0.141064 0.2146974
#> [2,]  0.4968919  7.2928197 -4.076767 1.2162810
#> [3,] -0.1410640 -4.0767675  8.077906 4.7540243
#> [4,]  0.2146974  1.2162810  4.754024 6.4849439
#> 
#> , , 4
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,] 0.26133802  0.02534335  0.4223071  0.4069123
#> [2,] 0.02534335  6.30295782 -6.4541388 -2.6281871
#> [3,] 0.42230707 -6.45413875  9.8645943  6.1048476
#> [4,] 0.40691233 -2.62818706  6.1048476  5.0003158
#> 
#> , , 5
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.3195240 -0.0904680  0.2374308 0.1298982
#> [2,] -0.0904680  2.9427594 -1.1713804 0.6582972
#> [3,]  0.2374308 -1.1713804  3.3907988 2.7736866
#> [4,]  0.1298982  0.6582972  2.7736866 3.4462426
#> 
#> 
#> $nu
#>          [,1]
#> [1,] 5.099916
#> [2,] 5.099916
#> [3,] 5.100216
#> [4,] 5.100574
#> [5,] 5.100700
#> 
#> $m
#>           [,1]
#> [1,] 0.2829799
#> [2,] 1.8593948
#> [3,] 0.8961513
#> [4,] 0.1536743
#> [5,] 1.3788928
#> 
#> $w
#>           [,1]
#> [1,] 222.75268
#> [2,]  11.82339
#> [3,]  49.61376
#> [4,]  29.72712
#> [5,]  15.84348
#> 
#> $s
#>           [,1]
#> [1,] 0.6360339
#> [2,] 1.1308499
#> [3,] 0.9021194
#> [4,] 0.5341974
#> [5,] 0.4694018
#> 
#> $scale
#>            [,1]
#> [1,] 0.01427928
#> [2,] 0.01427928
#> [3,] 0.01580176
#> [4,] 0.01732519
#> [5,] 0.01926532
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>              [,1]         [,2]        [,3]         [,4]
#> [1,]  0.041542740 -0.000644604  0.01906059  0.001850792
#> [2,] -0.000644604  0.254844581 -0.25312279 -0.139918721
#> [3,]  0.019060590 -0.253122788  0.44685556  0.281508015
#> [4,]  0.001850792 -0.139918721  0.28150801  0.306963932
#> 
#> , , 2, 1
#> 
#>             [,1]        [,2]         [,3]         [,4]
#> [1,] 0.036809494  0.00126066  0.005355864  0.007048713
#> [2,] 0.001260660  0.17837609 -0.163593071 -0.025808892
#> [3,] 0.005355864 -0.16359307  0.427226466  0.287579497
#> [4,] 0.007048713 -0.02580889  0.287579497  0.317813690
#> 
#> , , 3, 1
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.09963400 -0.39274506  0.1955158 -0.07481495
#> [2,] -0.39274506  8.76261498 -5.5784059  0.09651279
#> [3,]  0.19551576 -5.57840591  5.7007536  2.41181752
#> [4,] -0.07481495  0.09651279  2.4118175  2.97857188
#> 
#> , , 4, 1
#> 
#>              [,1]         [,2]        [,3]         [,4]
#> [1,]  0.029603122 -0.012379914  0.01077328 -0.004600031
#> [2,] -0.012379914  0.210016847 -0.19782596  0.002496617
#> [3,]  0.010773278 -0.197825962  0.34180057  0.158381201
#> [4,] -0.004600031  0.002496617  0.15838120  0.235898523
#> 
#> , , 5, 1
#> 
#>             [,1]       [,2]       [,3]      [,4]
#> [1,]  0.04819711 -0.1538509  0.1888201  0.102247
#> [2,] -0.15385091  3.4268377 -3.3677262 -1.539466
#> [3,]  0.18882015 -3.3677262  4.0899566  2.371304
#> [4,]  0.10224705 -1.5394656  2.3713038  1.739601
#> 
#> , , 1, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.01271997 -0.02667613  0.03125326  0.02584314
#> [2,] -0.02667613  0.31679561 -0.35750319 -0.25833994
#> [3,]  0.03125326 -0.35750319  0.61006360  0.52412577
#> [4,]  0.02584314 -0.25833994  0.52412577  0.51394162
#> 
#> , , 2, 2
#> 
#>             [,1]         [,2]        [,3]         [,4]
#> [1,]  0.02132061 -0.019381602  0.03964789  0.021182179
#> [2,] -0.01938160  0.199199590 -0.15634276 -0.001898778
#> [3,]  0.03964789 -0.156342761  0.45736596  0.366153448
#> [4,]  0.02118218 -0.001898778  0.36615345  0.423394019
#> 
#> , , 3, 2
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.01313385 -0.1248958  0.03749691 -0.04455246
#> [2,] -0.12489578  7.8117474 -5.32981389 -0.45520030
#> [3,]  0.03749691 -5.3298139  6.24449499  3.28819628
#> [4,] -0.04455246 -0.4552003  3.28819628  3.46603871
#> 
#> , , 4, 2
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.008344360 -0.003475793  0.007154946  0.002872979
#> [2,] -0.003475793  0.250202898 -0.226218104 -0.030257041
#> [3,]  0.007154946 -0.226218104  0.318954050  0.144951105
#> [4,]  0.002872979 -0.030257041  0.144951105  0.136488411
#> 
#> , , 5, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.02264114 -0.03463882  0.05158365  0.02879387
#> [2,] -0.03463882  2.45447687 -2.02113937 -0.62086306
#> [3,]  0.05158365 -2.02113937  2.39355796  1.29259889
#> [4,]  0.02879387 -0.62086306  1.29259889  1.02625402
#> 
#> , , 1, 3
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.006293439  0.006547921 -0.003483345  0.004059655
#> [2,]  0.006547921  0.190488141 -0.159233377 -0.069895098
#> [3,] -0.003483345 -0.159233377  0.371424368  0.318022677
#> [4,]  0.004059655 -0.069895098  0.318022677  0.322244677
#> 
#> , , 2, 3
#> 
#>               [,1]         [,2]         [,3]          [,4]
#> [1,]  0.0065946380 -0.002877201  0.001803909  0.0002783832
#> [2,] -0.0028772008  0.159260873 -0.144982947 -0.0328900005
#> [3,]  0.0018039085 -0.144982947  0.342581002  0.2691180659
#> [4,]  0.0002783832 -0.032890001  0.269118066  0.2882555908
#> 
#> , , 3, 3
#> 
#>              [,1]       [,2]        [,3]         [,4]
#> [1,]  0.010005931 -0.1175071  0.06836347 -0.006938031
#> [2,] -0.117507065  7.3648705 -4.64623888  0.202757522
#> [3,]  0.068363473 -4.6462389  4.44140629  1.643749412
#> [4,] -0.006938031  0.2027575  1.64374941  2.103695256
#> 
#> , , 4, 3
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.003138380 -0.001083318  0.009732933  0.009976369
#> [2,] -0.001083318  0.205936657 -0.177624896 -0.025883936
#> [3,]  0.009732933 -0.177624896  0.435887526  0.315141047
#> [4,]  0.009976369 -0.025883936  0.315141047  0.310635429
#> 
#> , , 5, 3
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.005891633 -0.07652647  0.04960172  0.004665317
#> [2,] -0.076526475  3.88005433 -2.67569603 -0.375144118
#> [3,]  0.049601718 -2.67569603  2.79064435  1.287004871
#> [4,]  0.004665317 -0.37514412  1.28700487  1.159231044
#> 
#> , , 1, 4
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.006719521 -0.007073932 -0.004440209 -0.009176685
#> [2,] -0.007073932  0.156491365 -0.137689434 -0.074574153
#> [3,] -0.004440209 -0.137689434  0.449843925  0.410972309
#> [4,] -0.009176685 -0.074574153  0.410972309  0.408157071
#> 
#> , , 2, 4
#> 
#>              [,1]        [,2]         [,3]         [,4]
#> [1,]  0.006645594 -0.00251093  0.009110996  0.007899364
#> [2,] -0.002510930  0.12647411 -0.142170573 -0.055450499
#> [3,]  0.009110996 -0.14217057  0.502044478  0.464766219
#> [4,]  0.007899364 -0.05545050  0.464766219  0.499978734
#> 
#> , , 3, 4
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.01058240 -0.14040698  0.1624283 0.08286753
#> [2,] -0.14040698  7.36650657 -4.8499337 0.05940998
#> [3,]  0.16242833 -4.84993372  6.6664242 4.13623886
#> [4,]  0.08286753  0.05940998  4.1362389 5.03552818
#> 
#> , , 4, 4
#> 
#>              [,1]         [,2]        [,3]         [,4]
#> [1,]  0.002676400 -0.009486109  0.01063755  0.003967767
#> [2,] -0.009486109  0.245777334 -0.21403422 -0.031805914
#> [3,]  0.010637546 -0.214034216  0.31331557  0.158317377
#> [4,]  0.003967767 -0.031805914  0.15831738  0.140412010
#> 
#> , , 5, 4
#> 
#>              [,1]        [,2]       [,3]         [,4]
#> [1,]  0.004629515 -0.06335449  0.0379478 -0.000104118
#> [2,] -0.063354495  2.51431653 -1.6108141 -0.107455320
#> [3,]  0.037947796 -1.61081409  1.5131606  0.610018342
#> [4,] -0.000104118 -0.10745532  0.6100183  0.616673623
#> 
#> , , 1, 5
#> 
#>              [,1]        [,2]         [,3]         [,4]
#> [1,]  0.005430582 -0.01208889  0.002944701 -0.001219104
#> [2,] -0.012088892  0.18682684 -0.189726180 -0.123809685
#> [3,]  0.002944701 -0.18972618  0.411344362  0.366446447
#> [4,] -0.001219104 -0.12380969  0.366446447  0.357064241
#> 
#> , , 2, 5
#> 
#>              [,1]         [,2]       [,3]        [,4]
#> [1,]  0.006608978 -0.006988461  0.0163900  0.01344893
#> [2,] -0.006988461  0.133426629 -0.1572697 -0.07080649
#> [3,]  0.016390001 -0.157269695  0.5814089  0.54574888
#> [4,]  0.013448930 -0.070806492  0.5457489  0.57848695
#> 
#> , , 3, 5
#> 
#>              [,1]        [,2]        [,3]       [,4]
#> [1,]  0.007494819  -0.1844649   0.2272513  0.1262418
#> [2,] -0.184464933   9.9557810 -10.0613063 -4.2588438
#> [3,]  0.227251312 -10.0613063  13.8164242  8.6713364
#> [4,]  0.126241840  -4.2588438   8.6713364  7.0707390
#> 
#> , , 4, 5
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.002416642 -0.007494433  0.01153087  0.00599700
#> [2,] -0.007494433  0.214685418 -0.22243325 -0.06041592
#> [3,]  0.011530875 -0.222433253  0.38426093  0.21875473
#> [4,]  0.005997000 -0.060415923  0.21875473  0.17678752
#> 
#> , , 5, 5
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.003740186 -0.04688375  0.03591887  0.00925269
#> [2,] -0.046883753  2.39417723 -1.89604156 -0.49450607
#> [3,]  0.035918873 -1.89604156  2.18744210  1.16212861
#> [4,]  0.009252690 -0.49450607  1.16212861  0.96933895
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  0.9343379   0.4371464 -0.07266752  0.06415354
#> [2,]  0.1815437   1.1121844 -0.90800719 -0.47341533
#> [3,]  0.3831676   0.2921370  0.06613539  0.11950056
#> [4,] -0.3372269  -0.2623023  0.63800011  0.69557044
#> [5,] -0.6297853 -11.2209318 19.31737663 11.48383369
#> 
#> , , 2, 1
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.1026691 -0.05684908 -0.2274579 -0.5415503
#> [2,] -0.2743627  1.10440232  0.6869187  1.6680355
#> [3,] -0.3230516  0.14012830  1.6486525  1.8259003
#> [4,]  0.3336476 -0.13081687 -0.6145733 -0.7628767
#> [5,] -2.9111568  0.68090606 -0.1048964  4.1367906
#> 
#> , , 3, 1
#> 
#>            [,1]      [,2]      [,3]       [,4]
#> [1,]  0.6645488 -1.201958  1.169125  0.2377392
#> [2,]  0.1737991  3.069811 -3.038640 -1.6383034
#> [3,]  0.2962465  3.746454 -3.828352 -2.2445139
#> [4,] -0.2134007 -3.106636  3.931221  2.7484695
#> [5,]  2.8719222 -8.702193 29.670026 28.9097972
#> 
#> , , 4, 1
#> 
#>            [,1]       [,2]        [,3]      [,4]
#> [1,]  0.9724287  0.0831181  0.47858862 0.3900962
#> [2,]  0.6519384  0.4082557  0.01873945 0.2994657
#> [3,]  0.8306910 -0.4585234  0.98751109 0.4747284
#> [4,] -0.8251874  0.4643794 -0.20765653 0.3355598
#> [5,]  0.2844223 -1.9379441  4.36259357 4.6727781
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.7421277 -0.4291040  0.5273557 -0.1109738
#> [2,]  0.7334604 -0.6732273  0.4748613 -0.3379423
#> [3,]  1.2046488 -2.4804954  1.5937699 -0.5536717
#> [4,] -1.1011883  2.5279636 -0.9317856  1.2473087
#> [5,]  0.6464465 10.9616623  4.7257362 21.5869900
#> 
#> , , 1, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9604533  0.1976341  0.2716997  0.2265384
#> [2,]  0.1401278  1.2612376 -0.9239992 -0.3709824
#> [3,]  0.3497543  0.5650917 -0.4276181 -0.2484476
#> [4,] -0.3081586 -0.5629125  1.2522094  1.1986963
#> [5,] -0.7607580 -4.4827678  3.9426682 -0.8824154
#> 
#> , , 2, 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9721915  0.10899486 -0.2728507 -0.3720678
#> [2,] -0.2169730  0.80142715  0.8034860  1.5095594
#> [3,] -0.2836692  0.02969517  1.6183240  1.8300908
#> [4,]  0.2886528  0.02029534 -0.7452326 -0.8448446
#> [5,]  0.3190589 -2.95379590 11.1814670  8.6409619
#> 
#> , , 3, 2
#> 
#>            [,1]        [,2]      [,3]       [,4]
#> [1,]  0.8316711  -0.6052515  1.090916  0.5090737
#> [2,]  0.2745658   2.1264971 -2.032237 -1.0834164
#> [3,]  0.4385307   2.7216521 -2.788229 -1.6191845
#> [4,] -0.3739130  -2.0034406  2.914455  2.2515318
#> [5,] -0.1232385 -21.4554609 23.764880 12.1444302
#> 
#> , , 4, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8364212  0.3250655  0.6023690  0.5955290
#> [2,]  0.5715642  0.4354126  0.1876879  0.3760506
#> [3,]  0.7045514 -0.1962942  1.2375252  0.8307299
#> [4,] -0.6670023  0.1760192 -0.5042869 -0.0489648
#> [5,]  1.3837945 -5.8155629  4.5577009  1.7918475
#> 
#> , , 5, 2
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.7234852  0.1958675  0.5027920  0.57770552
#> [2,]  0.7054600 -0.6150664  0.9836588  0.57230605
#> [3,]  1.1412685 -2.3528989  2.2041292  0.70722547
#> [4,] -1.0384790  2.2783192 -1.4079223  0.07064651
#> [5,]  1.0851660  1.4176735 -3.5997662 -3.23235031
#> 
#> , , 1, 3
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.9388686   0.2376461  0.09595173  0.1398277
#> [2,]  0.1440969   1.4144465 -1.06834487 -0.4935672
#> [3,]  0.3325042   0.6005864 -0.44888940 -0.2454639
#> [4,] -0.2933050  -0.4654114  1.11680929  1.0576739
#> [5,] -0.3155219 -12.9076946 16.40815013  8.4361321
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.0514558  0.2920013 -0.3022168 -0.2032758
#> [2,] -0.2803012  0.9198505  0.8278290  1.6144107
#> [3,] -0.3581925 -0.1983933  1.8493912  1.7610646
#> [4,]  0.3293794  0.1988401 -0.9159551 -0.8069249
#> [5,]  1.0303089 -8.4878640  9.6185519  4.0853635
#> 
#> , , 3, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8206436 -2.0093908   2.735792   1.471723
#> [2,]  0.2466167  1.0967808  -1.345534  -1.151823
#> [3,]  0.3970685  0.8667666  -1.335676  -1.478218
#> [4,] -0.3397246 -0.6025335   1.984377   2.366569
#> [5,]  0.6127692 36.6006961 -42.221836 -22.676204
#> 
#> , , 4, 3
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.8900241  0.2110643  0.8176840  0.89514117
#> [2,]  0.5503311  0.8923831 -0.4938628  0.04846628
#> [3,]  0.7019030  0.4208675  0.6128259  0.78246032
#> [4,] -0.6737400 -0.3989712  0.1104373 -0.01349059
#> [5,]  0.7513874 -6.1070269  0.1536629 -4.44611637
#> 
#> , , 5, 3
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  0.7340461 -0.8884187  1.410430  0.8116860
#> [2,]  0.7510882 -0.9251454  1.542260  0.8981231
#> [3,]  1.2285495 -2.5835279  2.895664  1.1600132
#> [4,] -1.1197694  3.0392843 -2.730814 -0.7056566
#> [5,]  0.5389740 -0.2952530  8.921380  9.3881026
#> 
#> , , 1, 4
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.9677874  0.3166334  0.03582999  0.1521153
#> [2,]  0.1150132  1.1854498 -1.20387927 -0.7533343
#> [3,]  0.3429570  0.5146834 -0.54110265 -0.3432653
#> [4,] -0.3148301 -0.4752188  1.16620666  1.0696016
#> [5,] -0.1422833 -8.5851424 20.54606544 14.0352701
#> 
#> , , 2, 4
#> 
#>              [,1]         [,2]       [,3]       [,4]
#> [1,]  0.993411519   0.01030105 -0.3731299 -0.4728363
#> [2,] -0.021733087   0.58072320  2.5708855  2.9121544
#> [3,]  0.001575784  -0.52153289  4.2191635  3.6810196
#> [4,]  0.014285297   0.67918593 -3.3442329 -2.6500026
#> [5,] -0.636785834 -12.14536355 17.0962393  7.6942481
#> 
#> , , 3, 4
#> 
#>            [,1]       [,2]          [,3]        [,4]
#> [1,]  0.8407598 -1.3594354   1.229098380  0.21478774
#> [2,]  0.2366007  0.8762550   0.228668697  0.54309046
#> [3,]  0.3829900  0.7730052   0.796075527  0.96291394
#> [4,] -0.3264365 -0.5686157  -0.006722341  0.02777256
#> [5,]  0.2040801 27.7270346 -21.475958187 -4.66120606
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9491812 -0.3631545  1.3790826  1.0234759
#> [2,]  0.6351228  1.1101224 -0.2266360  0.5940219
#> [3,]  0.7969549  0.4693626  0.9634061  1.2854731
#> [4,] -0.7866164 -0.3531592 -0.3279467 -0.5248487
#> [5,]  0.5258754  1.1083093 -7.6176831 -7.4909588
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.7246831 -0.6652379  0.5312248  0.01018663
#> [2,]  0.7222596 -0.1131253  0.7897397  0.65009064
#> [3,]  1.1881202 -1.4467517  2.0685492  1.07557348
#> [4,] -1.0768896  1.8764963 -1.5163138 -0.18604174
#> [5,]  0.6476090 -5.4913029 11.4345745  6.67823131
#> 
#> , , 1, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9752203  0.2961211  0.2735981  0.4379345
#> [2,]  0.1318272  1.1952403 -1.1777774 -0.7190539
#> [3,]  0.3583352  0.4616066 -0.4138475 -0.2327734
#> [4,] -0.3277040 -0.3883105  1.0131824  0.9639540
#> [5,] -0.4931435 -9.9705535 16.5855056  7.4649823
#> 
#> , , 2, 5
#> 
#>              [,1]         [,2]      [,3]      [,4]
#> [1,]  0.969493086   0.36411762 -1.294317 -1.226961
#> [2,]  0.001609237  -0.02158199  4.984115  5.203191
#> [3,]  0.034144821  -1.18781659  7.375120  6.770905
#> [4,] -0.020367282   1.27770657 -6.244023 -5.507972
#> [5,]  0.132082776 -14.08205348 20.323050  9.915308
#> 
#> , , 3, 5
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.9836682 -3.102078   1.476273  -0.840080
#> [2,]  0.1156095  2.311096   2.515826   4.391088
#> [3,]  0.1447899  3.831733   3.414260   6.456982
#> [4,] -0.1296521 -2.828362  -2.583682  -4.793343
#> [5,] -0.9272626 26.054789 -40.142004 -28.617844
#> 
#> , , 4, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8548089  0.1974781  1.0217821  1.1346882
#> [2,]  0.5869971  1.1817064 -1.0488130 -0.1702802
#> [3,]  0.7364343  0.8325856 -0.2856511  0.3226658
#> [4,] -0.6928110 -0.8784502  0.9962299  0.3841941
#> [5,]  0.4428923 -0.5155509 -4.4292415 -6.2014670
#> 
#> , , 5, 5
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.7352377  -0.4927480 -0.4805769 -0.9954967
#> [2,]  0.6825491  -0.6397695  1.5208193  1.1075512
#> [3,]  1.1202244  -2.2704652  3.2109787  1.7898759
#> [4,] -1.0267376   2.7954015 -2.7261029 -0.9211810
#> [5,]  1.4670192 -15.7594983 42.2548094 34.5429166
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_last_draw`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|