R6 Class Representing PosteriorBVARPANEL

The class PosteriorBVARPANEL contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvarPANEL$new()

specify_posterior_bvarPANEL$get_posterior()

specify_posterior_bvarPANEL$get_last_draw()

specify_posterior_bvarPANEL$clone()

Method new()

Create a new posterior output PosteriorBVARPANEL.

Usage

specify_posterior_bvarPANEL$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARPANEL.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvarPANEL$get_posterior()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvarPANEL$get_last_draw()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvarPANEL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARPANEL" "R6"                

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_posterior`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $A
#> , , 1
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  0.1768795 -0.04455398  1.83026957  0.03030539
#> [2,] -1.1182534  0.53655978 -0.41968345 -1.72586694
#> [3,]  1.0536658 -0.54220007  0.24886484  0.68990628
#> [4,]  0.3581811 -0.81578665  0.03725682  0.20748442
#> [5,] -1.3756012 -0.22966951  2.84842339  6.67871538
#> 
#> , , 2
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  0.7911302 -0.4091359 0.9267116 0.4560054
#> [2,]  0.5351159  0.5245311 0.2098262 0.1899578
#> [3,]  0.8717264  0.2800738 0.7710791 0.5606866
#> [4,] -0.6966136 -0.3189863 0.3088969 0.6120977
#> [5,] -1.8161269 -0.3126471 2.9099258 5.2350228
#> 
#> , , 3
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.7858461 -0.4357353 -0.04633301 -0.62005153
#> [2,]  0.5626931  0.4041920  0.05675721  0.11963791
#> [3,]  0.7352732  0.7386965  0.18869829 -0.05440647
#> [4,] -0.7173023 -0.4754943 -0.60812334 -0.31303479
#> [5,] -1.5934667 -0.6156272  3.66167036  7.50312172
#> 
#> , , 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8457930  0.7668844 -0.6064006 -0.4059733
#> [2,]  0.4962671  1.2512133 -0.1397483  0.4045272
#> [3,]  0.6888955  0.5841958  1.2350373  1.2273892
#> [4,] -0.6656594 -0.5035507 -0.4369031 -0.1645064
#> [5,] -1.5880630 -5.1306054  2.8675860  3.1787301
#> 
#> , , 5
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.9608776 -0.23816606  0.03552303 -0.2558462
#> [2,]  0.5853164  0.64064432  0.05431254  0.3197661
#> [3,]  0.8213477  0.02285527  1.97462474  1.7426067
#> [4,] -0.7739899  0.28835195 -1.40942395 -0.9661466
#> [5,] -1.9684455 -1.88991080 -1.66142764 -0.8087764
#> 
#> 
#> $V
#> , , 1
#> 
#>            [,1]        [,2]       [,3]        [,4]       [,5]
#> [1,]  2.6866622  0.53483647 -0.6518719  0.30165583  0.1807091
#> [2,]  0.5348365  3.15654494 -0.5754568 -0.02243911 -3.5871519
#> [3,] -0.6518719 -0.57545675  3.2439370  0.35509548  2.9700443
#> [4,]  0.3016558 -0.02243911  0.3550955  3.50620896 -1.3956293
#> [5,]  0.1807091 -3.58715191  2.9700443 -1.39562927 26.5475659
#> 
#> , , 2
#> 
#>             [,1]        [,2]        [,3]        [,4]        [,5]
#> [1,]  0.56905568 -0.02125948 -0.27205700  0.04760963 -0.31346782
#> [2,] -0.02125948  0.77120280 -0.01644690 -0.31053021  0.05754729
#> [3,] -0.27205700 -0.01644690  1.28022279 -0.57495991  0.07395465
#> [4,]  0.04760963 -0.31053021 -0.57495991  2.08808730  0.61633808
#> [5,] -0.31346782  0.05754729  0.07395465  0.61633808  9.54203459
#> 
#> , , 3
#> 
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,]  3.5670392 -1.1171311  1.4124261  0.5322469 -6.9069833
#> [2,] -1.1171311  3.0665206  0.3353776 -0.6037798  0.8274064
#> [3,]  1.4124261  0.3353776  5.2838018 -1.5791722 -8.9775047
#> [4,]  0.5322469 -0.6037798 -1.5791722  3.3702774  2.9514489
#> [5,] -6.9069833  0.8274064 -8.9775047  2.9514489 63.6195808
#> 
#> , , 4
#> 
#>            [,1]       [,2]        [,3]       [,4]       [,5]
#> [1,]  1.6540491   0.374660   0.1805426 -0.3957939  -1.612446
#> [2,]  0.3746600   8.217899   6.9747140 -5.7256735 -15.752449
#> [3,]  0.1805426   6.974714   9.6749048 -6.3489277 -15.219526
#> [4,] -0.3957939  -5.725674  -6.3489277  6.3924429  15.718624
#> [5,] -1.6124458 -15.752449 -15.2195263 15.7186241  91.440760
#> 
#> , , 5
#> 
#>            [,1]        [,2]        [,3]       [,4]       [,5]
#> [1,]  2.1996448  -0.4344035  -0.5442546  0.1112688  -3.109858
#> [2,] -0.4344035   4.1950895   4.4499025 -3.5996606 -10.004412
#> [3,] -0.5442546   4.4499025  10.0440824 -5.8752987 -14.743848
#> [4,]  0.1112688  -3.5996606  -5.8752987  5.6423076  14.596761
#> [5,] -3.1098580 -10.0044119 -14.7438476 14.5967607 109.055657
#> 
#> 
#> $Sigma
#> , , 1
#> 
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 10.114114 -1.429931  1.382976 -2.066493
#> [2,] -1.429931 20.108957 -5.006968 -5.754565
#> [3,]  1.382976 -5.006968 14.473224  2.285149
#> [4,] -2.066493 -5.754565  2.285149 13.952223
#> 
#> , , 2
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  3.9964971  0.4811231 -0.587797  1.020654
#> [2,]  0.4811231 11.2081541 -8.506521 -1.222984
#> [3,] -0.5877970 -8.5065208 22.490555 10.447299
#> [4,]  1.0206545 -1.2229839 10.447299 17.394548
#> 
#> , , 3
#> 
#>             [,1]        [,2]         [,3]        [,4]
#> [1,]  0.74706548  -0.5980733   0.03503003  0.02314834
#> [2,] -0.59807326  19.1819712 -15.78702081 -3.62635969
#> [3,]  0.03503003 -15.7870208  25.71259657 14.78159500
#> [4,]  0.02314834  -3.6263597  14.78159500 13.50731211
#> 
#> , , 4
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.28126304 -0.4755565  0.1548701  0.04687147
#> [2,] -0.47555650  8.4615000 -7.6558175 -2.52106324
#> [3,]  0.15487015 -7.6558175 15.4636026 11.18052882
#> [4,]  0.04687147 -2.5210632 11.1805288 11.08587502
#> 
#> , , 5
#> 
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  0.4624282 -0.409591 -0.2048325 -0.292154
#> [2,] -0.4095910  6.136201 -2.3629995  1.912907
#> [3,] -0.2048325 -2.362999  9.8174955  8.710915
#> [4,] -0.2921540  1.912907  8.7109153 11.149096
#> 
#> 
#> $nu
#>          [,1]
#> [1,] 5.099823
#> [2,] 5.099859
#> [3,] 5.099808
#> [4,] 5.099744
#> [5,] 5.099503
#> 
#> $m
#>            [,1]
#> [1,]  1.4489040
#> [2,]  0.2501060
#> [3,]  0.7564447
#> [4,]  1.1176312
#> [5,] -0.1554274
#> 
#> $w
#>          [,1]
#> [1,] 52.02458
#> [2,] 17.02613
#> [3,] 52.82913
#> [4,] 34.02773
#> [5,] 26.75791
#> 
#> $s
#>           [,1]
#> [1,] 0.5766302
#> [2,] 1.2502552
#> [3,] 1.9930100
#> [4,] 1.3503348
#> [5,] 0.9220370
#> 
#> $scale
#>            [,1]
#> [1,] 0.01427928
#> [2,] 0.01427928
#> [3,] 0.01671019
#> [4,] 0.01897676
#> [5,] 0.02116359
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.20162208 -0.01372080  0.01043414  0.11007890
#> [2,] -0.01372080  0.27551630 -0.15371462 -0.08535248
#> [3,]  0.01043414 -0.15371462  0.31222165  0.23438614
#> [4,]  0.11007890 -0.08535248  0.23438614  0.33459738
#> 
#> , , 2, 1
#> 
#>             [,1]         [,2]        [,3]         [,4]
#> [1,]  0.27874942 -0.086941815  0.06475942  0.013661286
#> [2,] -0.08694181  0.235354555 -0.14404133 -0.008065417
#> [3,]  0.06475942 -0.144041333  0.47283895  0.327753121
#> [4,]  0.01366129 -0.008065417  0.32775312  0.400612004
#> 
#> , , 3, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.15985554 -0.5197276  0.3140968 -0.09690311
#> [2,] -0.51972756  8.5360846 -4.8468527  1.47470317
#> [3,]  0.31409680 -4.8468527  4.6918899  1.49687849
#> [4,] -0.09690311  1.4747032  1.4968785  3.28718793
#> 
#> , , 4, 1
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.05444652 -0.01964156  0.01452654 0.01075716
#> [2,] -0.01964156  0.18424513 -0.08399631 0.00856729
#> [3,]  0.01452654 -0.08399631  0.30006146 0.22016189
#> [4,]  0.01075716  0.00856729  0.22016189 0.31151607
#> 
#> , , 5, 1
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.05027292 -0.08050027  0.08486041  0.05540144
#> [2,] -0.08050027  4.44134221 -4.07239662 -1.70335453
#> [3,]  0.08486041 -4.07239662  4.67733316  2.54493212
#> [4,]  0.05540144 -1.70335453  2.54493212  1.79476918
#> 
#> , , 1, 2
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.050205042 -0.001190363  0.01964931  0.04954347
#> [2,] -0.001190363  0.232603085 -0.23673634 -0.13437501
#> [3,]  0.019649313 -0.236736344  0.40700340  0.31676763
#> [4,]  0.049543471 -0.134375011  0.31676763  0.33816024
#> 
#> , , 2, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.04329529 -0.02559305 -0.04214146 -0.04626158
#> [2,] -0.02559305  0.17412704 -0.06975696  0.05011799
#> [3,] -0.04214146 -0.06975696  0.37971151  0.33300719
#> [4,] -0.04626158  0.05011799  0.33300719  0.40591406
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.0245797 -0.1110624  0.2163619  0.166471
#> [2,] -0.1110624  6.8943834 -5.6448504 -1.086817
#> [3,]  0.2163619 -5.6448504 12.7502704 10.187240
#> [4,]  0.1664710 -1.0868167 10.1872395 10.875862
#> 
#> , , 4, 2
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.009901866  0.003115041 -0.002413866  0.004676676
#> [2,]  0.003115041  0.250902495 -0.217958789 -0.026736975
#> [3,] -0.002413866 -0.217958789  0.361048462  0.199490864
#> [4,]  0.004676676 -0.026736975  0.199490864  0.226300858
#> 
#> , , 5, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.01000609 -0.02040625  0.04420173  0.03786319
#> [2,] -0.02040625  2.74392975 -2.04618153 -0.52603766
#> [3,]  0.04420173 -2.04618153  2.31157418  1.27831977
#> [4,]  0.03786319 -0.52603766  1.27831977  1.15095226
#> 
#> , , 1, 3
#> 
#>              [,1]       [,2]        [,3]         [,4]
#> [1,]  0.029007874  0.0143977 -0.01756142  0.004753024
#> [2,]  0.014397703  0.2453283 -0.18433993 -0.063904100
#> [3,] -0.017561417 -0.1843399  0.31031306  0.238434980
#> [4,]  0.004753024 -0.0639041  0.23843498  0.249893885
#> 
#> , , 2, 3
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.02299417 -0.02345808 -0.03669213 -0.04418573
#> [2,] -0.02345808  0.14940619 -0.02682125  0.07802897
#> [3,] -0.03669213 -0.02682125  0.34364790  0.34351722
#> [4,] -0.04418573  0.07802897  0.34351722  0.43633500
#> 
#> , , 3, 3
#> 
#>             [,1]      [,2]       [,3]       [,4]
#> [1,]  0.01293094 -0.119365  0.2439778  0.1965066
#> [2,] -0.11936502  9.560579 -9.4809127 -3.6614035
#> [3,]  0.24397783 -9.480913 14.4599411  9.5962222
#> [4,]  0.19650657 -3.661403  9.5962222  8.4605794
#> 
#> , , 4, 3
#> 
#>              [,1]         [,2]        [,3]         [,4]
#> [1,]  0.004960231 -0.005978392  0.01126296  0.009276828
#> [2,] -0.005978392  0.282047655 -0.23190327 -0.031727460
#> [3,]  0.011262962 -0.231903266  0.40705084  0.247004857
#> [4,]  0.009276828 -0.031727460  0.24700486  0.239830759
#> 
#> , , 5, 3
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.004049771 -0.06014819  0.05618239  0.02425179
#> [2,] -0.060148190  3.66843907 -2.68091139 -0.54171130
#> [3,]  0.056182391 -2.68091139  2.68502774  1.18327930
#> [4,]  0.024251787 -0.54171130  1.18327930  0.94763136
#> 
#> , , 1, 4
#> 
#>              [,1]         [,2]        [,3]         [,4]
#> [1,]  0.018087977  0.005631908 -0.01994423 -0.009285448
#> [2,]  0.005631908  0.189657114 -0.14515877 -0.060087485
#> [3,] -0.019944229 -0.145158767  0.25550691  0.206880157
#> [4,] -0.009285448 -0.060087485  0.20688016  0.213133793
#> 
#> , , 2, 4
#> 
#>              [,1]        [,2]         [,3]         [,4]
#> [1,]  0.015734235 -0.02788811  0.006709026 -0.008253394
#> [2,] -0.027888113  0.19350565 -0.156933395 -0.029103318
#> [3,]  0.006709026 -0.15693340  0.408454241  0.331992019
#> [4,] -0.008253394 -0.02910332  0.331992019  0.351048943
#> 
#> , , 3, 4
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.02205023 -0.04728194  0.3469292  0.3725809
#> [2,] -0.04728194 10.99104619 -6.8434295  0.4888847
#> [3,]  0.34692923 -6.84342948 12.4745178  9.3888843
#> [4,]  0.37258090  0.48888472  9.3888843 11.4902109
#> 
#> , , 4, 4
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.002487934 -0.005704235  0.002632544 -0.001528447
#> [2,] -0.005704235  0.156117031 -0.139287156 -0.020536210
#> [3,]  0.002632544 -0.139287156  0.277638739  0.182327489
#> [4,] -0.001528447 -0.020536210  0.182327489  0.185535696
#> 
#> , , 5, 4
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.005305711 -0.08613707  0.07083122  0.02255539
#> [2,] -0.086137068  3.28651331 -2.43036822 -0.51149833
#> [3,]  0.070831216 -2.43036822  2.63425903  1.30912885
#> [4,]  0.022555391 -0.51149833  1.30912885  1.12023868
#> 
#> , , 1, 5
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.02564301  0.01082943 -0.02762579 -0.01331826
#> [2,]  0.01082943  0.24163732 -0.20870753 -0.09204244
#> [3,] -0.02762579 -0.20870753  0.32063808  0.23255557
#> [4,] -0.01331826 -0.09204244  0.23255557  0.21470817
#> 
#> , , 2, 5
#> 
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.012793376 -0.01565302 -0.004836173 -0.01466823
#> [2,] -0.015653024  0.14645521 -0.162413805 -0.05837555
#> [3,] -0.004836173 -0.16241381  0.644900968  0.58154267
#> [4,] -0.014668228 -0.05837555  0.581542666  0.60329710
#> 
#> , , 3, 5
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.01605639 -0.1284435  0.2450121  0.1892779
#> [2,] -0.12844351  6.8446447 -5.1511122 -0.7735315
#> [3,]  0.24501211 -5.1511122  8.3122692  5.8483912
#> [4,]  0.18927792 -0.7735315  5.8483912  6.3604503
#> 
#> , , 4, 5
#> 
#>              [,1]          [,2]         [,3]         [,4]
#> [1,] 0.0024874425  0.0008910944  0.001379895 0.0014997519
#> [2,] 0.0008910944  0.1746177923 -0.128388679 0.0006474625
#> [3,] 0.0013798946 -0.1283886791  0.336718191 0.2538636000
#> [4,] 0.0014997519  0.0006474625  0.253863600 0.2717720718
#> 
#> , , 5, 5
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.00383175 -0.04868002  0.04863922  0.02229762
#> [2,] -0.04868002  2.88805185 -2.31120200 -0.62171917
#> [3,]  0.04863922 -2.31120200  2.72662931  1.47022418
#> [4,]  0.02229762 -0.62171917  1.47022418  1.21572362
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.6743615  0.15943442  0.1325378 -0.07197774
#> [2,]  0.6532055  0.89186049  0.0406092  0.14118901
#> [3,]  1.2412443 -0.17536753  2.1455101  1.28622473
#> [4,] -1.0523568  0.10553003 -1.1665830 -0.28822390
#> [5,] -1.9774840 -0.03482156  1.8420683  5.33014575
#> 
#> , , 2, 1
#> 
#>            [,1]        [,2]       [,3]      [,4]
#> [1,]  0.8165565  0.11874585  0.1949796 0.1002304
#> [2,]  1.0001314  0.04495297  0.7231401 0.5048710
#> [3,]  0.7123504 -1.14776957  1.4427838 0.3165240
#> [4,] -0.7410160  1.12695457 -0.5997608 0.5369213
#> [5,] -0.6985312 -0.13243055  0.9015679 4.4546435
#> 
#> , , 3, 1
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.7860821 -1.369112  0.8192354 -0.1014297
#> [2,]  0.4314755  2.805017 -1.1138309  0.1836361
#> [3,]  0.7170137  3.750018 -1.1250319  0.5271069
#> [4,] -0.5789170 -3.177011  1.6389555  0.3144695
#> [5,] -3.2900982  4.313655  7.6713089 13.9677823
#> 
#> , , 4, 1
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.7982004 -0.3751005  0.88227479  0.28639305
#> [2,]  0.2897002  1.0337478 -0.33384800 -0.04412284
#> [3,]  0.6553898  0.3130072  0.64776352  0.49787968
#> [4,] -0.5527171 -0.1882788  0.02907384  0.34349012
#> [5,] -2.0560315  0.8279931  2.03857514  6.00464669
#> 
#> , , 5, 1
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.8720708  0.10684347  0.5112227  0.3747028
#> [2,]  0.4495026  0.78297126  0.2257544  0.5971350
#> [3,]  0.7318521 -0.05777651  1.1236897  0.9531596
#> [4,] -0.6533384  0.17859122 -0.6212386 -0.3214037
#> [5,] -1.3357711 -8.80330307 15.2999393 12.3442837
#> 
#> , , 1, 2
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.7116259  0.2737978 -0.3567679 -0.33833299
#> [2,]  0.7279054  0.9937486  0.1013232  0.41495713
#> [3,]  1.4407224  0.0961699  1.3466394  1.21805472
#> [4,] -1.2431204 -0.1468288 -0.2334900 -0.07242974
#> [5,] -3.0059533 -3.4298264  3.3395177  2.25693911
#> 
#> , , 2, 2
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.7010029  0.1405521  0.29401054  0.2407123
#> [2,]  0.9058646  0.1754652 -0.17952495 -0.3541256
#> [3,]  1.2949516 -0.8706821 -0.02520569 -1.1708716
#> [4,] -1.1516381  0.8429478  0.75384143  1.8829628
#> [5,] -2.2004292  1.2518323  3.81216929  7.4347571
#> 
#> , , 3, 2
#> 
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  0.9226076 -2.678805  0.2410846 -1.970584
#> [2,]  0.2718242  3.676201  1.1597808  3.634656
#> [3,]  0.4207317  4.986340  2.6362899  6.108506
#> [4,] -0.3603082 -3.891594 -1.9009989 -4.534843
#> [5,] -2.1448175  1.157326  6.8761763 10.843082
#> 
#> , , 4, 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.8794082 -0.61738859  0.8831984 0.11548117
#> [2,]  0.7154672  0.83036646 -0.4744765 0.01572792
#> [3,]  0.9215185  0.03186522  0.4824105 0.26482369
#> [4,] -0.8555296  0.17399288  0.1936942 0.64675748
#> [5,] -1.8380564  0.83569407  1.9708038 4.36191071
#> 
#> , , 5, 2
#> 
#>            [,1]         [,2]       [,3]       [,4]
#> [1,]  0.8844676 -0.818277640  0.8454049  0.2434811
#> [2,]  0.4610429  0.751313298  0.7211243  1.1597019
#> [3,]  0.7597713  0.007535879  1.8399790  1.7657239
#> [4,] -0.6801471  0.471090243 -1.3667917 -1.0040571
#> [5,] -1.6844239 -4.003676717  6.6421295  6.8032064
#> 
#> , , 1, 3
#> 
#>            [,1]          [,2]        [,3]       [,4]
#> [1,]  0.7787837  0.1972530289 -0.09413159 -0.2184713
#> [2,]  0.8312788  0.9430955206  0.04337099  0.3751963
#> [3,]  1.6927268 -0.0297694110  1.63784425  1.3712724
#> [4,] -1.4652390 -0.0008594764 -0.65633572 -0.3138549
#> [5,] -5.8597195 -2.6756982095  4.99544539  4.4029983
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.7909348  0.0638917  0.29439286  0.2336139
#> [2,]  1.0357305  0.1089141 -0.07623621 -0.1811965
#> [3,]  1.2701832 -0.9848552  0.80000603 -0.3122707
#> [4,] -1.1650822  1.0058153  0.03438881  1.1499456
#> [5,] -3.8207807 -1.1046021  3.95631514  5.3901905
#> 
#> , , 3, 3
#> 
#>            [,1]      [,2]      [,3]      [,4]
#> [1,]  0.8138897 -1.685485 -1.298788 -3.122507
#> [2,]  0.3053926  2.001941  2.584214  3.938837
#> [3,]  0.4562516  2.562163  4.570620  6.411260
#> [4,] -0.3952371 -1.950725 -3.292067 -4.575045
#> [5,]  0.1421761 10.549558  5.871843 19.789855
#> 
#> , , 4, 3
#> 
#>            [,1]        [,2]        [,3]      [,4]
#> [1,]  0.9096694 -0.03707038  1.02500446 0.6329286
#> [2,]  0.7022843  0.86858440 -0.34670227 0.1312398
#> [3,]  0.9220796  0.39031145  0.62502337 0.6156980
#> [4,] -0.8669835 -0.33883696  0.01695664 0.1523525
#> [5,] -1.8119078 -1.92227641  0.71629719 2.0423618
#> 
#> , , 5, 3
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  0.9229843 -1.4434101  1.847231  0.8328916
#> [2,]  0.4761271  0.2238933  1.028934  1.1325601
#> [3,]  0.7831072 -0.8357785  2.312519  1.7410961
#> [4,] -0.7110534  1.2396051 -2.018069 -1.1822981
#> [5,] -2.2769992 17.0649365 -9.134707  3.7189661
#> 
#> , , 1, 4
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  0.7871644  0.21070555  0.01587086 -0.03698748
#> [2,]  0.7864477  1.00525815  0.03888290  0.39002273
#> [3,]  1.6678141 -0.09472974  2.08278970  1.91553491
#> [4,] -1.4556949  0.07746843 -1.08405313 -0.84580906
#> [5,] -5.0304582 -4.33425708  3.33310860  1.55840709
#> 
#> , , 2, 4
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.7986091  0.09400518 -0.01595854 -0.1082066
#> [2,]  1.0139116  0.17744935 -0.26985001 -0.3485139
#> [3,]  1.3858796 -1.10880147  0.93090639 -0.1856847
#> [4,] -1.2521925  1.07702471  0.06385092  1.1753410
#> [5,] -4.4368501 -0.40551674  4.25129836  6.6605323
#> 
#> , , 3, 4
#> 
#>            [,1]        [,2]      [,3]      [,4]
#> [1,]  0.7750163  -0.8294958 -2.901893 -4.319920
#> [2,]  0.3983716   3.2594402  3.193060  5.624444
#> [3,]  0.6212556   4.4493497  5.758029  9.277664
#> [4,] -0.5370752  -3.3899961 -4.514885 -7.143685
#> [5,] -0.2766722 -37.3236025 45.736351 29.627108
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.9446975  0.6306625  0.4825508  0.68959826
#> [2,]  0.7548140  1.2766236 -0.3985998  0.44021259
#> [3,]  0.9662555  0.8972190  0.3730116  0.82665954
#> [4,] -0.9299231 -1.0469301  0.4631855 -0.04766734
#> [5,] -1.3318142 -4.0988987 -0.1773703 -0.57192486
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.8821939 -1.2160342   0.8779641 -0.07184231
#> [2,]  0.4792656  0.1269603   1.4450008  1.53079967
#> [3,]  0.8003518 -1.1457689   3.2797285  2.61213112
#> [4,] -0.7149170  1.4456492  -2.4416543 -1.52464484
#> [5,] -1.9176786 16.1429778 -14.1385956 -2.63980101
#> 
#> , , 1, 5
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.7842779  0.1000625  0.33540727  0.2551122
#> [2,]  0.7779846  1.1336272 -0.02619236  0.3609367
#> [3,]  1.7184609  0.2822021  1.96373850  1.9732267
#> [4,] -1.5142047 -0.2567442 -1.01573335 -0.9455087
#> [5,] -4.3966310 -3.4205365 -1.42080227 -2.6303003
#> 
#> , , 2, 5
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.7837663  0.06538231  0.3100216 0.20313921
#> [2,]  0.9943396  0.04778211  0.2764083 0.10113493
#> [3,]  1.4061322 -1.01709040  1.0782940 0.04139819
#> [4,] -1.2769338  1.08793719 -0.3169146 0.78625858
#> [5,] -3.1741647 -4.19448693  7.0599842 7.04335061
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.8698756 -2.0603503 -0.7224463 -2.722802
#> [2,]  0.2539271  1.0538705  2.7093190  3.470518
#> [3,]  0.3916120  1.0728776  4.9355605  5.895543
#> [4,] -0.3484836 -0.5195471 -3.8756628 -4.330642
#> [5,]  0.1134506 23.0783579  7.2874332 29.156633
#> 
#> , , 4, 5
#> 
#>            [,1]        [,2]       [,3]      [,4]
#> [1,]  0.9647211 -0.01666868  0.5934201 0.3902406
#> [2,]  0.7893379  0.88336527 -0.2854976 0.3338782
#> [3,]  0.9925512  0.40063449  0.4358011 0.6123244
#> [4,] -0.9646259 -0.34271660  0.3011606 0.2175428
#> [5,] -1.2490104 -2.84448794  4.2267653 3.2415478
#> 
#> , , 5, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9213139 -1.4619025   1.598065  0.5781545
#> [2,]  0.4675955  0.4641137   0.979192  1.2396400
#> [3,]  0.7625505 -0.6283737   2.309143  1.8792592
#> [4,] -0.6995000  0.9807293  -1.871688 -1.2039022
#> [5,] -1.7326214 19.7002408 -10.707412  3.3049730
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_last_draw`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|