R6 Class Representing PosteriorBVARPANEL

The class PosteriorBVARPANEL contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvarPANEL$new()

specify_posterior_bvarPANEL$get_posterior()

specify_posterior_bvarPANEL$get_last_draw()

specify_posterior_bvarPANEL$clone()

Method new()

Create a new posterior output PosteriorBVARPANEL.

Usage

specify_posterior_bvarPANEL$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARPANEL.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvarPANEL$get_posterior()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvarPANEL$get_last_draw()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvarPANEL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARPANEL" "R6"                

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_posterior`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $A
#> , , 1
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.8313683 -0.10131088 -0.8550040 -0.1335969
#> [2,] -0.6923711  0.28490129  0.4750058 -0.5411260
#> [3,]  0.6005988 -0.54227877  1.4570229  1.4240324
#> [4,] -1.7785199 -0.07479582  1.5395671  0.1684289
#> [5,]  0.9338683  0.29510895  0.6296473 -0.8590804
#> 
#> , , 2
#> 
#>             [,1]        [,2]       [,3]      [,4]
#> [1,]  0.82765979  0.03905905  0.2041633 0.1107199
#> [2,]  0.13109047  0.58625336  0.2254701 0.5575242
#> [3,]  0.17229325 -0.41896496  0.9319803 0.6804866
#> [4,] -0.06010564  0.64356604 -0.3517449 0.2435981
#> [5,]  0.18468845  1.80312615  2.7819320 0.7521842
#> 
#> , , 3
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  1.05784330  0.1117767  0.63649232  0.6402111
#> [2,] -0.25460053  0.8000787 -0.22337872  0.3099866
#> [3,] -0.08300452  0.2604392  0.72035914  0.8012294
#> [4,]  0.13110602 -0.3510117 -0.09477719 -0.4212310
#> [5,]  1.58738398  2.6741335  0.37427952 -1.8233370
#> 
#> , , 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8841396 -0.7125813  0.7278642  0.2507407
#> [2,] -0.3148243  0.2378700  0.9350262  1.1590156
#> [3,] -0.2471456 -0.7839409  2.5889584  2.0946680
#> [4,]  0.2636071  1.0350106 -1.7320736 -1.0030955
#> [5,]  1.8059946  2.0786499  0.5703923 -1.7949320
#> 
#> , , 5
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9402393 -0.13613464  0.6021895  0.6045736
#> [2,] -0.1444827  0.92718193  1.4679541  2.1779462
#> [3,] -0.2038920 -0.09837387  2.4918046  2.4706116
#> [4,]  0.2373256 -0.28009148 -1.4140771 -1.5626317
#> [5,]  1.0767423 -1.82734038  1.6045128 -3.5157698
#> 
#> 
#> $V
#> , , 1
#> 
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,]  6.4822498  0.1751257  1.1298833 -2.1647551 -4.757456
#> [2,]  0.1751257 11.2642653 -0.2332166  2.6650931 -7.218188
#> [3,]  1.1298833 -0.2332166  8.2679225  0.2096118 -2.467620
#> [4,] -2.1647551  2.6650931  0.2096118  9.4075027 -1.959169
#> [5,] -4.7574564 -7.2181885 -2.4676197 -1.9591693 55.543412
#> 
#> , , 2
#> 
#>            [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,]  2.1665354 -0.5704026 -1.2860682  1.029270 -1.6673112
#> [2,] -0.5704026  1.5835655  0.6703304 -0.804303 -0.3032216
#> [3,] -1.2860682  0.6703304  2.1131944 -1.171385 -0.4241404
#> [4,]  1.0292699 -0.8043030 -1.1713854  1.809639 -1.1983946
#> [5,] -1.6673112 -0.3032216 -0.4241404 -1.198395 33.3955215
#> 
#> , , 3
#> 
#>            [,1]      [,2]       [,3]       [,4]       [,5]
#> [1,]  5.9034664 -1.536803 -1.0800448 -0.7644546  1.3455788
#> [2,] -1.5368029  4.557006  1.7431141 -1.6502859 -1.2314462
#> [3,] -1.0800448  1.743114  3.5407589 -2.0344333 -0.5353785
#> [4,] -0.7644546 -1.650286 -2.0344333  5.6613713  3.9304277
#> [5,]  1.3455788 -1.231446 -0.5353785  3.9304277 52.4182948
#> 
#> , , 4
#> 
#>           [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,]  2.086677   1.124861   1.172189 -1.604797  -7.765252
#> [2,]  1.124861   4.757656   4.956316 -6.172921 -11.113150
#> [3,]  1.172189   4.956316   7.473462 -9.070741 -10.344178
#> [4,] -1.604797  -6.172921  -9.070741 12.459264  14.786223
#> [5,] -7.765252 -11.113150 -10.344178 14.786223  73.106446
#> 
#> , , 5
#> 
#>             [,1]       [,2]        [,3]       [,4]       [,5]
#> [1,]  2.34960616  0.7131245  0.08982953 -0.4791720 -4.0066781
#> [2,]  0.71312453  5.4242297  3.85997604 -3.4932328 -4.4192732
#> [3,]  0.08982953  3.8599760  5.67720252 -4.0823953 -3.1146907
#> [4,] -0.47917202 -3.4932328 -4.08239525  6.8266189  0.3034812
#> [5,] -4.00667814 -4.4192732 -3.11469066  0.3034812 47.9274025
#> 
#> 
#> $Sigma
#> , , 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  8.5113702 -0.7711811  1.6127384 -2.1111173
#> [2,] -0.7711811 11.1638077  0.7577557  0.6395437
#> [3,]  1.6127384  0.7577557  6.4140457 -0.9289771
#> [4,] -2.1111173  0.6395437 -0.9289771  5.3496610
#> 
#> , , 2
#> 
#>            [,1]      [,2]      [,3]       [,4]
#> [1,]  4.2673881 -2.496269  2.016860  0.8068195
#> [2,] -2.4962688  8.801564 -6.163446 -1.5324568
#> [3,]  2.0168604 -6.163446 13.359598  8.4934163
#> [4,]  0.8068195 -1.532457  8.493416 10.1721150
#> 
#> , , 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.0961784 -0.3457323  0.5396784  0.4413476
#> [2,] -0.3457323  8.6338186 -6.5560374 -0.8258586
#> [3,]  0.5396784 -6.5560374 10.5035421  7.1595655
#> [4,]  0.4413476 -0.8258586  7.1595655  8.4993102
#> 
#> , , 4
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.4984900 -0.3268941  0.6001632 0.3787733
#> [2,] -0.3268941  4.4710904 -2.5527877 0.5352699
#> [3,]  0.6001632 -2.5527877  4.1685599 2.8175285
#> [4,]  0.3787733  0.5352699  2.8175285 3.9734692
#> 
#> , , 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.4024266 -0.4595491  0.2154088 -0.1602562
#> [2,] -0.4595491  4.6170473 -2.2722993  0.6428091
#> [3,]  0.2154088 -2.2722993  7.3194926  6.6246398
#> [4,] -0.1602562  0.6428091  6.6246398  8.0388258
#> 
#> 
#> $nu
#>          [,1]
#> [1,] 5.100026
#> [2,] 5.100035
#> [3,] 5.100046
#> [4,] 5.099878
#> [5,] 5.100019
#> 
#> $m
#>           [,1]
#> [1,] 1.1122551
#> [2,] 2.6037912
#> [3,] 0.3854686
#> [4,] 1.7833411
#> [5,] 2.2127243
#> 
#> $w
#>           [,1]
#> [1,] 174.96066
#> [2,]  17.22652
#> [3,]  48.21964
#> [4,]  19.31817
#> [5,]  35.86598
#> 
#> $s
#>           [,1]
#> [1,] 0.2502062
#> [2,] 0.8173755
#> [3,] 0.8465817
#> [4,] 0.4852323
#> [5,] 0.4280774
#> 
#> $scale
#>            [,1]
#> [1,] 0.01427928
#> [2,] 0.01427928
#> [3,] 0.01666118
#> [4,] 0.01892110
#> [5,] 0.02100835
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.047135066 -0.009246054 -0.004372143 -0.001181954
#> [2,] -0.009246054  0.301591106 -0.302667076 -0.144637334
#> [3,] -0.004372143 -0.302667076  0.520785051  0.353957965
#> [4,] -0.001181954 -0.144637334  0.353957965  0.337599725
#> 
#> , , 2, 1
#> 
#>               [,1]        [,2]          [,3]        [,4]
#> [1,]  0.0492572461 -0.03955165 -0.0002748318 -0.02268875
#> [2,] -0.0395516462  0.20945106 -0.0746854942  0.05118045
#> [3,] -0.0002748318 -0.07468549  0.2929969590  0.26612094
#> [4,] -0.0226887490  0.05118045  0.2661209363  0.37642596
#> 
#> , , 3, 1
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.04598905  0.05826551 -0.01424102 0.01594363
#> [2,]  0.05826551  7.44601333 -3.84598253 1.38635205
#> [3,] -0.01424102 -3.84598253  2.97099778 0.43440128
#> [4,]  0.01594363  1.38635205  0.43440128 1.71172933
#> 
#> , , 4, 1
#> 
#>             [,1]         [,2]        [,3]         [,4]
#> [1,]  0.06263883  0.019835433 -0.01299595  0.014750660
#> [2,]  0.01983543  0.172244638 -0.10045028 -0.007130609
#> [3,] -0.01299595 -0.100450277  0.28977254  0.172757147
#> [4,]  0.01475066 -0.007130609  0.17275715  0.168063889
#> 
#> , , 5, 1
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.050498152 -0.004843998  0.05731564  0.06671521
#> [2,] -0.004843998  2.816935367 -2.00099468 -0.26129637
#> [3,]  0.057315637 -2.000994676  2.67341768  1.43625112
#> [4,]  0.066715214 -0.261296374  1.43625112  1.32518464
#> 
#> , , 1, 2
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.017605903 -0.001630641 -0.02303463 -0.02287360
#> [2,] -0.001630641  0.176047153 -0.19070347 -0.09605327
#> [3,] -0.023034626 -0.190703466  0.53617449  0.45573526
#> [4,] -0.022873600 -0.096053266  0.45573526  0.43994830
#> 
#> , , 2, 2
#> 
#>             [,1]         [,2]        [,3]         [,4]
#> [1,]  0.01988269 -0.028230179 -0.01602766 -0.034884608
#> [2,] -0.02823018  0.182805381 -0.11819643 -0.004699712
#> [3,] -0.01602766 -0.118196425  0.59791503  0.593807168
#> [4,] -0.03488461 -0.004699712  0.59380717  0.686318873
#> 
#> , , 3, 2
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.0234710 -0.215545 -0.1186376 -0.3073557
#> [2,] -0.2155450 12.029132 -2.9533563  5.8914200
#> [3,] -0.1186376 -2.953356  7.3190639  6.3251121
#> [4,] -0.3073557  5.891420  6.3251121 12.0745982
#> 
#> , , 4, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.04954454 -0.02754047 -0.01280369 -0.03910081
#> [2,] -0.02754047  0.18763879 -0.18391509 -0.04701568
#> [3,] -0.01280369 -0.18391509  0.40341089  0.27865869
#> [4,] -0.03910081 -0.04701568  0.27865869  0.26731719
#> 
#> , , 5, 2
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.037829426 -0.06012048  0.04154537 -0.006659055
#> [2,] -0.060120476  3.66082125 -2.55231615 -0.370332556
#> [3,]  0.041545375 -2.55231615  2.62343718  1.207353976
#> [4,] -0.006659055 -0.37033256  1.20735398  1.120924657
#> 
#> , , 1, 3
#> 
#>              [,1]       [,2]        [,3]         [,4]
#> [1,]  0.011422561 -0.0186473  0.01189199  0.006252093
#> [2,] -0.018647298  0.3276548 -0.30432147 -0.161370029
#> [3,]  0.011891992 -0.3043215  0.51993984  0.408113156
#> [4,]  0.006252093 -0.1613700  0.40811316  0.371760232
#> 
#> , , 2, 3
#> 
#>              [,1]         [,2]         [,3]        [,4]
#> [1,]  0.007331460 -0.007747925 -0.002699221 -0.01074417
#> [2,] -0.007747925  0.166866256 -0.153218824 -0.02926884
#> [3,] -0.002699221 -0.153218824  0.643770906  0.60969569
#> [4,] -0.010744167 -0.029268844  0.609695692  0.68257283
#> 
#> , , 3, 3
#> 
#>              [,1]       [,2]        [,3]        [,4]
#> [1,]  0.004356474 -0.0801977  0.04244799 -0.01310478
#> [2,] -0.080197704  8.9292759 -5.43235822  0.74978476
#> [3,]  0.042447986 -5.4323582  5.07863478  1.69065627
#> [4,] -0.013104779  0.7497848  1.69065627  2.68593959
#> 
#> , , 4, 3
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.013513328 -0.02321152  0.01697805 -0.000829361
#> [2,] -0.023211522  0.19567498 -0.16739035 -0.021156193
#> [3,]  0.016978050 -0.16739035  0.36723238  0.249980599
#> [4,] -0.000829361 -0.02115619  0.24998060  0.245542307
#> 
#> , , 5, 3
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.01071563 -0.1022131  0.1297871  0.07316258
#> [2,] -0.10221313  2.7243115 -2.4484216 -0.86885230
#> [3,]  0.12978709 -2.4484216  2.8067298  1.45376262
#> [4,]  0.07316258 -0.8688523  1.4537626  1.02776441
#> 
#> , , 1, 4
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.009888142 -0.03470149  0.05010492  0.03941745
#> [2,] -0.034701489  0.34430499 -0.41759267 -0.30771847
#> [3,]  0.050104918 -0.41759267  0.76557120  0.67178211
#> [4,]  0.039417447 -0.30771847  0.67178211  0.62400233
#> 
#> , , 2, 4
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.007483165 -0.01498653  0.01185152 -0.000212579
#> [2,] -0.014986525  0.11444540 -0.07389115  0.014130492
#> [3,]  0.011851523 -0.07389115  0.54617136  0.586571635
#> [4,] -0.000212579  0.01413049  0.58657164  0.715820629
#> 
#> , , 3, 4
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.004556456 -0.03829433  0.02937211  0.004882322
#> [2,] -0.038294335  8.51880046 -6.06207033 -0.441859367
#> [3,]  0.029372113 -6.06207033  8.68797130  5.581579703
#> [4,]  0.004882322 -0.44185937  5.58157970  6.384392973
#> 
#> , , 4, 4
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.01639200 -0.03198615  0.0372651  0.01229278
#> [2,] -0.03198615  0.16714307 -0.1733535 -0.04285555
#> [3,]  0.03726510 -0.17335351  0.3325565  0.19872585
#> [4,]  0.01229278 -0.04285555  0.1987259  0.16860042
#> 
#> , , 5, 4
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.02093601 -0.2032786  0.2679346  0.1528864
#> [2,] -0.20327864  3.9161210 -3.6382641 -1.3331110
#> [3,]  0.26793463 -3.6382641  4.6610545  2.6392165
#> [4,]  0.15288639 -1.3331110  2.6392165  1.9894153
#> 
#> , , 1, 5
#> 
#>               [,1]         [,2]          [,3]         [,4]
#> [1,]  0.0072124669 -0.008291907 -0.0002311256 -0.005183544
#> [2,] -0.0082919072  0.135984783 -0.1017336090 -0.049775143
#> [3,] -0.0002311256 -0.101733609  0.2544626615  0.244647336
#> [4,] -0.0051835435 -0.049775143  0.2446473356  0.260927589
#> 
#> , , 2, 5
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.008103470 -0.02160824  0.02468969  0.009396757
#> [2,] -0.021608236  0.17510363 -0.15410108 -0.023469584
#> [3,]  0.024689694 -0.15410108  0.41979956  0.365909573
#> [4,]  0.009396757 -0.02346958  0.36590957  0.424414986
#> 
#> , , 3, 5
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.003958349 -0.07893639  0.03877242 -0.01393717
#> [2,] -0.078936392 10.24910971 -8.40046804 -1.94495946
#> [3,]  0.038772416 -8.40046804  8.91101541  3.98830985
#> [4,] -0.013937167 -1.94495946  3.98830985  3.21263839
#> 
#> , , 4, 5
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.006679614  0.00597552 -0.02514145 -0.02185346
#> [2,]  0.005975520  0.16306271 -0.15052355 -0.03022197
#> [3,] -0.025141450 -0.15052355  0.41169532  0.30758548
#> [4,] -0.021853460 -0.03022197  0.30758548  0.29360264
#> 
#> , , 5, 5
#> 
#>              [,1]       [,2]       [,3]        [,4]
#> [1,]  0.006314052 -0.1218632  0.1215021  0.05285109
#> [2,] -0.121863209  5.1050407 -3.9834468 -0.98586322
#> [3,]  0.121502136 -3.9834468  3.8678881  1.62327711
#> [4,]  0.052851087 -0.9858632  1.6232771  1.15212045
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.98638274  0.3257214 -0.2080843  0.01786841
#> [2,]  0.09927283  0.8316396 -0.4587468 -0.28652216
#> [3,]  0.27955669 -0.3526774  0.6324740  0.23240694
#> [4,] -0.25246416  0.3192598  0.2538671  0.66856447
#> [5,] -0.60824251 -5.8062860 12.3266880  7.41959808
#> 
#> , , 2, 1
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.8383943  0.10808727  0.2727209 0.20760055
#> [2,]  0.3705465 -0.07258629  0.6707505 0.74811800
#> [3,]  0.5601164 -1.40342044  1.6796755 0.84823101
#> [4,] -0.5203628  1.38387248 -0.8255161 0.06187829
#> [5,]  1.8747804 -1.16491342  1.8421650 0.01881948
#> 
#> , , 3, 1
#> 
#>            [,1]      [,2]      [,3]       [,4]
#> [1,]  1.1580857 -2.405009  1.987514  0.1439447
#> [2,] -0.1982729  2.988329 -1.725733  0.3088153
#> [3,] -0.3011319  4.051605 -2.269352  0.5432015
#> [4,]  0.2572053 -2.980181  2.431718  0.4892302
#> [5,] -0.7348582 -2.199625  0.704097 -5.2645096
#> 
#> , , 4, 1
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.5136555 -0.2287041  0.7801720  0.62028213
#> [2,]  0.2664903  0.8430254 -0.1429117  0.33082193
#> [3,]  0.3944425  0.6167829  0.5214594  0.76426420
#> [4,] -0.2506104 -0.5204323  0.2213140  0.07471779
#> [5,]  1.8669907  0.1715753 -1.1703638 -3.24499019
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.1407087 -1.4660648  1.6532417  0.6663759
#> [2,] -0.5351068  1.3771646 -1.0473641 -0.1597141
#> [3,] -0.8735363  0.7887573 -0.7608388 -0.1959340
#> [4,]  0.7670620 -0.2135991  0.9914916  0.8692489
#> [5,]  2.7579946  5.7811886  1.8581544  2.3289790
#> 
#> , , 1, 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.8432668 -0.05496115  0.8820606  0.8086929
#> [2,]  0.2487668  1.09773795 -1.6560505 -1.2972880
#> [3,]  0.4051202  0.03863805 -1.0540646 -1.2983199
#> [4,] -0.3267090 -0.03065234  1.5827980  1.9204034
#> [5,] -0.5386725  0.38618174  7.3074684  4.6310674
#> 
#> , , 2, 2
#> 
#>            [,1]        [,2]         [,3]       [,4]
#> [1,]  1.0521093 -0.29790839  0.006093333 -0.3048207
#> [2,] -0.3347644  0.86148080  2.757063078  3.7270154
#> [3,] -0.3699586 -0.02721979  3.774140006  4.0623047
#> [4,]  0.3492672  0.08634348 -2.859476375 -3.0262310
#> [5,]  1.2109800  4.64124844 -3.955278714 -5.3596387
#> 
#> , , 3, 2
#> 
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  1.0194856 -2.199247 -0.6514693 -2.634117
#> [2,] -0.1011978  2.669695  2.0660313  4.244077
#> [3,] -0.1681106  3.767491  3.6470686  6.861970
#> [4,]  0.1470444 -2.883772 -2.8667549 -5.306948
#> [5,]  0.8239090  6.479718 20.2172623 24.452328
#> 
#> , , 4, 2
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.30055611 -0.2263924  1.2286341  1.1120427
#> [2,]  0.17872500  0.3433017 -0.3977456 -0.1782501
#> [3,]  0.26991415 -0.5298538  0.3171370 -0.1351004
#> [4,] -0.06342195  0.5519384  0.2845692  0.7767129
#> [5,]  2.65180328  4.9372770 -2.1148117 -1.5664586
#> 
#> , , 5, 2
#> 
#>            [,1]      [,2]      [,3]        [,4]
#> [1,]  1.0998151 -2.699723  2.296256  0.67219372
#> [2,] -0.4801919  1.987061 -1.057708  0.18608507
#> [3,] -0.7782424  2.177217 -1.270241  0.08431503
#> [4,]  0.7285073 -1.156028  1.313749  0.63994890
#> [5,]  0.4554284 13.353985 -6.281163 -2.37250389
#> 
#> , , 1, 3
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.95925358  0.13249463 -0.07514195 -0.09387768
#> [2,] -0.01277784  0.94620553 -0.07412385  0.28024498
#> [3,] -0.07161682  0.01498709  1.72026251  1.58904793
#> [4,]  0.06850860 -0.03843403 -0.78821035 -0.55937918
#> [5,]  0.94242579 -1.39099036  8.23470201  4.52717873
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  1.0455772 -0.1273549 -0.01652353 -0.2237376
#> [2,] -0.3495922  0.9762553  2.52118669  3.5553921
#> [3,] -0.4636362  0.3155359  3.88704165  4.5061532
#> [4,]  0.4340055 -0.2023982 -2.94662253 -3.3887722
#> [5,]  1.0718971 -1.3403041 -0.03432354 -5.2317503
#> 
#> , , 3, 3
#> 
#>             [,1]      [,2]        [,3]       [,4]
#> [1,]  1.01638874 -2.767209   1.1910226 -0.8553175
#> [2,] -0.07682787  1.749809   0.2901054  1.3205450
#> [3,] -0.12465516  2.241123   0.9393101  2.3058512
#> [4,]  0.10296377 -1.549682  -0.1803725 -0.9448915
#> [5,]  0.99490743 31.043912 -18.0654476 -1.3127478
#> 
#> , , 4, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.6895623 -0.3093896  1.3240877  1.1380816
#> [2,] -0.6407429  0.5081193 -0.1567959  0.2399951
#> [3,] -0.7524205 -0.4070042  1.3565224  1.0883939
#> [4,]  0.8298908  0.5318361 -0.8518666 -0.4593595
#> [5,]  2.3807947 -0.8981898  4.2950258  0.1402409
#> 
#> , , 5, 3
#> 
#>            [,1]      [,2]       [,3]        [,4]
#> [1,]  1.1435589 -2.570310   2.772694   1.2222277
#> [2,] -0.4199079  2.225915  -1.189713   0.2298871
#> [3,] -0.6986673  2.542314  -1.506694   0.1201382
#> [4,]  0.6175246 -1.631497   1.492974   0.4773460
#> [5,]  1.0339935 17.094294 -16.445138 -10.0360308
#> 
#> , , 1, 4
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.92213396  0.55661263 -0.6132065 -0.4735641
#> [2,]  0.05635484 -0.04717994  1.3308068  1.3498740
#> [3,]  0.14440416 -2.40875852  4.7496171  3.7656235
#> [4,] -0.11859566  2.11638751 -3.4273966 -2.4328755
#> [5,]  0.73942089  0.43338017  2.2284551 -1.5014738
#> 
#> , , 2, 4
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9593750 -0.03080049 -0.6169839 -0.7967207
#> [2,] -0.2271984  0.78198419  3.2180130  4.1361942
#> [3,] -0.2306154 -0.22955635  4.9971315  5.1945030
#> [4,]  0.2446728  0.25876215 -3.8855594 -3.9761102
#> [5,]  0.7870178 -0.83789393  4.2955815  0.5305653
#> 
#> , , 3, 4
#> 
#>             [,1]      [,2]        [,3]       [,4]
#> [1,]  0.97198473 -2.412713   1.2619542 -0.4949005
#> [2,] -0.03033186  1.437071   0.5202898  1.3574910
#> [3,] -0.05575321  2.020035   0.9129410  2.1073110
#> [4,]  0.04603579 -1.486275  -0.1731126 -0.8947578
#> [5,]  1.22472894 35.389193 -22.3363853 -2.9118581
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  0.7374264 -0.1813725  1.575758  1.544441
#> [2,] -0.6374252  0.3238048  2.769318  3.046605
#> [3,] -0.8878588 -0.3309664  4.490337  4.315329
#> [4,]  0.9422230  0.4143302 -3.973561 -3.724032
#> [5,]  2.6724681 -0.6623288 -4.402171 -8.475732
#> 
#> , , 5, 4
#> 
#>            [,1]      [,2]      [,3]      [,4]
#> [1,]  1.1794270 -2.887000  2.898369  1.113637
#> [2,] -0.5208008  5.273754 -4.181299 -1.011852
#> [3,] -0.8860271  7.777074 -6.637332 -2.005190
#> [4,]  0.7752355 -6.378556  6.094008  2.362810
#> [5,]  1.7399707 -2.258454 11.085863  7.244433
#> 
#> , , 1, 5
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  1.00782981  0.2837637 -0.23275387 -0.1214591
#> [2,] -0.01197414  1.2023928 -0.05954727  0.3061538
#> [3,]  0.02180710  0.5520035  1.57004909  1.4748639
#> [4,] -0.02167972 -0.5177165 -0.58617475 -0.3936217
#> [5,]  0.06523901 -7.5392811  8.71270449  2.1302198
#> 
#> , , 2, 5
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  1.0552597 -0.1241020 -0.4135681 -0.678383
#> [2,] -0.3341660  1.2705778  2.4562039  3.698889
#> [3,] -0.4536578  0.5131134  3.7416285  4.421740
#> [4,]  0.4157850 -0.3729834 -2.7064740 -3.178263
#> [5,]  1.3438648 -5.7747521  1.7554943 -7.015927
#> 
#> , , 3, 5
#> 
#>             [,1]        [,2]         [,3]      [,4]
#> [1,]  0.96836691   0.4283956 -1.499940831 -1.490916
#> [2,] -0.02479043   2.6905764 -0.006576668  1.669433
#> [3,] -0.03926127   3.2576340  0.816216151  2.910437
#> [4,]  0.03182248  -2.8023420  0.151455231 -1.496893
#> [5,]  1.22493950 -32.6018338 34.480437669 10.644276
#> 
#> , , 4, 5
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  0.8455013 -0.2783935  1.580264  1.451938
#> [2,] -0.3890740 -0.1248993  2.632994  2.648529
#> [3,] -0.4323826 -1.0387275  4.122081  3.490735
#> [4,]  0.4715610  1.1662600 -3.616535 -2.869662
#> [5,]  1.2088469 -1.7331129 -4.016220 -8.888064
#> 
#> , , 5, 5
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  1.0701936  -1.811740  2.091802 1.0653122
#> [2,] -0.1823095   2.902256 -1.364908 0.2830889
#> [3,] -0.3068118   3.381098 -1.508629 0.3350483
#> [4,]  0.2492094  -2.329425  1.296876 0.1140681
#> [5,]  1.6013616 -13.979262 16.195112 4.2906296
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_last_draw`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|