R6 Class Representing PosteriorBVARPANEL

The class PosteriorBVARPANEL contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvarPANEL$new()

specify_posterior_bvarPANEL$get_posterior()

specify_posterior_bvarPANEL$get_last_draw()

specify_posterior_bvarPANEL$clone()

Method new()

Create a new posterior output PosteriorBVARPANEL.

Usage

specify_posterior_bvarPANEL$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARPANEL.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvarPANEL$get_posterior()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvarPANEL$get_last_draw()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvarPANEL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARPANEL" "R6"                

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_posterior`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $A
#> , , 1
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,] -0.1267858 -0.0203413 -0.56310472 -0.19275242
#> [2,]  0.4444274 -0.4990100 -0.23339755 -0.06298278
#> [3,] -0.5678657  1.4321335  0.43574859 -0.80747958
#> [4,] -2.0254454  0.6656910  0.06517147  0.31685521
#> [5,] -0.4460804 -2.3920864 -1.34863546  0.00633962
#> 
#> , , 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.7798891  0.04298984  0.7370560  0.7170323
#> [2,]  0.7568276  0.46528477  0.8877532  1.0398086
#> [3,]  0.5922327 -0.13118065  1.5264219  1.0147003
#> [4,] -0.4535895  0.10945113 -0.6959442 -0.1589159
#> [5,] -1.8944266 -0.10782957 -5.0033955 -1.9704579
#> 
#> , , 3
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.8880932 -0.17108427  0.6973567  0.5817348
#> [2,]  0.3095799  0.16607588  0.9514146  0.8290969
#> [3,]  0.5205483 -1.04574408  2.8902749  1.8254748
#> [4,] -0.4422339  0.92579721 -2.1622820 -1.4177182
#> [5,] -1.4626956 -0.09215346 -6.1245085 -4.8093881
#> 
#> , , 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8901513  0.7376971 -0.6896511 -0.4404547
#> [2,]  0.3238070 -0.1332535  1.0061286  0.6678609
#> [3,]  0.4951926 -1.1873435  2.2838027  1.2405292
#> [4,] -0.3729532  0.9654670 -0.7189132  0.2965735
#> [5,] -1.1575898 -2.3465723 -2.1806891 -2.7438484
#> 
#> , , 5
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9324431  0.70799598  0.0582352  0.4767747
#> [2,]  0.3622526 -0.06678619  0.8539696  0.5940853
#> [3,]  0.4988670 -0.48803369  2.3933455  1.8099964
#> [4,] -0.5454719  0.67466719 -1.1556497 -0.3859239
#> [5,] -1.3275848 -2.75970677 -0.0191987 -0.9409994
#> 
#> 
#> $V
#> , , 1
#> 
#>            [,1]        [,2]        [,3]        [,4]       [,5]
#> [1,]  2.0209494  0.41462185 -0.60498032 -0.73040104 -2.3836854
#> [2,]  0.4146218  2.19176338 -0.69929388  0.01868822  0.4807533
#> [3,] -0.6049803 -0.69929388  3.25183000 -0.07792695 -0.7559317
#> [4,] -0.7304010  0.01868822 -0.07792695  3.91762587  3.3921290
#> [5,] -2.3836854  0.48075331 -0.75593169  3.39212900 25.6965546
#> 
#> , , 2
#> 
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,]  1.4160799  0.2317108  0.7382256 -0.5116889 -5.214070
#> [2,]  0.2317108  1.8110998  0.2339121 -0.1195670 -2.870080
#> [3,]  0.7382256  0.2339121  1.8085792 -1.3089721 -5.995561
#> [4,] -0.5116889 -0.1195670 -1.3089721  2.3608638  6.465370
#> [5,] -5.2140704 -2.8700804 -5.9955613  6.4653695 49.728122
#> 
#> , , 3
#> 
#>             [,1]        [,2]       [,3]       [,4]       [,5]
#> [1,]  2.36667285  0.05660001  0.6676421 -0.6590067 -8.3749538
#> [2,]  0.05660001  1.56928873  0.9867901 -0.7741610 -0.9950211
#> [3,]  0.66764214  0.98679012  3.0091745 -1.7615992 -7.1564622
#> [4,] -0.65900670 -0.77416102 -1.7615992  2.5433826  6.0928059
#> [5,] -8.37495384 -0.99502108 -7.1564622  6.0928059 65.7011182
#> 
#> , , 4
#> 
#>            [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,]  2.8409063 -0.4483474 -1.254012 -0.4566522  0.9041011
#> [2,] -0.4483474  3.4204445  2.151984 -1.9922553 -3.8971885
#> [3,] -1.2540124  2.1519837  5.180581 -1.8331172 -4.3606941
#> [4,] -0.4566522 -1.9922553 -1.833117  4.5703334  2.9933884
#> [5,]  0.9041011 -3.8971885 -4.360694  2.9933884 32.7473031
#> 
#> , , 5
#> 
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,]  2.6260608  0.5814588  0.2880913 -0.7397783 -1.0416015
#> [2,]  0.5814588  4.0350564  3.1475809 -3.0240629 -0.6263376
#> [3,]  0.2880913  3.1475809  7.0375386 -4.8342742  2.4311006
#> [4,] -0.7397783 -3.0240629 -4.8342742  6.2565682 -0.9939417
#> [5,] -1.0416015 -0.6263376  2.4311006 -0.9939417 34.2112667
#> 
#> 
#> $Sigma
#> , , 1
#> 
#>           [,1]      [,2]     [,3]      [,4]
#> [1,] 11.657862 -2.673477 2.143084  2.665873
#> [2,] -2.673477 17.872312 6.123839  2.386515
#> [3,]  2.143084  6.123839 7.811121  2.998997
#> [4,]  2.665873  2.386515 2.998997 10.433866
#> 
#> , , 2
#> 
#>           [,1]       [,2]      [,3]       [,4]
#> [1,]  5.248132 -2.2927758  2.911164  2.5193982
#> [2,] -2.292776  8.8517257 -3.995774  0.7427214
#> [3,]  2.911164 -3.9957741 15.907109 12.2221140
#> [4,]  2.519398  0.7427214 12.222114 15.3844912
#> 
#> , , 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.7770480 -0.7238815  0.8179051  0.8105485
#> [2,] -0.7238815  8.4878617 -5.8491634  0.2148748
#> [3,]  0.8179051 -5.8491634 12.7541034  8.9901001
#> [4,]  0.8105485  0.2148748  8.9901001 11.0493380
#> 
#> , , 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.2957611 -0.6805523 -0.1005909 -0.4409986
#> [2,] -0.6805523  9.1674385 -4.9079775  1.2153104
#> [3,] -0.1005909 -4.9079775 12.4105681  9.1915362
#> [4,] -0.4409986  1.2153104  9.1915362 10.6204629
#> 
#> , , 5
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.16195715 -0.2743613 -0.06181631 -0.2661508
#> [2,] -0.27436128  5.6874174 -3.58935850  0.2558214
#> [3,] -0.06181631 -3.5893585 13.08346701 11.0303371
#> [4,] -0.26615084  0.2558214 11.03033710 11.7743127
#> 
#> 
#> $nu
#>          [,1]
#> [1,] 5.100276
#> [2,] 5.100428
#> [3,] 5.100334
#> [4,] 5.100071
#> [5,] 5.100464
#> 
#> $m
#>            [,1]
#> [1,] 0.70575936
#> [2,] 1.52073420
#> [3,] 0.41363053
#> [4,] 1.18665441
#> [5,] 0.06644609
#> 
#> $w
#>          [,1]
#> [1,] 56.30687
#> [2,] 20.71113
#> [3,] 27.14017
#> [4,] 43.90062
#> [5,] 55.91678
#> 
#> $s
#>           [,1]
#> [1,] 0.7345497
#> [2,] 1.1734411
#> [3,] 0.9384960
#> [4,] 0.8151476
#> [5,] 1.5807455
#> 
#> $scale
#>            [,1]
#> [1,] 0.01427928
#> [2,] 0.01427928
#> [3,] 0.01671019
#> [4,] 0.01897676
#> [5,] 0.02078904
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.11607212 -0.08839332  0.1368651  0.1018475
#> [2,] -0.08839332  0.42462252 -0.4279172 -0.2215028
#> [3,]  0.13686514 -0.42791720  0.8471445  0.5537127
#> [4,]  0.10184748 -0.22150277  0.5537127  0.4591349
#> 
#> , , 2, 1
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.100773043 -0.06612375  0.02095303 -0.001493026
#> [2,] -0.066123748  0.19826556 -0.06560172 -0.010075575
#> [3,]  0.020953034 -0.06560172  0.33651132  0.316669642
#> [4,] -0.001493026 -0.01007557  0.31666964  0.401695776
#> 
#> , , 3, 1
#> 
#>            [,1]        [,2]        [,3]      [,4]
#> [1,] 0.09117570  0.01686435  0.04581472 0.1026235
#> [2,] 0.01686435  9.25732946 -5.49645558 1.1287732
#> [3,] 0.04581472 -5.49645558  4.91391838 1.1024823
#> [4,] 0.10262350  1.12877316  1.10248232 2.2214279
#> 
#> , , 4, 1
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.11552056 -0.06436259  0.07833928  0.07816805
#> [2,] -0.06436259  0.32100860 -0.19378150 -0.04764815
#> [3,]  0.07833928 -0.19378150  0.33355569  0.19402897
#> [4,]  0.07816805 -0.04764815  0.19402897  0.20068102
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.1535533 -0.1354510  0.1391143  0.1512771
#> [2,] -0.1354510  3.4912324 -2.5881261 -0.6122338
#> [3,]  0.1391143 -2.5881261  2.8024460  1.4233111
#> [4,]  0.1512771 -0.6122338  1.4233111  1.3039023
#> 
#> , , 1, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.03427097 -0.05851598  0.07664726  0.04620631
#> [2,] -0.05851598  0.34202450 -0.41042375 -0.23753089
#> [3,]  0.07664726 -0.41042375  0.80377839  0.60692142
#> [4,]  0.04620631 -0.23753089  0.60692142  0.52090292
#> 
#> , , 2, 2
#> 
#>              [,1]         [,2]         [,3]        [,4]
#> [1,]  0.020133777 -0.006884366 -0.003689041 -0.00563566
#> [2,] -0.006884366  0.123325276 -0.125847124 -0.03722594
#> [3,] -0.003689041 -0.125847124  0.594646287  0.56317602
#> [4,] -0.005635660 -0.037225940  0.563176018  0.63278869
#> 
#> , , 3, 2
#> 
#>             [,1]        [,2]       [,3]      [,4]
#> [1,]  0.03874961  -0.6105018   0.520332  0.139150
#> [2,] -0.61050177  17.5158200 -14.076195 -2.741931
#> [3,]  0.52033198 -14.0761951  13.643298  4.906160
#> [4,]  0.13915003  -2.7419315   4.906160  3.626177
#> 
#> , , 4, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.01759873 -0.02416633  0.03754046  0.02458017
#> [2,] -0.02416633  0.19991559 -0.18582222 -0.03127459
#> [3,]  0.03754046 -0.18582222  0.35365766  0.22042997
#> [4,]  0.02458017 -0.03127459  0.22042997  0.23262895
#> 
#> , , 5, 2
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.02428617 -0.1330629  0.1163483  0.04071535
#> [2,] -0.13306290  2.9867191 -2.4193487 -0.67551097
#> [3,]  0.11634830 -2.4193487  2.7272071  1.38615368
#> [4,]  0.04071535 -0.6755110  1.3861537  1.08903093
#> 
#> , , 1, 3
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.01022112 -0.02549795  0.03496594  0.01652556
#> [2,] -0.02549795  0.24759186 -0.24282519 -0.10383377
#> [3,]  0.03496594 -0.24282519  0.48395515  0.36195117
#> [4,]  0.01652556 -0.10383377  0.36195117  0.33804957
#> 
#> , , 2, 3
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.008965535 -0.01281688  0.02138124  0.02109764
#> [2,] -0.012816877  0.11918983 -0.12780017 -0.04395734
#> [3,]  0.021381236 -0.12780017  0.79426241  0.79873739
#> [4,]  0.021097641 -0.04395734  0.79873739  0.89264045
#> 
#> , , 3, 3
#> 
#>              [,1]       [,2]        [,3]        [,4]
#> [1,]  0.006416794 -0.1027860  0.05078436 -0.02044553
#> [2,] -0.102786035  7.8448798 -5.32201053 -0.08092410
#> [3,]  0.050784363 -5.3220105  6.15146634  2.97362010
#> [4,] -0.020445535 -0.0809241  2.97362010  3.39160253
#> 
#> , , 4, 3
#> 
#>             [,1]         [,2]         [,3]         [,4]
#> [1,] 0.004900560  0.004565379  0.001385642  0.006520528
#> [2,] 0.004565379  0.225146328 -0.210690331 -0.039418423
#> [3,] 0.001385642 -0.210690331  0.401483669  0.252864997
#> [4,] 0.006520528 -0.039418423  0.252864997  0.243133932
#> 
#> , , 5, 3
#> 
#>              [,1]        [,2]      [,3]        [,4]
#> [1,]  0.004514068 -0.07033749  0.073159  0.03376975
#> [2,] -0.070337487  3.81741955 -3.056236 -0.80880225
#> [3,]  0.073158997 -3.05623570  3.235595  1.51954539
#> [4,]  0.033769752 -0.80880225  1.519545  1.14189894
#> 
#> , , 1, 4
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.005627932 -0.01116578  0.01871809  0.01085979
#> [2,] -0.011165782  0.25288296 -0.29098466 -0.17365570
#> [3,]  0.018718087 -0.29098466  0.63354266  0.50776011
#> [4,]  0.010859793 -0.17365570  0.50776011  0.44908573
#> 
#> , , 2, 4
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.010411917  0.002670875 -0.01986201 -0.01314891
#> [2,]  0.002670875  0.178378597 -0.09682285  0.05203426
#> [3,] -0.019862015 -0.096822851  0.37123609  0.31392562
#> [4,] -0.013148909  0.052034262  0.31392562  0.39320561
#> 
#> , , 3, 4
#> 
#>              [,1]        [,2]        [,3]       [,4]
#> [1,]  0.004787206 -0.07337192  0.07318964  0.0275484
#> [2,] -0.073371915  6.56686115 -4.64769690 -0.3180627
#> [3,]  0.073189639 -4.64769690  5.90396723  3.1971623
#> [4,]  0.027548397 -0.31806272  3.19716234  3.4196017
#> 
#> , , 4, 4
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.002273029 -0.01156692  0.01458256  0.006552259
#> [2,] -0.011566916  0.27353086 -0.25103993 -0.045354694
#> [3,]  0.014582558 -0.25103993  0.36314624  0.180526104
#> [4,]  0.006552259 -0.04535469  0.18052610  0.156450545
#> 
#> , , 5, 4
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.003032764 -0.05322837  0.05900252  0.03016956
#> [2,] -0.053228375  2.62916538 -2.27567873 -0.76257110
#> [3,]  0.059002518 -2.27567873  2.98837073  1.79351127
#> [4,]  0.030169561 -0.76257110  1.79351127  1.48541934
#> 
#> , , 1, 5
#> 
#>              [,1]       [,2]         [,3]         [,4]
#> [1,]  0.006055237 -0.0110747  0.009758237  0.003862514
#> [2,] -0.011074705  0.1939180 -0.240064880 -0.161198710
#> [3,]  0.009758237 -0.2400649  0.514843471  0.437737982
#> [4,]  0.003862514 -0.1611987  0.437737982  0.402414693
#> 
#> , , 2, 5
#> 
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.011256140 -0.02321911  0.002138618 -0.01422681
#> [2,] -0.023219110  0.19406505 -0.102125998  0.04213523
#> [3,]  0.002138618 -0.10212600  0.425393511  0.38514921
#> [4,] -0.014226813  0.04213523  0.385149207  0.46167090
#> 
#> , , 3, 5
#> 
#>              [,1]       [,2]      [,3]        [,4]
#> [1,]  0.004664559 -0.1277997  0.166273  0.09859572
#> [2,] -0.127799694  9.1246570 -8.335926 -2.84232885
#> [3,]  0.166272998 -8.3359264 11.492872  7.13009951
#> [4,]  0.098595724 -2.8423288  7.130100  6.22469067
#> 
#> , , 4, 5
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.003588561 -0.01080885  0.01353634  0.004710496
#> [2,] -0.010808852  0.24807593 -0.21070991 -0.016547720
#> [3,]  0.013536337 -0.21070991  0.50856365  0.361587923
#> [4,]  0.004710496 -0.01654772  0.36158792  0.369326448
#> 
#> , , 5, 5
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.003732399 -0.08631694  0.08551258  0.03696481
#> [2,] -0.086316939  3.57535815 -2.96860759 -0.90572875
#> [3,]  0.085512577 -2.96860759  3.31267376  1.69411278
#> [4,]  0.036964811 -0.90572875  1.69411278  1.27828009
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.7382634  0.6271583 -0.49905036 -0.28484241
#> [2,]  0.3219877  1.0316608 -0.09589014  0.04918701
#> [3,]  0.3964921  0.8248282  0.14781837  0.21550344
#> [4,] -0.2409506 -0.9279646  1.01996524  0.89461909
#> [5,] -2.3646529 -6.4677381  0.89544450  1.64429337
#> 
#> , , 2, 1
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.7405759 -0.01722476  0.2057522 0.08641241
#> [2,]  0.3789514  0.28142237  0.5405894 0.65926072
#> [3,]  0.6913083 -0.97947690  1.5197789 0.72209626
#> [4,] -0.5556654  0.95521994 -0.6334956 0.15911652
#> [5,] -1.4408715  1.58335107  1.3473794 5.24802866
#> 
#> , , 3, 1
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.9299794 -3.519531   3.080371  0.4614814
#> [2,]  0.4019713  4.590256  -2.884103 -0.1868705
#> [3,]  0.6025075  6.691218  -4.151310 -0.1531150
#> [4,] -0.5300212 -5.359613   4.090664  1.0329415
#> [5,] -3.3509796  7.546671 -10.449329 -2.0428826
#> 
#> , , 4, 1
#> 
#>            [,1]        [,2]         [,3]       [,4]
#> [1,]  0.6707956  0.23891533  0.674436978  0.6167524
#> [2,]  0.4278432  0.86832001  0.001691694  0.6524108
#> [3,]  0.6575330  0.09343625  1.346784110  1.3423600
#> [4,] -0.5261166 -0.16619566 -0.510692697 -0.5073595
#> [5,] -1.2166056 -0.23606653 -4.403161711 -2.5202083
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.7973877 -0.3734513   0.8840719  0.4433276
#> [2,]  0.3651439 -1.0445114   1.8386165  1.0051520
#> [3,]  0.6483947 -3.0004585   3.6839256  1.5497887
#> [4,] -0.5530881  3.1293018  -2.9428880 -0.7213830
#> [5,] -0.0239143  4.4655498 -10.3652587 -2.4876054
#> 
#> , , 1, 2
#> 
#>              [,1]        [,2]       [,3]       [,4]
#> [1,]  0.867170406   0.6263876 -0.8721182 -0.6770037
#> [2,]  0.083025138   1.2813893 -0.1466827  0.2626806
#> [3,]  0.053167491   1.2956997 -0.1927985  0.2219135
#> [4,] -0.001524365  -1.3198882  1.3436278  0.9549136
#> [5,]  0.246563013 -10.4174352  9.3646623  5.7902513
#> 
#> , , 2, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8477458  0.1595446  0.1480351 0.08426510
#> [2,]  0.4267055  0.3106458  0.5631791 0.73814676
#> [3,]  0.5987362 -1.0087205  1.7691550 0.91208410
#> [4,] -0.5321458  0.9354549 -0.7827021 0.04979475
#> [5,] -0.7219066  0.2729575 -2.2442976 0.10512896
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  0.9653328  -2.933911  3.305654  1.262979
#> [2,]  0.1136688   5.562985 -5.714839 -2.866241
#> [3,]  0.1657337   7.819306 -8.366467 -4.400285
#> [4,] -0.1392252  -6.477436  7.840424  4.717705
#> [5,] -0.9748387 -11.480772 17.332411 15.437619
#> 
#> , , 4, 2
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.9686806 -0.15754490   1.6771775  1.4840429
#> [2,]  0.6071686  0.43152924   0.7093033  0.9940723
#> [3,]  0.7393704 -0.03956686   1.8396796  1.7531805
#> [4,] -0.7016271  0.10612286  -1.2653418 -1.1301693
#> [5,] -2.1483071  0.33688055 -11.4539987 -9.0916979
#> 
#> , , 5, 2
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.9519330 -0.8743875   1.370779  0.684248
#> [2,]  0.4838610 -0.9448668   1.825591  1.194460
#> [3,]  0.7948443 -2.9026804   3.727925  1.851419
#> [4,] -0.7229489  3.1193772  -3.122064 -1.090472
#> [5,] -3.0407483 12.0428677 -14.348452 -4.398480
#> 
#> , , 1, 3
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.01440061  0.2021732  0.3172754  0.3708335
#> [2,]  0.07768604  1.3822666 -0.5912467 -0.1714509
#> [3,]  0.15888435  1.0780097  0.1709473  0.4478126
#> [4,] -0.12196205 -1.0152141  0.6669018  0.4461072
#> [5,] -2.01975426 -6.3404624  1.7935209 -0.5069628
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.9032031  0.2817879 -0.1455873 -0.08593238
#> [2,]  0.4337959  0.3178535  0.2707989  0.48413194
#> [3,]  0.5829475 -0.7668308  1.1708040  0.53593742
#> [4,] -0.5202960  0.7234652 -0.1933076  0.44482787
#> [5,] -2.1672772 -2.2218757  2.7623395  1.96422778
#> 
#> , , 3, 3
#> 
#>             [,1]      [,2]      [,3]       [,4]
#> [1,]  0.96717127 -2.706909  1.836520 -0.1177462
#> [2,]  0.04888985  5.194749 -2.710860  0.3276038
#> [3,]  0.06134111  7.529035 -3.696432  0.7615042
#> [4,] -0.04722769 -6.331840  3.815536  0.1816955
#> [5,] -0.23409373 -4.543866  9.042001  8.2282747
#> 
#> , , 4, 3
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9827351 -0.06610461  1.6235553  1.4498853
#> [2,]  0.5266911  0.73873118  0.4105428  0.9359343
#> [3,]  0.6987231  0.16744556  1.5498294  1.6311837
#> [4,] -0.6761797 -0.07995481 -1.0647855 -1.0457249
#> [5,] -1.2181039 -3.83575468 -3.1450703 -5.4101342
#> 
#> , , 5, 3
#> 
#>            [,1]      [,2]      [,3]       [,4]
#> [1,]  0.9477506 -1.058709  1.568840  0.8602722
#> [2,]  0.4642694 -1.781827  2.770407  1.5796037
#> [3,]  0.7599812 -3.899203  4.878789  2.2988827
#> [4,] -0.7055054  4.474219 -4.558209 -1.6211609
#> [5,] -1.8895394 -2.076191 -5.419458 -5.5980793
#> 
#> , , 1, 4
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.97994098  0.4272328 -0.1832153 -0.05894124
#> [2,]  0.09711767  0.6712021  0.3430469  0.46858485
#> [3,]  0.20981399 -0.5786580  2.0936363  1.55966730
#> [4,] -0.17716905  0.4364928 -0.9321475 -0.39645654
#> [5,] -1.02511396 -2.5689007 -2.0994559 -4.23380064
#> 
#> , , 2, 4
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.9213421  0.02033289  0.1110478 -0.07126994
#> [2,]  0.4260244  0.17058966  0.5660148  0.63953338
#> [3,]  0.6214278 -0.85611715  1.6292799  1.00761384
#> [4,] -0.5586965  0.91428502 -0.6774093  0.06057647
#> [5,] -2.0191891 -2.27347800 -0.4002845 -1.40098752
#> 
#> , , 3, 4
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.9868178  -0.2733491 0.04037508 -0.37786863
#> [2,]  0.1177852   1.3932196 0.05026337  0.63864865
#> [3,]  0.1719084   1.4337983 0.82543131  1.35623992
#> [4,] -0.1469655  -0.9422394 0.10145088 -0.01197249
#> [5,] -1.3761901 -15.6620617 1.16117651 -8.88553303
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9932493  0.3853448  0.7309074  0.8318945
#> [2,]  0.5545049  0.3842108  0.7297578  1.0409911
#> [3,]  0.6977552  0.0400619  1.6288618  1.6162039
#> [4,] -0.6743200 -0.1403300 -0.8543365 -0.8641309
#> [5,] -1.6070853 -0.9955355 -2.2431374 -2.1738657
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]      [,3]        [,4]
#> [1,]  0.9613999 -0.9290763  1.057740  0.38397989
#> [2,]  0.2764192 -0.4102585  1.044057  0.58160429
#> [3,]  0.4482933 -1.7778661  2.214946  0.77289068
#> [4,] -0.4211826  2.3254241 -1.835172 -0.03704042
#> [5,] -0.6323126 -4.6402878  5.978897  4.60755250
#> 
#> , , 1, 5
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.9314333  0.5182374 -0.3177747 -0.09815695
#> [2,]  0.1869617  0.5688391  0.4035783  0.44444244
#> [3,]  0.3952029 -0.4884750  1.7971240  1.26488614
#> [4,] -0.3257173  0.2991592 -0.6067157 -0.11717615
#> [5,] -1.6176465 -1.3466133 -2.1130013 -3.71758091
#> 
#> , , 2, 5
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.8924452  0.02655011 -0.02127951 -0.1763657
#> [2,]  0.4073227  0.15737614  0.54717349  0.6145006
#> [3,]  0.5920751 -1.02814042  1.12024828  0.2526396
#> [4,] -0.5250699  1.04147352 -0.25031864  0.6669416
#> [5,] -1.6501373 -0.94762184  3.22168620  3.6903371
#> 
#> , , 3, 5
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.9772237 -1.618731  0.6792234 -0.6543756
#> [2,]  0.0893794  2.047698 -0.7883241  0.1540534
#> [3,]  0.1324565  2.557506 -0.6158253  0.4983904
#> [4,] -0.1054087 -2.078869  1.2939761  0.5544872
#> [5,] -1.1875995 15.857870  1.2376757 14.3904594
#> 
#> , , 4, 5
#> 
#>            [,1]       [,2]        [,3]      [,4]
#> [1,]  1.0183840  0.0829046  1.25788222  1.268090
#> [2,]  0.5401488  0.8788040  0.09680986  0.711558
#> [3,]  0.6957646  0.2183129  1.58147379  1.718662
#> [4,] -0.6864731 -0.1830619 -0.92509840 -1.016437
#> [5,] -1.1716654 -4.2984571 -5.75045069 -8.989778
#> 
#> , , 5, 5
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.9119804 -0.3511154 -0.6420472 -1.070385
#> [2,]  0.4052171 -0.8480102  4.3757185  3.901080
#> [3,]  0.6640961 -2.5837178  7.8930805  6.387420
#> [4,] -0.6056831  2.8444307 -6.7882816 -5.058145
#> [5,] -1.1559458 -3.7049778  9.2527430  8.775221
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_last_draw`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|