R6 Class Representing PosteriorBVARPANEL

The class PosteriorBVARPANEL contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvarPANEL$new()

specify_posterior_bvarPANEL$get_posterior()

specify_posterior_bvarPANEL$get_last_draw()

specify_posterior_bvarPANEL$clone()

Method new()

Create a new posterior output PosteriorBVARPANEL.

Usage

specify_posterior_bvarPANEL$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARPANEL.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvarPANEL$get_posterior()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvarPANEL$get_last_draw()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvarPANEL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARPANEL" "R6"                

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_posterior`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $A
#> , , 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.0547213  1.0011152  0.3886203 -0.1689910
#> [2,]  0.1537751  0.7964868  0.4413507 -1.0377214
#> [3,]  0.7342987 -0.1178609  0.1491472  0.9195905
#> [4,]  1.2145557 -1.1369699 -0.3437289  0.3108160
#> [5,] -0.9758545 -3.0698201  3.7115541 -1.0192800
#> 
#> , , 2
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.04842591  0.1043002  0.3636906  0.1096719
#> [2,]  0.02722788  0.7369590  0.2863162  0.4674912
#> [3,]  0.13712078 -0.4573414  1.3470631  0.7055137
#> [4,] -0.03289211  0.7956617 -0.4546729  0.2503214
#> [5,] -0.30129004  3.3446949  0.8358000 -2.8982320
#> 
#> , , 3
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.9912965  0.1234523  1.01532133  1.2350210
#> [2,]  0.1865770  0.4008763  0.36635538  0.8146273
#> [3,]  0.3509388  0.2533579  0.08214285  0.4304927
#> [4,] -0.2162185  0.2777162 -0.55295544 -0.4052303
#> [5,] -1.2394291 -0.1295821 -4.51353747 -9.8495717
#> 
#> , , 4
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  1.0317291 -0.3455629  0.88181287  0.67669021
#> [2,]  0.0739439  0.1428560 -0.03758556  0.07942182
#> [3,]  0.4809623  0.2528161  0.78093714  1.09360035
#> [4,] -0.2939819 -0.2495128  0.22590878  0.06424030
#> [5,] -1.4644017  1.9693086 -3.32770068 -7.66088910
#> 
#> , , 5
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9571749 -0.17584718  0.3597657  0.2826820
#> [2,]  0.1922557  0.02301838 -0.2279268 -0.2851994
#> [3,]  0.3313363  0.05244055  1.1898552  1.3482630
#> [4,] -0.3325007  0.43001905  0.1024194  0.4372704
#> [5,] -1.5015065  1.65839330  0.2809852 -4.7187698
#> 
#> 
#> $V
#> , , 1
#> 
#>           [,1]      [,2]       [,3]       [,4]       [,5]
#> [1,]  6.038322 -1.549037  3.9242254 -0.7537020 -2.4774025
#> [2,] -1.549037  7.762418 -3.6547881 -1.3101571  2.1398858
#> [3,]  3.924225 -3.654788 11.6142067  0.1425780  0.1000343
#> [4,] -0.753702 -1.310157  0.1425780  6.3341872 -0.6650445
#> [5,] -2.477402  2.139886  0.1000343 -0.6650445 65.0520238
#> 
#> , , 2
#> 
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,]  2.1206110 -0.1778616 -0.2963519  0.6596142  -5.256294
#> [2,] -0.1778616  1.8866786  0.8621101 -0.6507849  -0.941649
#> [3,] -0.2963519  0.8621101  2.5445044 -0.9037008   1.829624
#> [4,]  0.6596142 -0.6507849 -0.9037008  2.3356029   3.893467
#> [5,] -5.2562938 -0.9416490  1.8296237  3.8934666 110.790927
#> 
#> , , 3
#> 
#>             [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,]   2.7525623 -0.2710912 -1.1372370  1.180295 -13.345453
#> [2,]  -0.2710912  2.0992037  0.8736468 -1.012141   1.236340
#> [3,]  -1.1372370  0.8736468  2.8401346 -1.185612   7.399803
#> [4,]   1.1802953 -1.0121413 -1.1856118  2.152752  -5.752268
#> [5,] -13.3454530  1.2363399  7.3998034 -5.752268 113.493766
#> 
#> , , 4
#> 
#>            [,1]      [,2]       [,3]      [,4]        [,5]
#> [1,]  2.5783176 -1.463190 -0.4920804  1.080771  -8.2689368
#> [2,] -1.4631899  2.952948  1.0537463 -1.715427   6.5661873
#> [3,] -0.4920804  1.053746  2.4709014 -2.086522   0.9461897
#> [4,]  1.0807707 -1.715427 -2.0865225  4.239014  -2.6000064
#> [5,] -8.2689368  6.566187  0.9461897 -2.600006 103.1226976
#> 
#> , , 5
#> 
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,]  1.5804420 -1.0447937  0.7663947  0.5824353 -4.659648
#> [2,] -1.0447937  4.8112076 -0.1096759 -3.2358793  1.926215
#> [3,]  0.7663947 -0.1096759  4.7665746 -1.0175081 -8.627714
#> [4,]  0.5824353 -3.2358793 -1.0175081  4.9904949  3.627432
#> [5,] -4.6596484  1.9262153 -8.6277143  3.6274322 65.582791
#> 
#> 
#> $Sigma
#> , , 1
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,] 13.6543650 2.9687889  0.6769788  2.9172183
#> [2,]  2.9687889 8.8061536  0.7794571  0.1691061
#> [3,]  0.6769788 0.7794571 12.6237052 -5.4651692
#> [4,]  2.9172183 0.1691061 -5.4651692 13.4709913
#> 
#> , , 2
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  3.9439626  0.66905910  0.3829912 -0.30615472
#> [2,]  0.6690591 10.09900397 -5.6028557 -0.01287639
#> [3,]  0.3829912 -5.60285567 10.7288090  3.56713623
#> [4,] -0.3061547 -0.01287639  3.5671362  7.29411900
#> 
#> , , 3
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,] 0.96908282  0.06822501  0.1274209  0.1157919
#> [2,] 0.06822501 10.17591306 -6.7298238 -1.8421220
#> [3,] 0.12742086 -6.72982381 14.3881437 12.2344584
#> [4,] 0.11579186 -1.84212197 12.2344584 14.3080258
#> 
#> , , 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.5825404  0.1621063 -0.1950292 -0.1430187
#> [2,]  0.1621063  7.3362108 -1.5012659  3.7789656
#> [3,] -0.1950292 -1.5012659 10.7881774 10.4793250
#> [4,] -0.1430187  3.7789656 10.4793250 14.4646318
#> 
#> , , 5
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.220124422  0.07902183 -0.05300057 -0.005438102
#> [2,]  0.079021828  6.36670365 -2.82365375  0.788069310
#> [3,] -0.053000569 -2.82365375  6.70401690  5.678569618
#> [4,] -0.005438102  0.78806931  5.67856962  7.023868186
#> 
#> 
#> $nu
#>          [,1]
#> [1,] 5.099570
#> [2,] 5.099574
#> [3,] 5.099834
#> [4,] 5.099584
#> [5,] 5.099502
#> 
#> $m
#>           [,1]
#> [1,] 0.9238091
#> [2,] 3.3186766
#> [3,] 0.4884743
#> [4,] 0.8118063
#> [5,] 1.0846734
#> 
#> $w
#>           [,1]
#> [1,] 166.27814
#> [2,]  21.18025
#> [3,]  20.71281
#> [4,]  47.67367
#> [5,]  35.46234
#> 
#> $s
#>           [,1]
#> [1,] 0.6234268
#> [2,] 0.8451490
#> [3,] 0.9329487
#> [4,] 1.5443822
#> [5,] 0.5584622
#> 
#> $scale
#>            [,1]
#> [1,] 0.01427928
#> [2,] 0.01427928
#> [3,] 0.01580739
#> [4,] 0.01795151
#> [5,] 0.02002019
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,] 0.07545955  0.02799409  0.02768113  0.0295711
#> [2,] 0.02799409  0.17495367 -0.15065655 -0.1001737
#> [3,] 0.02768113 -0.15065655  0.46584753  0.4275734
#> [4,] 0.02957110 -0.10017367  0.42757343  0.5204795
#> 
#> , , 2, 1
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.07586903 -0.01517507  0.01045488 0.04853319
#> [2,] -0.01517507  0.16405675 -0.08568822 0.01094291
#> [3,]  0.01045488 -0.08568822  0.29295672 0.20897507
#> [4,]  0.04853319  0.01094291  0.20897507 0.32167674
#> 
#> , , 3, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.04665292 -0.1689787  0.1853582  0.09563441
#> [2,] -0.16897873 10.3169246 -7.0793941 -0.18958548
#> [3,]  0.18535822 -7.0793941  6.3588746  1.99619826
#> [4,]  0.09563441 -0.1895855  1.9961983  2.60334847
#> 
#> , , 4, 1
#> 
#>              [,1]         [,2]         [,3]        [,4]
#> [1,]  0.067433988  0.005577417 -0.008318276 -0.01866694
#> [2,]  0.005577417  0.219122432 -0.183817308 -0.06159234
#> [3,] -0.008318276 -0.183817308  0.419886770  0.23611878
#> [4,] -0.018666938 -0.061592341  0.236118785  0.30180359
#> 
#> , , 5, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.05313276 -0.0205076  0.0501001  0.02895634
#> [2,] -0.02050760  3.0442211 -2.5918321 -0.83679991
#> [3,]  0.05010010 -2.5918321  3.0252043  1.50476283
#> [4,]  0.02895634 -0.8367999  1.5047628  1.12538796
#> 
#> , , 1, 2
#> 
#>             [,1]         [,2]         [,3]        [,4]
#> [1,] 0.020762985  0.007107044  0.008592106  0.02081483
#> [2,] 0.007107044  0.138012885 -0.105824779 -0.03626081
#> [3,] 0.008592106 -0.105824779  0.283422469  0.27728272
#> [4,] 0.020814828 -0.036260815  0.277282720  0.34558777
#> 
#> , , 2, 2
#> 
#>             [,1]         [,2]         [,3]       [,4]
#> [1,] 0.023122818  0.007931739  0.009787947 0.02380239
#> [2,] 0.007931739  0.179807066 -0.078101770 0.05988272
#> [3,] 0.009787947 -0.078101770  0.272421986 0.24724140
#> [4,] 0.023802394  0.059882719  0.247241398 0.36528722
#> 
#> , , 3, 2
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.02470659 -0.1818714  0.08268883 -0.02930086
#> [2,] -0.18187142  8.2681762 -4.61982094  0.81183951
#> [3,]  0.08268883 -4.6198209  4.78037523  2.11289923
#> [4,] -0.02930086  0.8118395  2.11289923  3.16373387
#> 
#> , , 4, 2
#> 
#>             [,1]         [,2]        [,3]        [,4]
#> [1,] 0.043362214  0.001296733  0.07092827  0.05851071
#> [2,] 0.001296733  0.345072987 -0.32505572 -0.07659054
#> [3,] 0.070928274 -0.325055719  0.67192581  0.39036796
#> [4,] 0.058510709 -0.076590538  0.39036796  0.32003222
#> 
#> , , 5, 2
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.01548165 -0.1051558  0.1325592  0.07284846
#> [2,] -0.10515583  3.1230427 -2.6752624 -0.93653867
#> [3,]  0.13255920 -2.6752624  3.4669743  2.15323318
#> [4,]  0.07284846 -0.9365387  2.1532332  1.87369585
#> 
#> , , 1, 3
#> 
#>              [,1]        [,2]        [,3]       [,4]
#> [1,]  0.008031649 -0.01225955  0.03275528  0.0333565
#> [2,] -0.012259546  0.24337438 -0.29418465 -0.2220911
#> [3,]  0.032755282 -0.29418465  0.84419670  0.8391153
#> [4,]  0.033356498 -0.22209106  0.83911527  0.8791052
#> 
#> , , 2, 3
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.015200529 -0.006629523  0.004535181  0.002427677
#> [2,] -0.006629523  0.084306133 -0.090825578 -0.031553587
#> [3,]  0.004535181 -0.090825578  0.359162449  0.341020262
#> [4,]  0.002427677 -0.031553587  0.341020262  0.379638960
#> 
#> , , 3, 3
#> 
#>              [,1]       [,2]        [,3]        [,4]
#> [1,]  0.008854896 -0.1060541  0.05094342 -0.01684059
#> [2,] -0.106054094  8.8817643 -5.26756893  0.50547126
#> [3,]  0.050943419 -5.2675689  5.06052632  1.93063420
#> [4,] -0.016840587  0.5054713  1.93063420  2.64314985
#> 
#> , , 4, 3
#> 
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.018777498  0.02521261 -0.008260273  0.01511445
#> [2,]  0.025212614  0.33100809 -0.259435014 -0.01636891
#> [3,] -0.008260273 -0.25943501  0.386611455  0.20179253
#> [4,]  0.015114448 -0.01636891  0.201792526  0.20220901
#> 
#> , , 5, 3
#> 
#>             [,1]         [,2]        [,3]       [,4]
#> [1,] 0.008524098  0.006467543  0.03711221 0.04435227
#> [2,] 0.006467543  3.612151114 -1.95732808 0.28735231
#> [3,] 0.037112209 -1.957328082  1.89944735 0.77896794
#> [4,] 0.044352269  0.287352308  0.77896794 1.07059773
#> 
#> , , 1, 4
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.01083425 -0.00488906  0.01229831  0.01005002
#> [2,] -0.00488906  0.25594667 -0.22292256 -0.11890090
#> [3,]  0.01229831 -0.22292256  0.49945049  0.45101720
#> [4,]  0.01005002 -0.11890090  0.45101720  0.45375297
#> 
#> , , 2, 4
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.011773488 -0.007947164  0.001055345 -0.003781642
#> [2,] -0.007947164  0.134482338 -0.091603200  0.005771509
#> [3,]  0.001055345 -0.091603200  0.433308395  0.420073593
#> [4,] -0.003781642  0.005771509  0.420073593  0.488855954
#> 
#> , , 3, 4
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.00629746 -0.0943369  0.07460599  0.02151871
#> [2,] -0.09433690 10.4422588 -7.00530813 -0.34202651
#> [3,]  0.07460599 -7.0053081  6.27386609  2.12822269
#> [4,]  0.02151871 -0.3420265  2.12822269  2.32699505
#> 
#> , , 4, 4
#> 
#>              [,1]        [,2]          [,3]       [,4]
#> [1,] 0.0125263341  0.01793319  0.0007646853 0.01884592
#> [2,] 0.0179331875  0.18861928 -0.0985396621 0.04388159
#> [3,] 0.0007646853 -0.09853966  0.2610152138 0.18297934
#> [4,] 0.0188459198  0.04388159  0.1829793381 0.21891001
#> 
#> , , 5, 4
#> 
#>              [,1]        [,2]        [,3]       [,4]
#> [1,]  0.005668256 -0.04841758  0.06587599  0.0406161
#> [2,] -0.048417577  2.45654520 -2.03449183 -0.5853669
#> [3,]  0.065875991 -2.03449183  2.86002007  1.8021224
#> [4,]  0.040616097 -0.58536689  1.80212243  1.6181863
#> 
#> , , 1, 5
#> 
#>              [,1]        [,2]         [,3]         [,4]
#> [1,]  0.003921253 -0.00506538  0.003558762  0.002000788
#> [2,] -0.005065380  0.18764560 -0.164293751 -0.090271076
#> [3,]  0.003558762 -0.16429375  0.265526434  0.216241408
#> [4,]  0.002000788 -0.09027108  0.216241408  0.200770396
#> 
#> , , 2, 5
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.009326993 -0.004323111 -0.004022804 -0.007407951
#> [2,] -0.004323111  0.125671511 -0.093387480 -0.002649565
#> [3,] -0.004022804 -0.093387480  0.316061199  0.282057616
#> [4,] -0.007407951 -0.002649565  0.282057616  0.321685822
#> 
#> , , 3, 5
#> 
#>              [,1]       [,2]        [,3]         [,4]
#> [1,]  0.004893979 -0.1072876  0.06506559 -0.002163271
#> [2,] -0.107287584  8.7334638 -6.08385171 -0.578928019
#> [3,]  0.065065586 -6.0838517  7.23288074  3.993234481
#> [4,] -0.002163271 -0.5789280  3.99323448  4.367056470
#> 
#> , , 4, 5
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,] 0.007436078  0.01128222  0.01513059 0.02469121
#> [2,] 0.011282217  0.21519073 -0.10271088 0.05426623
#> [3,] 0.015130586 -0.10271088  0.36276246 0.28946830
#> [4,] 0.024691208  0.05426623  0.28946830 0.33509670
#> 
#> , , 5, 5
#> 
#>              [,1]        [,2]       [,3]        [,4]
#> [1,]  0.003716163 -0.06273252  0.0646346  0.03039409
#> [2,] -0.062732516  4.88556440 -3.7867803 -0.92361877
#> [3,]  0.064634602 -3.78678026  3.7308016  1.61242578
#> [4,]  0.030394095 -0.92361877  1.6124258  1.18648990
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.7951399  0.19222991 -0.1766094 -0.08240888
#> [2,]  0.2768676  0.97041439 -0.1895989 -0.03780580
#> [3,]  0.4407640  0.03445327  1.1678339  0.76121282
#> [4,] -0.3466540 -0.04209212 -0.2170618  0.20470384
#> [5,] -0.2073462 -3.55292727  8.5868409  6.84601593
#> 
#> , , 2, 1
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.9331941 -0.1434224  0.02543143 -0.2339639
#> [2,]  0.3948793  0.6958605  0.76867903  1.2982910
#> [3,]  0.5835126 -0.4420039  1.52999545  1.3485626
#> [4,] -0.5350795  0.4680768 -0.71807524 -0.3852944
#> [5,] -1.1019606  0.9661656  7.81311876  4.2950023
#> 
#> , , 3, 1
#> 
#>            [,1]         [,2]       [,3]        [,4]
#> [1,]  0.8688097   1.80912392 -1.6413966 -0.67450394
#> [2,]  0.1942482   0.55398346 -0.2714796 -0.09442985
#> [3,]  0.3044295   0.09802151  0.3024498  0.17714478
#> [4,] -0.2725303   0.30508736  0.2728729  0.80591626
#> [5,]  0.9693944 -57.40399136 60.9301957 19.99815330
#> 
#> , , 4, 1
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.4867174 -0.08551156  0.7092865  1.0202022
#> [2,] -0.4774748 -1.06545111  0.3905677 -0.4238150
#> [3,] -0.5412942 -2.32791893  1.7775022  0.1431419
#> [4,]  0.4197400  2.35099281 -1.0626883  0.5234269
#> [5,] -3.3245378  1.22222322  3.8227491  0.2262905
#> 
#> , , 5, 1
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  1.0700649 -2.27085444  1.98034615  0.6794674
#> [2,]  0.1175074  1.05446854 -0.08424423  0.6812901
#> [3,]  0.1765304  0.69651582  0.26978079  0.8468867
#> [4,] -0.1988507  0.05713483 -0.17144790 -0.2111918
#> [5,] -0.6065821 19.29235227 -2.55558327  1.9069329
#> 
#> , , 1, 2
#> 
#>            [,1]        [,2]         [,3]       [,4]
#> [1,]  0.9503343  0.26620767 -0.453194959 -0.4862806
#> [2,]  0.1838326  0.96793171  0.007566778  0.5550218
#> [3,]  0.4861375  0.01306174  1.546612002  1.7526393
#> [4,] -0.4369944 -0.03801718 -0.523813156 -0.5612267
#> [5,] -0.7266905 -4.34318008 11.020860047  4.1547750
#> 
#> , , 2, 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9484641 -0.06302658  0.1232348 -0.0282580
#> [2,]  0.4411334  0.76738285  0.7884663  1.4591968
#> [3,]  0.5939642 -0.29289460  1.7238170  1.5776007
#> [4,] -0.5443241  0.33756959 -0.8014349 -0.5293505
#> [5,] -2.2979631 -1.87696910 -0.9131883 -7.1880328
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.8371568 -0.8568902  1.1307920   0.3091552
#> [2,]  0.2131517  0.5496123  0.3999572   0.7673247
#> [3,]  0.3099926  0.2946021  1.0283127   1.2895507
#> [4,] -0.2762753  0.3357361 -0.4984553  -0.2030999
#> [5,]  1.3217322 -8.4808101 -3.6122837 -12.1156008
#> 
#> , , 4, 2
#> 
#>            [,1]       [,2]          [,3]        [,4]
#> [1,]  1.0555121 -0.3154742   1.775677298   1.7746035
#> [2,] -0.4655328 -0.3894995  -0.410557373  -0.7658686
#> [3,] -0.5367638 -0.8706572   0.564210633  -0.1176430
#> [4,]  0.5243972  0.9111407   0.001845376   0.7179775
#> [5,] -0.3024689  6.7979476 -12.817333537 -14.0659562
#> 
#> , , 5, 2
#> 
#>            [,1]        [,2]         [,3]        [,4]
#> [1,]  0.9456546  -0.8132553  0.457108482 -0.01769135
#> [2,]  0.1149492   1.2779437 -0.303923299  0.50685650
#> [3,]  0.1871798   0.9126950  0.053312745  0.62066936
#> [4,] -0.1800534  -0.1630685  0.004304307 -0.10682425
#> [5,]  1.0187964 -20.3685190 41.818599388 28.66234105
#> 
#> , , 1, 3
#> 
#>            [,1]         [,2]       [,3]        [,4]
#> [1,]  0.9700638  0.315813995 -0.1204283  0.01663306
#> [2,]  0.2182546  0.861935322 -0.1285686  0.33751277
#> [3,]  0.5134751 -0.087314861  1.5427287  1.74780844
#> [4,] -0.4562355 -0.004664684 -0.6307154 -0.73589131
#> [5,] -1.7230189 -1.899552849  9.6490211  2.72581726
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8283088 -0.2068781  0.1809535 -0.1032628
#> [2,]  0.3868226  0.6756969  0.9957890  1.6234494
#> [3,]  0.5950023 -0.1895109  1.3700002  1.2897844
#> [4,] -0.5179007  0.2853759 -0.5763161 -0.3345825
#> [5,] -0.4768705  0.5858109 -0.3403380 -4.5349913
#> 
#> , , 3, 3
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.9653023 -2.6964713   1.6584093 -0.23415688
#> [2,]  0.2029133 -0.2829474   0.7937376  0.59239206
#> [3,]  0.2997289 -0.6497958   1.4855449  1.10472271
#> [4,] -0.2584212  0.8718527  -0.7196501 -0.04294604
#> [5,] -1.9905291 62.2458517 -31.7245638  2.76378810
#> 
#> , , 4, 3
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  1.1709040 -0.7084155   1.9798269   1.6483096
#> [2,] -0.2765888 -0.6394294  -0.3162963  -0.8104200
#> [3,] -0.1999703 -1.2870302   0.7193716  -0.1963759
#> [4,]  0.1716024  1.4415876  -0.2226779   0.8174320
#> [5,] -1.9676612  8.4111846 -13.0871944 -12.5061863
#> 
#> , , 5, 3
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  0.9998695 -1.72023468   1.8326949   0.8021207
#> [2,]  0.1580022  1.00746479   0.1990893   0.8936519
#> [3,]  0.2531663  0.18252954   1.0556760   1.2882226
#> [4,] -0.2267813  0.06677588  -0.4532371  -0.4417832
#> [5,] -1.6342927 31.61387391 -26.4415318 -13.3518051
#> 
#> , , 1, 4
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  0.9792637  0.18396164 -0.31303517 -0.25932254
#> [2,]  0.2098993  0.92388922  0.04821615  0.54109774
#> [3,]  0.5009072 -0.16012184  1.71493115  1.89272067
#> [4,] -0.4507102  0.09110645 -0.65596693 -0.70993040
#> [5,] -1.6470794 -0.97016023  6.61140530  0.04804158
#> 
#> , , 2, 4
#> 
#>            [,1]       [,2]       [,3]         [,4]
#> [1,]  0.9231371 -0.1307535  0.1714191 -0.007539436
#> [2,]  0.4178639  0.7000692  0.7562288  1.395357928
#> [3,]  0.5957997 -0.4307958  1.7325822  1.507544974
#> [4,] -0.5381048  0.4233577 -0.8496456 -0.534204697
#> [5,] -1.8488557  3.2336117  1.5210854 -1.788618259
#> 
#> , , 3, 4
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9666002 -1.89765652   1.175725 -0.2374391
#> [2,]  0.1957762 -0.06538634   1.019275  1.0658451
#> [3,]  0.2991987 -0.51180681   2.046531  1.9334247
#> [4,] -0.2511689  0.89008411  -1.245916 -0.6967567
#> [5,] -2.2930997 33.00370476 -21.390179 -7.2570457
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.9776320 -0.3324337   1.1437074  0.9600168
#> [2,] -0.2477700 -0.6677470  -0.6911623 -1.2518561
#> [3,] -0.1494001 -1.4570090   0.2320640 -0.8854841
#> [4,]  0.1653062  1.4710369   0.5347470  1.7055456
#> [5,] -0.2536473  9.1379762 -11.8541079 -9.8459698
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9836899 -1.1937096  1.2425435  0.5153852
#> [2,]  0.1553210  1.0947199  0.2028556  0.9698881
#> [3,]  0.2470379  0.3512306  0.9087245  1.2365294
#> [4,] -0.2237731  0.0309184 -0.3898837 -0.4020226
#> [5,] -1.0030718  9.7670567 -6.4770690 -5.8707274
#> 
#> , , 1, 5
#> 
#>            [,1]        [,2]         [,3]        [,4]
#> [1,]  0.9306164  0.36140953 -0.072106697  0.05982227
#> [2,]  0.2520790  0.86626981  0.006539374  0.47695382
#> [3,]  0.5252370 -0.14058404  1.744369158  1.91137307
#> [4,] -0.4440905  0.03531864 -0.730102167 -0.79494842
#> [5,] -2.1925730 -2.65375964  3.670586137 -3.44629107
#> 
#> , , 2, 5
#> 
#>            [,1]         [,2]       [,3]       [,4]
#> [1,]  0.9189515  0.005633948 -0.3187583 -0.4504288
#> [2,]  0.4381160  0.699788770  0.6048253  1.2280032
#> [3,]  0.5751753 -0.230640259  1.6248196  1.5366089
#> [4,] -0.5214356  0.262213956 -0.6318707 -0.4136942
#> [5,] -2.0371057 -0.567088042  6.2261850  0.7128747
#> 
#> , , 3, 5
#> 
#>            [,1]         [,2]       [,3]        [,4]
#> [1,]  0.9408366   0.49560684 -1.3630670 -1.39654650
#> [2,]  0.1947669  -0.04794405  0.6593284  0.69145290
#> [3,]  0.2914053  -0.72602670  1.4733146  1.13828872
#> [4,] -0.2518359   1.00127888 -0.7135960 -0.04093955
#> [5,] -1.2729483 -17.78780122 39.1670733 25.73186622
#> 
#> , , 4, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.0930934  0.1270212  0.5089519  0.8227126
#> [2,] -0.3268937 -0.6508104 -0.3933534 -0.9011497
#> [3,] -0.3607082 -1.5420398  0.7078100 -0.3993224
#> [4,]  0.3269622  1.4085836  0.1169669  1.1295016
#> [5,]  0.2930315  8.0229997  0.5479939  0.5751494
#> 
#> , , 5, 5
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.9886354 -0.4961727  1.0032187   0.7614723
#> [2,]  0.1635907  0.8184925  0.7275498   1.3703858
#> [3,]  0.2788601 -0.1039623  2.0706846   2.2426101
#> [4,] -0.2496595  0.3497223 -1.3717051  -1.2913179
#> [5,] -1.3845025 -0.2762635 -9.4144454 -17.4351529
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_last_draw`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|