R6 Class Representing PosteriorBVARPANEL

The class PosteriorBVARPANEL contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvarPANEL$new()

specify_posterior_bvarPANEL$get_posterior()

specify_posterior_bvarPANEL$get_last_draw()

specify_posterior_bvarPANEL$clone()

Method new()

Create a new posterior output PosteriorBVARPANEL.

Usage

specify_posterior_bvarPANEL$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARPANEL.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvarPANEL$get_posterior()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvarPANEL$get_last_draw()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvarPANEL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARPANEL" "R6"                

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_posterior`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $A
#> , , 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.6800815  1.3099204  0.8180461  0.8162666
#> [2,]  0.8131801  1.1927670 -0.1708331 -0.8006308
#> [3,] -0.7079766 -1.0267312 -0.7117171 -0.6417245
#> [4,] -0.9772367  0.4448568 -0.7086750 -1.3494435
#> [5,] -4.8916840 -2.4207416  2.0584405  3.6904489
#> 
#> , , 2
#> 
#>              [,1]        [,2]        [,3]       [,4]
#> [1,]  0.966585953  0.02007789  0.73746314  0.4421675
#> [2,]  0.006358996  0.63514850  0.05129154  0.3708912
#> [3,] -0.064054938 -0.19264006  0.57046534  0.5475853
#> [4,]  0.073767188  0.16058755 -0.31959608 -0.1497296
#> [5,] -4.220883568  0.22355941 -1.23706551  1.8053586
#> 
#> , , 3
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.94868259  0.5750524 -0.26680693 -0.1484242
#> [2,]  0.08243244  0.8293855  0.00385103  0.3959273
#> [3,]  0.29784669 -0.6139337  1.07169091  0.7716398
#> [4,] -0.14726776  0.5336887 -0.59277784 -0.3274706
#> [5,] -4.32243515 -3.1895619  3.26976942  4.1999446
#> 
#> , , 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.0279222  0.2940349 -0.1037854 0.01599661
#> [2,]  0.1320411  0.3378471  0.3356529 0.49163123
#> [3,]  0.0496751  0.3450629  1.0328820 1.03100307
#> [4,] -0.1454711 -0.5768082  0.3416741 0.24100062
#> [5,] -3.5010317 -6.9407662  7.0278193 5.24425530
#> 
#> , , 5
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  1.02601043  0.0697262  0.1303839 -0.03279871
#> [2,]  0.09767452  0.4854321  0.3392156  0.58833283
#> [3,]  0.19763343 -0.4787360  0.7879872  0.35917884
#> [4,] -0.07147344  0.8665608 -0.6150078  0.05831735
#> [5,] -3.00687398 -6.3084728 11.1187012  9.01346134
#> 
#> 
#> $V
#> , , 1
#> 
#>           [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,]  6.887888 -1.7128280 -3.1177381  1.0546902  4.136431
#> [2,] -1.712828  5.1818853  1.2721303  0.9074447 -2.697600
#> [3,] -3.117738  1.2721303  5.9976883  0.5076875 -4.961151
#> [4,]  1.054690  0.9074447  0.5076875  5.9427586 -1.075092
#> [5,]  4.136431 -2.6976003 -4.9611508 -1.0750920 52.143074
#> 
#> , , 2
#> 
#>             [,1]        [,2]        [,3]        [,4]       [,5]
#> [1,]  1.20680914  0.18059673 -0.08745323  0.16546364  0.1587940
#> [2,]  0.18059673  1.82438054  0.40260377 -0.08001125  0.7613832
#> [3,] -0.08745323  0.40260377  3.33799192 -1.34684720 -3.3296815
#> [4,]  0.16546364 -0.08001125 -1.34684720  1.96938523  4.8519999
#> [5,]  0.15879400  0.76138323 -3.32968153  4.85199986 50.0759056
#> 
#> , , 3
#> 
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,]  2.0621794 -0.3918304 -0.4398127  0.5509549 -0.2804561
#> [2,] -0.3918304  2.2862005  1.0449986 -0.7755447 -1.1794423
#> [3,] -0.4398127  1.0449986  4.6577830 -2.9469504 -1.3973784
#> [4,]  0.5509549 -0.7755447 -2.9469504  3.8681642 -2.1445136
#> [5,] -0.2804561 -1.1794423 -1.3973784 -2.1445136 60.1741975
#> 
#> , , 4
#> 
#>            [,1]       [,2]       [,3]       [,4]        [,5]
#> [1,]  2.1155614 -0.4445398  -1.137163  0.1436725   0.5491031
#> [2,] -0.4445398  3.7597203   2.655587 -2.6270665  -3.5475803
#> [3,] -1.1371627  2.6555874   6.254752 -3.4975091 -13.4339810
#> [4,]  0.1436725 -2.6270665  -3.497509  4.6451168   8.5796608
#> [5,]  0.5491031 -3.5475803 -13.433981  8.5796608  93.6210844
#> 
#> , , 5
#> 
#>             [,1]        [,2]        [,3]        [,4]       [,5]
#> [1,]  1.09823750  -0.0550783  -0.3333771  0.07713165  -3.118746
#> [2,] -0.05507830   3.8990142   3.6801343 -3.56192255 -10.068635
#> [3,] -0.33337706   3.6801343   7.7229185 -5.66687448 -18.654406
#> [4,]  0.07713165  -3.5619226  -5.6668745  6.32799426  12.691316
#> [5,] -3.11874616 -10.0686346 -18.6544062 12.69131629 147.056310
#> 
#> 
#> $Sigma
#> , , 1
#> 
#>            [,1]      [,2]      [,3]       [,4]
#> [1,]  8.4328783 -2.905372 -1.412230 -0.6330499
#> [2,] -2.9053719 10.757296  1.890219 -1.1000167
#> [3,] -1.4122296  1.890219  2.117394  2.9025537
#> [4,] -0.6330499 -1.100017  2.902554  6.6021049
#> 
#> , , 2
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  2.18010618 -0.3313507  0.09225533  0.7385199
#> [2,] -0.33135069 17.4017968 -7.16084545  0.5737707
#> [3,]  0.09225533 -7.1608454 13.46484182  9.2334185
#> [4,]  0.73851992  0.5737707  9.23341850 10.9790026
#> 
#> , , 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.4524408 -0.1291063 -0.5148214 -0.6365702
#> [2,] -0.1291063  8.2292901 -4.0957895  1.0614809
#> [3,] -0.5148214 -4.0957895  8.2628389  6.7401395
#> [4,] -0.6365702  1.0614809  6.7401395  9.5855432
#> 
#> , , 4
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.41440669 -0.4013824  0.3032507 -0.08043514
#> [2,] -0.40138243  5.9619911 -0.8861421  3.58586287
#> [3,]  0.30325070 -0.8861421  8.9522271  8.75565390
#> [4,] -0.08043514  3.5858629  8.7556539 12.43714607
#> 
#> , , 5
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.2209735 -0.192993 -0.3360141 -0.6592238
#> [2,] -0.1929930  6.504349 -2.3103272  2.1800477
#> [3,] -0.3360141 -2.310327  7.9791618  7.0921948
#> [4,] -0.6592238  2.180048  7.0921948  9.8839983
#> 
#> 
#> $nu
#>          [,1]
#> [1,] 5.100001
#> [2,] 5.100101
#> [3,] 5.099932
#> [4,] 5.100097
#> [5,] 5.100287
#> 
#> $m
#>            [,1]
#> [1,]  0.9146422
#> [2,] -0.5427897
#> [3,]  2.0970343
#> [4,]  1.6296786
#> [5,]  2.7544943
#> 
#> $w
#>           [,1]
#> [1,] 105.07613
#> [2,]  29.90662
#> [3,]  41.50658
#> [4,]  35.92115
#> [5,]  24.78660
#> 
#> $s
#>           [,1]
#> [1,] 0.3478831
#> [2,] 1.3242693
#> [3,] 1.2764205
#> [4,] 0.7955746
#> [5,] 0.9791129
#> 
#> $scale
#>            [,1]
#> [1,] 0.01427928
#> [2,] 0.01427928
#> [3,] 0.01671019
#> [4,] 0.01884990
#> [5,] 0.02089174
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>              [,1]          [,2]        [,3]         [,4]
#> [1,] 0.0521064466  0.0004989528  0.01209893  0.008660407
#> [2,] 0.0004989528  0.3035591172 -0.22492252 -0.111303659
#> [3,] 0.0120989278 -0.2249225242  0.58081517  0.472555391
#> [4,] 0.0086604067 -0.1113036592  0.47255539  0.441901382
#> 
#> , , 2, 1
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.06159292 -0.02501564  0.04164123  0.02345633
#> [2,] -0.02501564  0.12313243 -0.09193968 -0.02217512
#> [3,]  0.04164123 -0.09193968  0.54131816  0.48792631
#> [4,]  0.02345633 -0.02217512  0.48792631  0.53460720
#> 
#> , , 3, 1
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.03753031 -0.06118032 -0.03650131 -0.09687123
#> [2,] -0.06118032  6.82453725 -4.67383244 -0.47851740
#> [3,] -0.03650131 -4.67383244  5.28501556  2.84060151
#> [4,] -0.09687123 -0.47851740  2.84060151  3.14078438
#> 
#> , , 4, 1
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.10518170  0.06831018 -0.07043271 -0.04353708
#> [2,]  0.06831018  0.24715570 -0.11400064  0.01438692
#> [3,] -0.07043271 -0.11400064  0.31524163  0.23329709
#> [4,] -0.04353708  0.01438692  0.23329709  0.25296354
#> 
#> , , 5, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.07334840 -0.1902408  0.1478059  0.03044491
#> [2,] -0.19024082  4.4633361 -3.6512497 -1.07243653
#> [3,]  0.14780589 -3.6512497  4.0204912  2.00521037
#> [4,]  0.03044491 -1.0724365  2.0052104  1.57680908
#> 
#> , , 1, 2
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.021857775  0.003763154 -0.01804269 -0.01687811
#> [2,]  0.003763154  0.168260066 -0.14053479 -0.08014829
#> [3,] -0.018042692 -0.140534788  0.45571592  0.37237276
#> [4,] -0.016878107 -0.080148291  0.37237276  0.34727115
#> 
#> , , 2, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.01118420 -0.01567393  0.02381962  0.01508133
#> [2,] -0.01567393  0.17863818 -0.18293093 -0.04839383
#> [3,]  0.02381962 -0.18293093  0.37662250  0.26302561
#> [4,]  0.01508133 -0.04839383  0.26302561  0.27632924
#> 
#> , , 3, 2
#> 
#>             [,1]         [,2]       [,3]        [,4]
#> [1,] 0.011504063  0.007296203  0.0314797  0.04863353
#> [2,] 0.007296203  8.038209963 -5.3557937 -0.34890477
#> [3,] 0.031479699 -5.355793733  5.4634516  2.59687676
#> [4,] 0.048633533 -0.348904767  2.5968768  3.01723567
#> 
#> , , 4, 2
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.013475290 -0.002795673 -0.003745744 -0.004799917
#> [2,] -0.002795673  0.260944063 -0.267353170 -0.097107948
#> [3,] -0.003745744 -0.267353170  0.598428474  0.410540830
#> [4,] -0.004799917 -0.097107948  0.410540830  0.370949604
#> 
#> , , 5, 2
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.01405225 -0.1249534  0.06537142 -0.01426982
#> [2,] -0.12495337  5.1405358 -3.66114150 -0.68363557
#> [3,]  0.06537142 -3.6611415  3.76693493  1.83641129
#> [4,] -0.01426982 -0.6836356  1.83641129  1.67652962
#> 
#> , , 1, 3
#> 
#>              [,1]         [,2]         [,3]        [,4]
#> [1,]  0.009444013 -0.005099183 -0.008867595 -0.01316817
#> [2,] -0.005099183  0.280570435 -0.254205640 -0.12342714
#> [3,] -0.008867595 -0.254205640  0.425740859  0.31535038
#> [4,] -0.013168166 -0.123427135  0.315350377  0.28091516
#> 
#> , , 2, 3
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.008155267 -0.012103337  0.02159989 0.012427632
#> [2,] -0.012103337  0.127443262 -0.09429493 0.003358652
#> [3,]  0.021599894 -0.094294926  0.31646847 0.271321985
#> [4,]  0.012427632  0.003358652  0.27132199 0.312301296
#> 
#> , , 3, 3
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.006389324 -0.04171692  0.03202105 0.007132393
#> [2,] -0.041716920  6.31194486 -3.80759516 0.405651681
#> [3,]  0.032021047 -3.80759516  4.32696896 2.267892485
#> [4,]  0.007132393  0.40565168  2.26789249 3.178831481
#> 
#> , , 4, 3
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.007292963 -0.005967318  0.004456785 -0.001805319
#> [2,] -0.005967318  0.155692488 -0.116397137  0.004389317
#> [3,]  0.004456785 -0.116397137  0.275275424  0.190781858
#> [4,] -0.001805319  0.004389317  0.190781858  0.207414651
#> 
#> , , 5, 3
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.008766207 -0.05332282  0.08422817  0.05673838
#> [2,] -0.053322823  3.31347999 -2.26895786 -0.31856086
#> [3,]  0.084228168 -2.26895786  2.74263269  1.54020234
#> [4,]  0.056738381 -0.31856086  1.54020234  1.50563409
#> 
#> , , 1, 4
#> 
#>             [,1]         [,2]         [,3]         [,4]
#> [1,] 0.006524045  0.003918072  0.000124948  0.001337607
#> [2,] 0.003918072  0.201672260 -0.232494919 -0.146916575
#> [3,] 0.000124948 -0.232494919  0.528808215  0.452297704
#> [4,] 0.001337607 -0.146916575  0.452297704  0.425646612
#> 
#> , , 2, 4
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.006326346 -0.01023806  0.01271088 0.004489827
#> [2,] -0.010238060  0.12288988 -0.07382674 0.017162544
#> [3,]  0.012710882 -0.07382674  0.28307899 0.259945476
#> [4,]  0.004489827  0.01716254  0.25994548 0.318496355
#> 
#> , , 3, 4
#> 
#>              [,1]        [,2]         [,3]       [,4]
#> [1,]  0.003597810  0.02924054 -0.009762384 0.01526885
#> [2,]  0.029240536 11.91845092 -6.135431820 2.19598175
#> [3,] -0.009762384 -6.13543182  4.796635211 0.93864785
#> [4,]  0.015268849  2.19598175  0.938647850 3.04256274
#> 
#> , , 4, 4
#> 
#>              [,1]          [,2]         [,3]          [,4]
#> [1,] 6.058316e-03  2.294796e-05  0.003614113  0.0004979001
#> [2,] 2.294796e-05  2.229059e-01 -0.171973258 -0.0098049577
#> [3,] 3.614113e-03 -1.719733e-01  0.323662170  0.2126732611
#> [4,] 4.979001e-04 -9.804958e-03  0.212673261  0.2254491457
#> 
#> , , 5, 4
#> 
#>              [,1]       [,2]       [,3]        [,4]
#> [1,]  0.009617843 -0.1153454  0.1050159  0.03578034
#> [2,] -0.115345410  4.2895198 -2.9420519 -0.34847535
#> [3,]  0.105015906 -2.9420519  3.0991259  1.43182712
#> [4,]  0.035780339 -0.3484754  1.4318271  1.34775114
#> 
#> , , 1, 5
#> 
#>               [,1]          [,2]        [,3]        [,4]
#> [1,]  0.0148712485  0.0004666264 -0.02798719 -0.03606339
#> [2,]  0.0004666264  0.1638462158 -0.18897616 -0.12792738
#> [3,] -0.0279871872 -0.1889761593  0.57331361  0.55274097
#> [4,] -0.0360633924 -0.1279273798  0.55274097  0.57125570
#> 
#> , , 2, 5
#> 
#>              [,1]         [,2]        [,3]         [,4]
#> [1,]  0.004483071 -0.005488353  0.01309404  0.009223884
#> [2,] -0.005488353  0.091123385 -0.07138416 -0.002624580
#> [3,]  0.013094040 -0.071384158  0.26047541  0.235641068
#> [4,]  0.009223884 -0.002624580  0.23564107  0.274301885
#> 
#> , , 3, 5
#> 
#>              [,1]        [,2]         [,3]         [,4]
#> [1,]  0.001923570  0.01584324 -0.005725721  0.004133352
#> [2,]  0.015843244  7.44309458 -5.008093199 -0.147667061
#> [3,] -0.005725721 -5.00809320  5.838197073  3.009613509
#> [4,]  0.004133352 -0.14766706  3.009613509  3.458172497
#> 
#> , , 4, 5
#> 
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.003553055 -0.00105551 -0.009713954 -0.01380929
#> [2,] -0.001055510  0.17723651 -0.198630765 -0.06753517
#> [3,] -0.009713954 -0.19863077  0.515413234  0.38824257
#> [4,] -0.013809293 -0.06753517  0.388242570  0.36412582
#> 
#> , , 5, 5
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.00751486 -0.06207407  0.07253154 0.03640779
#> [2,] -0.06207407  2.91027752 -1.71905696 0.03434960
#> [3,]  0.07253154 -1.71905696  2.07166717 1.13653710
#> [4,]  0.03640779  0.03434960  1.13653710 1.27087360
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.9043442  0.1638990  0.45869509  0.46946715
#> [2,]  0.3404833  1.1269777 -0.75850515 -0.41021542
#> [3,]  0.6778455  0.6724353  0.03267721  0.22070785
#> [4,] -0.5618039 -0.6736841  0.80952903  0.63800664
#> [5,] -3.4237585 -1.9892875 -0.76459899 -0.06325994
#> 
#> , , 2, 1
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  1.02174103 -0.01641185  0.15361990 -0.03822023
#> [2,]  0.02949762  0.88932812 -0.01594829  0.74325262
#> [3,] -0.05600755 -0.10210830  0.53092959  0.60673062
#> [4,]  0.09762208  0.17416955  0.29076185  0.28947856
#> [5,] -4.93000072 -4.67646908  4.35565427  4.30171042
#> 
#> , , 3, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.0598052 -0.8712125  0.6084162 -0.3314679
#> [2,]  0.2600997  0.9017134  0.2554414  0.7007267
#> [3,]  0.3834066  0.5832061  0.9723100  1.2127103
#> [4,] -0.3087157 -0.3431108 -0.2871394 -0.2955492
#> [5,] -6.2477130 12.8935617  0.4630106 12.9110370
#> 
#> , , 4, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.7030865 -0.1709196  1.5836372  1.1394883
#> [2,]  0.4429508  1.2189758  0.2653319  0.9448583
#> [3,]  0.6444959  0.5366924  1.2078527  1.4039437
#> [4,] -0.4934749 -0.4992423 -0.7070634 -0.7938136
#> [5,] -3.5849131  2.0379709 -3.5345396  0.3261158
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  1.2021723  -2.062066  1.843628  0.3822149
#> [2,] -0.2878415   2.504051 -1.252518  0.6003166
#> [3,] -0.4726868   2.852201 -1.336214  0.9474845
#> [4,]  0.4136370  -1.646612  1.227830 -0.2882409
#> [5,] -1.7247032 -15.273767 16.026397 10.2702034
#> 
#> , , 1, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.0344105  0.1746260  0.3262781  0.2412197
#> [2,]  0.3534542  1.2337877 -1.0709080 -0.6567118
#> [3,]  0.8512805  0.4038356 -0.5226691 -0.5417990
#> [4,] -0.7339268 -0.3714937  1.2231275  1.3203352
#> [5,] -5.8237992 -5.7362227  9.3439224  8.0028669
#> 
#> , , 2, 2
#> 
#>              [,1]        [,2]        [,3]       [,4]
#> [1,]  1.048691792  0.02751907 -0.02941048 -0.2074852
#> [2,] -0.008427645  0.90926234  0.44074646  1.1808405
#> [3,] -0.038797098 -0.35548073  1.51330194  1.2807368
#> [4,]  0.092771820  0.29446111 -0.52286918 -0.3160792
#> [5,] -5.668073051  0.86420989  0.72418960  4.4933204
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]        [,3]      [,4]
#> [1,]  1.0102384 -2.5262935   0.8952134 -1.395914
#> [2,]  0.2016433  0.5859380   1.2342638  1.721262
#> [3,]  0.3035634  0.6273585   2.2006073  2.868219
#> [4,] -0.2510167 -0.0677397  -1.4139389 -1.547881
#> [5,] -3.6526381 36.6314368 -15.6688461 12.883454
#> 
#> , , 4, 2
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.09315056 -0.2087789  1.2607588  0.9794849
#> [2,] -0.01554261  0.7481766  0.2672623  0.8720377
#> [3,]  0.05499286  0.2626631  1.8070712  2.0421477
#> [4,] -0.04244277 -0.1838582 -1.1725663 -1.3347403
#> [5,] -3.23064845  0.6374212 -4.1899166 -1.5754992
#> 
#> , , 5, 2
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  1.0606218 -0.9121866   1.149449   0.4556643
#> [2,] -0.1290826  0.9313465   1.994970   2.7583986
#> [3,] -0.2017005  0.2210898   3.923224   4.3653774
#> [4,]  0.2245354  0.1399772  -3.226653  -3.3918743
#> [5,] -2.9079081  4.8744257 -14.956725 -11.8905133
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.9380551  0.2859087  0.03764546  0.03988732
#> [2,]  0.3792840  1.1188478 -0.94287907 -0.57074678
#> [3,]  0.7685119  0.5393582 -0.70489805 -0.62585654
#> [4,] -0.6445428 -0.5321810  1.46425796  1.44369878
#> [5,] -4.4706620 -5.7248890 11.36828702  9.64127851
#> 
#> , , 2, 3
#> 
#>             [,1]         [,2]       [,3]       [,4]
#> [1,]  1.01729978 -0.143297647  0.1430398 -0.1747462
#> [2,] -0.08302783  0.926837419  0.4595209  1.2269379
#> [3,] -0.01472246  0.002771361  1.2139993  1.2831785
#> [4,]  0.07009191  0.139665418 -0.3818010 -0.3067314
#> [5,] -3.48547977 -6.233959765  4.5708032  2.1147955
#> 
#> , , 3, 3
#> 
#>            [,1]      [,2]        [,3]      [,4]
#> [1,]  1.0009709 -1.730996 -0.03888532 -1.788796
#> [2,]  0.2370036  2.804989 -0.35371852  1.530557
#> [3,]  0.3560340  3.754577  0.04884503  2.699844
#> [4,] -0.3009461 -2.903831  0.52653889 -1.420543
#> [5,] -3.6295266 -3.456696 22.20282407 25.456729
#> 
#> , , 4, 3
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.0703075 -0.01745892  0.6994740  0.3984485
#> [2,]  0.0957217  0.86333728  0.1735749  0.7154659
#> [3,]  0.1382098  0.09027456  1.1472870  1.0552573
#> [4,] -0.1307480 -0.04331594 -0.4163020 -0.2275613
#> [5,] -2.3873130 -2.53270843  2.0971734  3.0674115
#> 
#> , , 5, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.0914908 -0.8876795  0.5435624 -0.2587902
#> [2,] -0.1002485  1.9281555  1.6211130  3.0345128
#> [3,] -0.1625966  1.7988193  3.3961381  4.8810392
#> [4,]  0.1925887 -1.2302081 -2.6917627 -3.7543481
#> [5,] -4.2402533 -8.8194712  1.7129996 -1.2378534
#> 
#> , , 1, 4
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.9711557  0.3643024 -0.05677208 -0.06011321
#> [2,]  0.3714763  1.0970110 -1.09062799 -0.72130974
#> [3,]  0.7933734  0.3918069 -0.76403085 -0.85955409
#> [4,] -0.6703666 -0.4151409  1.51568823  1.68372023
#> [5,] -5.0109833 -6.3954286 14.89921798 12.06964913
#> 
#> , , 2, 4
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  1.04961674 -0.01171315 -0.1753829 -0.4392176
#> [2,] -0.04885110  0.72671455  0.5052577  1.1429242
#> [3,] -0.07955478  0.03114941  1.2477758  1.3448784
#> [4,]  0.10641840  0.10243462 -0.3455650 -0.2923906
#> [5,] -3.47391986 -5.10473476  6.6606752  4.9635592
#> 
#> , , 3, 4
#> 
#>            [,1]      [,2]       [,3]        [,4]
#> [1,]  1.1098051 -4.135675   3.614580   0.7953454
#> [2,]  0.1091245  4.705734  -2.326671   0.5003423
#> [3,]  0.1439263  6.816092  -3.258239   0.8792011
#> [4,] -0.1210428 -5.545198   3.372769   0.1412195
#> [5,] -4.1860511 24.096044 -33.238228 -19.3135235
#> 
#> , , 4, 4
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.14036500 -0.1217409  0.9131395  0.5899482
#> [2,]  0.02233462  0.8472821 -0.7398324 -0.1057588
#> [3,]  0.04761786  0.4000568 -0.1307805  0.1271644
#> [4,] -0.07035109 -0.2912443  0.7351702  0.6170382
#> [5,] -1.87029012 -4.1984812  6.1540494  4.6196028
#> 
#> , , 5, 4
#> 
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  1.1426263 -2.276326  1.4508356 -0.142142
#> [2,] -0.2653332  3.992495 -0.4877954  2.090592
#> [3,] -0.4365637  5.307262 -0.3474373  3.084532
#> [4,]  0.4080251 -4.101097  0.3194888 -2.339774
#> [5,] -2.0894942 -9.538571 20.2388503 17.442086
#> 
#> , , 1, 5
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.8708142  0.18761479  0.2244257  0.3281845
#> [2,]  0.3542027  1.23384387 -1.2910518 -0.8428669
#> [3,]  0.7179327  0.16573817 -0.7373861 -0.7592094
#> [4,] -0.6057190 -0.09247284  1.3689734  1.5061884
#> [5,] -2.3972997 -8.93221786 15.9081128  8.2378098
#> 
#> , , 2, 5
#> 
#>             [,1]         [,2]       [,3]       [,4]
#> [1,]  1.01776672  0.008161884 -0.1638001 -0.3713072
#> [2,] -0.08038968  0.973426794  0.4139307  1.2407263
#> [3,] -0.04884010 -0.133153325  1.0049654  0.9325267
#> [4,]  0.08159059  0.232113094 -0.2873859 -0.1185590
#> [5,] -2.22734085 -9.095274599 19.0255707 14.8161539
#> 
#> , , 3, 5
#> 
#>             [,1]      [,2]      [,3]       [,4]
#> [1,]  1.03986371 -4.812420  4.631138  1.5736166
#> [2,]  0.05127646  6.925575 -5.259167 -1.4544105
#> [3,]  0.06676512 10.571697 -8.332408 -2.5360938
#> [4,] -0.05564947 -8.644803  7.511184  2.9291266
#> [5,] -1.71415440  1.702329 -2.402394 -0.5659088
#> 
#> , , 4, 5
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  1.0651424 -0.03915358  0.5742678  0.38061756
#> [2,]  0.1088552  0.76202298 -0.8168353 -0.21485883
#> [3,]  0.1697703  0.38546696 -0.2912071 -0.04182998
#> [4,] -0.1721609 -0.30172382  0.9716564  0.82614225
#> [5,] -1.4299371 -4.04226316  8.9263707  6.93939433
#> 
#> , , 5, 5
#> 
#>              [,1]       [,2]       [,3]         [,4]
#> [1,]  1.024437897  -1.660578  1.1566668  0.006022992
#> [2,] -0.222883906   3.427592  0.3991248  2.682767783
#> [3,] -0.361674047   4.491773  1.2295406  4.273794202
#> [4,]  0.351281194  -3.454582 -0.8430973 -3.186351984
#> [5,]  0.005952966 -15.126804  3.5334184 -6.231089782
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_last_draw`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|