R6 Class Representing PosteriorBVARPANEL

The class PosteriorBVARPANEL contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvarPANEL$new()

specify_posterior_bvarPANEL$get_posterior()

specify_posterior_bvarPANEL$get_last_draw()

specify_posterior_bvarPANEL$clone()

Method new()

Create a new posterior output PosteriorBVARPANEL.

Usage

specify_posterior_bvarPANEL$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARPANEL.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvarPANEL$get_posterior()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvarPANEL$get_last_draw()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvarPANEL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARPANEL" "R6"                

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_posterior`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $A
#> , , 1
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,] -0.21102094 -0.59817688 -0.37343794 -0.3252091
#> [2,]  0.29479247  0.07813029 -0.67637211  0.5461679
#> [3,] -0.07028739 -0.06761494  0.42547816 -0.5939923
#> [4,]  0.09250135 -2.08876854  1.03006061 -0.5029481
#> [5,]  4.65273039 -1.08553070 -0.06249677  0.5603375
#> 
#> , , 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.5549592 -0.5979822 0.94335082 0.36930858
#> [2,]  0.4055926  0.8141558 0.62517777 1.00971932
#> [3,]  0.6527925  0.3129333 0.45109432 0.54477832
#> [4,] -0.3640230 -0.1505244 0.01906171 0.07253341
#> [5,]  4.3221029 -3.0418161 4.41746328 3.52824098
#> 
#> , , 3
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  0.7140733 -0.1386995 0.3738449 0.08701643
#> [2,]  0.4214097  0.6158159 0.3110532 0.63121468
#> [3,]  0.7636865  0.2419256 0.4474132 0.51632948
#> [4,] -0.6532447 -0.3289742 0.4672667 0.34548494
#> [5,]  5.2494293 -4.6399936 8.8206678 7.44068182
#> 
#> , , 4
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.6564202  0.3322361 -0.09506144 -0.05060036
#> [2,]  0.3457996  1.0008025 -0.04989293  0.67624268
#> [3,]  0.5991228  0.3464754  1.34076340  1.53674697
#> [4,] -0.5311832 -0.4893626 -0.22438594 -0.41285780
#> [5,]  3.5344065 -6.9098995  7.57173678  4.59375331
#> 
#> , , 5
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.6657533  0.15370085 -0.3170726 -0.3374872
#> [2,]  0.1897268  1.01617123  1.4835408  2.1940218
#> [3,]  0.5269268  0.25797374  1.1724086  1.2490647
#> [4,] -0.6199356 -0.02482425 -0.6975506 -0.5713343
#> [5,]  2.7507574 -7.66056237 11.1043696  6.6662394
#> 
#> 
#> $V
#> , , 1
#> 
#>            [,1]       [,2]        [,3]        [,4]       [,5]
#> [1,]  4.3706865  0.2965730 -0.33117557 -0.38860864 -4.6299917
#> [2,]  0.2965730  2.3853685 -0.10190517  0.44900457 -1.2534478
#> [3,] -0.3311756 -0.1019052  2.56093988 -0.05129898  1.6732285
#> [4,] -0.3886086  0.4490046 -0.05129898  4.36144105 -0.5031079
#> [5,] -4.6299917 -1.2534478  1.67322845 -0.50310794 39.7275400
#> 
#> , , 2
#> 
#>             [,1]        [,2]       [,3]        [,4]       [,5]
#> [1,]  1.10878855 -0.14564883 -0.4957657 -0.01856652 -0.7640366
#> [2,] -0.14564883  1.60657612  0.4561010  0.05367229 -0.6584377
#> [3,] -0.49576571  0.45610099  1.2759351 -0.24442120  1.5389693
#> [4,] -0.01856652  0.05367229 -0.2444212  1.49973660 -3.4782726
#> [5,] -0.76403662 -0.65843768  1.5389693 -3.47827256 47.0606146
#> 
#> , , 3
#> 
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,]  1.3501290 -0.1702457 -1.6073265  0.9527629 -3.515178
#> [2,] -0.1702457  1.3276825  0.6786225 -0.2587889  1.238505
#> [3,] -1.6073265  0.6786225  6.3506508 -3.0703198  6.993161
#> [4,]  0.9527629 -0.2587889 -3.0703198  2.5693927 -2.824736
#> [5,] -3.5151781  1.2385047  6.9931613 -2.8247363 75.504251
#> 
#> , , 4
#> 
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,]  1.1076878  0.1680254 -0.3512702  0.1464657  0.4255297
#> [2,]  0.1680254  1.5137887  1.0730098 -0.7083787  3.4616009
#> [3,] -0.3512702  1.0730098  3.2194354 -1.7364152  1.0682415
#> [4,]  0.1464657 -0.7083787 -1.7364152  1.7855113 -1.2930257
#> [5,]  0.4255297  3.4616009  1.0682415 -1.2930257 72.8047149
#> 
#> , , 5
#> 
#>            [,1]         [,2]          [,3]       [,4]       [,5]
#> [1,]  2.6352653 -0.823962009   0.381694738 -0.1434388  -1.760318
#> [2,] -0.8239620  4.526859965  -0.009657409  0.8737151   3.371616
#> [3,]  0.3816947 -0.009657409   7.069826622 -5.5673329 -13.253601
#> [4,] -0.1434388  0.873715127  -5.567332858  8.5452576  12.483190
#> [5,] -1.7603177  3.371616040 -13.253601224 12.4831904  89.839249
#> 
#> 
#> $Sigma
#> , , 1
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,] 15.9858144  4.202563 -1.1434568  0.9598129
#> [2,]  4.2025633 15.068055 -5.6636452 -2.0420447
#> [3,] -1.1434568 -5.663645 10.5835250  0.4955069
#> [4,]  0.9598129 -2.042045  0.4955069  5.7709641
#> 
#> , , 2
#> 
#>           [,1]       [,2]      [,3]      [,4]
#> [1,] 4.4703009  0.2248897  2.160326  2.631849
#> [2,] 0.2248897  7.2401475 -5.866892 -1.160012
#> [3,] 2.1603257 -5.8668917 14.064187  8.845690
#> [4,] 2.6318489 -1.1600118  8.845690 10.840761
#> 
#> , , 3
#> 
#>           [,1]       [,2]      [,3]       [,4]
#> [1,] 2.0231442  0.1069567  1.192273  1.5494808
#> [2,] 0.1069567 10.3963802 -5.277404  0.8779304
#> [3,] 1.1922732 -5.2774037 16.028488 13.9594530
#> [4,] 1.5494808  0.8779304 13.959453 16.7010084
#> 
#> , , 4
#> 
#>           [,1]       [,2]       [,3]       [,4]
#> [1,] 0.7150429  0.4318641  0.4250336  0.8306437
#> [2,] 0.4318641  7.8105900 -0.9770071  4.3393319
#> [3,] 0.4250336 -0.9770071  8.8717937  8.6999977
#> [4,] 0.8306437  4.3393319  8.6999977 13.0047920
#> 
#> , , 5
#> 
#>             [,1]        [,2]         [,3]      [,4]
#> [1,] 0.617498773  0.01856483  0.008769986  0.183023
#> [2,] 0.018564831 10.99890805 -4.236585346  2.717269
#> [3,] 0.008769986 -4.23658535 13.033097737 10.821903
#> [4,] 0.183023002  2.71726923 10.821902996 13.254675
#> 
#> 
#> $nu
#>          [,1]
#> [1,] 5.100113
#> [2,] 5.100200
#> [3,] 5.100128
#> [4,] 5.100370
#> [5,] 5.100489
#> 
#> $m
#>            [,1]
#> [1,]  1.2075435
#> [2,]  0.2530813
#> [3,] -1.2273085
#> [4,]  1.1972278
#> [5,]  1.4999158
#> 
#> $w
#>          [,1]
#> [1,] 59.24092
#> [2,] 19.73919
#> [3,] 21.71618
#> [4,] 15.31682
#> [5,] 70.67521
#> 
#> $s
#>           [,1]
#> [1,] 0.5241626
#> [2,] 1.1395217
#> [3,] 2.0888522
#> [4,] 1.0375126
#> [5,] 1.1761258
#> 
#> $scale
#>            [,1]
#> [1,] 0.01427928
#> [2,] 0.01427928
#> [3,] 0.01671019
#> [4,] 0.01841954
#> [5,] 0.02040663
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,] 0.09088701  0.05372832  0.03261613  0.06777226
#> [2,] 0.05372832  0.40476269 -0.27081303 -0.09722487
#> [3,] 0.03261613 -0.27081303  0.48351332  0.34758105
#> [4,] 0.06777226 -0.09722487  0.34758105  0.37047224
#> 
#> , , 2, 1
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,] 0.12650240  0.07440359  0.04407724 0.06568729
#> [2,] 0.07440359  0.34131173 -0.08475882 0.11033536
#> [3,] 0.04407724 -0.08475882  0.35726222 0.30903321
#> [4,] 0.06568729  0.11033536  0.30903321 0.46096934
#> 
#> , , 3, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.2000372 -0.8425424  0.6659865  0.1062414
#> [2,] -0.8425424  9.3761843 -7.0646359 -0.7956437
#> [3,]  0.6659865 -7.0646359  6.4777974  1.9499561
#> [4,]  0.1062414 -0.7956437  1.9499561  1.7575669
#> 
#> , , 4, 1
#> 
#>             [,1]         [,2]         [,3]         [,4]
#> [1,] 0.062467375  0.002259561  0.006493947  0.007937263
#> [2,] 0.002259561  0.340174254 -0.193089184 -0.008213239
#> [3,] 0.006493947 -0.193089184  0.305291968  0.174074317
#> [4,] 0.007937263 -0.008213239  0.174074317  0.194233161
#> 
#> , , 5, 1
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.04557142 -0.02040523  0.06599615  0.07693892
#> [2,] -0.02040523  2.67961979 -1.88115112 -0.25693623
#> [3,]  0.06599615 -1.88115112  2.57260477  1.65707113
#> [4,]  0.07693892 -0.25693623  1.65707113  1.85061980
#> 
#> , , 1, 2
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.036458979  0.02258997 -0.01767191  0.007038182
#> [2,]  0.022589967  0.31321352 -0.24682261 -0.082908798
#> [3,] -0.017671913 -0.24682261  0.34146549  0.229726121
#> [4,]  0.007038182 -0.08290880  0.22972612  0.224866338
#> 
#> , , 2, 2
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.08072826 -0.03646942  0.09463905 0.05944649
#> [2,] -0.03646942  0.23966343 -0.13743132 0.03733567
#> [3,]  0.09463905 -0.13743132  0.56372601 0.49692608
#> [4,]  0.05944649  0.03733567  0.49692608 0.57968928
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  0.1826826 -1.0020904  1.007822  0.4266981
#> [2,] -1.0020904 14.5600313 -9.995020 -0.6810484
#> [3,]  1.0078219 -9.9950196  9.266888  3.3367721
#> [4,]  0.4266981 -0.6810484  3.336772  3.4847440
#> 
#> , , 4, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.02255533 -0.02242141  0.04073694  0.03155567
#> [2,] -0.02242141  0.20634822 -0.20418928 -0.06792809
#> [3,]  0.04073694 -0.20418928  0.35883946  0.23594091
#> [4,]  0.03155567 -0.06792809  0.23594091  0.21871815
#> 
#> , , 5, 2
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.02157271 -0.2159033  0.1805366  0.05399022
#> [2,] -0.21590333  5.1147834 -3.9346563 -0.87908713
#> [3,]  0.18053657 -3.9346563  3.8961582  1.64274856
#> [4,]  0.05399022 -0.8790871  1.6427486  1.23713909
#> 
#> , , 1, 3
#> 
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.029446876  0.01041838 -0.007959158  0.01413873
#> [2,]  0.010418381  0.18857578 -0.212140013 -0.13495768
#> [3,] -0.007959158 -0.21214001  0.487408953  0.42759328
#> [4,]  0.014138733 -0.13495768  0.427593282  0.42015561
#> 
#> , , 2, 3
#> 
#>             [,1]        [,2]         [,3]        [,4]
#> [1,] 0.032180204  0.01228987  0.004153454 0.009418213
#> [2,] 0.012289867  0.15523102 -0.108201008 0.005782880
#> [3,] 0.004153454 -0.10820101  0.377007208 0.314431673
#> [4,] 0.009418213  0.00578288  0.314431673 0.357700070
#> 
#> , , 3, 3
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.09958721 -0.6168613  0.3995401 -0.0257390
#> [2,] -0.61686134  8.2325686 -4.9952245  0.6020188
#> [3,]  0.39954009 -4.9952245  4.6754817  1.5268805
#> [4,] -0.02573900  0.6020188  1.5268805  2.2492189
#> 
#> , , 4, 3
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.006304130 -0.008367482  0.01761131  0.01301292
#> [2,] -0.008367482  0.129134229 -0.12936694 -0.03297456
#> [3,]  0.017611309 -0.129366936  0.32711311  0.23963131
#> [4,]  0.013012924 -0.032974564  0.23963131  0.22921182
#> 
#> , , 5, 3
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.007681408 -0.05754646  0.08470505  0.05661266
#> [2,] -0.057546464  2.64071177 -2.56941181 -1.09449812
#> [3,]  0.084705050 -2.56941181  3.58426630  2.26152664
#> [4,]  0.056612657 -1.09449812  2.26152664  1.78185147
#> 
#> , , 1, 4
#> 
#>               [,1]          [,2]         [,3]         [,4]
#> [1,]  0.0201404061 -0.0001913856 -0.004547809  0.005258111
#> [2,] -0.0001913856  0.1502400039 -0.156813508 -0.104073155
#> [3,] -0.0045478088 -0.1568135077  0.389142865  0.368078123
#> [4,]  0.0052581107 -0.1040731553  0.368078123  0.382703594
#> 
#> , , 2, 4
#> 
#>              [,1]         [,2]        [,3]       [,4]
#> [1,]  0.025361536 -0.003032186  0.04270842 0.04093578
#> [2,] -0.003032186  0.140487491 -0.04470363 0.05404508
#> [3,]  0.042708418 -0.044703626  0.43832390 0.42490957
#> [4,]  0.040935780  0.054045082  0.42490957 0.49118485
#> 
#> , , 3, 4
#> 
#>              [,1]       [,2]       [,3]         [,4]
#> [1,]  0.074704971 -0.5180608  0.3484314  0.005871213
#> [2,] -0.518060751 10.6542304 -7.1417398 -0.139253598
#> [3,]  0.348431359 -7.1417398  6.3683662  1.977823864
#> [4,]  0.005871213 -0.1392536  1.9778239  2.277764057
#> 
#> , , 4, 4
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.007281320 -0.007699787  0.02736038  0.02397859
#> [2,] -0.007699787  0.204152255 -0.22635427 -0.08169442
#> [3,]  0.027360385 -0.226354273  0.51413983  0.36525213
#> [4,]  0.023978592 -0.081694420  0.36525213  0.32347486
#> 
#> , , 5, 4
#> 
#>              [,1]       [,2]       [,3]         [,4]
#> [1,]  0.008488123 -0.1203896  0.0753642  0.006284737
#> [2,] -0.120389611  4.1046598 -2.7457994 -0.325791034
#> [3,]  0.075364204 -2.7457994  2.4320226  0.868687523
#> [4,]  0.006284737 -0.3257910  0.8686875  0.741747368
#> 
#> , , 1, 5
#> 
#>               [,1]         [,2]          [,3]          [,4]
#> [1,]  0.0114286835 -0.009276493 -0.0002896081  0.0001645899
#> [2,] -0.0092764933  0.303370337 -0.2071480796 -0.0819938327
#> [3,] -0.0002896081 -0.207148080  0.3414787100  0.2741428866
#> [4,]  0.0001645899 -0.081993833  0.2741428866  0.2650678417
#> 
#> , , 2, 5
#> 
#>             [,1]         [,2]        [,3]       [,4]
#> [1,] 0.020296574  0.001896606  0.05603335 0.05507588
#> [2,] 0.001896606  0.166797399 -0.06964154 0.05075283
#> [3,] 0.056033348 -0.069641543  0.51240915 0.46681265
#> [4,] 0.055075877  0.050752826  0.46681265 0.51910283
#> 
#> , , 3, 5
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.03425534 -0.3824419  0.3241688  0.07721916
#> [2,] -0.38244192 11.8581982 -8.3981516 -0.58462907
#> [3,]  0.32416880 -8.3981516  7.6588481  2.41547971
#> [4,]  0.07721916 -0.5846291  2.4154797  2.38944591
#> 
#> , , 4, 5
#> 
#>              [,1]         [,2]         [,3]        [,4]
#> [1,]  0.003389598 -0.002812256  0.005561876 0.005008305
#> [2,] -0.002812256  0.173465281 -0.107028213 0.023605040
#> [3,]  0.005561876 -0.107028213  0.341304486 0.270268214
#> [4,]  0.005008305  0.023605040  0.270268214 0.300472360
#> 
#> , , 5, 5
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.004858786 -0.06788097  0.05512706  0.01672057
#> [2,] -0.067880973  4.03991468 -2.84949729 -0.43127191
#> [3,]  0.055127063 -2.84949729  2.81587658  1.20506503
#> [4,]  0.016720575 -0.43127191  1.20506503  1.05527320
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.8971438  0.3140955  0.07004349  0.03297886
#> [2,]  0.1688465  0.9578926 -0.20447955  0.20920992
#> [3,]  0.6009857  0.3638863  1.11204106  1.23849071
#> [4,] -0.5898017 -0.4637757 -0.14314951 -0.20764598
#> [5,]  2.8314634 -0.7655076  2.02006134  0.86845966
#> 
#> , , 2, 1
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.8746870 -0.35972746  0.3142496 -0.2089913
#> [2,] -0.3992081  0.16083393 -0.5074871 -0.2062736
#> [3,] -0.2618785  0.03682527 -0.3751887 -0.1700563
#> [4,]  0.2793350  0.28263143  1.1091756  1.1695540
#> [5,]  5.0805467 -1.56089775  3.7933455  6.4070070
#> 
#> , , 3, 1
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.2631432  -0.3647905 -0.1681786 -0.54812666
#> [2,]  0.2358082   2.4462158 -1.0847952  0.06762092
#> [3,]  0.4707159   2.8189682 -0.9355492  0.22570115
#> [4,] -0.4004516  -2.2221147  1.4592856  0.72282349
#> [5,] 13.2744879 -22.8304765 30.4461508 16.34790172
#> 
#> , , 4, 1
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.7644661 -0.02832524  1.0071544  0.5210827
#> [2,]  0.3526489  0.66355717  0.1624848  0.8466925
#> [3,]  0.4627453  0.39191427  0.9317550  1.2236860
#> [4,] -0.3958581 -0.39355973 -0.2231178 -0.3830901
#> [5,]  1.1312480  2.42099740 -4.6089876 -1.2188342
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.7087640 -2.0305565  1.9249729  0.6135991
#> [2,]  0.3835623  1.0941785  0.4484631  1.1780558
#> [3,]  0.6322349  0.5919052  1.3099149  1.6982677
#> [4,] -0.5714571  0.2608829 -1.1890946 -1.0708391
#> [5,]  4.2006549  5.2695268 -0.6637521  4.8869989
#> 
#> , , 1, 2
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.7940176 -0.02662283  0.1226718 -0.01241014
#> [2,]  0.3166265  1.31997415 -0.4636882  0.11587241
#> [3,]  0.6368374  0.39555705  1.0168514  1.22976667
#> [4,] -0.5781730 -0.32256447 -0.1222330 -0.20259498
#> [5,]  1.8730732 -3.82298092  5.7164801  2.47413550
#> 
#> , , 2, 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.0388670  0.06753551  0.5081901  0.3717961
#> [2,] -0.2456386  0.60267164 -0.3714335  0.1564396
#> [3,] -0.4732802 -0.30462025 -0.5419516 -0.7186913
#> [4,]  0.4390165  0.36270167  1.1707776  1.4346458
#> [5,] -0.2154654 -3.56412791  2.6963680  0.9839620
#> 
#> , , 3, 2
#> 
#>            [,1]        [,2]         [,3]       [,4]
#> [1,]  0.1961255   0.1930945 -0.577819058 -0.6356867
#> [2,]  0.6711173   1.9310254  0.003790147  0.9825868
#> [3,]  1.0553330   1.9518070  0.599607124  1.4710496
#> [4,] -0.9236613  -1.4715125  0.037547753 -0.4581715
#> [5,]  9.9179145 -29.7442943 31.027460031 11.8899927
#> 
#> , , 4, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8051831 -0.3643006  0.9930806  0.5950391
#> [2,]  0.4322042  0.9341493 -0.1199625  0.6185563
#> [3,]  0.4554847  0.4260278  0.8404220  1.1529104
#> [4,] -0.4410754 -0.2169668 -0.2552981 -0.3869730
#> [5,]  3.8805791 -5.5397854  5.8641091  3.1213292
#> 
#> , , 5, 2
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.6870127 -1.1469079  0.9783342  0.08165071
#> [2,]  0.3441814  0.8344007  0.6076998  1.09672137
#> [3,]  0.5615186  0.5097953  1.2232966  1.49943380
#> [4,] -0.5188451  0.1394403 -0.9351580 -0.75787723
#> [5,]  5.7788816 -3.3512532 11.8972403 11.66084392
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]        [,3]         [,4]
#> [1,]  0.8769637  0.1461235 -0.11953838 -0.145989953
#> [2,]  0.2600067  1.0870284 -0.25992491  0.163776593
#> [3,]  0.6844983  0.2428378  0.94544364  1.059630471
#> [4,] -0.6538310 -0.2372424  0.04419289  0.006829769
#> [5,]  1.9201785 -3.2785942  5.17093110  3.011816363
#> 
#> , , 2, 3
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  1.0120661 -0.04770097  0.5384060  0.35433127
#> [2,] -0.2507877  0.85147813 -0.6162671  0.04983095
#> [3,] -0.4103828 -0.26054511 -0.2319896 -0.41386061
#> [4,]  0.3567968  0.33023450  0.8291909  1.08785075
#> [5,]  2.8343258 -5.09806580 12.3657755 10.34837033
#> 
#> , , 3, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.2230982  0.2491746 -0.3714732 -0.3524412
#> [2,]  0.6017311 -0.3102586  0.9563179  0.4105979
#> [3,]  0.9837704 -1.1205247  1.9679887  0.7635953
#> [4,] -0.8494283  1.2515232 -1.1878449  0.1619775
#> [5,]  9.5671533 -3.3821206 15.6004275 13.2296929
#> 
#> , , 4, 3
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.8184101  0.05578072 -0.1806662 -0.29635296
#> [2,]  0.3885862  0.86533027 -0.3091150  0.32479367
#> [3,]  0.4424546  0.43979299  0.6363732  0.93627334
#> [4,] -0.4275824 -0.34851959  0.2405621  0.04088619
#> [5,]  3.5448159 -7.30525618 14.4567527 10.71245735
#> 
#> , , 5, 3
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.7179351 -0.471646  0.1998361 -0.2985503
#> [2,]  0.4034411 -0.488164  1.6544390  1.3790039
#> [3,]  0.6784525 -2.159518  3.4017560  2.1113207
#> [4,] -0.6267563  2.469694 -2.9415328 -1.4069279
#> [5,]  4.5578585 -3.931743 25.0457753 24.8631249
#> 
#> , , 1, 4
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.8911638  0.09375874  0.3371434  0.39103805
#> [2,]  0.2887420  1.16417638 -0.5192155 -0.06061967
#> [3,]  0.7303697  0.23365838  1.1424348  1.33661703
#> [4,] -0.6763009 -0.15552881 -0.3782066 -0.46031568
#> [5,]  0.4383135 -6.05206236  7.8629910  2.25870512
#> 
#> , , 2, 4
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  1.0016295  0.1405117 -0.02419968 -0.1168356
#> [2,] -0.3109841  0.6845164 -0.87811712 -0.3333471
#> [3,] -0.3374328 -0.3670767  0.09669132 -0.1156938
#> [4,]  0.3034060  0.4419898  0.82197242  1.1433834
#> [5,]  3.4022388 -8.6731751  9.84737580  4.7136525
#> 
#> , , 3, 4
#> 
#>            [,1]          [,2]       [,3]       [,4]
#> [1,]  0.2374969  5.126673e-01 -0.4510771 -0.2856234
#> [2,]  0.6247982  1.744587e-04  1.4811475  1.1342283
#> [3,]  1.0028290 -8.189068e-01  2.8301145  1.8394949
#> [4,] -0.8660686  1.022130e+00 -1.8726405 -0.6785351
#> [5,]  8.8647633 -1.529117e+01  6.5747313 -3.4435890
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.7360069 -0.1709031  0.7784424  0.50408374
#> [2,]  0.3004390  1.0574951 -0.2797602  0.49424714
#> [3,]  0.3851234  0.6389930  0.3833086  0.84182265
#> [4,] -0.3324352 -0.4585036  0.2200799 -0.06693018
#> [5,]  2.9444058 -8.3364625  9.6581995  4.53030968
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.7727851 -1.0048411  0.4803025 -0.3170067
#> [2,]  0.3957259 -0.2899409  1.9074900  1.7932364
#> [3,]  0.6534294 -1.6056280  3.6827073  2.7959547
#> [4,] -0.6076577  2.0068440 -3.1207604 -1.9221022
#> [5,]  3.3757550  5.5297648 10.9700781 15.1445458
#> 
#> , , 1, 5
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.9087799  0.3671482 -0.1526593 -0.04054195
#> [2,]  0.2689418  1.1054877 -0.2159701  0.26893012
#> [3,]  0.7440995  0.3623843  1.0187549  1.26897299
#> [4,] -0.6951708 -0.3457458 -0.1192564 -0.26302730
#> [5,]  0.4923534 -8.4326043  9.6510740  2.98525748
#> 
#> , , 2, 5
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  0.9470284   0.2945046  0.06055425  0.06959869
#> [2,] -0.2360352   0.4842619 -0.73305095 -0.32673598
#> [3,] -0.2354566  -0.5284854  0.07093975 -0.26465889
#> [4,]  0.1970453   0.5902420  0.77689818  1.22030960
#> [5,]  5.2264553 -10.3199799 10.56173578  3.64473591
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  0.6524039   1.137074 -1.184218 -0.6603663
#> [2,]  0.2579230   2.286400 -1.288821 -0.2584482
#> [3,]  0.3936466   2.759706 -1.429465 -0.2757007
#> [4,] -0.3543063  -1.878269  1.705315  1.1642535
#> [5,]  5.3482227 -73.811065 68.722046 23.7195495
#> 
#> , , 4, 5
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.8577012  0.4324029  0.8581439  1.08848508
#> [2,]  0.4914353  1.4613483 -1.1523548 -0.06463597
#> [3,]  0.5635288  1.0947825 -0.5788243  0.18734307
#> [4,] -0.5407717 -1.1765485  1.2006484  0.39984366
#> [5,]  1.8946943 -4.3100349  6.5937842  3.75666479
#> 
#> , , 5, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8497462 -1.1540608  0.9703627  0.1545162
#> [2,]  0.3501861  0.3053320  1.3109168  1.5219061
#> [3,]  0.5794834 -0.6890969  2.7186320  2.3253498
#> [4,] -0.5437042  1.2247895 -2.3093755 -1.5394278
#> [5,]  1.9345281  1.1718430  7.1680080  7.7792064
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_last_draw`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|