R6 Class Representing PosteriorBVARPANEL

The class PosteriorBVARPANEL contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvarPANEL$new()

specify_posterior_bvarPANEL$get_posterior()

specify_posterior_bvarPANEL$get_last_draw()

specify_posterior_bvarPANEL$clone()

Method new()

Create a new posterior output PosteriorBVARPANEL.

Usage

specify_posterior_bvarPANEL$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARPANEL.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvarPANEL$get_posterior()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvarPANEL$get_last_draw()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvarPANEL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARPANEL" "R6"                

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_posterior`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $A
#> , , 1
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  -3.101861  2.4717740 -3.4113287 -4.711677
#> [2,]   1.074473 -5.7346986  2.2464140  7.725279
#> [3,]  -6.722068  2.8974794  0.5807068  5.911780
#> [4,] -10.080887  0.9472938 -2.5703935 -4.356170
#> [5,] -16.545891  4.5665852  2.2150997 19.129094
#> 
#> , , 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.7310070 -0.06428756  0.2588496  0.1027316
#> [2,] -0.2428132 -0.51142666  1.5426785  1.3965642
#> [3,] -0.2083215 -1.62797420  2.2325211  1.3693015
#> [4,]  0.2548957  1.51074882 -1.0760026 -0.1061787
#> [5,]  6.0191328  9.33042960  4.9094479  7.4353788
#> 
#> , , 3
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.9341534 -0.5792319  0.05644837 -0.4802316
#> [2,] -0.2971947  1.3035567 -0.06774913  0.7824324
#> [3,] -0.1373208  1.6123463  1.47293406  2.9271638
#> [4,]  0.2007100 -1.1988449 -0.41618408 -1.4432013
#> [5,]  2.4289241 -3.4375891 15.51573201  7.4260140
#> 
#> , , 4
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.8448647   0.3939028 -0.4337285 -0.2308758
#> [2,] -0.3724110  -0.2160656 -0.1412543 -0.4504109
#> [3,] -0.3689082  -0.2779794  0.6032769  0.3104395
#> [4,]  0.3794825   0.1958716  0.1711403  0.4860617
#> [5,]  4.7347490 -10.6089414 25.1566877 12.8890528
#> 
#> , , 5
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.84674726  0.05641097 -0.05403518  0.01874529
#> [2,] -0.07087738  1.05114109 -0.49197023 -0.06167609
#> [3,] -0.01986757  0.23200373  2.17599137  2.12195553
#> [4,]  0.10102712 -0.21169027 -0.94482248 -0.81591371
#> [5,]  6.32028095 -4.90991475 27.64557659 20.92082241
#> 
#> 
#> $V
#> , , 1
#> 
#>              [,1]        [,2]      [,3]       [,4]       [,5]
#> [1,]  480.0808013   0.4080859 -71.19826 -100.93569  152.30095
#> [2,]    0.4080859 272.0927912  19.75102   60.67233  164.19123
#> [3,]  -71.1982584  19.7510229 506.99047   50.07381  -34.69213
#> [4,] -100.9356890  60.6723297  50.07381  433.05810  556.16220
#> [5,]  152.3009517 164.1912262 -34.69213  556.16220 4317.70168
#> 
#> , , 2
#> 
#>               [,1]        [,2]          [,3]       [,4]        [,5]
#> [1,]   1.782967972  -0.3662845   0.001182646   0.117917  -39.389319
#> [2,]  -0.366284495   8.1077894  10.788561123 -10.229146   -4.777994
#> [3,]   0.001182646  10.7885611  18.540685497 -17.245367  -32.752899
#> [4,]   0.117917036 -10.2291464 -17.245367276  17.344909   23.169941
#> [5,] -39.389319188  -4.7779937 -32.752898610  23.169941 1651.119692
#> 
#> , , 3
#> 
#>             [,1]        [,2]       [,3]       [,4]       [,5]
#> [1,]   6.8832790  -0.9210524  -5.288933   5.481263  -43.01986
#> [2,]  -0.9210524   8.3230088  10.061964  -9.426651  -23.71215
#> [3,]  -5.2889330  10.0619644  21.650044 -18.918356  -42.70520
#> [4,]   5.4812626  -9.4266515 -18.918356  20.300747   31.67318
#> [5,] -43.0198624 -23.7121451 -42.705196  31.673182 1251.08556
#> 
#> , , 4
#> 
#>            [,1]       [,2]      [,3]       [,4]        [,5]
#> [1,]   5.808989 -10.160045 -11.82666  11.083959  -29.374586
#> [2,] -10.160045  33.475765  39.78309 -33.850987    7.778521
#> [3,] -11.826663  39.783088  54.01795 -43.641832  -29.315451
#> [4,]  11.083959 -33.850987 -43.64183  39.240706   -6.135903
#> [5,] -29.374586   7.778521 -29.31545  -6.135903 1239.386157
#> 
#> , , 5
#> 
#>           [,1]       [,2]        [,3]      [,4]        [,5]
#> [1,]  6.005181  -2.812321   -8.657881  10.05732   -7.400798
#> [2,] -2.812321  25.376519   28.684804 -26.61678  -72.163997
#> [3,] -8.657881  28.684804   55.011465 -52.19887 -140.556649
#> [4,] 10.057316 -26.616782  -52.198869  52.30450  126.969578
#> [5,] -7.400798 -72.163997 -140.556649 126.96958 1243.264035
#> 
#> 
#> $Sigma
#> , , 1
#> 
#>             [,1]        [,2]        [,3]      [,4]
#> [1,]  4.66680674 -0.42028601 -0.01459697 0.3135623
#> [2,] -0.42028601  3.78775253 -0.05360427 1.0843217
#> [3,] -0.01459697 -0.05360427  2.69169727 1.6189142
#> [4,]  0.31356226  1.08432174  1.61891416 4.8611760
#> 
#> , , 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.1000274 -0.3974063 -0.1413494 -0.4209503
#> [2,] -0.3974063  8.9609306 -3.5164323  1.8055009
#> [3,] -0.1413494 -3.5164323  5.1147509  2.9258670
#> [4,] -0.4209503  1.8055009  2.9258670  5.3422352
#> 
#> , , 3
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 1.0075640  0.5239243  0.4939861  1.055792
#> [2,] 0.5239243  5.0255811 -1.0817996  2.518029
#> [3,] 0.4939861 -1.0817996  8.5490401  8.993216
#> [4,] 1.0557923  2.5180289  8.9932164 12.441472
#> 
#> , , 4
#> 
#>             [,1]        [,2]       [,3]      [,4]
#> [1,]  0.47719269 -0.06205841  0.2610444 0.3821429
#> [2,] -0.06205841  3.78821693 -2.0931438 0.4725394
#> [3,]  0.26104444 -2.09314378  4.8476033 3.9603593
#> [4,]  0.38214291  0.47253944  3.9603593 5.0281087
#> 
#> , , 5
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.2588828 -0.1707855  0.2352692 0.1443335
#> [2,] -0.1707855  3.6548882 -1.2696854 1.1679148
#> [3,]  0.2352692 -1.2696854  3.6626592 3.2912114
#> [4,]  0.1443335  1.1679148  3.2912114 4.6851409
#> 
#> 
#> $nu
#>          [,1]
#> [1,] 5.100221
#> [2,] 5.100264
#> [3,] 5.100264
#> [4,] 5.100506
#> [5,] 5.100923
#> 
#> $m
#>           [,1]
#> [1,] 0.9431663
#> [2,] 2.2442621
#> [3,] 1.7067711
#> [4,] 1.6864638
#> [5,] 1.4379900
#> 
#> $w
#>            [,1]
#> [1,] 9426.68509
#> [2,]   12.53413
#> [3,]   42.65399
#> [4,]   37.63953
#> [5,]   31.63675
#> 
#> $s
#>           [,1]
#> [1,] 0.1876297
#> [2,] 0.4171976
#> [3,] 0.5780544
#> [4,] 0.4344424
#> [5,] 0.4614962
#> 
#> $scale
#>            [,1]
#> [1,] 0.01427928
#> [2,] 0.01427928
#> [3,] 0.01613942
#> [4,] 0.01810643
#> [5,] 0.02000144
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>               [,1]          [,2]        [,3]       [,4]
#> [1,]  0.0454262185 -0.0001065369  0.01966620 0.02071896
#> [2,] -0.0001065369  0.1199379535 -0.03941229 0.01381594
#> [3,]  0.0196661978 -0.0394122865  0.20533280 0.21168598
#> [4,]  0.0207189631  0.0138159367  0.21168598 0.26225859
#> 
#> , , 2, 1
#> 
#>              [,1]        [,2]         [,3]         [,4]
#> [1,]  0.020807934 -0.02046234  0.005545498 -0.007249765
#> [2,] -0.020462337  0.13258124 -0.106897208 -0.020066554
#> [3,]  0.005545498 -0.10689721  0.356619033  0.328453352
#> [4,] -0.007249765 -0.02006655  0.328453352  0.393609536
#> 
#> , , 3, 1
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,] 0.04874366  0.04661451  0.03280572  0.04487385
#> [2,] 0.04661451  5.54645659 -3.76448702 -0.23207572
#> [3,] 0.03280572 -3.76448702  4.25149679  2.09295775
#> [4,] 0.04487385 -0.23207572  2.09295775  2.28262162
#> 
#> , , 4, 1
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.03684258 -0.01586554  0.02386327  0.01315390
#> [2,] -0.01586554  0.17424931 -0.15013004 -0.01637942
#> [3,]  0.02386327 -0.15013004  0.31833395  0.17793289
#> [4,]  0.01315390 -0.01637942  0.17793289  0.17338792
#> 
#> , , 5, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.04497923 -0.1230379  0.1083074  0.04882971
#> [2,] -0.12303794  3.1304673 -2.5425435 -0.78682081
#> [3,]  0.10830740 -2.5425435  2.5544312  1.20696975
#> [4,]  0.04882971 -0.7868208  1.2069698  0.88116844
#> 
#> , , 1, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.02180161 -0.01060116 -0.01530129 -0.02175895
#> [2,] -0.01060116  0.34105786 -0.21236737 -0.04567966
#> [3,] -0.01530129 -0.21236737  0.29121928  0.20793632
#> [4,] -0.02175895 -0.04567966  0.20793632  0.21432633
#> 
#> , , 2, 2
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.014914370 -0.02072194  0.02344332 0.006988174
#> [2,] -0.020721940  0.22044865 -0.13212231 0.044683215
#> [3,]  0.023443315 -0.13212231  0.39218038 0.354076558
#> [4,]  0.006988174  0.04468321  0.35407656 0.480216966
#> 
#> , , 3, 2
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.02820459 -0.3135192  0.3440853  0.1551798
#> [2,] -0.31351921 10.8720924 -9.0327502 -2.1026501
#> [3,]  0.34408533 -9.0327502  9.7994126  4.4412774
#> [4,]  0.15517981 -2.1026501  4.4412774  3.5990386
#> 
#> , , 4, 2
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.01863116 -0.01027801  0.0914225  0.0856947
#> [2,] -0.01027801  0.19882900 -0.1868063 -0.0383964
#> [3,]  0.09142250 -0.18680627  0.7867339  0.6596329
#> [4,]  0.08569470 -0.03839640  0.6596329  0.6489319
#> 
#> , , 5, 2
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.01892810 -0.0504008  0.06009861  0.03502323
#> [2,] -0.05040080  2.2888710 -1.84818354 -0.50230312
#> [3,]  0.06009861 -1.8481835  2.01382264  0.99485890
#> [4,]  0.03502323 -0.5023031  0.99485890  0.78040168
#> 
#> , , 1, 3
#> 
#>               [,1]          [,2]         [,3]         [,4]
#> [1,]  0.0091278415 -0.0009111649  0.006128653  0.007544987
#> [2,] -0.0009111649  0.1565630209 -0.162731470 -0.099952321
#> [3,]  0.0061286527 -0.1627314697  0.358946980  0.316563287
#> [4,]  0.0075449874 -0.0999523214  0.316563287  0.305763507
#> 
#> , , 2, 3
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.008341869 -0.009974557  0.03060194  0.02785037
#> [2,] -0.009974557  0.121858419 -0.10118386 -0.01386245
#> [3,]  0.030601942 -0.101183861  0.39426697  0.39129504
#> [4,]  0.027850369 -0.013862450  0.39129504  0.46985671
#> 
#> , , 3, 3
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.03942466 -0.3354982  0.4888743  0.3141692
#> [2,] -0.33549824  7.8657457 -7.5633639 -2.7675855
#> [3,]  0.48887434 -7.5633639  9.8877424  5.7439001
#> [4,]  0.31416923 -2.7675855  5.7439001  4.6303168
#> 
#> , , 4, 3
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.015315807 -0.009563933  0.06577609  0.06231197
#> [2,] -0.009563933  0.113519026 -0.11856180 -0.03127345
#> [3,]  0.065776091 -0.118561796  0.63359561  0.55615856
#> [4,]  0.062311973 -0.031273449  0.55615856  0.54921068
#> 
#> , , 5, 3
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.01038781 -0.06257914  0.08052966  0.04831003
#> [2,] -0.06257914  3.49882041 -2.55781074 -0.48718057
#> [3,]  0.08052966 -2.55781074  2.59523899  1.18297439
#> [4,]  0.04831003 -0.48718057  1.18297439  1.01330227
#> 
#> , , 1, 4
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.006638985 -0.005385861  0.005631786  0.002695394
#> [2,] -0.005385861  0.186581188 -0.091295000 -0.013631733
#> [3,]  0.005631786 -0.091295000  0.218880677  0.203588237
#> [4,]  0.002695394 -0.013631733  0.203588237  0.225898401
#> 
#> , , 2, 4
#> 
#>              [,1]         [,2]        [,3]         [,4]
#> [1,]  0.006344267 -0.007666977  0.01055004  0.004270121
#> [2,] -0.007666977  0.114026793 -0.08826276 -0.001669227
#> [3,]  0.010550044 -0.088262756  0.33208978  0.321595359
#> [4,]  0.004270121 -0.001669227  0.32159536  0.391583751
#> 
#> , , 3, 4
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.02008147 -0.1851489  0.2914674  0.1967419
#> [2,] -0.18514892  6.3885016 -4.9706100 -0.8471166
#> [3,]  0.29146737 -4.9706100  7.0819027  4.4310557
#> [4,]  0.19674186 -0.8471166  4.4310557  4.5551441
#> 
#> , , 4, 4
#> 
#>              [,1]          [,2]        [,3]          [,4]
#> [1,]  0.009638317 -0.0084111206  0.06121735  0.0555376002
#> [2,] -0.008411121  0.1579487639 -0.12316044 -0.0008352712
#> [3,]  0.061217353 -0.1231604399  0.69419435  0.6073412622
#> [4,]  0.055537600 -0.0008352712  0.60734126  0.6170863051
#> 
#> , , 5, 4
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.006740939 -0.06169364  0.04506303  0.01053986
#> [2,] -0.061693638  3.50160482 -2.95212446 -0.92597544
#> [3,]  0.045063034 -2.95212446  3.34917047  1.76533284
#> [4,]  0.010539862 -0.92597544  1.76533284  1.37420611
#> 
#> , , 1, 5
#> 
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.006159821 -0.00625922  0.004297206 -0.00102371
#> [2,] -0.006259220  0.12439547 -0.071979757 -0.01946861
#> [3,]  0.004297206 -0.07197976  0.186862072  0.17160337
#> [4,] -0.001023710 -0.01946861  0.171603373  0.18357788
#> 
#> , , 2, 5
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.004288103 -0.008610496  0.009965321  0.004018691
#> [2,] -0.008610496  0.214639915 -0.183774254 -0.023146868
#> [3,]  0.009965321 -0.183774254  0.414739284  0.334128576
#> [4,]  0.004018691 -0.023146868  0.334128576  0.387919695
#> 
#> , , 3, 5
#> 
#>             [,1]       [,2]      [,3]       [,4]
#> [1,]  0.01553577 -0.1391626  0.169296 0.08539981
#> [2,] -0.13916257  9.4652977 -6.244819 0.04512190
#> [3,]  0.16929602 -6.2448186  5.931170 2.14348425
#> [4,]  0.08539981  0.0451219  2.143484 2.63426724
#> 
#> , , 4, 5
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.006296538 -0.02160850  0.04851166  0.03254773
#> [2,] -0.021608499  0.26717930 -0.30207414 -0.09453187
#> [3,]  0.048511658 -0.30207414  0.66956963  0.44457162
#> [4,]  0.032547725 -0.09453187  0.44457162  0.38266109
#> 
#> , , 5, 5
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.009124972 -0.09119968  0.05714088  0.004008237
#> [2,] -0.091199678  2.85233472 -2.16959606 -0.496387363
#> [3,]  0.057140882 -2.16959606  2.34110791  1.166817501
#> [4,]  0.004008237 -0.49638736  1.16681750  0.991294080
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.9237584  0.3176184  -0.2939912  -0.1052225
#> [2,] -0.3767617 -0.2653304   2.6038140   2.8418388
#> [3,] -0.9015574 -2.9796740   7.2364553   6.7680106
#> [4,]  0.8095701  2.6970454  -5.5090447  -4.9291441
#> [5,]  5.7784923  5.1825879 -24.8282765 -34.1370920
#> 
#> , , 2, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9158771  0.0688558 -0.3028265 -0.3347265
#> [2,] -0.3527865  0.5261305  0.4874389  0.9125092
#> [3,] -0.2387234 -0.5612327  1.1811499  0.9339116
#> [4,]  0.2612423  0.5433607 -0.6000636 -0.2399614
#> [5,]  3.1531164  0.2784740 33.9624026 28.5039256
#> 
#> , , 3, 1
#> 
#>            [,1]      [,2]       [,3]         [,4]
#> [1,]  0.6291251 -4.914519   4.243588   1.25158031
#> [2,] -0.4035496  2.494532  -1.720192  -0.14762554
#> [3,] -0.6220316  3.978023  -2.617441  -0.04363269
#> [4,]  0.4897016 -3.307937   3.066271   1.10146635
#> [5,] 16.9259135 86.517843 -71.623089 -30.52116506
#> 
#> , , 4, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8789507  0.1186262  0.3747961  0.3480881
#> [2,] -0.1710843 -1.5910463  0.2609468 -0.1997769
#> [3,] -0.2230117 -2.6599020  1.1929236  0.1470277
#> [4,]  0.1981026  2.7522891 -0.4134238  0.7531385
#> [5,]  5.1706481 -8.8175430  6.9005135 -0.6349374
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.6226406  -4.082291   1.556776 -1.321828
#> [2,]  0.8993752  -1.539946   3.035408  2.179109
#> [3,]  1.4732627  -3.567699   5.584855  3.543018
#> [4,] -1.3284410   3.578511  -4.879894 -2.808123
#> [5,]  1.3608066 112.382247 -25.873828 51.694264
#> 
#> , , 1, 2
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.8614715  0.2626392   0.2676366   0.4751594
#> [2,] -0.3537642  0.6297165   1.9186270   2.6196328
#> [3,] -0.7743063 -1.0670679   6.4299276   6.9592823
#> [4,]  0.6777589  0.9933366  -5.0016384  -5.3189772
#> [5,]  7.6882994 -3.4643615 -21.3760477 -35.8606726
#> 
#> , , 2, 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.0325460  0.01608323 -0.1468528 -0.1979904
#> [2,] -0.4242670  0.81474246  1.7916788  2.7162127
#> [3,] -0.4900834 -0.22216629  3.3722032  3.6174606
#> [4,]  0.4710888  0.18032115 -2.4332042 -2.5975649
#> [5,]  1.4932423  2.99285012  8.9469414  4.5049477
#> 
#> , , 3, 2
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.5330506 -3.272549  -1.291501  -4.123121
#> [2,] -0.3946698  3.684049  -4.766586  -2.973232
#> [3,] -0.5758675  5.073393  -6.484917  -3.999611
#> [4,]  0.4515851 -4.079116   5.883148   4.037076
#> [5,] 18.9190220 20.381155 130.125415 160.906291
#> 
#> , , 4, 2
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  0.9556238 -0.3727344  1.413626  1.072129
#> [2,] -1.0261340  1.0620586 -6.806836 -6.267221
#> [3,] -1.4571007  0.5273199 -8.356807 -8.308879
#> [4,]  1.3823277 -0.4040667  8.453167  8.526058
#> [5,]  6.6850003  1.1454481 31.376791 31.714815
#> 
#> , , 5, 2
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.7021634 -1.5784874 -0.6669451 -1.807955
#> [2,]  0.4220162  1.3978674  0.4954964  1.531078
#> [3,]  0.6896587  1.1020025  1.4979543  2.464489
#> [4,] -0.6727623 -0.5867234 -1.1117581 -1.678504
#> [5,]  6.9866223 14.4496763 52.9516987 61.296088
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.9792232  0.3184536   0.0751472   0.2451399
#> [2,] -0.4397268  0.7265166   1.9636890   2.7408957
#> [3,] -0.8998612 -0.8684513   6.6909166   7.3556161
#> [4,]  0.7813072  0.8491214  -5.2582243  -5.6810414
#> [5,]  6.2423525 -7.3164988 -16.3894485 -31.5205122
#> 
#> , , 2, 3
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.0279137 -0.01654963 -0.4779219 -0.6295964
#> [2,] -0.3611918  1.09500893  2.3232906  3.6621966
#> [3,] -0.4496924  0.20029141  3.7958865  4.5990678
#> [4,]  0.4346754 -0.10346066 -2.8209907 -3.3860821
#> [5,]  0.7455035 -6.07039349 10.6937962 -3.1347565
#> 
#> , , 3, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.4636506  -1.801681  -2.836163  -4.696384
#> [2,] -0.3280844   2.790617  -3.241187  -2.032857
#> [3,] -0.4704671   3.760300  -4.202837  -2.569727
#> [4,]  0.3628199  -2.772291   3.783628   2.770519
#> [5,] 19.4537543 -10.285828 151.394606 162.875611
#> 
#> , , 4, 3
#> 
#>           [,1]       [,2]      [,3]      [,4]
#> [1,]  1.179742 -0.4173993  3.293661  2.938007
#> [2,] -1.343586  1.0883394 -7.219364 -6.729363
#> [3,] -1.729283  0.5216087 -8.848088 -8.874844
#> [4,]  1.606101 -0.3704069  8.545702  8.711873
#> [5,]  4.767281  0.1486965 12.447307 11.740486
#> 
#> , , 5, 3
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.5752603 -1.5907246 -0.4740389 -1.567716
#> [2,]  0.4956433  0.5137182  1.4820744  2.033484
#> [3,]  0.8435393 -0.5651861  3.3151206  3.363491
#> [4,] -0.7815763  0.7905446 -2.6304594 -2.447847
#> [5,]  7.8268359 30.5508457 31.2980057 47.590140
#> 
#> , , 1, 4
#> 
#>            [,1]         [,2]        [,3]        [,4]
#> [1,]  0.9127343   0.34817766  0.04105692   0.2480794
#> [2,] -0.3429016   1.15172738  1.56804229   2.4893301
#> [3,] -0.7136697  -0.06248393  6.10765813   7.0933477
#> [4,]  0.6254964   0.23140656 -4.83851424  -5.5104626
#> [5,]  6.2657096 -17.05473509 -7.04087116 -26.2853752
#> 
#> , , 2, 4
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9871157   0.1510725 -0.6251528 -0.6427566
#> [2,] -0.5953566   1.4376976  1.9895721  3.5955897
#> [3,] -0.7664256   0.2649324  3.4280778  4.2731111
#> [4,]  0.6605776  -0.1650494 -2.5152888 -3.1131839
#> [5,]  8.5980158 -15.3559360 20.0301497 -1.7230947
#> 
#> , , 3, 4
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.4289024  -0.8777867  -3.887324  -5.239835
#> [2,] -0.2031791   5.4444554  -5.392542  -2.449287
#> [3,] -0.2915534   7.9614705  -7.636085  -3.267482
#> [4,]  0.2036748  -6.2337543   6.669366   3.419903
#> [5,] 18.7802895 -78.1892975 209.622566 177.862250
#> 
#> , , 4, 4
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,]  1.214141 -0.1246835   2.497895  2.419087
#> [2,] -1.405887  1.8768968  -8.948952 -7.773100
#> [3,] -1.829754  1.6563730 -10.490582 -9.560965
#> [4,]  1.703335 -1.6124041  10.382550  9.496140
#> [5,]  4.131545  0.8947960  19.756667 19.287365
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.6324871  -3.084416   1.159065 -0.828074
#> [2,]  0.6289834  -2.265936   3.879989  2.695991
#> [3,]  1.0501752  -4.838602   6.919291  4.277333
#> [4,] -0.9595624   4.543299  -5.920100 -3.392209
#> [5,]  4.5224470 104.324303 -34.184713 28.151397
#> 
#> , , 1, 5
#> 
#>            [,1]       [,2]         [,3]          [,4]
#> [1,]  0.9522642  0.1770790  -0.09125172   0.008262259
#> [2,] -0.2247011  0.6715038   2.03970653   2.719489038
#> [3,] -0.5462980 -0.6939181   6.38216483   7.026897081
#> [4,]  0.4947780  0.5801095  -4.91874594  -5.392550162
#> [5,]  3.3050605  1.3066515 -15.88818456 -25.423203426
#> 
#> , , 2, 5
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.9819213 -0.1150641 -0.8447847 -1.134314
#> [2,] -0.5909728  1.1544736  1.7297671  3.036099
#> [3,] -0.7541999  0.2805735  3.1828032  3.978170
#> [4,]  0.6601384 -0.2052592 -2.3677926 -2.961599
#> [5,]  7.8942603 -2.0368546 34.0926905 26.855664
#> 
#> , , 3, 5
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.5776031 -3.787185 -0.1822904 -3.1347230
#> [2,] -0.1640935  4.628507 -3.3394918 -0.6615546
#> [3,] -0.2035787  6.457345 -4.1930304 -0.3743366
#> [4,]  0.1559125 -5.214935  4.0864452  1.1566031
#> [5,] 13.1714003 16.624005 73.2293534 91.2045495
#> 
#> , , 4, 5
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9141649  -0.1155163  0.9922254  0.9291438
#> [2,] -0.5362622   3.5509821 -4.4357347 -1.8871572
#> [3,] -0.7994902   3.7723770 -5.1718385 -2.5024475
#> [4,]  0.7606672  -3.5632548  5.5809767  3.0840651
#> [5,]  5.0886435 -12.3030433 18.7034085  7.6869207
#> 
#> , , 5, 5
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  0.6447673 -3.574139  1.6549380 -0.5464694
#> [2,]  0.5844216  1.782771 -0.9073542  0.1720491
#> [3,]  0.9717333  1.909274 -0.9722594  0.1728653
#> [4,] -0.8851301 -1.350673  1.0410202  0.2801143
#> [5,]  4.3897196 66.167434  9.0374869 47.2797816
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_last_draw`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|