R6 Class Representing PosteriorBVARPANEL

The class PosteriorBVARPANEL contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvarPANEL$new()

specify_posterior_bvarPANEL$get_posterior()

specify_posterior_bvarPANEL$get_last_draw()

specify_posterior_bvarPANEL$clone()

Method new()

Create a new posterior output PosteriorBVARPANEL.

Usage

specify_posterior_bvarPANEL$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARPANEL.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvarPANEL$get_posterior()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvarPANEL$get_last_draw()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvarPANEL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARPANEL" "R6"                

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_posterior`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $A
#> , , 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.2034904 -0.3636442 -1.3276322 -0.1291760
#> [2,]  0.4161627  1.0186568  0.2154541 -0.8013564
#> [3,] -1.5823294 -0.0597053 -0.2351007  1.1949381
#> [4,] -0.4877118  0.2672443 -0.4062023  0.1804745
#> [5,]  1.4666676  1.0803295  2.1306699 -1.9431801
#> 
#> , , 2
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.97240077 -0.06243839  0.26990913 0.04249356
#> [2,] -0.01994121  0.61855886 -0.08366136 0.14242448
#> [3,]  0.31723450  0.03717878  0.51345228 0.24216079
#> [4,] -0.28032967  0.06390188  0.40499767 0.73550826
#> [5,]  2.90998756  2.51610173  4.37392895 1.82524798
#> 
#> , , 3
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.91059331 0.008481324  0.49875901  0.51526281
#> [2,]  0.00481271 0.595079614 -0.35583407  0.13596591
#> [3,]  0.13176097 0.403206453 -0.01591455  0.33136172
#> [4,] -0.06855197 0.373559704 -0.24941268 -0.08194199
#> [5,]  3.37590918 4.583365439  4.77540950  1.79286945
#> 
#> , , 4
#> 
#>              [,1]       [,2]        [,3]       [,4]
#> [1,]  0.962061362 -0.4350133  0.74097711 0.49721799
#> [2,] -0.028309846  0.5537590 -0.06258337 0.35407609
#> [3,]  0.021945033  0.1865437  0.45006098 0.55892621
#> [4,]  0.009720675  0.1225395  0.12626124 0.07445326
#> [5,]  2.872028204  2.2829038  7.03248397 3.14225803
#> 
#> , , 5
#> 
#>              [,1]        [,2]        [,3]      [,4]
#> [1,]  0.903947973 -0.06172217  0.30384588 0.2589866
#> [2,] -0.015541074  0.32212577 -0.03692625 0.1962698
#> [3,] -0.005865684 -0.34925601  1.47343857 1.3763533
#> [4,]  0.046782629  0.24702157  0.08557949 0.3716073
#> [5,]  2.559368953 -0.15260180  9.18978060 4.2906713
#> 
#> 
#> $V
#> , , 1
#> 
#>            [,1]       [,2]        [,3]        [,4]      [,5]
#> [1,]  3.2510768  0.1444904  0.47709327 -0.28928456 -2.162021
#> [2,]  0.1444904  5.9369192 -1.39666140  1.14679046 -3.080150
#> [3,]  0.4770933 -1.3966614  8.01156923  0.01768735  1.453489
#> [4,] -0.2892846  1.1467905  0.01768735  4.45221098 -4.666281
#> [5,] -2.1620214 -3.0801503  1.45348899 -4.66628134 33.077509
#> 
#> , , 2
#> 
#>              [,1]        [,2]         [,3]        [,4]       [,5]
#> [1,]  0.917100161  0.06766853  0.004866449 -0.05455378 -2.8560928
#> [2,]  0.067668532  0.77552458  0.224662691  0.09204718 -0.3863873
#> [3,]  0.004866449  0.22466269  1.064135302 -0.35188041  0.4772860
#> [4,] -0.054553781  0.09204718 -0.351880413  1.09109547 -0.0464334
#> [5,] -2.856092777 -0.38638734  0.477285968 -0.04643340 47.0013188
#> 
#> , , 3
#> 
#>            [,1]       [,2]        [,3]       [,4]       [,5]
#> [1,]  2.4470295 -0.8398546   1.1475214 -0.8584959  -3.835931
#> [2,] -0.8398546  4.4800707  -0.5196993  0.7071848  -0.546993
#> [3,]  1.1475214 -0.5196993   8.7478339 -0.6081000 -10.752843
#> [4,] -0.8584959  0.7071848  -0.6081000  2.9352262   4.967695
#> [5,] -3.8359310 -0.5469930 -10.7528434  4.9676951 148.157820
#> 
#> , , 4
#> 
#>             [,1]        [,2]       [,3]       [,4]        [,5]
#> [1,]  1.93778229  0.03237579 -0.9091055  0.8685025   0.8106837
#> [2,]  0.03237579  2.07205915  1.1313925 -2.0186450  -6.2364977
#> [3,] -0.90910555  1.13139254  3.0191195 -2.1513420  -7.4536193
#> [4,]  0.86850246 -2.01864499 -2.1513420  6.0892158   9.2021129
#> [5,]  0.81068373 -6.23649767 -7.4536193  9.2021129 114.0582086
#> 
#> , , 5
#> 
#>            [,1]       [,2]       [,3]       [,4]        [,5]
#> [1,]  1.9296873 -0.4774967 -0.8563882 -0.1513412  -4.8699718
#> [2,] -0.4774967  2.7248803  1.0351416 -1.7570645   4.8795934
#> [3,] -0.8563882  1.0351416  4.1261034 -1.4229633  -0.6322128
#> [4,] -0.1513412 -1.7570645 -1.4229633  4.5148720   1.0880086
#> [5,] -4.8699718  4.8795934 -0.6322128  1.0880086 167.5197032
#> 
#> 
#> $Sigma
#> , , 1
#> 
#>           [,1]       [,2]       [,3]       [,4]
#> [1,] 14.744219 -3.3314116 -2.9452051 -7.2688656
#> [2,] -3.331412 10.0563773  0.3739866 -1.6173933
#> [3,] -2.945205  0.3739866 11.6391123 -0.6964412
#> [4,] -7.268866 -1.6173933 -0.6964412  8.8621277
#> 
#> , , 2
#> 
#>           [,1]      [,2]      [,3]      [,4]
#> [1,]  3.756526 -1.397332 -2.166893 -1.851918
#> [2,] -1.397332 19.137152 -9.993484 -1.569521
#> [3,] -2.166893 -9.993484 21.349327 11.025629
#> [4,] -1.851918 -1.569521 11.025629 11.887388
#> 
#> , , 3
#> 
#>             [,1]        [,2]       [,3]      [,4]
#> [1,]  0.43146426 -0.01973863 -0.1470088 0.2462473
#> [2,] -0.01973863 10.15845396 -3.9694129 1.3359643
#> [3,] -0.14700876 -3.96941294  8.5100721 5.6759517
#> [4,]  0.24624729  1.33596425  5.6759517 7.8077995
#> 
#> , , 4
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.18356157 -0.2417834  0.03130135 -0.06559191
#> [2,] -0.24178344  8.4008185 -5.45674920 -0.27522823
#> [3,]  0.03130135 -5.4567492  9.20062414  5.92022780
#> [4,] -0.06559191 -0.2752282  5.92022780  6.41241379
#> 
#> , , 5
#> 
#>            [,1]        [,2]       [,3]     [,4]
#> [1,] 0.19276081  0.06184696  0.2619114 0.251839
#> [2,] 0.06184696  5.51582538 -2.3645383 1.017279
#> [3,] 0.26191141 -2.36453825  6.7501642 5.646665
#> [4,] 0.25183901  1.01727873  5.6466651 6.874466
#> 
#> 
#> $nu
#>          [,1]
#> [1,] 5.100155
#> [2,] 5.100347
#> [3,] 5.100166
#> [4,] 5.100063
#> [5,] 5.099816
#> 
#> $m
#>            [,1]
#> [1,]  2.0535437
#> [2,]  1.7085872
#> [3,] -0.3026712
#> [4,]  2.3707093
#> [5,]  1.0872878
#> 
#> $w
#>          [,1]
#> [1,] 72.09815
#> [2,] 17.46966
#> [3,] 63.29332
#> [4,] 22.02691
#> [5,] 38.78871
#> 
#> $s
#>           [,1]
#> [1,] 0.5130851
#> [2,] 1.4373191
#> [3,] 1.2941985
#> [4,] 0.9570138
#> [5,] 0.6405089
#> 
#> $scale
#>            [,1]
#> [1,] 0.01427928
#> [2,] 0.01427928
#> [3,] 0.01671019
#> [4,] 0.01897676
#> [5,] 0.02116359
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.04601436 -0.01362547 -0.02358744 -0.0467826
#> [2,] -0.01362547  0.28484533 -0.25964664 -0.1277259
#> [3,] -0.02358744 -0.25964664  0.54631597  0.4069350
#> [4,] -0.04678260 -0.12772587  0.40693502  0.4459990
#> 
#> , , 2, 1
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.089696422 -0.006719562 -0.01944627 -0.02458240
#> [2,] -0.006719562  0.225871228 -0.13969883 -0.03946108
#> [3,] -0.019446275 -0.139698831  0.36261630  0.30873817
#> [4,] -0.024582397 -0.039461081  0.30873817  0.41264736
#> 
#> , , 3, 1
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.04995986 -0.2227662  0.07986127 -0.07842888
#> [2,] -0.22276618 12.3304224 -5.96973737  2.36169068
#> [3,]  0.07986127 -5.9697374  4.48035605  0.77227968
#> [4,] -0.07842888  2.3616907  0.77227968  2.93438321
#> 
#> , , 4, 1
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.059437398 -0.008766044  0.02412969 -0.01619653
#> [2,] -0.008766044  0.157756738 -0.09676656 -0.03997548
#> [3,]  0.024129694 -0.096766562  0.37563786  0.23706104
#> [4,] -0.016196535 -0.039975480  0.23706104  0.25660273
#> 
#> , , 5, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.06254058 -0.2106280  0.1508798  0.03612088
#> [2,] -0.21062804  3.7879311 -3.0897258 -0.86243286
#> [3,]  0.15087977 -3.0897258  3.3597621  1.53211390
#> [4,]  0.03612088 -0.8624329  1.5321139  1.11072591
#> 
#> , , 1, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.03263450 -0.00482875 -0.04561791 -0.04285006
#> [2,] -0.00482875  0.27444238 -0.24461688 -0.15615794
#> [3,] -0.04561791 -0.24461688  0.55824101  0.49365191
#> [4,] -0.04285006 -0.15615794  0.49365191  0.48495707
#> 
#> , , 2, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.02068476 -0.01210096 -0.02558664 -0.03301700
#> [2,] -0.01210096  0.12979011 -0.12426544 -0.05509263
#> [3,] -0.02558664 -0.12426544  0.65994149  0.70021938
#> [4,] -0.03301700 -0.05509263  0.70021938  0.83558435
#> 
#> , , 3, 2
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.01719025 -0.11757646  0.0679974 -0.00777212
#> [2,] -0.11757646  6.85814910 -4.4770924 -0.02305591
#> [3,]  0.06799740 -4.47709238  4.5451055  1.95232263
#> [4,] -0.00777212 -0.02305591  1.9523226  2.36826186
#> 
#> , , 4, 2
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.01111363 -0.00403906  0.01316232 0.01118495
#> [2,] -0.00403906  0.20377698 -0.10979508 0.02376981
#> [3,]  0.01316232 -0.10979508  0.30644190 0.21933842
#> [4,]  0.01118495  0.02376981  0.21933842 0.24047671
#> 
#> , , 5, 2
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.02031983 -0.0932504  0.08806615  0.03162838
#> [2,] -0.09325040  4.6102663 -3.11404345 -0.24252093
#> [3,]  0.08806615 -3.1140435  3.09271434  1.25701536
#> [4,]  0.03162838 -0.2425209  1.25701536  1.26535668
#> 
#> , , 1, 3
#> 
#>             [,1]         [,2]         [,3]         [,4]
#> [1,] 0.009510436  0.001434693  0.005445435  0.007969518
#> [2,] 0.001434693  0.276090874 -0.220756668 -0.109272886
#> [3,] 0.005445435 -0.220756668  0.359548438  0.288985123
#> [4,] 0.007969518 -0.109272886  0.288985123  0.275769991
#> 
#> , , 2, 3
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.010265816 -0.02857242  0.01510303 -0.004248498
#> [2,] -0.028572425  0.23864642 -0.17748369 -0.015791859
#> [3,]  0.015103029 -0.17748369  0.34651761  0.260613949
#> [4,] -0.004248498 -0.01579186  0.26061395  0.302809074
#> 
#> , , 3, 3
#> 
#>              [,1]       [,2]        [,3]       [,4]
#> [1,]  0.003465392 -0.1030502  0.08242944  0.0183616
#> [2,] -0.103050209 11.3769062 -8.43932133 -1.1044186
#> [3,]  0.082429440 -8.4393213  7.61851419  2.3683022
#> [4,]  0.018361598 -1.1044186  2.36830215  1.9216300
#> 
#> , , 4, 3
#> 
#>              [,1]         [,2]        [,3]       [,4]
#> [1,]  0.003390385 -0.001359601  0.01206090 0.01098612
#> [2,] -0.001359601  0.129601197 -0.08032834 0.01604685
#> [3,]  0.012060897 -0.080328344  0.28228125 0.21765569
#> [4,]  0.010986116  0.016046854  0.21765569 0.23537921
#> 
#> , , 5, 3
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.004058439 -0.08919944  0.07669436  0.02656527
#> [2,] -0.089199442  3.32564691 -2.36059888 -0.38400986
#> [3,]  0.076694361 -2.36059888  2.88543953  1.59592832
#> [4,]  0.026565268 -0.38400986  1.59592832  1.50074351
#> 
#> , , 1, 4
#> 
#>              [,1]        [,2]         [,3]         [,4]
#> [1,]  0.008798027 -0.00599077 -0.001358539 -0.003680507
#> [2,] -0.005990770  0.26566303 -0.265403230 -0.159799838
#> [3,] -0.001358539 -0.26540323  0.446914587  0.360237349
#> [4,] -0.003680507 -0.15979984  0.360237349  0.321874779
#> 
#> , , 2, 4
#> 
#>               [,1]         [,2]         [,3]          [,4]
#> [1,]  0.0060897304 -0.008365036  0.005724428 -0.0001079878
#> [2,] -0.0083650358  0.098461816 -0.089925923 -0.0224680357
#> [3,]  0.0057244285 -0.089925923  0.303722582  0.2794797449
#> [4,] -0.0001079878 -0.022468036  0.279479745  0.3107444273
#> 
#> , , 3, 4
#> 
#>              [,1]       [,2]        [,3]        [,4]
#> [1,]  0.003833605 -0.1002128  0.04913875 -0.01664896
#> [2,] -0.100212752  6.9676383 -3.72471292  0.92952891
#> [3,]  0.049138750 -3.7247129  3.80090622  1.63212112
#> [4,] -0.016648957  0.9295289  1.63212112  2.65361449
#> 
#> , , 4, 4
#> 
#>              [,1]          [,2]       [,3]         [,4]
#> [1,]  0.001968683 -1.009164e-02  0.0112226 3.184031e-03
#> [2,] -0.010091642  2.213782e-01 -0.1653802 5.302797e-05
#> [3,]  0.011222596 -1.653802e-01  0.2607666 1.368144e-01
#> [4,]  0.003184031  5.302797e-05  0.1368144 1.408136e-01
#> 
#> , , 5, 4
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.002468202 -0.04062951  0.03917285  0.01606428
#> [2,] -0.040629508  2.52129458 -2.19241179 -0.71862512
#> [3,]  0.039172845 -2.19241179  2.76528517  1.59722156
#> [4,]  0.016064281 -0.71862512  1.59722156  1.30903833
#> 
#> , , 1, 5
#> 
#>              [,1]         [,2]        [,3]         [,4]
#> [1,]  0.008333861 -0.009061566  0.01088983  0.008559059
#> [2,] -0.009061566  0.201934435 -0.16953110 -0.085934014
#> [3,]  0.010889833 -0.169531099  0.36926303  0.334021343
#> [4,]  0.008559059 -0.085934014  0.33402134  0.340922492
#> 
#> , , 2, 5
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.005424488 -0.01473134  0.02602894  0.01712199
#> [2,] -0.014731336  0.17142450 -0.19777692 -0.08302800
#> [3,]  0.026028945 -0.19777692  0.48690124  0.39294424
#> [4,]  0.017121992 -0.08302800  0.39294424  0.38408048
#> 
#> , , 3, 5
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.003254468 -0.08905871  0.03772256 -0.02397323
#> [2,] -0.089058710  7.44677015 -4.14443363  0.76826496
#> [3,]  0.037722556 -4.14443363  4.31805833  1.95113926
#> [4,] -0.023973230  0.76826496  1.95113926  2.92256504
#> 
#> , , 4, 5
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.001396993 -0.004300236  0.006298288  0.002709943
#> [2,] -0.004300236  0.250531261 -0.257808598 -0.068323346
#> [3,]  0.006298288 -0.257808598  0.477991911  0.285954463
#> [4,]  0.002709943 -0.068323346  0.285954463  0.239317713
#> 
#> , , 5, 5
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.002909746 -0.07705324  0.07271013  0.02968707
#> [2,] -0.077053237  3.76149006 -2.89273154 -0.70805552
#> [3,]  0.072710126 -2.89273154  3.21073316  1.64503553
#> [4,]  0.029687070 -0.70805552  1.64503553  1.36376063
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.9580679  0.28057090 -0.0452111 -0.01095476
#> [2,]  0.1518311  0.91572267 -0.3345792 -0.01297067
#> [3,]  0.5448117 -0.09854835  0.6161216  0.59465979
#> [4,] -0.5392599  0.04780529  0.2757370  0.36223178
#> [5,]  1.6353907 -3.57622932  6.9283615  4.23794743
#> 
#> , , 2, 1
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.83031086 -0.03758100  0.1945060  0.0707661
#> [2,]  0.02551163  1.21241463  0.4566902  1.0621893
#> [3,]  0.29690483  0.47924606  1.1044083  1.1001687
#> [4,] -0.19632109 -0.43416027 -0.3514964 -0.1954388
#> [5,] -0.15571551 -0.03063727  7.8833326  1.8206845
#> 
#> , , 3, 1
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.97115597 -0.6072133  0.5097250 -0.2244584
#> [2,]  0.03956836 -0.1266906  0.5035923  0.3678780
#> [3,]  0.06375945 -0.8505911  1.2080317  0.6508392
#> [4,] -0.06512501  1.0968826 -0.5389618  0.3177321
#> [5,]  0.74272330  8.7642892  1.9153033  6.9881429
#> 
#> , , 4, 1
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  1.00645070  0.1113116  0.76437886  0.74262742
#> [2,]  0.05171946  0.2852093 -0.56162126 -0.19362323
#> [3,]  0.02244485 -0.5800368  0.55383148  0.06399127
#> [4,] -0.04906126  0.5433712  0.19257354  0.71757441
#> [5,]  1.84223046  0.5653636  0.02303119 -2.06619764
#> 
#> , , 5, 1
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.83148204 -1.4580391   1.4372369  0.6874358
#> [2,]  0.04244915  0.8824944  -0.1704669  0.3941271
#> [3,]  0.12221031 -0.1742083   0.7792625  0.7036693
#> [4,] -0.07660470  0.4439709  -0.1956894  0.1166258
#> [5,]  2.09893019 22.6963325 -12.8239024 -7.1733671
#> 
#> , , 1, 2
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  1.0571874  0.2774907 -0.4046937 -0.30807556
#> [2,]  0.1661624  0.8400148 -0.4596944 -0.08314934
#> [3,]  0.5852394 -0.1816094  0.6412472  0.65756866
#> [4,] -0.5633195  0.1269417  0.1797696  0.26707931
#> [5,] -1.4018049 -3.0711969 20.0191692 13.89304760
#> 
#> , , 2, 2
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.96130254 -0.1115326 -0.04274123 -0.2955799
#> [2,] -0.14729473  0.4931468  1.53100054  2.1929535
#> [3,] -0.09736406 -0.2780840  1.90095224  1.8567907
#> [4,]  0.10618313  0.2901432 -1.12850803 -0.9752537
#> [5,]  1.52476596  6.5943872  4.63477551  3.5845168
#> 
#> , , 3, 2
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.87951640 -1.4948613   1.6265335   0.5397871
#> [2,]  0.02848095  0.4874598   0.5273105   0.9587147
#> [3,]  0.04977133  0.1026818   1.1394360   1.3979722
#> [4,] -0.05007948  0.2371757  -0.4459150  -0.3216901
#> [5,]  2.86675076 23.0480577 -26.1573476 -18.9453946
#> 
#> , , 4, 2
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.99675115  0.29409597  0.5143204  0.61179926
#> [2,]  0.08621677  0.21517838 -0.5128763 -0.36645646
#> [3,]  0.03450826 -0.17125508  0.1462888 -0.04897872
#> [4,] -0.05677151  0.07700472  0.6499365  0.86391320
#> [5,]  1.70763902  1.10070917  1.8764240 -0.99727514
#> 
#> , , 5, 2
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.90058568 -0.7947616  0.55606397 0.02003319
#> [2,]  0.03442381  0.7522737 -0.02632832 0.47773694
#> [3,]  0.07329479 -0.1375614  0.69863462 0.68706315
#> [4,] -0.08715732  0.3501436 -0.08024777 0.15293939
#> [5,]  3.63498540 10.0588933  6.64828051 8.50719020
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]        [,3]         [,4]
#> [1,]  0.8622575  0.1448952 -0.12536305 -0.158244173
#> [2,]  0.2068227  0.8823095 -0.43052079  0.003518036
#> [3,]  0.4467286 -0.2315866  0.87390172  0.844097830
#> [4,] -0.3815999  0.1110927  0.09071847  0.217188921
#> [5,]  0.2224166  2.2958285  7.73508714  4.112188647
#> 
#> , , 2, 3
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.9613993 -0.09588046 -0.04274645 -0.2553915
#> [2,] -0.1870306  0.48374221  1.70809295  2.4197044
#> [3,] -0.2329973 -0.40809002  2.94333188  3.0546627
#> [4,]  0.2157228  0.41304577 -2.00200857 -1.9668531
#> [5,]  2.4415673  5.39729643  1.97450126 -1.6599953
#> 
#> , , 3, 3
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.91026999 -2.72553975   1.5361778 -0.4212170
#> [2,]  0.02521452  0.40601964   0.5301761  0.8658445
#> [3,]  0.03710959  0.09432378   1.2509274  1.4579021
#> [4,] -0.03850287 -0.05147016  -0.3237831 -0.3539012
#> [5,]  2.12202864 70.09468959 -36.6190425  3.8177564
#> 
#> , , 4, 3
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.98481396 -0.1510133  1.2643786  1.3314740
#> [2,]  0.05759247  0.3001644 -0.5422687 -0.3333646
#> [3,]  0.02180021 -0.2629535  0.3498124  0.1596028
#> [4,] -0.03998064  0.3057535  0.2120931  0.4216258
#> [5,]  1.72894731  1.8319986  0.7051873 -1.6034927
#> 
#> , , 5, 3
#> 
#>               [,1]       [,2]        [,3]       [,4]
#> [1,]  0.9214935909 -2.1499847  2.00830066  0.7493225
#> [2,]  0.0001165229  1.5343864 -0.38994605  0.6005523
#> [3,]  0.0053938243  1.3473198 -0.02049089  0.8883734
#> [4,] -0.0246049383 -0.3930492  0.03420738 -0.2351745
#> [5,]  3.3380477138  2.6855135  3.88147387  0.6485486
#> 
#> , , 1, 4
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.9246569 -0.1611689  0.3604128  0.27494374
#> [2,]  0.2155235  0.8379468 -0.3515683  0.05978821
#> [3,]  0.4668620 -0.2476713  1.0904330  1.09552132
#> [4,] -0.4130546  0.1654743 -0.2442050 -0.15254075
#> [5,] -0.7322852  8.1669996  1.8550058 -0.13782271
#> 
#> , , 2, 4
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  1.0116899 -0.1456227 -0.1865777 -0.414224
#> [2,] -0.2187641  0.6690345  1.6079430  2.373108
#> [3,] -0.2516186 -0.2702789  2.6409124  2.723839
#> [4,]  0.2282867  0.3302682 -1.7946596 -1.732769
#> [5,]  1.9786346  0.9154393 10.7173797  6.202080
#> 
#> , , 3, 4
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.91216372 -2.39672645  0.5786130 -1.2842879
#> [2,]  0.03088447  0.54771383  0.2817916  0.6986244
#> [3,]  0.04203933  0.31913399  1.0206535  1.4065567
#> [4,] -0.04244800 -0.01109648 -0.2537496 -0.2654191
#> [5,]  1.98464194 46.59474334 -3.7699705 23.0318815
#> 
#> , , 4, 4
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,] 0.900104080  0.10860889  0.4466309  0.64929145
#> [2,] 0.047017305  0.26933179 -0.5042552 -0.32922648
#> [3,] 0.008348097 -0.15602324  0.0832420 -0.01083593
#> [4,] 0.004012646  0.06296757  0.7938368  0.82943357
#> [5,] 1.611224673  5.68999028 -2.6971365 -2.31840431
#> 
#> , , 5, 4
#> 
#>              [,1]       [,2]        [,3]       [,4]
#> [1,]  0.917556876 -1.2510035  1.33155458  0.6231978
#> [2,]  0.004781189  1.3441066 -0.29432800  0.5849692
#> [3,]  0.009282555  0.9400326  0.13789396  0.7825603
#> [4,] -0.026264438 -0.2546854 -0.04713352 -0.2238717
#> [5,]  3.271126906 -5.5669147 17.05251784  8.9855592
#> 
#> , , 1, 5
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  0.9118135  0.31607650 -0.20792288 -0.08585134
#> [2,]  0.1932262  0.94105629 -0.35665406  0.10076109
#> [3,]  0.4593944  0.04842782  0.85178949  1.00829537
#> [4,] -0.4092177 -0.10776814  0.01587475 -0.02841713
#> [5,] -0.1047585 -3.62364693 12.92174694  5.59030201
#> 
#> , , 2, 5
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  0.9546882  0.1858191 -0.501836 -0.4981717
#> [2,] -0.1565741  0.5080877  1.631869  2.2644529
#> [3,] -0.1572339 -0.6760912  2.785890  2.5310375
#> [4,]  0.1571301  0.6607144 -1.947199 -1.6130687
#> [5,]  1.7595153 -3.8827000 20.367533 13.1368079
#> 
#> , , 3, 5
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.92363929 -2.55395150   1.6321644  -0.1342660
#> [2,]  0.02872422  0.37524390   0.6825559   1.0636766
#> [3,]  0.03967889  0.13848538   1.2861592   1.6254154
#> [4,] -0.04379223  0.07949749  -0.3916789  -0.3855257
#> [5,]  1.95765671 55.64413728 -38.4411813 -12.5520572
#> 
#> , , 4, 5
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.92309021  0.2473363  0.6220173  0.93884805
#> [2,]  0.01279712  0.5104628 -0.7985272 -0.45109521
#> [3,]  0.02300422 -0.2910119  0.2431381  0.03921401
#> [4,] -0.01557544  0.2517235  0.5030222  0.69342930
#> [5,]  1.49585183 -2.8442652  3.5797954 -3.07198670
#> 
#> , , 5, 5
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.92639709 -1.4297784  0.37666015 -0.5684598
#> [2,]  0.01482583  1.1155209 -0.09956881  0.6494178
#> [3,]  0.02170908  0.7396619  0.44221883  0.9902520
#> [4,] -0.02929586 -0.1752039 -0.08786042 -0.2162384
#> [5,]  2.42688222  7.8746239 26.38729002 28.4700705
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_last_draw`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|