R6 Class Representing PosteriorBVARPANEL

The class PosteriorBVARPANEL contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvarPANEL$new()

specify_posterior_bvarPANEL$get_posterior()

specify_posterior_bvarPANEL$get_last_draw()

specify_posterior_bvarPANEL$clone()

Method new()

Create a new posterior output PosteriorBVARPANEL.

Usage

specify_posterior_bvarPANEL$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARPANEL with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARPANEL.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvarPANEL$get_posterior()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARPANEL with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvarPANEL$get_last_draw()

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvarPANEL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARPANEL" "R6"                

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_posterior`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $A
#> , , 1
#> 
#>             [,1]       [,2]      [,3]       [,4]
#> [1,] -0.20658328  0.8910084 0.1376799  0.4677773
#> [2,]  0.49238474 -0.4939147 0.1136720  0.2776548
#> [3,] -0.04546558  0.2636655 0.3752455  0.5613779
#> [4,]  0.46124787  0.1157365 0.2211876  0.4353054
#> [5,] -0.16882918  1.7554577 0.7227908 -1.0486862
#> 
#> , , 2
#> 
#>             [,1]      [,2]        [,3]        [,4]
#> [1,]  0.80674887 0.4293576  0.03484794  0.16866625
#> [2,] -0.09388351 0.3826840 -0.17411994  0.13532217
#> [3,]  0.09200641 0.3428010  0.14810764  0.36851826
#> [4,]  0.02062104 0.5515339 -0.36408130 -0.02213668
#> [5,]  1.42944175 3.6228219  0.62700394 -0.73406824
#> 
#> , , 3
#> 
#>              [,1]       [,2]       [,3]       [,4]
#> [1,]  0.982195043  0.2479871  0.3375151  0.4597004
#> [2,]  0.071114892  0.6659621 -0.8266044 -0.4743624
#> [3,] -0.190576427 -0.6970126  0.8802147  0.2849453
#> [4,]  0.005071187  0.9396524  0.3795320  1.1619523
#> [5,]  1.321962725  0.9530918  5.2619213  2.4132175
#> 
#> , , 4
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.91554725  0.1483192  0.2815387  0.4147093
#> [2,] -0.01345927  0.5781822  0.4076070  0.7579882
#> [3,]  0.04442234 -0.3261601  1.3184260  1.1272762
#> [4,] -0.06143329  0.3554920 -0.4530154 -0.1989294
#> [5,]  0.90806020  1.3177839  6.3127069  3.7265040
#> 
#> , , 5
#> 
#>             [,1]       [,2]      [,3]      [,4]
#> [1,]  0.96702160  0.2076452 0.1166592 0.2597829
#> [2,] -0.05182942 -0.3408289 0.9177419 0.6891494
#> [3,] -0.07030498 -0.5900565 2.0762776 1.8988960
#> [4,] -0.06748663  0.1143440 0.1964424 0.1517786
#> [5,]  1.02944221  0.9803078 6.8353169 3.3620496
#> 
#> 
#> $V
#> , , 1
#> 
#>            [,1]        [,2]       [,3]        [,4]        [,5]
#> [1,]  1.6058619 -0.12740898 0.18681021 -0.15490115  2.04792582
#> [2,] -0.1274090  1.05075915 0.04713884  0.39049618  0.22739980
#> [3,]  0.1868102  0.04713884 1.01655847  0.04015443  0.02982256
#> [4,] -0.1549012  0.39049618 0.04015443  1.01412135 -0.18067073
#> [5,]  2.0479258  0.22739980 0.02982256 -0.18067073 17.20832248
#> 
#> , , 2
#> 
#>             [,1]       [,2]       [,3]        [,4]        [,5]
#> [1,]  2.11636173 -0.6848408 -0.7433430 -0.07661119  2.15087468
#> [2,] -0.68484076  1.9242860  0.4984614 -0.27061625 -1.20897817
#> [3,] -0.74334295  0.4984614  1.6011592  0.18239823 -0.79472739
#> [4,] -0.07661119 -0.2706163  0.1823982  0.94147524  0.03647911
#> [5,]  2.15087468 -1.2089782 -0.7947274  0.03647911 18.40908080
#> 
#> , , 3
#> 
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,]  4.332920 -1.666858 -1.730723  4.129424  5.401521
#> [2,] -1.666858  4.031313  1.414321 -3.392386 -6.180814
#> [3,] -1.730723  1.414321  3.703982 -3.148281 -4.324016
#> [4,]  4.129424 -3.392386 -3.148281  7.211267  7.447126
#> [5,]  5.401521 -6.180814 -4.324016  7.447126 49.897755
#> 
#> , , 4
#> 
#>             [,1]        [,2]       [,3]       [,4]      [,5]
#> [1,]  1.81123556  0.09394319 -0.1880779  0.5047493  6.783412
#> [2,]  0.09394319  1.79571268  1.5943372 -0.9601863  0.842933
#> [3,] -0.18807789  1.59433723  2.8826054 -1.8042592 -3.115502
#> [4,]  0.50474930 -0.96018629 -1.8042592  2.7928953  4.724211
#> [5,]  6.78341195  0.84293299 -3.1155018  4.7242109 59.523948
#> 
#> , , 5
#> 
#>            [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,]  1.8196045 -1.172010 -0.9642646 -0.5198123  1.017630
#> [2,] -1.1720102  4.130920  1.2589661  1.6550403  1.372249
#> [3,] -0.9642646  1.258966  3.8548739 -0.4544678 -2.430114
#> [4,] -0.5198123  1.655040 -0.4544678  5.2723825  4.561404
#> [5,]  1.0176297  1.372249 -2.4301144  4.5614043 57.207451
#> 
#> 
#> $Sigma
#> , , 1
#> 
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 15.609723  1.135515 -2.233064 -2.134976
#> [2,]  1.135515  9.073750 -3.054123 -2.601093
#> [3,] -2.233064 -3.054123 11.850229  3.428613
#> [4,] -2.134976 -2.601093  3.428613 11.140185
#> 
#> , , 2
#> 
#>           [,1]       [,2]      [,3]       [,4]
#> [1,]  2.895824 -1.2531329 -1.106843 -1.1569595
#> [2,] -1.253133  7.7282757 -4.121456  0.1858661
#> [3,] -1.106843 -4.1214557 16.478824 14.0248769
#> [4,] -1.156960  0.1858661 14.024877 16.6642659
#> 
#> , , 3
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.6372294   0.202453  -0.2025365  0.2431088
#> [2,]  0.2024530   9.326015 -10.3234715 -3.5294304
#> [3,] -0.2025365 -10.323472  22.7978806 15.7751696
#> [4,]  0.2431088  -3.529430  15.7751696 14.5695474
#> 
#> , , 4
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.14033494 -0.1137443  0.07903523  0.1127814
#> [2,] -0.11374429  8.2598899 -8.68293759 -2.9773214
#> [3,]  0.07903523 -8.6829376 16.82490944 11.2756591
#> [4,]  0.11278142 -2.9773214 11.27565910 10.1091484
#> 
#> , , 5
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.13840559 -0.2949187  0.04988641 -0.08534794
#> [2,] -0.29491870  6.5387640 -0.48936508  3.93883436
#> [3,]  0.04988641 -0.4893651 10.70544541 11.80819922
#> [4,] -0.08534794  3.9388344 11.80819922 16.33589200
#> 
#> 
#> $nu
#>          [,1]
#> [1,] 5.100064
#> [2,] 5.099809
#> [3,] 5.100114
#> [4,] 5.100001
#> [5,] 5.099898
#> 
#> $m
#>           [,1]
#> [1,] 0.6509184
#> [2,] 0.5403983
#> [3,] 1.0301302
#> [4,] 1.1319582
#> [5,] 0.6717043
#> 
#> $w
#>          [,1]
#> [1,] 30.87963
#> [2,] 27.00237
#> [3,] 49.74787
#> [4,] 19.00692
#> [5,] 46.95727
#> 
#> $s
#>           [,1]
#> [1,] 0.5143587
#> [2,] 1.6637193
#> [3,] 1.8614132
#> [4,] 1.3951575
#> [5,] 1.2899146
#> 
#> $scale
#>            [,1]
#> [1,] 0.01427928
#> [2,] 0.01427928
#> [3,] 0.01575508
#> [4,] 0.01789210
#> [5,] 0.01995393
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.10500245 -0.01106237  0.02122905  0.02585281
#> [2,] -0.01106237  0.54960690 -0.49288170 -0.31723127
#> [3,]  0.02122905 -0.49288170  0.64238234  0.46125464
#> [4,]  0.02585281 -0.31723127  0.46125464  0.43408363
#> 
#> , , 2, 1
#> 
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.100575693 -0.02878779  0.003689797 -0.00339391
#> [2,] -0.028787785  0.27233121 -0.226577340 -0.05412159
#> [3,]  0.003689797 -0.22657734  0.632801970  0.49748263
#> [4,] -0.003393910 -0.05412159  0.497482632  0.51079063
#> 
#> , , 3, 1
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.07168906 -0.07858544 -0.09867804 -0.1523201
#> [2,] -0.07858544  9.42253469 -5.62735563  0.6214703
#> [3,] -0.09867804 -5.62735563  5.04673198  1.4802371
#> [4,] -0.15232013  0.62147029  1.48023711  2.1737542
#> 
#> , , 4, 1
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.046735590 -0.01206629  0.01345995  0.001375622
#> [2,] -0.012066293  0.19073569 -0.15762160 -0.050447769
#> [3,]  0.013459947 -0.15762160  0.28388783  0.159483854
#> [4,]  0.001375622 -0.05044777  0.15948385  0.180887825
#> 
#> , , 5, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.05667578 -0.1359569  0.1006443  0.04309641
#> [2,] -0.13595685  2.4419035 -2.0733335 -0.69515530
#> [3,]  0.10064428 -2.0733335  2.8420878  1.72436612
#> [4,]  0.04309641 -0.6951553  1.7243661  1.46504620
#> 
#> , , 1, 2
#> 
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.017403168  0.005288644 -0.009054461 -0.002417725
#> [2,]  0.005288644  0.255327698 -0.248968639 -0.135409149
#> [3,] -0.009054461 -0.248968639  0.446727302  0.339860318
#> [4,] -0.002417725 -0.135409149  0.339860318  0.307511632
#> 
#> , , 2, 2
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.017759297 -0.03838209  0.03030839  0.005444592
#> [2,] -0.038382091  0.27768570 -0.23822245 -0.048823248
#> [3,]  0.030308393 -0.23822245  0.55971152  0.450906576
#> [4,]  0.005444592 -0.04882325  0.45090658  0.492641672
#> 
#> , , 3, 2
#> 
#>               [,1]          [,2]         [,3]        [,4]
#> [1,]  0.0112182373  5.591906e-04  -0.02519599 -0.02882556
#> [2,]  0.0005591906  2.023131e+01 -15.14718360 -2.15173623
#> [3,] -0.0251959921 -1.514718e+01  13.05469182  3.66243280
#> [4,] -0.0288255595 -2.151736e+00   3.66243280  2.71711926
#> 
#> , , 4, 2
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.007140543 -0.005935004  0.01561584  0.01424880
#> [2,] -0.005935004  0.171086605 -0.13673232 -0.01403848
#> [3,]  0.015615836 -0.136732320  0.31270073  0.21479591
#> [4,]  0.014248796 -0.014038483  0.21479591  0.21781343
#> 
#> , , 5, 2
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.018815738 -0.04143572  0.01102041 -0.007999338
#> [2,] -0.041435716  1.81765759 -1.44466681 -0.399151234
#> [3,]  0.011020412 -1.44466681  2.02519037  1.275292078
#> [4,] -0.007999338 -0.39915123  1.27529208  1.145615262
#> 
#> , , 1, 3
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.006899088 -0.006160137  0.01331609  0.01400139
#> [2,] -0.006160137  0.196669954 -0.21636515 -0.13613198
#> [3,]  0.013316092 -0.216365146  0.38038724  0.30860680
#> [4,]  0.014001394 -0.136131981  0.30860680  0.28191583
#> 
#> , , 2, 3
#> 
#>              [,1]         [,2]        [,3]       [,4]
#> [1,]  0.009484894 -0.006192777  0.01252098 0.01303928
#> [2,] -0.006192777  0.100169496 -0.04221514 0.04719536
#> [3,]  0.012520978 -0.042215135  0.36655301 0.39266799
#> [4,]  0.013039275  0.047195361  0.39266799 0.51783643
#> 
#> , , 3, 3
#> 
#>              [,1]       [,2]         [,3]        [,4]
#> [1,]  0.003274528 -0.0393360  0.006585953 -0.02133267
#> [2,] -0.039335999  9.3874399 -6.886203018 -0.66268589
#> [3,]  0.006585953 -6.8862030  6.506662168  2.20410830
#> [4,] -0.021332670 -0.6626859  2.204108297  2.11592891
#> 
#> , , 4, 3
#> 
#>              [,1]        [,2]         [,3]         [,4]
#> [1,]  0.004044678 -0.01595761  0.008188624 -0.003328559
#> [2,] -0.015957613  0.23569852 -0.125097319  0.054962070
#> [3,]  0.008188624 -0.12509732  0.250232594  0.154734158
#> [4,] -0.003328559  0.05496207  0.154734158  0.200042293
#> 
#> , , 5, 3
#> 
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.007004715 -0.02475975 -0.005338521 -0.02016491
#> [2,] -0.024759753  2.47626123 -2.336120220 -0.91782684
#> [3,] -0.005338521 -2.33612022  3.038776910  1.77287270
#> [4,] -0.020164912 -0.91782684  1.772872701  1.33168798
#> 
#> , , 1, 4
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.003627658 -0.006665584  0.01305776  0.01172835
#> [2,] -0.006665584  0.183643961 -0.21518677 -0.14373871
#> [3,]  0.013057758 -0.215186765  0.50236844  0.44236925
#> [4,]  0.011728347 -0.143738713  0.44236925  0.41861589
#> 
#> , , 2, 4
#> 
#>              [,1]        [,2]         [,3]         [,4]
#> [1,]  0.009331971 -0.01278606  0.003530118 -0.003762851
#> [2,] -0.012786059  0.17943651 -0.163984815 -0.033997633
#> [3,]  0.003530118 -0.16398482  0.411577569  0.339914813
#> [4,] -0.003762851 -0.03399763  0.339914813  0.376958881
#> 
#> , , 3, 4
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.002371757 -0.07273311  0.04597995 -0.002276042
#> [2,] -0.072733109  8.49743929 -5.78173502 -0.207479756
#> [3,]  0.045979947 -5.78173502  5.86373342  2.425604060
#> [4,] -0.002276042 -0.20747976  2.42560406  2.735376944
#> 
#> , , 4, 4
#> 
#>              [,1]         [,2]        [,3]         [,4]
#> [1,]  0.001874263 -0.009369596  0.01329382  0.007560659
#> [2,] -0.009369596  0.181595222 -0.18056954 -0.049209792
#> [3,]  0.013293818 -0.180569538  0.46483581  0.339902253
#> [4,]  0.007560659 -0.049209792  0.33990225  0.313245460
#> 
#> , , 5, 4
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.002854460 -0.02736346  0.01415438 -0.002186353
#> [2,] -0.027363461  1.95296772 -1.45968469 -0.307658036
#> [3,]  0.014154383 -1.45968469  2.53816355  1.852813398
#> [4,] -0.002186353 -0.30765804  1.85281340  1.871835557
#> 
#> , , 1, 5
#> 
#>              [,1]        [,2]         [,3]         [,4]
#> [1,]  0.004610191 -0.01263538  0.004279843 -0.002131768
#> [2,] -0.012635376  0.18705763 -0.156374219 -0.083841129
#> [3,]  0.004279843 -0.15637422  0.311626915  0.279702991
#> [4,] -0.002131768 -0.08384113  0.279702991  0.284875912
#> 
#> , , 2, 5
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.006816484 -0.01046727  0.01239222 0.006305745
#> [2,] -0.010467270  0.14087020 -0.09751484 0.013147169
#> [3,]  0.012392223 -0.09751484  0.34996555 0.319261126
#> [4,]  0.006305745  0.01314717  0.31926113 0.386975291
#> 
#> , , 3, 5
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.002261825 -0.03191111  0.02220927  0.003839823
#> [2,] -0.031911113  7.29661728 -5.02585722 -0.350456886
#> [3,]  0.022209273 -5.02585722  4.82761910  1.903819335
#> [4,]  0.003839823 -0.35045689  1.90381933  2.071859769
#> 
#> , , 4, 5
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.002842037 -0.01294238  0.02071861  0.01177235
#> [2,] -0.012942376  0.27599881 -0.26065020 -0.05597672
#> [3,]  0.020718611 -0.26065020  0.48687345  0.29601765
#> [4,]  0.011772355 -0.05597672  0.29601765  0.26061175
#> 
#> , , 5, 5
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.003639336 -0.06147711  0.07287003  0.04219728
#> [2,] -0.061477110  3.28442329 -2.31864804 -0.37369558
#> [3,]  0.072870030 -2.31864804  2.93229776  1.75732994
#> [4,]  0.042197281 -0.37369558  1.75732994  1.76936409
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>              [,1]       [,2]        [,3]       [,4]
#> [1,]  1.066777168  0.9647585 -0.47538168 -0.1776249
#> [2,] -0.081720964  0.1762090  0.38460551  0.4452108
#> [3,] -0.049070800 -0.3050941  1.21676438  1.0911863
#> [4,] -0.000793863 -0.1472248  0.04978859  0.1010514
#> [5,]  1.285544288  4.7748559 -4.15306669 -4.5767417
#> 
#> , , 2, 1
#> 
#>              [,1]       [,2]        [,3]        [,4]
#> [1,]  1.022720428  0.0272519  0.16765949  0.05338268
#> [2,]  0.038609818  0.3196439  0.46291507  0.80106425
#> [3,] -0.013401468 -0.5403760  0.78856086  0.39974533
#> [4,]  0.004612444  0.5755000  0.03903712  0.53000951
#> [5,] -0.606270023  0.9950725 -1.37038565 -4.23096334
#> 
#> , , 3, 1
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.89759353 -0.27312046  0.3010142 -0.0219845
#> [2,]  0.05166317 -0.09575798  0.4622702  0.2701056
#> [3,]  0.10246516 -0.55066358  0.9288346  0.4292997
#> [4,] -0.09331674  0.86352041 -0.2634772  0.5677321
#> [5,]  1.98694802 -0.11160725  4.9389015  0.4210753
#> 
#> , , 4, 1
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.62248146  0.10387919  0.9382829  0.64391243
#> [2,]  0.09122351  0.38541099 -0.5058367 -0.03106631
#> [3,] -0.05521745 -0.07175140  0.1820996  0.29420781
#> [4,]  0.15532697  0.07520751  0.4191435  0.46449623
#> [5,]  2.04709944 -1.70449165  5.9335406  2.44417904
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.8314051 -2.2482901  2.3852682  1.00560309
#> [2,]  0.1281907  1.3574030 -0.6448014  0.16300987
#> [3,]  0.2606976  0.9656435 -0.5855816 -0.04165242
#> [4,] -0.2134137  0.0734193  0.3867036  0.52124304
#> [5,]  1.9615276 -0.1772622  5.3715553  2.68089129
#> 
#> , , 1, 2
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.89695041  0.2827470 -0.0953742 -0.05581034
#> [2,]  0.02490899  0.2936258  0.1462801  0.29736914
#> [3,]  0.03246385 -1.3485724  1.9095517  1.46117170
#> [4,] -0.01270777  1.1189839 -0.8456348 -0.38736564
#> [5,]  1.46037997  5.0426209  1.2593803  0.83216782
#> 
#> , , 2, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.98794052 -0.03534234  0.08803877 -0.03777217
#> [2,]  0.02082697  0.40325975  0.25789584  0.62352350
#> [3,]  0.04134728 -0.69259939  1.36538725  0.90162803
#> [4,] -0.03682964  0.62371886 -0.35403167  0.14636256
#> [5,]  0.18034560  7.18523116 -2.72325512 -1.74724150
#> 
#> , , 3, 2
#> 
#>              [,1]       [,2]        [,3]       [,4]
#> [1,]  0.972196788 -1.6004206   1.1742849 -0.1301547
#> [2,] -0.003310544  0.1917844   0.4596017  0.5470744
#> [3,] -0.007351402  0.2518101   0.8320283  1.0188929
#> [4,]  0.002495274  0.2538273  -0.1218129  0.2015683
#> [5,]  0.950315282 19.2766159 -16.6646823 -8.8759693
#> 
#> , , 4, 2
#> 
#>            [,1]         [,2]        [,3]       [,4]
#> [1,]  0.8947388  0.031156180  0.08353536 0.13704049
#> [2,] -0.1817096  0.782185994 -0.53368675 0.07507673
#> [3,] -0.1624512 -0.030238649  0.48451803 0.48452076
#> [4,]  0.1819506  0.002423171  0.44214482 0.46727829
#> [5,]  1.4514192  1.836473644  3.79965947 0.90030500
#> 
#> , , 5, 2
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  0.8109953 -2.1549524  2.445264  1.1666022
#> [2,]  0.2184239  1.2138229 -1.196412 -0.4627466
#> [3,]  0.3729991  0.8283619 -1.317108 -0.8380467
#> [4,] -0.3250621  0.2618681  1.073851  1.2912581
#> [5,]  2.2929027 -5.0394292  7.496774  1.4519444
#> 
#> , , 1, 3
#> 
#>              [,1]       [,2]       [,3]       [,4]
#> [1,]  0.951327297  0.3595819 -0.3387643 -0.2919691
#> [2,]  0.003235882  0.3608623  0.1584253  0.3942523
#> [3,]  0.047079376 -1.3671980  1.8137506  1.2977373
#> [4,] -0.047283543  1.1883968 -0.7555492 -0.1668066
#> [5,]  1.341102621  0.3566444  6.5326134  2.0413139
#> 
#> , , 2, 3
#> 
#>             [,1]        [,2]       [,3]      [,4]
#> [1,]  0.95866746 -0.04062278  0.2520913 0.1053814
#> [2,]  0.01080308  0.46538885  0.2308860 0.6500730
#> [3,]  0.05223200 -0.64878849  1.1332590 0.7008257
#> [4,] -0.04575810  0.64379627 -0.2975489 0.2230273
#> [5,]  1.03245291  1.91121483  4.4111975 1.4961660
#> 
#> , , 3, 3
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.94557311 -0.4359988 -0.1995445 -0.7459990
#> [2,]  0.01418385  1.2968188 -0.4915138  0.3142670
#> [3,]  0.01096910  1.4157932 -0.1388282  0.8165660
#> [4,] -0.01213191 -0.9638386  0.7851285  0.2808221
#> [5,]  1.24547559 -8.7930326 22.6115104 13.5316568
#> 
#> , , 4, 3
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.97808720 -0.1538654  0.8019981 0.74424889
#> [2,] -0.03251476  0.9563359 -0.6715322 0.01941741
#> [3,] -0.08541142  0.3126819  0.3288633 0.54077781
#> [4,]  0.07187963 -0.2706949  0.3462730 0.18177567
#> [5,]  1.56545532  1.4316288  4.4077793 2.81123852
#> 
#> , , 5, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8674979 -1.8288667  2.3141235  1.2662715
#> [2,]  0.2633148  1.3749157 -0.9250552 -0.1274405
#> [3,]  0.4455620  1.1407145 -1.0357870 -0.4245726
#> [4,] -0.3959338 -0.2553499  0.8897298  0.8536914
#> [5,]  0.6596320 -1.3574146  3.9961143 -0.8970545
#> 
#> , , 1, 4
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.94106294  0.1295580 -0.1186610 -0.1466776
#> [2,]  0.01195305  0.3762453  0.2218612  0.4414940
#> [3,]  0.03883598 -1.4192585  1.8155940  1.3171174
#> [4,] -0.03653129  1.2624758 -0.7440371 -0.1837231
#> [5,]  1.35969060  4.1316877  0.5603232 -1.5393199
#> 
#> , , 2, 4
#> 
#>              [,1]       [,2]       [,3]       [,4]
#> [1,]  0.972071604 -0.1998388  0.1639050 -0.1361101
#> [2,] -0.002421202  0.4311252  0.3311944  0.7187600
#> [3,]  0.055360063 -0.4736138  1.0035141  0.7158519
#> [4,] -0.055423304  0.5518798 -0.1666775  0.2783261
#> [5,]  1.431554273  2.2235630  3.0477859  0.9037528
#> 
#> , , 3, 4
#> 
#>              [,1]       [,2]       [,3]       [,4]
#> [1,]  0.972248936 -1.7603101  1.1465125 -0.2171495
#> [2,]  0.008058861  1.0008291  0.2516153  0.9756701
#> [3,]  0.007360885  0.7771799  0.9711793  1.6718047
#> [4,] -0.011139236 -0.3217445 -0.3373383 -0.5658850
#> [5,]  0.839197157 21.0979595 -8.5980633 -0.4039726
#> 
#> , , 4, 4
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.99879983 -0.5322488  1.1373333  0.8303755
#> [2,] -0.06192353  1.0168989 -0.2382992  0.5032921
#> [3,] -0.11170174  0.5778473  0.3923839  0.8252946
#> [4,]  0.09716617 -0.4061587  0.1796586 -0.1121226
#> [5,]  1.11951421  1.9585890  2.4082296  0.5990304
#> 
#> , , 5, 4
#> 
#>              [,1]       [,2]       [,3]       [,4]
#> [1,]  0.951914966  -1.775655  1.3072049  0.1806274
#> [2,] -0.005486683   2.682573  0.5314713  2.3928786
#> [3,] -0.006847406   3.027539  1.5055157  3.7000710
#> [4,]  0.011626244  -1.894762 -1.4782577 -2.9045787
#> [5,]  1.039345868 -19.869198 22.0901973  6.3202222
#> 
#> , , 1, 5
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.97089442  0.1591232 -0.09225257 -0.05027004
#> [2,]  0.02262857  0.3535484 -0.01559497  0.17869665
#> [3,]  0.06162631 -1.5050175  1.86144400  1.34114663
#> [4,] -0.04803445  1.2864068 -0.88376536 -0.33228544
#> [5,]  0.15878757  6.1059737  6.46353840  4.27925279
#> 
#> , , 2, 5
#> 
#>              [,1]       [,2]       [,3]       [,4]
#> [1,]  0.934410577  0.1035204 -0.2828186 -0.3463931
#> [2,]  0.002241603  0.4319939  0.1588979  0.5279494
#> [3,]  0.090601938 -0.7807688  1.2598959  0.7499560
#> [4,] -0.062872649  0.7144811 -0.3298550  0.2283801
#> [5,]  0.631435409  1.8807363 12.9366768 10.7645380
#> 
#> , , 3, 5
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.97734471 -1.93433831   1.1960504 -0.1939564
#> [2,]  0.01400155  0.62981949   0.5547326  1.0602190
#> [3,]  0.01367214  0.35600738   1.2425083  1.6725403
#> [4,] -0.01267765  0.06172716  -0.5925701 -0.5826732
#> [5,]  0.41230349 28.21927222 -12.8511250 -1.9821352
#> 
#> , , 4, 5
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.98921314  0.3014602  0.20693442  0.5234251
#> [2,] -0.06686970  0.7824499 -0.02277941  0.5460028
#> [3,] -0.09917945  0.1087482  0.99711782  1.0968939
#> [4,]  0.09367703 -0.1901419 -0.14740523 -0.2824388
#> [5,]  0.72663603 -1.0764176  6.36607025  1.3298927
#> 
#> , , 5, 5
#> 
#>              [,1]       [,2]       [,3]      [,4]
#> [1,]  0.957211269 -1.4405080  1.2545294  0.376042
#> [2,] -0.003751298  1.4187399  0.7550127  1.790669
#> [3,] -0.006634212  1.2534402  1.6584791  2.672113
#> [4,]  0.007521034 -0.6796158 -1.4192888 -2.029977
#> [5,]  1.116056450  7.2930310  8.6326605  9.628068
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_last_draw`
## ------------------------------------------------

specification = specify_bvarPANEL$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|