R6 Class Representing PosteriorBVARs

The class PosteriorBVARs contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvars$new()

specify_posterior_bvars$get_posterior()

specify_posterior_bvars$get_last_draw()

specify_posterior_bvars$clone()

Method new()

Create a new posterior output PosteriorBVARs.

Usage

specify_posterior_bvars$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARs with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARs.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvars$get_posterior()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvars$get_last_draw()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvars$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARs" "R6"            

## ------------------------------------------------
## Method `specify_posterior_bvars$get_posterior`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $nu
#>      [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,]  5.1 5.100000 5.100000 5.100000 5.100000
#> [2,]  5.1 5.100000 5.100000 5.079436 5.079436
#> [3,]  5.1 5.100000 5.100000 5.100000 5.100000
#> [4,]  5.1 5.001131 5.001131 5.001131 5.001131
#> [5,]  5.1 5.034205 5.034205 5.034205 5.034205
#> 
#> $m
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.8495975 0.8682247 0.7639118 1.0095890 1.2004548
#> [2,] 0.1281337 0.8407688 0.8563481 1.1168946 0.9626927
#> [3,] 0.3578316 0.4926332 0.1445306 0.4946215 1.0593604
#> [4,] 1.1938285 0.9798738 1.0490564 0.8187018 0.7589881
#> [5,] 0.1616546 0.9021193 0.7436686 0.4730325 0.7180821
#> 
#> $w
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.25662651 0.06802740 0.14679021 0.07406320 0.03903621
#> [2,] 0.04219498 0.07144417 0.09336787 0.07929608 0.11391884
#> [3,] 0.04002768 0.08354273 0.14773708 0.08511706 0.19060316
#> [4,] 0.15471867 0.18480453 0.10876687 0.19204352 0.11311116
#> [5,] 0.04346377 0.10782900 0.09161363 0.12935369 0.08515449
#> 
#> $s
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 139.24830  46.24218  53.47676 104.17525  59.59214
#> [2,] 835.77905  25.53658 139.90893  30.38242 102.70131
#> [3,]  81.46635  49.90824  55.86722  55.65314  37.30102
#> [4,]  24.60240 148.44889  46.33455  64.65289  59.86091
#> [5,] 222.69913  36.67961  51.61319  54.19050  76.18713
#> 
#> $scale
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
#> [4,]    0
#> [5,]    0
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  5.08689299 -0.8407560  0.1542650 -0.02152868
#> [2,] -0.84075601  5.1709941 -1.7802468 -0.66586435
#> [3,]  0.15426502 -1.7802468  5.7620224  0.86291840
#> [4,] -0.02152868 -0.6658644  0.8629184  3.86987995
#> 
#> , , 2, 1
#> 
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 28.682658 -1.558275 -7.582660  4.756147
#> [2,] -1.558275 24.740815 -3.522011 -1.958386
#> [3,] -7.582660 -3.522011 22.377108 -1.492157
#> [4,]  4.756147 -1.958386 -1.492157 28.062153
#> 
#> , , 3, 1
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  3.3916532 -0.2935344  1.601153 0.5913579
#> [2,] -0.2935344 12.7913756 -4.414127 2.3148893
#> [3,]  1.6011528 -4.4141270  7.360438 2.0299907
#> [4,]  0.5913579  2.3148893  2.029991 5.0889813
#> 
#> , , 4, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.97048860 -0.1641253  0.4574059 -0.08411293
#> [2,] -0.16412529  0.9107985 -0.3210008 -0.21223414
#> [3,]  0.45740595 -0.3210008  1.0980628  0.38666941
#> [4,] -0.08411293 -0.2122341  0.3866694  0.87630718
#> 
#> , , 5, 1
#> 
#>             [,1]        [,2]       [,3]      [,4]
#> [1,]  6.52150823 -0.01312817  0.1188668 0.8213156
#> [2,] -0.01312817  9.33880976 -2.7360083 0.9247816
#> [3,]  0.11886675 -2.73600834  8.5500888 1.9375843
#> [4,]  0.82131556  0.92478161  1.9375843 6.8631800
#> 
#> , , 1, 2
#> 
#>              [,1]          [,2]       [,3]          [,4]
#> [1,] 1.035306e+00  8.420698e-05  0.1412672  0.0009551398
#> [2,] 8.420698e-05  2.655727e+00 -0.0941932 -1.3491949122
#> [3,] 1.412672e-01 -9.419320e-02  1.4121781  0.5425639161
#> [4,] 9.551398e-04 -1.349195e+00  0.5425639  3.7512564444
#> 
#> , , 2, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8377160  0.2852245 -0.3320211 -0.3593279
#> [2,]  0.2852245  1.0760848 -0.4301104 -0.4964190
#> [3,] -0.3320211 -0.4301104  1.3253317  1.0104256
#> [4,] -0.3593279 -0.4964190  1.0104256  2.1858945
#> 
#> , , 3, 2
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,]  2.062551  -1.682078   3.035443  1.447656
#> [2,] -1.682078  13.031845 -10.222675 -2.303505
#> [3,]  3.035443 -10.222675  14.704625  5.735839
#> [4,]  1.447656  -2.303505   5.735839  5.242661
#> 
#> , , 4, 2
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  4.1209483 -0.9201750 1.06865503 -0.30147160
#> [2,] -0.9201750  4.5081095 0.23260983  0.32608923
#> [3,]  1.0686550  0.2326098 4.66956973  0.02900489
#> [4,] -0.3014716  0.3260892 0.02900489  4.21159418
#> 
#> , , 5, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.0294702  0.2104471 -0.4844714  0.2135988
#> [2,]  0.2104471  4.0487092 -1.8741400 -0.2995485
#> [3,] -0.4844714 -1.8741400  3.9818668  1.1557146
#> [4,]  0.2135988 -0.2995485  1.1557146  1.6075915
#> 
#> , , 1, 3
#> 
#>             [,1]        [,2]      [,3]       [,4]
#> [1,]  1.24969276 -0.04342977 0.4431811 -0.1152837
#> [2,] -0.04342977  1.44140077 0.2283571  0.0843089
#> [3,]  0.44318112  0.22835713 2.2115798  0.1558194
#> [4,] -0.11528373  0.08430890 0.1558194  2.0303686
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  6.0342399  1.6483193 -0.2352416 -0.5786638
#> [2,]  1.6483193  6.2265907  0.3613647 -2.0352641
#> [3,] -0.2352416  0.3613647  4.7960080 -0.8752841
#> [4,] -0.5786638 -2.0352641 -0.8752841  5.2458079
#> 
#> , , 3, 3
#> 
#>           [,1]       [,2]      [,3]     [,4]
#> [1,] 3.4100174  0.8159719  2.312237 2.470904
#> [2,] 0.8159719 10.8239065 -5.493427 1.970670
#> [3,] 2.3122370 -5.4934275 10.565511 5.262841
#> [4,] 2.4709036  1.9706703  5.262841 9.806671
#> 
#> , , 4, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.8382500 -0.4166732  0.5164718  0.1781285
#> [2,] -0.4166732  2.7157383 -0.8537897 -0.1695782
#> [3,]  0.5164718 -0.8537897  1.7622806  0.1585451
#> [4,]  0.1781285 -0.1695782  0.1585451  1.6801872
#> 
#> , , 5, 3
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  1.39332472 -0.1688535  0.06867563 -0.2647434
#> [2,] -0.16885350  4.6882566 -2.29062604  0.1066842
#> [3,]  0.06867563 -2.2906260  4.67117152  0.7500667
#> [4,] -0.26474341  0.1066842  0.75006669  1.9863207
#> 
#> , , 1, 4
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  2.75348881  0.4053450  0.06528303 -0.21405532
#> [2,]  0.40534501  3.6117052 -1.05236595 -0.41535269
#> [3,]  0.06528303 -1.0523659  3.14765580  0.04746048
#> [4,] -0.21405532 -0.4153527  0.04746048  2.56683287
#> 
#> , , 2, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9638329  0.2159094 -0.0752556 -0.2763092
#> [2,]  0.2159094  1.3227590 -0.5105467 -0.4617324
#> [3,] -0.0752556 -0.5105467  1.5785412  0.8032185
#> [4,] -0.2763092 -0.4617324  0.8032185  1.1254684
#> 
#> , , 3, 4
#> 
#>           [,1]       [,2]      [,3]     [,4]
#> [1,] 3.6451138  0.9747772  1.772335 2.295122
#> [2,] 0.9747772 10.0821148 -5.503130 1.840813
#> [3,] 1.7723354 -5.5031304  7.925546 2.402241
#> [4,] 2.2951224  1.8408128  2.402241 5.989143
#> 
#> , , 4, 4
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  2.32532447 -0.03780023 -0.1145324  0.3746790
#> [2,] -0.03780023  3.17653175  1.1784123 -0.4131481
#> [3,] -0.11453240  1.17841227  3.6209845  0.5053245
#> [4,]  0.37467898 -0.41314808  0.5053245  2.2837157
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.4537982  0.2921388  0.1033012 -0.3196783
#> [2,]  0.2921388  8.2879409 -5.6078485 -1.2667623
#> [3,]  0.1033012 -5.6078485  6.5336294  1.6106308
#> [4,] -0.3196783 -1.2667623  1.6106308  2.0213459
#> 
#> , , 1, 5
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,] 1.48430587 0.50605516 0.05459283 0.0282480
#> [2,] 0.50605516 2.39039889 0.06808628 0.2891137
#> [3,] 0.05459283 0.06808628 1.56722336 0.2916156
#> [4,] 0.02824800 0.28911369 0.29161556 1.9183029
#> 
#> , , 2, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.8993473  0.3675867 -0.5066207 -0.1112436
#> [2,]  0.3675867  3.5913713 -0.2206541 -0.4297706
#> [3,] -0.5066207 -0.2206541  3.0036206  0.3217448
#> [4,] -0.1112436 -0.4297706  0.3217448  2.9675964
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.2516785 -0.2040644  0.3636175  0.6970637
#> [2,] -0.2040644 10.1279506 -6.0518260 -1.9751432
#> [3,]  0.3636175 -6.0518260  6.3963002  3.3400882
#> [4,]  0.6970637 -1.9751432  3.3400882  4.7336884
#> 
#> , , 4, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.3054752 -0.6928484  0.1454327 -0.1811175
#> [2,] -0.6928484  2.0628241 -0.4185093 -0.5076050
#> [3,]  0.1454327 -0.4185093  2.3415650  0.8098505
#> [4,] -0.1811175 -0.5076050  0.8098505  3.0104384
#> 
#> , , 5, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.1940590 -0.1763117  0.1412932 -0.2058467
#> [2,] -0.1763117  3.9184714 -1.5883588 -1.0523772
#> [3,]  0.1412932 -1.5883588  4.8099380  1.1361850
#> [4,] -0.2058467 -1.0523772  1.1361850  2.8035049
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  1.5987244 -0.04005507 -0.12735484  0.1657819
#> [2,] -0.1944307  0.85980546 -0.02540139  0.3659525
#> [3,]  0.6322887 -1.18817527  2.03386249  1.2805013
#> [4,] -0.7979614  1.23354733 -0.92484499 -0.3350355
#> [5,] -2.1512367 -4.68069834  2.78347780  0.4707550
#> 
#> , , 2, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.5444390 -0.4642077  0.2935320  0.2736167
#> [2,] -0.6031712  1.0351124  0.4721838  0.6606052
#> [3,]  0.7211623  0.4374799  1.0703864  1.6144261
#> [4,] -0.3383689 -0.1872511 -0.2826571 -0.5805243
#> [5,]  0.5286287 -0.5325410  2.3427089 -0.4853125
#> 
#> , , 3, 1
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,] -0.06064851 -0.74717766 -0.3283988 -0.9336209
#> [2,]  0.53869680  0.70896634  0.3636242  0.4033849
#> [3,]  1.02513166  0.37915280  1.1821431  0.8199009
#> [4,] -0.61917974  0.05066187 -0.1524461  0.5297212
#> [5,]  1.34802397 -0.61639296  2.0788848  2.1182279
#> 
#> , , 4, 1
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.1571263  0.2826602 -0.77583810 -0.1348249
#> [2,]  0.1330714  1.1251550  0.02923913 -0.5829925
#> [3,] -0.2263628  0.2949786  0.86072011 -0.4437115
#> [4,]  0.4922452 -0.3867083  0.41527179  1.5247751
#> [5,]  0.9578444 -0.3573128 -1.18167944 -1.7934360
#> 
#> , , 5, 1
#> 
#>              [,1]       [,2]        [,3]       [,4]
#> [1,]  0.551719273 -1.1173710  0.96900903 -0.8311049
#> [2,]  0.042372450  0.7533959 -0.09189942  0.4812597
#> [3,]  0.247938668 -0.1133430  0.72863772  0.9226531
#> [4,] -0.006630554  0.6318003 -0.11391157  0.4689519
#> [5,] -1.572972499  0.4366482 -2.49560529 -0.3729583
#> 
#> , , 1, 2
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.79778468  0.2579027 -0.2432952 -0.49874301
#> [2,]  0.19271032  0.5220833  0.2637756  0.64376014
#> [3,]  0.20732079 -0.8167066  1.5557109  1.09346142
#> [4,] -0.11951381  0.7302408 -0.4513936  0.11769704
#> [5,]  0.09428293 -0.6082403  0.1162190  0.08173841
#> 
#> , , 2, 2
#> 
#>             [,1]         [,2]        [,3]       [,4]
#> [1,]  0.56366370 -0.267854578  0.44513363  0.5399762
#> [2,] -0.47950933  0.733147635  0.33813951  0.5528164
#> [3,] -0.02741615  0.187630514  0.74237960 -0.1674293
#> [4,]  0.25127623 -0.003524348  0.01105889  0.8463266
#> [5,]  1.08674625 -0.785254404 -0.71478762  0.1472758
#> 
#> , , 3, 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.3353306 -0.69679308  0.1220455 -0.3981005
#> [2,]  1.0292094  0.57799151  0.8689654  1.0343531
#> [3,]  1.6280436  0.38520720  1.8182462  1.7362257
#> [4,] -1.4000940  0.09473132 -1.0126617 -0.6231816
#> [5,]  0.8759372 -2.26236800  2.4187302  2.4981765
#> 
#> , , 4, 2
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,] -0.02268471 -0.02939119 -0.6411944 -0.10869883
#> [2,] -0.71382680  0.91536413  0.6775033  0.19383638
#> [3,] -0.51039382  0.61472467  1.0692731  0.02532967
#> [4,]  0.89046189 -0.61490471  0.1694432  0.95984930
#> [5,] -1.50124460  2.37727565 -2.3607712  3.92393339
#> 
#> , , 5, 2
#> 
#>              [,1]        [,2]        [,3]      [,4]
#> [1,]  1.114037401 -1.00681288  1.16910198 0.5631610
#> [2,] -0.001706124  0.64348532 -0.04541168 0.3488063
#> [3,]  0.004711892 -0.05644135  0.53546099 0.4885293
#> [4,] -0.083420060  0.55106238 -0.09504551 0.2349790
#> [5,]  1.764728787  1.00553509 -0.77811226 1.0198545
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  1.0497309 -0.1849985  0.15550016  0.13984920
#> [2,] -0.2015697  1.0885620  0.01072221 -0.01754677
#> [3,] -0.3350773 -0.3916415  1.56040941  0.45207363
#> [4,]  0.3106461  0.4029647 -0.61340063  0.48466298
#> [5,]  0.5684009  1.7041112  0.64623034  1.02886277
#> 
#> , , 2, 3
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.69384418 -0.5453756  0.74319934  0.5383564
#> [2,] -0.26497494  0.9752232  0.03240591  0.2790843
#> [3,]  0.32354198  0.4346008  0.91285926 -0.1749873
#> [4,] -0.05357033 -0.1793323 -0.17539657  0.8913243
#> [5,] -4.82275632 -0.2182183  1.00099981  1.7368524
#> 
#> , , 3, 3
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.1477270 -0.89392447  0.05828931 -0.8093019
#> [2,]  0.8814261  0.74189306  0.77916922  0.9643891
#> [3,]  1.4198220  0.58048136  1.50520080  1.6973555
#> [4,] -1.1235394 -0.09359393 -0.69654087 -0.3825799
#> [5,]  1.5878577  1.65847035  1.27129160 -0.3536352
#> 
#> , , 4, 3
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,] -0.07165513  0.09608699 -0.3840078 -0.31337155
#> [2,]  0.47368776  1.78876034 -0.2316028  0.08787527
#> [3,]  0.31074776  0.98207351  0.1476246  0.22847407
#> [4,]  0.04142907 -1.03679085  0.9571781  0.91688547
#> [5,]  1.12566725  1.50674956  0.9573976 -3.09309392
#> 
#> , , 5, 3
#> 
#>            [,1]       [,2]        [,3]      [,4]
#> [1,]  0.8219340 -0.4639069  0.97316605 0.7314333
#> [2,]  0.1220811  0.7313453 -0.20770156 0.3428325
#> [3,]  0.3144079 -0.1942680  0.54306800 0.3686094
#> [4,] -0.1945537  0.3942140  0.06019911 0.2784768
#> [5,] -1.6744021  2.4460652 -3.23066531 0.4153140
#> 
#> , , 1, 4
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.67203303  0.01555798 -0.2608835  0.07738636
#> [2,]  0.30116228  0.99721022  0.1605589 -0.46342805
#> [3,]  0.16884916 -0.05556979  1.1063217 -0.43987166
#> [4,] -0.06636707  0.04972309 -0.0212792  1.43088044
#> [5,]  0.66716648  0.05328869 -0.1797359  0.63331595
#> 
#> , , 2, 4
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  0.6817680 -0.23444557 -0.03044250  0.55703204
#> [2,] -0.2992227  0.65150205  0.62060274  0.54003443
#> [3,] -0.1020009 -0.20891390  0.75319793  0.01386569
#> [4,]  0.2459655  0.32207532  0.08704604  0.69716920
#> [5,]  0.4080080 -0.03882738 -0.17330569 -0.11414430
#> 
#> , , 3, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.2037624 -0.5290492 -0.2765778 -0.8635522
#> [2,]  0.6834715  0.9069822  0.4426620  0.8094196
#> [3,]  1.0087202  0.7918026  1.3566178  1.6054284
#> [4,] -0.8005091 -0.3966959 -0.3873729 -0.2786149
#> [5,]  4.2886150 -3.3232564  4.9020831  2.3297191
#> 
#> , , 4, 4
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,] 0.23229976  0.00283252 -0.6978078  0.07493173
#> [2,] 0.71221830  1.18570372 -0.2391063 -0.38956078
#> [3,] 0.03243621  0.72913659  0.5923565 -0.60442166
#> [4,] 0.18586475 -0.71067694  0.6178181  1.54238276
#> [5,] 2.13212514 -0.44720795  2.3422063  2.85999953
#> 
#> , , 5, 4
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.73206601 -0.02692687  0.90121672  0.57545490
#> [2,]  0.06164654  0.88706999 -0.01054811 -0.07335276
#> [3,]  0.21349910  0.17394404  0.54942578 -0.13341620
#> [4,] -0.09673508 -0.01617469 -0.04567415  0.91365152
#> [5,]  0.75853556 -5.54461485  2.47285544 -2.16222793
#> 
#> , , 1, 5
#> 
#>              [,1]        [,2]        [,3]       [,4]
#> [1,]  1.259220323  0.05659643  0.06196316  0.1641865
#> [2,] -0.171880402  1.17933532 -0.16288213 -0.2851683
#> [3,]  0.004672815  0.33571823  0.93954170 -0.2616378
#> [4,] -0.078451706 -0.35357282  0.05528620  1.1950393
#> [5,] -0.580945834 -0.31707525 -0.06627702  0.1786312
#> 
#> , , 2, 5
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  1.0467520 -0.6038637 -0.4248320  0.30382526
#> [2,] -0.4397968  0.4590555  1.1712751  0.26146732
#> [3,] -0.3377222 -0.1275229  1.5221805  0.08648848
#> [4,]  0.3022059  0.4193114 -0.5329299  0.81385624
#> [5,]  4.5437159  0.0471289 -1.0201441 -3.81445693
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.3226345 -0.6121589  0.02769584 -0.7616524
#> [2,]  0.5801279  0.7890130  0.69948729  0.9647592
#> [3,]  1.0215554  0.5400422  1.73661688  1.7110276
#> [4,] -0.8070887 -0.1508068 -0.78598666 -0.3980075
#> [5,]  3.3089831 -0.6089832 -1.86149130 -0.1995715
#> 
#> , , 4, 5
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.05945378  0.1677705  0.3096784  0.2132793
#> [2,] -0.04497754  0.6582011 -0.3949692 -0.5836340
#> [3,]  0.27344407  0.2756603  0.6913566 -0.1836050
#> [4,]  0.05454775 -0.3377458  0.1914266  1.1388575
#> [5,]  0.09717716  1.7552741  0.9062654 -0.7334278
#> 
#> , , 5, 5
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.7186698 -0.17771405  1.0893769  0.1529921
#> [2,]  0.2617184  1.08045157 -0.4813574 -0.7031931
#> [3,]  0.4096646  0.46145690 -0.1141647 -0.9993535
#> [4,] -0.2787951 -0.34027930  0.5945817  1.9322166
#> [5,] -0.4876987  0.06958336  0.6972269  1.5093343
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvars$get_last_draw`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|