R6 Class Representing PosteriorBVARs

The class PosteriorBVARs contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvars$new()

specify_posterior_bvars$get_posterior()

specify_posterior_bvars$get_last_draw()

specify_posterior_bvars$clone()

Method new()

Create a new posterior output PosteriorBVARs.

Usage

specify_posterior_bvars$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARs with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARs.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvars$get_posterior()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvars$get_last_draw()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvars$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARs" "R6"            

## ------------------------------------------------
## Method `specify_posterior_bvars$get_posterior`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $nu
#>          [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,] 5.100000 5.100000 5.100000 5.100000 5.100000
#> [2,] 5.100000 5.100000 5.047733 5.047733 5.047733
#> [3,] 5.100000 5.100000 5.100000 5.100000 5.100000
#> [4,] 5.100000 5.076783 5.076783 5.076783 5.076783
#> [5,] 5.062302 5.062302 5.062302 5.062302 5.062302
#> 
#> $m
#>             [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,]  1.26254740 1.2654480 1.1062000 0.7879457 0.4826240
#> [2,]  0.94000304 0.7649120 0.7341031 0.2878214 0.5206422
#> [3,]  0.97962237 1.0401757 0.9443234 0.8579481 0.6058998
#> [4,] -0.09718435 1.0169199 0.5977836 0.7856812 1.0978106
#> [5,]  0.61392721 0.4542307 1.3235426 0.9001978 0.5353064
#> 
#> $w
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.18048172 0.20615453 0.07892005 0.08193152 0.13042147
#> [2,] 0.09333035 0.26282839 0.13377311 0.18647254 0.15931539
#> [3,] 0.73262483 0.40354627 0.13292615 0.07395895 0.09489119
#> [4,] 0.04910223 0.09396001 0.16079808 0.10546899 0.12030201
#> [5,] 0.06732733 0.13899589 0.14101253 0.31016665 0.15983911
#> 
#> $s
#>           [,1]      [,2]     [,3]     [,4]     [,5]
#> [1,]  62.89257  73.42451 46.58467 97.15351 65.93455
#> [2,]  47.36233 103.47627 69.86294 58.81769 61.31910
#> [3,] 202.17807  37.95226 50.20046 53.76704 44.35619
#> [4,] 485.20717  41.59145 52.83048 74.41204 72.95379
#> [5,]  51.63329  50.65602 77.03194 43.37183 75.80075
#> 
#> $scale
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
#> [4,]    0
#> [5,]    0
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>           [,1]       [,2]      [,3]       [,4]
#> [1,] 1.2171182 0.12971293 0.3348748 0.11373436
#> [2,] 0.1297129 1.77200875 0.2493534 0.02169235
#> [3,] 0.3348748 0.24935338 2.5867046 0.77363796
#> [4,] 0.1137344 0.02169235 0.7736380 2.31019347
#> 
#> , , 2, 1
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  1.22224569 -0.05007458  0.07690683  0.01895656
#> [2,] -0.05007458  1.84702331 -0.08380911 -0.54959986
#> [3,]  0.07690683 -0.08380911  1.56182924  0.57969715
#> [4,]  0.01895656 -0.54959986  0.57969715  1.89335186
#> 
#> , , 3, 1
#> 
#>           [,1]       [,2]      [,3]      [,4]
#> [1,] 8.0547411  0.9095342  2.794031  3.393679
#> [2,] 0.9095342 12.1547645 -2.299733 -1.688621
#> [3,] 2.7940310 -2.2997333  8.761908  2.550158
#> [4,] 3.3936790 -1.6886209  2.550158  7.801035
#> 
#> , , 4, 1
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,] 13.2954635  1.599275 -0.8873954  1.8848041
#> [2,]  1.5992752 13.218789  1.4561221  1.2194064
#> [3,] -0.8873954  1.456122 12.7675752 -0.8589475
#> [4,]  1.8848041  1.219406 -0.8589475 10.1091393
#> 
#> , , 5, 1
#> 
#>            [,1]      [,2]        [,3]     [,4]
#> [1,] 1.85258147  1.022043  0.01642602 1.341198
#> [2,] 1.02204303 10.553730 -1.17051684 1.951407
#> [3,] 0.01642602 -1.170517  7.44551058 3.969286
#> [4,] 1.34119785  1.951407  3.96928637 5.565252
#> 
#> , , 1, 2
#> 
#>              [,1]       [,2]       [,3]         [,4]
#> [1,]  2.916962212  0.0509453 -0.3181779  0.006892285
#> [2,]  0.050945304  1.6917858  0.1860893 -0.113712303
#> [3,] -0.318177889  0.1860893  3.0420439  1.039571284
#> [4,]  0.006892285 -0.1137123  1.0395713  2.383795123
#> 
#> , , 2, 2
#> 
#>              [,1]        [,2]      [,3]        [,4]
#> [1,]  2.951852002 -0.02770634 0.2915232 0.007487671
#> [2,] -0.027706342  3.90373555 0.6978194 0.216405949
#> [3,]  0.291523208  0.69781938 3.8242346 0.644145742
#> [4,]  0.007487671  0.21640595 0.6441457 2.661661828
#> 
#> , , 3, 2
#> 
#>           [,1]      [,2]       [,3]      [,4]
#> [1,] 1.5669923  1.715057  0.2051089 0.9754555
#> [2,] 1.7150569 14.207481 -5.4011421 3.3220387
#> [3,] 0.2051089 -5.401142  6.6078193 1.6322345
#> [4,] 0.9754555  3.322039  1.6322345 3.9720333
#> 
#> , , 4, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.3285270 -0.1266990  0.2589574 -0.1216832
#> [2,] -0.1266990  1.1002895 -0.4526487 -0.2746094
#> [3,]  0.2589574 -0.4526487  1.7779852  0.6403761
#> [4,] -0.1216832 -0.2746094  0.6403761  1.9834923
#> 
#> , , 5, 2
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  1.2539919  0.14572025 -0.3451888 -0.11663506
#> [2,]  0.1457203  4.40632977 -1.2995514 -0.01625762
#> [3,] -0.3451888 -1.29955137  3.5094002  1.54725557
#> [4,] -0.1166351 -0.01625762  1.5472556  3.53304373
#> 
#> , , 1, 3
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  1.13695894 -0.10073257 0.04220625  0.66985128
#> [2,] -0.10073257  0.99183400 0.06559526 -0.06529005
#> [3,]  0.04220625  0.06559526 2.65259837  0.27702012
#> [4,]  0.66985128 -0.06529005 0.27702012  2.43824019
#> 
#> , , 2, 3
#> 
#>           [,1]        [,2]       [,3]        [,4]
#> [1,]  3.856485 -1.04064281  1.2314937 -0.32202503
#> [2,] -1.040643  2.76906289 -0.3643781  0.05844608
#> [3,]  1.231494 -0.36437807  3.3335089  0.30044490
#> [4,] -0.322025  0.05844608  0.3004449  1.63404424
#> 
#> , , 3, 3
#> 
#>             [,1]      [,2]        [,3]        [,4]
#> [1,]  1.27105748  0.275989  0.08833102 -0.01192445
#> [2,]  0.27598900 10.296811 -7.11312991 -1.41594589
#> [3,]  0.08833102 -7.113130 11.17216590  5.25171058
#> [4,] -0.01192445 -1.415946  5.25171058  5.36309526
#> 
#> , , 4, 3
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,] 1.65543764  0.05680475  0.2791276  0.16290986
#> [2,] 0.05680475  1.43513113 -0.6401562  0.09830029
#> [3,] 0.27912757 -0.64015622  1.8506643 -0.20645470
#> [4,] 0.16290986  0.09830029 -0.2064547  1.53261425
#> 
#> , , 5, 3
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  2.79709392 -0.9673267  0.4486343 -0.01407213
#> [2,] -0.96732675  8.0521072 -3.2563935  0.77710338
#> [3,]  0.44863430 -3.2563935  7.5599720  2.69455629
#> [4,] -0.01407213  0.7771034  2.6945563  4.00794991
#> 
#> , , 1, 4
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  2.8480554 -0.3147091  0.68388793 -0.20179599
#> [2,] -0.3147091  3.0889273 -0.77921453  0.57207354
#> [3,]  0.6838879 -0.7792145  3.90787034  0.08029043
#> [4,] -0.2017960  0.5720735  0.08029043  2.98423115
#> 
#> , , 2, 4
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  1.58263134  0.1172108  0.09949944 -0.05963007
#> [2,]  0.11721081  2.1367132 -0.47071552  0.32188623
#> [3,]  0.09949944 -0.4707155  2.20977085  0.50978764
#> [4,] -0.05963007  0.3218862  0.50978764  2.29124973
#> 
#> , , 3, 4
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  2.01288393 -0.6184569  0.1035081 -0.01187467
#> [2,] -0.61845692 16.2784746 -7.5221458 -0.32722113
#> [3,]  0.10350809 -7.5221458  7.0229230  2.42617948
#> [4,] -0.01187467 -0.3272211  2.4261795  3.85239981
#> 
#> , , 4, 4
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  1.89986151  0.6179331 -0.10506250 -0.08008348
#> [2,]  0.61793309  2.5713770 -0.26167188 -0.65899315
#> [3,] -0.10506250 -0.2616719  1.98103664  0.09850825
#> [4,] -0.08008348 -0.6589932  0.09850825  2.94104703
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.2259560 -0.4745849  0.7391078  0.2634378
#> [2,] -0.4745849  3.6272523 -2.6329423 -0.3146513
#> [3,]  0.7391078 -2.6329423  4.6495465  1.9740070
#> [4,]  0.2634378 -0.3146513  1.9740070  2.8211598
#> 
#> , , 1, 5
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  2.6407569 -0.2016063 -0.33691296 -0.37331972
#> [2,] -0.2016063  2.3113410  0.56927150 -0.26253106
#> [3,] -0.3369130  0.5692715  2.37507674  0.09332044
#> [4,] -0.3733197 -0.2625311  0.09332044  2.25754886
#> 
#> , , 2, 5
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  1.70973791 -0.06659057 -0.07572757 -0.3599960
#> [2,] -0.06659057  1.69155238  0.03707392  0.3081897
#> [3,] -0.07572757  0.03707392  1.97239658  0.6143045
#> [4,] -0.35999601  0.30818975  0.61430454  2.4437534
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  2.0732567  0.3239909 -0.011033 -0.4224522
#> [2,]  0.3239909 10.9444358 -5.102250  1.4917146
#> [3,] -0.0110330 -5.1022503  5.995466  1.1060886
#> [4,] -0.4224522  1.4917146  1.106089  3.8539771
#> 
#> , , 4, 5
#> 
#>            [,1]          [,2]          [,3]        [,4]
#> [1,]  2.6883569 -0.1510524489  0.4276881436 -0.45092304
#> [2,] -0.1510524  3.1679263232 -0.0006403914  0.75056241
#> [3,]  0.4276881 -0.0006403914  2.0071907588 -0.05519769
#> [4,] -0.4509230  0.7505624123 -0.0551976911  2.36130024
#> 
#> , , 5, 5
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  1.63723881  0.3345642 -0.03948649  0.09118519
#> [2,]  0.33456418  6.5664010 -3.12661370 -0.39590203
#> [3,] -0.03948649 -3.1266137  4.82158650  1.20196577
#> [4,]  0.09118519 -0.3959020  1.20196577  2.78897175
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  1.1031644 -0.05000031 -0.16808090 -0.3729899
#> [2,] -0.1854077  0.82321792 -0.34138136  0.1136551
#> [3,] -0.4245783 -0.38106086 -0.04652109 -0.1914036
#> [4,]  0.4053490  0.43516821  1.04626469  1.2816280
#> [5,] -1.3212662 -0.70362554  2.21483962  2.3598774
#> 
#> , , 2, 1
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.7555350  0.02964537 -0.05250283 0.23059615
#> [2,] -0.2367797  1.04264320  0.38165800 0.89972756
#> [3,] -0.3040390 -0.16969135  1.06059465 0.81591125
#> [4,]  0.3521099  0.08960033 -0.12614418 0.05416481
#> [5,]  2.6758779  1.96818148  0.81642818 0.18797280
#> 
#> , , 3, 1
#> 
#>           [,1]      [,2]      [,3]       [,4]
#> [1,]  1.315634 -1.770798  1.347617  0.1299795
#> [2,] -1.616595  1.589878 -1.990703 -1.0347050
#> [3,] -2.367870  2.120439 -2.595317 -1.4787631
#> [4,]  2.060182 -1.393121  2.805945  2.1903509
#> [5,] 13.685731  8.237601 13.667567 16.8522450
#> 
#> , , 4, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.3438421  0.2405208 -1.1949401 -0.1998325
#> [2,] -0.3041656 -0.1639058 -0.8138573  0.3915590
#> [3,] -0.2435641 -1.4646133  0.1307326  0.0532660
#> [4,]  0.7043002  1.3945406  1.3480825  0.9961688
#> [5,]  0.1218139 -1.1598370 -3.6471872  1.4404133
#> 
#> , , 5, 1
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.17864579 -0.5852572  1.9081118  1.9040402
#> [2,] -0.11197353 -0.4528106 -0.5446756 -0.5403234
#> [3,]  0.01645219 -1.8455354 -0.3026944 -0.9277032
#> [4,] -0.05905004  2.1972164  0.4023489  1.0688818
#> [5,] -0.45078500 -1.3593636  1.2111793  1.5913103
#> 
#> , , 1, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.99523480  0.09036727 -0.48403912 -0.24673277
#> [2,]  0.11865161  0.89535846  0.02159326 -0.05709159
#> [3,] -0.13611149 -0.03430291  0.34927580  0.11533105
#> [4,]  0.07980012 -0.01161886  0.85676768  0.97748476
#> [5,]  0.95201304  0.99134112 -1.87816428  1.80529397
#> 
#> , , 2, 2
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.83211200 -0.4200142  0.5718876  0.5246218
#> [2,]  0.19643656  0.8218744  0.4585998  1.0232737
#> [3,] -0.17499516  0.4291478  0.5509125  0.6259714
#> [4,]  0.15631993 -0.1156213  0.1336596  0.1298052
#> [5,]  0.08483839 -5.2083335 -3.1466461 -3.2530737
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8355230 -0.4704584  0.8795598  0.5760225
#> [2,] -0.3548003  0.1168795 -1.0392030 -1.1039336
#> [3,] -0.3974014 -0.4397642 -0.9107114 -1.4271405
#> [4,]  0.4567203  0.8226197  1.4868061  2.2441019
#> [5,]  1.9147129 -2.0850345  2.6541026  1.1642824
#> 
#> , , 4, 2
#> 
#>             [,1]        [,2]       [,3]         [,4]
#> [1,]  0.06440620  0.19233064 -0.3774543 -0.533039635
#> [2,]  0.39194769  0.25425814  0.3926551 -0.001558559
#> [3,]  0.25340597 -0.07138875  0.8028759  0.131566088
#> [4,]  0.07336868  0.05022423  0.2820076  1.051073357
#> [5,] -0.57408874 -1.76753041  2.2122852 -0.282965422
#> 
#> , , 5, 2
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.05367316 -0.1041117  1.7942081  1.2593086
#> [2,] -0.01963374  0.5129147 -0.2528804 -0.3087593
#> [3,] -0.06188252 -0.5229972 -0.0429500 -1.0301227
#> [4,]  0.01456339  0.6164406  0.1925174  1.4058166
#> [5,]  1.27299174 -0.9376948 -0.7013887  0.5663905
#> 
#> , , 1, 3
#> 
#>              [,1]        [,2]        [,3]       [,4]
#> [1,]  1.064326795  0.02294975 -0.15812378  0.1327622
#> [2,] -0.203176180  0.92997786 -0.23107806 -0.4836337
#> [3,] -0.405779592 -0.21904946  0.62027559  0.3531297
#> [4,]  0.371998921  0.20027829  0.47135558  0.6290056
#> [5,]  0.005734242  0.08663214  0.03301912  3.2781356
#> 
#> , , 2, 3
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.65713221 -0.1646209  0.9787238  0.15600286
#> [2,]  0.07919818  0.7382616  0.4997502  0.70691929
#> [3,]  0.06736512  0.1183267  0.3695095  0.94689098
#> [4,]  0.04902950 -0.0298982  0.1356217  0.01895205
#> [5,] -0.87280505  2.9439870 -1.7563111 -1.08719594
#> 
#> , , 3, 3
#> 
#>            [,1]         [,2]        [,3]        [,4]
#> [1,]  0.8033821 -0.654208125  0.53709264  0.09320557
#> [2,]  0.1485994  0.706984309 -0.04385069 -0.07405331
#> [3,]  0.3468045  0.433039497  0.64272630  0.07036822
#> [4,] -0.2700460  0.007079035  0.06444050  0.91121994
#> [5,]  0.7924794 -2.651565600  2.02381579  0.54536195
#> 
#> , , 4, 3
#> 
#>            [,1]        [,2]       [,3]      [,4]
#> [1,]  0.4708629 -0.07583789  0.3578034 0.2164124
#> [2,]  0.5378393  0.61990543  0.2842378 0.1225149
#> [3,]  0.2997266  0.06300704  1.1634401 0.3547298
#> [4,] -0.1294717 -0.03274401 -0.2927438 0.5676706
#> [5,]  0.4340691  0.62971228  0.4165326 0.5569761
#> 
#> , , 5, 3
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.3627743 -0.47683806  1.2877309  1.1337203
#> [2,]  0.3226670 -0.07428104 -0.6071794 -0.3309056
#> [3,]  0.4082570 -1.42696812 -0.3770782 -1.1014733
#> [4,] -0.5761337  1.75233007  0.6812118  1.5200740
#> [5,]  0.0956879 -4.96724869  4.3953834  1.1934852
#> 
#> , , 1, 4
#> 
#>            [,1]        [,2]       [,3]         [,4]
#> [1,]  1.3902445 -0.03612218  0.3759036  0.170558024
#> [2,] -0.2473384  0.60269262 -0.7791456 -0.327173077
#> [3,] -0.1589681 -0.15182631  0.3271031 -0.001445131
#> [4,] -0.0362184  0.24005493  0.4827492  0.975017042
#> [5,]  1.5683071 -0.33967142  3.6108235 -0.395385203
#> 
#> , , 2, 4
#> 
#>              [,1]        [,2]        [,3]        [,4]
#> [1,]  0.820414392  0.09121942 -0.07584041  0.46492186
#> [2,] -0.222910732  1.33361554  0.15831001  0.74951045
#> [3,] -0.001509797 -0.15084326  1.18011568  0.82379617
#> [4,]  0.122165860  0.03058474 -0.18292724 -0.00788754
#> [5,] -1.027085546 -0.70095803  1.68790940  0.93721394
#> 
#> , , 3, 4
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.78935960 -0.5231740  0.8307326  0.4153408
#> [2,]  0.05980891  0.9313207 -0.5245274 -0.4806725
#> [3,]  0.10322236  0.7095991 -0.4133620 -0.5497432
#> [4,] -0.01636498 -0.3406261  0.9743943  1.3946935
#> [5,] -0.81377563 -2.6938524  2.2804213  1.7849732
#> 
#> , , 4, 4
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  1.05630581 -0.2196932  0.19767722  0.49410761
#> [2,]  0.13696383  0.8714472 -0.05567866  0.17553724
#> [3,] -0.12742768  0.3536976  0.76293031  0.01571495
#> [4,]  0.08824434 -0.2454752  0.14887987  0.82122489
#> [5,]  0.95319721 -1.7954847  0.91082752 -0.68502057
#> 
#> , , 5, 4
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.7252110  0.03564648  0.86651624  0.3517828
#> [2,] -0.1563067  0.81850980 -0.97114700 -0.8110919
#> [3,] -0.1732366  0.03201982 -0.99150966 -1.4960973
#> [4,]  0.2914409  0.01371732  1.56558519  2.2877352
#> [5,]  0.6108211 -1.28422772  0.05444958  2.5803225
#> 
#> , , 1, 5
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.9917136  0.3967190  0.2754802  0.12146203
#> [2,]  0.1187984  0.5849679 -0.4989876 -0.07709115
#> [3,] -0.1047411 -0.5672192  0.6167597  0.46588389
#> [4,]  0.1283009  0.3547825  0.3001704  0.54415921
#> [5,] -2.1448450  1.9914339 -0.1317739 -1.01083255
#> 
#> , , 2, 5
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.82702654 -0.16370257  0.37618096  0.1684659
#> [2,] -0.02121599  1.18423347  0.32448121  1.2668101
#> [3,] -0.10174769  0.06897822  0.86565041  0.9506474
#> [4,]  0.15324834 -0.03461643 -0.05828183 -0.0973956
#> [5,] -0.62762694 -0.63421373 -1.48116238 -1.6291013
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9995674 -0.2965500  0.9286783  0.6140247
#> [2,] -0.1666244  0.9165917 -0.7267488 -0.5848033
#> [3,] -0.0546503  0.6935323 -0.3824047 -0.5448386
#> [4,]  0.1023516 -0.3594292  0.9331211  1.3912495
#> [5,] -0.3697030 -5.8575422  4.4345169 -1.2169063
#> 
#> , , 4, 5
#> 
#>             [,1]         [,2]       [,3]       [,4]
#> [1,]  1.56213393 -0.007318809  0.5859871  0.5130220
#> [2,]  0.03406263  0.756429059  0.6044112 -0.0108897
#> [3,]  0.23340495 -0.445130693  1.1813565 -0.4298563
#> [4,] -0.40622051  0.437283902 -0.4125976  1.2526648
#> [5,] -0.83558387  0.259237779  1.3339874 -0.5472630
#> 
#> , , 5, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.4780925 -0.3059996  1.5504289  1.0145833
#> [2,] -0.1713720  0.5481406 -0.4483970 -0.3726840
#> [3,] -0.1503665 -0.2896063 -0.2385731 -0.8468488
#> [4,]  0.3588160  0.4073248  0.5529134  1.4116954
#> [5,]  1.9129721  4.7807545 -3.5665961 -2.7558589
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvars$get_last_draw`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|