R6 Class Representing PosteriorBVARs

The class PosteriorBVARs contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvars$new()

specify_posterior_bvars$get_posterior()

specify_posterior_bvars$get_last_draw()

specify_posterior_bvars$clone()

Method new()

Create a new posterior output PosteriorBVARs.

Usage

specify_posterior_bvars$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARs with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARs.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvars$get_posterior()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvars$get_last_draw()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvars$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARs" "R6"            

## ------------------------------------------------
## Method `specify_posterior_bvars$get_posterior`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $nu
#>      [,1] [,2] [,3]     [,4]     [,5]
#> [1,]  5.1  5.1  5.1 5.100000 5.100000
#> [2,]  5.1  5.1  5.1 5.119945 5.119945
#> [3,]  5.1  5.1  5.1 5.230650 5.230650
#> [4,]  5.1  5.1  5.1 5.100000 5.100000
#> [5,]  5.1  5.1  5.1 5.100000 5.081335
#> 
#> $m
#>             [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,]  1.16784129 1.0106740 1.0284236 0.7370142 0.6945989
#> [2,] -0.05345312 0.7553113 0.5882981 0.8918219 0.6098565
#> [3,]  0.18124651 0.4754992 0.6744253 0.8778214 0.6683411
#> [4,]  0.33861328 1.2401797 0.9010557 0.5867028 0.7128186
#> [5,]  0.78160142 1.1842921 1.1529461 1.2921505 0.8845885
#> 
#> $w
#>            [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.48009820 0.12883894 0.1724482 0.20003860 0.19051380
#> [2,] 0.05085382 0.07826823 0.1177261 0.07022986 0.09013887
#> [3,] 0.07363499 0.10817387 0.2878685 0.12990620 0.09178337
#> [4,] 0.20686026 0.09615377 0.2049521 0.14567971 0.08705124
#> [5,] 0.15828585 0.22558635 0.1506545 0.10362208 0.14721993
#> 
#> $s
#>           [,1]     [,2]      [,3]     [,4]      [,5]
#> [1,]  175.2424 35.28445 132.05653 24.79301  68.30916
#> [2,] 6831.7036 18.53436 108.65859 39.20047 122.63526
#> [3,]  209.2595 39.31381  54.94917 67.77378  56.52352
#> [4,]  811.6396 24.77370  91.62843 72.39735  68.56965
#> [5,]  113.7682 29.26502  84.09107 34.13220  57.61306
#> 
#> $scale
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
#> [4,]    0
#> [5,]    0
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  5.31095202 -0.9446609 -0.2621553  0.09234773
#> [2,] -0.94466091  4.1270835  0.6829409 -0.18160948
#> [3,] -0.26215529  0.6829409  4.6332272 -1.12871312
#> [4,]  0.09234773 -0.1816095 -1.1287131  6.72780886
#> 
#> , , 2, 1
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,] 141.866534  -3.306681  -9.866889 -26.88760
#> [2,]  -3.306681 168.072046 -20.155400 -26.71069
#> [3,]  -9.866889 -20.155400 153.272565  17.76276
#> [4,] -26.887597 -26.710693  17.762759 191.68820
#> 
#> , , 3, 1
#> 
#>           [,1]       [,2]      [,3]       [,4]
#> [1,]  7.395754 -2.4237374  3.709719  1.1618200
#> [2,] -2.423737 15.0563497 -8.198930 -0.7960513
#> [3,]  3.709719 -8.1989296 14.836110  6.2751853
#> [4,]  1.161820 -0.7960513  6.275185 15.4190623
#> 
#> , , 4, 1
#> 
#>            [,1]      [,2]      [,3]       [,4]
#> [1,] 13.0966817 -1.378695 -1.310017 -0.7645589
#> [2,] -1.3786953 32.517671  7.189675 -4.1339192
#> [3,] -1.3100171  7.189675 27.363199 -2.0461638
#> [4,] -0.7645589 -4.133919 -2.046164 18.3225210
#> 
#> , , 5, 1
#> 
#>          [,1]      [,2]       [,3]      [,4]
#> [1,] 5.817280 1.9151625  6.2842745  7.183714
#> [2,] 1.915163 6.4124155  0.7649424  3.314741
#> [3,] 6.284275 0.7649424 27.3753125 24.839227
#> [4,] 7.183714 3.3147408 24.8392269 31.438256
#> 
#> , , 1, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.78764009  0.05542174 -0.16002149 -0.01651123
#> [2,]  0.05542174  1.16219136 -0.12578822 -0.20121776
#> [3,] -0.16002149 -0.12578822  1.50711686 -0.05071931
#> [4,] -0.01651123 -0.20121776 -0.05071931  1.32846995
#> 
#> , , 2, 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.5381686 -0.19395364  0.3873872 0.32781069
#> [2,] -0.1939536  1.14919296 -0.1896024 0.04851518
#> [3,]  0.3873872 -0.18960237  0.9287727 0.36971521
#> [4,]  0.3278107  0.04851518  0.3697152 0.97734460
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  1.3626616 -0.8075108  1.189798 0.5053831
#> [2,] -0.8075108 13.0082158 -5.949426 0.6518441
#> [3,]  1.1897982 -5.9494256  7.094887 3.0579840
#> [4,]  0.5053831  0.6518441  3.057984 4.6109167
#> 
#> , , 4, 2
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  1.04040388 -0.07784625 -0.12401824  0.01934152
#> [2,] -0.07784625  1.26549866 -0.05160837 -0.03763376
#> [3,] -0.12401824 -0.05160837  1.48638508  0.33524326
#> [4,]  0.01934152 -0.03763376  0.33524326  0.85057718
#> 
#> , , 5, 2
#> 
#>          [,1]      [,2]      [,3]     [,4]
#> [1,] 1.773850  1.044392  1.447673 2.371737
#> [2,] 1.044392  4.453101 -1.418904 1.187920
#> [3,] 1.447673 -1.418904  6.537586 5.047569
#> [4,] 2.371737  1.187920  5.047569 6.930865
#> 
#> , , 1, 3
#> 
#>             [,1]       [,2]        [,3]     [,4]
#> [1,]  3.21896341 -0.1803205 -0.04967198 0.718995
#> [2,] -0.18032052  3.3101237  0.79463187 0.876906
#> [3,] -0.04967198  0.7946319  6.14091275 3.218340
#> [4,]  0.71899498  0.8769060  3.21833989 8.808016
#> 
#> , , 2, 3
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  3.42358392 -1.4866881  0.01931306  0.8688558
#> [2,] -1.48668814  4.1572782 -0.52395021 -2.1187724
#> [3,]  0.01931306 -0.5239502  3.01377964  0.3467647
#> [4,]  0.86885580 -2.1187724  0.34676472  4.4172681
#> 
#> , , 3, 3
#> 
#>           [,1]      [,2]       [,3]     [,4]
#> [1,] 1.5628387  1.514520  0.8976073 1.590674
#> [2,] 1.5145203 16.624649 -6.3672758 4.001611
#> [3,] 0.8976073 -6.367276  8.5996191 2.929298
#> [4,] 1.5906742  4.001611  2.9292976 6.525489
#> 
#> , , 4, 3
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  1.70355511 -0.01160303  0.1338421  0.4059038
#> [2,] -0.01160303  1.45837936 -0.1363991 -0.1480913
#> [3,]  0.13384213 -0.13639914  3.3232016  0.5213587
#> [4,]  0.40590379 -0.14809125  0.5213587  2.4582710
#> 
#> , , 5, 3
#> 
#>           [,1]      [,2]      [,3]     [,4]
#> [1,]  2.734475 -0.197571  2.770441 2.080330
#> [2,] -0.197571  7.711113 -4.466874 2.467361
#> [3,]  2.770441 -4.466874 12.739637 3.692008
#> [4,]  2.080330  2.467361  3.692008 8.541158
#> 
#> , , 1, 4
#> 
#>             [,1]         [,2]        [,3]         [,4]
#> [1,] 0.797468841  0.002269078  0.08114882  0.048727139
#> [2,] 0.002269078  1.161867368 -0.01648086 -0.005318761
#> [3,] 0.081148823 -0.016480862  2.15869070  0.531379211
#> [4,] 0.048727139 -0.005318761  0.53137921  1.202792130
#> 
#> , , 2, 4
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  1.2199825 -0.21030621  0.19717635 -0.1644666
#> [2,] -0.2103062  1.51928608 -0.06937157  0.3095783
#> [3,]  0.1971763 -0.06937157  1.40067870  0.2502112
#> [4,] -0.1644666  0.30957832  0.25021124  1.5677201
#> 
#> , , 3, 4
#> 
#>            [,1]      [,2]      [,3]      [,4]
#> [1,]  2.6203130 -1.375847  1.802886 0.5266236
#> [2,] -1.3758473 12.158050 -5.518804 1.3107384
#> [3,]  1.8028856 -5.518804  7.941846 2.5042301
#> [4,]  0.5266236  1.310738  2.504230 5.0031274
#> 
#> , , 4, 4
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  2.48504871 0.16826891 0.45813445 -0.01397724
#> [2,]  0.16826891 1.68760445 0.02351736  0.30386514
#> [3,]  0.45813445 0.02351736 2.07165155  0.30289000
#> [4,] -0.01397724 0.30386514 0.30289000  2.52010127
#> 
#> , , 5, 4
#> 
#>           [,1]       [,2]       [,3]     [,4]
#> [1,] 1.2509939  0.7800114  0.7868223 1.354584
#> [2,] 0.7800114  5.1851291 -1.0215145 2.625029
#> [3,] 0.7868223 -1.0215145  4.9880841 3.821221
#> [4,] 1.3545844  2.6250293  3.8212208 7.259761
#> 
#> , , 1, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.2995496 -0.1022447  0.3220171  0.3331252
#> [2,] -0.1022447  1.7027646 -0.2340639 -0.9187180
#> [3,]  0.3220171 -0.2340639  2.6212811  0.6728140
#> [4,]  0.3331252 -0.9187180  0.6728140  3.3258106
#> 
#> , , 2, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  3.4123440 -0.4013158 -0.6252539 -1.0240275
#> [2,] -0.4013158  3.7762974  0.2428145  0.1562540
#> [3,] -0.6252539  0.2428145  2.9581708  0.1248335
#> [4,] -1.0240275  0.1562540  0.1248335  5.3992460
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  1.7592041 -0.5855452 -0.413437  0.4486322
#> [2,] -0.5855452 15.7560099 -7.011428 -1.9876164
#> [3,] -0.4134370 -7.0114285  7.487115  3.2782571
#> [4,]  0.4486322 -1.9876164  3.278257  6.0965786
#> 
#> , , 4, 5
#> 
#>            [,1]         [,2]         [,3]       [,4]
#> [1,]  2.9255404  0.226067336  0.418784598 -1.5058622
#> [2,]  0.2260673  1.784053115 -0.009397199 -0.5709156
#> [3,]  0.4187846 -0.009397199  2.088780249 -0.6086740
#> [4,] -1.5058622 -0.570915598 -0.608673989  3.0510291
#> 
#> , , 5, 5
#> 
#>           [,1]      [,2]       [,3]       [,4]
#> [1,] 2.4120205  1.372280  0.2463972  0.8773403
#> [2,] 1.3722801  8.300462 -3.4202482 -0.1479590
#> [3,] 0.2463972 -3.420248  5.7318132  3.2995459
#> [4,] 0.8773403 -0.147959  3.2995459  4.9900752
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>             [,1]       [,2]        [,3]         [,4]
#> [1,]  0.26617986  0.3785997  0.13139367 -0.007011951
#> [2,]  0.49481060  0.4750618 -0.04427937  0.211252187
#> [3,]  0.01905465 -0.3172185  1.08192863  0.910596587
#> [4,]  0.29988894  0.1709415 -0.03083783  0.297296310
#> [5,] -1.86368537  1.2047691 -4.66413308 -7.599663995
#> 
#> , , 2, 1
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.5839099 -1.7463892  1.5879723 -0.04420542
#> [2,]  1.2641568 -1.7821131  2.5418814 -0.86349488
#> [3,]  1.5788154  1.8867958 -0.1595963 -6.43873016
#> [4,] -1.4643446 -0.4868227 -0.0374221  6.45279613
#> [5,]  0.3539643  3.9971503 -1.8854091  5.99088058
#> 
#> , , 3, 1
#> 
#>            [,1]         [,2]       [,3]       [,4]
#> [1,]  0.8407042  0.003198072  0.5611605  0.6587342
#> [2,] -0.2925606  0.306302155 -0.8559980 -0.9772234
#> [3,] -0.4613777 -0.263460465 -0.7930269 -1.0333776
#> [4,]  0.5174832  0.433306523  1.4157291  1.8515329
#> [5,]  0.1669872 -1.558402080  6.1686140  2.4420826
#> 
#> , , 4, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.2888409 -0.1876799 -0.5420128  0.5126913
#> [2,] -1.0900450  1.7997801 -2.3892785 -1.4475181
#> [3,] -0.1216302  0.2230943 -0.9915615 -0.3858365
#> [4,]  0.2857377 -0.2717188  2.1561348  1.3332798
#> [5,]  8.4122354  7.0340878  4.3146317 -5.8667813
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  2.3293442  0.4544222  5.3312019  4.911471
#> [2,]  0.6518663  1.2598559  0.7611671  1.919884
#> [3,]  1.0081296  0.6266220  1.3421598  2.212157
#> [4,] -1.7175428 -0.8757714 -3.0050020 -3.781018
#> [5,]  6.8509759  4.5539556 12.9686055 18.598658
#> 
#> , , 1, 2
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.7433202  0.56897721 -0.59381856 -0.3469387
#> [2,]  0.2271928  0.51354379  0.07135208  0.2321313
#> [3,]  0.2205995 -0.42198913  0.81756379  0.2967346
#> [4,] -0.1024036  0.19714054  0.40601308  0.8530816
#> [5,] -0.3150522 -0.08945535  1.16165157 -0.2671414
#> 
#> , , 2, 2
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.4613548  0.5499030 -0.11439921 -0.1788174
#> [2,]  0.8112360  0.8733000  0.51717547  0.4211626
#> [3,]  0.1761542  0.2337925  0.80473310  0.2772757
#> [4,] -0.1666443 -0.3651768  0.09859607  0.7422512
#> [5,]  1.6941374  1.4991233 -0.65285264 -0.3848423
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  1.3140436 -0.1790629  1.278733  0.4489098
#> [2,] -0.3619453  0.4239861 -1.098450 -0.7715990
#> [3,] -0.5729859 -0.1370032 -1.074211 -1.0342656
#> [4,]  0.5037716  0.2719741  1.627416  1.9014646
#> [5,] -2.5785576  4.1760445 -5.520412 -0.3758520
#> 
#> , , 4, 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.7785061 -0.39131560  1.5808046  0.3899520
#> [2,]  0.5933394  0.90694582 -0.3664425 -0.3106126
#> [3,]  0.3415614 -0.04695878  0.3471738 -0.2275593
#> [4,] -0.2591711  0.16191716  0.1108965  1.0935073
#> [5,] -1.3178969  1.60861252 -0.3370449  0.3785353
#> 
#> , , 5, 2
#> 
#>             [,1]       [,2]       [,3]      [,4]
#> [1,]  1.32836283 -0.5812208  1.5379251 1.4447063
#> [2,]  0.13063002  1.3434488  0.4405328 0.3000551
#> [3,]  0.05747316  0.9564584  1.1780969 0.1862538
#> [4,] -0.21516998 -0.6660769 -1.0070747 0.0391530
#> [5,] -0.18713209 -0.2738576  5.5333924 5.8088203
#> 
#> , , 1, 3
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  1.28956401 -0.09105235  1.1147436  0.3607137
#> [2,] -0.50126052  1.04232891 -1.1391458 -0.4928494
#> [3,] -0.18899506 -0.26194795 -0.1139423  0.2927124
#> [4,]  0.03402605  0.31523632  0.6945663  0.5294255
#> [5,]  3.65510356 -1.54486455 -0.7438509  5.1788254
#> 
#> , , 2, 3
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.32539518 -0.2660938 -0.04557897 -0.1372109
#> [2,]  0.08451457  0.6579761  0.44541082  0.6183391
#> [3,]  0.50466835 -0.4194857  1.17746843  0.1160419
#> [4,] -0.25449060  0.5226353 -0.22107247  0.8329197
#> [5,]  2.52147939 -0.9463033 -0.47539579 -2.0210668
#> 
#> , , 3, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.0320726 -0.1486700  0.9901490  0.6294227
#> [2,] -0.3516786  0.2592382 -0.9895295 -0.9104680
#> [3,] -0.4979913 -0.3209261 -0.9103832 -1.1378304
#> [4,]  0.4796746  0.5858803  1.4241856  1.9414620
#> [5,]  1.2957905 -2.6565405  3.7843031  1.2064165
#> 
#> , , 4, 3
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.16275427 -0.37698198  1.2082599 1.221863718
#> [2,]  0.26193528  0.57327728  0.1747680 0.006209074
#> [3,]  0.02969698 -0.07618704  0.2874327 0.338786495
#> [4,]  0.26603133  0.18526246  0.3506116 0.249143600
#> [5,] -1.13122291  2.36255510 -5.2739884 0.406297845
#> 
#> , , 5, 3
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  1.7785027 -0.2335658  2.0443819  1.59377688
#> [2,]  0.2157430  1.3607001  0.3748440  0.06749516
#> [3,]  0.1280322  0.6646481  0.8924007 -0.38774263
#> [4,] -0.4945011 -0.5549127 -1.0189813  0.51377481
#> [5,]  0.3433428 -0.3417318  9.0976153  7.64169545
#> 
#> , , 1, 4
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.6756449 -0.1745788  0.1501708  0.49966658
#> [2,]  0.0859040  1.2139138 -0.1443487 -0.23065884
#> [3,] -0.4485004  0.1106803  1.4164466  0.80173075
#> [4,]  0.5786322 -0.1134903 -0.4827226  0.01093887
#> [5,] -0.9464919  3.1040890  1.6384285  1.73321585
#> 
#> , , 2, 4
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.7320006  0.3200122 -0.16633425  0.1025159
#> [2,]  0.3152095  0.5124246  0.09454223  0.6059726
#> [3,]  0.6044001 -0.2519872  0.90137307  0.3241940
#> [4,] -0.4841395  0.2144119  0.10778401  0.5699759
#> [5,] -0.3643965 -0.5348845  0.29229440 -0.1707561
#> 
#> , , 3, 4
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  1.2067454 -0.5584430  0.96783113  0.4472677
#> [2,] -0.1757480  0.5171854 -0.74435277 -0.7849592
#> [3,] -0.3809106 -0.2854727 -0.81805560 -1.0508911
#> [4,]  0.2615520  0.5460594  1.34291880  1.8612826
#> [5,]  1.6734752  2.2834369 -0.07428454  2.6167723
#> 
#> , , 4, 4
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  1.44796548 -0.07793136  0.5843134 -0.3307905
#> [2,]  0.02271899  0.45223755 -0.2055943 -0.1543019
#> [3,]  0.23395560  0.09353834  0.6845491  0.3287706
#> [4,] -0.33401192 -0.03204789  0.1073249  0.7997036
#> [5,] -3.34027463 -1.45567236  0.7798807 -0.3658054
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.3092046  0.6527851  0.4048220  1.1358763
#> [2,]  0.2177192  1.4459184  0.3316459  0.4694082
#> [3,]  0.2985964  0.8698232  1.3038826  0.5908335
#> [4,] -0.5030687 -1.1742891 -0.5623172 -0.1642476
#> [5,]  3.6791087  1.8681340  2.7937798  2.3973617
#> 
#> , , 1, 5
#> 
#>             [,1]        [,2]         [,3]        [,4]
#> [1,]  0.59343885  0.17916917 -0.275821074 -0.09288836
#> [2,] -0.01455619  0.71353470 -0.008719854 -0.05193422
#> [3,] -0.40858221  0.09758484  0.796580465  0.08394887
#> [4,]  0.50278173 -0.14178124  0.318293295  1.02831399
#> [5,]  3.68839931  1.30999912 -0.202444816 -2.38646665
#> 
#> , , 2, 5
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.04504855 -0.1530119 -0.3050071  0.2874411
#> [2,]  0.85705409  0.3448859 -0.1815032  0.1352252
#> [3,] -0.09639001  0.4131981  0.6340427  0.4555457
#> [4,] -0.08489753 -0.1668776  0.4407922  0.5182084
#> [5,] -1.97948473  0.7826742 -0.1134325 -1.5025361
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]       [,3]         [,4]
#> [1,]  0.9806030 -0.7179450 0.62603575 -0.177489796
#> [2,] -0.1464699  0.7041638 0.13884806 -0.001598335
#> [3,] -0.2824679  0.4972604 0.53732858  0.236402121
#> [4,]  0.2745275 -0.1234049 0.09285701  0.850272777
#> [5,]  0.5337217  2.9552022 0.97353838  0.640070928
#> 
#> , , 4, 5
#> 
#>             [,1]        [,2]         [,3]       [,4]
#> [1,]  0.78935182 -0.09819062 -0.282119299  0.2769206
#> [2,] -0.21838676  0.13709020 -0.006034121  0.4661252
#> [3,]  0.36369126 -0.58009894  0.860113430  0.3655139
#> [4,] -0.27489171  0.59775110  0.237341193  0.5309327
#> [5,]  0.09829925  2.46923128 -0.365186856 -0.2522294
#> 
#> , , 5, 5
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.3037939 -0.73180595  0.3335159  0.8235563
#> [2,]  0.1480798  1.22083978  0.8733392  0.6632491
#> [3,]  0.2294428 -0.06612921  2.4691625  0.9549265
#> [4,] -0.3333516  0.29521492 -1.6329524 -0.4106379
#> [5,] -1.9700635  2.81289646  0.9897270  4.3999488
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvars$get_last_draw`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|