R6 Class Representing PosteriorBVARs

The class PosteriorBVARs contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvars$new()

specify_posterior_bvars$get_posterior()

specify_posterior_bvars$get_last_draw()

specify_posterior_bvars$clone()

Method new()

Create a new posterior output PosteriorBVARs.

Usage

specify_posterior_bvars$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARs with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARs.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvars$get_posterior()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvars$get_last_draw()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvars$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARs" "R6"            

## ------------------------------------------------
## Method `specify_posterior_bvars$get_posterior`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $nu
#>      [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,]  5.1 5.100000 5.100000 5.100000 5.100000
#> [2,]  5.1 5.100000 5.088290 5.088290 5.088290
#> [3,]  5.1 5.053993 5.053993 5.053993 5.053993
#> [4,]  5.1 5.100000 5.025128 5.025128 5.025128
#> [5,]  5.1 5.100000 5.046106 5.046106 5.046106
#> 
#> $m
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.4266466 1.1431262 0.9805690 1.2806442 1.2301546
#> [2,] 1.1347053 1.1542571 0.9678268 0.7478020 0.7870473
#> [3,] 1.0981628 0.7569538 0.7704725 0.6604323 1.0789738
#> [4,] 0.3615120 0.7099913 0.5222068 0.9819775 0.5279750
#> [5,] 0.1057444 0.7752284 0.7278230 0.9547358 1.1673988
#> 
#> $w
#>            [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.19011050 0.22196906 0.3709378 0.18129204 0.08729231
#> [2,] 0.21337405 0.19549668 0.1127561 0.13494031 0.11168093
#> [3,] 0.11180305 0.14632129 0.1339854 0.07857831 0.04941308
#> [4,] 0.06935781 0.09545092 0.2273247 0.12016401 0.08518210
#> [5,] 0.03556293 0.06951306 0.1452484 0.11596260 0.22466527
#> 
#> $s
#>           [,1]     [,2]      [,3]      [,4]     [,5]
#> [1,] 137.67028 37.57971  73.58986  61.11930 73.52721
#> [2,]  68.74045 76.12363  53.73865 120.57988 38.46700
#> [3,]  69.22232 40.41551  57.06005  73.91995 44.85972
#> [4,]  93.01318 34.84179 118.57253  43.16877 59.14601
#> [5,] 106.80696 53.14444  46.10218  52.92022 74.15197
#> 
#> $scale
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
#> [4,]    0
#> [5,]    0
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>           [,1]      [,2]      [,3]       [,4]
#> [1,] 5.5637448 0.3495279  2.458820  4.3874575
#> [2,] 0.3495279 4.8600383  1.714408  0.2422879
#> [3,] 2.4588196 1.7144078 11.160834  8.5837676
#> [4,] 4.3874575 0.2422879  8.583768 12.5918117
#> 
#> , , 2, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.6185116 -0.7837884  0.5050367  0.2417051
#> [2,] -0.7837884  2.9696843 -0.6225026 -1.0360582
#> [3,]  0.5050367 -0.6225026  2.1102944  0.8366637
#> [4,]  0.2417051 -1.0360582  0.8366637  2.9480736
#> 
#> , , 3, 1
#> 
#>            [,1]       [,2]        [,3]      [,4]
#> [1,] 2.58201228  0.3870400  0.03381975 0.1441499
#> [2,] 0.38704000 16.3323947 -8.18413654 0.4318057
#> [3,] 0.03381975 -8.1841365  7.56491599 1.5023740
#> [4,] 0.14414993  0.4318057  1.50237400 4.2984767
#> 
#> , , 4, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  2.84673272 -0.5822555 -0.2863739  0.07598618
#> [2,] -0.58225552  3.3520580  0.1979500 -0.33052175
#> [3,] -0.28637389  0.1979500  3.5892940  0.27256890
#> [4,]  0.07598618 -0.3305217  0.2725689  3.48249143
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  3.9708285 -0.5490104  2.782053  3.1690323
#> [2,] -0.5490104  5.0201340 -1.762896 -0.5508701
#> [3,]  2.7820533 -1.7628961  7.800352  5.0423652
#> [4,]  3.1690323 -0.5508701  5.042365  9.3594055
#> 
#> , , 1, 2
#> 
#>           [,1]      [,2]      [,3]     [,4]
#> [1,] 2.4094430 0.2722746 2.2112288 3.082596
#> [2,] 0.2722746 1.4215984 0.6543554 1.318834
#> [3,] 2.2112288 0.6543554 4.9845156 5.811350
#> [4,] 3.0825956 1.3188336 5.8113498 9.155763
#> 
#> , , 2, 2
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  2.4966783 -0.49583228 -0.2751316 -0.30541272
#> [2,] -0.4958323  2.48794281 -0.2353884 -0.09018039
#> [3,] -0.2751316 -0.23538842  3.0422320  0.70835545
#> [4,] -0.3054127 -0.09018039  0.7083555  2.77576754
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  1.1523047  0.9243851 -0.6444467 0.4348702
#> [2,]  0.9243851  9.1367315 -5.2012052 1.2259271
#> [3,] -0.6444467 -5.2012052  6.3287804 1.1084094
#> [4,]  0.4348702  1.2259271  1.1084094 3.9756534
#> 
#> , , 4, 2
#> 
#>              [,1]         [,2]      [,3]       [,4]
#> [1,]  1.339473421 -0.005522867 0.7706458  0.3420775
#> [2,] -0.005522867  0.986102747 0.3922280 -0.2342349
#> [3,]  0.770645827  0.392228005 2.9523992  0.4795350
#> [4,]  0.342077502 -0.234234850 0.4795350  0.9862596
#> 
#> , , 5, 2
#> 
#>             [,1]        [,2]      [,3]       [,4]
#> [1,]  2.58338789 -0.07845119  1.796346  1.4748992
#> [2,] -0.07845119  6.87883833 -4.567503 -0.7237573
#> [3,]  1.79634576 -4.56750289  9.423004  5.6142648
#> [4,]  1.47489916 -0.72375734  5.614265  7.1045269
#> 
#> , , 1, 3
#> 
#>           [,1]       [,2]       [,3]       [,4]
#> [1,] 2.5964122  0.2108529  0.1888562  0.5347519
#> [2,] 0.2108529  1.8478374 -0.3001112 -0.0696465
#> [3,] 0.1888562 -0.3001112  4.3087682  1.0507786
#> [4,] 0.5347519 -0.0696465  1.0507786  3.4685767
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.2574967 -0.3027802  0.5699427  0.1064453
#> [2,] -0.3027802  1.4499302 -0.2424929 -0.2862536
#> [3,]  0.5699427 -0.2424929  1.7161848  0.4913224
#> [4,]  0.1064453 -0.2862536  0.4913224  1.3102774
#> 
#> , , 3, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.9363863 -1.1060904  0.5279275 -0.5819190
#> [2,] -1.1060904 13.8309842 -8.3193180  0.5176894
#> [3,]  0.5279275 -8.3193180  8.9536246  2.0675869
#> [4,] -0.5819190  0.5176894  2.0675869  4.0404443
#> 
#> , , 4, 3
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  4.29204000 -0.13506589 0.88770780 0.04593722
#> [2,] -0.13506589  4.50139331 0.03652307 0.23800087
#> [3,]  0.88770780  0.03652307 6.82856898 1.99079318
#> [4,]  0.04593722  0.23800087 1.99079318 3.84445344
#> 
#> , , 5, 3
#> 
#>           [,1]      [,2]      [,3]      [,4]
#> [1,]  3.381195 -1.155643  1.833438  1.681901
#> [2,] -1.155643  3.852133 -2.629599 -0.792769
#> [3,]  1.833438 -2.629599  6.916002  3.423362
#> [4,]  1.681901 -0.792769  3.423362  4.505789
#> 
#> , , 1, 4
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  1.63047670  0.09060886 -0.1713944  0.01921227
#> [2,]  0.09060886  1.34236668 -0.2585644 -0.43640688
#> [3,] -0.17139445 -0.25856436  2.1393432  0.23431838
#> [4,]  0.01921227 -0.43640688  0.2343184  2.45739170
#> 
#> , , 2, 4
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  2.9744429 -0.1268847 0.6493343 -0.7993486
#> [2,] -0.1268847  4.2111857 0.6595016  0.3222768
#> [3,]  0.6493343  0.6595016 4.4231232  0.6747691
#> [4,] -0.7993486  0.3222768 0.6747691  6.8338630
#> 
#> , , 3, 4
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  1.8118482  0.4256754 -1.053662 -0.2142933
#> [2,]  0.4256754 12.6503955 -7.528863  0.9065338
#> [3,] -1.0536617 -7.5288631 11.465321  2.4913346
#> [4,] -0.2142933  0.9065338  2.491335  5.2387983
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.5638411 -0.2401539  0.3992894  0.1369457
#> [2,] -0.2401539  1.5289280 -0.6481349 -0.2887437
#> [3,]  0.3992894 -0.6481349  1.8121435  0.0736951
#> [4,]  0.1369457 -0.2887437  0.0736951  1.7185402
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  2.4055286 -0.3297348  2.047219  1.9675066
#> [2,] -0.3297348  3.9569043 -1.688420 -0.2282814
#> [3,]  2.0472190 -1.6884204  4.850090  3.4310580
#> [4,]  1.9675066 -0.2282814  3.431058  4.5346373
#> 
#> , , 1, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.3052454 -0.1577407 -0.4438731 -0.3081220
#> [2,] -0.1577407  2.2017749 -0.4931634 -0.5117764
#> [3,] -0.4438731 -0.4931634  2.7005844  0.3904082
#> [4,] -0.3081220 -0.5117764  0.3904082  2.4158912
#> 
#> , , 2, 5
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  1.6580900 -0.3802090 0.1041129 -0.1568766
#> [2,] -0.3802090  1.5336120 0.1567670  0.3544304
#> [3,]  0.1041129  0.1567670 1.9384211  0.5327918
#> [4,] -0.1568766  0.3544304 0.5327918  2.2102596
#> 
#> , , 3, 5
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  1.11988421 -0.6437598  0.04779337 -0.3167788
#> [2,] -0.64375977  7.8221885 -3.96394684  0.1425987
#> [3,]  0.04779337 -3.9639468  5.92830652  1.6843738
#> [4,] -0.31677881  0.1425987  1.68437381  4.8195242
#> 
#> , , 4, 5
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  2.96069867  0.5360255 -0.7529794  0.07206919
#> [2,]  0.53602546  1.5629000 -0.5911456 -0.14802832
#> [3,] -0.75297936 -0.5911456  2.7047080  0.26748814
#> [4,]  0.07206919 -0.1480283  0.2674881  1.47099924
#> 
#> , , 5, 5
#> 
#>              [,1]       [,2]       [,3]         [,4]
#> [1,]  1.544319066 -0.6699535  0.1151932 -0.003256982
#> [2,] -0.669953504  4.8534056 -3.0979909 -0.612960867
#> [3,]  0.115193234 -3.0979909  8.6618238  3.193826391
#> [4,] -0.003256982 -0.6129609  3.1938264  4.668127607
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,] -0.4540552  0.0756080 -2.273542 -2.342117
#> [2,]  1.5592092  1.1491407  2.415330  2.497060
#> [3,]  1.5589573  0.1396857  3.865388  3.457510
#> [4,] -1.2393618 -0.3083925 -2.203920 -1.829929
#> [5,] 11.3688844  4.5414693 10.273295 15.764655
#> 
#> , , 2, 1
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  1.2224189 -0.7039647  0.38922153 0.44688370
#> [2,]  0.2909520  0.2411160  0.72585458 0.94210862
#> [3,]  0.5562151 -0.3248463  1.07815730 0.66958931
#> [4,] -0.5821092  0.6368311 -0.34459722 0.07586819
#> [5,] -1.8267751  1.6665035 -0.02574874 1.16772328
#> 
#> , , 3, 1
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.0077241  0.07229194  0.8318057  0.6561781
#> [2,]  0.2482701  0.10223221 -0.2275683 -0.2251865
#> [3,]  0.4763439 -0.49538007  0.2467881 -0.2555159
#> [4,] -0.4412168  0.71329583  0.3797118  1.0714172
#> [5,] -1.5965599 -5.74265988 -0.2356227 -3.2241159
#> 
#> , , 4, 1
#> 
#>              [,1]        [,2]       [,3]        [,4]
#> [1,] -0.022045732  0.16324363 -0.5246737 -0.07825518
#> [2,]  0.236874883  0.61758440  0.0893922  0.43744933
#> [3,]  0.004876353 -0.03239747  0.4937438 -0.05289670
#> [4,]  0.330635111 -0.02654488  0.6425811  1.04249950
#> [5,]  0.753926894  1.20661078  1.9497662  1.51555746
#> 
#> , , 5, 1
#> 
#>              [,1]         [,2]       [,3]       [,4]
#> [1,] -0.668984931 -0.032802970 -1.5464737 -1.4830616
#> [2,]  0.005457423  0.893166702  0.1474155  0.3553747
#> [3,]  0.291987653  0.061845879  1.0902487  0.7217845
#> [4,]  0.433817778 -0.002075923  0.5436363  0.8905117
#> [5,]  2.771906205 -1.293553715  1.8134826  3.5994255
#> 
#> , , 1, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.7392541  0.1001417 -0.5110229 -0.4862646
#> [2,]  0.8279277  1.2296980  1.1537761  1.3760058
#> [3,]  2.0697452  0.7394976  4.5798163  4.4612674
#> [4,] -2.0488295 -0.7759531 -3.4236542 -3.3593284
#> [5,]  6.2349069  0.4735444  8.0875175 11.8911648
#> 
#> , , 2, 2
#> 
#>            [,1]       [,2]         [,3]        [,4]
#> [1,]  0.5982622  0.7809438 -0.177654144 -0.06241688
#> [2,]  0.1572561  0.3594557  0.517793222  1.00317798
#> [3,]  0.5345290 -0.9337204  0.999098240  0.34407707
#> [4,] -0.4005679  0.6836167 -0.006070339  0.52691804
#> [5,]  3.8669267 -1.2471168 -3.239904896 -0.56599349
#> 
#> , , 3, 2
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.8064742 -0.11904130  0.8345854  0.47486654
#> [2,]  0.2507692  0.51994777 -0.3969341 -0.11414656
#> [3,]  0.3202382  0.22035365 -0.1989681  0.07182029
#> [4,] -0.2512071  0.05731241  0.7576217  0.87947792
#> [5,] -0.6486140 -3.54116558  1.5926693 -5.64499329
#> 
#> , , 4, 2
#> 
#>             [,1]       [,2]         [,3]       [,4]
#> [1,] -0.14232887  0.0287134 -0.909668671 -0.5075442
#> [2,] -0.12290841  0.9508758  0.552377907  0.1882175
#> [3,]  0.05813791  0.2646461  1.275804975  0.4984384
#> [4,]  0.31995844 -0.2546174 -0.006332768  0.6481354
#> [5,]  1.50633968 -1.0464669  2.181131196  2.6877764
#> 
#> , , 5, 2
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,] -0.6259144  0.03411003 -1.48777769 -1.3120115
#> [2,]  0.0456838  0.37954456  0.09050135 -0.2175329
#> [3,]  0.2278020 -0.72312961  1.01295635 -0.3673770
#> [4,]  0.4692186  0.79884587  0.57032503  1.9292430
#> [5,]  2.2708385 -2.92583814  3.66078736  1.9063091
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.1381193  0.1351137 -0.5151590 -0.6329105
#> [2,]  0.1193042  1.1151448  0.5724058  0.5840387
#> [3,]  1.1300322  0.7252691  2.5690342  1.0613294
#> [4,] -1.1328541 -0.7104561 -1.3989857  0.1936423
#> [5,]  0.7790376 -1.7664705  4.2175547  0.8643778
#> 
#> , , 2, 3
#> 
#>             [,1]         [,2]        [,3]       [,4]
#> [1,]  0.91436982 -0.003849529 -0.02607664 -0.1065141
#> [2,] -0.04935998  0.747716262  0.17131575  0.3969238
#> [3,] -0.77453558 -0.152464884  1.18143410  0.2745411
#> [4,]  0.72423740  0.156328104 -0.18768990  0.7023908
#> [5,] -2.75660860  1.862104577  0.82128469  1.2774590
#> 
#> , , 3, 3
#> 
#>            [,1]        [,2]         [,3]       [,4]
#> [1,]  0.5847005 -0.11681448  0.623178884  0.2979701
#> [2,]  0.5278187  0.03593713  0.053842806 -0.1959380
#> [3,]  0.7367350 -0.57381126  0.662098959 -0.1163263
#> [4,] -0.6159999  0.83768483 -0.003511115  1.0101790
#> [5,]  0.3729902 -2.39396767  1.082784629  1.7326457
#> 
#> , , 4, 3
#> 
#>             [,1]         [,2]        [,3]       [,4]
#> [1,] -0.06215806  0.290634787 -0.35872326 -0.0468620
#> [2,] -0.20263433  0.008567227  0.09779963 -0.4133459
#> [3,] -0.24807190 -0.419214503  0.20759662 -0.6342748
#> [4,]  0.69284667  0.339493509  0.86871338  1.6637477
#> [5,] -6.10691668  0.013343759  1.97706895 -1.0821143
#> 
#> , , 5, 3
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,] -0.5769979 -0.02352979 -1.20122601 -1.1681685
#> [2,] -0.1338551  1.07262958 -0.00638048  0.2771286
#> [3,] -0.1230784  0.41316619  0.73601900  0.2658944
#> [4,]  0.7512344 -0.34958513  0.70553817  1.1960882
#> [5,]  5.0335636 -1.58724983  3.87837817  2.6919263
#> 
#> , , 1, 4
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  1.28874332 -0.3908556  0.2537475  0.29612750
#> [2,] -0.25611768  1.4909771 -0.2214799 -0.53730458
#> [3,] -0.13841002  0.0421799  1.2804951 -0.05226211
#> [4,]  0.03754734  0.1028703 -0.3599634  1.02368730
#> [5,] -0.35916067 -1.5949196  0.4270258 -0.88536943
#> 
#> , , 2, 4
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  0.9432706 -0.03505686  0.03972535  0.55473458
#> [2,]  0.6830681  0.75555315  1.32079433  0.24508131
#> [3,]  0.2610190 -0.04113894  1.51411319  0.98992516
#> [4,] -0.3569410  0.09945179 -0.67078492 -0.06715606
#> [5,]  0.9710060  0.62688712 -3.37808890  0.93698875
#> 
#> , , 3, 4
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.82113405 -0.42113023  0.57392840 -0.01190999
#> [2,]  0.02986512  0.47621867  0.03044489 -0.13995737
#> [3,]  0.12069037 -0.05709787  0.56587542  0.04010281
#> [4,] -0.03027858  0.30945567  0.23263936  1.00860961
#> [5,] -0.69698073  3.55053944 -5.89366260 -0.10132990
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] 0.34976931  0.4733508 -0.3935990 -0.2680767
#> [2,] 0.80480562  0.5024115  0.5106439  0.5404968
#> [3,] 0.11854833 -0.2695851  0.6472158  0.4399546
#> [4,] 0.06707683  0.1358493  0.4514942  0.6458968
#> [5,] 1.08627772 -1.2920960  0.6854625 -0.2376829
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.4637110 -0.1905227 -0.9983515 -1.2125955
#> [2,]  0.5437193  0.5636977  0.6179972  0.5578052
#> [3,]  1.0720171 -0.2492564  1.8151088  1.0435758
#> [4,] -0.4306041  0.4167445 -0.4445432  0.4757659
#> [5,]  1.8070364 -1.9411591  3.2000744  2.3855355
#> 
#> , , 1, 5
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  1.16247116 -0.09545064 -0.26730153  0.1413606
#> [2,] -0.18746809  1.43753299  0.06366867 -0.0636936
#> [3,]  0.45329144  0.40198683  1.21797985  0.3200313
#> [4,] -0.45508446 -0.40118994 -0.05530774  0.6524328
#> [5,]  0.04361028 -0.11581778 -0.91472786 -0.0215026
#> 
#> , , 2, 5
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  1.16332025 -0.48118317  0.1199230 -0.3491581
#> [2,] -0.06049193  0.76384701  0.4551677  1.1230462
#> [3,] -0.10157513  0.33181053  1.4611059  1.3819937
#> [4,]  0.03222843 -0.06503155 -0.5492038 -0.2839748
#> [5,]  1.13001088 -0.35622184  1.8855921 -0.2943908
#> 
#> , , 3, 5
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  0.92060103 -0.5899764  0.73856626  0.14127127
#> [2,]  0.07081060  0.4954949  0.08310789 -0.03100034
#> [3,]  0.10505221  0.1826194  0.51333587  0.11645696
#> [4,] -0.06971621  0.2003972  0.12625277  0.87745270
#> [5,] -0.08231128 -0.4453281 -1.03664487 -1.19480190
#> 
#> , , 4, 5
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,] -0.03328128  0.12461725 -0.4537261 -0.02833387
#> [2,]  0.21261664  0.33763333  0.4313092  0.67988402
#> [3,]  0.06301811  0.06979263  0.8603304  0.12467839
#> [4,]  0.31522401 -0.10198864  0.2645810  0.86775766
#> [5,] -2.42963263  0.75660096  1.3857512  0.04112332
#> 
#> , , 5, 5
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.7443876 -0.7621541 -0.1353438 -0.07024492
#> [2,] -0.1659424  0.7063044  0.7727836  0.12504173
#> [3,] -0.1362185  0.1355597  1.7471668  0.19673213
#> [4,]  0.2936423  0.2633354 -0.7335664  0.81860762
#> [5,] -2.0535759  0.6607448 -1.2058032  0.93602809
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvars$get_last_draw`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|