R6 Class Representing PosteriorBVARs

The class PosteriorBVARs contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvars$new()

specify_posterior_bvars$get_posterior()

specify_posterior_bvars$get_last_draw()

specify_posterior_bvars$clone()

Method new()

Create a new posterior output PosteriorBVARs.

Usage

specify_posterior_bvars$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARs with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARs.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvars$get_posterior()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvars$get_last_draw()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvars$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARs" "R6"            

## ------------------------------------------------
## Method `specify_posterior_bvars$get_posterior`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $nu
#>      [,1] [,2] [,3] [,4]     [,5]
#> [1,]  5.1  5.1  5.1  5.1 5.100000
#> [2,]  5.1  5.1  5.1  5.1 5.100000
#> [3,]  5.1  5.1  5.1  5.1 5.012686
#> [4,]  5.1  5.1  5.1  5.1 5.100000
#> [5,]  5.1  5.1  5.1  5.1 5.100000
#> 
#> $m
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1329308 0.9655289 0.7218623 1.1322474 0.9339704
#> [2,] 0.2928495 0.5174386 0.6174691 1.1313118 0.8877731
#> [3,] 1.2712393 1.4445221 0.8778843 1.1982708 1.3392336
#> [4,] 0.4904019 0.9217065 0.9770706 0.6694024 0.5295832
#> [5,] 0.6313116 0.8423154 0.9256168 0.6801884 0.6644997
#> 
#> $w
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.05055630 0.14158761 0.10183953 0.09049902 0.18965696
#> [2,] 0.11260873 0.04075269 0.10267833 0.10049846 0.10465471
#> [3,] 0.42511309 0.11979257 0.09403462 0.06303240 0.06583463
#> [4,] 0.11308057 0.08844309 0.12117714 0.09437331 0.08893491
#> [5,] 0.08607951 0.09726960 0.17895515 0.23596683 0.23572457
#> 
#> $s
#>           [,1]     [,2]      [,3]     [,4]     [,5]
#> [1,] 498.85861 20.15818 113.75529 39.38333 84.73904
#> [2,] 411.03592 31.12115 106.32018 58.51929 53.51531
#> [3,] 168.59752 47.35422  79.23888 42.92635 46.22140
#> [4,]  75.82511 70.08236  55.42057 50.82512 54.63130
#> [5,]  73.56684 58.91985  50.90445 56.03662 70.72556
#> 
#> $scale
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
#> [4,]    0
#> [5,]    0
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] 16.4149789 -1.5377578 -0.7764206 -0.8933836
#> [2,] -1.5377578 15.8721124  2.6125868 -0.7335203
#> [3,] -0.7764206  2.6125868 20.3505137  3.1968439
#> [4,] -0.8933836 -0.7335203  3.1968439 11.6314162
#> 
#> , , 2, 1
#> 
#>            [,1]      [,2]       [,3]      [,4]
#> [1,] 20.0641417 -1.645026  0.2078544  2.410570
#> [2,] -1.6450265 10.519655  3.3803826  3.617331
#> [3,]  0.2078544  3.380383 15.1332858  1.423795
#> [4,]  2.4105701  3.617331  1.4237952 10.491483
#> 
#> , , 3, 1
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  6.5163222 -1.852220  2.1586687 -0.3374158
#> [2,] -1.8522203 12.345171 -5.3859705 -1.7025492
#> [3,]  2.1586687 -5.385970  9.2788055  0.4172644
#> [4,] -0.3374158 -1.702549  0.4172644  6.4753059
#> 
#> , , 4, 1
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  2.7118958 -0.21954569 -0.24829435 -0.12854098
#> [2,] -0.2195457  2.19098994 -0.02697894  0.60847713
#> [3,] -0.2482944 -0.02697894  2.15364242  0.05219036
#> [4,] -0.1285410  0.60847713  0.05219036  2.18272541
#> 
#> , , 5, 1
#> 
#>            [,1]        [,2]      [,3]      [,4]
#> [1,] 3.28355869  0.05473635  1.554065  1.933962
#> [2,] 0.05473635  8.09947372 -6.384010 -2.063737
#> [3,] 1.55406535 -6.38401029 11.066604  4.646684
#> [4,] 1.93396212 -2.06373690  4.646684  4.896401
#> 
#> , , 1, 2
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.618235684 -0.04678705 -0.06179297 -0.006282505
#> [2,] -0.046787051  0.88716472 -0.31997682 -0.151106125
#> [3,] -0.061792966 -0.31997682  0.91560836  0.478264803
#> [4,] -0.006282505 -0.15110612  0.47826480  1.367722061
#> 
#> , , 2, 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.9408721 -0.23046916 -0.1422764 0.15397779
#> [2,] -0.2304692  1.09091134 -0.3060840 0.03863042
#> [3,] -0.1422764 -0.30608401  1.2909563 0.21020542
#> [4,]  0.1539778  0.03863042  0.2102054 1.65108785
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.6933395 -0.8528261  0.3329776 -0.2512176
#> [2,] -0.8528261 13.7938899 -6.0632399  3.2996358
#> [3,]  0.3329776 -6.0632399  5.4700084  0.6091345
#> [4,] -0.2512176  3.2996358  0.6091345  4.8172426
#> 
#> , , 4, 2
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  1.75887058 0.02374568 -0.53196740 -0.10912031
#> [2,]  0.02374568 2.52003979  0.15000750  0.54785799
#> [3,] -0.53196740 0.15000750  2.04314703 -0.02078044
#> [4,] -0.10912031 0.54785799 -0.02078044  2.52561639
#> 
#> , , 5, 2
#> 
#>            [,1]        [,2]      [,3]        [,4]
#> [1,]  2.3162570 -0.64949459  1.989863  1.00031215
#> [2,] -0.6494946  4.01719705 -2.131554 -0.08657929
#> [3,]  1.9898629 -2.13155379  6.891058  3.67469641
#> [4,]  1.0003121 -0.08657929  3.674696  5.36893337
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.1365514  0.6134259 -0.8880071 -0.2666223
#> [2,]  0.6134259  3.8384230 -1.1824696 -0.5197036
#> [3,] -0.8880071 -1.1824696  4.0488632  0.8370492
#> [4,] -0.2666223 -0.5197036  0.8370492  2.4541617
#> 
#> , , 2, 3
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  3.43301904 -0.02354933 -0.93648437 -0.4527807
#> [2,] -0.02354933  2.39499420 -0.03214092  0.4781953
#> [3,] -0.93648437 -0.03214092  2.82323462  0.1975511
#> [4,] -0.45278073  0.47819530  0.19755110  6.2532699
#> 
#> , , 3, 3
#> 
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  2.3436672 -1.522581 -0.3069659 -1.106361
#> [2,] -1.5225812 12.815676 -4.2219251  1.869602
#> [3,] -0.3069659 -4.221925  5.7418471  1.386270
#> [4,] -1.1063606  1.869602  1.3862697  4.343937
#> 
#> , , 4, 3
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  1.44583350 -0.2569978 -0.1242298 -0.07721552
#> [2,] -0.25699780  2.2842649  0.2083807  0.36996769
#> [3,] -0.12422978  0.2083807  1.8704295  0.31413006
#> [4,] -0.07721552  0.3699677  0.3141301  2.44024996
#> 
#> , , 5, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  3.4440744  1.1361910 -0.6654861  0.2151693
#> [2,]  1.1361910  4.7024984 -2.5825998 -0.7188395
#> [3,] -0.6654861 -2.5825998  5.3524924  1.9278498
#> [4,]  0.2151693 -0.7188395  1.9278498  3.1244445
#> 
#> , , 1, 4
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,] 1.19016219  0.04094015  0.14898954  0.1565683
#> [2,] 0.04094015  1.49746814 -0.06349539 -0.1306994
#> [3,] 0.14898954 -0.06349539  1.36008661  0.2238794
#> [4,] 0.15656828 -0.13069936  0.22387941  1.5084153
#> 
#> , , 2, 4
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  1.6953860 -0.1714602 -0.4404295 0.1139040
#> [2,] -0.1714602  1.4290693  0.1238300 0.4966484
#> [3,] -0.4404295  0.1238300  2.9790333 0.6117476
#> [4,]  0.1139040  0.4966484  0.6117476 4.2927397
#> 
#> , , 3, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.1371664  0.2506003 -0.4867261 -0.3098372
#> [2,]  0.2506003 12.6783607 -8.0495266  0.2391891
#> [3,] -0.4867261 -8.0495266  8.3358158  2.0560599
#> [4,] -0.3098372  0.2391891  2.0560599  4.8880563
#> 
#> , , 4, 4
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  1.89444231 -0.28381508 -0.48857774  0.04371304
#> [2,] -0.28381508  1.82261814 -0.08763781 -0.63883762
#> [3,] -0.48857774 -0.08763781  2.00137559  0.98471270
#> [4,]  0.04371304 -0.63883762  0.98471270  2.52086573
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.3888469 -0.4760012  0.3248469 -0.1428078
#> [2,] -0.4760012  4.5485097 -2.3596486 -0.2933679
#> [3,]  0.3248469 -2.3596486  4.6747139  1.4849995
#> [4,] -0.1428078 -0.2933679  1.4849995  2.1691303
#> 
#> , , 1, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.7893494  0.2461768  0.1955512 -0.3260275
#> [2,]  0.2461768  2.9406311 -0.3915668 -0.2949718
#> [3,]  0.1955512 -0.3915668  3.4370116  0.2612689
#> [4,] -0.3260275 -0.2949718  0.2612689  2.5119764
#> 
#> , , 2, 5
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  1.1999170 -0.1820669 -0.1340955 0.1323901
#> [2,] -0.1820669  1.4875549  0.1175739 0.4771650
#> [3,] -0.1340955  0.1175739  1.5138623 0.5675163
#> [4,]  0.1323901  0.4771650  0.5675163 2.6821207
#> 
#> , , 3, 5
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  1.3964248 -1.571738  0.3592941  0.2647801
#> [2,] -1.5717379 10.994838 -6.2280132 -1.0356141
#> [3,]  0.3592941 -6.228013  7.7680302  1.4902438
#> [4,]  0.2647801 -1.035614  1.4902438  3.6186153
#> 
#> , , 4, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.9953946  0.4573257 -0.4448410 -0.4048262
#> [2,]  0.4573257  1.6208202 -0.1105913  0.2436194
#> [3,] -0.4448410 -0.1105913  2.0235859  0.5146038
#> [4,] -0.4048262  0.2436194  0.5146038  2.3433392
#> 
#> , , 5, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.8166638 -0.1037176  0.3291528  1.1538033
#> [2,] -0.1037176  3.8234347 -1.5787700 -0.9336189
#> [3,]  0.3291528 -1.5787700  3.8974877  1.2288946
#> [4,]  1.1538033 -0.9336189  1.2288946  3.7563444
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -1.6225796 -1.1123764 -0.6143965  1.0307167
#> [2,]  1.8030427  1.4197644  0.7977665 -0.2883491
#> [3,]  0.4848786  0.0997163  0.8085150  0.4383684
#> [4,]  0.4452595  0.2825303  0.4076771  0.2135408
#> [5,]  4.6092719  4.3529970 -2.4740426 -3.3790332
#> 
#> , , 2, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  1.07214490 -0.3897462 -0.4570104  0.16179104
#> [2,] -1.20845404  0.2627656 -0.6685090  0.30006618
#> [3,] -0.67172498  0.4530483  0.6519425  1.06524038
#> [4,]  0.78761725 -0.1194501  0.5724588  0.01626423
#> [5,] -0.03056958  1.7329528 -0.1000840 -1.46620346
#> 
#> , , 3, 1
#> 
#>             [,1]        [,2]       [,3]         [,4]
#> [1,]  0.08693719  0.19875325  0.2885183  0.590514476
#> [2,]  0.75191227  0.25941210  0.2992486  0.116059926
#> [3,]  1.02200640  0.22411259  0.6235023  0.003127583
#> [4,] -0.74583195  0.01301488  0.1204825  0.816424878
#> [5,]  2.69228350 -3.69780604 -0.9409386 -4.751349529
#> 
#> , , 4, 1
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.99849842  0.23863972  0.4749306  0.9415175
#> [2,]  0.15613778  0.02436024 -0.7640818 -0.4279371
#> [3,] -0.07475220 -0.08123783  0.2648025  0.3260830
#> [4,]  0.05138793  0.03424402  0.5747409  0.3883894
#> [5,]  1.22736328 -0.67544269  0.3889108 -0.9761063
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -0.1766509 -0.8304852 -0.2639758 -0.5173782
#> [2,] -0.2467058  1.6087136 -1.1555319 -0.3790079
#> [3,] -0.3090200  1.0506562 -0.8606750 -0.5606000
#> [4,]  0.8270848 -0.7402508  2.0153292  1.7901181
#> [5,]  0.4331812  3.0732104 -1.2061229  0.3219993
#> 
#> , , 1, 2
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.48972029  0.1332656  0.08021458  0.1155335
#> [2,]  0.33746862  0.8409847 -0.22490344  0.1363466
#> [3,]  0.07645706 -0.1871460  0.78115972  0.7202024
#> [4,]  0.12416863  0.1745494  0.17160089  0.2752268
#> [5,] -0.09680539 -1.5498423  0.86820029 -1.1881207
#> 
#> , , 2, 2
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.59244655  0.08362577  0.22155770  0.4261251
#> [2,] -0.15514778  0.53453474  0.05373654 -0.3171254
#> [3,]  0.02156804  0.42719591  0.74762135  0.6512262
#> [4,]  0.15395568 -0.28399530  0.12924191  0.3848221
#> [5,] -0.33001298  0.19734532 -0.10907508  0.2832636
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.6884321 -0.5733967  0.5205479 0.2306889
#> [2,]  0.4140468  0.7101360  0.3669229 0.3066958
#> [3,]  0.5060352  0.5420130  1.0675196 0.5285951
#> [4,] -0.4214507 -0.1261252 -0.3274664 0.3729826
#> [5,]  0.5614564 -1.4317504 -2.0908557 0.4241209
#> 
#> , , 4, 2
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.24842896  0.3263917  0.7898740  1.0356752
#> [2,]  0.33602285  0.8608812 -0.3018418 -0.2672795
#> [3,] -0.02522162  0.4996385  0.4029726  0.4341163
#> [4,] -0.06606549 -0.6215680  0.2956086  0.2411909
#> [5,] -0.46998690  1.9441397  2.1596463 -0.7353334
#> 
#> , , 5, 2
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,] -0.1815131 -0.59415044 -1.15068152 -0.75957710
#> [2,] -0.1324372  0.87546816 -0.06245823  0.01149628
#> [3,] -0.2316013  0.43289637  0.39179759 -0.14122824
#> [4,]  0.7242514 -0.02743386  0.94446815  1.38213673
#> [5,]  1.1860736 -4.47497341  6.50045425  4.65041944
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.7832305  0.2435422 -0.47566421  0.5209286
#> [2,]  0.3810292  0.9648350 -0.02406065 -0.6285274
#> [3,]  0.2565671  0.1840789  0.33208940  0.2117494
#> [4,] -0.1620762 -0.2558947  0.83026935  0.6447404
#> [5,] -1.4778983 -1.2518449  0.82688966  0.9975178
#> 
#> , , 2, 3
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.13939690 -0.1151149 -0.1354879  0.2613252
#> [2,]  0.05011019  0.6923653 -0.2113199 -0.6548837
#> [3,] -0.14326279  0.5241007  0.7927284  0.5855789
#> [4,]  0.04082115 -0.3478251  0.2869619  0.6198990
#> [5,]  1.98157125  0.5921550 -1.8907237 -3.8358886
#> 
#> , , 3, 3
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.9998934 -0.81470124  0.6592390 -0.02673414
#> [2,] -0.2589173  0.60464969  0.2434525  0.36831909
#> [3,] -0.3609556  0.56018177  0.7721380  0.74449369
#> [4,]  0.3753087  0.02159191 -0.1356073  0.31023629
#> [5,] -0.1986027 -5.03698555  1.0047623 -1.25608614
#> 
#> , , 4, 3
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.7464171  0.01561701  1.12776513  0.9845189
#> [2,]  0.2131778  0.61160048 -0.09624835 -0.3073436
#> [3,]  0.5980596 -0.23346520  0.64605753 -0.5406307
#> [4,] -0.5058817  0.21704048 -0.04308003  1.1811459
#> [5,]  0.2647398  1.27389901  0.24761012  1.5956790
#> 
#> , , 5, 3
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.6414606 -0.59286515  0.2890546 -0.02025588
#> [2,] -0.5967427  0.99062014 -0.4827219 -0.11335568
#> [3,] -1.0033484  0.27763274 -0.2599329 -0.34430991
#> [4,]  1.1187664  0.08315953  1.0027098  1.30693450
#> [5,]  2.8369165 -3.57293443  4.2576395  1.70647200
#> 
#> , , 1, 4
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  1.00661627  0.09177729 -0.0444472  0.10548750
#> [2,] -0.09978225  1.13848215 -0.1062565 -0.08135058
#> [3,] -0.09314923  0.24531171  0.5593642  0.49208190
#> [4,]  0.10147025 -0.32526151  0.4725719  0.50321835
#> [5,]  0.05546989  1.49066958 -1.4986052  0.36814981
#> 
#> , , 2, 4
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9817717 -0.09626852  0.3377107  0.1093920
#> [2,]  0.3332244  0.55465012 -0.2259486 -0.1335961
#> [3,]  0.0422575  0.50506754  0.7057800  0.7848409
#> [4,] -0.1240575 -0.31557493  0.2253077  0.3285807
#> [5,]  1.1421237  1.57675236 -2.9501179  1.1357669
#> 
#> , , 3, 4
#> 
#>              [,1]       [,2]       [,3]       [,4]
#> [1,]  0.848467607 -0.7986174 0.45994460  0.3081219
#> [2,]  0.019869341  1.3134012 0.08202898 -0.1045840
#> [3,]  0.070997457  1.5663437 0.38532910 -0.1698463
#> [4,]  0.005027841 -1.0834096 0.31340145  1.0196352
#> [5,] -0.669519572 -1.5975139 1.23081197  2.4000168
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.5204406  0.4266268  0.85272973  1.0576159
#> [2,]  0.1432689  0.3330256 -0.57729393 -0.1082619
#> [3,]  0.2932289 -0.2483417  0.68252401  0.1224822
#> [4,] -0.1302956  0.1371944  0.03836603  0.5089093
#> [5,]  0.6167821 -1.3972495 -0.31722901  1.0992839
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.8253868 -0.1112515  0.2829376  0.07881672
#> [2,] -0.4063484  0.4453722 -0.1594332  0.05337285
#> [3,] -0.6621234 -0.3680370  0.1498792 -0.02069457
#> [4,]  0.7055405  0.4662056  0.5603121  0.95165556
#> [5,]  2.0265906  0.6656895  6.1830657  1.46522145
#> 
#> , , 1, 5
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  1.05403287  0.05448737  0.9284921  0.29936422
#> [2,] -0.01426091  1.23697552 -1.2008716 -0.56752431
#> [3,]  0.12820129  0.40124155  0.2510529  0.08622183
#> [4,] -0.11107095 -0.47529622  0.4351668  0.76434906
#> [5,] -0.95371610  1.72638040  0.4428883  4.68804237
#> 
#> , , 2, 5
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.6820987 -0.3477825 -0.03417076 -0.3126096
#> [2,]  0.3306535  1.2025007 -0.07062009  0.3920667
#> [3,]  0.1893387  0.5158786  0.67317736  0.8182862
#> [4,] -0.1099928 -0.3768379  0.30898448  0.3416198
#> [5,] -1.2033221  1.0876968 -0.43713702 -0.6673425
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.0807230 -0.2688981  0.3915051  0.2861453
#> [2,] -0.2399873  0.4131383  0.3730314 -0.2632237
#> [3,] -0.3032859  0.1475097  1.0864785 -0.2994327
#> [4,]  0.2742583  0.1309936 -0.3397117  1.1676560
#> [5,]  0.8840323 -0.3273810  1.3894955  2.6517951
#> 
#> , , 4, 5
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.7874361  0.45502742  0.57490074  1.0628107
#> [2,]  0.6122938  0.64533168 -0.85724698 -0.1855498
#> [3,]  0.2848397 -0.18837006 -0.01190328 -0.3552314
#> [4,] -0.2451102  0.04921587  0.79862804  0.9550029
#> [5,]  1.5285175 -0.70726436  1.55824563  2.8737372
#> 
#> , , 5, 5
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  1.8575437 -1.38087947  0.89276851  0.8673738
#> [2,] -0.4218741  0.67753081  0.16495622 -0.3076465
#> [3,] -0.7741725 -0.02850317  0.55416929 -0.6786707
#> [4,]  0.3536977  0.67008073 -0.05003207  1.2588189
#> [5,]  2.2675591  1.72220259  1.49463890  1.4673616
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvars$get_last_draw`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|