R6 Class Representing PosteriorBVARs

The class PosteriorBVARs contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvars$new()

specify_posterior_bvars$get_posterior()

specify_posterior_bvars$get_last_draw()

specify_posterior_bvars$clone()

Method new()

Create a new posterior output PosteriorBVARs.

Usage

specify_posterior_bvars$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARs with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARs.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvars$get_posterior()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvars$get_last_draw()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvars$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARs" "R6"            

## ------------------------------------------------
## Method `specify_posterior_bvars$get_posterior`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $nu
#>      [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,]  5.1 5.108052 5.108052 5.108052 5.035413
#> [2,]  5.1 5.100000 5.100000 5.100000 5.100000
#> [3,]  5.1 5.100000 5.100000 5.100000 5.100000
#> [4,]  5.1 5.100000 5.100000 5.100000 5.100000
#> [5,]  5.1 5.100000 5.100000 5.100000 5.100000
#> 
#> $m
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.03258962 1.2214315 0.8024290 1.1505021 0.8488256
#> [2,] 0.70344328 0.7697624 0.7517628 1.0694209 0.9924634
#> [3,] 0.19477817 0.9468251 0.7921490 0.8613656 1.1460122
#> [4,] 0.46519771 1.1452910 0.6915521 0.7769309 1.3099754
#> [5,] 0.05266881 0.5369111 0.8227531 0.5337735 0.6512235
#> 
#> $w
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.04630883 0.06160728 0.07970241 0.06411287 0.05087420
#> [2,] 0.17171098 0.05738415 0.12371291 0.10213471 0.11901195
#> [3,] 0.05621183 0.09823041 0.13733188 0.04244179 0.05707869
#> [4,] 0.28999989 0.17260980 0.14856895 0.19790865 0.26347766
#> [5,] 0.03380618 0.11806940 0.17673644 0.08128494 0.07145465
#> 
#> $s
#>           [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,] 1240.5506 21.75104 122.6932 50.26376 79.53124
#> [2,]  187.8732 25.13721 144.0263 51.51721 58.34924
#> [3,]  156.9156 20.77378 109.9159 48.00917 49.14660
#> [4,]  345.6580 23.29100 154.5460 28.83495 73.37438
#> [5,]  113.3398 41.14326  53.0068 60.33144 82.43192
#> 
#> $scale
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
#> [4,]    0
#> [5,]    0
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>           [,1]       [,2]      [,3]       [,4]
#> [1,] 34.242194 -3.9550523  3.654595 14.5472220
#> [2,] -3.955052 29.9244810  3.423894 -0.2354667
#> [3,]  3.654595  3.4238938 36.544009  5.0806511
#> [4,] 14.547222 -0.2354667  5.080651 51.4717664
#> 
#> , , 2, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  3.9171773 -0.2678846  0.6043228  0.9725593
#> [2,] -0.2678846  5.6028974 -1.9456257 -0.6020596
#> [3,]  0.6043228 -1.9456257  7.6416367  2.4155913
#> [4,]  0.9725593 -0.6020596  2.4155913  7.4782832
#> 
#> , , 3, 1
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,] 11.3909952  0.9998108 14.0604836 14.679514
#> [2,]  0.9998108 11.2277941  0.3475835  6.248362
#> [3,] 14.0604836  0.3475835 40.0525057 31.110086
#> [4,] 14.6795143  6.2483616 31.1100856 36.368717
#> 
#> , , 4, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  8.0368375  1.7463235  0.5955574 -0.4323305
#> [2,]  1.7463235 10.9620258  3.0519385 -0.5663983
#> [3,]  0.5955574  3.0519385 11.5359820 -1.5877483
#> [4,] -0.4323305 -0.5663983 -1.5877483 13.5972993
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.8747507 -0.6699634  0.8114944 -0.8573404
#> [2,] -0.6699634  7.6162412 -0.8868132  1.0229062
#> [3,]  0.8114944 -0.8868132  4.9676338  1.8102749
#> [4,] -0.8573404  1.0229062  1.8102749  4.7929555
#> 
#> , , 1, 2
#> 
#>           [,1]        [,2]        [,3]       [,4]
#> [1,] 2.4032735  0.14336785  1.03568733  0.2037929
#> [2,] 0.1433679  0.55398735 -0.05692238 -0.2334827
#> [3,] 1.0356873 -0.05692238  2.01870099  0.6302379
#> [4,] 0.2037929 -0.23348268  0.63023791  0.8362726
#> 
#> , , 2, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8489970 -0.1187183 -0.2163982  0.4476054
#> [2,] -0.1187183  0.9989551 -0.1661953 -0.2932626
#> [3,] -0.2163982 -0.1661953  1.4453864  0.1405464
#> [4,]  0.4476054 -0.2932626  0.1405464  1.1335054
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]     [,3]       [,4]
#> [1,]  9.1783909 -0.7523795 16.27020 13.5811836
#> [2,] -0.7523795  8.1814187 -6.35308 -0.7594716
#> [3,] 16.2701979 -6.3530798 35.60886 26.5729360
#> [4,] 13.5811836 -0.7594716 26.57294 23.2385870
#> 
#> , , 4, 2
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 0.8465745  0.3606768  0.1455334 0.3782887
#> [2,] 0.3606768  1.7198022 -0.2109223 0.6253619
#> [3,] 0.1455334 -0.2109223  0.9737850 0.2686167
#> [4,] 0.3782887  0.6253619  0.2686167 1.2740901
#> 
#> , , 5, 2
#> 
#>            [,1]      [,2]      [,3]       [,4]
#> [1,]  1.3747513 -1.107019  1.066334  0.6357551
#> [2,] -1.1070187  5.889637 -3.979975 -1.4634111
#> [3,]  1.0663340 -3.979975  6.840488  3.3631223
#> [4,]  0.6357551 -1.463411  3.363122  3.3433664
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  4.9661861  0.8815475  0.2800947 -0.4158751
#> [2,]  0.8815475  3.9200610 -0.1487975 -0.8670515
#> [3,]  0.2800947 -0.1487975  3.0015011 -0.1059844
#> [4,] -0.4158751 -0.8670515 -0.1059844  3.0856034
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  3.6184885 -1.1132246 -0.3265555 -0.1044926
#> [2,] -1.1132246  3.8759510 -0.5131726  0.9133411
#> [3,] -0.3265555 -0.5131726  4.9126601  0.1669871
#> [4,] -0.1044926  0.9133411  0.1669871  5.3270764
#> 
#> , , 3, 3
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  3.0247355 -1.1370169  1.163480  0.1051462
#> [2,] -1.1370169 10.5508269 -5.162913 -0.2271892
#> [3,]  1.1634802 -5.1629131  6.607481  1.9358474
#> [4,]  0.1051462 -0.2271892  1.935847  5.2851613
#> 
#> , , 4, 3
#> 
#>            [,1]      [,2]       [,3]     [,4]
#> [1,]  6.4592591 -2.234043 -0.5545285 0.379734
#> [2,] -2.2340430  4.496582  1.2798543 1.575783
#> [3,] -0.5545285  1.279854  6.7285434 1.860133
#> [4,]  0.3797340  1.575783  1.8601334 6.218120
#> 
#> , , 5, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.4048727 -0.7095276  0.7703291  0.7735159
#> [2,] -0.7095276  4.4634500 -2.3587243 -1.1948751
#> [3,]  0.7703291 -2.3587243  3.4001428  1.5476959
#> [4,]  0.7735159 -1.1948751  1.5476959  3.0555004
#> 
#> , , 1, 4
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 1.7161517  0.1022508  0.2279183 0.2691964
#> [2,] 0.1022508  1.1927828 -0.1647408 0.1192013
#> [3,] 0.2279183 -0.1647408  1.8527011 0.2435456
#> [4,] 0.2691964  0.1192013  0.2435456 1.5618982
#> 
#> , , 2, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] 1.29807489  0.1537830 0.01980572  0.2425236
#> [2,] 0.15378301  2.2390579 0.72566480 -0.1515822
#> [3,] 0.01980572  0.7256648 2.23684311  0.4136094
#> [4,] 0.24252364 -0.1515822 0.41360939  2.8126378
#> 
#> , , 3, 4
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  1.28691500  0.06163173 -0.06797624  0.5155272
#> [2,]  0.06163173 14.76957219 -9.16425861 -1.7881275
#> [3,] -0.06797624 -9.16425861  9.00900843  2.4163616
#> [4,]  0.51552725 -1.78812751  2.41636159  3.5239354
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  1.0520886 -0.1258884  0.4154972 0.3908050
#> [2,] -0.1258884  1.1428449 -0.1683071 0.3969291
#> [3,]  0.4154972 -0.1683071  2.2222091 0.5418152
#> [4,]  0.3908050  0.3969291  0.5418152 1.8424166
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.4521027 -0.3129811  0.4572338 -0.1749308
#> [2,] -0.3129811  4.7520543 -3.0208250 -0.2027240
#> [3,]  0.4572338 -3.0208250  5.7047350  1.2885235
#> [4,] -0.1749308 -0.2027240  1.2885235  2.1955477
#> 
#> , , 1, 5
#> 
#>             [,1]       [,2]        [,3]      [,4]
#> [1,]  1.99484199  0.2065535 -0.08525657 0.1781209
#> [2,]  0.20655353  2.3634986 -0.76016293 0.5529560
#> [3,] -0.08525657 -0.7601629  2.64668889 0.1590913
#> [4,]  0.17812086  0.5529560  0.15909132 1.7621303
#> 
#> , , 2, 5
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  2.0048220 -0.16195071 -0.16056768 -0.1149136
#> [2,] -0.1619507  1.31625800  0.06853514  0.2255500
#> [3,] -0.1605677  0.06853514  1.34545318 -0.2926913
#> [4,] -0.1149136  0.22554998 -0.29269132  2.0422330
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  1.5046142 -0.1305396 -0.3075712 -0.443580
#> [2,] -0.1305396 13.2622449 -6.9168407 -1.130452
#> [3,] -0.3075712 -6.9168407  6.4092102  1.577744
#> [4,] -0.4435800 -1.1304522  1.5777438  4.279658
#> 
#> , , 4, 5
#> 
#>              [,1]         [,2]       [,3]         [,4]
#> [1,]  2.067563269 -0.009583528 -0.1731108 -0.130733291
#> [2,] -0.009583528  1.521903066 -0.2523575  0.002282911
#> [3,] -0.173110835 -0.252357537  2.4150930  0.933130283
#> [4,] -0.130733291  0.002282911  0.9331303  2.563053105
#> 
#> , , 5, 5
#> 
#>           [,1]       [,2]      [,3]       [,4]
#> [1,] 2.8422713  0.1605005  0.755490  0.6628353
#> [2,] 0.1605005  4.5056225 -1.213632 -0.1188518
#> [3,] 0.7554900 -1.2136322  4.881411  1.5250639
#> [4,] 0.6628353 -0.1188518  1.525064  3.8321049
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.5010824 -0.3891840 -0.4010135 -0.1199125
#> [2,] -0.4487598  1.3991097  0.1770979 -0.1163843
#> [3,] -0.5911252  0.3076881  0.4786864  0.1528137
#> [4,]  0.9323912 -0.2582497  0.5729562  1.1283334
#> [5,] -2.0475488  6.0552275  2.7509068 -5.7228073
#> 
#> , , 2, 1
#> 
#>            [,1]        [,2]      [,3]        [,4]
#> [1,]  1.2366849  0.11073112 0.3382216 -0.67547673
#> [2,]  0.8644138  0.77282872 0.1765969  0.53398076
#> [3,]  0.6552995  0.10904956 0.3332140 -0.07910077
#> [4,] -0.7894760 -0.09639633 0.4042539  1.19783746
#> [5,] -1.9271050  1.91181398 0.6457528 -2.92545654
#> 
#> , , 3, 1
#> 
#>              [,1]       [,2]       [,3]       [,4]
#> [1,]  0.003288198 -0.7640537 -1.3360206 -1.6781538
#> [2,] -0.187352294  0.1643762 -0.6543066 -0.6447078
#> [3,] -0.823368618 -0.1152278 -1.1691732 -1.4314484
#> [4,]  0.959205015  0.6141292  2.2856517  2.8084629
#> [5,]  9.961992785 -0.6066181 13.7618848 12.0115833
#> 
#> , , 4, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9106106  0.9316913 -0.9375216 -1.6221079
#> [2,] -0.7241436 -1.1755603 -0.7205245 -0.4646257
#> [3,] -0.5748744 -2.7706449  0.8308063  1.2670496
#> [4,]  0.6241637  2.4829072  0.4806998  0.2613365
#> [5,] -1.1723921 -2.3167274  2.9975331  4.8643455
#> 
#> , , 5, 1
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.22367850  0.20020076  0.67440557  0.6817886
#> [2,] -0.10406891  0.04033385 -0.20095933 -0.2104477
#> [3,]  0.06120314 -1.36898442  0.64711513 -0.1836419
#> [4,]  0.27280119  1.26221529  0.01409411  0.8754769
#> [5,]  1.68116935  3.05881429  2.01528439  1.7716328
#> 
#> , , 1, 2
#> 
#>            [,1]         [,2]         [,3]         [,4]
#> [1,]  0.1145167 -0.117033537 -0.584660207 -0.000122762
#> [2,]  0.4056409  1.101236345  0.005773568 -0.210204545
#> [3,] -0.2102197 -0.002910421  0.737333036  0.014258395
#> [4,]  0.5760807  0.046363716  0.540852968  1.017922216
#> [5,] -0.6998301  0.108278940 -0.743356562  0.125357683
#> 
#> , , 2, 2
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.8508573  0.07608836  0.29282232 0.24799837
#> [2,]  0.1257195  0.92834772  0.01930136 0.06766524
#> [3,]  0.4564310  0.09939031  0.31183415 0.32502871
#> [4,] -0.3715389 -0.09763416  0.48354907 0.63097006
#> [5,]  0.8391452  0.50675023 -0.18317145 0.28231570
#> 
#> , , 3, 2
#> 
#>             [,1]       [,2]      [,3]       [,4]
#> [1,]  0.02073152 -0.9288236 -0.704770 -1.1101394
#> [2,] -0.42109349  1.4615947 -1.349080 -0.8508382
#> [3,] -0.89352630  1.0734440 -1.685811 -1.6061514
#> [4,]  1.02586436 -0.6895858  2.475657  2.6318394
#> [5,] 12.53649555  1.5363829 23.521536 21.2394365
#> 
#> , , 4, 2
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  1.1868963  0.1364629  0.38996553  0.5133717
#> [2,] -0.4359872  0.0423444 -0.91663178 -1.0950867
#> [3,]  0.1283329  0.2796941  0.67839309  0.2711091
#> [4,] -0.2013168 -0.3000801  0.20595797  0.5648708
#> [5,]  2.1653730  1.0739401  0.03418362  2.0500260
#> 
#> , , 5, 2
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.57019205  0.6376484  0.1678313  0.6755906
#> [2,] -0.04355147  0.1010948 -0.2799794 -0.1997710
#> [3,] -0.04771455 -1.3003025  0.3636576 -0.5731670
#> [4,]  0.25858187  1.0886233  0.4456963  1.2166970
#> [5,] -1.01250652 -3.2112559  6.3130817  2.5295866
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.6593127 -0.2180457 -0.3649461  0.20914474
#> [2,]  0.3290945  1.0459177 -0.1490222 -0.16549177
#> [3,]  0.3714280 -0.2688765  0.9280703  0.01487478
#> [4,] -0.1965201  0.3653344  0.2556824  0.92945507
#> [5,] -1.4607463 -0.5033173  0.9584041 -1.32704870
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.6256108  0.2189306 -0.1615041  0.1183390
#> [2,] -0.6370990  0.3899252 -0.5727550 -0.0288697
#> [3,]  0.3884619 -0.3392060  0.4495036 -0.6607988
#> [4,] -0.1001150  0.3607052  0.6440582  1.4423614
#> [5,]  2.5999623 -1.6751982 -0.4996481  5.8803210
#> 
#> , , 3, 3
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.8758317 -0.9546624  0.3072725 -0.14161588
#> [2,]  0.1391920  0.7522025  0.2169001  0.08192053
#> [3,]  0.2757390  0.3068918  0.9155453  0.24260521
#> [4,] -0.2428366  0.2067509 -0.1142219  0.76752279
#> [5,]  1.6644742 -2.3601264  0.3212905  3.34831438
#> 
#> , , 4, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.5426071  0.2483249  2.0335877  1.1482360
#> [2,] -0.5279465  0.4949133 -1.3132561 -0.6120776
#> [3,] -0.2541984  0.4451226  0.3670491  1.2660318
#> [4,]  0.4511883 -0.4995006 -0.1063445 -0.5643207
#> [5,] -2.4393879 -0.3258087  5.6757434 -3.8459310
#> 
#> , , 5, 3
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.5710211 -0.3792362  0.8565182 0.5224976
#> [2,]  0.5305956  0.5406155  0.2766342 0.4787983
#> [3,]  0.8446317 -0.6744647  1.2726665 0.6201050
#> [4,] -0.6600860  0.8358782 -0.6313477 0.1120313
#> [5,]  0.1322059  0.9573244 -2.4867420 0.5496667
#> 
#> , , 1, 4
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.68993542  0.18174946 -0.08972494  0.04787965
#> [2,]  0.16882089  0.94270280 -0.24804613 -0.29740859
#> [3,] -0.01367105  0.03055954  0.85316502  0.14439996
#> [4,]  0.10183333 -0.10560139  0.19863425  0.90604621
#> [5,]  1.63224207  0.21207789  0.65987166 -0.82164745
#> 
#> , , 2, 4
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.88350632  0.1901153  0.4917449  0.40165602
#> [2,] -0.10230081  0.6990327 -0.3150242  0.15674544
#> [3,]  0.18362890 -0.2023229  0.2915171  0.09956356
#> [4,] -0.06671329  0.1609152  0.5193005  0.78062529
#> [5,] -2.17545295  0.8618820 -1.6367264 -2.43020528
#> 
#> , , 3, 4
#> 
#>             [,1]         [,2]        [,3]         [,4]
#> [1,]  1.08923408 -0.821241824  0.78614655  0.260373001
#> [2,]  0.08884821  0.514402303  0.10018041 -0.087342044
#> [3,]  0.10238333  0.004895404  0.72381368  0.044522553
#> [4,] -0.15469392  0.421906077 -0.07226645  0.898075767
#> [5,]  0.73194112  1.073892115 -1.13331763  0.004682584
#> 
#> , , 4, 4
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.94443528  0.6948444  1.2565446  0.81742291
#> [2,] -0.31049640  0.5615890  0.3703632 -0.86724928
#> [3,]  0.04743839  0.2907438  1.8902288  1.06685064
#> [4,] -0.04228188 -0.5095388 -1.3001295 -0.29265248
#> [5,]  1.90583102 -0.1128731 -0.5602143  0.04475365
#> 
#> , , 5, 4
#> 
#>            [,1]        [,2]       [,3]      [,4]
#> [1,] 0.25078901 -0.01908579 0.08412758 0.3673642
#> [2,] 0.07808328  0.38032014 0.11792690 0.2526109
#> [3,] 0.14111298 -0.47890864 0.82883142 0.2671029
#> [4,] 0.17289812  0.56531631 0.08320651 0.5486212
#> [5,] 1.38908678 -1.96019448 0.51322794 0.1705062
#> 
#> , , 1, 5
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.70634556  0.07619325 -0.14904016 -0.17240933
#> [2,]  0.25934424  1.02378067 -0.11300766  0.01696221
#> [3,]  0.14357386  0.21979627  0.76054838  0.09947346
#> [4,] -0.04056517 -0.23010706  0.30597418  1.01730660
#> [5,]  0.36403634 -0.40824394 -0.05984137 -0.96316453
#> 
#> , , 2, 5
#> 
#>             [,1]       [,2]       [,3]         [,4]
#> [1,]  1.25322292 -0.1267447  0.2301900 -0.003718292
#> [2,]  0.48880699  0.5661504 -0.5318310 -0.348604203
#> [3,] -0.03402286 -0.2614699  0.5571210  0.196932783
#> [4,] -0.17262640  0.3377812  0.3875611  0.898585077
#> [5,]  1.16303261  0.9289953  2.0855155  0.555149035
#> 
#> , , 3, 5
#> 
#>              [,1]       [,2]        [,3]        [,4]
#> [1,]  1.134477209 -0.3721426  0.61043048  0.24106897
#> [2,]  0.023394298  0.7326781 -0.20408404 -0.05844575
#> [3,]  0.015425103  0.7817520  0.04246042 -0.04456509
#> [4,] -0.070764055 -0.3836337  0.57964314  0.95532017
#> [5,]  0.004420039 -2.8569734  2.44310789 -0.13114266
#> 
#> , , 4, 5
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.81192809  0.9728903  2.23605010  2.6143380
#> [2,] -0.44632191  1.0110932 -0.31426195 -1.3739497
#> [3,]  0.13298010  0.4932808  0.26007317 -0.9337999
#> [4,] -0.03854348 -0.8084762 -0.01979708  1.0224994
#> [5,] -1.17894834 -0.1141032 -1.13465245  2.9239458
#> 
#> , , 5, 5
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.74076841 -0.07992968  0.37703869  0.03137492
#> [2,] -0.05134246  0.40714454  0.08278294  0.37617758
#> [3,] -0.03065061 -0.56947215  0.91214053  0.71297451
#> [4,]  0.16759612  0.62859055 -0.10771627  0.30911140
#> [5,] -0.73357857  0.71121622  0.41206112 -1.97615152
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvars$get_last_draw`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|