R6 Class Representing PosteriorBVARs

The class PosteriorBVARs contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvars$new()

specify_posterior_bvars$get_posterior()

specify_posterior_bvars$get_last_draw()

specify_posterior_bvars$clone()

Method new()

Create a new posterior output PosteriorBVARs.

Usage

specify_posterior_bvars$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARs with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARs.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvars$get_posterior()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvars$get_last_draw()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvars$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARs" "R6"            

## ------------------------------------------------
## Method `specify_posterior_bvars$get_posterior`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $nu
#>      [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,]  5.1 5.100000 5.100000 5.100000 5.100000
#> [2,]  5.1 5.100000 5.100000 5.100000 5.100000
#> [3,]  5.1 5.057568 5.102918 5.102918 5.102918
#> [4,]  5.1 5.100000 5.100000 5.100000 5.100000
#> [5,]  5.1 5.100000 5.100000 5.100000 5.000609
#> 
#> $m
#>              [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,]  0.520547953 0.7787633 0.9633304 0.6362416 1.0445436
#> [2,]  0.057605091 0.2865048 0.6220552 1.2743330 1.1497069
#> [3,] -0.077604373 0.6690344 0.8688207 0.9163693 0.8623567
#> [4,]  0.118409825 0.4666839 0.2419398 0.3360296 0.3477420
#> [5,] -0.009994883 0.8308557 1.0206075 1.0194180 0.9400937
#> 
#> $w
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.11249931 0.09365494 0.20385721 0.19575723 0.07545807
#> [2,] 0.03557386 0.05667276 0.18079503 0.16960078 0.18657309
#> [3,] 0.05536531 0.12870161 0.15383893 0.13934933 0.09997117
#> [4,] 0.04158850 0.06223981 0.06348967 0.09166903 0.18057956
#> [5,] 0.04447417 0.07442858 0.13207215 0.07945171 0.12376908
#> 
#> $s
#>            [,1]     [,2]      [,3]     [,4]     [,5]
#> [1,]  129.03775 55.51051  85.05769 61.73875 66.53290
#> [2,]  414.64385 30.08000  81.17600 55.68054 65.04848
#> [3,]  127.26065 28.46269  73.60854 40.55991 54.95781
#> [4,]   60.62477 65.03184  46.62029 72.75253 79.30624
#> [5,] 3600.16976 17.89108 129.76881 45.50223 52.63064
#> 
#> $scale
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
#> [4,]    0
#> [5,]    0
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  2.9874728 -0.34482942 -0.0515920  0.27270590
#> [2,] -0.3448294  3.93749456 -0.2038089 -0.07889908
#> [3,] -0.0515920 -0.20380889  2.2608158 -0.12433409
#> [4,]  0.2727059 -0.07889908 -0.1243341  4.61226912
#> 
#> , , 2, 1
#> 
#>             [,1]       [,2]        [,3]      [,4]
#> [1,] 14.32523437  1.1138320  0.06139259  2.298101
#> [2,]  1.11383205 18.1896778 -0.11919186 -3.075847
#> [3,]  0.06139259 -0.1191919 17.11883042  2.788538
#> [4,]  2.29810122 -3.0758471  2.78853775 12.578649
#> 
#> , , 3, 1
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  4.3314468 -0.4379093 -0.1844804 -1.716357
#> [2,] -0.4379093 11.4211555 -2.5272011  2.478321
#> [3,] -0.1844804 -2.5272011  9.9311966  2.992106
#> [4,] -1.7163568  2.4783209  2.9921057 13.262190
#> 
#> , , 4, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  2.29434763  0.3225417 -0.4631420 -0.06348282
#> [2,]  0.32254174  1.8293697 -0.5491336  0.36791334
#> [3,] -0.46314198 -0.5491336  2.7742038  0.22007909
#> [4,] -0.06348282  0.3679133  0.2200791  2.81755313
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]        [,3]      [,4]
#> [1,] 98.5259185   9.286671   0.7494369  10.76001
#> [2,]  9.2866710 114.906937  41.1966092 -10.26008
#> [3,]  0.7494369  41.196609  98.3644367 -17.65602
#> [4,] 10.7600100 -10.260076 -17.6560221 157.20083
#> 
#> , , 1, 2
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  1.14245409  0.17068972 -0.03752007  0.1144161
#> [2,]  0.17068972  1.34180629 -0.02572554 -0.1646516
#> [3,] -0.03752007 -0.02572554  1.74144725  0.3105014
#> [4,]  0.11441614 -0.16465157  0.31050143  1.3318382
#> 
#> , , 2, 2
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  1.34744861 -0.05199228 -0.3738086 -0.23653241
#> [2,] -0.05199228  1.22593793  0.1393930  0.06183885
#> [3,] -0.37380860  0.13939302  1.2766182  0.41864800
#> [4,] -0.23653241  0.06183885  0.4186480  1.25978887
#> 
#> , , 3, 2
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  1.3383976 -2.278961  0.1125654 -1.2995083
#> [2,] -2.2789606 19.767467 -6.8740720  6.2353258
#> [3,]  0.1125654 -6.874072  6.2734808  0.5306167
#> [4,] -1.2995083  6.235326  0.5306167  5.3041997
#> 
#> , , 4, 2
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  2.2725008  0.3841329 -1.267460 -0.4885940
#> [2,]  0.3841329  2.9244569 -1.389358  0.2874183
#> [3,] -1.2674603 -1.3893575  4.795144  2.0303622
#> [4,] -0.4885940  0.2874183  2.030362  6.4920804
#> 
#> , , 5, 2
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  1.0948786 -0.2025421  1.616477  0.9018833
#> [2,] -0.2025421  3.3158672 -2.562891 -1.0254581
#> [3,]  1.6164772 -2.5628908  6.642070  3.7175739
#> [4,]  0.9018833 -1.0254581  3.717574  3.1727143
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  3.0450482 -0.2212505 -0.8469489 -0.6731160
#> [2,] -0.2212505  2.6864324 -0.2812246  1.3516180
#> [3,] -0.8469489 -0.2812246  2.9400278  0.7559284
#> [4,] -0.6731160  1.3516180  0.7559284  4.7099438
#> 
#> , , 2, 3
#> 
#>             [,1]         [,2]         [,3]       [,4]
#> [1,]  2.00875675  0.094838533 -1.204211259 -0.5644886
#> [2,]  0.09483853  1.634446611  0.007087329 -0.3334809
#> [3,] -1.20421126  0.007087329  3.312479980  0.9216138
#> [4,] -0.56448862 -0.333480873  0.921613815  2.7856381
#> 
#> , , 3, 3
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  1.9174847  0.3778241 -0.5410519 -0.654160
#> [2,]  0.3778241 10.5007842 -2.4774779  1.781753
#> [3,] -0.5410519 -2.4774779  7.8882873  3.458901
#> [4,] -0.6541600  1.7817534  3.4589007  5.245177
#> 
#> , , 4, 3
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  1.8078519 -0.3869910  0.6325955 0.2918955
#> [2,] -0.3869910  1.0106904 -0.1822397 0.1308185
#> [3,]  0.6325955 -0.1822397  2.3317394 0.7302448
#> [4,]  0.2918955  0.1308185  0.7302448 1.4516215
#> 
#> , , 5, 3
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  2.96333819 -0.3668425  0.8289643  0.02091587
#> [2,] -0.36684253  5.3064389 -3.5073760 -1.25018099
#> [3,]  0.82896431 -3.5073760  9.1327178  2.05387788
#> [4,]  0.02091587 -1.2501810  2.0538779  4.05604826
#> 
#> , , 1, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.0102078  0.4069881 -0.6987538 -0.8598255
#> [2,]  0.4069881  2.5708041 -0.7484072 -0.8863509
#> [3,] -0.6987538 -0.7484072  2.0773741  1.0632871
#> [4,] -0.8598255 -0.8863509  1.0632871  2.9905907
#> 
#> , , 2, 4
#> 
#>             [,1]        [,2]        [,3]      [,4]
#> [1,]  1.98862303  0.05270455 -0.25139805 0.4456747
#> [2,]  0.05270455  2.52891629 -0.08757452 0.4614706
#> [3,] -0.25139805 -0.08757452  1.78915125 0.3946787
#> [4,]  0.44567467  0.46147062  0.39467873 2.4837205
#> 
#> , , 3, 4
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.8605472 -0.2869137  0.2852273 0.2636349
#> [2,] -0.2869137  8.3827137 -4.3493500 0.2386858
#> [3,]  0.2852273 -4.3493500  6.0568376 2.9804644
#> [4,]  0.2636349  0.2386858  2.9804644 4.2084335
#> 
#> , , 4, 4
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  1.76060527  0.1827122 -0.03953852 -0.04936201
#> [2,]  0.18271215  2.6852580 -0.42724421 -0.18422905
#> [3,] -0.03953852 -0.4272442  2.74413558  0.84312698
#> [4,] -0.04936201 -0.1842290  0.84312698  2.96839065
#> 
#> , , 5, 4
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  1.62353405 -0.2240584  0.3674797  0.05279372
#> [2,] -0.22405838  2.9158233 -1.4965601 -0.64969721
#> [3,]  0.36747973 -1.4965601  4.7077842  1.13831965
#> [4,]  0.05279372 -0.6496972  1.1383196  2.46420909
#> 
#> , , 1, 5
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  2.14263675  0.04382613 -0.09021035 -0.1535250
#> [2,]  0.04382613  1.74795707 -0.81926636 -0.5713942
#> [3,] -0.09021035 -0.81926636  3.17142427  1.6664345
#> [4,] -0.15352500 -0.57139422  1.66643450  3.4278219
#> 
#> , , 2, 5
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  1.6155361 -0.2081333 0.6433610 0.6165456
#> [2,] -0.2081333  2.1210075 0.2411434 0.2395484
#> [3,]  0.6433610  0.2411434 2.6574309 1.3829411
#> [4,]  0.6165456  0.2395484 1.3829411 3.2744691
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  2.0265581 -0.9115304  1.425997  1.033102
#> [2,] -0.9115304 13.2429367 -7.919378 -1.706333
#> [3,]  1.4259975 -7.9193775  8.326009  3.005152
#> [4,]  1.0331019 -1.7063328  3.005152  5.465344
#> 
#> , , 4, 5
#> 
#>           [,1]       [,2]       [,3]       [,4]
#> [1,] 2.1323754  0.1818622  0.6956870  0.2353725
#> [2,] 0.1818622  1.7577550 -0.1270786 -0.6815106
#> [3,] 0.6956870 -0.1270786  2.1939183  0.2636215
#> [4,] 0.2353725 -0.6815106  0.2636215  2.3260616
#> 
#> , , 5, 5
#> 
#>            [,1]      [,2]       [,3]       [,4]
#> [1,]  1.9723753 -0.363601 -0.7559877 -0.1427312
#> [2,] -0.3636010  6.136996 -2.3653811  0.8815920
#> [3,] -0.7559877 -2.365381  3.9184394  0.5819932
#> [4,] -0.1427312  0.881592  0.5819932  1.8672717
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.5046348  0.4319024 -0.02700659 -0.1839241
#> [2,]  0.5541671  0.9458902 -0.14379230  0.6214321
#> [3,]  0.4786753  0.5041496  0.71515865  1.5174191
#> [4,] -0.2347889 -0.6734265  0.27817339 -0.4365584
#> [5,] -2.5696695  0.5638629  0.85772455  0.8162255
#> 
#> , , 2, 1
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.01280378  0.42711495  0.7600735  0.90741742
#> [2,]  0.50360743 -0.09209277 -0.2168440 -0.07038468
#> [3,]  0.28065886  0.94902366 -0.1953971  0.24733412
#> [4,] -0.03447158 -0.69377908  0.8009374  0.52770852
#> [5,] -0.59929331  0.13814781 -1.8181239 -1.31660101
#> 
#> , , 3, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8675943  0.1890148  1.2926308  0.9856878
#> [2,]  0.1951953 -0.2501136 -0.3633484 -0.4355467
#> [3,]  0.3243754 -0.4428881 -0.4207754 -1.3474033
#> [4,] -0.2696948  0.5900734  0.7859418  1.7585912
#> [5,] -0.4977151  3.0063414  0.1081045  5.9403073
#> 
#> , , 4, 1
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.1905849  0.07003445 -0.3340774  0.2073634
#> [2,] -0.3442959  0.22595665 -0.2336281 -0.1069225
#> [3,]  0.1189126  0.12002943  0.7865314  0.7742409
#> [4,]  0.1906325 -0.11825581  0.3127443  0.1640228
#> [5,] -1.5675671  0.66852478  1.2140590  1.3680235
#> 
#> , , 5, 1
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]   1.4571584 -0.37566300   0.7702105 -3.2748772
#> [2,]  -0.9608805  0.33057801  -0.8662669  0.7671675
#> [3,]  -1.1956692 -0.00310455  -0.9315742  1.9490606
#> [4,]   1.2516458  0.22660647   1.7618711  0.5626311
#> [5,] -14.6730530  0.56889563 -12.3355213 -1.1797349
#> 
#> , , 1, 2
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,] 0.57273215  0.2118732 -0.02640052  0.15015299
#> [2,] 0.18688169  1.0646985 -0.20431686 -0.23128529
#> [3,] 0.02126231  0.6303167  1.00350160  0.65097874
#> [4,] 0.14728072 -0.7098003  0.01770954  0.36678145
#> [5,] 0.70260222  0.8591581  0.86244935  0.08394885
#> 
#> , , 2, 2
#> 
#>             [,1]        [,2]      [,3]       [,4]
#> [1,]  0.45810110 -0.16136752 0.7140513  0.5889777
#> [2,] -0.24694918  0.68331953 0.1200716  0.5042931
#> [3,]  0.24060833  0.25941031 0.5541988  0.3163269
#> [4,]  0.03557887 -0.08420137 0.1127340  0.4464372
#> [5,] -1.12666683 -0.75364550 0.6197062 -0.2647711
#> 
#> , , 3, 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.8825889  0.05507117  1.3124665  1.0773267
#> [2,]  0.1389496  0.44616725 -0.6636381 -0.7961179
#> [3,]  0.3168428 -0.72390772 -0.3956898 -1.3276937
#> [4,] -0.2417948  0.66679357  0.8508070  1.8837789
#> [5,] -1.2471498  5.90731694 -0.6145161  2.5502478
#> 
#> , , 4, 2
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.09911365  0.03185404  0.08911503  0.2538997
#> [2,] -0.70679360  0.60716175 -0.32253390 -0.2066524
#> [3,] -0.05487055  0.31131042  0.68206783  0.7492685
#> [4,]  0.38171451 -0.30075685  0.28478675  0.2014122
#> [5,]  0.32148674  0.05797874 -0.30306279 -0.9199470
#> 
#> , , 5, 2
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.6425944 -0.76771847  0.1097500  0.09221476
#> [2,] -0.4113888  0.80602715 -0.6696887 -0.35451321
#> [3,] -0.6449239 -0.01339449 -0.2102305 -0.55690705
#> [4,]  0.7969052  0.34990963  1.1294654  1.51007077
#> [5,]  0.5680585  2.20938828  1.0604611  0.69033377
#> 
#> , , 1, 3
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.82024565  0.2068051 -0.2073521 -0.4682113
#> [2,] -0.03720054  0.8966568 -0.1731638  0.5642917
#> [3,] -0.46140737  0.4051844  1.0513389  1.6240522
#> [4,]  0.38445589 -0.4708017  0.1084055 -0.3441623
#> [5,]  5.35726808  1.1346568 -0.8060684 -0.3139660
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.6336348 -0.3435054  0.4597423 0.4722623
#> [2,] -0.7472471  0.9206317  0.8975876 0.6733711
#> [3,] -0.5810729  0.3817444  1.7182756 0.2553148
#> [4,]  0.7925067 -0.2260989 -0.9141227 0.4676690
#> [5,] -2.5284579  2.6446295 -2.0830066 1.3360432
#> 
#> , , 3, 3
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.9276784  0.03097903  1.01899563  0.8632253
#> [2,]  0.1417801 -0.65676485 -0.16655535 -0.6778988
#> [3,]  0.1953103 -1.33933456 -0.04167658 -1.2166441
#> [4,] -0.1606590  1.49293776  0.58476922  1.9093184
#> [5,] -0.6849640  3.83376430 -3.35212616 -1.8926107
#> 
#> , , 4, 3
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.7798228  0.63208277  0.5074219  0.19466078
#> [2,] -0.1598573  0.40325461 -0.3950002  0.09700504
#> [3,] -0.1336452 -0.08462617  0.4362717  0.87754809
#> [4,]  0.1848381 -0.10369099  0.3750811  0.08529289
#> [5,]  1.9837558 -0.83566308  1.0773980 -0.71842292
#> 
#> , , 5, 3
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  1.1181654 -0.18210180  0.72690068  0.4781423
#> [2,] -0.1101340  0.69283334 -0.24035951 -0.2946048
#> [3,] -0.4329848  0.03804318  0.03460183 -0.6258080
#> [4,]  0.3180062  0.12890776  0.58549697  1.3827168
#> [5,]  1.9401272 -2.08111936 -0.07725331  0.8312197
#> 
#> , , 1, 4
#> 
#>             [,1]       [,2]        [,3]        [,4]
#> [1,]  1.21156500  0.8702326 -0.33573781 -0.26322217
#> [2,] -0.37077503  0.3949912 -0.02528397  0.08453527
#> [3,] -0.02510196 -0.2374652  1.08657277  0.08974587
#> [4,] -0.04678945 -0.1197733  0.04208561  0.97786257
#> [5,]  1.85569611  1.1791082  1.86568134  2.02940617
#> 
#> , , 2, 4
#> 
#>             [,1]        [,2]       [,3]      [,4]
#> [1,]  0.57784469 -0.41323742  0.3583719 0.5911702
#> [2,] -0.40954595  1.12168707  0.9159227 0.4917454
#> [3,]  0.41966933  0.13772048  1.3915118 0.6621630
#> [4,] -0.12591882  0.02862985 -0.6653705 0.1036248
#> [5,]  0.04765104 -2.90349587  2.0318755 3.4001816
#> 
#> , , 3, 4
#> 
#>              [,1]       [,2]         [,3]       [,4]
#> [1,]  1.010310670 -0.2465864  0.717161158  0.4329623
#> [2,] -0.055786491  0.3301206 -0.240009913 -0.4615942
#> [3,]  0.043547992 -0.3312251 -0.009855594 -0.7539799
#> [4,] -0.004586518  0.4756113  0.636926031  1.5600924
#> [5,] -1.236188651  3.8308942  0.537356574  1.4716140
#> 
#> , , 4, 4
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.07279064  0.72291543  0.5995615 -0.18093677
#> [2,]  0.73277415  0.07751907 -0.6220914  0.27769969
#> [3,]  0.56967817 -0.27355715  0.3208990  1.14680766
#> [4,] -0.27395784  0.05031959  0.4873401 -0.03037112
#> [5,]  1.08164678 -0.40271325 -0.6563519 -2.42512373
#> 
#> , , 5, 4
#> 
#>             [,1]        [,2]        [,3]         [,4]
#> [1,]  0.65867611 -0.58331384 -0.11835672 -0.412423747
#> [2,]  0.08435380  0.91714661  0.12482489  0.002194196
#> [3,]  0.20032704  0.26978360  1.02411502 -0.024532586
#> [4,] -0.06734675  0.05951035  0.05731875  1.204859769
#> [5,]  1.37877972 -1.29602371 -2.67801507 -0.063794042
#> 
#> , , 1, 5
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  1.24118995  0.43460956 -0.23643844  0.15642493
#> [2,] -0.19938730  0.60941037  0.10493538 -0.16545007
#> [3,]  0.08382198 -0.10017519  1.17061502  0.02397843
#> [4,] -0.17178375 -0.03798096 -0.07816084  0.94202312
#> [5,]  0.35615291 -0.54249273  1.43827782 -0.59937842
#> 
#> , , 2, 5
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.5386953 -0.22236246  0.5733459  0.51105744
#> [2,] -0.2824350  0.70473866  0.6157316  0.62138145
#> [3,]  0.4883806  0.13057101  1.3498490  0.06341684
#> [4,] -0.1884078  0.01966422 -0.5831426  0.67372205
#> [5,] -1.0407867  0.72531827 -1.9368329 -2.02760091
#> 
#> , , 3, 5
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.91408134 -0.26390819  1.1407574  0.66574671
#> [2,]  0.07490937  0.67604589 -0.2785732  0.01925859
#> [3,]  0.29390693 -0.06112007  0.4365621  0.21079136
#> [4,] -0.22677792  0.25018756  0.1532223  0.56749231
#> [5,] -0.43529831  0.18077781 -3.1839248 -0.02631500
#> 
#> , , 4, 5
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  1.3405142 -0.7403210 0.53981799 1.6805248
#> [2,]  1.1944827  0.7432767 0.25868771 0.2800840
#> [3,]  0.2175754  0.7195987 0.78610935 0.2871976
#> [4,] -0.3675025 -0.4220923 0.01369561 0.1314172
#> [5,] -0.3329383 -1.1988007 0.06466219 0.3988921
#> 
#> , , 5, 5
#> 
#>              [,1]       [,2]        [,3]        [,4]
#> [1,]  0.741958336 -1.2761379  0.63230263  0.65267011
#> [2,] -0.019621148  0.5251373 -0.03216662  0.02314996
#> [3,]  0.125232873 -0.4141638  0.47375862 -0.15070492
#> [4,] -0.007205198  1.0403923  0.17936919  0.83897732
#> [5,]  0.964777583 -1.6607877  0.89908759  0.39908347
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvars$get_last_draw`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|