R6 Class Representing PosteriorBVARs

The class PosteriorBVARs contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvars$new()

specify_posterior_bvars$get_posterior()

specify_posterior_bvars$get_last_draw()

specify_posterior_bvars$clone()

Method new()

Create a new posterior output PosteriorBVARs.

Usage

specify_posterior_bvars$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARs with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARs.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvars$get_posterior()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvars$get_last_draw()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvars$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARs" "R6"            

## ------------------------------------------------
## Method `specify_posterior_bvars$get_posterior`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $nu
#>      [,1] [,2] [,3]     [,4]     [,5]
#> [1,]  5.1  5.1  5.1 5.100000 5.100000
#> [2,]  5.1  5.1  5.1 5.100000 5.100000
#> [3,]  5.1  5.1  5.1 5.100000 5.100000
#> [4,]  5.1  5.1  5.1 5.100000 5.100000
#> [5,]  5.1  5.1  5.1 5.086577 5.086577
#> 
#> $m
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.13411041 0.4628842 1.1991257 0.8140111 0.6629442
#> [2,] 0.06171044 0.5483313 0.7265511 0.9272284 0.7483728
#> [3,] 1.16558794 0.6306807 0.8611906 0.9181724 0.8625465
#> [4,] 0.52678984 0.4367852 0.3332391 0.5038167 0.4427902
#> [5,] 0.62977953 1.3512887 1.0082253 1.4426378 0.4639004
#> 
#> $w
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.08134198 0.23251459 0.18160811 0.23357564 0.09221963
#> [2,] 0.05159186 0.05737959 0.07009808 0.09198115 0.07735102
#> [3,] 0.26457199 0.16123555 0.06207743 0.10060676 0.13970380
#> [4,] 0.12673690 0.06158314 0.06431362 0.07242450 0.13048219
#> [5,] 0.33965263 0.34212588 0.30858211 0.25471759 0.09489273
#> 
#> $s
#>            [,1]     [,2]      [,3]     [,4]      [,5]
#> [1,]   99.70695 40.88198 127.59817 38.82669  83.83347
#> [2,] 1076.77950 24.94386 125.53698 46.12404 104.12294
#> [3,]   90.36595 41.86572  86.58981 64.10533  50.17603
#> [4,]  148.46400 45.63566  82.45340 67.83489  40.86281
#> [5,]  358.89941 27.68012  85.87431 30.40502  76.96216
#> 
#> $scale
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
#> [4,]    0
#> [5,]    0
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] 4.52520781  0.5516210  0.8714277 0.05373858
#> [2,] 0.55162104  3.2156830 -0.3715094 0.24848003
#> [3,] 0.87142767 -0.3715094  4.0040328 1.64547739
#> [4,] 0.05373858  0.2484800  1.6454774 3.87754143
#> 
#> , , 2, 1
#> 
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 23.264648 -2.567363 -1.023280  2.587722
#> [2,] -2.567363 32.415613 14.263221  7.388474
#> [3,] -1.023280 14.263221 31.321886  6.315056
#> [4,]  2.587722  7.388474  6.315056 25.064773
#> 
#> , , 3, 1
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 4.5235190  0.6373694  0.6724568 0.5213125
#> [2,] 0.6373694  8.6517309 -2.0461786 1.3735739
#> [3,] 0.6724568 -2.0461786  5.4713902 2.1635427
#> [4,] 0.5213125  1.3735739  2.1635427 6.0311569
#> 
#> , , 4, 1
#> 
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  5.4604321 1.6163492 -0.4140788 1.2317431
#> [2,]  1.6163492 5.8385341  0.2250014 0.0841567
#> [3,] -0.4140788 0.2250014  5.2779142 0.6681544
#> [4,]  1.2317431 0.0841567  0.6681544 4.6961547
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,] 11.3429486  0.7805658 -2.20368662  0.88044654
#> [2,]  0.7805658 14.2770813 -2.30524232 -2.38576619
#> [3,] -2.2036866 -2.3052423  9.66449182  0.01400067
#> [4,]  0.8804465 -2.3857662  0.01400067 14.05097368
#> 
#> , , 1, 2
#> 
#>            [,1]        [,2]      [,3]       [,4]
#> [1,] 1.65980743  0.09347025 0.2828233  0.1632905
#> [2,] 0.09347025  0.93478796 0.1252682 -0.2323680
#> [3,] 0.28282333  0.12526817 2.2623042  0.1230841
#> [4,] 0.16329047 -0.23236802 0.1230841  2.1482153
#> 
#> , , 2, 2
#> 
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.495588402 -0.09616551  0.005307419 -0.05571317
#> [2,] -0.096165506  1.67634458 -0.323353848  0.49164757
#> [3,]  0.005307419 -0.32335385  1.061227989  0.37585183
#> [4,] -0.055713172  0.49164757  0.375851833  1.62422212
#> 
#> , , 3, 2
#> 
#>           [,1]       [,2]      [,3]      [,4]
#> [1,]  1.460151 -1.9837852  1.881829 0.3010760
#> [2,] -1.983785 14.6096962 -8.901373 0.5877578
#> [3,]  1.881829 -8.9013735  9.169005 1.9948308
#> [4,]  0.301076  0.5877578  1.994831 3.8484503
#> 
#> , , 4, 2
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 1.3596546  0.0166310  0.0728546 0.2609229
#> [2,] 0.0166310  1.5206909 -0.2027718 0.1207857
#> [3,] 0.0728546 -0.2027718  2.6058244 1.1278802
#> [4,] 0.2609229  0.1207857  1.1278802 2.2665037
#> 
#> , , 5, 2
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.77109584 -0.8961348  0.08951381 -0.3247099
#> [2,] -0.89613482  6.2096403 -1.83362469  0.4508819
#> [3,]  0.08951381 -1.8336247  2.35159714  1.0278734
#> [4,] -0.32470994  0.4508819  1.02787341  1.6416176
#> 
#> , , 1, 3
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  3.34388952 -0.05438191 -1.0642605  0.03394041
#> [2,] -0.05438191  3.88013173  0.3428347 -1.85589236
#> [3,] -1.06426048  0.34283474  4.4588798  0.11538225
#> [4,]  0.03394041 -1.85589236  0.1153823  4.93832975
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  3.2457196  0.3633209 -0.32857911 -0.28377978
#> [2,]  0.3633209  4.5453231  0.71529453 -1.13233167
#> [3,] -0.3285791  0.7152945  3.63354163 -0.05924337
#> [4,] -0.2837798 -1.1323317 -0.05924337  5.47282066
#> 
#> , , 3, 3
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  2.5725906 -1.23795823  0.1519506 -0.48281075
#> [2,] -1.2379582 12.07906353 -6.3371784 -0.09022128
#> [3,]  0.1519506 -6.33717843  6.5796836  1.42407484
#> [4,] -0.4828107 -0.09022128  1.4240748  4.59066092
#> 
#> , , 4, 3
#> 
#>           [,1]        [,2]        [,3]       [,4]
#> [1,] 3.9752610  0.18931513  1.31285455  0.8278966
#> [2,] 0.1893151  2.55913761 -0.07683013 -0.4724878
#> [3,] 1.3128546 -0.07683013  2.72006118  0.1332462
#> [4,] 0.8278966 -0.47248785  0.13324621  2.5529014
#> 
#> , , 5, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.7175721 -1.0402697  0.6673323 -0.6508529
#> [2,] -1.0402697  7.3613465 -0.8582960  0.1241289
#> [3,]  0.6673323 -0.8582960  5.2542972  2.1772321
#> [4,] -0.6508529  0.1241289  2.1772321  5.3818774
#> 
#> , , 1, 4
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.90098176 -0.07234373 0.08597718  0.2061633
#> [2,] -0.07234373  1.87171105 0.58271619 -0.3266919
#> [3,]  0.08597718  0.58271619 1.97345228  0.3727000
#> [4,]  0.20616327 -0.32669190 0.37270001  1.4773722
#> 
#> , , 2, 4
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  1.33948732  0.20898515 -0.4095823 -0.09264293
#> [2,]  0.20898515  1.29633090 -0.4384980  0.00256348
#> [3,] -0.40958235 -0.43849801  2.0203358  0.15068385
#> [4,] -0.09264293  0.00256348  0.1506838  1.15796271
#> 
#> , , 3, 4
#> 
#>             [,1]         [,2]        [,3]         [,4]
#> [1,]  1.58253477 -0.429388103  0.06080232 -0.012645471
#> [2,] -0.42938810  8.577930033 -6.63149435 -0.002387891
#> [3,]  0.06080232 -6.631494354 11.37850079  3.342897624
#> [4,] -0.01264547 -0.002387891  3.34289762  6.250550964
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.1782006 -0.8273779  0.1131744  1.0217240
#> [2,] -0.8273779  3.1791672 -0.6276845 -0.4774978
#> [3,]  0.1131744 -0.6276845  2.1351201  0.2955927
#> [4,]  1.0217240 -0.4774978  0.2955927  3.0670413
#> 
#> , , 5, 4
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,] 1.46595707  0.03679488  0.4068059  0.3589054
#> [2,] 0.03679488  5.08688360 -3.6856427 -0.5220545
#> [3,] 0.40680586 -3.68564274  5.4756055  2.2647409
#> [4,] 0.35890542 -0.52205450  2.2647409  3.2920427
#> 
#> , , 1, 5
#> 
#>              [,1]         [,2]       [,3]      [,4]
#> [1,]  2.155138951 -0.007538923 -0.6416321 0.5146593
#> [2,] -0.007538923  2.804011134 -0.1250663 0.2184643
#> [3,] -0.641632137 -0.125066321  2.7725230 0.2684644
#> [4,]  0.514659327  0.218464279  0.2684644 3.4458620
#> 
#> , , 2, 5
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  4.68415591 -0.93285596 -0.06855414  0.02972493
#> [2,] -0.93285596  3.25613334  0.41747988 -0.07260786
#> [3,] -0.06855414  0.41747988  3.51206427  0.17137361
#> [4,]  0.02972493 -0.07260786  0.17137361  2.37023354
#> 
#> , , 3, 5
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.56363489 -0.1762171  0.7099997 0.05137257
#> [2,] -0.17621710 10.4351367 -6.1509712 0.42558467
#> [3,]  0.70999969 -6.1509712  7.2109940 1.78062028
#> [4,]  0.05137257  0.4255847  1.7806203 3.51332028
#> 
#> , , 4, 5
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 0.9227776 0.15550186 0.13356269 0.4023311
#> [2,] 0.1555019 1.14544991 0.08384711 0.2784088
#> [3,] 0.1335627 0.08384711 0.99920924 0.1011128
#> [4,] 0.4023311 0.27840882 0.10111284 1.4382680
#> 
#> , , 5, 5
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  2.7633503 -1.8691970  2.363309 0.2153406
#> [2,] -1.8691970  5.5951105 -3.323977 0.1291417
#> [3,]  2.3633094 -3.3239767  7.129429 2.0021263
#> [4,]  0.2153406  0.1291417  2.002126 3.5050642
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.86781097  0.7444781  0.7519076  0.5380911
#> [2,]  0.15101359  0.1912677 -0.9014054 -0.5302007
#> [3,] -0.48562438 -0.3508854 -0.4120694 -0.4126862
#> [4,]  0.49033839  0.1152907  1.0400839  1.1493085
#> [5,] -0.01140852 -0.6493351  2.1630486  2.4662320
#> 
#> , , 2, 1
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,] -1.0176402 -0.7666140  0.29305469 -0.19336471
#> [2,]  0.7821145  0.6169132  0.24836337  0.37077770
#> [3,]  1.8894251  1.4816492  0.87464055  1.20554556
#> [4,] -1.1056235 -0.8954333 -0.03133142 -0.07037766
#> [5,]  1.2657066 -2.6915431 -2.30810693  2.50977485
#> 
#> , , 3, 1
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.3177344 -0.1231916  0.2374395 -0.05899595
#> [2,]  0.8486037  0.2849430  0.7985416  0.47351903
#> [3,]  1.5103404 -0.4611785  1.9114006  0.87011434
#> [4,] -1.2130278  0.6060677 -1.0504120  0.11800984
#> [5,] -0.8529838  1.1923455 -1.1328072  2.98808940
#> 
#> , , 4, 1
#> 
#>             [,1]        [,2]      [,3]      [,4]
#> [1,]  0.03547513  0.40352013 0.1375843 0.1997981
#> [2,]  0.08002368  0.05425324 0.1774109 0.5127645
#> [3,]  0.26723222 -0.04257168 0.8097942 0.1887441
#> [4,]  0.07867028 -0.09192119 0.1230675 0.6902352
#> [5,] -0.61093879  1.78357579 0.8485660 3.8161076
#> 
#> , , 5, 1
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,]  1.182712  0.6228702  1.2469244  1.963601
#> [2,]  1.926401 -0.6300033 -0.1781240 -0.266629
#> [3,]  2.587364 -2.6653597  0.6933341 -1.087368
#> [4,] -2.811156  2.4378268 -0.3191084  1.180973
#> [5,]  4.132301 -3.9965754  5.5619735 -1.381511
#> 
#> , , 1, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.3285956  0.1064756 -0.3545454 -0.1042730
#> [2,]  0.3359758  0.8359584 -0.3489047 -0.2866748
#> [3,] -0.3366706 -0.2147077 -0.1034761 -0.2449974
#> [4,]  0.6004126  0.1391436  1.2172785  1.3261953
#> [5,] -0.6617468  1.9834402  1.0098959  0.1158997
#> 
#> , , 2, 2
#> 
#>             [,1]         [,2]       [,3]        [,4]
#> [1,]  0.89903836 -0.318496582 -0.1820259 -0.13126883
#> [2,]  0.02125831 -0.004824654  0.2239924 -0.08220793
#> [3,]  0.09152901  0.103229701  0.8447736  0.26464028
#> [4,] -0.05465839  0.232246225  0.1391190  0.82021688
#> [5,]  0.57893201 -0.152738117  0.2645049  1.28559731
#> 
#> , , 3, 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.5072663  0.52844516 -0.1659858  0.2056752
#> [2,]  0.2528776  0.30846819  0.4654169  0.2270274
#> [3,]  0.4628770 -0.04477439  1.1145274  0.4595387
#> [4,] -0.3001072  0.11406119 -0.2309547  0.5168098
#> [5,]  1.7986345 -5.68781883  4.6802502 -2.4299679
#> 
#> , , 4, 2
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.34680136  0.03941343 -0.26161985  0.1403636
#> [2,]  0.15981657  0.27212051  0.01163820  0.3547344
#> [3,]  0.29210163 -0.35581745  1.18058078  0.8544697
#> [4,] -0.07289883  0.37280211 -0.08667523  0.1258601
#> [5,]  0.56266287 -1.07647983  0.13573900 -1.3368680
#> 
#> , , 5, 2
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.5923409  0.7836645  1.33919954  1.6255487
#> [2,] -0.4062996  0.3335742 -0.14560571 -0.5052816
#> [3,] -0.4891875 -0.9388486  0.31737273 -1.1220975
#> [4,]  0.6720090  0.6331494  0.06162432  1.3753000
#> [5,]  0.6747170 -0.8752025 -1.06555594  0.1275393
#> 
#> , , 1, 3
#> 
#>            [,1]          [,2]        [,3]        [,4]
#> [1,]  0.5477609 -0.0003457972 -0.31774281 -0.19053727
#> [2,]  0.1582251  0.9207000108 -0.07834823 -0.07928227
#> [3,] -0.3286206 -0.2606457161 -0.24349157 -0.29803484
#> [4,]  0.5024362  0.1606130852  1.28490195  1.40476217
#> [5,] -0.2632865  4.1940432134  0.28961678 -0.69538249
#> 
#> , , 2, 3
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.86361194  0.41965319 0.79114133 -0.03519781
#> [2,] -0.02903551  0.06577917 0.02006284 -0.08253500
#> [3,] -0.16605449 -0.51118504 0.48791343 -0.02725803
#> [4,]  0.18361019  0.52332914 0.14043684  1.03334627
#> [5,]  0.67884822 -2.41393708 2.36091477  1.23568601
#> 
#> , , 3, 3
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,] 0.911010530 -0.4178813  1.15116101  0.6644098
#> [2,] 0.030263795  1.0269708 -0.62986886 -0.6124566
#> [3,] 0.010890884  0.5933699 -0.09354843 -0.8499388
#> [4,] 0.005371919 -0.3942887  0.70582071  1.6326318
#> [5,] 1.016492861  3.3178984 -3.21974187 -0.6433891
#> 
#> , , 4, 3
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.6346086 -0.02591917 -0.05525502 0.06057359
#> [2,]  0.2029120  0.13164002  0.09007786 0.46662883
#> [3,]  0.3947006 -0.09798794  1.06765079 0.57937656
#> [4,] -0.2766490  0.18845982 -0.06143351 0.39067041
#> [5,]  0.2398381 -4.40276838  0.24930911 0.64590307
#> 
#> , , 5, 3
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.73457265 -0.4497510  2.1760321  1.6274611
#> [2,]  0.02080182 -0.9638306 -0.4320668 -0.4067882
#> [3,] -0.01629123 -2.4858202 -0.4887552 -1.0954866
#> [4,]  0.15786600  2.8000807  0.4486237  1.2505528
#> [5,] -1.77680750 -1.9572135 -0.1653967  4.9168987
#> 
#> , , 1, 4
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.81612696 -0.27658190 -0.01693062  0.19375110
#> [2,] -0.09221368  1.17222732 -0.32539460 -0.40763995
#> [3,] -0.60248035  0.08606095  0.97761328 -0.08400052
#> [4,]  0.66231907  0.07030038  0.10263201  1.07478744
#> [5,] -0.20394220 -1.77094038 -0.89496812 -0.91144394
#> 
#> , , 2, 4
#> 
#>             [,1]        [,2]       [,3]         [,4]
#> [1,]  0.91493654 -0.01763881 0.08832381  0.154443332
#> [2,]  0.11608868  0.76300090 0.48215009 -0.270824669
#> [3,]  0.13054795  0.13028575 0.54449979 -0.001442509
#> [4,] -0.09107582 -0.04540097 0.23134012  0.986500261
#> [5,] -1.64432470 -0.66806392 1.51341673  1.848001437
#> 
#> , , 3, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.1318060 -0.7510633  0.7196198  0.1011798
#> [2,] -0.2679425  0.3887242  0.1474860 -0.1745539
#> [3,] -0.5222955 -0.1758479  0.6773959 -0.3223865
#> [4,]  0.4540187  0.6443337 -0.1202353  1.2554319
#> [5,]  0.3817229 -2.6476324  3.7115633  0.9663152
#> 
#> , , 4, 4
#> 
#>              [,1]        [,2]       [,3]       [,4]
#> [1,]  0.321024564  0.87020584 -0.1452649 0.16390207
#> [2,]  0.387074340  0.26303651  0.2538423 0.83379881
#> [3,] -0.004912771  0.04220952  0.8140313 0.04839585
#> [4,]  0.243780583 -0.30781222  0.2161528 0.85326865
#> [5,] -1.628306811 -0.57954758  0.4455653 1.62934762
#> 
#> , , 5, 4
#> 
#>              [,1]       [,2]       [,3]        [,4]
#> [1,]  0.684828349 -1.1500501  1.0470505  0.67238811
#> [2,] -0.004833239  0.8204854  0.2480789  0.02234988
#> [3,]  0.149981579  0.1370781  1.1368223 -0.09079419
#> [4,]  0.020993928  0.4046612 -0.6776814  0.71489028
#> [5,] -1.294145701  0.3474413  2.8775352  3.56669357
#> 
#> , , 1, 5
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.9438233 -0.5975370 -0.09648432  0.1180205
#> [2,] -0.2531633  1.2078249  0.10580676 -0.1728517
#> [3,] -0.4872538 -0.3419003  1.29903365  0.3770985
#> [4,]  0.5247093  0.5627490 -0.23476327  0.6568946
#> [5,] -0.3156766  0.6414849 -1.14853595 -1.8731601
#> 
#> , , 2, 5
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.36522412 -0.1710766 0.02614446 -0.3421648
#> [2,] -0.41543560  1.5103146 0.25092553 -0.5907817
#> [3,]  0.07046034  0.5265163 0.79804842 -0.3044241
#> [4,] -0.12115255 -0.4732167 0.09989827  1.4800555
#> [5,]  2.29647563 -1.4901914 0.29872395  0.6189869
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.6383832 -1.2062517  1.8508080  0.4188848
#> [2,] -0.4635706  0.9727044 -0.4910509  0.1161254
#> [3,] -0.7598746  0.9249144 -0.5223653  0.1597972
#> [4,]  0.5403383 -0.3088137  0.7806603  0.7597792
#> [5,] -1.4386891 -0.7984233 -4.4747549 -4.8685027
#> 
#> , , 4, 5
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.29976287 -0.1167771  0.4133609 -0.4530927
#> [2,]  0.08206447  0.1686149  0.1594470  0.3005333
#> [3,]  0.41707144 -0.2974858  1.1826242  0.6276975
#> [4,] -0.18476128  0.3430158 -0.3342176  0.5139580
#> [5,]  1.18329156  1.0250608  0.6165479  1.9743943
#> 
#> , , 5, 5
#> 
#>             [,1]         [,2]       [,3]       [,4]
#> [1,]  0.28343414 -0.793742876  0.9634847 0.50483519
#> [2,]  0.44580113  0.007846816  0.8874599 0.45805974
#> [3,]  0.80502220 -1.076818931  2.1273930 0.69432205
#> [4,] -0.47772606  1.463638419 -1.6229207 0.03140286
#> [5,]  0.01842151  1.154780862  2.3672878 3.30370400
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvars$get_last_draw`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|