R6 Class Representing PosteriorBVARs

The class PosteriorBVARs contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvars$new()

specify_posterior_bvars$get_posterior()

specify_posterior_bvars$get_last_draw()

specify_posterior_bvars$clone()

Method new()

Create a new posterior output PosteriorBVARs.

Usage

specify_posterior_bvars$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARs with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARs.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvars$get_posterior()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvars$get_last_draw()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvars$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARs" "R6"            

## ------------------------------------------------
## Method `specify_posterior_bvars$get_posterior`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $nu
#>      [,1] [,2] [,3]     [,4]     [,5]
#> [1,]  5.1  5.1  5.1 5.123394 5.143642
#> [2,]  5.1  5.1  5.1 5.100000 5.100000
#> [3,]  5.1  5.1  5.1 5.100000 5.100000
#> [4,]  5.1  5.1  5.1 5.090035 5.090035
#> [5,]  5.1  5.1  5.1 5.100000 5.100000
#> 
#> $m
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08343964 1.0515031 0.7739068 1.1001293 0.6674816
#> [2,] 0.78486174 0.7476969 0.3970339 0.8198663 0.9014390
#> [3,] 0.04644208 0.6208918 0.7339423 1.0460748 1.2684791
#> [4,] 0.61703399 0.5105634 1.0403655 0.9895968 0.9229582
#> [5,] 0.68146964 1.3017485 1.1418674 0.9305531 0.9522250
#> 
#> $w
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.06058857 0.12416546 0.07190060 0.09805918 0.14681302
#> [2,] 0.11872698 0.17987558 0.16031522 0.18912023 0.13749684
#> [3,] 0.11774075 0.07926549 0.05371738 0.06634388 0.07732836
#> [4,] 0.04460732 0.14389876 0.12088589 0.22605215 0.17493922
#> [5,] 0.36323688 0.32203897 0.15874633 0.14293268 0.11079231
#> 
#> $s
#>           [,1]     [,2]      [,3]     [,4]      [,5]
#> [1,] 790.78465 25.74401  93.91523 38.65917 109.98403
#> [2,]  92.39107 54.08184  80.11387 47.84752  77.79894
#> [3,] 170.69764 35.72816  56.74625 34.09090  64.58943
#> [4,]  51.21840 91.79366  46.46041 45.21078  74.14976
#> [5,] 370.28146 24.70767 100.60389 36.82341  69.00456
#> 
#> $scale
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
#> [4,]    0
#> [5,]    0
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 30.070142  3.663594  2.718339 -5.840346
#> [2,]  3.663594 21.540793  2.394953  0.085079
#> [3,]  2.718339  2.394953 19.533997  1.085430
#> [4,] -5.840346  0.085079  1.085430 24.019208
#> 
#> , , 2, 1
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  2.62392545 -0.04020554  0.4656462  0.5439299
#> [2,] -0.04020554  1.84879931 -0.2935592 -0.5209171
#> [3,]  0.46564625 -0.29355923  3.6672465  1.9843135
#> [4,]  0.54392985 -0.52091706  1.9843135  4.6031657
#> 
#> , , 3, 1
#> 
#>           [,1]      [,2]      [,3]      [,4]
#> [1,]  6.602160  1.894200 -2.484430 -1.646553
#> [2,]  1.894200 14.976464 -8.553059 -2.627716
#> [3,] -2.484430 -8.553059 11.322360  3.013962
#> [4,] -1.646553 -2.627716  3.013962  7.421785
#> 
#> , , 4, 1
#> 
#>             [,1]      [,2]        [,3]      [,4]
#> [1,]  1.44733415 0.3101745 -0.00589441 0.1900609
#> [2,]  0.31017446 1.8670747  0.27628686 1.3201996
#> [3,] -0.00589441 0.2762869  1.29712902 1.0247815
#> [4,]  0.19006094 1.3201996  1.02478147 4.7417087
#> 
#> , , 5, 1
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 10.360318  2.9987823  2.5602328  5.385389
#> [2,]  2.998782 13.3162624  0.3468727  3.378581
#> [3,]  2.560233  0.3468727 14.2444082  8.492410
#> [4,]  5.385389  3.3785806  8.4924105 26.929339
#> 
#> , , 1, 2
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  1.46212925  0.08823508  0.0879263 -0.2964542
#> [2,]  0.08823508  0.64787856 -0.2143389 -0.2634746
#> [3,]  0.08792630 -0.21433890  0.8968150  0.3532879
#> [4,] -0.29645422 -0.26347458  0.3532879  1.3429130
#> 
#> , , 2, 2
#> 
#>              [,1]        [,2]       [,3]         [,4]
#> [1,]  1.181910679 -0.02517107 -0.1049211 -0.007355683
#> [2,] -0.025171067  2.28323014 -0.7643654  0.143693502
#> [3,] -0.104921087 -0.76436536  2.0597255  0.360294488
#> [4,] -0.007355683  0.14369350  0.3602945  2.430341023
#> 
#> , , 3, 2
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 0.7840941  0.1262284  0.4983287 0.4580784
#> [2,] 0.1262284 13.8884021 -5.4153018 1.6412324
#> [3,] 0.4983287 -5.4153018  7.9395515 3.3203284
#> [4,] 0.4580784  1.6412324  3.3203284 5.6018606
#> 
#> , , 4, 2
#> 
#>             [,1]        [,2]       [,3]      [,4]
#> [1,]  3.50439379 -0.06868815 -0.3247804 1.2624439
#> [2,] -0.06868815  3.52265676 -0.1052999 0.8976767
#> [3,] -0.32478044 -0.10529987  3.5329977 0.1098620
#> [4,]  1.26244388  0.89767667  0.1098620 4.2289359
#> 
#> , , 5, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8578491  0.3557020 -0.1574853 -0.4861850
#> [2,]  0.3557020  3.0859431 -2.0058147 -0.9124436
#> [3,] -0.1574853 -2.0058147  2.3298280  1.2297736
#> [4,] -0.4861850 -0.9124436  1.2297736  2.4331811
#> 
#> , , 1, 3
#> 
#>             [,1]       [,2]        [,3]      [,4]
#> [1,]  3.25383318 1.04129726 -0.07782694 0.5010790
#> [2,]  1.04129726 3.44711411  0.09941885 1.0551390
#> [3,] -0.07782694 0.09941885  2.53439977 0.1587111
#> [4,]  0.50107903 1.05513905  0.15871112 3.3968973
#> 
#> , , 2, 3
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,] 2.72988980  0.01453668  0.1566966  0.4266220
#> [2,] 0.01453668  2.23319421 -0.1685964 -0.0340416
#> [3,] 0.15669656 -0.16859644  2.4597962  0.4670095
#> [4,] 0.42662199 -0.03404160  0.4670095  2.5875658
#> 
#> , , 3, 3
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  2.10739898  0.1458951 -0.2981538 0.06410204
#> [2,]  0.14589506 11.2894404 -5.9897854 0.28286802
#> [3,] -0.29815381 -5.9897854  6.7154321 1.41527939
#> [4,]  0.06410204  0.2828680  1.4152794 3.03544737
#> 
#> , , 4, 3
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,] 1.99836536  0.16065527  0.01933161 0.11889677
#> [2,] 0.16065527  1.30754453 -0.13983524 0.09460732
#> [3,] 0.01933161 -0.13983524  2.33669240 1.16355455
#> [4,] 0.11889677  0.09460732  1.16355455 3.18155259
#> 
#> , , 5, 3
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  3.0477793  0.7592368 -1.695691 -0.3117723
#> [2,]  0.7592368  6.0862522 -2.557785 -0.6178953
#> [3,] -1.6956912 -2.5577850 10.028459  3.1914446
#> [4,] -0.3117723 -0.6178953  3.191445  5.5636785
#> 
#> , , 1, 4
#> 
#>             [,1]       [,2]      [,3]        [,4]
#> [1,]  1.39104323 -0.1345668 0.4359417  0.09355107
#> [2,] -0.13456677  0.9241117 0.0916093 -0.09927540
#> [3,]  0.43594174  0.0916093 1.3295083  0.33808273
#> [4,]  0.09355107 -0.0992754 0.3380827  1.06594161
#> 
#> , , 2, 4
#> 
#>             [,1]         [,2]        [,3]         [,4]
#> [1,]  1.28756665  0.167142433  0.03961068 -0.239945887
#> [2,]  0.16714243  2.173129614 -0.18373066 -0.006268753
#> [3,]  0.03961068 -0.183730658  1.72752932  0.051224745
#> [4,] -0.23994589 -0.006268753  0.05122474  2.313106291
#> 
#> , , 3, 4
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,] 1.37703442  0.08348745  0.3146602  0.8023204
#> [2,] 0.08348745  9.60705092 -6.5257849 -0.7578334
#> [3,] 0.31466024 -6.52578493  7.8404349  2.3788894
#> [4,] 0.80232037 -0.75783343  2.3788894  3.5891630
#> 
#> , , 4, 4
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  1.249268516 -0.03010412  0.14580890  0.002823457
#> [2,] -0.030104115  2.15144362 -0.05188379 -0.486585204
#> [3,]  0.145808903 -0.05188379  1.51785143  0.117483090
#> [4,]  0.002823457 -0.48658520  0.11748309  1.231677919
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  1.1165750 -0.2541354 -0.518840 -0.6315480
#> [2,] -0.2541354  3.9714197 -1.324823  0.1183726
#> [3,] -0.5188400 -1.3248234  3.212064  2.0337891
#> [4,] -0.6315480  0.1183726  2.033789  3.6361611
#> 
#> , , 1, 5
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  4.21330749 -0.04282325 1.29577269  0.07911359
#> [2,] -0.04282325  3.73281094 0.11317267 -0.10404884
#> [3,]  1.29577269  0.11317267 3.97226353  0.05901886
#> [4,]  0.07911359 -0.10404884 0.05901886  3.57679398
#> 
#> , , 2, 5
#> 
#>              [,1]         [,2]        [,3]       [,4]
#> [1,]  3.120408467 -0.005618462 -0.05728169 -1.0893600
#> [2,] -0.005618462  3.362459532  0.64252831  0.6324848
#> [3,] -0.057281686  0.642528315  2.56363447  0.4785000
#> [4,] -1.089360044  0.632484796  0.47850001  3.1159843
#> 
#> , , 3, 5
#> 
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  2.7343573  0.835277 -0.6248547 0.3896288
#> [2,]  0.8352770 12.823967 -7.5498301 2.0350068
#> [3,] -0.6248547 -7.549830  8.0946316 1.0982454
#> [4,]  0.3896288  2.035007  1.0982454 4.4511719
#> 
#> , , 4, 5
#> 
#>              [,1]       [,2]      [,3]         [,4]
#> [1,]  1.521097320 -0.3421161 0.3275691 -0.006363214
#> [2,] -0.342116053  1.9830079 0.5771992  0.215318760
#> [3,]  0.327569143  0.5771992 2.3189459  0.164863542
#> [4,] -0.006363214  0.2153188 0.1648635  2.044871388
#> 
#> , , 5, 5
#> 
#>           [,1]        [,2]      [,3]       [,4]
#> [1,] 1.7605510  0.19109093  0.165578 0.36385997
#> [2,] 0.1910909  5.24212102 -2.207603 0.09104784
#> [3,] 0.1655780 -2.20760313  4.275090 1.17439481
#> [4,] 0.3638600  0.09104784  1.174395 3.39841424
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]        [,2]          [,3]        [,4]
#> [1,] 0.05494131 -0.01746692  0.0636204612 -0.08427820
#> [2,] 0.68777872  1.33265154  0.0008299849 -0.17283321
#> [3,] 0.07743732  0.65992513 -0.3618472619  0.01230593
#> [4,] 0.24519400 -0.54154098  1.1832223249  1.08210673
#> [5,] 2.66216566 -4.89425683  0.8058259728 -2.64770170
#> 
#> , , 2, 1
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.53147769 -0.3311989  0.4166752 0.05664712
#> [2,]  0.16903895  1.2361942  0.6691225 1.35409716
#> [3,]  0.20439715  1.2471016 -0.3110587 0.11842520
#> [4,] -0.05864933 -0.9670630  0.8045290 0.55399434
#> [5,]  0.14679148 -1.6773121  1.5424776 3.43508652
#> 
#> , , 3, 1
#> 
#>            [,1]        [,2]       [,3]      [,4]
#> [1,]  0.8176377 -0.51206983  1.0003208 0.3160291
#> [2,] -0.2253798  0.44737291 -0.4315325 0.2010740
#> [3,] -0.1488735  0.31778733 -0.2859445 0.1619401
#> [4,]  0.1579619 -0.05307595  0.9033499 0.6818847
#> [5,]  6.2963635  7.53383955 -5.6032699 0.3465264
#> 
#> , , 4, 1
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,] -0.26948428 -0.35946389  0.04298267 -0.3927318
#> [2,] -0.12260424  0.12547481 -0.48006965 -1.0359294
#> [3,]  0.41389096  0.12303042  1.20926894  0.8173237
#> [4,]  0.02663936  0.04764898 -0.21873554  0.3869949
#> [5,]  0.63374431 -1.17523292  1.06385488 -1.6572879
#> 
#> , , 5, 1
#> 
#>           [,1]      [,2]      [,3]       [,4]
#> [1,]  1.360819 -1.190795  1.224666  0.2652236
#> [2,] -1.366214  0.248318 -3.897984 -3.4545171
#> [3,] -2.463069 -1.108858 -6.117575 -5.8535390
#> [4,]  2.062625  1.567917  6.210483  6.2969257
#> [5,] 12.896333  4.862884 16.814138 25.6377374
#> 
#> , , 1, 2
#> 
#>            [,1]        [,2]          [,3]        [,4]
#> [1,]  0.1252594  0.09366114 -0.3204348215 -0.09411518
#> [2,]  0.3524479  0.96383887  0.0009452932  0.15734080
#> [3,] -0.3092248  0.13588895  0.8748129830  0.21123024
#> [4,]  0.6464995 -0.15111332  0.2562079155  0.82885385
#> [5,]  0.1816792 -0.42884974  0.0866476045 -0.80596161
#> 
#> , , 2, 2
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.83471868 -0.5916288  0.4377358 0.61230156
#> [2,] -0.21242040  0.7283961  0.9153019 0.58312402
#> [3,]  0.22929094 -1.0547476  1.3605411 0.02984584
#> [4,] -0.09552756  1.1138675 -0.6480195 0.64303759
#> [5,]  0.40069466  1.5184822  0.7180400 0.83270336
#> 
#> , , 3, 2
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.01366548 -0.5320815  1.2301156 0.60742521
#> [2,]  0.07110464  1.0229533 -0.3724578 0.03555584
#> [3,]  0.07211929  0.7427597 -0.2425336 0.04342652
#> [4,] -0.07906520 -0.4365415  0.6431346 0.72371477
#> [5,] -0.48088371  1.6370396  1.0033620 0.21144035
#> 
#> , , 4, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.2176399 -0.5725756 -0.1826127  0.1245301
#> [2,]  0.3501360  0.7391753 -0.2905335 -0.5189617
#> [3,]  0.7965566  0.4675226  1.3376590  1.1857550
#> [4,] -0.4678210 -0.2866978 -0.2760695 -0.1722122
#> [5,] -4.0963505  2.6199020  1.1945079 -0.6000127
#> 
#> , , 5, 2
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.82148449 -1.08067476  1.05105767  0.6965820
#> [2,] -0.01559882  1.10411786 -0.18200640 -0.3007417
#> [3,]  0.08059623  0.44934321  0.50654641 -0.4942591
#> [4,]  0.02995625  0.06005451  0.06701932  1.2388967
#> [5,] -0.89596176 -0.16315733 -3.27558612 -3.0756288
#> 
#> , , 1, 3
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.53746618  0.04852806 -0.15307880 -0.1424062
#> [2,]  0.19699476  0.81945796 -0.18192151 -0.1128465
#> [3,] -0.05620899  0.10845065  0.87800345  0.5794436
#> [4,]  0.25264748 -0.09397676  0.22248820  0.6144514
#> [5,]  0.39948030  0.30000209 -0.08952718 -2.4487353
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.4819789 -0.3891834  0.7217937 -0.4251330
#> [2,]  0.4544683  0.2566281  0.5566578  1.8275042
#> [3,]  0.7889783 -0.3868967  0.7436618  1.2490559
#> [4,] -0.5737931  0.6244342 -0.1009521 -0.2765581
#> [5,] -1.0608566 -1.1245592 -3.1199649 -1.3959398
#> 
#> , , 3, 3
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  1.0482367 -0.16689569  0.82607262  0.6058633
#> [2,]  0.1632918  0.64339129  0.01191276 -0.1379630
#> [3,]  0.1463107  0.26205280  0.37194753 -0.2056422
#> [4,] -0.1983915 -0.07044825  0.23334491  0.9975308
#> [5,]  1.0222980  0.60033727 -1.02839503 -1.0644922
#> 
#> , , 4, 3
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  1.0366936  0.5299503  0.4024742  0.23054413
#> [2,]  0.3191193  0.4322323 -0.4907341 -0.70029937
#> [3,]  0.6096880  0.2726417  1.0604285  1.02231314
#> [4,] -0.6093444 -0.4152812 -0.2098311 -0.08038365
#> [5,] -0.7575856 -0.8030316  2.0300146  1.94968158
#> 
#> , , 5, 3
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.2799528 -0.23434259  0.1144399  0.1915132
#> [2,] -0.1288794  0.95085001  0.3636604 -0.5759939
#> [3,] -0.5292914  0.06198239  1.5638644 -0.6181631
#> [4,]  0.3741609  0.03462888 -0.5501225  1.5558828
#> [5,] -0.2967120  1.55268287 -4.1117065  1.0877158
#> 
#> , , 1, 4
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.53443206  0.26947149 -0.31586612 -0.06129872
#> [2,]  0.25467888  0.77015936  0.11898421 -0.19209001
#> [3,]  0.05615232  0.03919977  1.24023870  0.05720691
#> [4,]  0.14026262 -0.14724981 -0.07674516  1.01494927
#> [5,] -0.22416320  1.28997984 -0.43023468 -0.31597090
#> 
#> , , 2, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8326618 -0.4068961  0.7848433  0.3244175
#> [2,] -0.5526705  0.6257315  0.6487430  0.5337107
#> [3,] -0.3295254  0.3417374  0.8970538 -0.1097982
#> [4,]  0.4208983 -0.1202236 -0.2966215  0.9115191
#> [5,]  1.7132357  3.2800255 -0.7595867 -2.9058119
#> 
#> , , 3, 4
#> 
#>              [,1]        [,2]       [,3]        [,4]
#> [1,]  1.103149480 -0.50658640  0.6808255  0.08833602
#> [2,]  0.126785022  0.56678306  0.3842409  0.32108485
#> [3,]  0.001163582  0.27347765  1.0013260  0.48044018
#> [4,] -0.052740443  0.08436196 -0.3537526  0.49768671
#> [5,] -0.912660611  0.35114081 -0.5262740 -0.91190722
#> 
#> , , 4, 4
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.11546557  0.3277329  0.6845750  0.6391751
#> [2,] -0.03211554  0.4499176 -0.4522667 -0.1622060
#> [3,] -0.20490975 -0.5762674  0.0333915 -0.1815980
#> [4,]  0.12657507  0.4719361  0.6991834  0.9742750
#> [5,]  2.48076791 -0.4004985  2.1412388 -0.9001187
#> 
#> , , 5, 4
#> 
#>             [,1]        [,2]       [,3]      [,4]
#> [1,]  0.89653622 -0.41348737  1.1627921  1.196288
#> [2,]  0.10418390  0.90092165 -0.3498124 -0.464978
#> [3,]  0.19072236  0.11672641  0.2770484 -0.719409
#> [4,] -0.14279502  0.09837388  0.1409021  1.166088
#> [5,]  0.05127598 -0.53728287  3.0515938  1.644614
#> 
#> , , 1, 5
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  0.76367950  0.22990721  0.1094939  0.2270846
#> [2,] -0.06605469  0.77641699 -0.1057001 -0.3456980
#> [3,]  0.45420729  0.08634091  1.4546574 -0.4254353
#> [4,] -0.30014139 -0.09611436 -0.5199299  1.3427628
#> [5,]  1.23019680 -2.46972962  3.1835556 -0.2491466
#> 
#> , , 2, 5
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.66891202 -0.2511197  0.08029076 -0.2308749
#> [2,] -0.16993534  0.9237945  0.96581878  0.1500411
#> [3,]  0.19777389 -0.6875875  1.70622245  0.3949619
#> [4,] -0.04741309  0.6991540 -0.80649707  0.7505783
#> [5,]  2.35551494 -1.8257630 -1.14632677 -3.3560217
#> 
#> , , 3, 5
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  1.3500256 -0.51392850  0.69227691  0.35570753
#> [2,] -0.1087717  0.49890423  0.15659057 -0.06616198
#> [3,] -0.4090218  0.05193676  0.60468176 -0.13382088
#> [4,]  0.2199589  0.28857345  0.08785116  1.01137536
#> [5,]  0.9447446 -0.96394243 -2.84213333 -0.85470658
#> 
#> , , 4, 5
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  1.40972961 -1.19695311  0.50915060  0.6890987
#> [2,]  0.48790221  1.39111931  0.07877555 -0.8461907
#> [3,]  0.05885974  0.47560909  1.33462054 -0.4399909
#> [4,] -0.21836558 -0.04558825 -0.46979135  1.2381218
#> [5,]  0.72718117 -1.62263210 -2.81722531 -1.3681180
#> 
#> , , 5, 5
#> 
#>             [,1]        [,2]         [,3]       [,4]
#> [1,]  0.57958766 -0.61589113  0.715406493 0.38955965
#> [2,]  0.02198514  1.07761645  0.009520219 0.31485917
#> [3,]  0.15017625  0.28685146  0.851202512 0.07824087
#> [4,] -0.01213427  0.08233657 -0.229021586 0.70006634
#> [5,]  3.35007102 -4.37626699  3.260542033 0.88627349
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvars$get_last_draw`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|