R6 Class Representing PosteriorBVARs

The class PosteriorBVARs contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

specify_posterior_bvars$new()

specify_posterior_bvars$get_posterior()

specify_posterior_bvars$get_last_draw()

specify_posterior_bvars$clone()

Method new()

Create a new posterior output PosteriorBVARs.

Usage

specify_posterior_bvars$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARs with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARs.

Method get_posterior()

Returns a list containing Bayesian estimation output.

Usage

specify_posterior_bvars$get_posterior()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

Method get_last_draw()

Returns an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

specify_posterior_bvars$get_last_draw()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

Method clone()

The objects of this class are cloneable with this method.

Usage

specify_posterior_bvars$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARs" "R6"            

## ------------------------------------------------
## Method `specify_posterior_bvars$get_posterior`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $nu
#>      [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,]  5.1 5.100000 5.100000 5.100000 5.100000
#> [2,]  5.1 5.100000 5.021932 5.021932 5.021932
#> [3,]  5.1 5.100000 5.100000 5.100000 5.100000
#> [4,]  5.1 5.100000 5.100000 5.100000 5.100000
#> [5,]  5.1 5.045609 5.045609 5.045609 5.045609
#> 
#> $m
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.5613640 0.2933729 0.7243536 0.8325434 1.2500060
#> [2,] 0.4443765 0.4977583 0.4869349 0.9525539 1.1797791
#> [3,] 0.6428730 1.1528701 0.7438036 0.7847814 0.8204669
#> [4,] 0.3989072 1.0780996 1.2145781 1.1611249 0.3560437
#> [5,] 0.6414850 0.8413493 1.0848475 0.5846717 0.6187250
#> 
#> $w
#>            [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.04782878 0.09559593 0.1451897 0.18084386 0.14464553
#> [2,] 0.10455075 0.17937258 0.5017863 0.21027857 0.28710528
#> [3,] 0.17911642 0.07668942 0.1078367 0.08200389 0.14254556
#> [4,] 0.07763874 0.13707454 0.1216487 0.15448169 0.09043211
#> [5,] 0.09912711 0.18190411 0.2551021 0.22907651 0.16550662
#> 
#> $s
#>           [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,]  91.11386 49.49357 62.34742 74.03673 50.31307
#> [2,]  86.93099 62.46505 52.24670 43.53386 82.13950
#> [3,] 123.94348 27.01018 70.42957 40.56146 68.43603
#> [4,]  97.10768 53.32954 59.28958 63.68573 63.68704
#> [5,]  54.57776 53.73122 50.39751 57.56760 59.11718
#> 
#> $scale
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
#> [4,]    0
#> [5,]    0
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  2.5146355 -0.278231  0.8206233  1.178720
#> [2,] -0.2782310  3.008120 -2.2467380 -2.133656
#> [3,]  0.8206233 -2.246738  6.5031794  4.759373
#> [4,]  1.1787199 -2.133656  4.7593731  7.797959
#> 
#> , , 2, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.5255640 -0.1918720 -0.4963968 -0.8572944
#> [2,] -0.1918720  2.2862194  0.3450324  0.7858094
#> [3,] -0.4963968  0.3450324  4.7637531  4.7494529
#> [4,] -0.8572944  0.7858094  4.7494529  9.0226857
#> 
#> , , 3, 1
#> 
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  4.1702785 -1.073180 -0.6525808 -1.042972
#> [2,] -1.0731796 15.940468 -6.0844721  2.527067
#> [3,] -0.6525808 -6.084472  9.9205006  5.480293
#> [4,] -1.0429723  2.527067  5.4802931 11.790506
#> 
#> , , 4, 1
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  2.3535847  0.5625872 0.2833452 -0.3450808
#> [2,]  0.5625872  2.4936328 0.4177893 -0.3797039
#> [3,]  0.2833452  0.4177893 2.5829603  0.7961015
#> [4,] -0.3450808 -0.3797039 0.7961015  3.9638791
#> 
#> , , 5, 1
#> 
#>           [,1]      [,2]      [,3]     [,4]
#> [1,] 4.1296666 0.7088587 3.1452112 3.535183
#> [2,] 0.7088587 4.2707393 0.2556604 1.110866
#> [3,] 3.1452112 0.2556604 8.7441516 7.117853
#> [4,] 3.5351828 1.1108664 7.1178534 8.763143
#> 
#> , , 1, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  3.1118818  0.7380750 -0.6249836 -0.9980817
#> [2,]  0.7380750  2.7318303 -0.3146736 -0.3446382
#> [3,] -0.6249836 -0.3146736  1.2044717  0.4838091
#> [4,] -0.9980817 -0.3446382  0.4838091  2.4926113
#> 
#> , , 2, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  3.4045770  0.2198437 -0.2713432 -1.4034100
#> [2,]  0.2198437  1.8237776 -0.3505550  0.2674153
#> [3,] -0.2713432 -0.3505550  4.3594225  3.0511971
#> [4,] -1.4034100  0.2674153  3.0511971  6.0905143
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.1981152 -0.8463302  0.8867596  1.3645978
#> [2,] -0.8463302  8.9994958 -5.1127657 -0.8027083
#> [3,]  0.8867596 -5.1127657  5.5894518  3.0939949
#> [4,]  1.3645978 -0.8027083  3.0939949  6.5553337
#> 
#> , , 4, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.0899845 -0.4836626  0.1156268  0.5089021
#> [2,] -0.4836626  2.4260118 -0.6126543 -0.2585652
#> [3,]  0.1156268 -0.6126543  3.5434452 -0.3325637
#> [4,]  0.5089021 -0.2585652 -0.3325637  2.1961970
#> 
#> , , 5, 2
#> 
#>          [,1]       [,2]      [,3]      [,4]
#> [1,] 2.387189  0.6507650  2.082294 1.7450914
#> [2,] 0.650765  4.8443455 -1.639920 0.3088597
#> [3,] 2.082294 -1.6399198 10.182670 6.7598721
#> [4,] 1.745091  0.3088597  6.759872 7.8308571
#> 
#> , , 1, 3
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  2.0792975  0.17542956 -0.22351176  0.23129570
#> [2,]  0.1754296  2.39163701 -0.07329105 -0.08997636
#> [3,] -0.2235118 -0.07329105  1.91230343  0.52961887
#> [4,]  0.2312957 -0.08997636  0.52961887  2.47347038
#> 
#> , , 2, 3
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  1.07250092 0.14017395 0.07400933 -0.07005990
#> [2,]  0.14017395 2.42450919 0.02237765  0.07666399
#> [3,]  0.07400933 0.02237765 3.83982446  2.82042758
#> [4,] -0.07005990 0.07666399 2.82042758  5.27187461
#> 
#> , , 3, 3
#> 
#>           [,1]       [,2]      [,3]       [,4]
#> [1,] 3.0371673  0.1062954  0.156001  3.1736833
#> [2,] 0.1062954 12.2367283 -5.976356  0.6295584
#> [3,] 0.1560010 -5.9763561  9.045164  3.2405215
#> [4,] 3.1736833  0.6295584  3.240521 11.2058425
#> 
#> , , 4, 3
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  1.75621432 -0.6970593 0.08218854 -0.1560034
#> [2,] -0.69705926  3.1632185 0.13727661  0.8129425
#> [3,]  0.08218854  0.1372766 2.05940383  0.7005369
#> [4,] -0.15600341  0.8129425 0.70053688  2.7027276
#> 
#> , , 5, 3
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,] 1.3066497  0.2850960  0.5404484 0.5857297
#> [2,] 0.2850960  5.5099765 -2.6654513 0.2161564
#> [3,] 0.5404484 -2.6654513  6.1232049 2.7877425
#> [4,] 0.5857297  0.2161564  2.7877425 4.6311252
#> 
#> , , 1, 4
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  2.4355178 0.24430690 -0.33655409 -0.97657997
#> [2,]  0.2443069 2.13383582  0.03271768  0.08483467
#> [3,] -0.3365541 0.03271768  2.85247273  1.82577707
#> [4,] -0.9765800 0.08483467  1.82577707  4.15080076
#> 
#> , , 2, 4
#> 
#>              [,1]      [,2]         [,3]       [,4]
#> [1,]  1.266747532 0.2110214 -0.002356657 -0.1875587
#> [2,]  0.211021408 1.1623397  0.466835702  0.7418048
#> [3,] -0.002356657 0.4668357  3.358676502  2.6995311
#> [4,] -0.187558692 0.7418048  2.699531109  4.0402019
#> 
#> , , 3, 4
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  1.59840145  0.06587389 -0.6879772 -0.1530334
#> [2,]  0.06587389  8.42933027 -5.1799331 -0.3686412
#> [3,] -0.68797716 -5.17993312  7.6455297  3.6656334
#> [4,] -0.15303342 -0.36864125  3.6656334  6.1029289
#> 
#> , , 4, 4
#> 
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  1.963231848 -0.43625893 0.303115085  0.009181291
#> [2,] -0.436258933  2.58264701 0.086036522 -0.045211494
#> [3,]  0.303115085  0.08603652 2.348801388  0.003946573
#> [4,]  0.009181291 -0.04521149 0.003946573  1.562196894
#> 
#> , , 5, 4
#> 
#>           [,1]      [,2]       [,3]     [,4]
#> [1,] 2.8619932  1.441916  0.4395004 1.577329
#> [2,] 1.4419156  6.015635 -2.1152171 1.152903
#> [3,] 0.4395004 -2.115217  5.9784471 2.996531
#> [4,] 1.5773288  1.152903  2.9965308 4.822295
#> 
#> , , 1, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.6760269 -0.2555890  0.1378689 -0.1021293
#> [2,] -0.2555890  1.4384216 -0.1936358 -0.4442652
#> [3,]  0.1378689 -0.1936358  1.6508491  0.2276359
#> [4,] -0.1021293 -0.4442652  0.2276359  2.2570805
#> 
#> , , 2, 5
#> 
#>           [,1]       [,2]       [,3]       [,4]
#> [1,] 1.7529025  0.1666594  0.2779032  0.1824827
#> [2,] 0.1666594  1.7938743 -0.7374194 -0.2856556
#> [3,] 0.2779032 -0.7374194  2.9079435 -0.4830011
#> [4,] 0.1824827 -0.2856556 -0.4830011  2.2563676
#> 
#> , , 3, 5
#> 
#>            [,1]        [,2]       [,3]      [,4]
#> [1,] 1.78980336  0.01398801  0.5321651  1.701394
#> [2,] 0.01398801  8.37117145 -4.6437235 -1.008610
#> [3,] 0.53216514 -4.64372354  8.2634139  5.218604
#> [4,] 1.70139356 -1.00860988  5.2186044  9.934603
#> 
#> , , 4, 5
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.4381202  0.2312548 -0.4930679 -0.3678188
#> [2,]  0.2312548  1.7628933 -0.2320201 -0.1265411
#> [3,] -0.4930679 -0.2320201  1.5524826  0.2516223
#> [4,] -0.3678188 -0.1265411  0.2516223  2.1797863
#> 
#> , , 5, 5
#> 
#>            [,1]        [,2]      [,3]       [,4]
#> [1,] 2.44953703  0.05143969  3.167375  2.5789281
#> [2,] 0.05143969  5.16443353 -2.680834 -0.9742608
#> [3,] 3.16737525 -2.68083361 10.097801  6.6872415
#> [4,] 2.57892812 -0.97426083  6.687242  7.5017009
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.60972360  0.5636386 -0.2820637  0.24145327
#> [2,]  0.25874213  0.1957914  0.6342903  0.65916913
#> [3,] -0.08600404 -0.6048046  1.3602900  0.87093026
#> [4,]  0.19964469  0.4768909 -0.3067818 -0.03547298
#> [5,]  1.49604781 -2.1365153 -0.7782360 -0.06375134
#> 
#> , , 2, 1
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.8554941  0.2222596  2.19570029  2.1507899
#> [2,]  0.1344096  0.7781761  0.80778694  1.4290890
#> [3,] -0.5370315 -0.2570142 -0.21150447 -1.0995084
#> [4,]  0.4538894  0.1592495  0.09631789  0.8283117
#> [5,]  0.7444705  2.6244208  2.87778769  7.2638848
#> 
#> , , 3, 1
#> 
#>            [,1]         [,2]        [,3]       [,4]
#> [1,]  0.7681685 -0.882949150  0.08378202 -0.8625881
#> [2,]  0.3569136  0.367381051  0.25377342  0.3557843
#> [3,]  0.5866462 -0.004995039  1.03200536  1.0387019
#> [4,] -0.5122061  0.438506002 -0.17113986  0.3304467
#> [5,]  0.5789190  4.399123057  1.80723818  1.6621804
#> 
#> , , 4, 1
#> 
#>             [,1]        [,2]        [,3]      [,4]
#> [1,]  0.12273086 -0.02240771 -0.02593594 0.1748680
#> [2,]  0.07369225 -0.04576926  0.10486409 0.3516778
#> [3,] -0.29620957 -0.01560593  0.59707666 0.3780147
#> [4,]  0.59671273  0.09277644  0.39366707 0.5518088
#> [5,] -0.53889701 -2.80986757  0.15563022 0.3896734
#> 
#> , , 5, 1
#> 
#>             [,1]       [,2]      [,3]       [,4]
#> [1,]  0.88311659 -0.1054753  1.499501  0.6355428
#> [2,] -0.90425539  0.3100616 -1.848455 -1.6693208
#> [3,] -1.43971179 -0.8755266 -2.518902 -3.0467837
#> [4,]  1.48935600  0.8876698  2.728481  3.6507093
#> [5,] -0.08761045  3.8346668  3.999930  4.8594131
#> 
#> , , 1, 2
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,] 0.04598335  0.1909637  0.5230328  0.26747190
#> [2,] 0.69115320  0.6397888 -0.4587016  0.23511787
#> [3,] 0.19234843 -0.7664368  1.1259031  0.91626112
#> [4,] 0.15477023  0.6585649 -0.2789737  0.02101817
#> [5,] 0.56798175  1.0560513 -0.7423577 -2.03022132
#> 
#> , , 2, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9981798  0.2396226  2.2138207  2.6912319
#> [2,] -0.5312149  1.0951818  0.4768155  1.2036876
#> [3,] -0.3857459  0.2914832 -0.8824784 -0.7092645
#> [4,]  0.4035157 -0.3558654  0.7344006  0.5150616
#> [5,]  2.1927567  0.8319581  2.3543694 -3.1891249
#> 
#> , , 3, 2
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9333355 -0.64350270  0.7480105  0.2656050
#> [2,]  0.4113832  0.31979988  0.4630991  0.6230205
#> [3,]  0.5378125 -0.07446034  1.1269727  1.0624055
#> [4,] -0.5033611  0.40149979 -0.5193764 -0.1882701
#> [5,] -1.7042015  3.56560732  0.7806000  1.6663546
#> 
#> , , 4, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.1490484  0.4825389 -0.1070199 -0.1715873
#> [2,]  0.3437771  0.8629741  0.3286557 -0.1751259
#> [3,] -0.0659030  0.2855417  1.6960171 -0.5476684
#> [4,]  0.3164741 -0.4117916 -0.6359537  1.5991517
#> [5,]  2.2424108 -2.2605422 -1.4796968 -0.1452711
#> 
#> , , 5, 2
#> 
#>            [,1]        [,2]      [,3]      [,4]
#> [1,]  1.8346956  0.06153231  2.226190  1.402963
#> [2,] -0.5303861  0.17664878 -1.070267 -1.200765
#> [3,] -1.1331113 -1.00933201 -1.486941 -2.523992
#> [4,]  0.7404275  1.04863930  1.356510  2.765128
#> [5,] -0.5345733 -1.84273868  3.962437  4.167722
#> 
#> , , 1, 3
#> 
#>            [,1]        [,2]       [,3]         [,4]
#> [1,]  1.1020288 -0.38917446  0.2308560  0.618809567
#> [2,] -0.3086299  1.13855281 -0.2867448 -0.116254675
#> [3,] -0.8906421 -0.02975707  1.7094322  0.842553193
#> [4,]  0.8447872  0.15604671 -0.6700996  0.007931203
#> [5,] -1.2126485  1.44699606 -1.9242137 -3.031766323
#> 
#> , , 2, 3
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.9381996  0.1083505  2.07140531  2.4134662
#> [2,] -0.2850977  1.6977399 -0.08904053 -0.9437194
#> [3,] -0.2679050  0.2153713 -0.60829927 -2.4771404
#> [4,]  0.3130358 -0.3627465  0.61323638  2.4076167
#> [5,] -0.6316524  0.1651691  6.27718400  6.6604975
#> 
#> , , 3, 3
#> 
#>            [,1]        [,2]       [,3]        [,4]
#> [1,]  0.8956301 -0.53648759  0.4661457 -0.02512903
#> [2,]  0.2904781  0.39324356  0.5137662  1.06793884
#> [3,]  0.4422979 -0.08625451  1.3074447  1.37099490
#> [4,] -0.3827740  0.40521912 -0.4859773 -0.34205135
#> [5,] -1.9899841  2.71188512 -6.0813547 -6.00742337
#> 
#> , , 4, 3
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.24385294  0.17311349 -0.04344049 -0.3020435
#> [2,]  0.41553424  0.26887396  0.28253976 -0.3768949
#> [3,] -0.07213085  0.05174733  1.52755201 -0.3422856
#> [4,]  0.34077549 -0.08454078 -0.56705417  1.4038406
#> [5,] -1.82830990  0.09294452  4.42430169  3.9918849
#> 
#> , , 5, 3
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.6086528 -0.72158958  0.6190662  0.6271648
#> [2,] -0.4652991  0.06385666 -0.6273339 -0.5027997
#> [3,] -0.7896854 -1.39702195 -0.1556419 -1.1075258
#> [4,]  0.4849387  1.69727270  0.7753132  1.7311382
#> [5,]  1.5392995  2.40229047  4.2905983  3.3062505
#> 
#> , , 1, 4
#> 
#>            [,1]       [,2]        [,3]       [,4]
#> [1,]  0.5702813 -0.1101456  0.36090891  0.9951314
#> [2,] -0.0473401  0.8092802 -0.15420708 -0.3929148
#> [3,] -0.6003958 -0.6798273  1.24171880  0.4854610
#> [4,]  0.7473935  0.7559265 -0.37111707  0.2387184
#> [5,]  1.0983096 -1.6545642 -0.07385012 -4.9967910
#> 
#> , , 2, 4
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.97577074  0.2502990  1.96359584  2.3127963
#> [2,] -0.23956656  1.3308607  0.97862628  1.1013843
#> [3,] -0.44022074  0.3860332 -0.09884361 -0.2739859
#> [4,]  0.42909844 -0.4947964  0.09548557  0.2494173
#> [5,] -0.07965121  1.4707126 -1.38353577  0.0257847
#> 
#> , , 3, 4
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.8043769 -0.88789645  0.4208101 -0.5727733
#> [2,]  0.3277416  0.85241505  0.2059650  0.7887191
#> [3,]  0.5047130  0.48177783  1.1275791  1.5706127
#> [4,] -0.4016199 -0.04861005 -0.4197477 -0.3897773
#> [5,] -2.0790633  1.44800916  4.7394902  3.8640482
#> 
#> , , 4, 4
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.02425968  0.4709587 -0.25509981 -0.2072563
#> [2,]  0.02270590  0.8703287  0.05475735 -0.1270196
#> [3,]  0.18982475  0.3112654  1.21770855 -0.1482714
#> [4,]  0.15579573 -0.4705367 -0.11597237  1.2017149
#> [5,] -0.63919492  0.7596447 -0.72512638  1.4352749
#> 
#> , , 5, 4
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.9542836 -0.78243000  0.9108473  0.7673148
#> [2,] -0.6421009  0.77752985 -1.4967421 -0.8095786
#> [3,] -1.1481160 -0.09264902 -1.8488939 -1.6622011
#> [4,]  1.1221565  0.44535061  2.4138202  2.2343004
#> [5,]  2.2028332  1.47114701 -0.7983133  2.9278222
#> 
#> , , 1, 5
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.55984674 -0.1228080  0.1164529  0.21144660
#> [2,]  0.09261113  0.9099905 -0.2814329 -0.27607738
#> [3,] -0.30206888 -0.6084102  0.8692296  0.08397282
#> [4,]  0.41406122  0.5953895  0.1280007  0.88734177
#> [5,]  3.12529102  1.6244868 -0.7177117 -1.37367909
#> 
#> , , 2, 5
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.37905411  0.1338363 -0.1107463  0.39241624
#> [2,]  0.09092618  0.9504079  0.6304196 -0.04294242
#> [3,]  0.19582319  0.1201086  1.4948440 -0.37418807
#> [4,] -0.01757231 -0.1857953 -0.4883856  1.22955912
#> [5,]  2.77066107  3.7620767 -1.1858597 -2.81317417
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.8299125 -0.7934015  0.5704611 -0.28088499
#> [2,]  0.3338071  0.8203587  0.4610057  0.88600624
#> [3,]  0.4839089  0.7270884  0.8824610  1.42612126
#> [4,] -0.4438677 -0.2699488 -0.1963796 -0.33122119
#> [5,]  0.9479653  0.8905340 -3.0505197 -0.04484968
#> 
#> , , 4, 5
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.11084981  0.04747172 -0.02927877 -0.1877554
#> [2,]  0.40285806  0.17554276  0.27322662  0.1917573
#> [3,]  0.38921501 -0.20181417  0.72042699  0.6577266
#> [4,] -0.05587714  0.22208945  0.25221279  0.3929370
#> [5,] -2.73237142 -1.56306787  2.20388024  1.7510544
#> 
#> , , 5, 5
#> 
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.2467533  0.36229623  0.5392653  0.5675912
#> [2,] -0.4812311  0.86291938 -1.8121978 -1.5456332
#> [3,] -0.8278803  0.03838317 -1.8373968 -2.3826038
#> [4,]  0.6714581 -0.15070738  2.5384765  3.1067216
#> [5,]  1.9128363 -0.98537824  3.6130802  1.9394199
#> 
#> 

## ------------------------------------------------
## Method `specify_posterior_bvars$get_last_draw`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|