R6 Class Representing PosteriorBVARs

The class PosteriorBVARs contains posterior output and the specification including the last MCMC draw for the Bayesian Panel VAR model. Note that due to the thinning of the MCMC output the starting value in element last_draw might not be equal to the last draw provided in element posterior.

See also

Public fields

last_draw

an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

posterior

a list containing Bayesian estimation output.

Methods

Public methods

PosteriorBVARs$new()

PosteriorBVARs$get_posterior()

PosteriorBVARs$get_last_draw()

PosteriorBVARs$clone()

PosteriorBVARs$new()

Create a new posterior output PosteriorBVARs.

Usage

PosteriorBVARs$new(specification_bvarPANEL, posterior_bvarPANEL)

Arguments

specification_bvarPANEL

an object of class BVARs with the last draw of the current MCMC run as the starting value.

posterior_bvarPANEL

a list containing Bayesian estimation output.

Returns

A posterior output PosteriorBVARs.

PosteriorBVARs$get_posterior()

Returns a list containing Bayesian estimation output.

Usage

PosteriorBVARs$get_posterior()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
posterior$get_posterior()

PosteriorBVARs$get_last_draw()

Returns an object of class BVARs with the last draw of the current MCMC run as the starting value to be passed to the continuation of the MCMC estimation using estimate().

Usage

PosteriorBVARs$get_last_draw()

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

PosteriorBVARs$clone()

The objects of this class are cloneable with this method.

Usage

PosteriorBVARs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
class(posterior)
#> [1] "PosteriorBVARs" "R6"            

## ------------------------------------------------
## Method `PosteriorBVARs$get_posterior()`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
posterior       = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior$get_posterior()
#> $A_c_cpp
#>      [,1]       
#> [1,] numeric,100
#> [2,] numeric,100
#> [3,] numeric,100
#> [4,] numeric,100
#> [5,] numeric,100
#> 
#> $Sigma_c_cpp
#>      [,1]      
#> [1,] numeric,80
#> [2,] numeric,80
#> [3,] numeric,80
#> [4,] numeric,80
#> [5,] numeric,80
#> 
#> $nu
#>          [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,] 5.100000 5.100000 5.100000 5.106178 5.106178
#> [2,] 5.100000 5.100000 5.209739 5.209739 5.209739
#> [3,] 5.100000 5.100000 5.100000 5.100000 5.100000
#> [4,] 5.100000 5.100000 5.100000 5.100000 5.100000
#> [5,] 5.085546 5.085546 5.085546 5.085546 5.068466
#> 
#> $m
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 1.0289963 0.8535821 0.6970159 0.3890592 0.8146473
#> [2,] 0.8217636 0.8330153 0.6899908 0.7520781 0.9693726
#> [3,] 0.9755457 1.1156229 0.5802595 0.8064154 0.8901187
#> [4,] 0.2999274 0.5331938 0.8034993 0.5042762 0.9045823
#> [5,] 0.5226304 1.0350667 0.8140604 0.6697936 1.1975751
#> 
#> $w
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.27047640 0.10664274 0.11075755 0.14920910 0.11074465
#> [2,] 0.19042605 0.10616340 0.05626761 0.04993345 0.05462572
#> [3,] 0.15857694 0.06563184 0.07750850 0.05794045 0.04352454
#> [4,] 0.04077182 0.11902268 0.15499282 0.29022811 0.50017510
#> [5,] 0.12783445 0.15907936 0.13336775 0.13952250 0.11402572
#> 
#> $s
#>           [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,]  76.46972 56.93489 93.52144 50.48280 90.69067
#> [2,]  92.39225 72.22866 59.76586 64.97174 67.46900
#> [3,] 122.13724 37.40156 74.52211 39.73472 66.24010
#> [4,]  60.37953 69.51325 65.81671 48.38556 76.65872
#> [5,] 171.70783 36.34097 66.65888 49.74742 45.02938
#> 
#> $scale
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
#> [4,]    0
#> [5,]    0
#> 
#> $Y
#>      [,1]       
#> [1,] numeric,700
#> [2,] numeric,700
#> [3,] numeric,700
#> [4,] numeric,700
#> [5,] numeric,700
#> 
#> $Sigma_c
#> , , 1, 1
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  2.38063916  0.07975577 -0.7575409 -0.2925794
#> [2,]  0.07975577  2.12963902  0.2429499 -0.4573482
#> [3,] -0.75754094  0.24294990  3.9145953  0.4858633
#> [4,] -0.29257940 -0.45734817  0.4858633  2.5978330
#> 
#> , , 2, 1
#> 
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  2.767055208  0.006113813 -0.35801766  0.07198315
#> [2,]  0.006113813  1.353157270  0.06846622 -0.04375209
#> [3,] -0.358017662  0.068466221  3.15695155  0.83059338
#> [4,]  0.071983154 -0.043752088  0.83059338  2.63749874
#> 
#> , , 3, 1
#> 
#>           [,1]        [,2]       [,3]       [,4]
#> [1,]  4.947176  -3.0659855   1.318637 -2.3482371
#> [2,] -3.065986  22.8676340 -12.004410  0.8229033
#> [3,]  1.318637 -12.0044103  12.499140  2.5236414
#> [4,] -2.348237   0.8229033   2.523641  8.3075144
#> 
#> , , 4, 1
#> 
#>           [,1]      [,2]       [,3]      [,4]
#> [1,] 2.0508119 0.3526033  2.7497948  2.311220
#> [2,] 0.3526033 1.7423736  0.9103203  1.195142
#> [3,] 2.7497948 0.9103203 12.4825725 10.252710
#> [4,] 2.3112196 1.1951423 10.2527096 10.727403
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  4.5003004 -0.5562947 -0.7196324 -1.091342
#> [2,] -0.5562947  9.3275582 -2.4555408 -1.176918
#> [3,] -0.7196324 -2.4555408  8.5683839  2.095925
#> [4,] -1.0913416 -1.1769181  2.0959253  5.453734
#> 
#> , , 1, 2
#> 
#>            [,1]      [,2]       [,3]      [,4]
#> [1,]  2.6040674 0.6098616 -0.2612397 0.4195638
#> [2,]  0.6098616 2.1475453  0.2038837 0.5542554
#> [3,] -0.2612397 0.2038837  1.9724322 0.8830305
#> [4,]  0.4195638 0.5542554  0.8830305 2.0441962
#> 
#> , , 2, 2
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  2.42077912 -0.5220004  0.7552795 -0.06424595
#> [2,] -0.52200042  2.0392099 -0.8506513 -0.21995773
#> [3,]  0.75527953 -0.8506513  2.6242244  1.18698150
#> [4,] -0.06424595 -0.2199577  1.1869815  4.26578476
#> 
#> , , 3, 2
#> 
#>            [,1]       [,2]        [,3]      [,4]
#> [1,]  3.3298779  -3.834975   0.1698808 -2.033179
#> [2,] -3.8349750  20.059188 -10.0082037  4.558986
#> [3,]  0.1698808 -10.008204  10.2063508  2.082259
#> [4,] -2.0331794   4.558986   2.0822588  6.294620
#> 
#> , , 4, 2
#> 
#>            [,1]        [,2]        [,3]      [,4]
#> [1,]  1.3819553 -0.33854259  0.48883528 0.3735411
#> [2,] -0.3385426  2.51751253 -0.06162183 0.4297931
#> [3,]  0.4888353 -0.06162183  6.37485819 4.4927513
#> [4,]  0.3735411  0.42979314  4.49275128 6.3275415
#> 
#> , , 5, 2
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.9284192 -0.2844311  0.3280344  0.3705217
#> [2,] -0.2844311  3.4893081 -1.7226758 -0.1438894
#> [3,]  0.3280344 -1.7226758  3.5674124  2.0286979
#> [4,]  0.3705217 -0.1438894  2.0286979  4.0958907
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  2.1202113 -0.1126960  0.7858960  0.3888058
#> [2,] -0.1126960  1.9979459 -0.3974881 -0.3922411
#> [3,]  0.7858960 -0.3974881  3.8907565  0.5233476
#> [4,]  0.3888058 -0.3922411  0.5233476  2.3502113
#> 
#> , , 2, 3
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  1.68876535 -0.6117002 -0.3761843 -0.03179762
#> [2,] -0.61170018  2.6043277 -0.2646078  0.36609274
#> [3,] -0.37618432 -0.2646078  2.1616395  0.41102523
#> [4,] -0.03179762  0.3660927  0.4110252  2.27445039
#> 
#> , , 3, 3
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  1.58093496 -0.5374299 -0.02292843 -0.5497016
#> [2,] -0.53742991 10.1556569 -5.58636403 -0.1331992
#> [3,] -0.02292843 -5.5863640  6.44396031  2.6764582
#> [4,] -0.54970160 -0.1331992  2.67645819  6.5396334
#> 
#> , , 4, 3
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  1.9933695 -0.5711376  1.0772343  1.099866
#> [2,] -0.5711376  2.5002482 -0.2907189 -1.034971
#> [3,]  1.0772343 -0.2907189  5.4629052  3.232124
#> [4,]  1.0998663 -1.0349705  3.2321245  5.646579
#> 
#> , , 5, 3
#> 
#>              [,1]       [,2]         [,3]       [,4]
#> [1,]  2.795839492  0.4388906  0.009050561 -0.8849924
#> [2,]  0.438890626  5.2722956 -2.551265342 -1.1445592
#> [3,]  0.009050561 -2.5512653  4.089824560  1.4773524
#> [4,] -0.884992357 -1.1445592  1.477352382  3.3768272
#> 
#> , , 1, 4
#> 
#>              [,1]         [,2]        [,3]       [,4]
#> [1,]  1.068140674  0.002719958 -0.04659786 0.11377952
#> [2,]  0.002719958  2.065799993 -0.29302871 0.08074243
#> [3,] -0.046597859 -0.293028712  1.84731277 0.72081421
#> [4,]  0.113779518  0.080742426  0.72081421 2.95414922
#> 
#> , , 2, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.8118943 -0.1699378  0.1184069 -0.3818873
#> [2,] -0.1699378  2.8109138 -0.7697505  0.9236656
#> [3,]  0.1184069 -0.7697505  3.0842434  0.1090868
#> [4,] -0.3818873  0.9236656  0.1090868  3.0873390
#> 
#> , , 3, 4
#> 
#>           [,1]       [,2]       [,3]      [,4]
#> [1,]  2.781636   1.449414  -1.708957 -0.026046
#> [2,]  1.449414  14.989135 -11.628939 -1.044324
#> [3,] -1.708957 -11.628939  13.012485  3.161335
#> [4,] -0.026046  -1.044324   3.161335  2.951803
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,] 1.59790651 0.08508404 0.6283869 0.7742658
#> [2,] 0.08508404 1.18974862 0.2654839 0.5890063
#> [3,] 0.62838693 0.26548389 3.3269535 1.2346060
#> [4,] 0.77426585 0.58900625 1.2346060 2.5887039
#> 
#> , , 5, 4
#> 
#>            [,1]       [,2]       [,3]     [,4]
#> [1,]  1.6659938 -0.3630282 -0.1788386 0.143073
#> [2,] -0.3630282  4.4470981 -1.0519553 0.263626
#> [3,] -0.1788386 -1.0519553  5.8228441 2.763137
#> [4,]  0.1430730  0.2636260  2.7631371 4.586752
#> 
#> , , 1, 5
#> 
#>            [,1]        [,2]        [,3]        [,4]
#> [1,]  2.0707945  0.45103806 -0.53376936 -0.20694249
#> [2,]  0.4510381  2.99770161 -0.90796266  0.02620834
#> [3,] -0.5337694 -0.90796266  3.08487658 -0.02206695
#> [4,] -0.2069425  0.02620834 -0.02206695  2.90252593
#> 
#> , , 2, 5
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  1.58778919 -0.08333205  0.12323727 -0.25647041
#> [2,] -0.08333205  2.14390317 -0.56289329  0.50771992
#> [3,]  0.12323727 -0.56289329  2.04670549  0.01161477
#> [4,] -0.25647041  0.50771992  0.01161477  2.41416554
#> 
#> , , 3, 5
#> 
#>             [,1]        [,2]       [,3]       [,4]
#> [1,]  1.94429735 -0.00882417 -0.3000349 -0.1720077
#> [2,] -0.00882417 11.94623754 -8.2602046 -0.7631309
#> [3,] -0.30003491 -8.26020464 10.2537066  3.1240098
#> [4,] -0.17200767 -0.76313087  3.1240098  5.3281434
#> 
#> , , 4, 5
#> 
#>              [,1]        [,2]       [,3]        [,4]
#> [1,]  1.629049724 -0.09497637 0.09722493 0.001401665
#> [2,] -0.094976375  2.48772725 0.32275740 0.054859855
#> [3,]  0.097224925  0.32275740 4.05329871 1.632209950
#> [4,]  0.001401665  0.05485986 1.63220995 3.975444865
#> 
#> , , 5, 5
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  1.41301227 -0.9074634  0.7064922  0.03973047
#> [2,] -0.90746336  6.5510124 -2.7522814 -1.17684668
#> [3,]  0.70649219 -2.7522814  4.5797663  1.57868573
#> [4,]  0.03973047 -1.1768467  1.5786857  2.63763922
#> 
#> 
#> $A_c
#> , , 1, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.1904722  0.1591577  0.8345344  0.2847645
#> [2,] -0.1129768  1.0413150 -0.9643392 -0.4539323
#> [3,]  0.1772130  0.3897294 -0.3257642 -0.3126453
#> [4,] -0.1827119 -0.3676608  0.9509539  1.2385780
#> [5,] -2.1561013 -3.0319235  1.2740636 -0.6869625
#> 
#> , , 2, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.8492155  0.2485910  0.3805800  0.1156576
#> [2,]  0.3699895  0.8689060  0.3285513  0.5088825
#> [3,]  0.2712818  0.4177077  0.2516721  0.3160693
#> [4,] -0.2833968 -0.4216579  0.4403845  0.5908548
#> [5,]  2.3537627  1.1975806 -0.2640185 -0.3867979
#> 
#> , , 3, 1
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.4259856  0.23473540  0.86513324  1.1023376
#> [2,]  0.5902545  0.08643897 -0.09719322 -0.8806646
#> [3,]  1.0358117 -0.17415806  0.45705678 -1.3264730
#> [4,] -0.7642743  0.21661740  0.16529408  1.9027562
#> [5,] -2.3533047  5.29498104 -1.04004117  1.3464683
#> 
#> , , 4, 1
#> 
#>            [,1]       [,2]      [,3]       [,4]
#> [1,]  1.2849396  0.1601185  3.062837  2.7646129
#> [2,] -0.2014341  0.5718334 -1.137835 -0.9409509
#> [3,] -0.6098928 -0.2224868 -1.684609 -1.8949740
#> [4,]  0.4829370  0.1625927  1.547943  1.8711034
#> [5,]  1.5864012  1.0943517  5.022359  5.0695637
#> 
#> , , 5, 1
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.0293564 -0.5961135  0.8156619  0.6822632
#> [2,]  0.1221348  0.7468151  0.2188828  0.5675937
#> [3,]  0.3279178 -0.2273412  1.0523143  0.9373368
#> [4,] -0.2776871  0.5157691 -0.4402140 -0.2268625
#> [5,] -3.0094618  0.3919018 -0.6154629 -1.1147663
#> 
#> , , 1, 2
#> 
#>             [,1]        [,2]       [,3]         [,4]
#> [1,]  1.03962861 -0.13354568  0.6379359  0.184352245
#> [2,] -0.05865501  1.22207423 -0.4184792 -0.005366688
#> [3,]  0.51879491  0.35985100  1.0315742  0.865457054
#> [4,] -0.47809014 -0.29372737 -0.3020742  0.110035607
#> [5,] -0.23869443  0.06916549  1.4188056  0.257778274
#> 
#> , , 2, 2
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  1.26444259 -0.04067502 -0.1122525 -0.62350204
#> [2,]  0.08489094  0.46766335  0.2960173 -0.08158275
#> [3,]  0.04153940 -0.10272571  1.2758070  0.45124606
#> [4,] -0.14698930  0.20650860 -0.2779951  0.80990510
#> [5,]  0.73498495  0.24521588  1.7774734  2.77751640
#> 
#> , , 3, 2
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,]  0.4047419  0.27741717  0.91624454  1.0752044
#> [2,]  0.4082616 -0.08554255  0.04332071 -0.7599594
#> [3,]  0.7268281 -0.83191157  0.47436305 -0.9755448
#> [4,] -0.4776899  0.78627116  0.12350463  1.5847179
#> [5,] -0.5700397  5.04938558 -2.15967589  1.5747441
#> 
#> , , 4, 2
#> 
#>            [,1]        [,2]      [,3]       [,4]
#> [1,]  0.9803531  0.07182301  1.406514  1.3830426
#> [2,] -0.2711069  0.78451535 -2.081052 -1.4436250
#> [3,] -0.6183273 -0.34805289 -1.801516 -2.2111003
#> [4,]  0.5818901  0.31304358  2.294028  2.7442523
#> [5,]  2.2155502  1.05347380  2.256183 -0.5928071
#> 
#> , , 5, 2
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.9912564 -0.0351604 -0.1835146 -0.25636961
#> [2,]  0.1330905  0.6677770  0.3044381  0.71552340
#> [3,]  0.2041601 -0.3639229  1.4675599  1.16400199
#> [4,] -0.2377413  0.4605281 -0.4207632 -0.02245861
#> [5,]  1.8751730 -3.4405096  1.7995061 -2.03489401
#> 
#> , , 1, 3
#> 
#>            [,1]       [,2]        [,3]        [,4]
#> [1,]  0.9439659  0.4650860 -0.36232585 -0.18554787
#> [2,] -0.1230928  0.6056309  0.19487806  0.07513690
#> [3,] -0.1216556 -0.6156093  1.08841597  1.03327541
#> [4,]  0.1472111  0.4247783  0.03406354  0.07225753
#> [5,]  0.5942666 -0.6657438  1.11053359  2.76190734
#> 
#> , , 2, 3
#> 
#>             [,1]        [,2]       [,3]         [,4]
#> [1,] 0.596450607 -0.09615341  0.2276953  0.191458340
#> [2,] 0.188676208  0.32828104  0.3822429 -0.044237755
#> [3,] 0.008391821  0.12489149  0.6648055  0.128319324
#> [4,] 0.077165884  0.06533606  0.1338210  0.838197961
#> [5,] 0.680025763  0.67971711 -1.0585903  0.002954144
#> 
#> , , 3, 3
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.99587097 -0.7237703  1.1018332  0.19275844
#> [2,]  0.08162849  0.8011103 -0.2667544 -0.04203466
#> [3,]  0.08906443  0.6052641  0.2453935  0.25201472
#> [4,] -0.11202640 -0.1873809  0.3037539  0.70937395
#> [5,]  0.98037862  0.4808391 -1.4868695  0.34945376
#> 
#> , , 4, 3
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  0.5196080  0.1398914  0.9542838  1.504616
#> [2,] -0.2915476  0.6915568 -1.6607134 -2.122659
#> [3,] -0.7787412 -0.2726748 -1.9443425 -1.989939
#> [4,]  0.9308137  0.2259336  2.6071964  2.528791
#> [5,]  0.1585308  0.6388086 -0.9058733 -3.045869
#> 
#> , , 5, 3
#> 
#>             [,1]        [,2]       [,3]        [,4]
#> [1,]  0.62612595  0.03829009  0.1518379  0.70118631
#> [2,]  0.11649849  0.36823521  0.5741648  0.43953633
#> [3,]  0.25476940 -0.57490224  1.7002691  0.68855956
#> [4,] -0.08420946  0.64795878 -0.7693021 -0.04789843
#> [5,] -0.18918404 -2.67369024 -1.2625415  2.76470844
#> 
#> , , 1, 4
#> 
#>               [,1]       [,2]        [,3]        [,4]
#> [1,]  0.9104052997  0.5874340 -0.04454699 -0.07919669
#> [2,] -0.0249894401  0.4375063  0.04282932  0.22424009
#> [3,]  0.0688695356 -0.4675598  1.42924606  1.31792198
#> [4,]  0.0008630811  0.2593556 -0.34588632 -0.22179679
#> [5,] -0.7799850194 -0.8549209 -1.51858402  0.39709414
#> 
#> , , 2, 4
#> 
#>             [,1]       [,2]        [,3]       [,4]
#> [1,]  0.95727422  0.1373162  0.21073979 -0.3604906
#> [2,]  0.29834763  0.8890840 -0.09502563 -0.3376175
#> [3,] -0.01940086  0.3124704  0.22521006 -0.0807149
#> [4,] -0.03380927 -0.2433859  0.56425650  1.2462795
#> [5,] -0.10347919 -2.3827903  2.38954044  0.3923427
#> 
#> , , 3, 4
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.8365849 -1.0006760  0.9632966  0.19784685
#> [2,]  0.2307466  0.9410368 -0.2766269 -0.10685783
#> [3,]  0.2697201  0.6775433  0.2418871 -0.01584053
#> [4,] -0.2481525 -0.2016047  0.3956382  0.95469146
#> [5,]  1.2144044  2.4090722 -3.3721199  0.54028781
#> 
#> , , 4, 4
#> 
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  1.2734838 -0.2489002  0.9492715  2.1561890
#> [2,]  0.1778982  0.9177652 -1.5118357 -0.6794479
#> [3,] -0.3027556 -0.1623053 -1.7463680 -1.4521868
#> [4,]  0.2059356  0.2270824  2.3626097  1.6534141
#> [5,] -0.7183706  1.4584305  3.1577183  3.5455884
#> 
#> , , 5, 4
#> 
#>             [,1]       [,2]       [,3]       [,4]
#> [1,]  0.77356569 -0.6729942  0.5788895  0.4334015
#> [2,]  0.01410300  0.4525924  0.4968346  0.6004597
#> [3,]  0.19827567 -0.4791546  1.4201289  0.8977872
#> [4,] -0.08256448  0.8613858 -0.6653227 -0.0991007
#> [5,] -0.16915865 -2.3965715 -3.3430035 -0.2243277
#> 
#> , , 1, 5
#> 
#>            [,1]        [,2]        [,3]       [,4]
#> [1,] 0.13407144 -0.84389186  0.28242015  0.6519701
#> [2,] 0.54628382  1.37131388 -0.15529804 -0.1029883
#> [3,] 0.24106668 -0.06082378  1.18712431  0.9203128
#> [4,] 0.09840496  0.38882996 -0.29143525 -0.1904468
#> [5,] 0.75092066  1.48208673 -0.05948568  1.6634488
#> 
#> , , 2, 5
#> 
#>             [,1]        [,2]        [,3]       [,4]
#> [1,]  0.62915785  0.03190645 -0.02113993 -0.1766041
#> [2,] -0.27812674  0.43064723  0.37661050  0.3444069
#> [3,]  0.16874296  0.32162184  1.02004982  0.2476794
#> [4,]  0.02372881 -0.15159443 -0.12743483  0.7663909
#> [5,]  1.18204996 -0.56590327  2.63523158  0.3753879
#> 
#> , , 3, 5
#> 
#>            [,1]       [,2]       [,3]        [,4]
#> [1,]  0.8898730 -1.0293693  1.1067419  0.15545430
#> [2,]  0.2823068  1.0419195 -0.3223885  0.05029568
#> [3,]  0.3164992  1.0951398 -0.0862053 -0.02138164
#> [4,] -0.2718812 -0.5536804  0.5445603  0.95336351
#> [5,] -1.2794029  0.8293864  1.2811732 -1.00085225
#> 
#> , , 4, 5
#> 
#>            [,1]       [,2]      [,3]      [,4]
#> [1,]  0.5458149 -0.5749867  1.574740  1.886278
#> [2,] -0.2258648  0.2079252 -1.897175 -1.196860
#> [3,] -0.5495698 -0.1423191 -1.723531 -1.584944
#> [4,]  0.6936754  0.3364794  2.233101  1.965282
#> [5,]  0.9015136  1.5372619 -4.559818 -2.198686
#> 
#> , , 5, 5
#> 
#>            [,1]       [,2]       [,3]      [,4]
#> [1,]  1.3791222 -0.7812241  0.8408929 0.4490999
#> [2,] -0.3855922  0.3709040  0.2846137 0.2511693
#> [3,] -0.6033430 -0.5986556  1.1120321 0.3779943
#> [4,]  0.4889129  0.9533179 -0.4944633 0.4003353
#> [5,] -2.8172770  2.2614867 -2.1297930 0.7746728
#> 
#> 

## ------------------------------------------------
## Method `PosteriorBVARs$get_last_draw()`
## ------------------------------------------------

specification = specify_bvars$new(
   data = ilo_dynamic_panel[1:5]
)
burn_in        = estimate(specification, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior      = estimate(burn_in, 5)
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|