Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARGROUPPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARGROUPPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarGroupPANEL$new(
      data = ilo_dynamic_panel[1:5],
      exogenous = ilo_exogenous_variables[1:5],
      group_allocation = country_grouping_region[1:5]
)
#> Country groupings have been pre-specified and will not be estimated.
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003601243 0.0006789333  0.00305624  0.004524408
#> 
#> $AFG$Sigma$equation2
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.005002588 0.003303576 -0.008801102 -0.001431545
#> Sigma[2,2]  0.187973518 0.052556805  0.120232057  0.234661475
#> 
#> $AFG$Sigma$equation3
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[3,1]  0.003592893 0.004287771 -0.0005121263  0.009118131
#> Sigma[3,2] -0.172537964 0.075554232 -0.2615684996 -0.103831188
#> Sigma[3,3]  0.368895452 0.130540918  0.2266034268  0.523386620
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.001703346 0.003963739 -0.002512758  0.006641887
#> Sigma[4,2] -0.105831679 0.068544304 -0.187827593 -0.052236501
#> Sigma[4,3]  0.346027472 0.128147220  0.197715614  0.492085493
#> Sigma[4,4]  0.354807107 0.129257389  0.205695981  0.493293232
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.97203937 0.01234740  0.96015449   0.98810597
#> lag1_var2  0.03482517 0.06770406 -0.05450674   0.09533010
#> lag1_var3  0.07547608 0.17624374 -0.15337689   0.24395971
#> lag1_var4 -0.06247441 0.15627795 -0.21420938   0.14014148
#> const      0.10206052 0.97515607 -0.87418581   1.31493990
#> exo1      -0.02274423 0.04907488 -0.05878430   0.04459528
#> exo2      -0.03641287 0.06095653 -0.10416342   0.01581691
#> exo3      -0.14948490 0.08538238 -0.22631718  -0.04073313
#> 
#> $AFG$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.44179495 0.1517562  0.25982547   0.61009783
#> lag1_var2 -0.46263791 0.4405804 -0.95602473  -0.01390307
#> lag1_var3 -2.97613530 1.0700314 -4.16588405  -1.72754351
#> lag1_var4  2.56992668 0.9749046  1.40665579   3.63788194
#> const      9.63406222 7.1878813  2.49098403  18.33389104
#> exo1      -0.03681705 0.3588077 -0.44777867   0.29649987
#> exo2       0.08261190 0.1790882 -0.09729186   0.30096459
#> exo3      -0.27268895 0.5636955 -0.83359670   0.44691571
#> 
#> $AFG$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.29951325 0.2802984  -0.6201397  -0.02799084
#> lag1_var2  1.29782228 0.5242427   0.6786413   1.83450254
#> lag1_var3  4.27512589 0.9823506   3.3930607   5.40624305
#> lag1_var4 -2.94918071 0.8198905  -3.9208446  -2.28458810
#> const     -6.62765808 6.4731926 -14.5514561   0.48630786
#> exo1       0.04718243 0.4268756  -0.2640801   0.62816662
#> exo2      -1.12817732 0.3394693  -1.5509045  -0.77361231
#> exo3      -0.08912271 0.4494011  -0.6067937   0.26528025
#> 
#> $AFG$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.1277281 0.2865181  -0.4571630    0.1889800
#> lag1_var2  0.9297150 0.4566364   0.3886511    1.3853432
#> lag1_var3  2.6015903 0.7645474   1.8113852    3.4187007
#> lag1_var4 -1.3739708 0.5996303  -2.0175595   -0.7300530
#> const     -5.4577871 4.1969535 -10.3266991   -0.5707551
#> exo1       0.0318806 0.4415169  -0.4502681    0.5732936
#> exo2      -1.2308430 0.4499885  -1.7906062   -0.8171287
#> exo3      -0.2073242 0.4442222  -0.7903161    0.2048778
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005931175 0.002431444 0.004615601   0.00925025
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01189683 0.009928046 -0.02550046 -0.004559434
#> Sigma[2,2]  0.13429024 0.038504303  0.09437012  0.177110822
#> 
#> $AGO$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01045929 0.005766488  0.007212016   0.01829362
#> Sigma[3,2] -0.05281997 0.020024584 -0.080090441  -0.03688348
#> Sigma[3,3]  0.03974058 0.013145260  0.028422094   0.05723150
#> 
#> $AGO$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.001094250 0.002579948 -0.001686783 4.021061e-03
#> Sigma[4,2]  0.052948294 0.018701601  0.034523007 7.660787e-02
#> Sigma[4,3] -0.003454488 0.003182723 -0.006655930 7.633241e-05
#> Sigma[4,4]  0.039163073 0.014470133  0.025772684 5.812982e-02
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  1.02659184 0.02870562  1.003706236   1.06040671
#> lag1_var2 -0.25869576 0.02189687 -0.287552163  -0.23898503
#> lag1_var3 -0.31106178 0.04562528 -0.358775843  -0.26189831
#> lag1_var4  0.29816874 0.03486046  0.258596810   0.33423330
#> const      0.69056406 0.29227564  0.363430646   1.02984461
#> exo1       0.06611192 0.06257506  0.002197453   0.14568101
#> exo2      -0.08786081 0.08444011 -0.195322415  -0.01195606
#> exo3       0.04312086 0.11039615 -0.073085894   0.18196501
#> 
#> $AGO$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.10144496 0.13701138 -0.04009602    0.2584677
#> lag1_var2 -0.02843831 0.15719395 -0.21756041    0.1056337
#> lag1_var3 -1.20362459 0.23762110 -1.47342819   -0.9238789
#> lag1_var4  1.18159370 0.16141895  1.03723669    1.3886067
#> const      0.76557861 5.82763047 -5.45990696    7.9564011
#> exo1       0.01999441 0.36047820 -0.33906275    0.4108827
#> exo2       0.58476694 0.08560864  0.47683530    0.6700796
#> exo3       0.08488117 0.65874930 -0.34283079    0.9848038
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.09063665 0.07693639  -0.1838369  -0.01506632
#> lag1_var2 -0.32832657 0.05728436  -0.3805238  -0.25820085
#> lag1_var3  0.56647479 0.09343992   0.4683716   0.68111827
#> lag1_var4  0.41619092 0.06177075   0.3352722   0.47458687
#> const      3.42998385 2.18197965   1.4958502   6.34010113
#> exo1       0.06502029 0.21008414  -0.1482880   0.32653099
#> exo2      -1.48034581 0.22222514  -1.7385548  -1.22999000
#> exo3       3.38431088 0.39995843   2.8401116   3.60101405
#> 
#> $AGO$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.1397257 0.04593066  -0.1985531  -0.09682027
#> lag1_var2 -0.3484757 0.07357181  -0.4346823  -0.27804791
#> lag1_var3 -0.3656938 0.12865174  -0.4772421  -0.21701086
#> lag1_var4  1.4197281 0.10647191   1.2960275   1.52307285
#> const      0.2308509 3.01739086  -2.7245718   4.01953601
#> exo1       0.1516009 0.25375205  -0.1550695   0.41466917
#> exo2      -1.3724206 0.17461261  -1.5520176  -1.16257404
#> exo3       3.7609402 0.14182478   3.6506552   3.92620714
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002069679 0.0006707216 0.001583412  0.002990209
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.08327785 0.04844976   -0.147504  -0.03880311
#> Sigma[2,2]  9.79621968 3.53181450    7.403207  14.65825776
#> 
#> $ALB$Sigma$equation3
#>                  mean       sd   5% quantile 95% quantile
#> Sigma[3,1]  0.0482681 0.038384   0.008801442   0.09707225
#> Sigma[3,2] -6.6811698 2.459293 -10.003298323  -4.86164656
#> Sigma[3,3]  6.3654680 1.770727   4.612215498   8.62230803
#> 
#> $ALB$Sigma$equation4
#>                    mean          sd 5% quantile  95% quantile
#> Sigma[4,1] -0.008700098 0.008777276 -0.02002789 -0.0009152203
#> Sigma[4,2] -0.189475224 0.823144457 -1.27083292  0.4971127586
#> Sigma[4,3]  2.306156340 0.815333214  1.64432018  3.4032771616
#> Sigma[4,4]  2.638857824 0.543278375  2.05323797  3.2277924273
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  1.007133933 0.02584280  0.97335094   1.02981289
#> lag1_var2 -0.006355029 0.02821347 -0.02544602   0.03254099
#> lag1_var3 -0.021738721 0.04467913 -0.05546371   0.03939266
#> lag1_var4  0.017189171 0.03713485 -0.03346467   0.04587835
#> const      0.049787457 0.41686149 -0.44697719   0.50076817
#> exo1      -0.016466011 0.03607808 -0.05619805   0.02310902
#> exo2      -0.061817336 0.05305172 -0.13066010  -0.01170126
#> exo3       0.078379897 0.04483535  0.03694699   0.13743508
#> 
#> $ALB$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -4.0877110  2.301343   -6.445020    -1.156096
#> lag1_var2  4.2767799  1.555835    2.637023     6.204756
#> lag1_var3  6.1898550  2.581713    3.429998     9.334120
#> lag1_var4 -5.0836598  2.238068   -7.819222    -2.717234
#> const     34.0733808 33.715377  -10.313137    63.531737
#> exo1       0.5509044  3.499734   -3.069167     4.932985
#> exo2       0.2740324  3.724005   -3.875078     4.451658
#> exo3      -0.5622875  4.820489   -6.943324     3.699488
#> 
#> $ALB$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1   3.2397140  1.725082    1.317868     5.208410
#> lag1_var2  -2.5470088  1.063793   -3.483796    -1.157014
#> lag1_var3  -3.6158055  1.885358   -5.098321    -1.062840
#> lag1_var4   3.7360130  1.618062    1.561638     5.101832
#> const     -25.0901431 36.192603  -62.510266    20.084026
#> exo1       -2.1430947  2.427691   -5.331774    -0.286453
#> exo2       -0.5943185  2.876489   -4.071862     2.213691
#> exo3       -0.3290401  3.402363   -2.553101     4.352958
#> 
#> $ALB$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.59554625  0.8131812  -0.4135955    1.3048804
#> lag1_var2 -0.12384237  1.1474604  -1.4420709    1.1448890
#> lag1_var3 -0.09723509  1.9654019  -2.2909024    2.1275057
#> lag1_var4  1.00521949  1.5611963  -0.7741346    2.7277606
#> const     -7.32901214 22.4415128 -35.2262059   15.9224702
#> exo1      -1.70201006  0.8819011  -2.5782224   -0.5782320
#> exo2      -0.39315908  1.0014024  -1.3241605    0.9139107
#> exo3      -0.71384990  1.5557889  -2.6698601    0.6961566
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001512964 0.0002048184 0.001297404  0.001738597
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.002095025 0.003066121 -0.006218729 0.0001656551
#> Sigma[2,2]  0.072991426 0.017083079  0.058523987 0.0922269875
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.005206056 0.001240193  0.003521759  0.005850474
#> Sigma[3,2] -0.057110239 0.023509637 -0.083647951 -0.029714869
#> Sigma[3,3]  0.206633137 0.029051559  0.185266332  0.245788269
#> 
#> $ARE$Sigma$equation4
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.0038818791 0.002198869  0.001184421  0.006040581
#> Sigma[4,2] -0.0005689037 0.016241285 -0.015432037  0.020359894
#> Sigma[4,3]  0.1651701737 0.018712170  0.148451750  0.188724760
#> Sigma[4,4]  0.1683527275 0.020165871  0.151202262  0.194047984
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.985390634 0.03216159  0.94315952   1.01115255
#> lag1_var2  0.018911402 0.08172240 -0.04151110   0.12953884
#> lag1_var3  0.006085448 0.10723357 -0.06098200   0.15357088
#> lag1_var4 -0.010034236 0.10871736 -0.15922416   0.06603895
#> const      0.692834807 0.44059398  0.26146045   1.26519064
#> exo1      -0.007603551 0.04839422 -0.05491802   0.05383896
#> exo2      -0.076662387 0.02583571 -0.10719252  -0.05366030
#> exo3       0.002398012 0.06094159 -0.04888871   0.07331513
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.03943226 0.4914087  -0.3214462    0.7065679
#> lag1_var2  0.56967333 0.6293759  -0.2821637    0.9980136
#> lag1_var3 -0.32095385 0.7540378  -1.3409148    0.1282559
#> lag1_var4  0.29913230 0.6183529  -0.1377468    1.1341498
#> const      1.09478862 1.8424900  -0.9649993    3.2115040
#> exo1       0.02775548 0.1552628  -0.1330580    0.2050218
#> exo2       1.91145504 0.4949717   1.3236773    2.3851318
#> exo3      -0.74238000 0.6096708  -1.5619673   -0.2613211
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.6623249 0.2559527   0.4212641    0.9980846
#> lag1_var2 -1.5101356 1.2885189  -2.4449473    0.2566229
#> lag1_var3 -1.1424588 1.8884937  -2.4032925    1.4151492
#> lag1_var4  1.9077307 1.8498737  -0.5551726    3.2046466
#> const      0.4286536 3.9826303  -4.0633240    5.1729011
#> exo1      -0.4572920 0.3937438  -0.9816271   -0.1819822
#> exo2      -2.3159334 0.6477988  -2.8566012   -1.4922178
#> exo3      -0.8757656 0.8845593  -1.8212539    0.2252592
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.7467329 0.2536927   0.5261413   1.08717196
#> lag1_var2 -1.1461109 0.8664332  -1.8617600  -0.06473598
#> lag1_var3 -1.4574605 1.3301432  -2.4683397   0.23284969
#> lag1_var4  2.2196619 1.3909842   0.4782169   3.31419587
#> const     -1.1493172 5.4706838  -7.3799082   5.24647127
#> exo1      -0.4067697 0.2895714  -0.7935812  -0.16350907
#> exo2      -0.8525059 0.3793766  -1.2090734  -0.35416063
#> exo3      -1.4540439 0.7091294  -2.4059001  -0.91302579
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003360365 0.001892173 0.001986198  0.005944746
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05823646 0.02291808 -0.08588523  -0.03391771
#> Sigma[2,2]  3.54202621 0.75329179  2.77929329   4.49446466
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04305244 0.01309221   0.0287252    0.0569249
#> Sigma[3,2] -2.48276503 0.46451554  -3.0669871   -2.0138423
#> Sigma[3,3]  1.96664680 0.28303816   1.6834735    2.2909405
#> 
#> $ARG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.009597809 0.004275854  0.004929089   0.01460474
#> Sigma[4,2] -0.364837703 0.120455123 -0.519577111  -0.24592858
#> Sigma[4,3]  0.506607932 0.082243680  0.421590931   0.60464613
#> Sigma[4,4]  0.317483848 0.030913468  0.276036148   0.34139748
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.86623516 0.05463166  0.79967650   0.92099855
#> lag1_var2  0.30459316 0.13873032  0.17036338   0.46489985
#> lag1_var3  0.49900983 0.22662439  0.27943326   0.75923061
#> lag1_var4 -0.45802077 0.20209239 -0.69584850  -0.26411945
#> const      1.12256142 0.25000315  0.90232225   1.44206090
#> exo1       0.04848891 0.10995109 -0.03793853   0.19582468
#> exo2      -0.13721389 0.06269224 -0.19642921  -0.06502127
#> exo3      -0.01412178 0.06176906 -0.08730113   0.04556040
#> 
#> $ARG$A$equation2
#>                mean        sd 5% quantile 95% quantile
#> lag1_var1 -1.983126  1.972470  -3.8703856    0.2210273
#> lag1_var2  2.639492  2.007247   0.4724237    5.0774934
#> lag1_var3  3.318020  3.475098  -0.5203451    7.4196656
#> lag1_var4 -2.309975  2.872141  -5.9101690    0.7055189
#> const     -3.849796 18.099704 -18.2070579   20.4940794
#> exo1      -2.199854  1.774163  -4.6148621   -0.8742603
#> exo2       1.977658  2.505934  -0.4166479    4.9774191
#> exo3       1.734306  3.676108  -1.0075295    6.6768383
#> 
#> $ARG$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  1.0951116 1.1763738  -0.2889787   2.02612304
#> lag1_var2 -1.0256318 0.7455707  -1.7150006  -0.06670721
#> lag1_var3 -1.0991201 1.3957826  -2.4419256   0.67129611
#> lag1_var4  1.3613110 0.9296750   0.1353279   2.23139158
#> const     12.7750786 9.7859922   1.3907053  23.44720524
#> exo1       0.5786336 1.3000378  -0.4935435   2.31073085
#> exo2      -4.7637770 1.9893070  -7.2051332  -3.00577058
#> exo3       1.9039403 2.4262424  -1.2456285   4.05943695
#> 
#> $ARG$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.1958699 0.3068902 -0.55127926    0.1599915
#> lag1_var2  0.4750788 0.8108766 -0.08668164    1.5650011
#> lag1_var3  0.7412849 1.3371058 -0.19275742    2.5416256
#> lag1_var4  0.2006080 1.2914052 -1.54868107    1.1037932
#> const      8.6515684 6.3762032  0.73026788   14.7429532
#> exo1      -0.8155012 0.4782276 -1.31579098   -0.2309625
#> exo2      -4.0483413 0.5569276 -4.75740569   -3.6033482
#> exo3       3.2938416 0.5226850  2.60975901    3.7791411
#> 
#> 
#>