Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARGROUPPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARGROUPPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarGroupPANEL$new(
      data = ilo_dynamic_panel[1:5],
      exogenous = ilo_exogenous_variables[1:5],
      group_allocation = country_grouping_region[1:5]
)
#> Country groupings have been pre-specified and will not be estimated.
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.007776469 0.001089364 0.006755411  0.008943431
#> 
#> $AFG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01468537 0.006899682 -0.02128398   -0.0059557
#> Sigma[2,2]  0.16519941 0.029524617  0.13340214    0.1987063
#> 
#> $AFG$Sigma$equation3
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[3,1]  0.007020349 0.005851169 -0.0007695042   0.01215010
#> Sigma[3,2] -0.116119007 0.017488719 -0.1385023929  -0.09891276
#> Sigma[3,3]  0.197427097 0.048608373  0.1559280816   0.26223843
#> 
#> $AFG$Sigma$equation4
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[4,1] -0.0005642188 0.004947961 -0.006386873  0.003950564
#> Sigma[4,2] -0.0480912069 0.016432728 -0.061372034 -0.026136531
#> Sigma[4,3]  0.1699288718 0.048056239  0.130751930  0.233991420
#> Sigma[4,4]  0.1749157639 0.050597725  0.136255899  0.243816071
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.95531703 0.02534583   0.9294965   0.98758501
#> lag1_var2 -0.69663160 0.15162353  -0.9019695  -0.58500609
#> lag1_var3 -1.53990859 0.33634747  -1.9982379  -1.29545988
#> lag1_var4  1.35448508 0.28796884   1.1432983   1.74758105
#> const      9.56955150 2.46575158   7.7332642  12.91496128
#> exo1      -0.03065964 0.15042151  -0.1643762   0.14871972
#> exo2      -0.19208972 0.04753236  -0.2387499  -0.13970032
#> exo3      -0.26569201 0.16177846  -0.4670142  -0.09557701
#> 
#> $AFG$A$equation2
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.324786681  0.1526576   0.1473763    0.4797251
#> lag1_var2  0.343357756  0.9966027  -0.9314579    1.3538947
#> lag1_var3 -1.341319348  2.1622859  -4.0229658    0.8983009
#> lag1_var4  1.142886819  1.8598004  -0.8095425    3.4101937
#> const      2.212447779 17.0825368 -14.2676427   24.2810288
#> exo1      -0.168549909  0.2873181  -0.4632359    0.1227571
#> exo2       0.142090168  0.3509107  -0.2303884    0.5760973
#> exo3       0.008552909  0.3860924  -0.4693791    0.4159388
#> 
#> $AFG$A$equation3
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1   0.01063894  0.1439535  -0.1696071    0.1563394
#> lag1_var2   4.70215752  1.1764267   3.5480897    6.2146829
#> lag1_var3  11.93005482  2.4157099   9.5452041   15.0473405
#> lag1_var4  -9.63016266  2.0478640 -12.2552499   -7.5829220
#> const     -57.85861467 17.7823352 -80.7425157  -40.8540849
#> exo1        0.10628737  0.2463274  -0.1772424    0.3942968
#> exo2       -0.43890450  0.3021583  -0.8362850   -0.1626945
#> exo3        0.78606342  0.2125101   0.5225858    1.0086649
#> 
#> $AFG$A$equation4
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1   0.18654826  0.1081297   0.04587095    0.2888212
#> lag1_var2   5.60178809  1.0220115   4.63900532    6.8621647
#> lag1_var3  13.01233948  2.0870550  11.01536303   15.6392087
#> lag1_var4 -10.46292011  1.7640608 -12.68095436   -8.7491341
#> const     -72.90520618 14.2222639 -91.23983266  -60.0566909
#> exo1        0.02667792  0.3479722  -0.26784452    0.4953898
#> exo2       -0.30430994  0.2856261  -0.57286354    0.0658426
#> exo3        1.03495040  0.2181167   0.76227955    1.2596536
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.007053259 0.001529004 0.005105813  0.008551004
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.02244334 0.007359813 -0.03231012  -0.01629626
#> Sigma[2,2]  0.16592362 0.054005243  0.11346827   0.23618849
#> 
#> $AGO$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01507814 0.005826311  0.009615037   0.02272842
#> Sigma[3,2] -0.08254836 0.033504192 -0.127642492  -0.05603751
#> Sigma[3,3]  0.05417434 0.024134422  0.037045899   0.08717403
#> 
#> $AGO$Sigma$equation4
#>                    mean           sd  5% quantile 95% quantile
#> Sigma[4,1] -0.002836416 0.0008410364 -0.003764957 -0.001855336
#> Sigma[4,2]  0.053569206 0.0135028855  0.037294807  0.065326357
#> Sigma[4,3] -0.012200367 0.0036733169 -0.015543677 -0.007369916
#> Sigma[4,4]  0.033758825 0.0083402827  0.024374639  0.041959190
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.04256297 0.04775648  0.99247022   1.09973215
#> lag1_var2 -0.49244162 0.34353093 -0.89381404  -0.10451288
#> lag1_var3 -0.65836226 0.46980943 -1.18670800  -0.11897297
#> lag1_var4  0.59179517 0.42173257  0.11906385   1.08700645
#> const      3.88100224 3.84898310 -0.08031954   7.93412511
#> exo1       0.06582083 0.08043072 -0.04206747   0.13499025
#> exo2      -0.04316239 0.09749602 -0.10755366   0.09046865
#> exo3       0.12369637 0.18091418 -0.11265319   0.28284671
#> 
#> $AGO$A$equation2
#>                 mean         sd  5% quantile 95% quantile
#> lag1_var1 -0.1567756  0.1978857  -0.37851340   0.03623397
#> lag1_var2  1.6379249  1.7086177  -0.43681291   3.48735271
#> lag1_var3  0.9854653  2.3296963  -1.79244639   3.47956991
#> lag1_var4 -0.8226052  2.0499472  -3.04924828   1.60091576
#> const     -6.7391941 17.5314282 -21.27707999  16.99900165
#> exo1       0.3183363  0.6564965  -0.17601192   1.21369151
#> exo2       0.5077667  0.4416479   0.04161914   1.03806388
#> exo3      -1.0228091  0.5137798  -1.51967874  -0.35296757
#> 
#> $AGO$A$equation3
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.084095635  0.1310830  -0.2218563   0.06237138
#> lag1_var2 -0.926509850  1.2476352  -2.0916921   0.58233389
#> lag1_var3 -0.305961894  1.7124754  -2.0101312   1.70711782
#> lag1_var4  1.120757239  1.4967066  -0.6378841   2.58944274
#> const     14.985407482 14.6083391  -2.7944349  28.66949348
#> exo1       0.001346558  0.3717199  -0.5092862   0.21082147
#> exo2      -1.484109045  0.3550450  -1.8211262  -1.05628716
#> exo3       3.573613819  0.3805983   3.0766863   3.88770770
#> 
#> $AGO$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.2502785 0.04154764  -0.3048999   -0.2145120
#> lag1_var2 -0.2071318 0.56501974  -0.7054626    0.4142024
#> lag1_var3 -0.2305461 0.75934008  -0.9516642    0.5898733
#> lag1_var4  1.2219543 0.64732475   0.5183528    1.8378732
#> const      7.2131400 8.35809769  -2.0089177   14.4596023
#> exo1       0.2875711 0.20179488   0.1018432    0.5093888
#> exo2      -1.3128338 0.26429423  -1.6738351   -1.1347258
#> exo3       3.3227288 0.35144435   2.9021842    3.7138988
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                  mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00230468 0.000341385 0.001906586  0.002685676
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.08209948 0.03359528  -0.1200578  -0.05006124
#> Sigma[2,2]  8.85975541 3.48094563   5.8586903  12.90472601
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04545346 0.02445948  0.02083491   0.07653777
#> Sigma[3,2] -6.12187175 2.16507801 -8.42139914  -3.88601347
#> Sigma[3,3]  6.38105739 1.97063109  3.94961593   8.10650455
#> 
#> $ALB$Sigma$equation4
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.009887305 0.01576708 -0.02952029  0.004331053
#> Sigma[4,2] -0.301965699 0.92806044 -1.49650997  0.614725421
#> Sigma[4,3]  2.750705074 0.98885101  1.59221343  3.878836464
#> Sigma[4,4]  3.035418453 1.03699280  1.85271352  4.259122474
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  1.03937313 0.01445315  1.022162767  1.054527841
#> lag1_var2  0.03937620 0.02731512  0.004449843  0.066323469
#> lag1_var3  0.04730468 0.04340616 -0.008219464  0.089683905
#> lag1_var4 -0.03818548 0.03428959 -0.070210352  0.006464261
#> const     -1.57111771 0.58591645 -2.090166303 -0.795657747
#> exo1       0.04489455 0.03252009  0.008470365  0.083147196
#> exo2      -0.05890442 0.03573057 -0.095388137 -0.014294302
#> exo3       0.12010725 0.01955873  0.100828676  0.142604730
#> 
#> $ALB$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  -5.0080285  1.6882671  -7.2604325    -3.619179
#> lag1_var2  -0.2476594  1.8788478  -1.9929332     2.223470
#> lag1_var3  -0.7713565  2.8512576  -3.4101983     2.965153
#> lag1_var4   0.6682996  2.4229296  -2.4843363     3.003371
#> const     132.6514547 46.9377788  91.3624757   194.757197
#> exo1       -1.5914896  1.7980655  -3.4341892     0.685092
#> exo2        1.9544122  4.0980425  -0.5990808     7.626778
#> exo3       -2.8965293  0.9560687  -4.1497949    -2.070705
#> 
#> $ALB$A$equation3
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1   2.3673466  1.708700    0.7474535   4.46308726
#> lag1_var2  -0.5837396  3.885663   -5.3233358   3.49670940
#> lag1_var3  -0.3939829  5.846782   -7.4783620   5.84188640
#> lag1_var4   1.0049724  4.928532   -4.2787324   6.95479997
#> const     -34.6150529 79.856389 -111.1585150  65.67560946
#> exo1       -1.2207559  1.604000   -3.1366791   0.40556996
#> exo2       -2.4410136  2.707607   -5.9974762   0.08567417
#> exo3        1.7942590  1.220727    0.6715620   3.38438280
#> 
#> $ALB$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -1.09951429  1.222142   -2.363074    0.3095567
#> lag1_var2 -1.15088548  3.255069   -4.824607    2.3640856
#> lag1_var3 -1.43790550  4.936404   -6.920229    4.0051655
#> lag1_var4  2.04129497  4.121450   -2.536033    6.6468054
#> const     53.37322899 64.382877   -3.744652  134.3125298
#> exo1      -2.21553421  2.973745   -5.410712    0.6353722
#> exo2      -1.29734482  1.516937   -3.379270   -0.2285089
#> exo3       0.00727191  1.000254   -1.117593    1.0083884
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                  mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00184325 0.0003829451 0.001401658  0.002293062
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd  5% quantile  95% quantile
#> Sigma[2,1] -0.001917908 0.001834385 -0.004281738 -9.727666e-05
#> Sigma[2,2]  0.053557157 0.014837076  0.037279060  7.151674e-02
#> 
#> $ARE$Sigma$equation3
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.0009879315 0.003018863 -0.001974411  0.004905312
#> Sigma[3,2] -0.0642296246 0.017527586 -0.083339508 -0.043941870
#> Sigma[3,3]  0.2076794519 0.073468640  0.143073982  0.303785487
#> 
#> $ARE$Sigma$equation4
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[4,1] -8.136009e-05 0.003771217 -0.004747421  0.003961409
#> Sigma[4,2] -2.312587e-02 0.008587153 -0.034050292 -0.014881643
#> Sigma[4,3]  1.595145e-01 0.065168367  0.096785206  0.242568846
#> Sigma[4,4]  1.434134e-01 0.062394219  0.082918113  0.222543399
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                    mean         sd 5% quantile 95% quantile
#> lag1_var1  0.9962618189 0.03114305  0.95634247  1.019849917
#> lag1_var2  0.5092557483 0.03692252  0.46565625  0.549801659
#> lag1_var3  0.6398122316 0.04095006  0.58444845  0.668790598
#> lag1_var4 -0.6277261968 0.04288618 -0.65993791 -0.569848899
#> const     -0.8291757786 0.63419521 -1.35653990 -0.005162530
#> exo1       0.0004569761 0.02331189 -0.03108765  0.016678211
#> exo2      -0.0892557680 0.02882341 -0.12615943 -0.062096530
#> exo3      -0.0112393835 0.01944970 -0.03253358  0.006627119
#> 
#> $ARE$A$equation2
#>                   mean        sd 5% quantile 95% quantile
#> lag1_var1  0.105685888 0.1487776 -0.09637447   0.22151725
#> lag1_var2  1.155918535 0.2886556  0.81249390   1.42444658
#> lag1_var3  0.514171923 0.4832620 -0.10314061   0.98361398
#> lag1_var4 -0.550187866 0.4962251 -1.06529167   0.06401774
#> const      0.515400456 1.6822190 -1.12344999   2.54299316
#> exo1       0.002912823 0.3234743 -0.27883668   0.41924949
#> exo2       2.107060583 0.2839602  1.82258948   2.44963977
#> exo3      -0.633069311 0.5327464 -1.09975331  -0.02603069
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.4561769 0.3573828   0.1022789    0.8575689
#> lag1_var2 -1.7740805 0.7169134  -2.6124351   -0.9860575
#> lag1_var3 -1.5077672 0.8416542  -2.3499543   -0.4401743
#> lag1_var4  2.3630376 0.7842569   1.3711038    3.1595632
#> const     -1.0476666 5.9142517  -8.9198230    4.2661428
#> exo1      -0.3350668 0.5189774  -0.8976883    0.3021341
#> exo2      -2.5328907 0.4526518  -2.9756395   -1.9582450
#> exo3      -1.3053788 0.7792526  -2.3528596   -0.6993172
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.5463392 0.3205735   0.2092631    0.9298073
#> lag1_var2 -0.9601837 0.5112671  -1.6012134   -0.4335743
#> lag1_var3 -1.2026574 0.4808247  -1.7587643   -0.6700355
#> lag1_var4  2.0529775 0.4319530   1.6060997    2.5929698
#> const     -2.6160484 6.4059471 -11.3653071    2.0468216
#> exo1      -0.3549136 0.4089773  -0.7354401    0.1151541
#> exo2      -0.9091101 0.5072852  -1.3731725   -0.2432247
#> exo3      -1.8387774 0.6280814  -2.5802498   -1.1863457
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003154994 0.0003159155 0.002757863  0.003444659
#> 
#> $ARG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.06346797 0.009711812 -0.07460329  -0.05284651
#> Sigma[2,2]  2.74354927 0.416476328  2.25465123   3.22422835
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04894335 0.01272511  0.03313107   0.06253908
#> Sigma[3,2] -1.91070725 0.28020087 -2.18895372  -1.57330058
#> Sigma[3,3]  1.67786973 0.28432037  1.37162650   2.01675056
#> 
#> $ARG$Sigma$equation4
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01220503 0.01081897  0.001098203   0.02588729
#> Sigma[4,2] -0.27253212 0.15176270 -0.462058525  -0.13182748
#> Sigma[4,3]  0.57180498 0.14740971  0.440190620   0.76296655
#> Sigma[4,4]  0.44867012 0.14995437  0.278198850   0.59658772
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.912951375 0.05668542  0.86282607   0.98010990
#> lag1_var2  0.053797651 0.11610138 -0.07777991   0.19079439
#> lag1_var3  0.087366683 0.19260007 -0.13137358   0.31389071
#> lag1_var4 -0.088418034 0.16648777 -0.28268441   0.10299159
#> const      2.416902502 2.17539722 -0.55836221   4.08861869
#> exo1       0.075092654 0.03162996  0.05110114   0.11803722
#> exo2      -0.131701910 0.06461006 -0.19620376  -0.04860178
#> exo3      -0.003472461 0.04257813 -0.04978722   0.04027879
#> 
#> $ARG$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.7687282  1.038947  -1.8312732    0.4448619
#> lag1_var2 -3.9234781  4.404315  -8.9291455    0.1835687
#> lag1_var3 -7.6515286  7.283724 -15.8918602   -0.9490907
#> lag1_var4  7.2987465  6.492583   1.2962453   14.8120783
#> const     44.6624480 33.778591  20.4695479   90.9176607
#> exo1      -2.0189597  1.133514  -3.4279022   -0.8096975
#> exo2       2.8206415  1.743468   0.5211989    4.3509072
#> exo3       1.6301725  2.635580  -1.2032826    4.8622879
#> 
#> $ARG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.03853448  1.2137616  -1.2174757     1.461871
#> lag1_var2  3.26358774  3.8128707  -1.2719730     6.986635
#> lag1_var3  6.03656075  6.3194641  -1.4533103    12.232253
#> lag1_var4 -5.15425839  5.4816973 -10.5171184     1.322174
#> const      6.35276150 49.6939506 -39.5185987    59.638388
#> exo1       0.31730477  0.7177531  -0.5073477     1.026737
#> exo2      -5.65142585  1.2980406  -7.2218964    -4.204975
#> exo3       0.76417511  1.8764761  -1.1373153     2.902816
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.5569650  0.8222656  -1.5239878    0.3703945
#> lag1_var2  0.6406259  1.5593053  -1.4913219    1.7491059
#> lag1_var3  0.9824349  2.6079110  -2.5883591    2.7858239
#> lag1_var4 -0.2928956  2.1859929  -1.8814223    2.6511306
#> const     33.5059418 40.9517485 -11.2812859   75.5447728
#> exo1      -0.9993601  0.4922887  -1.5324856   -0.4366399
#> exo2      -4.4616681  0.7484534  -5.2883969   -3.5529605
#> exo3       1.9949588  1.1165146   0.6582054    3.1297749
#> 
#> 
#>