Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARGROUPPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARGROUPPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarGroupPANEL$new(
      data = ilo_dynamic_panel[1:5],
      exogenous = ilo_exogenous_variables[1:5],
      group_allocation = country_grouping_region[1:5]
)
#> Country groupings have been pre-specified and will not be estimated.
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004957324 0.00202864 0.002765502  0.007452111
#> 
#> $AFG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01471858 0.00735714 -0.02383658  -0.00722322
#> Sigma[2,2]  0.20306940 0.06090180  0.16131441   0.28553078
#> 
#> $AFG$Sigma$equation3
#>                  mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.0182527 0.008376845  0.008926492   0.02743014
#> Sigma[3,2] -0.1896146 0.043304546 -0.247598749  -0.15305464
#> Sigma[3,3]  0.3065807 0.058117332  0.243695592   0.37497343
#> 
#> $AFG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01345453 0.006494958  0.006224432   0.02129416
#> Sigma[4,2] -0.11314009 0.024096978 -0.137100172  -0.08462107
#> Sigma[4,3]  0.25294281 0.049698394  0.200045832   0.31471017
#> Sigma[4,4]  0.23121966 0.044409012  0.185493645   0.28649835
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.99082128 0.01307620  0.98147077   1.00858615
#> lag1_var2 -0.02761961 0.24637823 -0.35765796   0.16980291
#> lag1_var3 -0.05741235 0.56704264 -0.80780027   0.41555335
#> lag1_var4  0.05613664 0.50137601 -0.36720497   0.71644271
#> const      0.34095744 3.07085316 -2.17553650   4.44363542
#> exo1       0.02815707 0.10963270 -0.09049091   0.16289727
#> exo2      -0.12655524 0.13682256 -0.31078865  -0.01080482
#> exo3      -0.16996928 0.05868848 -0.24528544  -0.12178209
#> 
#> $AFG$A$equation2
#>                 mean          sd 5% quantile 95% quantile
#> lag1_var1  0.3004413  0.08098129   0.1933496    0.3690668
#> lag1_var2 -0.3916445  1.68465256  -2.0206832    1.4312884
#> lag1_var3 -2.9562824  3.93723163  -6.6696906    1.2932027
#> lag1_var4  2.5364300  3.50859016  -1.2961235    5.8244395
#> const     12.9342080 20.07783385 -10.1302055   33.9548195
#> exo1      -0.3968551  0.75972710  -1.3066381    0.4567243
#> exo2      -0.1650955  0.79000967  -0.8531192    0.8437779
#> exo3      -0.4067085  0.49289777  -0.9716918    0.1448903
#> 
#> $AFG$A$equation3
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1  -0.04929812  0.1840048  -0.27409470   0.15031940
#> lag1_var2   3.11142919  2.6501141  -0.24858339   5.45972780
#> lag1_var3   8.28155657  5.9902234   0.91099742  13.62966068
#> lag1_var4  -6.39281164  5.2806294 -11.11807734   0.02969308
#> const     -38.14462853 32.0458368 -67.37397268   2.49962987
#> exo1        0.36372333  0.6336682  -0.23727839   1.04242720
#> exo2       -0.81786254  1.0074657  -2.11812833   0.18076825
#> exo3        0.36952384  0.6152316  -0.08319129   1.17578801
#> 
#> $AFG$A$equation4
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1   0.09884151  0.1960526  -0.12698970    0.3260634
#> lag1_var2   3.38834365  2.1948909   0.49309894    5.2189307
#> lag1_var3   7.96652364  4.8896630   1.64851669   12.1176226
#> lag1_var4  -6.00263553  4.2890243  -9.66393178   -0.5070839
#> const     -44.65582704 26.4831839 -66.83199792  -10.0932623
#> exo1        0.22783211  0.5075712  -0.21889959    0.8932847
#> exo2       -0.90980540  0.7948099  -1.88127450   -0.0486427
#> exo3        0.32683793  0.5226111  -0.08701331    1.0261162
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005736053 0.001970241 0.004095735  0.008375829
#> 
#> $AGO$Sigma$equation2
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.008575195 0.003010588 -0.01228689 -0.005716553
#> Sigma[2,2]  0.113377678 0.025841109  0.08132010  0.136943315
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.007558505 0.002982226  0.004766484   0.01089745
#> Sigma[3,2] -0.048670584 0.013757560 -0.064336222  -0.03317456
#> Sigma[3,3]  0.033481359 0.009365944  0.023196374   0.04361467
#> 
#> $AGO$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.001827266 0.003088200 -0.001623108 0.0053056433
#> Sigma[4,2]  0.045895684 0.012939465  0.028046968 0.0543680888
#> Sigma[4,3] -0.005028759 0.005212541 -0.010872784 0.0005712344
#> Sigma[4,4]  0.036962427 0.007050727  0.027729247 0.0433372771
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.99688961 0.02308374  0.967693028   1.01872184
#> lag1_var2  0.08020669 0.04328113  0.030694270   0.12624471
#> lag1_var3  0.15460605 0.02567164  0.126475033   0.18174333
#> lag1_var4 -0.10482446 0.03766629 -0.147388524  -0.06362312
#> const     -3.08282920 1.35028054 -4.559718275  -1.65407621
#> exo1       0.03629307 0.05628323 -0.013482932   0.11147336
#> exo2       0.06798237 0.04858818  0.007424425   0.11676383
#> exo3       0.07964720 0.09483220 -0.025483466   0.18688113
#> 
#> $AGO$A$equation2
#>                   mean          sd  5% quantile 95% quantile
#> lag1_var1  -0.08038454  0.15007113  -0.25402324   0.09332862
#> lag1_var2   1.81753974  0.18320751   1.58421009   1.94909536
#> lag1_var3   1.10836766  0.09823455   0.97827614   1.19841424
#> lag1_var4  -0.89580963  0.14169358  -1.04856735  -0.74977109
#> const     -13.80095770 12.26942431 -28.08365715   0.20509081
#> exo1        0.21557218  0.21300631  -0.02899065   0.44917551
#> exo2       -0.18571994  0.33620232  -0.62344028   0.12966943
#> exo3       -0.84449262  0.42839655  -1.39658177  -0.44771806
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.15291907 0.04877300  -0.2124712   -0.1014834
#> lag1_var2 -0.94632597 0.04943780  -1.0085549   -0.8945845
#> lag1_var3 -0.20389805 0.04003727  -0.2434365   -0.1538647
#> lag1_var4  1.00981027 0.06334867   0.9434665    1.0895397
#> const     18.94727062 6.93875861  11.0173769   26.6416966
#> exo1      -0.01353062 0.13506524  -0.1631539    0.1433126
#> exo2      -1.16238113 0.15167678  -1.3492255   -1.0165155
#> exo3       3.35921133 0.28831264   3.0383933    3.6755827
#> 
#> $AGO$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.2965473 0.09715964 -0.40044437   -0.1810797
#> lag1_var2  0.4883320 0.12019988  0.34428584    0.5995674
#> lag1_var3  0.7223821 0.03843762  0.67268017    0.7590684
#> lag1_var4  0.3917302 0.10719801  0.29012850    0.5073951
#> const     -0.4711065 6.93617722 -9.54941552    4.5504218
#> exo1       0.1900982 0.17871627 -0.01202705    0.3833468
#> exo2      -1.6139467 0.25004677 -1.89056035   -1.3133103
#> exo3       3.1365509 0.27770744  2.93822216    3.5110329
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002098667 0.00090122 0.001125515  0.003136253
#> 
#> $ALB$Sigma$equation2
#>                   mean        sd 5% quantile 95% quantile
#> Sigma[2,1] -0.06285921 0.0473995  -0.1146306  -0.01788928
#> Sigma[2,2]  7.98236826 3.1529815   4.8976337  11.50270395
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd   5% quantile 95% quantile
#> Sigma[3,1]  0.02388743 0.02754777  9.458578e-05   0.05993477
#> Sigma[3,2] -4.97268389 1.73900772 -7.277746e+00  -3.39411126
#> Sigma[3,3]  5.18776934 2.03961467  3.748591e+00   7.98444414
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.02131228 0.01607648 -0.04340847  -0.01190403
#> Sigma[4,2]  0.47320685 1.14720703 -0.38066778   2.04609008
#> Sigma[4,3]  2.17539988 1.47014596  0.55523031   3.88699686
#> Sigma[4,4]  2.96570155 1.20140941  1.87822551   4.55968096
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  1.027420388 0.04294957  0.97397034   1.06903745
#> lag1_var2  0.002950910 0.02046688 -0.01865536   0.02728290
#> lag1_var3 -0.007991207 0.03031185 -0.04305215   0.02684186
#> lag1_var4  0.008571244 0.02471364 -0.01857530   0.03829836
#> const     -0.739670874 1.32049646 -2.19794135   0.75401436
#> exo1       0.054241622 0.07522581 -0.03377641   0.13698515
#> exo2      -0.048803998 0.07055214 -0.12075065   0.03134489
#> exo3       0.069426240 0.03126066  0.04350278   0.11068702
#> 
#> $ALB$A$equation2
#>                mean        sd 5% quantile 95% quantile
#> lag1_var1 -4.698801  3.167004   -9.043375   -2.7442720
#> lag1_var2  3.052628  1.633602    1.252444    5.0346612
#> lag1_var3  4.304522  2.659714    1.418114    7.5757535
#> lag1_var4 -3.648721  2.355706   -6.601162   -1.1880389
#> const     76.416447 65.910653   28.320080  166.3562887
#> exo1      -1.861539  2.220728   -3.800657    0.8982963
#> exo2       0.145346  1.824561   -1.875534    2.1984601
#> exo3      -1.279201  3.329383   -3.489653    3.3190669
#> 
#> $ALB$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  2.2035660  2.448859  -0.6148538    5.0730125
#> lag1_var2 -2.9532940  4.282771  -8.2533284    1.3550258
#> lag1_var3 -3.9590115  6.506113 -12.0455650    2.6434377
#> lag1_var4  4.0509747  5.508055  -1.4934433   10.9558940
#> const      3.7361779 77.793055 -55.6667192  108.8243265
#> exo1      -0.4311089  2.825501  -3.3615944    3.0897477
#> exo2      -1.7853164  2.621212  -5.3140410    0.3033779
#> exo3       0.1055583  2.982821  -3.9088732    2.4301962
#> 
#> $ALB$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -1.0028952  1.869877   -3.370520    0.7233159
#> lag1_var2 -1.4311615  3.734943   -6.029268    2.2051656
#> lag1_var3 -1.7906984  5.540736   -8.625289    3.6948869
#> lag1_var4  2.3558117  4.618346   -2.232327    8.0875903
#> const     54.6185654 91.378986  -32.205404  160.3765175
#> exo1      -1.5431128  1.935178   -2.877649    1.0933442
#> exo2      -1.9115124  3.302630   -6.238602    1.2196358
#> exo3      -0.8253237  1.636173   -2.800277    0.4798258
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001363062 0.0002458059 0.001116828  0.001657346
#> 
#> $ARE$Sigma$equation2
#>                    mean           sd  5% quantile 95% quantile
#> Sigma[2,1] -0.002868597 0.0009338481 -0.003924826 -0.001867626
#> Sigma[2,2]  0.053550546 0.0071641659  0.045870247  0.061016687
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.004336004 0.003480805  0.001604083  0.009015409
#> Sigma[3,2] -0.043414073 0.020366232 -0.068194231 -0.025010695
#> Sigma[3,3]  0.146830481 0.052099244  0.092285963  0.208443815
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.002291587 0.003030556  7.319242e-05   0.00635814
#> Sigma[4,2] -0.002490646 0.018623840 -2.521136e-02   0.01420370
#> Sigma[4,3]  0.115965812 0.036916317  7.508554e-02   0.15843630
#> Sigma[4,4]  0.116667955 0.024937892  8.487395e-02   0.14108842
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.98820514 0.03462034  0.945611345   1.02400453
#> lag1_var2  0.47552559 0.05420400  0.419625566   0.54075994
#> lag1_var3  0.62507541 0.04171277  0.580365709   0.67361077
#> lag1_var4 -0.61191646 0.03036981 -0.650240506  -0.58413573
#> const     -0.65268946 0.56828683 -1.259492039  -0.04169041
#> exo1       0.02063365 0.02415488 -0.002546955   0.05215366
#> exo2      -0.09208361 0.03869750 -0.138104997  -0.04973347
#> exo3       0.05672793 0.03989965  0.004326184   0.09153293
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.16308551 0.0790973  -0.2405040  -0.06495415
#> lag1_var2  1.85504140 0.2642542   1.5018536   2.06479827
#> lag1_var3  1.29575083 0.2080811   1.0255004   1.48591003
#> lag1_var4 -1.23401430 0.2035761  -1.4281446  -0.97136333
#> const     -0.08910219 1.6899175  -1.7752149   2.05980565
#> exo1      -0.16297708 0.1727205  -0.3353705   0.05064471
#> exo2       1.91828258 0.2187875   1.6480832   2.15003919
#> exo3      -0.85449229 0.1905387  -1.0801459  -0.63701749
#> 
#> $ARE$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.23148323 0.4394819  -0.3471027    0.6283601
#> lag1_var2  0.19780457 0.7278723  -0.5518955    1.1232461
#> lag1_var3  1.01883092 0.4755340   0.5097172    1.6040346
#> lag1_var4 -0.06975231 0.4132259  -0.5511396    0.4083876
#> const     -2.43288572 3.4398050  -5.8519413    1.4591257
#> exo1      -0.31999727 0.5842677  -1.0445505    0.3061111
#> exo2      -2.35695538 0.3546719  -2.8111736   -1.9997623
#> exo3      -1.26436947 0.7883263  -2.1961716   -0.5271582
#> 
#> $ARE$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.09440716 0.3993599  -0.4327241    0.4435274
#> lag1_var2  1.66279261 0.6139964   0.9325152    2.3467958
#> lag1_var3  2.06377997 0.3977841   1.5540202    2.4666842
#> lag1_var4 -1.03898191 0.3540931  -1.3182391   -0.5597292
#> const     -4.36053361 3.3414178  -8.0919345   -1.2807003
#> exo1      -0.47345369 0.6342368  -1.2962259    0.1054431
#> exo2      -0.87369573 0.3678460  -1.3106718   -0.5022093
#> exo3      -1.97740552 0.6793838  -2.8203810   -1.4038547
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                  mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00303712 0.001004765  0.00198349  0.004298648
#> 
#> $ARG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.07146974 0.009005828 -0.08069436  -0.06029119
#> Sigma[2,2]  3.64060021 0.528491246  3.11325068   4.31667648
#> 
#> $ARG$Sigma$equation3
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.05535793 0.008648249  0.04468346   0.06459117
#> Sigma[3,2] -2.62408003 0.259493301 -2.94653423  -2.36580853
#> Sigma[3,3]  2.11647186 0.112663174  1.99232620   2.23905828
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01381294 0.006072167  0.007845756   0.02093605
#> Sigma[4,2] -0.45425450 0.120918910 -0.556655404  -0.30208041
#> Sigma[4,3]  0.57936439 0.108148623  0.453601492   0.68620632
#> Sigma[4,4]  0.34001888 0.063742415  0.261397439   0.40790787
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.91726036 0.01639600  0.90484360   0.93958743
#> lag1_var2  0.21901013 0.12286027  0.04967888   0.29861570
#> lag1_var3  0.36163584 0.20127106  0.08438258   0.48816399
#> lag1_var4 -0.32646567 0.17684278 -0.43721242  -0.08364671
#> const      0.11559195 1.21631083 -0.99027227   1.69246603
#> exo1       0.04076576 0.05713837 -0.03143034   0.09840038
#> exo2      -0.08290653 0.01265239 -0.09995678  -0.07335720
#> exo3       0.04408630 0.05958678 -0.01477328   0.11988649
#> 
#> $ARG$A$equation2
#>                mean         sd 5% quantile 95% quantile
#> lag1_var1 -2.978400  0.9260357  -4.1250650   -2.0733299
#> lag1_var2  1.555133  5.3884317  -3.6896797    8.5280669
#> lag1_var3  1.257476  8.8918466  -7.4429645   12.7486446
#> lag1_var4 -1.066391  7.9976427 -11.2392418    7.0675922
#> const     70.964968 53.1942239   0.3066533  119.4747676
#> exo1      -1.475803  2.1341425  -4.1939861    0.3392154
#> exo2       1.512652  1.4389726   0.1022664    3.2328109
#> exo3      -2.252091  3.5237239  -4.9654318    2.5226567
#> 
#> $ARG$A$equation3
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1   1.13585463  0.4027095   0.7775743     1.618158
#> lag1_var2   1.79464903  4.2894381  -3.9836459     5.203177
#> lag1_var3   3.73309043  7.0487784  -5.7674918     9.368062
#> lag1_var4  -2.69078497  6.3201432  -7.9310588     5.748820
#> const     -34.55347018 43.9320885 -73.3618970    24.086776
#> exo1        0.05236511  1.3147060  -1.3389761     1.657747
#> exo2       -4.99013531  1.1559391  -6.5248315    -3.958585
#> exo3        3.96635986  2.5668236   0.4599594     5.745768
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.7945043  0.3503697   -1.177319   -0.4485419
#> lag1_var2  3.0025389  1.4919335    1.120779    4.5122448
#> lag1_var3  4.9252693  2.4510965    1.849638    7.4412711
#> lag1_var4 -3.6502207  2.2061463   -5.895880   -0.8749571
#> const      4.7256475 13.4544174   -5.315173   22.4437652
#> exo1      -0.9134874  0.5444343   -1.451704   -0.2717025
#> exo2      -4.6041940  0.6792890   -5.291453   -3.8369814
#> exo3       2.8418217  0.5853083    2.228043    3.5570087
#> 
#> 
#>