Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARGROUPPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARGROUPPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarGroupPANEL$new(
      data = ilo_dynamic_panel[1:5],
      exogenous = ilo_exogenous_variables[1:5],
      group_allocation = country_grouping_region[1:5]
)
#> Country groupings have been pre-specified and will not be estimated.
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004060212 0.0009651453 0.002996854  0.005174017
#> 
#> $AFG$Sigma$equation2
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.005635548 0.005033642 -0.01042775 0.0006838477
#> Sigma[2,2]  0.166036057 0.044472818  0.11187565 0.2122048225
#> 
#> $AFG$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.004122554 0.009759644 -0.007393692   0.01413837
#> Sigma[3,2] -0.155146555 0.038078913 -0.205332826  -0.12094765
#> Sigma[3,3]  0.313951216 0.081641457  0.224002702   0.41239428
#> 
#> $AFG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.00203000 0.008722064 -0.008344944   0.01072122
#> Sigma[4,2] -0.09341091 0.032431996 -0.130462101  -0.05511634
#> Sigma[4,3]  0.27963397 0.075997928  0.189330013   0.36376012
#> Sigma[4,4]  0.27228566 0.069846473  0.187731667   0.34647269
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1  0.99463223 0.018556726  0.96918466   1.00816193
#> lag1_var2  0.23868663 0.005392442  0.23400544   0.24500935
#> lag1_var3  0.52935916 0.019113409  0.50555290   0.54940048
#> lag1_var4 -0.45494052 0.021297673 -0.47694813  -0.42891168
#> const     -3.21552578 0.448076085 -3.60913011  -2.70230758
#> exo1       0.01322114 0.038595878 -0.02171158   0.05715170
#> exo2      -0.04413327 0.080750651 -0.13361690   0.05319615
#> exo3      -0.10365785 0.067916451 -0.17040020  -0.01769299
#> 
#> $AFG$A$equation2
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.254814359 0.08210334  0.17045680    0.3568673
#> lag1_var2  0.836609844 0.14595793  0.66850955    0.9910216
#> lag1_var3 -0.334186903 0.26164637 -0.54832502   -0.0343191
#> lag1_var4  0.290977634 0.21989363  0.03773902    0.4906970
#> const     -3.677394561 3.93751962 -7.92952470    1.0885591
#> exo1       0.224447684 0.32532629 -0.14356334    0.5828933
#> exo2      -0.003486967 0.49306734 -0.62222195    0.5034180
#> exo3      -0.110694915 0.45082349 -0.69350919    0.3019432
#> 
#> $AFG$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.11917262 0.2064288  -0.3859051   0.08379564
#> lag1_var2 -0.89625633 0.1523700  -1.0726582  -0.73312600
#> lag1_var3 -0.24681812 0.4024635  -0.7133106   0.15487130
#> lag1_var4  0.96649989 0.4108730   0.5375411   1.43048664
#> const     18.31281222 2.6552205  15.4601036  21.27490963
#> exo1      -0.49905649 0.8136500  -1.1476673   0.56581863
#> exo2      -1.62964000 0.4573794  -2.2479561  -1.25884721
#> exo3      -0.05018415 0.5749603  -0.5549371   0.70972127
#> 
#> $AFG$A$equation4
#>                   mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.001439353 0.2192739  -0.2788749    0.2182892
#> lag1_var2 -0.578808551 0.1009609  -0.7017422   -0.4768252
#> lag1_var3 -0.473430899 0.3331453  -0.8471123   -0.1359000
#> lag1_var4  1.289829082 0.3655111   0.9162341    1.7010307
#> const     11.488083801 3.2895657   8.1676324   15.1268383
#> exo1      -0.439029214 0.8041929  -1.1594051    0.5827100
#> exo2      -1.732688982 0.5110307  -2.4085632   -1.3331773
#> exo3       0.022482421 0.5098283  -0.5084994    0.6572349
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005914 0.001151029 0.004598627  0.007278849
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01151828 0.005955855 -0.01653065 -0.003751477
#> Sigma[2,2]  0.16598316 0.027948450  0.13113634  0.194692171
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.008398069 0.002565860  0.005082098   0.01087903
#> Sigma[3,2] -0.065691303 0.006972522 -0.073924614  -0.05777460
#> Sigma[3,3]  0.038791379 0.005788171  0.034452938   0.04666175
#> 
#> $AGO$Sigma$equation4
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  1.579804e-05 0.003101437 -0.003356674  0.003787488
#> Sigma[4,2]  7.160680e-02 0.021129058  0.048192097  0.097332268
#> Sigma[4,3] -1.257014e-02 0.004534884 -0.017024781 -0.007103794
#> Sigma[4,4]  5.158631e-02 0.016590557  0.034866594  0.071994367
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1  0.95976568 0.030132297   0.9265659   0.99620658
#> lag1_var2  0.31424516 0.009201067   0.3045902   0.32481577
#> lag1_var3  0.46642661 0.026922312   0.4352219   0.49751935
#> lag1_var4 -0.39436487 0.011416929  -0.4065109  -0.38021458
#> const     -3.78379862 2.660908147  -6.8744280  -1.06068607
#> exo1       0.05921867 0.119235759  -0.0825171   0.18370649
#> exo2      -0.12452266 0.050481391  -0.1800869  -0.06525472
#> exo3      -0.02893347 0.103300942  -0.1608751   0.07234395
#> 
#> $AGO$A$equation2
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.04968027 0.06679307  -0.02203138   0.13177629
#> lag1_var2  0.82226738 0.08552759   0.71912123   0.90468456
#> lag1_var3 -0.29039421 0.11697381  -0.44252725  -0.17661655
#> lag1_var4  0.34455075 0.09094758   0.23785090   0.44663186
#> const     -6.19769928 3.99882783 -10.34358624  -1.61986060
#> exo1       0.16967113 0.35941170  -0.32393055   0.43466947
#> exo2       0.62251924 0.52996898   0.08778442   1.30132822
#> exo3      -0.76528371 0.53764243  -1.32370811  -0.08745335
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.23287981 0.04807380  -0.2719871   -0.1686642
#> lag1_var2 -0.38270960 0.03577637  -0.4180706   -0.3454388
#> lag1_var3  0.52677195 0.05924536   0.4782114    0.6053011
#> lag1_var4  0.34033865 0.02471468   0.3102517    0.3652463
#> const     16.20680846 4.19163589  10.8298210   20.4293695
#> exo1      -0.01401019 0.22261336  -0.2328719    0.2609136
#> exo2      -1.58724574 0.22615403  -1.8341352   -1.3263913
#> exo3       3.20526181 0.27570152   2.9090487    3.5451640
#> 
#> $AGO$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.2748974 0.06360536  -0.3304156   -0.1918633
#> lag1_var2  0.2601239 0.06386425   0.1985714    0.3359542
#> lag1_var3  0.3282543 0.08666512   0.2149925    0.4111740
#> lag1_var4  0.7132403 0.07925780   0.6385400    0.8171701
#> const      3.3349762 4.81991681  -3.1167012    7.5260912
#> exo1       0.1634817 0.25061495  -0.1100990    0.4523574
#> exo2      -1.4184328 0.27621292  -1.5959404   -1.0437956
#> exo3       3.0417708 0.19507981   2.8931721    3.3089022
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001862974 0.0006532222 0.001364301  0.002750403
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.03007942 0.01124462 -0.03962318  -0.01675405
#> Sigma[2,2]  7.30529643 1.83050868  5.36944898   9.33178985
#> 
#> $ALB$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.00645492 0.008934146 -0.004198638   0.01588469
#> Sigma[3,2] -4.74055262 0.914775762 -5.471746174  -3.50666597
#> Sigma[3,3]  4.87785223 1.357938780  3.655504312   6.67911194
#> 
#> $ALB$Sigma$equation4
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[4,1] -0.01601705 0.004187714 -0.02064548  -0.01176737
#> Sigma[4,2]  0.14342492 0.897477392 -1.00258415   1.04985988
#> Sigma[4,3]  2.10455651 1.615612306  0.86913465   4.28380640
#> Sigma[4,4]  2.69881033 1.365069799  1.50142421   4.50925451
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  1.03693818 0.01556695  1.016869618  1.052169593
#> lag1_var2  0.01724325 0.01301507  0.004042570  0.033771715
#> lag1_var3  0.01454367 0.02036141 -0.007474535  0.039083805
#> lag1_var4 -0.01218678 0.01808319 -0.032886573  0.008569025
#> const     -1.08663398 0.47167053 -1.674677030 -0.589838703
#> exo1       0.02873061 0.03745154 -0.009523770  0.070736309
#> exo2      -0.02578252 0.02610852 -0.054626001  0.001502772
#> exo3       0.05301639 0.02858210  0.016765213  0.082608625
#> 
#> $ALB$A$equation2
#>                mean        sd 5% quantile 95% quantile
#> lag1_var1 -3.777868  1.148916  -5.0205152   -2.5231072
#> lag1_var2  1.786081  1.370927   0.5761922    3.6214882
#> lag1_var3  2.250658  2.274249   0.1632867    5.2662434
#> lag1_var4 -1.770233  1.887948  -4.2603862   -0.0154381
#> const     66.609914 20.143773  42.9882683   87.4516133
#> exo1      -1.428824  3.940117  -5.0553470    3.1019554
#> exo2      -1.613703  2.890274  -4.6727771    1.6473686
#> exo3      -1.375001  4.018621  -4.2131528    4.1232411
#> 
#> $ALB$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1   2.8518158  2.153043   0.5576707    5.2980977
#> lag1_var2  -0.5322445  1.520107  -1.6374044    1.5545349
#> lag1_var3  -0.4117265  2.459768  -2.0818727    2.9847716
#> lag1_var4   1.0480724  2.136144  -1.9001953    2.5095536
#> const     -48.5751460 41.988941 -96.0612170    1.1723674
#> exo1       -1.2652355  1.563275  -2.8185172    0.6469135
#> exo2        0.2431346  2.797766  -2.4906333    3.8625788
#> exo3        1.3160791  3.299277  -2.9418882    4.4583973
#> 
#> $ALB$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1   0.4364784  2.307774   -2.097568     2.723872
#> lag1_var2   0.4468297  1.943598   -1.537074     2.813540
#> lag1_var3   0.8126194  3.055651   -2.365231     4.591450
#> lag1_var4   0.2548185  2.641808   -3.014617     3.035250
#> const     -13.4996075 51.736809  -80.029553    36.124789
#> exo1       -2.3017494  1.736268   -4.353307    -0.520164
#> exo2       -0.7089739  2.497967   -3.367231     2.004158
#> exo3        0.5980455  1.714637   -1.129574     2.782335
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005163834 0.002720797 0.002548599  0.008397419
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.004439506 0.006031387 -0.01220621  0.001352916
#> Sigma[2,2]  0.066953124 0.024147129  0.04599654  0.098271789
#> 
#> $ARE$Sigma$equation3
#>                    mean         sd  5% quantile 95% quantile
#> Sigma[3,1]  0.006223712 0.01340988 -0.006772196   0.02296021
#> Sigma[3,2] -0.080239854 0.04906309 -0.146024703  -0.04248180
#> Sigma[3,3]  0.231821996 0.08891149  0.152023799   0.34989649
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[4,1]  0.003075112 0.008932803 -0.00526706   0.01403671
#> Sigma[4,2] -0.029711680 0.034394798 -0.07632138  -0.00441144
#> Sigma[4,3]  0.172044079 0.055283465  0.11734871   0.24341917
#> Sigma[4,4]  0.151094723 0.039393904  0.10828148   0.19290731
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.78660621 0.11137513   0.6615917   0.91010320
#> lag1_var2  0.86362400 0.18631153   0.6570555   1.05010395
#> lag1_var3  0.98356832 0.21979332   0.7422353   1.20845395
#> lag1_var4 -0.91236188 0.19303016  -1.1245507  -0.70438830
#> const     -0.06678818 0.36998861  -0.4301046   0.40810773
#> exo1      -0.07107587 0.08944604  -0.1748635   0.01843202
#> exo2      -0.14982611 0.08121716  -0.2216723  -0.05918204
#> exo3      -0.17363824 0.06982859  -0.2337566  -0.08002989
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.05347743 0.1644518  -0.1222949    0.2555861
#> lag1_var2  0.27522528 0.7454057  -0.5965115    0.9813631
#> lag1_var3 -0.48377359 0.8986574  -1.5369549    0.2934307
#> lag1_var4  0.44189395 0.8839871  -0.2892075    1.4746218
#> const      2.63337770 1.6616380   0.8769561    4.4416537
#> exo1      -0.07186604 0.2349611  -0.2809601    0.2317036
#> exo2       2.07993874 0.2581585   1.7722399    2.3527794
#> exo3      -0.51603531 0.2555942  -0.8415120   -0.2718612
#> 
#> $ARE$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.29672519 0.1679772   0.0674431    0.4215381
#> lag1_var2  0.49463755 1.3276043  -1.0867498    1.8299335
#> lag1_var3  1.02549682 1.6357557  -0.9651999    2.6329056
#> lag1_var4 -0.06709327 1.6210196  -1.6963889    1.8830604
#> const     -5.46862201 1.8691479  -7.6006425   -3.4750308
#> exo1      -0.42530278 0.5901367  -1.1756712    0.1711930
#> exo2      -2.62265789 0.2558347  -2.9492900   -2.4202755
#> exo3      -1.97861031 0.3540769  -2.3882820   -1.6714025
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.3185460 0.2241007  0.09088292    0.6052296
#> lag1_var2  0.6877088 0.8683843 -0.38517225    1.5669383
#> lag1_var3  0.6156128 1.1316566 -0.82666884    1.6782869
#> lag1_var4  0.3403275 1.1535405 -0.77201381    1.7822751
#> const     -5.2389678 2.7565703 -8.36503054   -2.8139596
#> exo1      -0.5009883 0.5557202 -1.07532714    0.2022653
#> exo2      -1.0462063 0.1464568 -1.17687174   -0.8507999
#> exo3      -2.4638663 0.3815036 -2.83281588   -2.0570840
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002932865 0.0005872642 0.002131533  0.003313252
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.06446837 0.01922298 -0.08811448  -0.04385589
#> Sigma[2,2]  3.33414533 0.90423207  2.64919290   4.55680198
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.05057857 0.01932293  0.03043546   0.07491718
#> Sigma[3,2] -2.41786371 0.77986622 -3.49411309  -1.88989187
#> Sigma[3,3]  1.98640146 0.71262756  1.36187990   2.92981916
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01290188 0.009989748  0.001837229   0.02382005
#> Sigma[4,2] -0.43021547 0.294721461 -0.785113813  -0.11174778
#> Sigma[4,3]  0.57206974 0.314404670  0.225039997   0.91889364
#> Sigma[4,4]  0.34715458 0.177614544  0.176386460   0.55789639
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.94059605 0.02545164  0.911359609   0.96905581
#> lag1_var2  0.04967278 0.02512190  0.029034992   0.08294014
#> lag1_var3  0.08210946 0.04161597  0.051306003   0.13754576
#> lag1_var4 -0.07385699 0.02970641 -0.111303880  -0.04509209
#> const      1.12297807 1.37048891 -0.500516075   2.58411929
#> exo1       0.01492337 0.02515397 -0.004978787   0.04922245
#> exo2      -0.12175969 0.07508200 -0.210088640  -0.04766866
#> exo3       0.07405001 0.01652050  0.054642204   0.09100513
#> 
#> $ARG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -1.9530072  1.5525712  -3.5885545   -0.1659573
#> lag1_var2  1.3125884  1.3264005  -0.3353090    2.6141053
#> lag1_var3  0.9921958  2.1485296  -1.7186274    3.0516308
#> lag1_var4 -0.3231289  1.8370656  -2.1965115    1.9250852
#> const     14.9768513 45.2983426 -46.7973408   49.7634624
#> exo1      -0.9455265  0.9610078  -1.8501483    0.2889623
#> exo2       1.7615895  2.0353353  -0.1428695    4.1326432
#> exo3       0.1561089  1.6884672  -1.9887289    1.7990832
#> 
#> $ARG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.70300538  1.1958792  -0.4869055    2.1341583
#> lag1_var2 -0.15427206  0.9070792  -1.0269373    0.9628248
#> lag1_var3  0.42805674  1.4938488  -1.0052440    2.2734023
#> lag1_var4 -0.04597721  1.2111335  -1.5027686    1.1961263
#> const     16.71120204 35.8533756 -13.7292293   65.0199133
#> exo1      -0.16009099  1.4315612  -2.1310600    0.7841765
#> exo2      -5.08759088  1.5848553  -7.0519953   -3.6286465
#> exo3       2.68521977  0.8238909   1.6666772    3.5730822
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.5756070  0.3852845  -0.9168298  -0.06516103
#> lag1_var2  0.7202704  0.1113326   0.5845089   0.83068612
#> lag1_var3  1.1466161  0.2133175   0.8771349   1.33653312
#> lag1_var4 -0.2719783  0.1056407  -0.4079657  -0.18108414
#> const     23.0198855 10.1940083  11.4572500  35.07724076
#> exo1      -0.8015389  1.0121936  -2.1660757   0.01643280
#> exo2      -4.5456125  0.8678036  -5.5498935  -3.53795197
#> exo3       3.0870244  0.5543427   2.5399599   3.74053375
#> 
#> 
#>