Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARGROUPPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARGROUPPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarGroupPANEL$new(
      data = ilo_dynamic_panel[1:5],
      exogenous = ilo_exogenous_variables[1:5],
      group_allocation = country_grouping_region[1:5]
)
#> Country groupings have been pre-specified and will not be estimated.
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004102731 0.001309932 0.003013016  0.005869442
#> 
#> $AFG$Sigma$equation2
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.008374336 0.006570413 -0.01739861 -0.003647906
#> Sigma[2,2]  0.161718759 0.038092282  0.11243057  0.197970043
#> 
#> $AFG$Sigma$equation3
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01034933 0.01079239  0.003781205   0.02515675
#> Sigma[3,2] -0.11388997 0.02884308 -0.150386086  -0.08588952
#> Sigma[3,3]  0.25212161 0.09158140  0.155384114   0.36510094
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.007221201 0.009335532  0.0008867132   0.02008803
#> Sigma[4,2] -0.048591130 0.019394504 -0.0712877076  -0.02724188
#> Sigma[4,3]  0.234739275 0.091418039  0.1341886110   0.34501465
#> Sigma[4,4]  0.250480445 0.095289551  0.1445005624   0.36472859
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1  0.962614853 0.02524741  0.939104042   0.99507782
#> lag1_var2  0.025978618 0.01404043  0.010886132   0.04327205
#> lag1_var3  0.077117288 0.06986341  0.006550898   0.16386825
#> lag1_var4 -0.066555928 0.07586398 -0.158375659   0.01417881
#> const      0.516698764 0.67469725 -0.216282254   1.32696016
#> exo1      -0.027797840 0.07538492 -0.110300301   0.06053093
#> exo2      -0.003828845 0.03931536 -0.047368002   0.04307692
#> exo3      -0.100863098 0.06304821 -0.178004263  -0.03327286
#> 
#> $AFG$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.4340687 0.1337439   0.2882694   0.58463912
#> lag1_var2 -0.5990192 0.1248586  -0.7213289  -0.45896236
#> lag1_var3 -3.3559090 0.5265257  -3.8414002  -2.66316508
#> lag1_var4  2.8738198 0.5403225   2.1516300   3.34907810
#> const     12.9481905 2.3107299   9.8297650  14.79894363
#> exo1      -0.3135283 0.7037475  -0.8962067   0.62724480
#> exo2      -0.3419504 0.2155798  -0.5914666  -0.09241765
#> exo3      -0.9618339 0.3529196  -1.2607798  -0.51419982
#> 
#> $AFG$A$equation3
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1  0.006589934  0.2363610  -0.22719130    0.2580963
#> lag1_var2  0.996198786  0.9331493  -0.17342536    1.7843598
#> lag1_var3  3.980003671  1.8921907   1.78216987    5.9369914
#> lag1_var4 -2.716943769  1.6504217  -4.48894600   -0.8676200
#> const     -9.370574378 15.6455859 -24.68173892    8.4986875
#> exo1       0.152660856  0.8260507  -0.77736116    1.1453909
#> exo2      -0.731916080  0.3726070  -1.11327796   -0.2936245
#> exo3       0.719898897  0.6154391  -0.09082169    1.2420440
#> 
#> $AFG$A$equation4
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1   0.22219473  0.2237793  -0.03205629    0.4617845
#> lag1_var2   0.81185481  1.0442274  -0.49596160    1.6528791
#> lag1_var3   2.70984696  2.0888206   0.20119816    4.7052585
#> lag1_var4  -1.50354179  1.7940732  -3.29126326    0.6189125
#> const     -11.15398535 17.9252742 -28.34176327    9.2736434
#> exo1        0.03373184  0.6387613  -0.53957002    0.8620856
#> exo2       -0.87059285  0.3667183  -1.21999534   -0.4005617
#> exo3        0.52780388  0.6418512  -0.30023537    1.1510062
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005581462 0.001626632  0.00408233  0.007733595
#> 
#> $AGO$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0117845 0.00906402 -0.02355217 -0.003398581
#> Sigma[2,2]  0.1555887 0.05020151  0.11142383  0.220974976
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.008189525 0.004953508  0.00378216   0.01472570
#> Sigma[3,2] -0.060619465 0.026878175 -0.09721180  -0.03948085
#> Sigma[3,3]  0.039100764 0.013973716  0.02732207   0.05795139
#> 
#> $AGO$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1] -0.001695059 0.003231669 -0.005999045  0.001123318
#> Sigma[4,2]  0.071253802 0.025602683  0.050874906  0.099369680
#> Sigma[4,3] -0.010862487 0.010684212 -0.025574103 -0.003493972
#> Sigma[4,4]  0.054461673 0.019283140  0.034873347  0.078909252
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.02415659 0.02443953  0.99890690    1.0548480
#> lag1_var2 -0.58055521 0.03926538 -0.62950293   -0.5397945
#> lag1_var3 -0.73568807 0.02950670 -0.77018730   -0.7047688
#> lag1_var4  0.66896352 0.04101191  0.62820689    0.7213262
#> const      4.80662869 1.30055954  3.45857223    6.3529281
#> exo1       0.05725810 0.11269644 -0.05944756    0.1933569
#> exo2       0.05730023 0.10274407 -0.05402003    0.1831220
#> exo3       0.18199736 0.06832909  0.08854673    0.2283858
#> 
#> $AGO$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.06973799 0.1714326  -0.2521572    0.1431746
#> lag1_var2  1.72751859 0.5792974   1.1972985    2.4734953
#> lag1_var3  1.02538049 0.6887690   0.3649453    1.8694109
#> lag1_var4 -0.92449362 0.6174472  -1.6874498   -0.3100844
#> const     -5.09233646 8.8062457 -12.0677651    6.8125528
#> exo1      -0.04884990 0.7128652  -0.9016555    0.6387473
#> exo2       0.08617833 0.3264040  -0.2607469    0.4136704
#> exo3      -1.12730053 0.4701613  -1.6648450   -0.5991479
#> 
#> $AGO$A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.1318098 0.05918082 -0.20140016  -0.06980786
#> lag1_var2 -0.5973678 0.23235061 -0.87802769  -0.33966497
#> lag1_var3  0.2100258 0.26415564 -0.06308441   0.53526054
#> lag1_var4  0.7257520 0.26524938  0.41270619   1.00768761
#> const      7.9424272 1.96005541  5.26520801   9.46581589
#> exo1       0.0947554 0.34701672 -0.20449085   0.55954267
#> exo2      -1.2640283 0.24488057 -1.48696423  -0.99542632
#> exo3       3.5847453 0.26154386  3.34068499   3.92433355
#> 
#> $AGO$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.26978869 0.12351750  -0.3553194  -0.09963833
#> lag1_var2  0.73728952 0.22437911   0.5247542   1.01917154
#> lag1_var3  1.05771915 0.29891034   0.7613675   1.40721871
#> lag1_var4  0.09818167 0.23652055  -0.1847318   0.34427727
#> const     -4.23640165 7.46705464 -10.7106989   5.31054583
#> exo1       0.03209586 0.29712325  -0.2632350   0.35914999
#> exo2      -1.51189512 0.06022168  -1.5818224  -1.44451318
#> exo3       3.21293285 0.30824255   2.8405516   3.54347458
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002690237 0.0007982054 0.001846353  0.003633839
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.06785315 0.01690802 -0.08990911  -0.05249922
#> Sigma[2,2]  7.51882557 2.68482548  5.56071441  11.20369883
#> 
#> $ALB$Sigma$equation3
#>                   mean        sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04925419 0.0196012  0.02654535   0.06826708
#> Sigma[3,2] -4.91933025 2.1710525 -7.57271624  -2.78716744
#> Sigma[3,3]  5.59607765 2.5231708  2.77593575   8.31276040
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] 0.003989684 0.02352539 -0.02351923   0.02982851
#> Sigma[4,2] 0.146080934 1.07183176 -1.01234107   1.38580649
#> Sigma[4,3] 2.748987172 1.38624271  1.08276958   4.04731898
#> Sigma[4,4] 3.425674803 1.16457725  2.43203531   4.95051437
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.97104213 0.01545409  0.95542491 0.9884070632
#> lag1_var2 -0.02886716 0.04761360 -0.08972871 0.0169817359
#> lag1_var3 -0.05232371 0.07168864 -0.14390752 0.0165064926
#> lag1_var4  0.04064070 0.06132034 -0.01832178 0.1189072165
#> const      1.35625239 0.60247876  0.65016678 1.9812318373
#> exo1      -0.05153630 0.03895874 -0.08470245 0.0006790531
#> exo2      -0.06084227 0.07986155 -0.15216557 0.0276297092
#> exo3       0.05075615 0.05407670 -0.01088795 0.1017726421
#> 
#> $ALB$A$equation2
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1 -2.964423903  0.4537404  -3.3503869   -2.3623517
#> lag1_var2  3.358109434  2.5756068   0.1684376    5.6190493
#> lag1_var3  4.828036328  3.8504201  -0.0479938    8.2316924
#> lag1_var4 -3.772635848  3.2734731  -6.6086851    0.3356477
#> const     13.283401027 43.1398008 -35.1757995   63.7427497
#> exo1      -0.008120943  2.6424255  -3.6202545    1.8783715
#> exo2      -1.047840743  1.8855706  -3.0802557    0.8736124
#> exo3      -0.060637460  4.4211456  -5.2209075    4.6886885
#> 
#> $ALB$A$equation3
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1   2.19087314  1.0015281   1.0465878    3.2545587
#> lag1_var2  -1.30510358  2.3994589  -3.2410848    1.6397451
#> lag1_var3  -1.67736362  3.6281531  -4.5788079    2.6734456
#> lag1_var4   1.99578916  3.1848365  -1.8226067    4.5067226
#> const     -13.88544524 18.2952414 -34.1459575    8.1073368
#> exo1       -2.16395992  2.4403366  -5.1713775    0.3458847
#> exo2       -0.05932211  0.9312027  -0.8463243    1.1737512
#> exo3       -0.36394848  2.8909198  -2.7527312    3.4480971
#> 
#> $ALB$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1   0.2897619  1.2077030  -1.0938122    1.3065396
#> lag1_var2   0.8018396  1.2637108   0.1371118    2.5109261
#> lag1_var3   1.4030902  1.9832135   0.1385497    4.0986703
#> lag1_var4  -0.2433948  1.7006882  -2.5815324    0.8064214
#> const     -15.1788427 29.8251886 -39.1567145   24.4423763
#> exo1       -2.1988519  1.8451980  -4.7422877   -0.9518430
#> exo2       -0.6990096  1.4049590  -2.4617628    0.7614941
#> exo3       -0.3966504  0.9851271  -1.4517255    0.6846979
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00442434 0.00114375 0.003291733  0.005886391
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.003892822 0.004424107 -0.007677749  0.002054758
#> Sigma[2,2]  0.061084393 0.014088190  0.043713621  0.072598694
#> 
#> $ARE$Sigma$equation3
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.01756241 0.006847939  0.01068971   0.02526145
#> Sigma[3,2] -0.05321631 0.010473071 -0.06738908  -0.04539923
#> Sigma[3,3]  0.21162992 0.056591801  0.14629121   0.27770336
#> 
#> $ARE$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01471380 0.005154123  0.008743619   0.02047110
#> Sigma[4,2] -0.00486961 0.012755086 -0.015152056   0.01125155
#> Sigma[4,3]  0.17395728 0.055841976  0.112572648   0.24124592
#> Sigma[4,4]  0.17711967 0.063256487  0.108688386   0.24996994
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.95362486 0.06860329  0.86357045   1.01490685
#> lag1_var2 -0.71065801 0.04805192 -0.76562151  -0.65493516
#> lag1_var3 -0.94826899 0.03637803 -0.99728094  -0.92022947
#> lag1_var4  0.93059001 0.03762135  0.89902655   0.98128765
#> const      2.56816401 1.09958494  1.14180540   3.52991760
#> exo1       0.01588857 0.07439729 -0.07599818   0.08979601
#> exo2      -0.10348155 0.09813486 -0.23280576  -0.01518890
#> exo3      -0.03401346 0.08637321 -0.14009138   0.05938682
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.06085369 0.2316746  -0.1903065   0.32235334
#> lag1_var2  0.66193557 0.2237183   0.4266924   0.92223016
#> lag1_var3 -0.18045323 0.2819284  -0.5310448   0.07933592
#> lag1_var4  0.13517454 0.3035916  -0.1599430   0.51910804
#> const      2.32872311 2.7966518  -0.5603291   5.39176680
#> exo1      -0.14298825 0.4610690  -0.6959671   0.33492376
#> exo2       2.12759868 0.2853000   1.8224427   2.45404938
#> exo3      -0.76904959 0.3453144  -1.1846812  -0.38823502
#> 
#> $ARE$A$equation3
#>                 mean        sd  5% quantile 95% quantile
#> lag1_var1  0.4336450 0.3208385   0.09225334    0.7667014
#> lag1_var2 -1.8635021 0.2704982  -2.14244292   -1.5359066
#> lag1_var3 -1.7100611 0.3988295  -2.09993367   -1.2496083
#> lag1_var4  2.5778161 0.4775872   1.98342082    3.0335268
#> const     -1.5602723 7.1357400 -10.03508673    4.4127770
#> exo1      -0.2954725 0.4575857  -0.82509114    0.1607141
#> exo2      -2.2072979 0.7122690  -2.93082350   -1.3264456
#> exo3      -1.2320138 0.8247287  -1.96745743   -0.1307004
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.4786090 0.2738984   0.2022406  0.809402227
#> lag1_var2 -1.4013979 0.1949486  -1.6275506 -1.210745644
#> lag1_var3 -1.8940481 0.2122383  -2.1310609 -1.667169164
#> lag1_var4  2.7585258 0.2718544   2.4532049  3.062073648
#> const     -2.1058991 6.9804592  -9.2608693  5.987058667
#> exo1      -0.3855417 0.3353647  -0.7727451  0.002259451
#> exo2      -0.5509169 0.5341292  -1.0111183  0.166718079
#> exo3      -1.8516727 0.6168772  -2.3331568 -1.013867624
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003108185 0.0007347824 0.002468374  0.003971336
#> 
#> $ARG$Sigma$equation2
#>                   mean        sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05938765 0.0282794  -0.0951294   -0.0342456
#> Sigma[2,2]  3.41992945 0.6906270   2.7954941    4.3302659
#> 
#> $ARG$Sigma$equation3
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.0484289 0.02288537   0.0266953    0.0758156
#> Sigma[3,2] -2.4786147 0.58401915  -3.1947400   -1.8503364
#> Sigma[3,3]  2.0198022 0.46391294   1.4889038    2.5423458
#> 
#> $ARG$Sigma$equation4
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01408536 0.00687009  0.007043132   0.02212303
#> Sigma[4,2] -0.44873682 0.25191131 -0.636346222  -0.12196922
#> Sigma[4,3]  0.57721929 0.15459153  0.403459347   0.72159397
#> Sigma[4,4]  0.34590976 0.05182807  0.284575720   0.39369133
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.93047799 0.04022214  0.88739702   0.97535876
#> lag1_var2 -0.13475724 0.03050536 -0.16250681  -0.10168710
#> lag1_var3 -0.22060891 0.05446775 -0.27345571  -0.15939838
#> lag1_var4  0.19156799 0.04215162  0.14443007   0.23769168
#> const      3.65531559 1.39535384  2.37316652   5.25216440
#> exo1       0.04079954 0.06868880 -0.02505184   0.11957715
#> exo2      -0.09209930 0.06770105 -0.16566764  -0.02262767
#> exo3       0.02537990 0.07481729 -0.04554597   0.11745138
#> 
#> $ARG$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -2.6823558  1.186927  -4.2776795   -1.7035381
#> lag1_var2  1.2848093  1.610401  -0.4456495    3.0728041
#> lag1_var3  1.0224742  2.722363  -2.0343217    3.8895008
#> lag1_var4 -0.3761169  2.457017  -3.1515441    2.1985113
#> const     36.4943065 38.865643  -5.8949815   76.4290410
#> exo1      -1.5029807  1.699042  -3.5919040    0.3213441
#> exo2       1.8413397  1.143228   0.5151864    3.0896123
#> exo3       2.4055670  2.264700  -0.3257430    4.9088464
#> 
#> $ARG$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  1.4684942  1.078805   0.4735588    2.8770522
#> lag1_var2 -0.8668593  1.359339  -2.4296550    0.5305218
#> lag1_var3 -0.7993345  2.299101  -3.3074494    1.7116267
#> lag1_var4  1.1431872  2.032643  -0.8886063    3.5099122
#> const     -2.0546305 34.788549 -35.7721312   34.2894537
#> exo1       0.4074729  1.168818  -0.8407313    1.8548853
#> exo2      -4.8023951  1.158771  -6.0485807   -3.5137017
#> exo3       1.6821499  1.757072  -0.2737487    3.5654775
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.1826310  0.4276655  -0.6385483    0.3478441
#> lag1_var2 -0.1005855  0.4255248  -0.6200547    0.3306541
#> lag1_var3 -0.2092821  0.7131841  -1.0429023    0.4932540
#> lag1_var4  1.0375901  0.6004547   0.4213889    1.7740669
#> const     15.0887004 13.7077287   4.4996421   32.9863021
#> exo1      -0.5829685  0.1968560  -0.7966395   -0.3518653
#> exo2      -4.2448911  0.7814112  -5.1007957   -3.2888530
#> exo3       3.5610067  1.0456480   2.5773664    4.9353031
#> 
#> 
#>