Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARGROUPPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARGROUPPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarGroupPANEL$new(
      data = ilo_dynamic_panel[1:5],
      exogenous = ilo_exogenous_variables[1:5],
      group_allocation = country_grouping_region[1:5]
)
#> Country groupings have been pre-specified and will not be estimated.
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004514197 0.001211388 0.003096646  0.005912799
#> 
#> $AFG$Sigma$equation2
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.008474015 0.005225405 -0.01440425 -0.002306097
#> Sigma[2,2]  0.190198192 0.031661165  0.14691165  0.213927575
#> 
#> $AFG$Sigma$equation3
#>                   mean         sd   5% quantile 95% quantile
#> Sigma[3,1]  0.01399906 0.01078355  0.0008317442   0.02565587
#> Sigma[3,2] -0.15719839 0.06151024 -0.2288985286  -0.08559100
#> Sigma[3,3]  0.30301613 0.12141512  0.1913587630   0.46126769
#> 
#> $AFG$Sigma$equation4
#>                   mean         sd   5% quantile 95% quantile
#> Sigma[4,1]  0.01259471 0.01053397 -0.0004691174   0.02377217
#> Sigma[4,2] -0.08415405 0.06551135 -0.1659223954  -0.01610788
#> Sigma[4,3]  0.27926614 0.11623728  0.1826118080   0.43593922
#> Sigma[4,4]  0.29748608 0.11209990  0.2074510788   0.44733260
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.96932016 0.02352526  0.95379109   1.00169845
#> lag1_var2  0.06377644 0.08197861 -0.02307204   0.16437886
#> lag1_var3  0.15989470 0.18637415 -0.03549073   0.38762116
#> lag1_var4 -0.14085673 0.16657118 -0.34702943   0.03181252
#> const     -0.03605173 0.76383065 -0.82653120   0.93135553
#> exo1      -0.03469108 0.02462551 -0.06710310  -0.01181275
#> exo2      -0.03688994 0.07256787 -0.09877857   0.06059341
#> exo3      -0.16195116 0.07966327 -0.24246682  -0.06032404
#> 
#> $AFG$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.3013772 0.1221340   0.1776897    0.4466020
#> lag1_var2  0.1959081 0.9819266  -1.0754571    1.1557033
#> lag1_var3 -1.5312374 2.2583100  -4.3795526    0.7824013
#> lag1_var4  1.2617277 2.0356055  -0.8244889    3.8121808
#> const      6.6208652 9.6675845  -4.0934356   18.1407217
#> exo1      -0.2178644 0.5902024  -0.8565723    0.4988122
#> exo2      -0.1107592 0.3213979  -0.5170091    0.2188091
#> exo3      -0.2775630 0.5568061  -0.8432822    0.4329822
#> 
#> $AFG$A$equation3
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  -0.15605779  0.3012381  -0.5181849    0.1604204
#> lag1_var2   2.20084234  1.6263899   1.0154504    4.4341912
#> lag1_var3   6.38290137  3.5384054   3.3613399   11.1491588
#> lag1_var4  -4.81792216  3.1010636  -8.9637215   -2.0723869
#> const     -20.52065366 16.3861258 -42.4983009   -6.8762356
#> exo1        0.05094388  0.7618521  -0.8722518    0.7647346
#> exo2       -1.20909470  0.5109162  -1.6476607   -0.5325919
#> exo3        0.46698377  0.8700164  -0.4329028    1.5699636
#> 
#> $AFG$A$equation4
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  -0.03421936  0.3217999  -0.4120908    0.2930814
#> lag1_var2   2.28276265  1.5491089   1.1927614    4.4086237
#> lag1_var3   5.80964489  3.3220883   3.3020220   10.3105304
#> lag1_var4  -4.26190783  2.9001044  -8.1629720   -1.9606048
#> const     -21.47767974 14.9141704 -42.0006059  -11.0299501
#> exo1       -0.03791454  0.6079399  -0.7569310    0.4409251
#> exo2       -1.38141818  0.6946522  -1.9371424   -0.4542304
#> exo3        0.42045246  0.8461155  -0.4167599    1.4642163
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                  mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.02438794 0.005626363  0.01791319   0.03089237
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.02733835 0.007511055 -0.03549704  -0.01941494
#> Sigma[2,2]  0.14290226 0.046347804  0.10624549   0.20607946
#> 
#> $AGO$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.06575909 0.01660858   0.0462836   0.08480932
#> Sigma[3,2] -0.10519924 0.02873855  -0.1387622  -0.07242989
#> Sigma[3,3]  0.21380242 0.04571904   0.1599255   0.26341436
#> 
#> $AGO$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] 0.048656899 0.01159821  0.03466454   0.06096884
#> Sigma[4,2] 0.008454394 0.02206897 -0.02094405   0.02604300
#> Sigma[4,3] 0.141883083 0.03188213  0.10852250   0.17947256
#> Sigma[4,4] 0.164883964 0.02065763  0.13940610   0.18541606
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.8396651 0.04428469   0.7822850   0.88082801
#> lag1_var2  0.6567751 0.18209699   0.4534091   0.86207698
#> lag1_var3  0.7404134 0.22232907   0.5308170   1.02851415
#> lag1_var4 -0.7550857 0.20490437  -1.0189945  -0.56974121
#> const      3.7039424 2.35450357   0.8426078   6.31870211
#> exo1       0.0986532 0.15772998  -0.1175542   0.21716209
#> exo2      -0.2277011 0.20226300  -0.4921705  -0.04595733
#> exo3      -0.3744312 0.19719283  -0.5909958  -0.15252199
#> 
#> $AGO$A$equation2
#>                   mean        sd  5% quantile 95% quantile
#> lag1_var1  0.129108400 0.1186716 -0.005943414    0.2316953
#> lag1_var2 -0.008157508 0.2807475 -0.317301886    0.2682727
#> lag1_var3 -1.161634249 0.3462134 -1.559585068   -0.7798654
#> lag1_var4  1.167174272 0.3096597  0.836904338    1.5186344
#> const     -1.781550634 4.5255809 -5.059380857    4.3368875
#> exo1      -0.220455842 0.4441036 -0.595318405    0.3692553
#> exo2       0.514315249 0.2233689  0.280047976    0.7684867
#> exo3       0.103811333 0.4203106 -0.311812658    0.6453205
#> 
#> $AGO$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.7008262 0.1336989  -0.8543989   -0.5606020
#> lag1_var2  2.8970936 0.5297046   2.3722146    3.5352466
#> lag1_var3  4.3764256 0.6512857   3.7176032    5.2067261
#> lag1_var4 -3.3083429 0.5928318  -4.0883391   -2.7652989
#> const      7.0936073 6.7525520  -0.1916047   14.0837949
#> exo1       0.3044273 0.3655757  -0.1080150    0.7391559
#> exo2      -2.0538942 0.3848885  -2.4828701   -1.5976073
#> exo3       1.8762924 0.5376893   1.2757993    2.3984912
#> 
#> $AGO$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.7427596 0.1529413  -0.8745778   -0.5450600
#> lag1_var2  2.6249369 0.6562144   1.8382280    3.2485349
#> lag1_var3  3.1444967 0.7813547   2.1612972    3.9295836
#> lag1_var4 -2.0287523 0.7130892  -2.8027756   -1.1680191
#> const      5.9352798 7.6924864  -3.9834079   12.9040559
#> exo1       0.1232145 0.4604985  -0.5099318    0.4495792
#> exo2      -1.9724504 0.3069879  -2.3277842   -1.6241543
#> exo3       2.3650845 0.4330527   1.7806265    2.6846536
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002299282 0.0009724701 0.001440099  0.003527759
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05820051 0.04795644  -0.1234617  -0.03343779
#> Sigma[2,2]  7.17413543 3.21605525   4.3626772  11.40173296
#> 
#> $ALB$Sigma$equation3
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.008269728 0.04607925 -0.04460036   0.06283154
#> Sigma[3,2] -4.340186898 2.38075560 -7.27824308  -1.91169967
#> Sigma[3,3]  5.289354869 1.99003594  3.83293194   7.90093214
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.03552375 0.03221577 -0.07802718 -0.005760738
#> Sigma[4,2]  0.55786194 0.69787228 -0.18994830  1.420603072
#> Sigma[4,3]  2.73782425 1.59329437  1.30046125  4.802865418
#> Sigma[4,4]  3.61517773 1.64204638  1.99096907  5.625735141
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                   mean          sd   5% quantile  95% quantile
#> lag1_var1  1.007496869 0.009960283  0.9965059299  1.0179219956
#> lag1_var2  0.020708590 0.018803109  0.0034826442  0.0453196492
#> lag1_var3  0.021741159 0.025615173 -0.0006129721  0.0555055451
#> lag1_var4 -0.018667292 0.021094692 -0.0464257737 -0.0009900906
#> const     -0.440218099 0.549904323 -1.1131260669  0.1515161496
#> exo1      -0.001984658 0.047281520 -0.0606610996  0.0439591234
#> exo2      -0.039453548 0.096092673 -0.1388207774  0.0820793945
#> exo3       0.077977959 0.011781669  0.0643441167  0.0917640882
#> 
#> $ALB$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -3.4229796  0.7729462  -4.2227493    -2.526024
#> lag1_var2  3.4372491  1.4810674   1.9208829     5.120674
#> lag1_var3  4.7859195  2.4393352   2.2604425     7.436121
#> lag1_var4 -3.9144912  1.9860396  -6.0737291    -1.849762
#> const     33.0256760 26.7405974  -0.3570131    59.922729
#> exo1      -0.8872356  3.6343124  -3.1019888     4.111792
#> exo2       0.8666582  3.9980998  -4.1428286     4.965040
#> exo3      -0.2572260  2.1727192  -3.1454053     1.647820
#> 
#> $ALB$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  2.6375463  1.164239    1.116940   3.74732925
#> lag1_var2 -4.0535276  2.446599   -6.367413  -1.13083069
#> lag1_var3 -5.7479368  3.801047   -9.279826  -1.23719754
#> lag1_var4  5.5040277  3.103367    1.780811   8.42241177
#> const     14.1689686 56.291031  -40.946370  82.55481458
#> exo1       0.1584827  3.397812   -3.465622   4.38064624
#> exo2      -2.4842766  1.820720   -3.991289  -0.02551466
#> exo3      -1.7872271  1.913105   -4.245451   0.10524428
#> 
#> $ALB$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.27583617  0.9804699  -0.9871153    1.2127105
#> lag1_var2 -2.49658277  1.9340149  -4.6984188   -0.3480471
#> lag1_var3 -3.60695704  2.9199846  -6.9508742   -0.4281975
#> lag1_var4  3.92204305  2.3951848   1.2843519    6.6937559
#> const     39.92603007 46.0516994  -2.4064982   97.9527242
#> exo1      -0.07988231  3.1132794  -2.6637941    4.0807275
#> exo2      -2.11333144  2.2389652  -3.9580694    0.8622331
#> exo3      -2.27567464  2.2684059  -4.8002944    0.3412177
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001546016 0.0003657259 0.001147219  0.001985904
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.003268546 0.002565564 -0.005899824 -0.000180404
#> Sigma[2,2]  0.056329400 0.015968617  0.035200155  0.070687430
#> 
#> $ARE$Sigma$equation3
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.01046426 0.007726712  0.00457831   0.02100676
#> Sigma[3,2] -0.06473236 0.040494533 -0.10703643  -0.02200214
#> Sigma[3,3]  0.26634306 0.135917498  0.13686483   0.42971638
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.008494645 0.006051502  0.003521528  0.016642168
#> Sigma[4,2] -0.022846079 0.032703066 -0.059669037  0.007416392
#> Sigma[4,3]  0.223916007 0.111482402  0.107626450  0.353794140
#> Sigma[4,4]  0.215159016 0.093302760  0.114205589  0.318570594
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.95912648 0.03268688  0.92758281   0.99482707
#> lag1_var2  0.36495894 0.23561926  0.18109303   0.62563487
#> lag1_var3  0.44363430 0.30836571  0.20336646   0.78216233
#> lag1_var4 -0.43496009 0.31183189 -0.77689949  -0.19559568
#> const      0.39131519 0.56119852 -0.27715863   0.82028675
#> exo1      -0.01709499 0.05264626 -0.07342033   0.04429170
#> exo2      -0.10730437 0.04091412 -0.15668155  -0.06463884
#> exo3       0.01553657 0.06820284 -0.05279034   0.09065971
#> 
#> $ARE$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.1676239 0.3994273 -0.15237164    0.7049583
#> lag1_var2 -0.4032692 1.0248685 -1.69549071    0.5045733
#> lag1_var3 -1.4484818 1.2084199 -2.90870020   -0.4069527
#> lag1_var4  1.3540556 1.1545624  0.40519758    2.7093970
#> const      3.4872206 3.5688362 -0.94707887    6.8370255
#> exo1       0.3467519 0.2975885  0.07039422    0.6803309
#> exo2       1.9834002 0.2846050  1.68236995    2.3371946
#> exo3      -0.6060985 0.2549300 -0.94772342   -0.3983089
#> 
#> $ARE$A$equation3
#>                   mean        sd 5% quantile 95% quantile
#> lag1_var1   0.02833195 0.5774790  -0.7046882    0.5842621
#> lag1_var2   5.27479953 3.4304249   2.5473562    9.4274940
#> lag1_var3   7.59064438 4.5085200   4.0855371   13.0462568
#> lag1_var4  -6.46119359 4.4771952 -11.8434257   -3.0500007
#> const     -11.01647496 8.3924659 -19.6211111   -1.0178404
#> exo1       -1.10516973 0.7952946  -1.9053385   -0.1250759
#> exo2       -2.24412743 0.4847572  -2.8207096   -1.7710657
#> exo3       -0.52311370 0.5745569  -0.8985460    0.2713204
#> 
#> $ARE$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.08782798 0.5608537  -0.6780138   0.51431902
#> lag1_var2  5.07484501 2.7493068   2.7155322   8.35591663
#> lag1_var3  6.62132954 3.6867603   3.7119312  11.11246410
#> lag1_var4 -5.53014925 3.6835587 -10.0213026  -2.74007151
#> const     -9.13605298 7.6262975 -16.5439426  -0.04553513
#> exo1      -0.81920588 0.6462206  -1.3863884   0.02476154
#> exo2      -0.70370475 0.5430112  -1.2146758  -0.02922014
#> exo3      -0.97608971 0.6138117  -1.4801909  -0.14704401
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002550206 0.0006496251 0.001887717  0.003380053
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05002908 0.03287485 -0.09454466  -0.03058215
#> Sigma[2,2]  3.03451009 1.17664790  2.19640236   4.61655501
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.03945196 0.02249791  0.02456856   0.07012006
#> Sigma[3,2] -2.22393652 0.75326470 -3.24784974  -1.74056300
#> Sigma[3,3]  1.87028807 0.44673977  1.49748186   2.47132247
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01098894 0.004605351  0.005165905   0.01556799
#> Sigma[4,2] -0.44558086 0.114337053 -0.551867154  -0.29616413
#> Sigma[4,3]  0.59374712 0.127496562  0.446322455   0.72778834
#> Sigma[4,4]  0.36596343 0.098356485  0.264476936   0.47381347
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.914119614 0.02172507  0.89298954   0.94184432
#> lag1_var2  0.237483628 0.05312968  0.17004735   0.28950092
#> lag1_var3  0.393049994 0.08853845  0.28192853   0.48265339
#> lag1_var4 -0.363774160 0.07956921 -0.43964106  -0.26022408
#> const      0.574353380 0.46872327  0.13781507   1.10948368
#> exo1       0.026001589 0.05587881 -0.04700783   0.07922821
#> exo2      -0.141596569 0.05963877 -0.19254250  -0.07346173
#> exo3      -0.004681989 0.02927165 -0.04474142   0.01615998
#> 
#> $ARG$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -1.40507338  1.144941   -2.803928   -0.1538222
#> lag1_var2  0.66370070  2.011503   -0.725562    3.4312823
#> lag1_var3 -0.08968873  3.367283   -2.476935    4.5344406
#> lag1_var4  0.60709210  2.913943   -3.413374    2.6145725
#> const      9.39189375 21.510018  -11.253125   33.7594293
#> exo1      -1.17698563  1.106117   -2.263620    0.2499412
#> exo2       1.30728306  1.538290   -0.281036    3.2540314
#> exo3       0.13620795  1.661775   -2.059122    1.4962711
#> 
#> $ARG$A$equation3
#>                 mean         sd  5% quantile 95% quantile
#> lag1_var1  0.5223147  0.8454679  -0.24206488    1.6363656
#> lag1_var2  2.0816668  1.6000834  -0.09829815    3.1646791
#> lag1_var3  4.0982221  2.7112172   0.41198844    5.9759228
#> lag1_var4 -3.3420938  2.3220936  -4.88491142   -0.1396931
#> const     -1.1689847 14.1350725 -15.64810193   16.3247886
#> exo1       0.1280512  0.9125937  -1.08920168    0.9292198
#> exo2      -4.9076864  1.3947126  -6.67545339   -3.5400082
#> exo3       2.2752610  1.8693561   0.55624235    4.7120218
#> 
#> $ARG$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.4699028 0.2930270   -0.769910  -0.10809977
#> lag1_var2  2.4847422 0.5684177    1.719360   2.93902670
#> lag1_var3  4.0390533 0.9724522    2.741087   4.87897503
#> lag1_var4 -2.9152028 0.8083797   -3.628550  -1.83835413
#> const      5.0278746 9.0277690   -6.042420  14.73392702
#> exo1      -0.6850221 0.4860700   -1.151697  -0.08378468
#> exo2      -4.6965962 0.5288929   -5.321242  -4.23741195
#> exo3       2.5451805 0.9682060    1.605330   3.77244069
#> 
#> 
#>