Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARGROUPPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARGROUPPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarGroupPANEL$new(
      data = ilo_dynamic_panel[1:5],
      exogenous = ilo_exogenous_variables[1:5],
      group_allocation = country_grouping_region[1:5]
)
#> Country groupings have been pre-specified and will not be estimated.
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#>  bpvars: Forecasting with Bayesian Panel VARs     |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004818668 0.001201781 0.003337909  0.005807678
#> 
#> $AFG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01165403 0.005240647 -0.01798289   -0.0060350
#> Sigma[2,2]  0.19516962 0.049547014  0.14143387    0.2519837
#> 
#> $AFG$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01063434 0.006639958  0.003761229   0.01763772
#> Sigma[3,2] -0.15116159 0.050109266 -0.189771732  -0.08670635
#> Sigma[3,3]  0.26081818 0.099996433  0.150809000   0.38081378
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.006528801 0.007611268 -0.001910033   0.01308494
#> Sigma[4,2] -0.074577298 0.035403126 -0.108297868  -0.02965701
#> Sigma[4,3]  0.223863379 0.095176555  0.127256616   0.34361053
#> Sigma[4,4]  0.221233104 0.094930670  0.129760581   0.34465988
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1  1.01099232 0.009563355   0.9989314   1.01961255
#> lag1_var2 -0.05695538 0.121858034  -0.1612842   0.10265879
#> lag1_var3 -0.06909474 0.260921402  -0.2918691   0.27562744
#> lag1_var4  0.05128105 0.228141497  -0.2518962   0.24584476
#> const      0.84606297 1.478418432  -0.9149449   2.42011825
#> exo1      -0.04896908 0.100644135  -0.1532417   0.06348588
#> exo2      -0.05227009 0.089178554  -0.1122607   0.07078137
#> exo3      -0.21092093 0.110959100  -0.3144306  -0.07801191
#> 
#> $AFG$A$equation2
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.243780476  0.2574968 -0.09453227    0.4465664
#> lag1_var2  0.004734126  0.9686957 -1.27552363    0.7892333
#> lag1_var3 -2.255796973  1.9632620 -4.91663612   -0.7997524
#> lag1_var4  1.959428678  1.6910203  0.71319225    4.2581848
#> const      7.866749180 10.4096967 -2.90380328   20.9883658
#> exo1      -0.027037582  0.5345234 -0.63308614    0.6102636
#> exo2      -0.190077654  0.2830486 -0.48641806    0.1491050
#> exo3      -0.217673614  0.9798061 -1.06125964    1.0979658
#> 
#> $AFG$A$equation3
#>                    mean          sd  5% quantile 95% quantile
#> lag1_var1  5.260951e-04  0.05597696  -0.07468683   0.04578961
#> lag1_var2  2.862032e+00  1.30873009   1.48832259   4.23582478
#> lag1_var3  8.039963e+00  2.58076685   5.23751818  10.74351034
#> lag1_var4 -6.249653e+00  2.16883799  -8.49864604  -3.85647619
#> const     -3.349182e+01 20.05605876 -55.91404838 -12.73785805
#> exo1      -1.948861e-01  0.33284290  -0.51201591   0.23982370
#> exo2      -8.052162e-01  0.54574054  -1.49280180  -0.23335354
#> exo3       3.753381e-01  0.45145025  -0.16472216   0.85726594
#> 
#> $AFG$A$equation4
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1   0.1263471  0.1256651  -0.01238789    0.2566486
#> lag1_var2   3.2043249  1.0707219   1.96478722    4.3679009
#> lag1_var3   7.8411332  2.1454174   5.24032199   10.1012475
#> lag1_var4  -5.9586582  1.8168272  -7.83938089   -3.7057378
#> const     -40.3831193 18.3655519 -59.59209857  -21.0706125
#> exo1       -0.2262097  0.4421558  -0.64427922    0.3045235
#> exo2       -0.9214930  0.5632121  -1.57224408   -0.2774162
#> exo3        0.4310411  0.2249214   0.15971836    0.6490357
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005257887 0.0009096135 0.004210036  0.006322198
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01357717 0.001386854 -0.01544376  -0.01260534
#> Sigma[2,2]  0.23115232 0.083996795  0.13809415   0.31321344
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.009793616 0.001642778  0.007810288   0.01141113
#> Sigma[3,2] -0.092299284 0.038450114 -0.133267176  -0.05067991
#> Sigma[3,3]  0.052978007 0.016236481  0.032910145   0.06906256
#> 
#> $AGO$Sigma$equation4
#>                     mean          sd 5% quantile 95% quantile
#> Sigma[4,1] -0.0007736433 0.001788335 -0.00238264  0.001403954
#> Sigma[4,2]  0.1012271095 0.032594670  0.06404910  0.132869548
#> Sigma[4,3] -0.0209378612 0.018587041 -0.03972575  0.002028300
#> Sigma[4,4]  0.0683470150 0.014966010  0.04940745  0.083589276
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.01869630 0.04093172  0.96486683    1.0567274
#> lag1_var2 -0.37688058 0.22780679 -0.65251875   -0.1434659
#> lag1_var3 -0.46434258 0.27755590 -0.80734354   -0.1950769
#> lag1_var4  0.44426261 0.24316463  0.20349218    0.7399940
#> const      1.44030380 2.93910681 -1.25062963    5.3259741
#> exo1       0.03912865 0.07165439 -0.05462256    0.1072376
#> exo2       0.06607052 0.07877364 -0.01974885    0.1621529
#> exo3       0.17664362 0.07247718  0.10488969    0.2640824
#> 
#> $AGO$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  -0.2975425  0.2482675  -0.5358354    0.0190943
#> lag1_var2   3.8636974  2.1549397   1.2812864    6.1388668
#> lag1_var3   4.0368034  3.0397632   0.3332689    7.1650279
#> lag1_var4  -3.5647412  2.7316926  -6.4112013   -0.2869505
#> const     -25.0793533 15.8181875 -40.0971198   -4.5213489
#> exo1       -0.1104594  0.4120113  -0.4403099    0.3883354
#> exo2       -0.2501398  0.4007822  -0.7796515    0.1185145
#> exo3       -1.0145471  1.0794839  -2.2410672    0.1659946
#> 
#> $AGO$A$equation3
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1 -0.08877448  0.05188039  -0.1381501  -0.02094458
#> lag1_var2 -1.40791320  0.82929463  -2.1759306  -0.37930086
#> lag1_var3 -0.94174187  1.19863178  -2.0390000   0.54761798
#> lag1_var4  1.70396040  1.03284557   0.4314931   2.68126073
#> const     19.01320048 10.64845596   5.4393867  27.35772923
#> exo1       0.12236608  0.23431694  -0.1363676   0.32371885
#> exo2      -1.17894213  0.14104879  -1.3171532  -1.02301868
#> exo3       3.48797796  0.35274246   3.0461568   3.81864637
#> 
#> $AGO$A$equation4
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1 -0.40485815 0.1613526  -0.54067113   -0.1981077
#> lag1_var2  1.39847917 0.9938575   0.33500847    2.5857608
#> lag1_var3  1.98685076 1.3740195   0.45999754    3.6146769
#> lag1_var4 -0.76281584 1.2705493  -2.24258894    0.6291214
#> const     -5.24785113 8.0480103 -14.08940871    3.0220238
#> exo1       0.04560949 0.1103493  -0.05961268    0.1864371
#> exo2      -1.67684606 0.2390671  -1.99602973   -1.4781221
#> exo3       3.16586186 0.6520559   2.44179127    3.9098519
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002321704 0.0006999951 0.001448651  0.002987598
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.03411657 0.03835826 -0.07648749  0.004877288
#> Sigma[2,2]  6.51561575 0.96114132  5.44324779  7.574610852
#> 
#> $ALB$Sigma$equation3
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[3,1] -0.002749526 0.02528897 -0.02935031    0.0289117
#> Sigma[3,2] -4.162180476 1.27085115 -5.77750670   -2.8854038
#> Sigma[3,3]  4.635278150 1.84797802  3.36879616    7.1748371
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.03100593 0.01559084 -0.05174788  -0.01689506
#> Sigma[4,2]  0.22671572 1.19192061 -1.09033446   1.64217969
#> Sigma[4,3]  2.22774235 1.25370256  1.33724410   3.94309540
#> Sigma[4,4]  2.86977005 1.01203031  1.94472456   4.10084838
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  1.03582931 0.03569337  0.991679449  1.074457176
#> lag1_var2 -0.03952559 0.04403953 -0.084926184  0.006396260
#> lag1_var3 -0.07269907 0.06492381 -0.138104114 -0.004556398
#> lag1_var4  0.06222229 0.05126983  0.007893640  0.114553349
#> const     -0.22131551 1.57603477 -2.038103068  1.563058952
#> exo1       0.06784996 0.04349781  0.009051681  0.098386452
#> exo2      -0.10067123 0.04571868 -0.154693916 -0.052233368
#> exo3       0.03197527 0.05652140 -0.020986360  0.105991127
#> 
#> $ALB$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  -5.307128  1.866189  -7.4206445   -3.1671700
#> lag1_var2   1.726696  1.915902  -0.2978297    3.6183320
#> lag1_var3   2.247084  2.890868  -0.8578390    5.0479183
#> lag1_var4  -1.950145  2.443431  -4.3081725    0.7682614
#> const     113.647348 55.066831  42.5518128  167.6745448
#> exo1       -3.768527  2.342177  -6.8578886   -1.5799684
#> exo2       -0.163923  2.869935  -3.2620285    3.3213237
#> exo3       -1.150575  2.532664  -3.7284724    1.9306486
#> 
#> $ALB$A$equation3
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1   3.2246248  2.827150   -0.6292204     5.398033
#> lag1_var2  -0.9592541  2.172165   -2.8316419     1.892772
#> lag1_var3  -1.0733171  3.302855   -3.9617322     3.317643
#> lag1_var4   1.6507476  2.756609   -2.0237045     4.075443
#> const     -53.2363342 80.399298 -102.8289397    55.881067
#> exo1       -0.7447745  2.987180   -4.5552186     2.050428
#> exo2       -1.0664117  1.720643   -2.2571810     1.246094
#> exo3        2.4575507  2.511783   -0.3397647     5.058885
#> 
#> $ALB$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.29228759  2.063357  -3.0959012   1.33651624
#> lag1_var2 -0.13762403  1.381320  -1.1494764   1.74009324
#> lag1_var3 -0.01788263  2.191528  -1.7147988   2.94318826
#> lag1_var4  0.87006038  1.788488  -1.5473149   2.30877150
#> const     17.40855655 55.774266 -25.3465177  93.42078559
#> exo1      -3.44967072  3.427917  -7.9823989  -0.44688102
#> exo2      -1.40360004  1.195862  -2.7485592   0.01298431
#> exo3       2.12809369  2.359127  -0.4238415   4.45386254
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001262667 0.0002810567 0.001029039  0.001637411
#> 
#> $ARE$Sigma$equation2
#>                    mean           sd  5% quantile  95% quantile
#> Sigma[2,1] -0.001796887 0.0009876446 -0.002965085 -0.0007209039
#> Sigma[2,2]  0.060035413 0.0236467040  0.031652045  0.0844526207
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.006743463 0.001782165  0.004626064  0.008213118
#> Sigma[3,2] -0.058123144 0.022418500 -0.083274935 -0.031141591
#> Sigma[3,3]  0.188250609 0.013344916  0.171726019  0.202365465
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.005409697 0.001233954  0.004106875  0.006642466
#> Sigma[4,2] -0.012075735 0.007642443 -0.021085165 -0.004465852
#> Sigma[4,3]  0.145631276 0.014531609  0.132045415  0.164800454
#> Sigma[4,4]  0.138573465 0.019388207  0.119519884  0.162840475
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.968813483 0.02669573  0.93750453   0.99561826
#> lag1_var2 -0.250119909 0.11121566 -0.37265345  -0.14766312
#> lag1_var3 -0.341778880 0.15700933 -0.51645430  -0.19727857
#> lag1_var4  0.337610758 0.15156477  0.20096709   0.50943195
#> const      1.154803548 0.41187021  0.70191886   1.59337340
#> exo1      -0.009100811 0.01603743 -0.02206271   0.01208177
#> exo2      -0.086897014 0.04047921 -0.11333552  -0.03148383
#> exo3      -0.027768255 0.03499345 -0.06283831   0.01627708
#> 
#> $ARE$A$equation2
#>                 mean         sd  5% quantile 95% quantile
#> lag1_var1  0.1355573 0.25611106 -0.154372058  0.422981166
#> lag1_var2 -1.3722799 1.41610148 -2.575579596  0.166534529
#> lag1_var3 -2.6344844 1.80578904 -4.200421819 -0.626735490
#> lag1_var4  2.5234122 1.75682863  0.584514499  4.035934594
#> const      5.7280212 3.00426831  2.532933390  9.290346839
#> exo1       0.1914594 0.15153609 -0.006182296  0.324956391
#> exo2       1.9985242 0.07980937  1.924412988  2.101604983
#> exo3      -0.2360091 0.32990254 -0.685900279 -0.003964011
#> 
#> $ARE$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1   0.1882604 0.4962291  -0.2885832   0.79492186
#> lag1_var2   3.1534244 2.6978409   0.2084799   5.43287454
#> lag1_var3   4.7972189 3.6874373   0.7419298   7.79940772
#> lag1_var4  -3.7038207 3.6591332  -6.5549612   0.33967942
#> const     -12.3926796 7.5650209 -19.0758649  -3.91859569
#> exo1       -0.6396863 0.5537188  -1.3081019  -0.03862947
#> exo2       -2.6064760 0.7246537  -3.2936402  -1.66821330
#> exo3       -1.6968033 0.8106203  -2.3479572  -0.66092973
#> 
#> $ARE$A$equation4
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1   0.2866361 0.4952770  -0.04567077   0.96603544
#> lag1_var2   2.1515561 1.6702803   0.34358875   3.61650622
#> lag1_var3   2.8351211 2.3822925   0.17646553   4.83428610
#> lag1_var4  -1.7985803 2.3789112  -3.67890543   0.86230548
#> const     -10.0329937 6.8260424 -16.99274972  -2.23153927
#> exo1       -0.5031599 0.4912185  -1.13559658  -0.03634366
#> exo2       -1.0566571 0.7540374  -1.80609691  -0.10390356
#> exo3       -1.9436720 0.6056571  -2.51943328  -1.25550366
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002883693 0.001524759 0.002044714   0.00495922
#> 
#> $ARG$Sigma$equation2
#>                   mean        sd 5% quantile 95% quantile
#> Sigma[2,1] -0.04598678 0.0190793 -0.06887215  -0.02466518
#> Sigma[2,2]  2.46298365 0.4943263  1.84604359   2.96609764
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.03515654 0.01382979  0.02297684   0.05304357
#> Sigma[3,2] -1.70482603 0.25315264 -1.93032299  -1.37307861
#> Sigma[3,3]  1.39267328 0.17047258  1.17004796   1.53214730
#> 
#> $ARG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.007831068 0.004171511  0.002782024   0.01218921
#> Sigma[4,2] -0.225378485 0.129017826 -0.341605380  -0.07797167
#> Sigma[4,3]  0.390406760 0.134089091  0.220539919   0.52643632
#> Sigma[4,4]  0.280477380 0.087182793  0.173118566   0.37362668
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.932018775 0.05239918  0.87216766   0.99149392
#> lag1_var2 -0.162880325 0.21019426 -0.42369114   0.06184623
#> lag1_var3 -0.260937290 0.34469630 -0.68751876   0.10839719
#> lag1_var4  0.238848662 0.31514643 -0.09731612   0.63061295
#> const      3.227950077 1.25629531  1.74779382   4.62292938
#> exo1      -0.002910133 0.07404046 -0.09373913   0.06208348
#> exo2      -0.081191799 0.04445171 -0.12828491  -0.03111903
#> exo3       0.163867892 0.08056460  0.08847250   0.26779367
#> 
#> $ARG$A$equation2
#>                   mean        sd 5% quantile 95% quantile
#> lag1_var1  -2.83560690  1.965302   -5.423778   -0.9910468
#> lag1_var2  -6.58007250  3.816533  -11.372667   -2.6273856
#> lag1_var3 -11.91301926  6.200585  -19.706686   -5.4665179
#> lag1_var4  10.98242403  5.743792    5.164351   18.2598849
#> const     136.14044421 49.826632   81.811989  197.2314826
#> exo1       -0.63242294  1.060769   -1.914143    0.2225595
#> exo2        1.40570378  2.259833   -1.008466    3.5832862
#> exo3       -0.01559956  3.071524   -3.877656    2.8363850
#> 
#> $ARG$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1   0.9644394  1.303384  -0.5650134     2.463707
#> lag1_var2   4.3806139  3.153242   1.5021651     8.494664
#> lag1_var3   7.9042307  5.115147   3.1843879    14.586704
#> lag1_var4  -6.5182285  4.793401 -12.7739655    -2.192494
#> const     -51.1569567 26.549100 -87.3750817   -30.287563
#> exo1       -0.2017756  1.295046  -1.6572554     1.350781
#> exo2       -4.6173641  2.073725  -6.7379131    -2.370468
#> exo3        3.0207112  2.440222   0.6203804     6.057506
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.8601193  0.4817742   -1.375790   -0.2625092
#> lag1_var2  0.2687220  1.1518157   -1.124355    1.5496423
#> lag1_var3  0.4737588  1.8868034   -1.832152    2.5457779
#> lag1_var4  0.4076330  1.7700028   -1.545430    2.5589576
#> const     30.5927082 11.8883907   14.389914   39.4196177
#> exo1      -0.6320364  0.9320438   -1.820119    0.3154556
#> exo2      -4.2704920  0.8319083   -5.186491   -3.3635742
#> exo3       3.3687831  0.6569179    2.682120    4.0850372
#> 
#> 
#>