Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004387646 0.001230972 0.003134479  0.005833945
#> 
#> $AFG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01346984 0.007020983 -0.02290498 -0.007931132
#> Sigma[2,2]  0.21893610 0.058062403  0.17836870  0.298181700
#> 
#> $AFG$Sigma$equation3
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01499633 0.01157574  0.002547805   0.02912240
#> Sigma[3,2] -0.14876571 0.05478495 -0.210355665  -0.09487794
#> Sigma[3,3]  0.38429067 0.11289618  0.259239393   0.51829954
#> 
#> $AFG$Sigma$equation4
#>                   mean         sd   5% quantile 95% quantile
#> Sigma[4,1]  0.01295906 0.01063350  0.0001534586   0.02441788
#> Sigma[4,2] -0.06644313 0.05882876 -0.1418669973  -0.01158887
#> Sigma[4,3]  0.39944478 0.12109055  0.2564106511   0.53675548
#> Sigma[4,4]  0.47050183 0.15589551  0.2842334627   0.64403174
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.97173877 0.02335219  0.942867750   0.99606420
#> lag1_var2  0.03987821 0.04404477 -0.008894178   0.08562180
#> lag1_var3  0.15021580 0.11875975  0.008707235   0.27447415
#> lag1_var4 -0.14661547 0.10881577 -0.260318036  -0.01580918
#> const      0.78929650 0.86940621 -0.133987064   1.78205715
#> exo1      -0.05504484 0.07226528 -0.132309156   0.03055888
#> exo2      -0.04973571 0.08745678 -0.158008694   0.04434237
#> exo3      -0.11063324 0.04651322 -0.174770152  -0.07991836
#> 
#> $AFG$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.43556118 0.1417808  0.27527851    0.5858144
#> lag1_var2  0.20869846 0.1243016  0.03898902    0.2887359
#> lag1_var3 -1.80888111 0.2147438 -2.10010750   -1.6596432
#> lag1_var4  1.61383850 0.1773831  1.46023314    1.8410646
#> const     -1.14288280 2.2735827 -3.66208550    1.4440533
#> exo1       0.02569260 0.5424560 -0.66883798    0.5579288
#> exo2      -0.06149396 0.2872701 -0.40956269    0.2573160
#> exo3      -0.65589324 0.4470496 -1.03896575   -0.1379894
#> 
#> $AFG$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  -0.3703482 0.2640463  -0.6710928  -0.06087343
#> lag1_var2   2.2172007 0.3969649   1.6767253   2.46382263
#> lag1_var3   6.9098924 1.2398965   5.2088115   7.78945393
#> lag1_var4  -5.4180537 1.1936023  -6.3297917  -3.78727105
#> const     -10.2262437 3.9262014 -14.6850777  -6.25377759
#> exo1        0.1726159 0.3635851  -0.3254965   0.44823571
#> exo2       -0.4944030 0.5790013  -1.2207273  -0.01409351
#> exo3        1.1165537 0.5529783   0.6896730   1.86455145
#> 
#> $AFG$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  -0.2791609 0.2932006  -0.6408872   0.03816175
#> lag1_var2   2.2646496 0.5820650   1.4647030   2.60271741
#> lag1_var3   6.1749882 1.7046222   3.8409610   7.47193072
#> lag1_var4  -4.6959849 1.6116642  -5.9949508  -2.50517902
#> const     -11.0402856 5.7651646 -17.7959215  -5.00307360
#> exo1        0.2533555 0.3872975  -0.1237547   0.75425396
#> exo2       -0.4630356 0.7219424  -1.2486712   0.18773634
#> exo3        0.9695905 0.4598800   0.6599543   1.60398136
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.007088164 0.002050373 0.005129544  0.009746661
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01204765 0.004017148 -0.01548758 -0.007231383
#> Sigma[2,2]  0.13133195 0.049654986  0.09585585  0.199611855
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.006213343 0.001235156  0.005014155  0.007651366
#> Sigma[3,2] -0.052321400 0.023939316 -0.084288814 -0.031596725
#> Sigma[3,3]  0.037291198 0.014961487  0.022391030  0.055293793
#> 
#> $AGO$Sigma$equation4
#>                     mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.0004267131 0.001081733 -0.0001679901  0.001884347
#> Sigma[4,2]  0.0500220616 0.019892639  0.0316826849  0.076220684
#> Sigma[4,3] -0.0059172710 0.011026914 -0.0186909392  0.006159580
#> Sigma[4,4]  0.0379042890 0.009995138  0.0291231570  0.051224504
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.88897709 0.02392623  0.87165822   0.92186462
#> lag1_var2  0.39656085 0.07937761  0.35118186   0.50377178
#> lag1_var3  0.53690260 0.06172920  0.49465793   0.62158518
#> lag1_var4 -0.50591536 0.07025521 -0.60099109  -0.45530283
#> const      0.65409784 0.81536583 -0.10088252   1.69408132
#> exo1       0.09946882 0.08505411 -0.01162168   0.18010385
#> exo2      -0.09087254 0.09041041 -0.16998018   0.02511181
#> exo3      -0.14448921 0.14047314 -0.29074340   0.01239763
#> 
#> $AGO$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.14011129 0.1090392  0.01386622   0.26488108
#> lag1_var2  0.44057769 0.1932985  0.30496530   0.70142229
#> lag1_var3 -0.75608892 0.2879962 -0.92530426  -0.36294821
#> lag1_var4  0.77454635 0.2471098  0.43448390   0.93142382
#> const     -5.36745071 4.3536009 -8.80473495   0.56679331
#> exo1      -0.05084327 0.1177480 -0.18636886   0.08174668
#> exo2       0.64868482 0.2247144  0.36769772   0.86361126
#> exo3      -0.38333805 0.3722798 -0.81531999  -0.10478880
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.17032972 0.03538451  -0.2096566   -0.1284591
#> lag1_var2  0.24450090 0.10487859   0.1314075    0.3539561
#> lag1_var3  1.38093605 0.06879498   1.3022927    1.4495914
#> lag1_var4 -0.28723131 0.11529549  -0.4098922   -0.1696087
#> const     -2.25895041 3.31779309  -4.9403531    2.1104406
#> exo1       0.06228297 0.09118429  -0.0405603    0.1530189
#> exo2      -1.43608390 0.16429038  -1.6134198   -1.2639376
#> exo3       3.46867499 0.50568714   3.0379289    4.1361432
#> 
#> $AGO$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.2743241 0.04797402 -0.33789969   -0.2351266
#> lag1_var2  0.4911755 0.21403029  0.31747971    0.7780184
#> lag1_var3  0.6814837 0.25085744  0.47713791    1.0158074
#> lag1_var4  0.4467398 0.25384104  0.11143599    0.6530739
#> const     -2.7415855 2.24545513 -4.34223130    0.3440115
#> exo1       0.1245797 0.07470856  0.05404548    0.2081456
#> exo2      -1.3254809 0.09422277 -1.44562363   -1.2359913
#> exo3       3.4213366 0.25815067  3.17367799    3.7302852
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002223392 0.0001802771 0.002013916  0.002408814
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.06726697 0.03734375  -0.1017288  -0.02248116
#> Sigma[2,2]  7.98048516 2.50852271   4.7142893  10.17076513
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[3,1]  0.02216153 0.02511843 -0.006142645   0.05213446
#> Sigma[3,2] -5.20672497 1.51644103 -6.902973622  -3.44627936
#> Sigma[3,3]  6.92061859 1.49924425  4.985629558   8.26081631
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.02680418 0.02407672 -0.05661232 -0.001844723
#> Sigma[4,2]  0.11550643 2.53384897 -1.64294954  3.605156819
#> Sigma[4,3]  4.12955789 1.69072056  2.59052090  6.278293910
#> Sigma[4,4]  5.04422657 2.24206360  2.30037693  6.854729449
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.97478184 0.01535692  0.963779037   0.99548503
#> lag1_var2 -0.05518109 0.02604396 -0.085495585  -0.03159243
#> lag1_var3 -0.09105644 0.04020944 -0.138644704  -0.05747013
#> lag1_var4  0.07749285 0.03410479  0.048376196   0.11871402
#> const      1.43375798 0.35976207  1.186383317   1.92982101
#> exo1       0.03933625 0.02944625  0.008471749   0.07499115
#> exo2      -0.04927365 0.04469330 -0.085968015   0.01099366
#> exo3       0.05816769 0.04397042  0.013804263   0.11292206
#> 
#> $ALB$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -2.3540886  1.450008   -4.201864   -0.9324334
#> lag1_var2  3.2097686  2.646555    1.125983    6.8193250
#> lag1_var3  4.3557348  4.093246    1.082732    9.9240547
#> lag1_var4 -3.5819495  3.586969   -8.462487   -0.7216098
#> const     14.1650920 23.149668   -2.751229   45.3988606
#> exo1      -1.6520111  2.488880   -4.806522    0.8307249
#> exo2       0.1873961  2.215778   -1.856981    3.0550173
#> exo3      -0.7971816  2.645983   -4.368377    1.4253603
#> 
#> $ALB$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1   1.4042464  1.694118  -0.3754378     3.216912
#> lag1_var2   0.4796450  2.541399  -1.7826649     3.514247
#> lag1_var3   1.3543644  4.094957  -2.0614077     6.268170
#> lag1_var4  -0.4867511  3.595047  -4.8476008     2.429618
#> const     -28.5673078 29.622325 -59.6710533     8.522230
#> exo1       -0.5332375  2.673312  -3.8659311     2.251921
#> exo2       -1.5234219  2.855306  -5.0762701     1.461780
#> exo3        1.7816454  2.430098  -1.0269072     4.394586
#> 
#> $ALB$A$equation4
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1  -0.4399434  1.034638  -1.45809017    0.7882338
#> lag1_var2   2.4666807  2.498807  -0.13977658    5.1267687
#> lag1_var3   4.1365422  3.997252  -0.03956505    8.4403084
#> lag1_var4  -2.6435554  3.508314  -6.40104222    1.0846027
#> const     -19.1860046 23.378137 -41.43762781   10.8173693
#> exo1       -1.6420380  1.715841  -3.79550641    0.1477083
#> exo2       -1.6043928  2.491534  -4.14054473    1.3137362
#> exo3        1.4362912  2.168071  -0.94932022    3.9636759
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd  5% quantile 95% quantile
#> Sigma[1,1] 0.001105877 0.0001882489 0.0009485841  0.001356825
#> 
#> $ARE$Sigma$equation2
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.00240205 0.001963928 -0.005083886 -0.001020159
#> Sigma[2,2]  0.07788692 0.016969260  0.055771716  0.092769629
#> 
#> $ARE$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.00432672 0.001600823  0.002596973  0.006280921
#> Sigma[3,2] -0.05366646 0.027263439 -0.080294383 -0.021357548
#> Sigma[3,3]  0.15956609 0.042188779  0.103298962  0.195600977
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.003145886 0.001011034  0.001845084  0.004084236
#> Sigma[4,2] -0.002815290 0.020649925 -0.018020650  0.025330768
#> Sigma[4,3]  0.124743054 0.038289902  0.074804575  0.158672573
#> Sigma[4,4]  0.127535027 0.043549126  0.075344622  0.175865933
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1  0.93888929 0.020008571  0.92673312   0.96642079
#> lag1_var2  0.07201362 0.008981644  0.06321042   0.08354802
#> lag1_var3  0.06868446 0.013697126  0.05658929   0.08575687
#> lag1_var4 -0.06194576 0.016874425 -0.08458058  -0.04996447
#> const      1.08361368 0.265609322  0.81797620   1.41219572
#> exo1      -0.01731694 0.041554724 -0.06337508   0.02570716
#> exo2      -0.08941729 0.044985129 -0.14340824  -0.04006967
#> exo3      -0.02087818 0.054069573 -0.08047011   0.04488737
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.39971786 0.2569463   0.1798960    0.6964598
#> lag1_var2 -0.38120375 0.1994294  -0.5844705   -0.1462260
#> lag1_var3 -1.35303133 0.2645493  -1.6144236   -1.0457746
#> lag1_var4  1.23445496 0.2193189   0.9641892    1.4370118
#> const     -0.61660592 2.6318174  -3.6566254    2.2871839
#> exo1       0.03141917 0.1673440  -0.1672379    0.2191703
#> exo2       2.07084695 0.3891589   1.5998376    2.5006292
#> exo3      -0.40544277 0.4936336  -0.7611516    0.2710573
#> 
#> $ARE$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.09699284 0.14540292  -0.2690859   0.04960317
#> lag1_var2  1.83843655 0.05943958   1.7678176   1.89230847
#> lag1_var3  3.04569277 0.24349774   2.8386689   3.36316641
#> lag1_var4 -1.95031551 0.26123596  -2.3038851  -1.77300735
#> const     -5.15234026 1.86381092  -7.3537234  -3.24904008
#> exo1      -0.63288409 0.35110607  -0.8713400  -0.15145828
#> exo2      -2.64914400 0.66509214  -3.5105760  -2.08190136
#> exo3      -1.65702339 0.71289379  -2.4050552  -0.77341481
#> 
#> $ARE$A$equation4
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.005536553 0.06034683 -0.07605774   0.05045099
#> lag1_var2  1.544957298 0.12790072  1.42026686   1.70976109
#> lag1_var3  1.901737908 0.37215284  1.66903804   2.41394460
#> lag1_var4 -0.860515067 0.39137070 -1.39803136  -0.60840373
#> const     -3.276246330 1.83415974 -5.29079460  -1.29534018
#> exo1      -0.639369132 0.27407059 -0.82911285  -0.26136046
#> exo2      -1.033496650 0.54826708 -1.72847284  -0.54218066
#> exo3      -2.135137327 0.56006097 -2.68748897  -1.41668892
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002545662 0.0004274297 0.002153565  0.003099963
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05493438 0.02344633   -0.086111  -0.03441144
#> Sigma[2,2]  3.28505215 1.09681797    2.194451   4.54755743
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04347336 0.02071497  0.02725358   0.07129527
#> Sigma[3,2] -2.32763146 0.86458152 -3.43245834  -1.51115817
#> Sigma[3,3]  1.89682522 0.69203557  1.36198356   2.83945522
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01189987 0.008753086  0.004782788   0.02369752
#> Sigma[4,2] -0.37248732 0.301089835 -0.787428324  -0.19793886
#> Sigma[4,3]  0.53266861 0.243480404  0.387445702   0.86811360
#> Sigma[4,4]  0.33886545 0.095434055  0.231267699   0.44628474
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.91932649 0.01465316  0.905781743   0.93729026
#> lag1_var2  0.16979251 0.04887318  0.131179924   0.23641417
#> lag1_var3  0.27991715 0.07711926  0.220656803   0.38543353
#> lag1_var4 -0.26048889 0.06619083 -0.350763569  -0.20931720
#> const      1.01667467 1.04679458 -0.427936010   1.72564256
#> exo1       0.02419698 0.01639101  0.003007137   0.03889699
#> exo2      -0.10837489 0.05474832 -0.160883522  -0.04942852
#> exo3       0.05497047 0.04934578  0.001309782   0.10224545
#> 
#> $ARG$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -1.24953258  0.2461044   -1.504197   -0.9388512
#> lag1_var2  1.07125534  1.7792464   -1.135020    2.7769995
#> lag1_var3  0.55390358  2.8061527   -2.925533    3.3158499
#> lag1_var4  0.02428109  2.6374813   -2.514087    3.3577110
#> const      1.48973227 18.5831020  -17.609974   23.2884317
#> exo1      -0.99407363  1.2851393   -2.465073    0.2936002
#> exo2       0.94406254  3.7516885   -3.441169    4.8003512
#> exo3      -1.27233914  2.2018378   -4.148648    0.8473569
#> 
#> $ARG$A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.5905340  0.3613488   0.2075549     1.038987
#> lag1_var2  1.9581909  1.3246423   0.8036311     3.557951
#> lag1_var3  3.9070409  2.0998502   2.1259638     6.456074
#> lag1_var4 -3.1364050  1.9166625  -5.5344861    -1.550388
#> const     -3.9563338 12.5662023 -18.2635100    10.666565
#> exo1      -0.2256198  1.0224588  -1.2969305     1.020032
#> exo2      -4.4137309  2.3776678  -6.7250249    -1.552689
#> exo3       2.9619586  1.3009825   1.3239311     4.300183
#> 
#> $ARG$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.4073072 0.3935677  -0.7825394   0.06923767
#> lag1_var2  2.6923966 0.4126847   2.1680324   3.06219642
#> lag1_var3  4.3785004 0.7043053   3.4812329   5.01289467
#> lag1_var4 -3.2060849 0.5440705  -3.6783614  -2.51993223
#> const      0.3247415 2.8877905  -2.7855862   3.83051192
#> exo1      -0.8614800 0.3528245  -1.2304110  -0.42549367
#> exo2      -4.3213531 0.4197292  -4.8712340  -3.97342532
#> exo3       2.2819421 0.6424966   1.5486198   2.84358379
#> 
#> 
#>