Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.006868507 0.003277347 0.004268531    0.0113331
#> 
#> $AFG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01813948 0.008356545 -0.02815544 -0.008724818
#> Sigma[2,2]  0.23095103 0.060447251  0.17365592  0.309808409
#> 
#> $AFG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.02121634 0.01279650  0.01122166   0.03856006
#> Sigma[3,2] -0.18607511 0.07335290 -0.27530899  -0.10723474
#> Sigma[3,3]  0.32466843 0.08562895  0.22199962   0.41106598
#> 
#> $AFG$Sigma$equation4
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01461232 0.01143685  0.005533539   0.02964301
#> Sigma[4,2] -0.09441645 0.06337029 -0.156018140  -0.01458607
#> Sigma[4,3]  0.27452162 0.06636939  0.195654509   0.34945333
#> Sigma[4,4]  0.26464556 0.04786740  0.216696223   0.32544025
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.98236889 0.02651016   0.9479057  1.006008694
#> lag1_var2 -0.48419553 0.02309573  -0.5147109 -0.463778798
#> lag1_var3 -1.04707946 0.04929219  -1.1140345 -1.018304131
#> lag1_var4  0.91919419 0.04448845   0.8876128  0.979167909
#> const      6.42034069 0.69146638   5.8437141  7.345003066
#> exo1      -0.02428594 0.09993725  -0.1552969  0.070350761
#> exo2      -0.10488226 0.11008957  -0.2279651  0.001308147
#> exo3      -0.30965911 0.05458352  -0.3556605 -0.236139222
#> 
#> $AFG$A$equation2
#>                    mean        sd 5% quantile 95% quantile
#> lag1_var1   0.272092020 0.1100897   0.1550313    0.4065232
#> lag1_var2   1.627252018 0.2787595   1.2556129    1.8760904
#> lag1_var3   1.261686251 0.3769520   0.7470925    1.5198481
#> lag1_var4  -1.056037717 0.3093768  -1.2753524   -0.6350613
#> const     -16.041725595 3.6886359 -19.9414765  -11.5186568
#> exo1       -0.205023896 0.2788764  -0.4724377    0.1547084
#> exo2        0.002113274 0.4911756  -0.6097924    0.4661645
#> exo3        0.172283399 0.8447887  -0.8902301    0.9507474
#> 
#> $AFG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.19922370 0.07573192 -0.27309481   -0.1099933
#> lag1_var2 -0.64769728 0.28309441 -0.96517727   -0.3438265
#> lag1_var3  0.26813504 0.55357955 -0.36578410    0.7748364
#> lag1_var4  0.51624300 0.50996240  0.04524295    1.1029307
#> const     16.93879934 2.09081742 14.63961129   19.1280167
#> exo1       0.14357129 0.65468563 -0.70925888    0.6911666
#> exo2      -1.32051005 0.52585374 -1.88306947   -0.7640743
#> exo3      -0.01084726 0.82803949 -0.76680163    1.0242066
#> 
#> $AFG$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.07770638 0.09133033 -0.15187248   0.04434347
#> lag1_var2  0.13134567 0.36901141 -0.31181254   0.47930776
#> lag1_var3  0.98671709 0.76025100  0.02327016   1.63612253
#> lag1_var4  0.03600936 0.68683572 -0.55661337   0.90328200
#> const      3.23899548 1.96561311  0.79164233   5.22017389
#> exo1       0.08235819 0.66272492 -0.75198784   0.72937145
#> exo2      -1.41128538 0.38558960 -1.82869913  -0.94980907
#> exo3       0.20069565 0.53222925 -0.29758018   0.88114691
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005582319 0.001013292 0.004960648  0.006980528
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01349697 0.005092046 -0.01984233 -0.008152566
#> Sigma[2,2]  0.12280208 0.010736734  0.11121784  0.136088298
#> 
#> $AGO$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01042808 0.002444499  0.008401719   0.01367230
#> Sigma[3,2] -0.05732558 0.006232199 -0.064453751  -0.04995669
#> Sigma[3,3]  0.04249044 0.002840160  0.039555427   0.04593305
#> 
#> $AGO$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1] 0.0001509951 0.002158031 -0.002217195  0.002770241
#> Sigma[4,2] 0.0409902451 0.006831039  0.034569598  0.048467630
#> Sigma[4,3] 0.0003515281 0.003130050 -0.003506391  0.003702966
#> Sigma[4,4] 0.0386204520 0.006384265  0.031268561  0.046155225
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  1.00426428 0.02140093  0.981292396   1.02841427
#> lag1_var2 -0.23934307 0.01695232 -0.261931057  -0.22517584
#> lag1_var3 -0.28763906 0.06154731 -0.358975473  -0.21554222
#> lag1_var4  0.26454847 0.03036916  0.227183158   0.29669664
#> const      2.00194906 1.99037700 -0.290689770   4.10914536
#> exo1       0.03011153 0.06108464 -0.051897316   0.07672944
#> exo2      -0.02443020 0.04879312 -0.076916096   0.03289534
#> exo3       0.12786423 0.10881832 -0.006534646   0.24341081
#> 
#> $AGO$A$equation2
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.006112493 0.07240737 -0.09186794   0.06822359
#> lag1_var2  0.255753338 0.05100025  0.20291292   0.32039504
#> lag1_var3 -0.972882697 0.24538631 -1.16533811  -0.64801680
#> lag1_var4  0.933483781 0.14460309  0.76406273   1.06961722
#> const      3.013406585 8.02965968 -7.46600067  10.07243781
#> exo1       0.118680160 0.27748349 -0.14895329   0.45740701
#> exo2       0.337944654 0.25294364  0.06642563   0.63495738
#> exo3      -0.544983438 0.29990986 -0.86502399  -0.17167852
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.26572456 0.05162777  -0.3190247   -0.2060936
#> lag1_var2  0.58735645 0.03946708   0.5377657    0.6274550
#> lag1_var3  1.79512705 0.14833634   1.6046602    1.9415448
#> lag1_var4 -0.77765053 0.09518602  -0.8755144   -0.6625578
#> const      5.54278781 3.79554683   2.1488029    9.6608172
#> exo1       0.05657108 0.17806562  -0.1219785    0.2657893
#> exo2      -1.59057247 0.20250855  -1.7886340   -1.3491381
#> exo3       3.05522818 0.27287269   2.7709872    3.3966295
#> 
#> $AGO$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.35539531 0.07238655 -0.43470048  -0.26759281
#> lag1_var2  0.91319081 0.02076620  0.89186429   0.93722231
#> lag1_var3  1.23119767 0.07748357  1.13982730   1.31029470
#> lag1_var4 -0.08916267 0.03145479 -0.11234947  -0.04623549
#> const     -1.76862818 5.20633360 -8.17703230   3.79232768
#> exo1       0.19023565 0.11271118  0.05214913   0.30141333
#> exo2      -1.64807685 0.05735098 -1.70994753  -1.57750542
#> exo3       3.03270597 0.22362759  2.79601451   3.26489643
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002484517 0.00066699 0.001917599  0.003375961
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05739647 0.02789309   -0.095740  -0.03779317
#> Sigma[2,2]  7.84723191 1.71067726    6.051018   9.95476998
#> 
#> $ALB$Sigma$equation3
#>                   mean        sd   5% quantile 95% quantile
#> Sigma[3,1]  0.02192314 0.0183562 -0.0002340787     0.042429
#> Sigma[3,2] -5.11273133 0.4145705 -5.4629542560    -4.608436
#> Sigma[3,3]  5.03089579 0.7066133  4.3749422649     5.860148
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.02025659 0.01201567 -0.03365708 -0.006638359
#> Sigma[4,2]  0.15379671 1.01423512 -0.85040281  1.461269674
#> Sigma[4,3]  1.96049012 0.94786949  0.94644517  3.104136906
#> Sigma[4,4]  2.52131292 0.48857085  1.94689026  3.071462969
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.99676034 0.03029883  0.96334423  1.026665100
#> lag1_var2 -0.01433574 0.02772713 -0.04908860  0.013646023
#> lag1_var3 -0.03170694 0.03997608 -0.08157958  0.009007659
#> lag1_var4  0.02527147 0.03623424 -0.01061639  0.071201299
#> const      0.42024661 0.74979718 -0.49895915  1.232489252
#> exo1       0.03097653 0.05570381 -0.02463527  0.098469587
#> exo2      -0.07862867 0.04928149 -0.14079485 -0.042638009
#> exo3       0.11718660 0.01533543  0.09836927  0.133840890
#> 
#> $ALB$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  -5.2519184  1.234462  -6.5627535   -3.8407712
#> lag1_var2   1.3269130  4.185308  -1.8796982    6.9519037
#> lag1_var3   1.6647287  6.359428  -3.0126659   10.2341851
#> lag1_var4  -1.3963384  5.501001  -8.8028116    2.7442199
#> const     114.3407867 50.777562  53.3478745  168.7159009
#> exo1       -0.5510888  2.302821  -2.7699592    2.4220711
#> exo2        1.7274849  1.691461  -0.3190667    3.5040892
#> exo3       -2.6382746  3.163834  -6.8269866    0.1838136
#> 
#> $ALB$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1   3.0306787  0.9870622   2.0193812   4.19113353
#> lag1_var2  -0.6598743  3.0055843  -4.8123903   1.34804590
#> lag1_var3  -0.6349016  4.6249268  -7.0244464   2.44951668
#> lag1_var4   1.2757767  3.9610337  -1.3942729   6.74341941
#> const     -52.6820691 39.2663656 -96.6960712  -5.45979268
#> exo1       -1.9741311  1.4662693  -3.6826385  -0.32965722
#> exo2       -1.7101161  1.4024421  -3.0626655   0.06000393
#> exo3        1.3378570  1.8368430  -0.1739508   3.81273147
#> 
#> $ALB$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.53256992  0.8967419  -1.5828165    0.4613601
#> lag1_var2  0.07998724  0.8362628  -0.5803373    1.1437329
#> lag1_var3  0.31670518  1.3033550  -0.7270701    1.9472940
#> lag1_var4  0.62753187  1.1347438  -0.7697898    1.5365250
#> const     17.26999185 12.5526654   3.9221422   30.4980537
#> exo1      -2.44096373  1.6900867  -4.7179268   -1.0655389
#> exo2      -0.58379968  1.9005577  -2.7810663    1.5424837
#> exo3      -0.35976323  0.6580107  -1.0042043    0.4809509
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001993362 0.0004807204 0.001429179  0.002543444
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.000384419 0.001183888 -0.001843882 0.0008375122
#> Sigma[2,2]  0.062473910 0.017974421  0.041500039 0.0820613847
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.003161699 0.001562919  0.001858722  0.005268753
#> Sigma[3,2] -0.050663255 0.020696999 -0.075270530 -0.029567734
#> Sigma[3,3]  0.170384169 0.018089635  0.147128917  0.187201638
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.002624154 0.001924025  0.0003475404  0.004710452
#> Sigma[4,2] -0.002783496 0.017012479 -0.0241359286  0.013873057
#> Sigma[4,3]  0.133400340 0.012130214  0.1200185257  0.143816386
#> Sigma[4,4]  0.133747778 0.017795546  0.1137105834  0.152681225
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.960948169 0.03490404  0.91305573   0.98424036
#> lag1_var2 -0.324006154 0.01772560 -0.34724586  -0.30760444
#> lag1_var3 -0.444279450 0.02247456 -0.47484386  -0.42704057
#> lag1_var4  0.436034057 0.01777644  0.42135072   0.45930729
#> const      1.656768709 0.72783522  0.80226641   2.40906530
#> exo1      -0.011573701 0.04904093 -0.05341092   0.05311218
#> exo2      -0.087373869 0.03488924 -0.12032418  -0.04993159
#> exo3      -0.004359079 0.03967571 -0.03448717   0.04899460
#> 
#> $ARE$A$equation2
#>                    mean        sd 5% quantile 95% quantile
#> lag1_var1  2.005127e-01 0.3449534 -0.05363266   0.67342170
#> lag1_var2 -1.395486e-01 0.2269720 -0.35445612   0.13127351
#> lag1_var3 -1.135899e+00 0.2590315 -1.45443430  -0.88290331
#> lag1_var4  1.026875e+00 0.2945386  0.75314363   1.37287332
#> const      3.671120e+00 5.3204639 -2.41846863   9.70063437
#> exo1       6.075776e-05 0.2714309 -0.33010308   0.29459905
#> exo2       2.046054e+00 0.1500700  1.86565819   2.21110779
#> exo3      -6.734110e-01 0.5912722 -1.26301112   0.07348412
#> 
#> $ARE$A$equation3
#>                 mean        sd   5% quantile 95% quantile
#> lag1_var1  0.2587540 0.2090099   0.008257957    0.4596987
#> lag1_var2  0.9840658 0.5708297   0.394000079    1.6023715
#> lag1_var3  1.8823619 0.7423621   1.256297329    2.7064974
#> lag1_var4 -0.8584111 0.7961141  -1.745716604   -0.1452045
#> const     -9.0809164 2.2951224 -11.639690537   -6.9547604
#> exo1      -0.7481464 0.7263141  -1.360278176    0.2189155
#> exo2      -2.6395242 0.3357434  -2.941094253   -2.2152346
#> exo3      -1.7603952 0.6103126  -2.544643600   -1.1691086
#> 
#> $ARE$A$equation4
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1  0.40010512 0.3042777   0.03367551   0.69446418
#> lag1_var2  0.91632269 0.5212416   0.39160923   1.50955350
#> lag1_var3  1.05681054 0.5802240   0.58711575   1.71392748
#> lag1_var4 -0.08383854 0.5891396  -0.72379614   0.43652079
#> const     -8.44377870 4.1695645 -13.31171427  -4.04714598
#> exo1      -0.77121122 0.6612597  -1.45587126   0.03341988
#> exo2      -1.07083453 0.2498033  -1.25762617  -0.76188989
#> exo3      -2.33603348 0.6922236  -3.17258770  -1.69231206
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003275632 0.0006365202 0.002477926  0.003799344
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.06368687 0.01958501 -0.08763911  -0.04348942
#> Sigma[2,2]  3.02051712 0.45505265  2.54706548   3.57179207
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04877787 0.01305542  0.03217083   0.06207674
#> Sigma[3,2] -2.16713995 0.26295279 -2.48878309  -1.89422980
#> Sigma[3,3]  1.76020042 0.15847677  1.62386361   1.93561034
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[4,1]  0.01103963 0.004584116  0.00551107   0.01604695
#> Sigma[4,2] -0.36090628 0.071009783 -0.43184499  -0.27079223
#> Sigma[4,3]  0.48761038 0.106818657  0.35713346   0.59070577
#> Sigma[4,4]  0.29925057 0.074840391  0.21693771   0.38751543
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.94759711 0.05604774  0.88621717   1.01220808
#> lag1_var2 -0.29402365 0.03229722 -0.32261249  -0.25173095
#> lag1_var3 -0.47960202 0.05616197 -0.53111306  -0.40676224
#> lag1_var4  0.42624785 0.05293802  0.35767827   0.47584754
#> const      4.68851176 1.54975843  3.00517694   6.59455292
#> exo1       0.01564843 0.06231999 -0.05775924   0.08316893
#> exo2      -0.10673933 0.05621326 -0.15374503  -0.03304387
#> exo3       0.07807575 0.04140822  0.03977592   0.12998470
#> 
#> $ARG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -3.0239705  1.8444161  -4.8111805   -0.6680593
#> lag1_var2  1.1363290  0.7878165   0.1661393    1.8854503
#> lag1_var3  0.7698801  1.4035920  -0.8655110    2.1740763
#> lag1_var4 -0.2945437  1.3919299  -1.7110952    1.4076418
#> const     55.8836215 59.7028379 -21.3241266  112.8011283
#> exo1      -0.9192956  0.8901082  -1.5764669    0.2903405
#> exo2       1.8527155  1.3049540   0.6006539    3.5516931
#> exo3       0.8381752  2.0280662  -1.9637183    2.1518614
#> 
#> $ARG$A$equation3
#>                  mean         sd   5% quantile 95% quantile
#> lag1_var1  1.43290081  1.1479369  6.715915e-04    2.5371076
#> lag1_var2  0.07711137  0.5524769 -4.561517e-01    0.7840680
#> lag1_var3  0.75617049  0.9693867 -2.175862e-01    1.9719625
#> lag1_var4 -0.27603194  0.9834614 -1.533210e+00    0.6516179
#> const     -9.21146802 41.7023510 -4.910964e+01   43.5867552
#> exo1       0.03693857  1.0609581 -1.238325e+00    1.0600873
#> exo2      -4.28625064  0.8200427 -5.394513e+00   -3.6367557
#> exo3       1.69093817  1.9791877 -2.179260e-01    4.2564280
#> 
#> $ARG$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.4650370 0.1853409  -0.6538984   -0.2331498
#> lag1_var2  0.8520506 0.1295391   0.6978884    0.9640276
#> lag1_var3  1.3518889 0.2425943   1.0515277    1.5729479
#> lag1_var4 -0.5002045 0.1840836  -0.7466337   -0.3547898
#> const     21.3056316 7.2425570  13.4689623   29.2310386
#> exo1      -0.5622163 0.6728790  -1.3772789    0.1139144
#> exo2      -3.5974980 0.3882440  -4.0433580   -3.1661343
#> exo3       2.4120009 1.0886396   1.0008244    3.4793432
#> 
#> 
#>