Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[1,1] 0.01469882 0.02030943 0.004777633    0.0420366
#> 
#> $AFG$Sigma$equation2
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.004524027 0.03018404 -0.02737295    0.0366309
#> Sigma[2,2]  0.431796293 0.21387407  0.22509739    0.7046896
#> 
#> $AFG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1] -0.01176859 0.05295649 -0.08460129   0.02316256
#> Sigma[3,2] -0.28747447 0.12891953 -0.40115692  -0.13449580
#> Sigma[3,3]  0.38426460 0.20985513  0.20376232   0.66520254
#> 
#> $AFG$Sigma$equation4
#>                   mean         sd 5% quantile  95% quantile
#> Sigma[4,1] -0.01595256 0.04662849 -0.07967298  0.0129938462
#> Sigma[4,2] -0.11276619 0.11010593 -0.25343329 -0.0007552898
#> Sigma[4,3]  0.30319093 0.19760253  0.16833091  0.5744995648
#> Sigma[4,4]  0.30051434 0.18049607  0.19966624  0.5485201741
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.93781653 0.02885884  0.91399148   0.97599215
#> lag1_var2  0.19131585 0.10438076  0.06831550   0.29823154
#> lag1_var3  0.43578041 0.27695200  0.14227698   0.75133503
#> lag1_var4 -0.38822630 0.26343477 -0.70547967  -0.12430333
#> const     -0.63489640 1.11943195 -1.74740807   0.75823053
#> exo1       0.06249123 0.03746911  0.01379976   0.09665209
#> exo2      -0.07473441 0.16083364 -0.29368113   0.01735709
#> exo3      -0.11437224 0.17700553 -0.26845155   0.10825939
#> 
#> $AFG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.6630908  0.3737610   0.3093744    1.0551968
#> lag1_var2 -1.0157093  1.0969625  -2.2833857    0.1934028
#> lag1_var3 -3.7494216  2.0069108  -6.1767450   -1.5683492
#> lag1_var4  3.0870612  1.6842742   1.2583954    5.1303121
#> const     18.1149733 10.9352933   6.3920498   28.8882270
#> exo1      -0.3892587  1.1640489  -1.7190977    0.9823487
#> exo2      -0.1437050  0.3261892  -0.5636470    0.1350363
#> exo3      -0.6641669  0.6308388  -1.5207264   -0.1786031
#> 
#> $AFG$A$equation3
#>                  mean        sd 5% quantile  95% quantile
#> lag1_var1  -0.2797007 0.1646852  -0.4487763  -0.100057142
#> lag1_var2   1.9916847 0.8363061   1.0462055   2.984458569
#> lag1_var3   5.6332877 1.8880392   3.4384808   7.815429720
#> lag1_var4  -4.0341844 1.7232363  -5.9779645  -2.008307797
#> const     -20.3333001 6.5212212 -29.1005056 -14.770059196
#> exo1        0.0454860 0.7173504  -0.8202533   0.847083370
#> exo2       -0.8831793 0.8299453  -1.7179657  -0.002974364
#> exo3        0.1369006 0.9140141  -1.0528406   1.011583396
#> 
#> $AFG$A$equation4
#>                   mean        sd 5% quantile 95% quantile
#> lag1_var1  -0.01093931 0.1149955  -0.1468608   0.11821455
#> lag1_var2   1.32201925 0.6010994   0.6964299   2.01604885
#> lag1_var3   3.52432901 1.4217937   1.9250128   5.09518624
#> lag1_var4  -2.15264375 1.3231016  -3.5678740  -0.62071065
#> const     -14.54568615 5.3916209 -21.3722728  -9.67165523
#> exo1       -0.12072537 0.2939770  -0.4498686   0.21658391
#> exo2       -0.99469152 0.9523165  -2.0364186   0.04203958
#> exo3       -0.15419956 0.9854489  -1.5137874   0.49910698
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.009848275 0.002301732 0.006673491   0.01144618
#> 
#> $AGO$Sigma$equation2
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.007084449 0.005943662 -0.01308393 0.0005133205
#> Sigma[2,2]  0.119153347 0.020570497  0.09591506 0.1415213681
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.006413462 0.003308054  0.002925746   0.01009665
#> Sigma[3,2] -0.049833047 0.008120199 -0.058723861  -0.04107713
#> Sigma[3,3]  0.034197208 0.006147799  0.026208464   0.04004631
#> 
#> $AGO$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.003504452 0.003458883  0.0008213882  0.008180151
#> Sigma[4,2]  0.051260306 0.010514075  0.0389420439  0.062451893
#> Sigma[4,3] -0.008072087 0.007688424 -0.0152969790  0.001862592
#> Sigma[4,4]  0.038404106 0.004503099  0.0326049849  0.042699510
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.9104932 0.02993527   0.8775663   0.93962929
#> lag1_var2  0.4662573 0.08220294   0.3878791   0.56943258
#> lag1_var3  0.6364010 0.11205393   0.5313385   0.75946482
#> lag1_var4 -0.5890678 0.10574408  -0.7128130  -0.49218217
#> const     -1.0138047 0.95353657  -2.1282207   0.01386799
#> exo1       0.0692661 0.03894137   0.0210587   0.11127498
#> exo2      -0.1098839 0.14146004  -0.2496587   0.04215046
#> exo3      -0.1816590 0.17144115  -0.3242261   0.04882611
#> 
#> $AGO$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.06081276 0.1325492 -0.17741088   0.09781109
#> lag1_var2  0.92208485 0.4533037  0.55251680   1.47406069
#> lag1_var3 -0.02435489 0.6720675 -0.61335935   0.78009446
#> lag1_var4  0.08362120 0.6104583 -0.65447783   0.62738389
#> const     -2.19110709 3.0393514 -4.95341436   1.74878826
#> exo1       0.10367647 0.3046760 -0.24637945   0.38490178
#> exo2       0.20870837 0.2073282  0.02149222   0.47728870
#> exo3      -0.68541548 0.3614132 -1.12365068  -0.30525437
#> 
#> $AGO$A$equation3
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.118333252 0.07535843 -0.19802568   -0.0263322
#> lag1_var2  0.059513648 0.12282954 -0.08077065    0.1966078
#> lag1_var3  1.099441692 0.22011185  0.86307550    1.3325001
#> lag1_var4 -0.058376763 0.19533566 -0.25530738    0.1574467
#> const     -0.003724951 2.33007433 -2.54881254    2.7279690
#> exo1       0.016116014 0.10514281 -0.09167347    0.1487756
#> exo2      -1.263687718 0.17244200 -1.42356433   -1.0375192
#> exo3       3.479132082 0.30871232  3.17124694    3.8765457
#> 
#> $AGO$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.2958795 0.04401515 -0.33653609   -0.2465178
#> lag1_var2  0.6666459 0.24243041  0.43480647    0.9286137
#> lag1_var3  0.9445705 0.33775311  0.61657254    1.3100332
#> lag1_var4  0.2090267 0.30816088 -0.13189772    0.5135257
#> const     -3.6823540 1.76807682 -5.89714299   -2.0396585
#> exo1       0.1777847 0.20863834 -0.09181932    0.3529759
#> exo2      -1.4566803 0.14085170 -1.58217125   -1.2829555
#> exo3       3.2448404 0.13223166  3.07178934    3.3654303
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003220013 0.0007663079 0.002224907  0.003966237
#> 
#> $ALB$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.1031385 0.05721412  -0.1784557  -0.05182315
#> Sigma[2,2]  9.7954077 4.58359189   7.1582838  16.06017348
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04001786 0.02127952   0.0133795   0.06234073
#> Sigma[3,2] -5.07822839 1.43854948  -6.6175738  -3.35747266
#> Sigma[3,3]  4.65285517 1.15837087   3.6453398   6.18483803
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.03306923 0.02625541 -0.06790702  -0.01018509
#> Sigma[4,2]  1.66474761 2.64207700 -0.89751568   5.08595814
#> Sigma[4,3]  1.54067189 1.05236960  0.32828500   2.76104713
#> Sigma[4,4]  3.16327539 1.09909378  2.23636709   4.57009370
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.982529837 0.03610094  0.94595181  1.024487957
#> lag1_var2 -0.022559961 0.01245883 -0.03234333 -0.006354767
#> lag1_var3 -0.045558122 0.02353476 -0.06344632 -0.014562249
#> lag1_var4  0.036947558 0.02124108  0.01079009  0.055273661
#> const      0.876039376 0.94131041 -0.34275265  1.771985210
#> exo1      -0.008667986 0.08808081 -0.09741375  0.100178425
#> exo2      -0.057481479 0.03933594 -0.09869363 -0.009746337
#> exo3       0.050054917 0.06837247 -0.04380067  0.093253666
#> 
#> $ALB$A$equation2
#>                mean        sd 5% quantile 95% quantile
#> lag1_var1 -2.376718  1.573622  -4.4704925   -1.0491806
#> lag1_var2  2.160216  2.536642  -0.3951972    5.3624615
#> lag1_var3  3.052114  4.096987  -0.8462533    8.3226750
#> lag1_var4 -2.337763  3.533201  -6.8884149    0.9631566
#> const     22.718616 23.483935  -3.7519690   45.1506515
#> exo1       2.496050  4.041359  -3.0281679    5.3186984
#> exo2      -1.336709  2.623826  -4.6645185    1.3693381
#> exo3      -3.713461  2.889683  -5.9556318    0.1654766
#> 
#> $ALB$A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  1.8968738  0.8561586   0.9323273    2.6974347
#> lag1_var2 -1.5572048  1.3115369  -3.2699683   -0.3190992
#> lag1_var3 -1.9440499  2.1953356  -4.7839925    0.1647109
#> lag1_var4  2.2499491  1.8520992   0.5739813    4.6901690
#> const     -5.4966524 21.5088863 -34.1788126   13.5573914
#> exo1      -3.1289671  2.0830517  -4.4990753   -0.2524282
#> exo2      -0.2054825  1.3065198  -1.7224333    1.2521119
#> exo3       2.5629175  1.5642526   0.6136848    4.2075103
#> 
#> $ALB$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.2188089  1.000383  -0.6448058    1.5565215
#> lag1_var2 -0.6021670  1.646663  -2.7603255    0.9652267
#> lag1_var3 -0.5497206  2.579468  -3.9455167    1.8777523
#> lag1_var4  1.3810874  2.131763  -0.6585945    4.1731722
#> const      9.3875885 14.377966 -10.1409590   20.4246825
#> exo1      -1.2525111  1.035056  -2.4548728   -0.1582065
#> exo2      -1.1694825  1.155438  -2.4050380    0.2577840
#> exo3       0.1446453  1.344035  -1.2077866    1.7540372
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.008006713 0.001075222 0.007166562  0.009438778
#> 
#> $ARE$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01405668 0.002712948 -0.01733118  -0.01114424
#> Sigma[2,2]  0.07422787 0.018453332  0.05482486   0.09694059
#> 
#> $ARE$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.01446575 0.00370125  0.01081610   0.01897964
#> Sigma[3,2] -0.06698413 0.01802788 -0.09126308  -0.05211083
#> Sigma[3,3]  0.16271248 0.03289983  0.13285631   0.20530783
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.002878832 0.002840658  0.001209098  0.006723946
#> Sigma[4,2] -0.008371153 0.005925423 -0.016279419 -0.003110708
#> Sigma[4,3]  0.113253941 0.026217716  0.086078343  0.141328240
#> Sigma[4,4]  0.110698672 0.024914536  0.085978610  0.140659711
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd 5% quantile  95% quantile
#> lag1_var1  0.54719220 0.03922527   0.4961462  5.830390e-01
#> lag1_var2 -0.09731108 0.10165145  -0.2250315  4.126214e-06
#> lag1_var3 -0.39419675 0.13020654  -0.5496405 -2.504337e-01
#> lag1_var4  0.49975892 0.12872710   0.3510214  6.453392e-01
#> const      3.37099603 0.73290826   2.6865517  4.284593e+00
#> exo1      -0.15415165 0.06316288  -0.2302740 -8.373160e-02
#> exo2      -0.19135432 0.10471470  -0.3287303 -9.842884e-02
#> exo3      -0.52040148 0.17931034  -0.7138871 -3.311986e-01
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.70227597 0.1561361  0.54530790    0.8638476
#> lag1_var2  0.43049158 0.7241349 -0.25034604    1.3755875
#> lag1_var3 -0.11927446 0.9198232 -1.00137018    1.0658657
#> lag1_var4 -0.08248108 0.9050242 -1.24581935    0.8064890
#> const     -1.85538019 1.0439151 -3.00768775   -0.7605070
#> exo1       0.26832339 0.1549479  0.05496396    0.3728936
#> exo2       2.21181680 0.2944986  1.88400989    2.5640459
#> exo3      -0.04310080 0.2209698 -0.25831441    0.1897489
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.1355042 0.3502522 -0.59034473    0.1906565
#> lag1_var2  1.0345763 1.0627565 -0.40882051    1.8542422
#> lag1_var3  1.8878473 1.4923044 -0.09066198    3.2571686
#> lag1_var4 -0.7685564 1.4847908 -2.17529981    1.1735764
#> const     -6.2248843 1.6930218 -8.16381328   -4.2805972
#> exo1      -0.6150675 0.2426734 -0.86054237   -0.3347867
#> exo2      -2.6160375 0.2651906 -2.93023762   -2.3733897
#> exo3      -1.6597857 0.8866492 -2.58470491   -0.6979405
#> 
#> $ARE$A$equation4
#>                 mean        sd  5% quantile 95% quantile
#> lag1_var1  0.4525611 0.2894070   0.06883672   0.69065148
#> lag1_var2  0.9455538 0.5746254   0.15838219   1.35992723
#> lag1_var3  1.2842589 0.8659445   0.15536935   2.10502780
#> lag1_var4 -0.3218738 0.8474254  -1.12433238   0.79164351
#> const     -8.7161644 2.4538756 -11.95261469  -6.62822972
#> exo1      -0.3981649 0.2673110  -0.66537767  -0.05423267
#> exo2      -0.8754942 0.2919638  -1.27066756  -0.65795493
#> exo3      -1.6738418 0.7437716  -2.41503386  -0.85226940
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004040003 0.001995452 0.001890246  0.006277895
#> 
#> $ARG$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.1005412 0.05813542  -0.1693003  -0.04297943
#> Sigma[2,2]  4.7065426 1.91306802   2.5431121   6.60960318
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.06874628 0.03542964  0.02856218    0.1072828
#> Sigma[3,2] -3.02859447 1.34566332 -4.36035908   -1.4428496
#> Sigma[3,3]  2.28600567 0.84497846  1.24669136    3.1959523
#> 
#> $ARG$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.008872702 0.006757764  0.0007963885   0.01416502
#> Sigma[4,2] -0.188641839 0.502579371 -0.6466383722   0.46898018
#> Sigma[4,3]  0.495160372 0.234235534  0.2525038092   0.75696637
#> Sigma[4,4]  0.423189957 0.197395917  0.2775751850   0.68971601
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.93446887 0.03052685  0.899407047   0.96625771
#> lag1_var2  0.13653505 0.10497669  0.007840467   0.22637607
#> lag1_var3  0.21721827 0.16710238  0.013300176   0.35144870
#> lag1_var4 -0.20501112 0.15243667 -0.334940222  -0.01767761
#> const      0.99805335 1.20604790 -0.383390868   2.39896046
#> exo1       0.05620785 0.10444201 -0.033298545   0.19435746
#> exo2      -0.13236489 0.04630903 -0.181184741  -0.08372834
#> exo3      -0.01721234 0.10813589 -0.133438016   0.10711737
#> 
#> $ARG$A$equation2
#>                mean        sd   5% quantile 95% quantile
#> lag1_var1 -1.343672  1.575150  -3.049144709    0.4579622
#> lag1_var2 -1.635675  4.064375  -6.973758414    1.5926258
#> lag1_var3 -3.477039  6.617685 -11.996432487    1.9672029
#> lag1_var4  4.025590  5.898722  -0.576546619   11.8203050
#> const      8.218783 51.744407 -49.399114876   70.2521521
#> exo1      -1.646230  3.246915  -5.753026041    1.5935514
#> exo2       1.885639  2.136025  -0.649701072    4.2932698
#> exo3       2.685759  3.151416   0.007593158    6.9048460
#> 
#> $ARG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.12025611  0.6019306   0.3992259     1.719953
#> lag1_var2  1.36289912  3.2116107  -1.2119955     5.734172
#> lag1_var3  2.65140169  5.0822892  -1.3534759     9.577403
#> lag1_var4 -2.21912209  4.6818105  -8.5784676     1.578446
#> const      0.64100479 40.1123874 -50.5064307    41.310000
#> exo1       0.06928127  2.0480723  -2.1836859     2.505282
#> exo2      -4.55888376  0.6214356  -5.2396838    -3.892883
#> exo3       1.06040036  1.7105994  -1.2589960     2.364696
#> 
#> $ARG$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.27903748  0.7911610  -0.3058657    1.3593969
#> lag1_var2 -0.01196834  2.2420502  -2.8502784    1.9107621
#> lag1_var3 -0.09697838  3.7045414  -4.7936276    3.0566589
#> lag1_var4  0.86960282  3.2366440  -1.9056763    4.9388488
#> const      5.82549996 11.3838505  -8.4912708   15.1614079
#> exo1      -1.00976635  0.4242919  -1.3543483   -0.4353853
#> exo2      -3.90571478  0.9855937  -5.0277164   -2.7509778
#> exo3       3.00698385  0.5531554   2.3121989    3.5796176
#> 
#> 
#>