Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004232954 0.000966797 0.003353157    0.0054807
#> 
#> $AFG$Sigma$equation2
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.006629874 0.003992462 -0.01180318 -0.003520965
#> Sigma[2,2]  0.220257735 0.038925661  0.17322789  0.258340171
#> 
#> $AFG$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.004659605 0.005608575 -0.001329294   0.01017382
#> Sigma[3,2] -0.214203789 0.052478931 -0.285312800  -0.17501068
#> Sigma[3,3]  0.378196428 0.080877469  0.283697240   0.47015263
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.002235555 0.004889763 -0.003239446  0.005972469
#> Sigma[4,2] -0.132563621 0.056374300 -0.204074232 -0.074923363
#> Sigma[4,3]  0.323105975 0.068901846  0.239005440  0.393502997
#> Sigma[4,4]  0.302972632 0.054575133  0.238031306  0.359765624
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.00880153 0.02067212  0.98247921   1.02933085
#> lag1_var2  0.20956346 0.02132983  0.18437496   0.23043744
#> lag1_var3  0.50691590 0.02082773  0.48084090   0.52581458
#> lag1_var4 -0.44354009 0.01116032 -0.45187536  -0.42833854
#> const     -2.86289760 0.85177337 -3.75822154  -1.98236420
#> exo1      -0.08546011 0.05618638 -0.15881822  -0.03335575
#> exo2      -0.00723910 0.03806863 -0.05223806   0.03607330
#> exo3      -0.07278702 0.05578793 -0.14357023  -0.02943606
#> 
#> $AFG$A$equation2
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1   0.31155229 0.13565032   0.2061070    0.4922934
#> lag1_var2   1.40420472 0.09779474   1.3080117    1.5288049
#> lag1_var3   0.98074810 0.14777336   0.8285880    1.1640002
#> lag1_var4  -0.86473399 0.18878644  -1.0492274   -0.6268511
#> const     -11.95782340 5.29002397 -17.0967316   -5.1840410
#> exo1        0.01403880 0.41596840  -0.4395134    0.5017914
#> exo2        0.54816414 0.16670425   0.3336705    0.6896316
#> exo3        0.01928125 0.54764866  -0.6078092    0.5747031
#> 
#> $AFG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.15620599 0.17381983  -0.3401404   0.05366126
#> lag1_var2 -0.77211077 0.17729179  -0.9362780  -0.55870170
#> lag1_var3 -0.01475852 0.18361808  -0.2272328   0.18057123
#> lag1_var4  0.74533601 0.09569207   0.6562370   0.86442455
#> const     18.44646369 5.90275334  14.3389970  26.57308737
#> exo1       0.28873588 0.38292775  -0.1945949   0.64826112
#> exo2      -1.60460945 0.47313532  -2.1938308  -1.09833331
#> exo3      -0.14074885 0.39110237  -0.6042225   0.27408873
#> 
#> $AFG$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.03158774 0.1595088 -0.19192252    0.1666137
#> lag1_var2 -0.12008270 0.1862149 -0.31492394    0.1062172
#> lag1_var3  0.52000665 0.2722152  0.19764813    0.7809345
#> lag1_var4  0.39833885 0.1991837  0.21926675    0.6235211
#> const      7.40675555 5.1390568  3.51004605   14.3841759
#> exo1       0.32530557 0.2998198 -0.08188847    0.5666531
#> exo2      -1.46419932 0.5757935 -2.10486858   -0.7817645
#> exo3      -0.00676524 0.3816484 -0.36516069    0.4861587
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005846478 0.001477151  0.00427164  0.007679723
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01148401 0.003278871 -0.01449948 -0.007771107
#> Sigma[2,2]  0.12229095 0.016972305  0.10172068  0.137303868
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.006371927 0.001648927  0.004323361  0.007986745
#> Sigma[3,2] -0.049605204 0.011208388 -0.059729575 -0.037377121
#> Sigma[3,3]  0.036186342 0.005392066  0.030392539  0.041727427
#> 
#> $AGO$Sigma$equation4
#>                    mean           sd  5% quantile 95% quantile
#> Sigma[4,1] -0.002123207 0.0007899135 -0.002824926 -0.001144526
#> Sigma[4,2]  0.048965757 0.0053198898  0.042570307  0.054283582
#> Sigma[4,3] -0.003519842 0.0042290809 -0.008758357  0.000564933
#> Sigma[4,4]  0.038071784 0.0082276274  0.029245531  0.047166990
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.99246858 0.01624050  0.97363248   1.00906155
#> lag1_var2  0.13041334 0.05636541  0.07880012   0.20502919
#> lag1_var3  0.21398096 0.08164263  0.14081363   0.31616529
#> lag1_var4 -0.15340959 0.07887210 -0.25644186  -0.07753709
#> const     -3.86982789 1.72656975 -5.96859848  -1.96583562
#> exo1       0.06616483 0.11742436 -0.04856256   0.21203172
#> exo2      -0.02150829 0.10104269 -0.14663618   0.08610503
#> exo3       0.10320812 0.15422682 -0.05550485   0.27345735
#> 
#> $AGO$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.02261915 0.1405381  -0.1709478  0.116788703
#> lag1_var2  1.18643775 0.1784142   0.9652116  1.378513479
#> lag1_var3  0.29291633 0.1031417   0.1813396  0.404051675
#> lag1_var4 -0.25773273 0.2108832  -0.4933049 -0.005696921
#> const     -1.62090748 9.2736656 -12.1168736  9.323718378
#> exo1       0.21124540 0.4252712  -0.2933854  0.636115835
#> exo2       0.36033919 0.1869668   0.1496323  0.572662963
#> exo3      -1.10390268 0.3100952  -1.3080519 -0.681292772
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.07818203 0.10702754  -0.1558467   0.06872809
#> lag1_var2 -1.59421713 0.09591086  -1.7170423  -1.50462309
#> lag1_var3 -1.10583630 0.05414757  -1.1795528  -1.06265810
#> lag1_var4  1.84420261 0.09386666   1.7486609   1.96204399
#> const     21.58773373 6.13710687  13.4133710  27.01584749
#> exo1       0.01021499 0.32334944  -0.3221382   0.42150868
#> exo2      -1.46771685 0.12925489  -1.6001187  -1.30691496
#> exo3       3.75396125 0.13860582   3.6317405   3.93561928
#> 
#> $AGO$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.1898715 0.09340408 -0.31724286   -0.1417826
#> lag1_var2 -0.8255796 0.09446231 -0.95050411   -0.7395234
#> lag1_var3 -1.0340815 0.04774061 -1.09149414   -0.9827087
#> lag1_var4  1.8979863 0.09641162  1.81979719    2.0278104
#> const     15.4958729 5.58287378 10.07500854   22.3585199
#> exo1       0.2297004 0.12814680  0.06747382    0.3626686
#> exo2      -1.5004844 0.13912575 -1.62697140   -1.3240944
#> exo3       3.3176827 0.23547547  3.10316592    3.6183171
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002220014 0.0005292572 0.001718392  0.002914797
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.04859899 0.01729382 -0.06925061  -0.03175583
#> Sigma[2,2]  7.25999107 1.78565773  5.33143520   9.27969534
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01851194 0.01806898  0.002050085   0.03920532
#> Sigma[3,2] -4.78588645 1.21466535 -6.185899032  -3.42229696
#> Sigma[3,3]  5.11197192 1.24826535  3.523175018   6.38494017
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.01729684 0.01142787 -0.03090617 -0.004539952
#> Sigma[4,2]  0.17157165 0.49958558 -0.42433893  0.672321992
#> Sigma[4,3]  2.19840454 0.77160504  1.27890579  2.978583917
#> Sigma[4,4]  2.74911275 0.52103358  2.09547604  3.235675278
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1  1.021449390 0.04302440  0.966566964   1.06387894
#> lag1_var2  0.011297227 0.02807340 -0.013977681   0.04798720
#> lag1_var3  0.004522326 0.04380548 -0.032968422   0.06319453
#> lag1_var4 -0.003053709 0.03659958 -0.051723567   0.02981103
#> const     -0.675484247 1.17960305 -1.756655560   0.87783076
#> exo1       0.052194190 0.04389111  0.003915011   0.10578571
#> exo2      -0.046410027 0.06369030 -0.118456234   0.01033567
#> exo3       0.095263983 0.06527251  0.013907359   0.16273797
#> 
#> $ALB$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -3.7962609  1.9582676   -5.924386    -1.461065
#> lag1_var2  4.2683971  0.8586256    3.199526     5.155942
#> lag1_var3  6.1455192  1.4881173    4.273258     7.592057
#> lag1_var4 -5.1109223  1.1597184   -6.141916    -3.657459
#> const     32.1241199 44.8140633  -20.820459    79.553152
#> exo1      -4.0518656  1.9644380   -6.482445    -2.123151
#> exo2      -0.9472775  4.8960307   -6.875325     3.513432
#> exo3      -2.7435183  3.5652636   -6.768986     1.015443
#> 
#> $ALB$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  2.4130259  2.187891  -0.3730198     4.603239
#> lag1_var2 -3.0532054  1.639806  -5.2135822    -1.664273
#> lag1_var3 -4.2254274  2.455983  -7.4666243    -2.001036
#> lag1_var4  4.2539024  1.987440   2.4502039     6.878771
#> const      1.1148878 58.923261 -66.6366489    61.158335
#> exo1       0.8639327  1.452058  -0.9218047     2.391253
#> exo2       0.3648201  4.758983  -5.0895714     5.814052
#> exo3       1.8495638  3.216357  -1.1026877     5.798756
#> 
#> $ALB$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.0907819  1.064184  -1.5121679   0.82456511
#> lag1_var2 -0.6638136  1.601957  -2.5854521   0.82338849
#> lag1_var3 -0.7503035  2.313469  -3.6089513   1.32438862
#> lag1_var4  1.5167457  1.994035  -0.2458111   4.00371126
#> const     19.2832049 36.006043 -24.2841108  50.84634044
#> exo1      -1.7175340  1.365310  -2.8237925   0.08151951
#> exo2      -0.2871083  2.305896  -3.3785435   1.52589955
#> exo3       0.0699801  1.776310  -2.2971862   1.62828973
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004158155 0.005962055  0.00125151   0.01223203
#> 
#> $ARE$Sigma$equation2
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[2,1] 0.001703029 0.008959173 -0.003452234   0.01382904
#> Sigma[2,2] 0.070368617 0.011294403  0.055618030   0.08114027
#> 
#> $ARE$Sigma$equation3
#>                    mean         sd  5% quantile 95% quantile
#> Sigma[3,1]  0.007155004 0.01249788 -0.001637779   0.02395488
#> Sigma[3,2] -0.058003959 0.02264224 -0.081556775  -0.02964106
#> Sigma[3,3]  0.236867566 0.06978173  0.159295397   0.31759847
#> 
#> $ARE$Sigma$equation4
#>                    mean         sd  5% quantile 95% quantile
#> Sigma[4,1]  0.008563036 0.01894863 -0.002916205   0.03432142
#> Sigma[4,2] -0.004393354 0.02400602 -0.027072369   0.02685320
#> Sigma[4,3]  0.195478214 0.06934619  0.117478320   0.26518618
#> Sigma[4,4]  0.195373821 0.07206718  0.117780571   0.27605903
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.98194262 0.01546840   0.9637546  0.999603180
#> lag1_var2  0.31155230 0.05821203   0.2422722  0.376480128
#> lag1_var3  0.40006402 0.09796175   0.3011488  0.522285221
#> lag1_var4 -0.38564979 0.10180113  -0.5099780 -0.278535025
#> const     -0.60171240 0.61402234  -1.3944652  0.005213566
#> exo1       0.02456789 0.11521332  -0.1011461  0.161426634
#> exo2      -0.12258143 0.05498536  -0.1952258 -0.074217604
#> exo3      -0.04480658 0.08222926  -0.1504963  0.034550067
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.18769827 0.2365929 -0.09784815   0.45281485
#> lag1_var2  0.97754410 0.3849640  0.59057541   1.47151190
#> lag1_var3  0.38489454 0.3106196  0.05912334   0.77361107
#> lag1_var4 -0.43873629 0.2318745 -0.70536069  -0.17438913
#> const     -0.07667258 3.4461894 -3.61585831   3.70992916
#> exo1       0.17269363 0.2572252 -0.08438628   0.48597541
#> exo2       2.01810084 0.3293596  1.59872962   2.34940262
#> exo3      -0.51946518 0.4460523 -1.10300577  -0.09343955
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.1048300 0.4410076  -0.3537318    0.5570740
#> lag1_var2  1.6303685 1.3519871  -0.2200674    2.3678738
#> lag1_var3  2.9467993 1.6779688   0.6450543    3.9786075
#> lag1_var4 -1.9000509 1.6192725  -2.8456028    0.3316572
#> const     -6.4984354 7.2526063 -14.4653769   -0.2312628
#> exo1      -0.5003440 0.5205927  -0.9077309    0.2017745
#> exo2      -2.5250079 0.3485281  -2.9369411   -2.1435040
#> exo3      -0.9242671 0.5106179  -1.4101268   -0.3356368
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.2287495 0.3932242  -0.1339817    0.6701308
#> lag1_var2  2.5029462 1.0761494   1.0237038    3.2567021
#> lag1_var3  3.4024034 1.4536567   1.4107566    4.4494702
#> lag1_var4 -2.3637512 1.4596260  -3.3446726   -0.3510149
#> const     -8.5726822 4.8901495 -14.2071310   -4.6114537
#> exo1      -0.3843501 0.4319522  -0.7896236    0.1826942
#> exo2      -0.9601286 0.3708963  -1.3064888   -0.4739824
#> exo3      -1.3491285 0.2522863  -1.5239323   -1.0059864
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                  mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00279953 0.0003533226 0.002394804   0.00320466
#> 
#> $ARG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.04775892 0.009703929 -0.05817577  -0.03591438
#> Sigma[2,2]  2.89639876 0.269171689  2.55230252   3.16920661
#> 
#> $ARG$Sigma$equation3
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.03772535 0.008825793  0.02830673   0.04872797
#> Sigma[3,2] -2.04966909 0.218890003 -2.31120099  -1.80720244
#> Sigma[3,3]  1.64401694 0.213105688  1.44563258   1.91607007
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01018772 0.004147628  0.006958102   0.01571664
#> Sigma[4,2] -0.31926953 0.105771801 -0.416489565  -0.17919225
#> Sigma[4,3]  0.44050757 0.101615684  0.329059788   0.56007281
#> Sigma[4,4]  0.27418581 0.055966460  0.234315154   0.35086946
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.910427471 0.03969125  0.86427248   0.95617914
#> lag1_var2  0.418687086 0.07847949  0.35636556   0.52520372
#> lag1_var3  0.689599988 0.12685895  0.58763616   0.86150460
#> lag1_var4 -0.624821050 0.11411999 -0.77975077  -0.53829085
#> const     -1.488377255 1.91544174 -3.76499838   0.67612207
#> exo1       0.046454819 0.08002361 -0.01936504   0.15231232
#> exo2      -0.106029137 0.06212702 -0.16952604  -0.03172087
#> exo3      -0.008454463 0.09098645 -0.13358983   0.05734916
#> 
#> $ARG$A$equation2
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1 -2.10919435  0.3271426  -2.40694562   -1.7291099
#> lag1_var2  0.69137023  0.6391534  -0.16539673    1.1977926
#> lag1_var3 -0.06724173  1.0929100  -1.52189127    0.8409837
#> lag1_var4  0.56580091  0.8541836  -0.26570572    1.6509727
#> const     29.23037836 32.2724055 -10.61305130   59.4530207
#> exo1      -0.26171210  1.6562168  -2.06562184    1.7156470
#> exo2       2.19870580  2.6669109  -0.06385814    5.5509489
#> exo3       0.43425442  1.9831855  -2.15809651    2.2929872
#> 
#> $ARG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.87456158  0.3866224   0.3813624    1.2300987
#> lag1_var2 -0.53768614  0.2885809  -0.9301708   -0.3065922
#> lag1_var3 -0.14363463  0.4838162  -0.7847637    0.2962102
#> lag1_var4  0.57503642  0.4137422   0.1531759    1.0909955
#> const      9.46691760 22.8937942 -16.0503734   32.4496092
#> exo1      -0.02999327  1.3974207  -1.6311533    1.2983836
#> exo2      -4.98191358  1.7527472  -7.1772524   -3.4193125
#> exo3       2.84031524  0.9196923   1.8411839    3.9526535
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.4996812  0.3056719  -0.8868442   -0.2391698
#> lag1_var2 -0.1505878  0.3052119  -0.4916470    0.1535629
#> lag1_var3 -0.2407136  0.4844395  -0.7692263    0.2505386
#> lag1_var4  1.0545084  0.4667052   0.5622386    1.5762801
#> const     24.9095878 11.3541332  15.1430008   39.0967647
#> exo1      -0.2288304  0.6067494  -0.9355783    0.4583882
#> exo2      -4.1067150  0.1702820  -4.2793186   -3.8980593
#> exo3       3.4604966  0.4726127   2.8721269    3.9654830
#> 
#> 
#>