Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004861271 0.0008419931 0.004220449  0.006013568
#> 
#> $AFG$Sigma$equation2
#>                  mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0113474 0.003490168 -0.01395458 -0.006607077
#> Sigma[2,2]  0.1809811 0.040882774  0.14435617  0.226060533
#> 
#> $AFG$Sigma$equation3
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.01376882 0.002146057  0.01103135   0.01588867
#> Sigma[3,2] -0.16776317 0.057618756 -0.23745503  -0.10877878
#> Sigma[3,3]  0.32157517 0.083391970  0.24488240   0.42370010
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.009870103 0.001298544  0.008714387   0.01160448
#> Sigma[4,2] -0.101171024 0.046586110 -0.158631114  -0.05189822
#> Sigma[4,3]  0.284330596 0.076060704  0.217032247   0.37086128
#> Sigma[4,4]  0.276353987 0.075420119  0.212794502   0.36611340
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.99615904 0.02033046   0.9703243   1.01426956
#> lag1_var2 -0.25980733 0.10767143  -0.3548208  -0.11653013
#> lag1_var3 -0.53018955 0.23932891  -0.7170585  -0.20647661
#> lag1_var4  0.45563618 0.20941207   0.1726392   0.61785274
#> const      3.73906769 1.70636107   1.4810585   5.06783999
#> exo1      -0.07390283 0.08742581  -0.1757829   0.02644380
#> exo2      -0.12602147 0.02788051  -0.1493959  -0.09061277
#> exo3      -0.22735749 0.06411610  -0.3059275  -0.15871263
#> 
#> $AFG$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.31827758 0.07252014   0.2233260   0.38480940
#> lag1_var2  0.74457735 0.23924317   0.5225429   1.05267677
#> lag1_var3 -0.52652953 0.30573555  -0.8216848  -0.13265101
#> lag1_var4  0.48401040 0.19383476   0.2294609   0.66577346
#> const     -5.10573520 5.52002351 -12.3501795  -0.07615631
#> exo1      -0.08667923 0.79716889  -0.7537454   0.98284356
#> exo2       0.11669061 0.39820398  -0.3300851   0.57169539
#> exo3       0.07594360 0.55915439  -0.3981175   0.80953347
#> 
#> $AFG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.11339061 0.09020093  -0.2076870  -0.01592052
#> lag1_var2  0.46768255 0.69615407   0.0303735   1.41858623
#> lag1_var3  2.62165485 1.41601264   1.7881524   4.56339742
#> lag1_var4 -1.55519457 1.22233896  -3.2315537  -0.77199606
#> const      1.69585788 9.11121005 -10.5787498   7.87746017
#> exo1      -0.10974861 0.80149191  -0.9802727   0.71521491
#> exo2      -1.51525704 0.41799585  -2.0572422  -1.11694084
#> exo3      -0.08318526 0.49713673  -0.5427058   0.49087281
#> 
#> $AFG$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.02800276 0.1132704  -0.0933404    0.1423777
#> lag1_var2  0.91319186 0.7224242   0.4782464    1.9055666
#> lag1_var3  2.66155385 1.4686657   1.6866396    4.6610670
#> lag1_var4 -1.45322005 1.2613356  -3.1656756   -0.5747169
#> const     -7.79998805 8.9271176 -19.9425458   -1.1699755
#> exo1      -0.16004045 0.6324784  -0.9076296    0.5605160
#> exo2      -1.55274725 0.3200364  -1.9675406   -1.2698050
#> exo3       0.05602435 0.5106893  -0.5057555    0.5587436
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                  mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00584477 0.001716105 0.003826045  0.007652538
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01485569 0.005608685 -0.02095822 -0.009124636
#> Sigma[2,2]  0.14111684 0.016803757  0.12061494  0.159405923
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.009721154 0.004803191  0.005125669   0.01488511
#> Sigma[3,2] -0.061576017 0.017551391 -0.083773522  -0.04368033
#> Sigma[3,3]  0.045262798 0.013611551  0.031833550   0.06127818
#> 
#> $AGO$Sigma$equation4
#>                    mean           sd 5% quantile  95% quantile
#> Sigma[4,1] -0.001243733 0.0009016272 -0.00233148 -0.0004696163
#> Sigma[4,2]  0.052621171 0.0106736744  0.03946292  0.0637550495
#> Sigma[4,3] -0.003773280 0.0066638630 -0.01121903  0.0031002449
#> Sigma[4,4]  0.041468548 0.0098229233  0.03253519  0.0523546554
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd   5% quantile 95% quantile
#> lag1_var1  0.97343157 0.01890202  0.9509991790   0.99418532
#> lag1_var2 -0.03704828 0.02364421 -0.0688610849  -0.01861540
#> lag1_var3 -0.01919012 0.03778512 -0.0661220220   0.02107220
#> lag1_var4  0.02978545 0.04628133 -0.0005293185   0.09374692
#> const      0.24714385 1.62099730 -1.6162126348   1.89023209
#> exo1       0.02734153 0.05002980 -0.0402698519   0.06465440
#> exo2      -0.08315019 0.02889306 -0.1160646735  -0.05114968
#> exo3       0.03064876 0.15944473 -0.1191102733   0.21799058
#> 
#> $AGO$A$equation2
#>                  mean        sd   5% quantile 95% quantile
#> lag1_var1  0.11841249 0.1546898  -0.032266545    0.2978868
#> lag1_var2  0.17391320 0.1529669   0.056876992    0.3821674
#> lag1_var3 -0.94974731 0.1131147  -1.072449453   -0.8132812
#> lag1_var4  0.93659282 0.1634149   0.735181179    1.1110403
#> const     -0.36528689 8.0151257 -10.712539387    7.2173970
#> exo1       0.04510982 0.2976763  -0.314455289    0.3695197
#> exo2       0.33222192 0.4402528   0.008147211    0.9227800
#> exo3      -0.59282345 0.7169641  -1.508082036    0.1330372
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.19921963 0.06028304  -0.2486690  -0.11848591
#> lag1_var2 -0.15506311 0.06477135  -0.2335148  -0.09132766
#> lag1_var3  0.75382752 0.05656764   0.6937751   0.81859048
#> lag1_var4  0.19476918 0.05526994   0.1265521   0.25063362
#> const      8.23138485 6.64332081   2.9081714  17.03099712
#> exo1      -0.02970332 0.13468459  -0.2064199   0.09457448
#> exo2      -1.30173470 0.23372996  -1.6172230  -1.13391149
#> exo3       3.32890250 0.33462040   3.0006838   3.75577460
#> 
#> $AGO$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.20717801 0.10051979 -0.32180291  -0.10107287
#> lag1_var2 -0.02019676 0.07098747 -0.06742665   0.07657614
#> lag1_var3  0.03380409 0.08580643 -0.06841901   0.11901425
#> lag1_var4  1.00968245 0.09561970  0.88691773   1.09847830
#> const      2.41218826 2.24551556  0.02882593   4.87674361
#> exo1       0.02852620 0.13880560 -0.15448393   0.14968624
#> exo2      -1.37117648 0.13943027 -1.50948902  -1.19284142
#> exo3       3.18511502 0.31635313  2.81737810   3.54323372
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd  5% quantile 95% quantile
#> Sigma[1,1] 0.001469439 0.0003908327 0.0009531118  0.001816472
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.04660977 0.02201802 -0.06889656  -0.01871738
#> Sigma[2,2]  8.48423312 2.33721081  5.73013156  10.66461310
#> 
#> $ALB$Sigma$equation3
#>                  mean        sd  5% quantile 95% quantile
#> Sigma[3,1]  0.0187138 0.0183255  0.002504224    0.0419941
#> Sigma[3,2] -5.2862259 1.5117720 -6.749242357   -3.5722802
#> Sigma[3,3]  5.7721180 1.8906700  3.943905435    8.1477051
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.01556523 0.01450015 -0.02850088  0.002151055
#> Sigma[4,2]  0.54315693 1.78109363 -1.32862486  2.765822554
#> Sigma[4,3]  2.64524462 1.67427436  1.55752268  4.938842822
#> Sigma[4,4]  3.64102840 1.49741835  2.20821220  5.302936355
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  1.002175674 0.01891468  0.97678366  1.018428938
#> lag1_var2 -0.016817890 0.01415888 -0.02989062 -0.001232021
#> lag1_var3 -0.035773832 0.02143500 -0.05786238 -0.013260369
#> lag1_var4  0.029809248 0.01633381  0.01228453  0.045283503
#> const      0.271097934 0.69848078 -0.40800206  1.079989817
#> exo1       0.009625907 0.03115559 -0.02407717  0.045162272
#> exo2      -0.051782652 0.03271247 -0.09432262 -0.022410150
#> exo3       0.060589154 0.01876896  0.04103585  0.084222628
#> 
#> $ALB$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -4.3279314  2.051063  -6.8380427  -2.62792428
#> lag1_var2  1.6706067  1.199119   0.1024038   2.78464782
#> lag1_var3  2.1052749  1.701681  -0.1374066   3.59246957
#> lag1_var4 -1.8008320  1.329474  -2.7814178  -0.02343054
#> const     89.9252210 73.356513  16.5260862 169.81410689
#> exo1      -1.0512375  4.608200  -6.7362167   3.22676295
#> exo2       0.7003319  2.043580  -1.5636123   2.99403958
#> exo3      -1.9228880  2.039392  -3.9861520   0.67911007
#> 
#> $ALB$A$equation3
#>                   mean        sd  5% quantile 95% quantile
#> lag1_var1   2.80348715  2.745979   -0.6772913     5.347260
#> lag1_var2  -0.26427000  1.335347   -1.5527798     1.391299
#> lag1_var3   0.05789407  2.228277   -2.0954793     2.687767
#> lag1_var4   0.71620713  1.965964   -1.4958692     2.659294
#> const     -54.98979856 81.633026 -132.8212427    38.577518
#> exo1       -1.60807767  3.751988   -5.7045903     2.606025
#> exo2       -1.13416932  2.429379   -3.9212993     1.339754
#> exo3        0.69891307  1.490017   -0.9230136     2.448310
#> 
#> $ALB$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.13317695  2.2085040  -2.8359897    2.2602111
#> lag1_var2  0.55948674  1.3416358  -0.7598114    2.2342102
#> lag1_var3  1.08666873  2.3712530  -1.1201044    4.1202219
#> lag1_var4 -0.02629705  2.0681150  -2.6684814    1.9913510
#> const      1.16616673 50.8809130 -50.0267722   57.6521281
#> exo1      -2.24611030  1.7033948  -4.0288543   -0.1520324
#> exo2      -0.71714727  2.6339611  -3.5166739    2.4075273
#> exo3      -0.63501033  0.5493846  -1.3886481   -0.2311513
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                  mean           sd  5% quantile 95% quantile
#> Sigma[1,1] 0.00123436 0.0001961898 0.0009807935  0.001428883
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.001991428 0.001413011 -0.003379537 -0.000180237
#> Sigma[2,2]  0.057764480 0.021201412  0.042369521  0.086354576
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[3,1]  0.004125825 0.003559462 -0.0002830937  0.007717655
#> Sigma[3,2] -0.056416417 0.019978616 -0.0813937044 -0.035311060
#> Sigma[3,3]  0.159341100 0.033982577  0.1257221466  0.197078185
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd   5% quantile  95% quantile
#> Sigma[4,1]  0.002659532 0.002663586 -0.0003157259  0.0057851889
#> Sigma[4,2] -0.011346231 0.008651318 -0.0206367829 -0.0006666849
#> Sigma[4,3]  0.117358829 0.028734378  0.0934896314  0.1561182505
#> Sigma[4,4]  0.110943106 0.027885120  0.0865531930  0.1476422366
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.970358555 0.03468474  0.92674269  1.001351898
#> lag1_var2 -0.042655035 0.01096790 -0.05546168 -0.030165716
#> lag1_var3 -0.081509133 0.02974992 -0.12230398 -0.060467473
#> lag1_var4  0.084070598 0.03429560  0.05967087  0.130915128
#> const      0.578875655 0.61500135 -0.19696278  1.181074724
#> exo1      -0.027693987 0.01375773 -0.03965815 -0.009408087
#> exo2      -0.090156671 0.03391432 -0.12738921 -0.061313182
#> exo3      -0.008848053 0.04196178 -0.05877966  0.038871778
#> 
#> $ARE$A$equation2
#>                 mean        sd  5% quantile 95% quantile
#> lag1_var1  0.1076160 0.1496296 -0.048655450    0.2949700
#> lag1_var2  0.5494344 0.1780271  0.403476920    0.7573773
#> lag1_var3 -0.2696078 0.1807545 -0.474449840   -0.1189276
#> lag1_var4  0.1962152 0.1955537  0.001294699    0.4379830
#> const      3.3226599 3.7004402 -0.408818283    7.9888646
#> exo1       0.1858013 0.1304801  0.040490510    0.3214627
#> exo2       1.9874020 0.2283790  1.681812525    2.1816800
#> exo3      -0.6235360 0.2756821 -0.881520634   -0.2631492
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.3417767 0.1492989   0.1559720   0.49332417
#> lag1_var2 -0.7960277 0.2712804  -1.0770733  -0.48524313
#> lag1_var3 -0.1692609 0.1879131  -0.3909988   0.01943756
#> lag1_var4  1.1220622 0.1841265   0.9407623   1.34710060
#> const     -5.3630975 3.3266097  -8.3759836  -0.98765652
#> exo1      -0.2585516 0.5074150  -0.7363254   0.37375986
#> exo2      -2.5005246 0.3751804  -2.9652899  -2.14160403
#> exo3      -0.8407344 0.5291643  -1.5581686  -0.49191964
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.4114339 0.1950085   0.2318185    0.6593472
#> lag1_var2 -0.3771089 0.1547069  -0.4970956   -0.1704706
#> lag1_var3 -0.3849461 0.1031777  -0.4940916   -0.2580909
#> lag1_var4  1.3108934 0.1108737   1.2065634    1.4440325
#> const     -4.7397981 2.6856888  -8.2804675   -2.4509939
#> exo1      -0.1108995 0.5462124  -0.5706555    0.6174666
#> exo2      -0.9698129 0.3741818  -1.3195812   -0.4995098
#> exo3      -1.3634437 0.3614922  -1.8351506   -1.0356080
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002595242 0.0004615447 0.002251561  0.003208036
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05338281 0.01321155 -0.06539524  -0.03658094
#> Sigma[2,2]  2.80652804 0.27544863  2.47884455   3.11473780
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04140517 0.01067768  0.03031018   0.05465316
#> Sigma[3,2] -1.97751026 0.40018170 -2.48388837  -1.58482480
#> Sigma[3,3]  1.71720207 0.47650720  1.28095280   2.33750764
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01003613 0.006954752  0.002546508  0.018667745
#> Sigma[4,2] -0.28541605 0.287840101 -0.647890986  0.007652954
#> Sigma[4,3]  0.56238848 0.269652491  0.344081485  0.914113432
#> Sigma[4,4]  0.43356485 0.144519323  0.268329973  0.578586303
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.92135171 0.01138256  0.90871306   0.93504352
#> lag1_var2  0.11805604 0.04935443  0.06393742   0.17486309
#> lag1_var3  0.19666639 0.08319690  0.10729431   0.29329621
#> lag1_var4 -0.18056592 0.07078751 -0.25815796  -0.09991055
#> const      1.17841636 0.64702794  0.37444556   1.78346236
#> exo1      -0.01553467 0.04326298 -0.07040011   0.02903496
#> exo2      -0.16777272 0.06390215 -0.23701432  -0.10914074
#> exo3       0.05920623 0.04465217  0.01202021   0.11471892
#> 
#> $ARG$A$equation2
#>                 mean         sd  5% quantile 95% quantile
#> lag1_var1 -1.9812560  0.9608421  -2.88626597   -0.7280868
#> lag1_var2  2.4170977  2.4823030  -0.01176609    5.5086442
#> lag1_var3  2.7429503  4.0504018  -1.24322098    7.8091625
#> lag1_var4 -1.8949802  3.6417308  -6.47231009    1.6955204
#> const      4.4035900 17.4585858 -14.37165395   23.8396158
#> exo1      -0.3314391  1.2798871  -1.29707320    1.4152249
#> exo2       3.2978306  1.1821298   2.14748059    4.8191477
#> exo3       0.5781638  2.2636846  -1.79410983    3.0226833
#> 
#> $ARG$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.7151177 0.6476686  0.16625088   1.58755890
#> lag1_var2  1.6379346 1.4967074 -0.07726341   3.15223048
#> lag1_var3  3.3890018 2.4049431  0.59115722   5.84491946
#> lag1_var4 -2.7331023 2.1642377 -4.99106964  -0.40511119
#> const     -0.1255207 6.5865849 -7.49304618   7.20174216
#> exo1      -0.5802167 0.6589132 -1.43282650   0.03631388
#> exo2      -6.1352033 0.7470696 -6.80113876  -5.17224854
#> exo3       1.8967537 1.6038707  0.16022419   3.54767102
#> 
#> $ARG$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.5760435 0.3831182  -0.9709292   -0.1374450
#> lag1_var2  3.4395812 1.1124246   2.3701221    4.8535655
#> lag1_var3  5.5923314 1.8136978   3.8422548    7.8952229
#> lag1_var4 -4.2914495 1.6063817  -6.3188227   -2.7619009
#> const     -3.0207296 6.7228324 -11.5270603    3.0762470
#> exo1      -0.8743401 0.3450139  -1.2535601   -0.5084896
#> exo2      -4.6909511 0.5422052  -5.3655142   -4.1357967
#> exo3       2.4919243 0.3471690   2.0354663    2.7945276
#> 
#> 
#>