Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005265431 0.002270318 0.003238217  0.008168866
#> 
#> $AFG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01558003 0.01208789 -0.03186448 -0.005659843
#> Sigma[2,2]  0.19540906 0.03102268  0.16394928  0.231809220
#> 
#> $AFG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.02220941 0.01331814  0.01195186   0.04040456
#> Sigma[3,2] -0.17464589 0.06940157 -0.25033449  -0.09443627
#> Sigma[3,3]  0.39818929 0.12331624  0.27425894   0.54146301
#> 
#> $AFG$Sigma$equation4
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[4,1]  0.01746106 0.009987996  0.00831322   0.03067443
#> Sigma[4,2] -0.10086624 0.067488764 -0.17225533  -0.02295895
#> Sigma[4,3]  0.35633636 0.111998278  0.23227895   0.47463767
#> Sigma[4,4]  0.34726923 0.106474784  0.22584185   0.44773644
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.98654315 0.04135392   0.9366563   1.02407377
#> lag1_var2 -0.40709196 0.03660540  -0.4513437  -0.36918535
#> lag1_var3 -0.85570293 0.04861971  -0.9144666  -0.80601997
#> lag1_var4  0.74115922 0.03528693   0.7043153   0.78246317
#> const      5.78713190 0.32599333   5.3686201   6.11565959
#> exo1      -0.04588327 0.12630754  -0.1759357   0.09479927
#> exo2      -0.11522322 0.06388465  -0.1732642  -0.03046334
#> exo3      -0.23392580 0.08369682  -0.3426691  -0.15665299
#> 
#> $AFG$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.3852706 0.1560717  0.19599612   0.55025064
#> lag1_var2  0.8356282 0.1002524  0.76041739   0.97033665
#> lag1_var3 -0.2570370 0.1312373 -0.42655777  -0.12701074
#> lag1_var4  0.2059537 0.1390012  0.05004302   0.37248518
#> const     -5.9460664 2.6638056 -8.54816280  -3.12569396
#> exo1      -0.6593427 0.2300736 -0.96636568  -0.45918509
#> exo2       0.3544885 0.2432656  0.14900279   0.68282117
#> exo3      -0.3070239 0.2561583 -0.59130727  -0.03347449
#> 
#> $AFG$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.2946817 0.2387067  -0.5992520  -0.05291462
#> lag1_var2 -0.7032432 0.1529049  -0.8884548  -0.53332830
#> lag1_var3 -0.2109783 0.1478771  -0.3702648  -0.03230666
#> lag1_var4  1.0438745 0.1680991   0.8193818   1.18692370
#> const     15.5234190 3.6768272  11.9436209  19.93638399
#> exo1       0.4920644 0.6960132  -0.2940431   1.32656533
#> exo2      -1.5013739 0.4359722  -2.0371934  -1.02486575
#> exo3       0.2056378 0.3746706  -0.1422646   0.70043703
#> 
#> $AFG$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.1447377 0.1953434 -0.40574198   0.02790535
#> lag1_var2 -0.3609466 0.1344752 -0.50108676  -0.19223780
#> lag1_var3 -0.3665245 0.1144103 -0.51110070  -0.24829572
#> lag1_var4  1.2982033 0.1543735  1.12029187   1.46678919
#> const      7.9401928 2.9566826  4.54113490  11.38771724
#> exo1       0.2202694 0.7269918 -0.61276611   1.07242941
#> exo2      -1.4329418 0.3651153 -1.87630902  -1.02830050
#> exo3       0.1933897 0.4636482 -0.07381898   0.83076086
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.006467209 0.001627243 0.004751425  0.008356176
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01071474 0.007067245 -0.02019766 -0.004567196
#> Sigma[2,2]  0.11380944 0.023248011  0.08860292  0.141666917
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.008723617 0.004217103  0.003927506   0.01321584
#> Sigma[3,2] -0.044302519 0.012986151 -0.061131812  -0.03187153
#> Sigma[3,3]  0.034433257 0.006967914  0.025758506   0.04139660
#> 
#> $AGO$Sigma$equation4
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.0003177960 0.004164708 -0.004333507  0.005236118
#> Sigma[4,2]  0.0468205619 0.007071766  0.040747430  0.055445362
#> Sigma[4,3] -0.0004267083 0.008415230 -0.009377832  0.009859732
#> Sigma[4,4]  0.0411177343 0.004427842  0.037904972  0.047105421
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.03611189 0.02150642  1.01224164   1.06166397
#> lag1_var2 -0.62004259 0.01268223 -0.63176098  -0.60419799
#> lag1_var3 -0.80661048 0.04521415 -0.86638566  -0.76812807
#> lag1_var4  0.71901716 0.04601891  0.68326322   0.78131853
#> const      5.88283090 1.24799335  4.38308392   7.16623157
#> exo1       0.03472702 0.10035418 -0.05651419   0.15683934
#> exo2       0.04664723 0.04724884 -0.01598163   0.08691126
#> exo3       0.12108670 0.04580582  0.08911629   0.18390361
#> 
#> $AGO$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.1345953 0.05769631  0.05738209   0.18040638
#> lag1_var2  0.5174061 0.22938139  0.22529817   0.72260287
#> lag1_var3 -0.5230817 0.23668314 -0.80275144  -0.29082110
#> lag1_var4  0.5666490 0.24036228  0.31969831   0.84374936
#> const     -5.4312895 2.76043060 -7.21638745  -1.64719834
#> exo1       0.2207371 0.32504921 -0.17113671   0.58203128
#> exo2       0.1945803 0.28981993 -0.18775696   0.44867695
#> exo3      -0.3976320 0.32974597 -0.79874581  -0.04644546
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.11800593 0.04024691  -0.1569830  -0.06665685
#> lag1_var2 -0.39708994 0.06115641  -0.4401013  -0.31298054
#> lag1_var3  0.45015884 0.06025433   0.3712681   0.50401193
#> lag1_var4  0.49334968 0.05456848   0.4419666   0.55983043
#> const      6.77504241 3.53069379   2.6172907  10.69157833
#> exo1       0.01081749 0.23713224  -0.2905659   0.21924731
#> exo2      -1.29157315 0.22952538  -1.5160594  -0.99578985
#> exo3       3.45137520 0.23776369   3.1576256   3.69855490
#> 
#> $AGO$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.14666121 0.08771040 -0.22139894  -0.03104967
#> lag1_var2  0.08908817 0.12107795 -0.05033416   0.20316435
#> lag1_var3  0.14594895 0.18688666 -0.10597750   0.30448905
#> lag1_var4  0.91649336 0.16670734  0.78121536   1.13778322
#> const     -0.95805930 1.95592711 -2.99120580   1.14998349
#> exo1       0.26614114 0.08269925  0.16198069   0.35152684
#> exo2      -1.49496781 0.11004505 -1.59990158  -1.36151096
#> exo3       3.40599636 0.20512768  3.24186623   3.63827060
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003737444 0.0009436371 0.002904422  0.004990317
#> 
#> $ALB$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.1079066 0.01331201  -0.1229717  -0.09660147
#> Sigma[2,2]  8.7559620 1.50080870   7.4867587  10.73457236
#> 
#> $ALB$Sigma$equation3
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.0675387 0.02781335  0.03867436   0.09671625
#> Sigma[3,2] -5.1055016 1.25878121 -6.28705022  -3.46077328
#> Sigma[3,3]  5.0300447 1.16744390  3.45971591   5.97964203
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.00440665 0.03202838 -0.03299736   0.03571665
#> Sigma[4,2]  0.86086206 0.87991152  0.13593947   2.03637188
#> Sigma[4,3]  1.97630987 0.77248650  1.25314547   2.99482967
#> Sigma[4,4]  3.09125760 0.75667333  2.39163857   4.09262976
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1  0.855896901 0.03724735  0.817691654   0.90168496
#> lag1_var2 -0.022341714 0.02853389 -0.044884460   0.01604327
#> lag1_var3 -0.029726961 0.04177368 -0.060233360   0.02767582
#> lag1_var4  0.017901297 0.03425826 -0.029193754   0.04141869
#> const      4.105083163 1.09878028  3.062600300   5.34238419
#> exo1       0.003639585 0.04646080 -0.028817464   0.06545080
#> exo2       0.025378701 0.02545857 -0.003954742   0.05539785
#> exo3       0.087991325 0.05720279  0.037937292   0.15882837
#> 
#> $ALB$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -2.4421789  1.736081  -4.4087365   -0.4349564
#> lag1_var2  1.9247585  1.903137   0.2665044    4.4435459
#> lag1_var3  2.4173934  3.072628  -0.1206163    6.5357727
#> lag1_var4 -1.9307755  2.533849  -5.3325438    0.1071423
#> const     35.1781833 35.003911 -10.2924892   66.8427805
#> exo1      -2.9305207  1.293688  -4.5363049   -1.6426124
#> exo2      -2.0180501  1.839666  -4.3503634   -0.2155377
#> exo3      -0.1292768  2.614612  -2.9984994    3.0409274
#> 
#> $ALB$A$equation3
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1   1.42197573  0.5468194   0.9772504     2.161400
#> lag1_var2  -0.74830054  2.3108987  -3.3768492     1.877065
#> lag1_var3  -0.58977462  3.5831740  -4.6680829     3.397267
#> lag1_var4   1.22042273  2.9857852  -2.1362782     4.578556
#> const     -13.94394014 28.0956376 -45.2468223    16.240892
#> exo1        0.01304152  1.2123677  -1.2702770     1.259225
#> exo2       -0.65360041  1.3915872  -2.0173319     1.141873
#> exo3       -0.20290479  3.3854472  -4.5828278     3.052987
#> 
#> $ALB$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.3849685  1.054887   -1.154223   1.06346019
#> lag1_var2  0.1841647  1.584748   -1.300651   2.09635987
#> lag1_var3  0.5818194  2.425390   -1.794488   3.46892136
#> lag1_var4  0.4597363  2.110415   -2.052930   2.50254280
#> const      8.7082151 24.331463  -10.594622  41.71491398
#> exo1      -1.8982608  1.587412   -3.710296  -0.04791747
#> exo2      -2.3022844  1.664258   -4.368431  -0.57967392
#> exo3      -0.2867033  2.126111   -3.055775   1.60412082
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                  mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00484682 0.001784914 0.002717682  0.006866139
#> 
#> $ARE$Sigma$equation2
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.008045692 0.00563918 -0.01346878 -0.001798721
#> Sigma[2,2]  0.079726805 0.02469102  0.05260005  0.109743599
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.009992911 0.006200598  0.002322025   0.01613522
#> Sigma[3,2] -0.057058687 0.028968691 -0.092945226  -0.02580665
#> Sigma[3,3]  0.184240878 0.068303995  0.120985618   0.27391141
#> 
#> $ARE$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1] 0.003764137 0.005334482 -0.001275909   0.01029516
#> Sigma[4,2] 0.004014664 0.025104573 -0.015089040   0.03799462
#> Sigma[4,3] 0.143208845 0.057419490  0.085940658   0.21475443
#> Sigma[4,4] 0.149726913 0.062479093  0.092610714   0.21716543
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.90629393 0.07126338  0.819356022   0.97637367
#> lag1_var2 -1.23293054 0.20350271 -1.496648284  -1.04561465
#> lag1_var3 -1.68990446 0.22735942 -1.987518136  -1.47776752
#> lag1_var4  1.65736201 0.20528944  1.479789259   1.93107846
#> const      4.85214856 0.50299708  4.280577604   5.36205286
#> exo1       0.04434118 0.03880340 -0.003889172   0.08277817
#> exo2      -0.06105623 0.07848641 -0.160858524   0.01747884
#> exo3      -0.14247418 0.05534825 -0.196316258  -0.07740097
#> 
#> $ARE$A$equation2
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1  0.09418413 0.1063499 -0.007884464   0.21645467
#> lag1_var2  1.91174513 0.8428179  1.090682347   2.89854131
#> lag1_var3  1.54812509 1.0714892  0.590544669   2.81852339
#> lag1_var4 -1.52251413 1.0454638 -2.756402894  -0.57696993
#> const     -3.80963002 3.2267150 -7.215219117   0.01950784
#> exo1      -0.10953334 0.1474318 -0.230537518   0.08976314
#> exo2       1.82384192 0.3203296  1.490302547   2.21300856
#> exo3      -0.57253978 0.2094018 -0.849649420  -0.38040126
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.1583659 0.3742548  -0.1214541    0.6685787
#> lag1_var2 -2.8629219 1.7735494  -4.9242688   -1.1046358
#> lag1_var3 -3.3381155 2.1686886  -5.8701751   -1.0739221
#> lag1_var4  4.2167226 2.0542322   2.0915497    6.5884516
#> const      4.2598544 5.2283191  -0.9858137   10.7624242
#> exo1      -0.3798548 0.5296690  -0.7797132    0.3329100
#> exo2      -2.3387822 0.4763964  -2.8950808   -1.8626682
#> exo3      -2.0073713 0.5199836  -2.6891436   -1.5465224
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.2246511 0.3347532 -0.08546514   0.66544295
#> lag1_var2 -1.4345495 1.2454806 -2.78399901   0.01081924
#> lag1_var3 -2.2089479 1.4846208 -3.81986979  -0.43495058
#> lag1_var4  3.1350813 1.3850073  1.47673266   4.60884036
#> const     -0.6566472 2.7785145 -2.88809634   3.07252951
#> exo1      -0.4777487 0.5110922 -0.80100292   0.22313913
#> exo2      -0.9450358 0.3516459 -1.39736259  -0.62323599
#> exo3      -2.5313667 0.4907818 -3.08306019  -2.00200554
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003046662 0.0004786642 0.002603267  0.003672513
#> 
#> $ARG$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0609488 0.01577898 -0.08244309  -0.04946758
#> Sigma[2,2]  3.0998730 1.01943779  2.29569583   4.46605918
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04265481 0.01004393  0.03515672   0.05636556
#> Sigma[3,2] -2.22641474 0.69769559 -3.14881211  -1.60616387
#> Sigma[3,3]  1.84509888 0.46869602  1.37723235   2.44921477
#> 
#> $ARG$Sigma$equation4
#>                    mean         sd  5% quantile 95% quantile
#> Sigma[4,1]  0.006682764 0.00203804  0.004066258  0.008656777
#> Sigma[4,2] -0.380712535 0.13160874 -0.497466020 -0.206887398
#> Sigma[4,3]  0.546631433 0.10308918  0.414406464  0.629233392
#> Sigma[4,4]  0.352317163 0.06475767  0.286848627  0.432810381
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.80989234 0.02087338  0.78547513  0.830023699
#> lag1_var2  0.11658637 0.02800564  0.07839880  0.132785407
#> lag1_var3  0.19474260 0.03960343  0.14049108  0.221508429
#> lag1_var4 -0.19793428 0.04957366 -0.23419855 -0.130392587
#> const      5.34169277 0.45771327  4.91516175  5.920067570
#> exo1       0.01857923 0.02669225 -0.01124306  0.042429540
#> exo2      -0.11902214 0.09848775 -0.21545911  0.004106525
#> exo3      -0.04696288 0.12047301 -0.16771322  0.107993699
#> 
#> $ARG$A$equation2
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1 -1.00456948  0.952487  -2.17284501   -0.1938010
#> lag1_var2  1.08278307  1.135098   0.06546481    2.5567094
#> lag1_var3  0.55123946  1.591870  -0.85345068    2.6393787
#> lag1_var4  0.07755106  1.874242  -2.38731482    1.7013912
#> const     -8.56388484 27.891323 -38.87113445   21.3089272
#> exo1      -2.05454066  1.690845  -3.50448472    0.1718451
#> exo2       3.20596820  1.321972   1.39229970    4.1625554
#> exo3       0.53881846  2.399050  -2.74161998    2.3598370
#> 
#> $ARG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.61789160  0.7298017  0.02800489    1.5391717
#> lag1_var2 -0.20902670  0.9630774 -1.48795856    0.6216841
#> lag1_var3  0.38207864  1.3911932 -1.47558959    1.5256279
#> lag1_var4  0.07047284  1.5916839 -1.27522675    2.2036483
#> const     15.05893256 20.1903207 -5.65271907   36.9773464
#> exo1       0.26265916  1.0718114 -1.09660729    1.3055325
#> exo2      -5.78587028  1.1055381 -6.81293398   -4.3444725
#> exo3       2.97743273  1.7502938  1.48462865    5.3364627
#> 
#> $ARG$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.08813403 0.2045194 -0.33249410    0.1351156
#> lag1_var2  0.46496562 0.3153945  0.03535009    0.6865202
#> lag1_var3  0.73764831 0.4930225  0.06321276    1.0572013
#> lag1_var4  0.18036394 0.5137265 -0.20568909    0.8800103
#> const      7.35493136 3.5573571  3.63141986   11.0879999
#> exo1      -1.06780287 0.2556261 -1.35393196   -0.7896252
#> exo2      -4.29395828 0.4845886 -4.83496103   -3.8194653
#> exo3       3.66237797 0.4082587  3.21073761    4.1104172
#> 
#> 
#>