Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004267575 0.001307135 0.002977529  0.005683347
#> 
#> $AFG$Sigma$equation2
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.005387568 0.002976459 -0.008989907 -0.002215212
#> Sigma[2,2]  0.184975093 0.056288953  0.120408791  0.244089021
#> 
#> $AFG$Sigma$equation3
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[3,1] -0.0000970187 0.003908566 -0.003788169  0.004348844
#> Sigma[3,2] -0.1556452810 0.032353789 -0.197562041 -0.123967669
#> Sigma[3,3]  0.2619758847 0.054705308  0.225489548  0.336816166
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1] -0.002552941 0.003323607 -0.005368814  0.001323099
#> Sigma[4,2] -0.086963029 0.036565080 -0.134943446 -0.054212789
#> Sigma[4,3]  0.223595238 0.052290988  0.183513033  0.293640011
#> Sigma[4,4]  0.214522585 0.048216941  0.169991378  0.275647112
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.979248925 0.02509178  0.94739152   1.00449385
#> lag1_var2  0.348694990 0.20676488  0.18645200   0.62981245
#> lag1_var3  0.760218869 0.41852134  0.41543659   1.32332700
#> lag1_var4 -0.656269225 0.35794057 -1.13584858  -0.35851921
#> const     -4.263620564 3.06922768 -8.36385201  -1.88958630
#> exo1      -0.047808177 0.06060225 -0.12160761   0.01779976
#> exo2      -0.004987165 0.04206296 -0.05100316   0.04629350
#> exo3      -0.093415451 0.04042184 -0.14582361  -0.06045442
#> 
#> $AFG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.3025735  0.2992221 -0.03187868    0.6567830
#> lag1_var2 -0.5308069  1.0691777 -1.86028468    0.5376325
#> lag1_var3 -3.3130102  2.1211610 -5.83743587   -1.1617339
#> lag1_var4  2.8670577  1.7973449  1.03271618    4.9617119
#> const     13.9530832 13.3489033 -0.18350129   29.0273170
#> exo1      -0.2469187  0.1472203 -0.42465702   -0.1061154
#> exo2      -0.2829364  0.6182614 -0.82250121    0.5219721
#> exo3      -0.5565186  0.1376841 -0.71339909   -0.3927908
#> 
#> $AFG$A$equation3
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  -0.1304364  0.2459549  -0.41030065   0.15670669
#> lag1_var2   2.9940854  1.9348253   1.04923069   5.18249259
#> lag1_var3   8.1405428  4.0733774   4.05154855  12.58139666
#> lag1_var4  -6.2963645  3.5227503 -10.08311887  -2.75460203
#> const     -33.8918003 25.6921115 -61.55412046  -8.96262606
#> exo1        0.1508547  0.2124643  -0.08218714   0.40815806
#> exo2       -0.5773910  0.7330420  -1.47396100   0.07428112
#> exo3        0.9055835  0.5482086   0.36306072   1.59656998
#> 
#> $AFG$A$equation4
#>                    mean         sd 5% quantile 95% quantile
#> lag1_var1  -0.009289361  0.1524202  -0.1691345    0.1590311
#> lag1_var2   3.132471022  1.7814155   1.3773907    5.0342386
#> lag1_var3   7.499523605  3.7552349   3.8576916   11.5743527
#> lag1_var4  -5.613285737  3.2555721  -9.1686613   -2.4601563
#> const     -38.329952602 23.9460691 -65.3272285  -15.7798092
#> exo1        0.031160929  0.2536133  -0.2800532    0.2998303
#> exo2       -0.719412789  0.5483860  -1.3624489   -0.2227128
#> exo3        0.845640556  0.5734273   0.2410462    1.5526064
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                  mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.01246643 0.002730126 0.009894819   0.01604231
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01568049 0.002867518 -0.01891284  -0.01246544
#> Sigma[2,2]  0.11495743 0.019383959  0.09039452   0.13150063
#> 
#> $AGO$Sigma$equation3
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.01592955 0.002420318  0.01285950   0.01832723
#> Sigma[3,2] -0.05311259 0.008576480 -0.06200087  -0.04359818
#> Sigma[3,3]  0.04126643 0.005550247  0.03581631   0.04743728
#> 
#> $AGO$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1] 0.006449014 0.003071215  0.002722379  0.009804965
#> Sigma[4,2] 0.038624412 0.008863522  0.026794435  0.045934920
#> Sigma[4,3] 0.002798036 0.004199872 -0.002951757  0.005979184
#> Sigma[4,4] 0.039051796 0.007729225  0.029809949  0.047732166
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1   0.8740982 0.04122313   0.83588092   0.92684524
#> lag1_var2   1.7929538 0.04618192   1.73138516   1.83163036
#> lag1_var3   2.3572854 0.04531563   2.31136428   2.41518656
#> lag1_var4  -2.0443386 0.02093589  -2.07117916  -2.02572629
#> const     -20.5959818 2.34164360 -22.20799860 -17.40966251
#> exo1        0.1056167 0.11700145  -0.03724949   0.21402951
#> exo2       -0.1501859 0.08805812  -0.23263175  -0.05008500
#> exo3       -0.2386163 0.15949300  -0.39695768  -0.04368533
#> 
#> $AGO$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.13617150  0.1431458 -0.04272541    0.2691784
#> lag1_var2 -1.28450229  0.3179090 -1.68123304   -0.9500668
#> lag1_var3 -2.86973013  0.5455545 -3.44956129   -2.2017602
#> lag1_var4  2.61357909  0.3686261  2.18041472    3.0015165
#> const     17.56507321 14.2369075  3.46715016   35.6342463
#> exo1      -0.03677451  0.1816784 -0.22138659    0.1700756
#> exo2       0.60537050  0.3387055  0.32941062    1.0477880
#> exo3      -0.05500249  0.5859915 -0.52437793    0.7134631
#> 
#> $AGO$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.3167122 0.0702493  -0.3700776   -0.2358682
#> lag1_var2  0.9076741 0.2326483   0.6573755    1.1947475
#> lag1_var3  2.1586535 0.3390869   1.8482216    2.5776812
#> lag1_var4 -1.1000668 0.2403655  -1.3945310   -0.8903818
#> const      2.9638093 8.2567323  -7.4538864   11.5697579
#> exo1       0.1666893 0.2321820  -0.1222777    0.3774155
#> exo2      -1.6440531 0.2102761  -1.8690901   -1.4236628
#> exo3       2.9109335 0.2850512   2.6118596    3.2623455
#> 
#> $AGO$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.2976065 0.0750044 -0.39022978  -0.24209072
#> lag1_var2 -0.2032718 0.1910996 -0.37689784   0.04858337
#> lag1_var3 -0.1948756 0.2570480 -0.39594547   0.15366533
#> lag1_var4  1.1595698 0.1686751  0.93479134   1.30432945
#> const     10.8096527 8.9401929 -0.16687856  19.72785201
#> exo1       0.1853037 0.1354749  0.00146791   0.28833324
#> exo2      -1.4745672 0.2505052 -1.81928582  -1.31716720
#> exo3       3.2899779 0.3763165  2.86996816   3.60363054
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004820008 0.002220981 0.002331044  0.007234588
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.09526599 0.04979173  -0.1563215  -0.05187957
#> Sigma[2,2]  7.43588048 1.96369105   6.0221472  10.12760673
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.02752732 0.03753612 -0.01847694   0.06379928
#> Sigma[3,2] -4.16711026 0.65302279 -4.84068869  -3.33747299
#> Sigma[3,3]  4.32052641 0.78676527  3.64298259   5.37620243
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.04654157 0.02421419 -0.07788442  -0.02680391
#> Sigma[4,2]  1.08640685 1.19291732 -0.13408154   2.58693201
#> Sigma[4,3]  1.84074067 1.47974412  0.74048843   3.87161376
#> Sigma[4,4]  3.19704509 1.85426954  1.87803248   5.75103235
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.96482117 0.04291785  0.91554612  1.006311409
#> lag1_var2 -0.07856737 0.06534524 -0.15916007 -0.007605265
#> lag1_var3 -0.13108649 0.10143896 -0.25686129 -0.022847807
#> lag1_var4  0.10832064 0.08080433  0.02192985  0.208240218
#> const      2.20088425 2.25039607 -0.14897450  4.907634505
#> exo1       0.03468264 0.06817732 -0.05091553  0.100902475
#> exo2      -0.09045626 0.08333796 -0.19856980 -0.012805286
#> exo3       0.04836492 0.03421747  0.01313983  0.090635144
#> 
#> $ALB$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -5.62454435  1.4940120  -7.1327100   -3.8338218
#> lag1_var2  5.13249514  1.3098997   3.6087456    6.6464947
#> lag1_var3  7.78217932  1.7891422   5.7433972    9.8821228
#> lag1_var4 -6.58585230  1.4585736  -8.1616407   -4.9077648
#> const     65.67661199 61.3324415 -10.9019576  126.9341533
#> exo1      -2.47004754  2.1526665  -5.1107472   -0.1765289
#> exo2      -0.04911157  0.8528094  -0.7261577    0.9702707
#> exo3      -0.83165222  1.4232042  -2.1608527    1.0022966
#> 
#> $ALB$A$equation3
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1   5.0802775   2.225597    3.3289507    8.1139620
#> lag1_var2  -1.9523301   2.820947   -4.2942574    1.8254682
#> lag1_var3  -2.9187539   4.176608   -6.3606978    2.6754658
#> lag1_var4   3.2559069   3.399504   -1.3192555    5.9231166
#> const     -82.8622373 102.433811 -219.5957099    7.5266993
#> exo1       -0.5788315   2.534224   -3.6915237    2.1838385
#> exo2       -0.2454047   0.655091   -1.0440061    0.4661791
#> exo3        0.3512570   1.320576   -0.9184652    2.0599577
#> 
#> $ALB$A$equation4
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1   1.7006798  1.893750    0.3638234    4.2506429
#> lag1_var2   1.4386885  2.692490   -0.5820638    5.1202027
#> lag1_var3   2.2638732  4.121465   -0.8232561    7.8988607
#> lag1_var4  -0.9759403  3.400659   -5.6307348    1.5444350
#> const     -57.2963875 89.233571 -178.5573190    9.6694233
#> exo1       -2.2126706  2.308747   -5.3989500   -0.6998666
#> exo2       -0.1697378  0.418242   -0.6396461    0.2525255
#> exo3       -0.2040504  1.737398   -1.9133302    1.8925102
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001530444 0.0001297403 0.001430796  0.001707083
#> 
#> $ARE$Sigma$equation2
#>                    mean           sd  5% quantile 95% quantile
#> Sigma[2,1] -0.002951799 0.0006565832 -0.003849439 -0.002517166
#> Sigma[2,2]  0.049339747 0.0088741827  0.038732689  0.057825025
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.005857168 0.002557117  0.003425665  0.009049622
#> Sigma[3,2] -0.060392492 0.020411482 -0.083486826 -0.036893529
#> Sigma[3,3]  0.172025874 0.033287728  0.131498194  0.208618754
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.003855773 0.002264848  0.001639353  0.006555812
#> Sigma[4,2] -0.023196742 0.014611393 -0.041297858 -0.007795574
#> Sigma[4,3]  0.129830132 0.018251258  0.106788099  0.148839990
#> Sigma[4,4]  0.116125095 0.008810557  0.104270599  0.122651227
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                    mean         sd 5% quantile 95% quantile
#> lag1_var1  0.9262152891 0.05342286  0.85849964   0.98147455
#> lag1_var2  0.8903613239 0.04818795  0.84464227   0.95284010
#> lag1_var3  1.1324945849 0.07578903  1.06200183   1.23062022
#> lag1_var4 -1.0981808358 0.08274643 -1.21042755  -1.03052794
#> const     -0.6985042349 0.44033403 -1.16594173  -0.18673778
#> exo1      -0.0007125296 0.02611427 -0.03237868   0.02544050
#> exo2      -0.0824853657 0.03311447 -0.12669872  -0.05464484
#> exo3       0.0407412099 0.05296842 -0.02785404   0.08901557
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.07872051 0.4026418  -0.2351465    0.5951887
#> lag1_var2 -0.72815414 0.1554531  -0.9228910   -0.5974491
#> lag1_var3 -1.95606458 0.3180568  -2.3744736   -1.6757232
#> lag1_var4  1.87299507 0.3943401   1.4560058    2.3416721
#> const      4.81279066 3.2560850   1.5668964    8.9342493
#> exo1       0.03976645 0.3729180  -0.3647219    0.4134447
#> exo2       1.99960851 0.1686049   1.7994425    2.1615997
#> exo3      -0.78313829 0.3521598  -1.1558242   -0.3611617
#> 
#> $ARE$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1   0.3449967 0.6378780  -0.3913764    0.9302573
#> lag1_var2   4.1056369 0.4154581   3.6746343    4.6291559
#> lag1_var3   5.9519436 0.7319773   5.4398002    6.9523617
#> lag1_var4  -4.8933739 0.8023016  -5.9430370   -4.1922211
#> const     -14.0619217 8.1815805 -25.1200621   -7.6382310
#> exo1       -0.3721320 0.6040368  -0.8245094    0.4303838
#> exo2       -2.4263692 0.3809923  -2.7363126   -1.9107805
#> exo3       -1.4373720 0.5296831  -1.8275853   -0.7144007
#> 
#> $ARE$A$equation4
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1   0.3731474 0.3523141  -0.02012987    0.7079087
#> lag1_var2   3.6221986 0.3053513   3.28152728    3.9921539
#> lag1_var3   4.5331101 0.4929900   4.17250224    5.2071401
#> lag1_var4  -3.5045004 0.5122976  -4.18927686   -3.1178379
#> const     -12.1346404 5.7846639 -19.82800130   -7.3852251
#> exo1       -0.3529540 0.4137708  -0.73386340    0.1879445
#> exo2       -0.9037849 0.3070202  -1.14969287   -0.4884161
#> exo3       -2.1108814 0.3477551  -2.39325148   -1.6449857
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003575472 0.001257068 0.002373239  0.005151975
#> 
#> $ARG$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0367128 0.01638649  -0.0557487  -0.01918167
#> Sigma[2,2]  2.8272598 0.83725051   1.8628680   3.75894647
#> 
#> $ARG$Sigma$equation3
#>                   mean        sd 5% quantile 95% quantile
#> Sigma[3,1]  0.03371713 0.0137613  0.01845149   0.04735589
#> Sigma[3,2] -2.01757481 0.5205178 -2.48635335  -1.35436722
#> Sigma[3,3]  1.68199423 0.3560158  1.22972596   2.03069980
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01377307 0.005713275  0.006407722   0.01925939
#> Sigma[4,2] -0.32648277 0.141142763 -0.484429165  -0.16757660
#> Sigma[4,3]  0.50656780 0.128773533  0.358495170   0.65352388
#> Sigma[4,4]  0.34810162 0.074269144  0.283815191   0.44772710
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.81355278 0.03906094  0.77489140  0.857841291
#> lag1_var2  0.71765314 0.27794149  0.33707061  0.918847670
#> lag1_var3  1.17734477 0.44994681  0.56132641  1.506846621
#> lag1_var4 -1.07561517 0.40829854 -1.36923418 -0.515697569
#> const     -1.14808634 2.45086726 -3.90040230  1.634185410
#> exo1       0.04839449 0.07334165 -0.02873929  0.126489438
#> exo2      -0.08510076 0.03572897 -0.11472444 -0.037487199
#> exo3      -0.06350517 0.05322012 -0.11979276  0.001664888
#> 
#> $ARG$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  -4.2773097  1.897090  -6.6082083    -2.446372
#> lag1_var2  -2.7657384  4.070643  -6.2984277     2.590050
#> lag1_var3  -5.5390533  6.809515 -11.1775202     3.521470
#> lag1_var4   5.4979875  5.945440  -2.2883387    10.741198
#> const     121.5160084 36.796390  89.0118675   170.447054
#> exo1       -1.1875912  1.794539  -2.5957885     1.251069
#> exo2        3.0101929  1.612846   0.8807326     4.438371
#> exo3        0.5335677  2.112221  -2.1761223     2.621230
#> 
#> $ARG$A$equation3
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1   1.6777702  1.196625   0.61468146    3.0097559
#> lag1_var2   3.7990580  2.989064   0.03453437    6.6449444
#> lag1_var3   6.8169074  4.993259   0.45818393   11.5382865
#> lag1_var4  -5.7688629  4.322439  -9.99653319   -0.3577383
#> const     -50.5450153 36.786492 -98.22584859  -15.8631702
#> exo1        0.3300913  1.387018  -1.56084922    1.3586526
#> exo2       -5.6032918  1.077894  -6.48690780   -4.2798420
#> exo3        2.5872718  1.362792   0.82042518    3.8920469
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -1.0720276  0.3357600   -1.448123   -0.6889254
#> lag1_var2  2.2523522  0.9797030    1.263767    3.3452446
#> lag1_var3  3.6631194  1.6149175    2.025353    5.4943978
#> lag1_var4 -2.5435168  1.4102763   -4.208078   -1.1488762
#> const     21.4584599 18.7770681    2.116897   45.0949300
#> exo1      -0.4433735  0.6204755   -1.176262    0.1632336
#> exo2      -4.2541342  0.7683301   -4.927151   -3.2491682
#> exo3       3.2450289  0.8029106    2.487266    4.2671312
#> 
#> 
#>