Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.007622949 0.001375821 0.006210125  0.009091321
#> 
#> $AFG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01913198 0.005494063 -0.02515031  -0.01269121
#> Sigma[2,2]  0.22466207 0.049760045  0.18604919   0.29137765
#> 
#> $AFG$Sigma$equation3
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.0289659 0.01522104  0.01258912   0.04730297
#> Sigma[3,2] -0.2187751 0.06249308 -0.27558363  -0.13584886
#> Sigma[3,3]  0.4044940 0.13262252  0.23230336   0.53404507
#> 
#> $AFG$Sigma$equation4
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[4,1]  0.02268075 0.01425999  0.007368961   0.03987846
#> Sigma[4,2] -0.13632487 0.05769809 -0.189230265  -0.06097257
#> Sigma[4,3]  0.35007062 0.12166457  0.193405125   0.46991669
#> Sigma[4,4]  0.33095772 0.10929545  0.192017584   0.44082055
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.957229187 0.01477884   0.9392363   0.97322501
#> lag1_var2 -0.447941531 0.02644843  -0.4748580  -0.41723427
#> lag1_var3 -0.954008943 0.03359232  -0.9936189  -0.91938714
#> lag1_var4  0.831039373 0.03129512   0.8039298   0.87248158
#> const      6.816367581 0.87676198   5.9058506   7.84814641
#> exo1      -0.006037454 0.10791318  -0.1422996   0.09807711
#> exo2      -0.103506796 0.13140796  -0.2625065   0.04123215
#> exo3      -0.257634694 0.10095248  -0.3654115  -0.14664243
#> 
#> $AFG$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1   0.5273468 0.1078630   0.4617969    0.6749471
#> lag1_var2   1.1902302 0.2313656   0.9291012    1.4406131
#> lag1_var3   0.5820774 0.2220887   0.3504196    0.8354651
#> lag1_var4  -0.5554180 0.1294504  -0.7039718   -0.4201923
#> const     -12.4651026 4.9262412 -17.5472496   -6.2547278
#> exo1       -0.6341785 0.7033002  -1.2287978    0.1570545
#> exo2       -0.2443418 0.5584857  -0.8837674    0.4121961
#> exo3       -0.1718594 0.4142161  -0.5910461    0.3149891
#> 
#> $AFG$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.3236902 0.1527894  -0.4538209   -0.1431163
#> lag1_var2 -0.5783562 0.4347454  -1.1054316   -0.1059932
#> lag1_var3  0.1554545 0.5386159  -0.5155298    0.7144263
#> lag1_var4  0.7257590 0.3827266   0.4148250    1.2396433
#> const     14.3545178 9.0755728   2.1795517   21.5882008
#> exo1       0.5197551 1.0031829  -0.5425912    1.6397968
#> exo2      -0.8630548 0.8912733  -1.7902020    0.1451300
#> exo3      -0.3624145 0.6229868  -0.9683474    0.3749654
#> 
#> $AFG$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.10620931 0.1471751  -0.2662158   0.06004187
#> lag1_var2  0.02633998 0.4346032  -0.5099201   0.49434549
#> lag1_var3  0.57064336 0.5445202  -0.1384612   1.10025847
#> lag1_var4  0.48726645 0.3913881   0.2037632   1.02399954
#> const      1.65668091 9.4593297 -11.3138024   8.50126652
#> exo1       0.33881025 0.8681036  -0.4249504   1.43579053
#> exo2      -0.98097351 0.7383471  -1.7157788  -0.12419483
#> exo3      -0.38286443 0.5992621  -1.0079499   0.30575140
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005407548 0.001082513 0.004334716   0.00679182
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01163408 0.003184351 -0.01352006  -0.00732289
#> Sigma[2,2]  0.15016244 0.032831473  0.11823626   0.19000808
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.008505236 0.001506993  0.006568014  0.009890252
#> Sigma[3,2] -0.068338283 0.016427555 -0.090453560 -0.054740927
#> Sigma[3,3]  0.043063604 0.008316263  0.035678030  0.054021031
#> 
#> $AGO$Sigma$equation4
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[4,1] -0.0004831083 0.002115636 -0.002285447  0.002184723
#> Sigma[4,2]  0.0532047551 0.006974450  0.044918651  0.059993144
#> Sigma[4,3] -0.0105772517 0.004430854 -0.015599429 -0.006557035
#> Sigma[4,4]  0.0381681514 0.002442518  0.035689951  0.041264920
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.99845769 0.01247316  0.98417570    1.0107007
#> lag1_var2 -0.69736428 0.03860214 -0.73697973   -0.6548748
#> lag1_var3 -0.88836491 0.02704627 -0.91948087   -0.8598994
#> lag1_var4  0.77149221 0.03437072  0.73605000    0.8060656
#> const      9.28879743 0.65724788  8.48341252   10.0063044
#> exo1       0.01271406 0.08334222 -0.06144414    0.1197343
#> exo2       0.05119054 0.11310564 -0.03256589    0.2063250
#> exo3       0.16738615 0.11560945  0.04847012    0.3030400
#> 
#> $AGO$A$equation2
#>                    mean        sd 5% quantile 95% quantile
#> lag1_var1   0.016198568 0.1392190  -0.1270923    0.1896264
#> lag1_var2   1.469928485 0.1161908   1.3206292    1.5818058
#> lag1_var3   0.615071215 0.2892409   0.3601327    0.9884156
#> lag1_var4  -0.392177595 0.1937415  -0.6502634   -0.2359217
#> const     -17.552228161 4.3830524 -22.5978042  -12.4062503
#> exo1       -0.008607991 0.1985813  -0.2314077    0.2282780
#> exo2        0.536498136 0.3428337   0.1719675    0.9592576
#> exo3       -0.792942452 0.5867807  -1.4292611   -0.1600571
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.23556741 0.06935848 -0.32195967  -0.16534798
#> lag1_var2 -0.87261734 0.09167466 -0.98343603  -0.77975122
#> lag1_var3 -0.12697236 0.12650539 -0.26449359   0.01722769
#> lag1_var4  0.90411489 0.09009605  0.79835958   0.97594910
#> const     22.96553040 2.64320444 19.76565138  25.68155637
#> exo1       0.08998682 0.13981613 -0.07171629   0.23801232
#> exo2      -1.37349859 0.19683564 -1.59556020  -1.13853242
#> exo3       3.50432795 0.24295635  3.22225010   3.76308553
#> 
#> $AGO$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.3058794 0.06311492 -0.37694102   -0.2459826
#> lag1_var2  0.2967670 0.08168288  0.20397898    0.3935284
#> lag1_var3  0.4228994 0.12916807  0.32241183    0.5979430
#> lag1_var4  0.6817420 0.11846142  0.52211166    0.7669471
#> const     -0.1860194 4.02824583 -5.35110995    3.8647442
#> exo1       0.1054112 0.15013099 -0.03746124    0.2971731
#> exo2      -1.2375239 0.14643357 -1.41598204   -1.0940727
#> exo3       3.3779159 0.30937127  3.06745842    3.7290741
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001690024 0.0007030637  0.00110393  0.002550072
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.04934324 0.03293761 -0.09262481  -0.02125745
#> Sigma[2,2]  8.94739952 3.27344731  6.06499176  13.27718562
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01222964 0.02647922 -0.009987848   0.04756607
#> Sigma[3,2] -4.97609792 1.96430632 -7.395196562  -2.95785404
#> Sigma[3,3]  4.82701786 1.31287899  3.150787667   6.09705871
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.02547223 0.01556685 -0.04575047    -0.010473
#> Sigma[4,2]  1.24138594 1.01069940  0.20211164     2.297187
#> Sigma[4,3]  1.77520026 1.05323155  0.42999250     2.759950
#> Sigma[4,4]  3.14089719 0.62488885  2.33890897     3.763901
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  1.021379311 0.03872126  0.97969115   1.06230847
#> lag1_var2  0.001233125 0.04212211 -0.04578742   0.04775547
#> lag1_var3 -0.009280564 0.06349113 -0.08021641   0.06019691
#> lag1_var4  0.009197554 0.05194108 -0.04841332   0.06739760
#> const     -0.550878532 1.59375284 -1.94809180   1.22603099
#> exo1       0.011522733 0.03158599 -0.02093328   0.05059384
#> exo2      -0.043354751 0.04625982 -0.09174173   0.01229436
#> exo3       0.045480217 0.03497081  0.01197348   0.08630147
#> 
#> $ALB$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  -7.1240844  2.117007   -9.719979   -5.4756677
#> lag1_var2   0.8708562  2.750259   -1.677601    4.2237883
#> lag1_var3   1.2209421  4.134421   -2.437541    6.3530418
#> lag1_var4  -1.2364803  3.466159   -5.546757    1.7658499
#> const     177.4270615 87.578150   89.834373  281.0611036
#> exo1       -2.6546347  4.094589   -8.174404    0.6675003
#> exo2       -1.1464041  3.066589   -4.643684    2.0982362
#> exo3       -1.8534321  4.747881   -7.754566    3.2398161
#> 
#> $ALB$A$equation3
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1   5.0585056  2.301446    2.0237986    7.1071516
#> lag1_var2  -1.3263520  1.651270   -2.4125813    0.9367442
#> lag1_var3  -1.9402863  2.523012   -3.5887803    1.5208208
#> lag1_var4   2.4169947  2.049946   -0.3762836    3.7577030
#> const     -91.2406506 78.467033 -180.6805440    1.9445634
#> exo1       -0.6373492  1.847334   -2.9122613    1.3479122
#> exo2       -0.5404529  2.787133   -2.6631929    3.2084029
#> exo3        1.1454729  3.335485   -2.1469089    5.2332231
#> 
#> $ALB$A$equation4
#>                 mean         sd   5% quantile 95% quantile
#> lag1_var1  0.2912810   2.478730  -2.729175930    2.6782568
#> lag1_var2 -1.2911965   2.495407  -4.408224651    0.6986178
#> lag1_var3 -1.9115283   3.736102  -6.498781353    1.0958717
#> lag1_var4  2.3787324   2.970425  -0.008976964    6.0573419
#> const     27.5575203 102.906888 -56.095131068  143.2783816
#> exo1      -2.4604831   1.676916  -4.568309156   -0.9113221
#> exo2      -1.3747094   1.927988  -3.592715520    0.8278282
#> exo3      -0.1494288   1.142902  -1.701070623    0.7641967
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[1,1] 0.001213902 0.000191429 0.0009985775  0.001437554
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd  5% quantile  95% quantile
#> Sigma[2,1] -0.001715748 0.000932696 -0.002696989 -0.0005969176
#> Sigma[2,2]  0.056306072 0.009117015  0.049292906  0.0685906664
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.005281915 0.001856046  0.002798442  0.006807614
#> Sigma[3,2] -0.065975145 0.014264285 -0.083901720 -0.052207544
#> Sigma[3,3]  0.199681321 0.037077981  0.160268313  0.243317461
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.004006609 0.001468894  0.002448637  0.005792012
#> Sigma[4,2] -0.021724134 0.009491953 -0.032628372 -0.012168882
#> Sigma[4,3]  0.150862284 0.028111831  0.123791281  0.186264330
#> Sigma[4,4]  0.138712757 0.023074956  0.119591978  0.164850143
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1  0.95390084 0.029550717  0.92623869  0.992710590
#> lag1_var2 -0.01283934 0.026158222 -0.04443005  0.012233461
#> lag1_var3 -0.04928226 0.007486708 -0.05766352 -0.041357064
#> lag1_var4  0.05151145 0.013856179  0.03839133  0.069229652
#> const      1.02404106 0.608630369  0.47240002  1.829383495
#> exo1      -0.01990808 0.036746325 -0.06068353  0.022944243
#> exo2      -0.12044329 0.033037119 -0.16086206 -0.085834396
#> exo3      -0.04344531 0.035655978 -0.07678967  0.003076715
#> 
#> $ARE$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.06017100 0.21620051 -0.10786434   0.35353689
#> lag1_var2  0.88988833 0.09324844  0.76224730   0.95123027
#> lag1_var3  0.13279590 0.08099536  0.03264166   0.21064182
#> lag1_var4 -0.15411436 0.14102212 -0.32806885  -0.01940264
#> const      0.54556974 2.29395579 -1.81254060   3.13285948
#> exo1       0.05201846 0.27033426 -0.26832301   0.35791283
#> exo2       2.02311361 0.11983544  1.89694599   2.16537766
#> exo3      -0.71911090 0.25360435 -0.98418616  -0.44895069
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.2524885 0.2868927 -0.05583367    0.5970200
#> lag1_var2 -0.8588259 0.3232105 -1.26361200   -0.5412140
#> lag1_var3 -0.5638601 0.1312667 -0.69600663   -0.3963889
#> lag1_var4  1.4592680 0.1889150  1.26146920    1.6932077
#> const      0.7160786 2.5169126 -2.50889313    3.2408553
#> exo1      -0.8481393 0.3032494 -1.14480054   -0.4988378
#> exo2      -2.1048669 0.1669395 -2.28669921   -1.9033722
#> exo3      -1.2432086 0.5725383 -1.67170882   -0.4707718
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.2515869 0.2413446  -0.0605483    0.4705215
#> lag1_var2 -0.2115137 0.3169865  -0.6184097    0.1021861
#> lag1_var3 -0.4958227 0.1214115  -0.6079377   -0.3372218
#> lag1_var4  1.4126684 0.1298527   1.2607172    1.5617155
#> const     -0.5029571 1.9722197  -2.2062974    1.9194745
#> exo1      -0.7877366 0.2770613  -1.1036513   -0.4583201
#> exo2      -0.5094158 0.2698472  -0.8202740   -0.2234664
#> exo3      -1.8210864 0.5309447  -2.3033132   -1.1441030
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003827415 0.0002487173 0.003657947  0.004168357
#> 
#> $ARG$Sigma$equation2
#>                  mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0682833 0.008641444 -0.07812588  -0.05868785
#> Sigma[2,2]  3.3175603 0.326969090  2.90880595   3.61394119
#> 
#> $ARG$Sigma$equation3
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.06263614 0.005433203  0.05535148   0.06709508
#> Sigma[3,2] -2.52926344 0.311751424 -2.80310501  -2.11848418
#> Sigma[3,3]  2.19569091 0.334948006  1.73755459   2.44332544
#> 
#> $ARG$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1]  0.02372016 0.00302270   0.0211806   0.02755882
#> Sigma[4,2] -0.58523382 0.17487960  -0.7122937  -0.36124878
#> Sigma[4,3]  0.74284436 0.16166874   0.5189709   0.84171398
#> Sigma[4,4]  0.43530505 0.07728302   0.3316113   0.49265377
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.92108991 0.06550563  0.84676660   0.99182606
#> lag1_var2 -0.19303586 0.02090578 -0.21715583  -0.17178061
#> lag1_var3 -0.31546325 0.03923894 -0.36162257  -0.27640355
#> lag1_var4  0.27428802 0.03221549  0.24355289   0.31538156
#> const      4.65314848 1.75608948  2.75127584   6.61000243
#> exo1       0.05982943 0.08316340 -0.02050423   0.15489606
#> exo2      -0.12144013 0.10864895 -0.22848592   0.01358995
#> exo3       0.06368578 0.05696475  0.02113076   0.13988024
#> 
#> $ARG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -2.7417373  0.6330019  -3.5266970    -2.161111
#> lag1_var2 -1.9542615  0.4919430  -2.5995282    -1.489260
#> lag1_var3 -4.3439020  0.8562146  -5.4360559    -3.486234
#> lag1_var4  4.2138587  0.8419437   3.5006021     5.348269
#> const     84.8648115 23.4769324  66.0896406   114.820652
#> exo1      -0.4511048  1.7939288  -2.2701084     1.683228
#> exo2       2.8380802  2.0218261   0.7484731     4.888048
#> exo3       1.1755438  2.6169980  -1.8730988     4.195358
#> 
#> $ARG$A$equation3
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1   0.95631943  0.5304415   0.3520950     1.553703
#> lag1_var2   0.85567992  0.2898569   0.5241009     1.197573
#> lag1_var3   2.11887183  0.5153236   1.4888613     2.672953
#> lag1_var4  -1.38669828  0.4134670  -1.9448879    -1.054332
#> const     -11.18872932 26.5032333 -39.1604370    19.136218
#> exo1       -0.02220497  1.5850387  -1.9030159     1.719611
#> exo2       -5.50256368  2.0547404  -7.6647588    -3.264179
#> exo3        2.11999812  1.9893522   0.1285714     4.610020
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.8140019  0.3751785  -1.1553696   -0.3745017
#> lag1_var2 -0.3577162  0.1098900  -0.4350606   -0.2073789
#> lag1_var3 -0.5643002  0.2115632  -0.7666414   -0.3007249
#> lag1_var4  1.2980600  0.2032876   1.0534298    1.5070221
#> const     38.3100540 18.9641403  18.1254700   61.7489113
#> exo1      -0.3902240  0.5926667  -1.0832445    0.2985009
#> exo2      -4.3672146  1.0165490  -5.5196377   -3.2354724
#> exo3       3.2143435  0.9986829   1.9129678    4.1295082
#> 
#> 
#>