Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004737351 0.0005804741 0.004038868  0.005316363
#> 
#> $AFG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01549192 0.005519112 -0.02288945  -0.01076127
#> Sigma[2,2]  0.21729576 0.050105477  0.17486513   0.28412382
#> 
#> $AFG$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01424392 0.005457594  0.009740074   0.02161391
#> Sigma[3,2] -0.18083487 0.065334604 -0.270554649  -0.13456929
#> Sigma[3,3]  0.28500444 0.060590905  0.230014559   0.36482209
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.008598014 0.003221345  0.006098992   0.01294596
#> Sigma[4,2] -0.094584361 0.047984550 -0.160138923  -0.06300054
#> Sigma[4,3]  0.234003134 0.042558727  0.193399903   0.28210137
#> Sigma[4,4]  0.222304679 0.039614949  0.189227783   0.27479486
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.99495977 0.02925271  0.95488746  1.017183039
#> lag1_var2 -0.05614594 0.07227363 -0.14294329 -0.002053505
#> lag1_var3 -0.10920861 0.15850613 -0.30423945  0.031737916
#> lag1_var4  0.10224554 0.13746486 -0.02238938  0.271717712
#> const      0.54967127 0.92141106 -0.48354997  1.426785979
#> exo1      -0.06296264 0.08140262 -0.14113845  0.042930379
#> exo2      -0.09813013 0.06416048 -0.18472303 -0.051538912
#> exo3      -0.16550061 0.11396445 -0.32094250 -0.078546263
#> 
#> $AFG$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.22769991 0.2905727  0.01203758    0.6201541
#> lag1_var2  0.71526505 0.4790756  0.16433294    1.2793108
#> lag1_var3 -0.56010162 0.9625742 -1.75135785    0.4619547
#> lag1_var4  0.45877482 0.8101501 -0.37359561    1.4828603
#> const     -0.24430629 8.1199897 -6.91074723   10.6458234
#> exo1      -0.52703055 0.2822156 -0.84226165   -0.1935305
#> exo2       0.37648632 0.2774847  0.01023126    0.6318653
#> exo3      -0.04580557 0.5148831 -0.70360997    0.4786195
#> 
#> $AFG$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.1759053 0.3789636  -0.6877030    0.1379688
#> lag1_var2  0.5522287 0.5234830  -0.1023394    1.0960140
#> lag1_var3  2.8352667 0.9624341   1.7259753    3.9548971
#> lag1_var4 -1.7211290 0.7988091  -2.6793636   -0.8419573
#> const      1.0759427 8.9545963  -8.4141839   11.5010123
#> exo1       0.3476495 0.5468533  -0.3596181    0.8675630
#> exo2      -1.1776605 0.5347740  -1.7318918   -0.5370550
#> exo3      -0.2090439 0.6319338  -0.7128613    0.6211986
#> 
#> $AFG$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.09055057 0.2988516  -0.4795105    0.1747927
#> lag1_var2  1.00269480 0.3620914   0.5343101    1.3456686
#> lag1_var3  2.93527815 0.5950286   2.1887728    3.5429939
#> lag1_var4 -1.70732766 0.4775797  -2.2132575   -1.1213827
#> const     -5.67065491 6.7625738 -11.7390239    2.6971028
#> exo1       0.12578085 0.5508394  -0.6266533    0.5552095
#> exo2      -1.01953571 0.4706980  -1.5133500   -0.5287608
#> exo3      -0.16866218 0.5275140  -0.7553545    0.4023536
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.008907998 0.002427314 0.006237734    0.0112852
#> 
#> $AGO$Sigma$equation2
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.006685144 0.003978825 -0.01069378 -0.001945862
#> Sigma[2,2]  0.149278511 0.015709947  0.13623632  0.170082122
#> 
#> $AGO$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01276996 0.003556773  0.009468016   0.01746104
#> Sigma[3,2] -0.04138774 0.013942092 -0.054798094  -0.02499528
#> Sigma[3,3]  0.03918922 0.011985288  0.024698564   0.05233726
#> 
#> $AGO$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1] 0.007432830 0.004155322  0.002755382   0.01221300
#> Sigma[4,2] 0.079038243 0.024027521  0.060446033   0.11139748
#> Sigma[4,3] 0.005164297 0.010156521 -0.005427797   0.01685743
#> Sigma[4,4] 0.070334389 0.030223654  0.044029257   0.11087699
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.01078826 0.03792885  0.96538442    1.0517937
#> lag1_var2 -0.41884161 0.06634756 -0.48958590   -0.3483360
#> lag1_var3 -0.47246112 0.09919368 -0.58372633   -0.3734464
#> lag1_var4  0.47497855 0.09606114  0.37349188    0.5802850
#> const      0.48018112 1.24037849 -0.65767986    2.0632279
#> exo1       0.08044133 0.10629552 -0.01695324    0.2211815
#> exo2       0.04026699 0.07876652 -0.04194922    0.1388973
#> exo3       0.28029511 0.18998415  0.01893541    0.4120551
#> 
#> $AGO$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.08689967 0.2169224 -0.18306234    0.3043913
#> lag1_var2 -0.06495499 0.4543434 -0.66990616    0.3296178
#> lag1_var3 -1.32216784 0.6603772 -2.14645711   -0.6390752
#> lag1_var4  1.21387754 0.5814763  0.64300620    1.9486619
#> const      6.93593721 1.3686373  5.37224509    8.3094921
#> exo1       0.30127853 0.4004463 -0.03104826    0.7971325
#> exo2       0.74630939 0.7566146  0.13375403    1.7713510
#> exo3      -0.34739397 0.6280713 -0.89742664    0.4926753
#> 
#> $AGO$A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.1286452 0.06165363  -0.2055848  -0.07432663
#> lag1_var2 -0.3743986 0.06046351  -0.4392555  -0.30094789
#> lag1_var3  0.5907562 0.09630413   0.4585547   0.65318012
#> lag1_var4  0.4295480 0.06764081   0.3816082   0.52266182
#> const      2.5115972 2.41188003  -0.3063512   4.61521096
#> exo1       0.0105585 0.27063698  -0.3008625   0.32641034
#> exo2      -1.2682496 0.25769135  -1.5513295  -0.98514755
#> exo3       3.7161067 0.18045126   3.5034672   3.92382113
#> 
#> $AGO$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.1535197 0.1632039 -0.32477429   0.05107535
#> lag1_var2 -0.4478974 0.3378357 -0.90473827  -0.17273503
#> lag1_var3 -0.4634055 0.4925764 -1.12089104  -0.03989785
#> lag1_var4  1.5040976 0.4474444  1.12888923   2.09390714
#> const      2.0278117 2.3594107 -0.49892349   4.70465340
#> exo1       0.3402199 0.2348971  0.07304314   0.60311574
#> exo2      -1.0086381 0.4213254 -1.30365317  -0.42852510
#> exo3       3.8549822 0.5373943  3.45219827   4.59011924
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005305979 0.001875387 0.003643427  0.007380266
#> 
#> $ALB$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.1362429 0.09057634  -0.2585455  -0.08745805
#> Sigma[2,2]  8.8067049 5.89081657   5.3556795  16.86142311
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.09441615 0.05544331    0.050709    0.1700146
#> Sigma[3,2] -5.59687595 3.38011731  -10.214162   -3.8273287
#> Sigma[3,3]  5.58424831 2.01118577    4.303070    8.3016828
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] 0.005825948 0.02345693 -0.01737761   0.03454557
#> Sigma[4,2] 0.351766780 0.77473756 -0.33796368   1.36293695
#> Sigma[4,3] 2.247424978 0.61594361  1.64724608   2.92990182
#> Sigma[4,4] 3.039651346 0.43514838  2.48045229   3.45235491
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.913928334 0.02153686   0.8846728   0.92940065
#> lag1_var2  0.146606110 0.01049327   0.1379502   0.15996048
#> lag1_var3  0.226272909 0.01925285   0.2076351   0.25034129
#> lag1_var4 -0.193195213 0.01506048  -0.2127967  -0.18047964
#> const     -0.135224188 0.68221269  -0.5256779   0.80520523
#> exo1      -0.010132635 0.09318121  -0.1228791   0.06634372
#> exo2      -0.004484593 0.07937668  -0.0582190   0.10472348
#> exo3       0.149658974 0.04485374   0.1129785   0.21050002
#> 
#> $ALB$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -1.3614830  2.320862  -3.1794699    1.7821184
#> lag1_var2  1.3514590  1.098864   0.4996573    2.8473388
#> lag1_var3  1.3981347  1.449592   0.2542180    3.3492650
#> lag1_var4 -0.9559488  1.270323  -2.6492387    0.1161476
#> const     12.3723027 59.280187 -65.2772865   67.9339230
#> exo1      -1.9153860  2.585363  -4.3900271    1.3839017
#> exo2      -2.7912852  4.081780  -8.2457418    0.7293469
#> exo3      -1.5348657  3.319101  -4.0444652    2.7110254
#> 
#> $ALB$A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.3197713  1.9448327  -2.3129129    1.3609724
#> lag1_var2  0.3220574  0.8767509  -0.4822981    1.4673582
#> lag1_var3  1.2284171  1.3394633   0.1110298    2.9910008
#> lag1_var4 -0.3743326  1.1513596  -1.9043882    0.6023321
#> const     -1.3802538 52.6333603 -38.7340903   70.3294562
#> exo1       0.2189373  2.6438818  -3.0510878    2.7423848
#> exo2       0.1844593  2.8364591  -2.1909694    3.9356582
#> exo3       1.7864402  2.2451269  -1.2419573    3.5482584
#> 
#> $ALB$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.7573741  0.8177794  -1.7011753  -0.06899782
#> lag1_var2  1.1197059  0.9821756   0.3387590   2.42025350
#> lag1_var3  2.1143520  1.4395598   1.0723556   4.05166221
#> lag1_var4 -0.8105347  1.2884812  -2.5402436   0.21737286
#> const      1.6581909 23.1681003 -20.0957750  31.99231705
#> exo1      -0.7841087  1.3636566  -1.9602210   0.97748881
#> exo2      -1.9590022  1.4055566  -3.6931581  -0.43531041
#> exo3       1.0478125  0.5699470   0.4679601   1.76771596
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001736959 0.0004126943  0.00119402  0.002078257
#> 
#> $ARE$Sigma$equation2
#>                    mean           sd 5% quantile 95% quantile
#> Sigma[2,1] -0.001250917 0.0006939324 -0.00219458 -0.000766254
#> Sigma[2,2]  0.048395983 0.0101678953  0.03596063  0.059468846
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.006358734 0.002333633  0.003667919  0.009075877
#> Sigma[3,2] -0.041292175 0.009072390 -0.053638316 -0.035489974
#> Sigma[3,3]  0.148788195 0.030280414  0.116943840  0.181412890
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.005688655 0.002567737  0.003200149  0.008795961
#> Sigma[4,2] -0.004200997 0.011064351 -0.015856042  0.009060069
#> Sigma[4,3]  0.117937359 0.027448947  0.086949894  0.145405237
#> Sigma[4,4]  0.116578904 0.024688531  0.087396872  0.140798710
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1  0.94603449 0.016953703  0.92956627  0.966180112
#> lag1_var2  0.37096479 0.010458146  0.35944188  0.383112357
#> lag1_var3  0.45291050 0.007237848  0.44328599  0.459373338
#> lag1_var4 -0.43786213 0.009163145 -0.44965883 -0.429048110
#> const      0.25316129 0.183599969  0.08083312  0.451748131
#> exo1      -0.03678084 0.038300616 -0.07029814  0.014299092
#> exo2      -0.06521841 0.069135948 -0.13500313  0.021486662
#> exo3      -0.02592804 0.033406293 -0.06713416  0.007893482
#> 
#> $ARE$A$equation2
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.152628781 0.14052819  0.05461468    0.3431788
#> lag1_var2  0.459001184 0.11168457  0.36752315    0.6075486
#> lag1_var3 -0.388952675 0.08500623 -0.48818061   -0.3100654
#> lag1_var4  0.293185890 0.10385639  0.16660341    0.3981364
#> const      3.876814286 1.91073749  1.79794818    5.8349272
#> exo1      -0.003591668 0.18467407 -0.16343038    0.2442845
#> exo2       2.125757385 0.19128732  1.86943627    2.2905629
#> exo3      -0.720294845 0.23892181 -1.02124378   -0.4740464
#> 
#> $ARE$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.07561147 0.2317994  -0.1724688    0.3103424
#> lag1_var2  0.39268438 0.2943138   0.1126960    0.7108666
#> lag1_var3  1.07677846 0.3257730   0.7759112    1.4913613
#> lag1_var4 -0.07662870 0.3419176  -0.5111293    0.2342039
#> const     -2.54843778 2.7993030  -5.9961770    0.4793840
#> exo1      -0.48966203 0.1726826  -0.6392417   -0.2581311
#> exo2      -2.30309511 0.2913299  -2.6190174   -1.9434082
#> exo3      -1.41765409 0.4065342  -1.9019424   -1.0302369
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.2487413 0.2135892 -0.03789844    0.4182111
#> lag1_var2  0.7717934 0.2483227  0.54949813    1.0436779
#> lag1_var3  0.8339352 0.2785695  0.58538246    1.1797635
#> lag1_var4  0.1137618 0.2735515 -0.24343185    0.3521215
#> const     -2.6191991 1.4019702 -4.07901898   -0.8695064
#> exo1      -0.5196369 0.2492737 -0.76141312   -0.2140084
#> exo2      -0.6579707 0.1928989 -0.87263439   -0.4607216
#> exo3      -2.0070751 0.3024268 -2.32547387   -1.6429792
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002998601 0.001988542 0.001621994  0.005687878
#> 
#> $ARG$Sigma$equation2
#>                   mean        sd 5% quantile 95% quantile
#> Sigma[2,1] -0.08336858 0.0766188  -0.1871675   -0.0382993
#> Sigma[2,2]  4.63564587 2.9114704   2.9907994    8.6082606
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.05946399 0.04688756  0.03244015    0.1233726
#> Sigma[3,2] -3.17555895 1.81010654 -5.64816315   -2.0300379
#> Sigma[3,3]  2.37011868 1.11361278  1.53570715    3.8857518
#> 
#> $ARG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.009816875 0.002136407  0.007236459   0.01211174
#> Sigma[4,2] -0.366217290 0.124430405 -0.479307450  -0.20317571
#> Sigma[4,3]  0.470119276 0.129841182  0.319781722   0.61677754
#> Sigma[4,4]  0.277813208 0.068968420  0.218626051   0.36529392
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.899456433 0.02243845  0.87451855   0.92644334
#> lag1_var2  0.226891371 0.13799354  0.13289806   0.41417184
#> lag1_var3  0.368288157 0.21543273  0.22294587   0.66080260
#> lag1_var4 -0.337642178 0.20879350 -0.61908515  -0.19085507
#> const      0.839575684 1.05528839  0.28318302   2.28876794
#> exo1       0.030669245 0.06997431 -0.05914648   0.09550883
#> exo2      -0.115700402 0.04325593 -0.16182379  -0.06312089
#> exo3       0.005391527 0.13978036 -0.17596385   0.14194002
#> 
#> $ARG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -1.0450209  0.9360599  -2.1991912   -0.1024767
#> lag1_var2  1.5407460  5.1354179  -5.4317287    5.6443170
#> lag1_var3  1.4487851  8.0932012  -9.4905646    8.1294018
#> lag1_var4 -0.9656367  7.7744237  -7.6184521    9.4818176
#> const      2.5395259 33.6082203 -43.1613791   28.2128330
#> exo1      -2.1826576  2.9688501  -5.6226656    1.2155448
#> exo2       1.8320814  1.1434144   0.6503401    3.0826850
#> exo3      -0.4965524  5.2723783  -5.4847226    6.4297075
#> 
#> $ARG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.02090932  0.5750726  -0.6868551    0.6409792
#> lag1_var2  1.50765940  3.3947163  -1.4435911    6.0433350
#> lag1_var3  3.12929802  5.3436025  -1.6460178   10.2307794
#> lag1_var4 -2.25278575  5.1964618  -9.0860942    2.5974044
#> const      5.81766994 24.7507210  -9.9832489   39.6282551
#> exo1       0.91107595  2.1063883  -1.5557273    3.2975624
#> exo2      -4.95445797  0.9383886  -5.8966010   -3.7597175
#> exo3       3.47252800  3.5989345  -1.0401844    7.2697294
#> 
#> $ARG$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.7585246 0.1881925  -0.9238731   -0.5211843
#> lag1_var2  2.6557219 0.3441286   2.1906146    2.9024914
#> lag1_var3  4.3526450 0.5287709   3.6379835    4.7190795
#> lag1_var4 -3.0684845 0.6059925  -3.4878001   -2.2416505
#> const      3.1947351 5.8767843  -1.9554201   10.9761918
#> exo1      -0.4816974 0.6356685  -1.2138856    0.1628791
#> exo2      -4.3780023 0.6630995  -4.9756858   -3.6476206
#> exo3       3.5167330 0.6685684   2.8962842    4.3543152
#> 
#> 
#>