Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003701637 0.0004419389 0.003198624  0.004177096
#> 
#> $AFG$Sigma$equation2
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.007226255 0.002915225 -0.009566785 -0.003325464
#> Sigma[2,2]  0.201281431 0.051571302  0.142617579  0.256975063
#> 
#> $AFG$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.005755737 0.007285161 -0.004142354   0.01010745
#> Sigma[3,2] -0.179501651 0.071050123 -0.272831820  -0.11240844
#> Sigma[3,3]  0.359820037 0.148171541  0.236481892   0.55768873
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.003472548 0.007351694 -0.006662871  0.007984498
#> Sigma[4,2] -0.102556115 0.063389777 -0.187776067 -0.056172433
#> Sigma[4,3]  0.320219893 0.135136343  0.198398083  0.497577718
#> Sigma[4,4]  0.316267304 0.124642193  0.191769836  0.474168404
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.980657777 0.02760029   0.9564791   1.01765417
#> lag1_var2 -0.009559398 0.06018765  -0.0767788   0.05515657
#> lag1_var3 -0.011877607 0.15626856  -0.1819697   0.15974323
#> lag1_var4  0.010715743 0.14409853  -0.1501353   0.16471762
#> const      0.596196452 0.93226630  -0.6365088   1.42801730
#> exo1       0.006824255 0.07640474  -0.0582974   0.10911776
#> exo2      -0.052454349 0.08402294  -0.1605704   0.01843157
#> exo3      -0.111491245 0.06619049  -0.1790527  -0.02898978
#> 
#> $AFG$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.21378559 0.2494006  -0.0675792   0.49069672
#> lag1_var2  0.57401081 0.8662277  -0.2395995   1.55759300
#> lag1_var3 -0.86440353 1.7527456  -2.6198208   1.03832745
#> lag1_var4  0.72616932 1.5061087  -0.8564619   2.28684103
#> const      1.89637884 8.2371369  -7.3575111  11.62636277
#> exo1      -0.27922212 0.4435520  -0.8501835   0.05868141
#> exo2       0.08367122 0.4405455  -0.3761434   0.62819476
#> exo3      -0.14399083 0.7954801  -0.8224320   0.91852801
#> 
#> $AFG$A$equation3
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1 -0.049964276  0.3099604  -0.30400811    0.3152374
#> lag1_var2  0.172561274  1.0310582  -0.95110572    1.3217954
#> lag1_var3  2.118047766  2.0128543   0.01556805    4.2838401
#> lag1_var4 -1.093307955  1.7045754  -2.87470868    0.6928175
#> const      2.631345859 10.7253357 -11.95528243   10.1573764
#> exo1       0.008252257  0.5786073  -0.78343631    0.4262752
#> exo2      -1.153615488  0.4684994  -1.57615911   -0.5482750
#> exo3      -0.023147193  0.7063938  -0.99254059    0.3719979
#> 
#> $AFG$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.02787276 0.2276269  -0.1561679    0.3276083
#> lag1_var2  0.48123573 0.7247414  -0.3228466    1.3500192
#> lag1_var3  1.88899625 1.3137456   0.5298903    3.4829970
#> lag1_var4 -0.78136389 1.0838607  -2.0673298    0.3396889
#> const     -2.56004586 8.6400916 -14.2577644    3.3870500
#> exo1      -0.08907565 0.5997647  -0.8859522    0.3855697
#> exo2      -1.13374677 0.3780961  -1.4736103   -0.6333073
#> exo3       0.04071751 0.4110469  -0.5208612    0.3286608
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                 mean         sd 5% quantile 95% quantile
#> Sigma[1,1] 0.0346491 0.02298315  0.01650542   0.06557222
#> 
#> $AGO$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0383276 0.04724318  -0.1031869 -0.003970156
#> Sigma[2,2]  0.1809911 0.06474411   0.1332184  0.269799733
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1] -0.006541799 0.003402463 -0.009786832 -0.002716037
#> Sigma[3,2] -0.042169539 0.015453954 -0.059798980 -0.024007999
#> Sigma[3,3]  0.051107335 0.012233441  0.034708200  0.061119180
#> 
#> $AGO$Sigma$equation4
#>                   mean         sd   5% quantile 95% quantile
#> Sigma[4,1] -0.03871190 0.04048771 -0.0943214651  -0.01075672
#> Sigma[4,2]  0.10827186 0.06715329  0.0590330306   0.20030375
#> Sigma[4,3]  0.02007786 0.02127808 -0.0001259725   0.04545909
#> Sigma[4,4]  0.11538994 0.07705640  0.0542365033   0.21898327
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.8068528 0.1318817   0.6263139   0.89156683
#> lag1_var2  0.6594931 0.7117230   0.1546141   1.62205468
#> lag1_var3  0.8885033 0.8604172   0.3096743   2.06755151
#> lag1_var4 -0.8326616 0.8017858  -1.9284161  -0.29745363
#> const      0.8571827 2.2603603  -1.9866876   3.16680889
#> exo1       0.2528116 0.1840909   0.1062317   0.50344701
#> exo2      -0.1462529 0.1361213  -0.2476099   0.03904698
#> exo3      -0.2076336 0.3283752  -0.5910390   0.13574816
#> 
#> $AGO$A$equation2
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1  0.04677524 0.3619968  -0.35850728    0.4470438
#> lag1_var2  0.59616296 1.7776514  -1.56356844    2.3656689
#> lag1_var3 -0.39113391 2.0944901  -2.92213379    1.7250074
#> lag1_var4  0.43806975 1.9783947  -1.50189230    2.8275991
#> const     -3.10863625 8.6879422 -14.00126608    5.5635107
#> exo1      -0.17365162 0.6991723  -0.94284377    0.6669817
#> exo2       0.52535832 0.5190402  -0.05174642    1.0761361
#> exo3      -0.43384223 0.7986634  -1.30578385    0.4891735
#> 
#> $AGO$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.07083690 0.1124788  -0.1999313   0.06048685
#> lag1_var2 -0.90122995 0.5068085  -1.4727727  -0.33119885
#> lag1_var3 -0.19322861 0.6826911  -0.9116649   0.63798116
#> lag1_var4  1.03873774 0.6415947   0.2502624   1.67618424
#> const     13.29564213 6.4947435   6.7849816  21.57860023
#> exo1       0.04881608 0.2629212  -0.2853016   0.29982321
#> exo2      -1.45576603 0.2316025  -1.7283137  -1.22765295
#> exo3       3.44505792 0.1217862   3.3003264   3.55742602
#> 
#> $AGO$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.09123808 0.2701215  -0.3296040    0.2647868
#> lag1_var2 -0.89107516 1.0885121  -2.3816730   -0.1109009
#> lag1_var3 -1.05869590 1.3191713  -2.8696603   -0.1804947
#> lag1_var4  1.99518357 1.2472698   1.1800514    3.7113265
#> const      8.28231260 0.6530149   7.7062564    9.1533272
#> exo1      -0.07464683 0.3473335  -0.4787087    0.3199643
#> exo2      -1.31427540 0.5548059  -1.9656568   -0.7420365
#> exo3       3.62119692 0.4984013   3.0755477    4.2194682
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003708635 0.002545786 0.001728938  0.007177274
#> 
#> $ALB$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0846315 0.04891507  -0.1502003  -0.04200577
#> Sigma[2,2]  7.5058549 1.19961382   5.9820021   8.52838181
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.02254571 0.06751714 -0.05061112    0.1052011
#> Sigma[3,2] -4.75095876 1.42272313 -5.86672521   -2.8401531
#> Sigma[3,3]  5.31832944 0.83617778  4.27657384    6.1777699
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.03924424 0.05020027  -0.1055464  0.007129829
#> Sigma[4,2]  0.31589727 0.99838553  -0.5870561  1.586285193
#> Sigma[4,3]  2.56000164 1.09637382   1.3198627  3.778369257
#> Sigma[4,4]  3.32890655 1.62420513   1.9517436  5.488895071
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.99722276 0.01981414  0.976000985  1.021614027
#> lag1_var2 -0.01615457 0.03658340 -0.066218893  0.003845888
#> lag1_var3 -0.04303983 0.06082408 -0.126144481 -0.003306875
#> lag1_var4  0.03357952 0.04878255  0.004168296  0.100012232
#> const      0.51229868 0.36851252  0.104248324  0.956616800
#> exo1       0.03656657 0.06920195 -0.054565838  0.095809823
#> exo2      -0.06254043 0.06316391 -0.143432226  0.001467499
#> exo3       0.04427376 0.04834242 -0.005247952  0.105367873
#> 
#> $ALB$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -3.80059089  1.445550   -5.451628   -2.0830169
#> lag1_var2  3.78441669  1.791623    1.682750    5.5433963
#> lag1_var3  5.57628749  2.726826    2.238910    8.3176571
#> lag1_var4 -4.62256081  2.315610   -6.993065   -1.7829060
#> const     38.91345277 32.090318    4.762824   77.3245727
#> exo1      -2.00177654  1.700899   -4.277979   -0.5008676
#> exo2      -0.01976934  4.076514   -4.138756    4.3367314
#> exo3       1.77879720  3.950509   -3.506468    5.1938408
#> 
#> $ALB$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1   3.0770724  2.0709962   0.9365066     5.667651
#> lag1_var2  -2.1351977  0.9388035  -3.2623707    -1.212174
#> lag1_var3  -2.9594461  1.6233323  -4.8961779    -1.361071
#> lag1_var4   3.1988118  1.3844531   1.8667229     4.903405
#> const     -28.8477202 44.0537796 -81.4200151    15.201516
#> exo1        0.4950196  2.2843184  -2.3305137     2.806734
#> exo2       -0.7699137  3.1481833  -4.2707318     2.252155
#> exo3       -1.7806093  2.9095642  -5.1079467     1.568870
#> 
#> $ALB$A$equation4
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.618788316  1.3771581  -0.5768062    2.4052832
#> lag1_var2  0.007218025  0.9201455  -0.6098295    1.2764135
#> lag1_var3  0.241693137  1.6539844  -0.7456020    2.5297703
#> lag1_var4  0.697591382  1.3362306  -1.1429335    1.5559771
#> const     -8.418159084 35.1885541 -51.0412331   30.1447646
#> exo1      -0.531771577  2.1645006  -3.3938164    1.4729399
#> exo2      -0.878213330  1.0145962  -1.7639021    0.4451795
#> exo3      -0.662751205  1.6057708  -2.4671111    1.2511983
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002721491 0.001283523  0.00166121  0.004458549
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.005336508 0.004512004 -0.01090673 -0.001198493
#> Sigma[2,2]  0.056619754 0.020432787  0.03722398  0.082185398
#> 
#> $ARE$Sigma$equation3
#>                    mean         sd   5% quantile 95% quantile
#> Sigma[3,1]  0.004178932 0.00487773 -0.0004239592   0.01048373
#> Sigma[3,2] -0.048064536 0.01615431 -0.0688912562  -0.03241931
#> Sigma[3,3]  0.152248072 0.01567137  0.1345769642   0.16843296
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.000144705 0.001671049 -0.001282175  0.002384795
#> Sigma[4,2] -0.004918573 0.008879714 -0.015323760  0.005345791
#> Sigma[4,3]  0.116622386 0.024648859  0.084964256  0.140970891
#> Sigma[4,4]  0.114104800 0.025419426  0.080746093  0.137982322
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.85419052 0.02532801  0.82037701   0.87659299
#> lag1_var2  0.61023076 0.02908896  0.57806866   0.64210956
#> lag1_var3  0.72644195 0.04484942  0.66820853   0.77002554
#> lag1_var4 -0.68489516 0.05053621 -0.73208061  -0.61883092
#> const      0.56398562 0.56023283 -0.13650397   1.00165383
#> exo1      -0.03093125 0.05711778 -0.09198757   0.02892216
#> exo2      -0.11857153 0.02619201 -0.14830267  -0.09137532
#> exo3      -0.09981766 0.08941905 -0.20817415  -0.00154853
#> 
#> $ARE$A$equation2
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.17552666 0.15830638 -0.010035864   0.32220971
#> lag1_var2  0.64377892 0.06581175  0.552999230   0.68753847
#> lag1_var3 -0.14139450 0.10422947 -0.271491519  -0.05535514
#> lag1_var4  0.10216500 0.12996549 -0.002238479   0.26781957
#> const     -1.02813069 2.59181641 -3.630188840   2.19342553
#> exo1      -0.09350933 0.19061974 -0.330657927   0.08869133
#> exo2       1.90061684 0.17583906  1.686942428   2.07634074
#> exo3      -0.59771516 0.30609616 -1.000874876  -0.34197248
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.2478675 0.1970692   0.0755772    0.5122001
#> lag1_var2 -1.3903643 0.4072603  -1.8956292   -0.9620457
#> lag1_var3 -1.0233730 0.6116699  -1.6679956   -0.2610933
#> lag1_var4  1.9346683 0.6108356   1.2014851    2.6110745
#> const      0.1586218 3.2324448  -4.0040934    3.2463851
#> exo1      -0.4556772 0.2646482  -0.6849630   -0.1252187
#> exo2      -2.1734902 0.5805608  -2.5298045   -1.3835327
#> exo3      -1.3083264 0.4186598  -1.8075254   -0.9267800
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.3716032 0.1854423   0.1567547   0.56154741
#> lag1_var2 -0.9381090 0.3994296  -1.4197385  -0.49419566
#> lag1_var3 -1.1951571 0.6437714  -1.8797497  -0.40310107
#> lag1_var4  2.1019270 0.6545086   1.3089922   2.82443093
#> const     -2.3155513 2.8838769  -6.1999407  -0.15455870
#> exo1      -0.5470002 0.2919925  -0.8312871  -0.16706079
#> exo2      -0.7056093 0.4723863  -0.9909067  -0.05190651
#> exo3      -1.8334489 0.3281527  -2.1745086  -1.44403132
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002825834 0.000632849 0.002028489   0.00348275
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.06111728 0.01592986  -0.0774431  -0.04387034
#> Sigma[2,2]  3.14017461 0.83462642   2.2026051   4.11410634
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04707681 0.01242549  0.03343578   0.06222355
#> Sigma[3,2] -2.29704781 0.49284551 -2.77304862  -1.67250542
#> Sigma[3,3]  1.92223070 0.33779361  1.46965297   2.20138429
#> 
#> $ARG$Sigma$equation4
#>                  mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.0112247 0.005138373  0.006497863   0.01788339
#> Sigma[4,2] -0.4160766 0.127850269 -0.552437323  -0.26211332
#> Sigma[4,3]  0.5705750 0.136654794  0.406753593   0.70664944
#> Sigma[4,4]  0.3501847 0.081218127  0.277096058   0.45618689
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.93106651 0.02199043  0.90534408   0.95430010
#> lag1_var2 -0.02043575 0.06981142 -0.11495382   0.01356511
#> lag1_var3 -0.03597435 0.11588972 -0.19251420   0.02510098
#> lag1_var4  0.02395616 0.10068826 -0.03233203   0.16206188
#> const      2.58693733 1.41845476  0.94235772   3.83561425
#> exo1       0.02620189 0.06321937 -0.01524946   0.11341407
#> exo2      -0.12162403 0.03926775 -0.16188269  -0.07601846
#> exo3       0.04113679 0.05468185 -0.01191139   0.10722893
#> 
#> $ARG$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -2.02198405  0.7228115  -2.9108451   -1.3041348
#> lag1_var2  2.65096463  2.6551113   1.2258063    6.2705632
#> lag1_var3  3.29766068  4.4359781   0.7745998    9.3726947
#> lag1_var4 -2.38682224  3.9899527  -7.8011778   -0.1003329
#> const      2.89953968 23.1073298 -26.7735262   21.9816582
#> exo1      -1.22026958  2.3435705  -4.2734693    0.9487584
#> exo2       2.26874905  1.4995511   1.0224831    4.2453451
#> exo3      -0.04662126  2.5821059  -2.7671208    2.7533006
#> 
#> $ARG$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.9492958 0.5563726   0.2317585    1.4407674
#> lag1_var2 -1.2103293 1.1241427  -2.7549337   -0.5159578
#> lag1_var3 -1.3819845 1.9594410  -4.0710445   -0.0380299
#> lag1_var4  1.6415253 1.7089282   0.4963936    3.9692373
#> const     16.9149294 8.3319695   6.2103089   25.0658492
#> exo1      -0.1715380 2.0004170  -1.8513031    2.3725052
#> exo2      -4.8993517 1.5671420  -7.0356984   -3.7545166
#> exo3       3.2499223 1.8600810   1.2607404    5.1682979
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.4030142  0.3928210 -0.77039815   0.07332784
#> lag1_var2  0.4188824  0.5066368  0.09517053   1.11062525
#> lag1_var3  0.6467020  0.7710332  0.17110568   1.70694423
#> lag1_var4  0.2473688  0.7270890 -0.74553585   0.66151489
#> const     17.0609331 10.3857476  5.27462620  27.72420981
#> exo1      -0.9890096  0.6430699 -1.61199326  -0.28882203
#> exo2      -3.9780927  0.7606118 -4.99909793  -3.37615664
#> exo3       3.5450257  0.7129741  2.86106945   4.36123719
#> 
#> 
#>