Provides posterior estimation summary for Bayesian Hierarchical Panel Vector Autoregressions

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARPANEL'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARPANEL obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                  mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00538212 0.001838525 0.003295371  0.007567622
#> 
#> $AFG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01426148 0.005411084 -0.01984475 -0.007603145
#> Sigma[2,2]  0.22377942 0.087983977  0.14250827  0.339834033
#> 
#> $AFG$Sigma$equation3
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.01899983 0.004689511  0.01270035   0.02291192
#> Sigma[3,2] -0.18882817 0.071204838 -0.26045447  -0.10071820
#> Sigma[3,3]  0.35263357 0.023373250  0.32419665   0.37499314
#> 
#> $AFG$Sigma$equation4
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[4,1]  0.01423706 0.003662737  0.01076175   0.01824057
#> Sigma[4,2] -0.10222294 0.044495380 -0.14517284  -0.04622410
#> Sigma[4,3]  0.30632567 0.016600176  0.28823618   0.32653767
#> Sigma[4,4]  0.29896889 0.041918879  0.26314777   0.35510772
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.98812113 0.03067250   0.9564601  1.026697565
#> lag1_var2 -0.17152936 0.07654786  -0.2309439 -0.068226724
#> lag1_var3 -0.36549236 0.16741575  -0.5123019 -0.143048675
#> lag1_var4  0.32014708 0.15032534   0.1237735  0.456585277
#> const      2.44909023 0.96885410   1.3572432  3.512040881
#> exo1      -0.03191247 0.07025678  -0.1049016  0.053657547
#> exo2      -0.07318480 0.06968252  -0.1538884  0.003675175
#> exo3      -0.19831993 0.04181721  -0.2429030 -0.155788142
#> 
#> $AFG$A$equation2
#>                 mean        sd  5% quantile 95% quantile
#> lag1_var1  0.3993933 0.2199639   0.19590698   0.68956018
#> lag1_var2  0.4684715 0.3592515   0.22503943   0.95792832
#> lag1_var3 -0.9787622 0.6718249  -1.51391149  -0.07827064
#> lag1_var4  0.8165615 0.5766668   0.07762192   1.34519424
#> const     -0.7416077 8.6446255 -10.40542987   8.35074536
#> exo1      -0.3477887 0.6574861  -0.90657062   0.51953339
#> exo2      -0.1724078 0.3855705  -0.59123744   0.24744390
#> exo3      -0.1631474 0.7494275  -1.08178819   0.58514408
#> 
#> $AFG$A$equation3
#>                   mean        sd 5% quantile 95% quantile
#> lag1_var1  -0.03292239 0.3060737  -0.4255912   0.24739090
#> lag1_var2   1.07893470 0.1835277   0.8808241   1.25639445
#> lag1_var3   3.98053414 0.6478462   3.1132583   4.47513146
#> lag1_var4  -2.69165315 0.7072577  -3.2063064  -1.73481189
#> const     -10.35700977 8.4204509 -16.7092345   0.93121768
#> exo1       -0.09945566 0.6466365  -0.9100504   0.53752545
#> exo2       -1.02259267 1.0361364  -2.3159725   0.07794533
#> exo3       -0.10508462 1.0412038  -1.1150469   1.24603318
#> 
#> $AFG$A$equation4
#>                   mean        sd 5% quantile 95% quantile
#> lag1_var1   0.17876029 0.2629818  -0.1109237   0.45863284
#> lag1_var2   1.52132397 0.3156853   1.2299681   1.92679877
#> lag1_var3   4.09957856 0.8145149   3.1496658   5.04378787
#> lag1_var4  -2.68182479 0.7960554  -3.5424665  -1.70118854
#> const     -20.48251071 8.7720833 -30.5850461 -10.46983403
#> exo1       -0.27278054 0.4492964  -0.7693532   0.19580671
#> exo2       -1.13446177 1.1341016  -2.6197810  -0.05150896
#> exo3       -0.05086959 0.7977797  -0.7751393   1.00977524
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005605952 0.000237931 0.005291779  0.005821405
#> 
#> $AGO$Sigma$equation2
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.009890036 0.003026826 -0.01337609 -0.006468747
#> Sigma[2,2]  0.127911064 0.028347239  0.09953957  0.160096515
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.007686159 0.002048215  0.005389534  0.009754031
#> Sigma[3,2] -0.050803389 0.005576821 -0.057653825 -0.044899689
#> Sigma[3,3]  0.035832145 0.005612826  0.030122303  0.042181395
#> 
#> $AGO$Sigma$equation4
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[4,1] -0.0002871491 0.001344351 -0.001943338  0.001011847
#> Sigma[4,2]  0.0566632231 0.019332201  0.037400716  0.079066963
#> Sigma[4,3] -0.0061329656 0.006000655 -0.011348768  0.001760185
#> Sigma[4,4]  0.0451576338 0.020038507  0.027072994  0.066703843
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  1.00291715 0.02297576  0.975298470   1.02791286
#> lag1_var2  0.11191712 0.11481276 -0.046781410   0.17600512
#> lag1_var3  0.18223675 0.15435915 -0.026537027   0.25907087
#> lag1_var4 -0.12066781 0.12371319 -0.187533034   0.04840713
#> const     -4.30369125 2.35665610 -6.513371809  -1.20426684
#> exo1       0.05577567 0.06897959  0.006398912   0.14937231
#> exo2      -0.05356074 0.12283403 -0.202839787   0.06986425
#> exo3       0.06698658 0.06324455 -0.001229830   0.14199062
#> 
#> $AGO$A$equation2
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1  0.08894063 0.1884975  -0.10226807    0.3244779
#> lag1_var2  0.89177936 0.2987833   0.66891412    1.2969257
#> lag1_var3 -0.18360022 0.2893423  -0.37269990    0.2121602
#> lag1_var4  0.22205057 0.2367243  -0.08911837    0.4236868
#> const     -5.75650478 8.9779832 -17.05152188    2.7203085
#> exo1       0.11932643 0.5482506  -0.43488246    0.8219311
#> exo2       0.25684885 0.5303087  -0.41280202    0.7892439
#> exo3      -0.46234024 0.2539269  -0.74323076   -0.1549220
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.17904546 0.10035048  -0.2810009  -0.05593599
#> lag1_var2 -0.83062072 0.12669557  -0.9956609  -0.72103487
#> lag1_var3 -0.06268066 0.10146557  -0.1893978   0.03361768
#> lag1_var4  0.89936917 0.05536555   0.8423919   0.95998406
#> const     17.11866987 6.98886340  10.7909398  26.38203011
#> exo1       0.06950743 0.28743723  -0.2590953   0.40711523
#> exo2      -1.29083878 0.23074648  -1.5205835  -0.99421930
#> exo3       3.42390503 0.26767449   3.0875209   3.68535438
#> 
#> $AGO$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.1892694 0.1011811 -0.29126861  -0.05940743
#> lag1_var2 -0.2115413 0.1528445 -0.38227924  -0.03147284
#> lag1_var3 -0.2507287 0.1474554 -0.33917374  -0.04955727
#> lag1_var4  1.2536535 0.1429956  1.06236776   1.37109188
#> const      4.6076422 3.8197768  1.28932852   9.72773010
#> exo1       0.2236478 0.1914157  0.03173666   0.46876953
#> exo2      -1.3738662 0.2176747 -1.65592347  -1.16257351
#> exo3       3.5897475 0.3600345  3.14205865   3.94838188
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001585368 0.0003087522 0.001316117  0.001995184
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.04994736 0.01324046 -0.06707947  -0.03743916
#> Sigma[2,2]  9.28497573 1.68007388  7.68073766  11.45022210
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.02429794 0.01231542  0.01321636   0.04042671
#> Sigma[3,2] -5.99532670 1.13693371 -7.25851819  -4.83184419
#> Sigma[3,3]  5.47811740 1.18681349  4.56102016   7.08971455
#> 
#> $ALB$Sigma$equation4
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[4,1] -0.01155988 0.006926453 -0.01762138 -0.003274149
#> Sigma[4,2]  0.28880008 1.079393501 -1.09335319  1.361650974
#> Sigma[4,3]  1.73782606 0.822762727  0.77496320  2.693878689
#> Sigma[4,4]  2.33432055 0.233652282  2.02425027  2.544246540
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.02544234 0.03334748  0.99550210   1.06769496
#> lag1_var2  0.02981135 0.01410663  0.01865013   0.04889461
#> lag1_var3  0.03684128 0.01877294  0.01953931   0.06158991
#> lag1_var4 -0.02845170 0.01561987 -0.04816320  -0.01268741
#> const     -1.16693509 1.00409664 -2.50680184  -0.36804451
#> exo1       0.01312363 0.06170342 -0.06259707   0.07647227
#> exo2      -0.05676370 0.03746606 -0.10810304  -0.02992031
#> exo3       0.10053009 0.05121724  0.04642028   0.16153964
#> 
#> $ALB$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  -6.0683124  2.647179  -8.9356096   -3.3783207
#> lag1_var2   3.4348243  1.510650   1.6152512    4.8885492
#> lag1_var3   4.9104494  2.202884   2.4721548    7.4293404
#> lag1_var4  -4.2023210  1.817778  -6.3214942   -2.3210167
#> const     104.5465157 75.405440  39.5585377  203.6479895
#> exo1       -2.0595749  2.581354  -5.3676196    0.5532024
#> exo2        3.0352667  2.867393  -0.1061907    6.2013335
#> exo3       -0.3309862  3.203489  -4.4237545    2.8837612
#> 
#> $ALB$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1   4.2723174  1.409215    2.731697     5.962382
#> lag1_var2  -3.1310295  1.722396   -5.453214    -1.827345
#> lag1_var3  -4.5212150  2.759476   -8.266971    -2.336795
#> lag1_var4   4.5876761  2.359619    2.674945     7.775515
#> const     -44.6802135 22.752998  -75.027701   -24.417943
#> exo1       -0.7852701  1.839461   -2.575772     1.187707
#> exo2       -4.5142115  2.479087   -6.682863    -1.333576
#> exo3       -0.5026423  2.148999   -2.266217     2.325990
#> 
#> $ALB$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.2466607  0.8730224  -0.9004764   1.00991660
#> lag1_var2 -1.4257620  1.3378376  -3.1678689  -0.17053605
#> lag1_var3 -2.0912786  2.0329353  -4.7341059  -0.21066646
#> lag1_var4  2.6449276  1.7032946   1.0643061   4.87725635
#> const     23.8753382 36.6440022  -4.1301942  70.19973216
#> exo1      -2.1553309  2.0397349  -3.5259733   0.62826565
#> exo2      -2.8576920  2.3059123  -6.0442282  -1.36187634
#> exo3      -0.8176951  0.7198421  -1.6397255  -0.04356012
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001320191 0.0002278342 0.001017306  0.001510259
#> 
#> $ARE$Sigma$equation2
#>                     mean         sd  5% quantile 95% quantile
#> Sigma[2,1] -0.0008093696 0.00158431 -0.002472382  0.001060648
#> Sigma[2,2]  0.0554275388 0.01490161  0.040409544  0.074363303
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1] -0.000556799 0.003797098 -0.005604589  0.002581801
#> Sigma[3,2] -0.070003409 0.038738977 -0.115209784 -0.030815609
#> Sigma[3,3]  0.253476030 0.083858715  0.143414401  0.320451887
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1] -0.001202782 0.003087863 -0.005397997 0.0007578099
#> Sigma[4,2] -0.026250426 0.027905060 -0.056680955 0.0017114146
#> Sigma[4,3]  0.204449072 0.068300565  0.116059204 0.2695792571
#> Sigma[4,4]  0.191374644 0.063250073  0.112278038 0.2581981814
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.936485231 0.02207435  0.90969863   0.95731614
#> lag1_var2 -0.007358181 0.04499033 -0.06159245   0.04016371
#> lag1_var3 -0.026737600 0.06000619 -0.09850074   0.04086572
#> lag1_var4  0.035098998 0.05058100 -0.02237900   0.09487581
#> const      1.039830114 0.41467277  0.62196771   1.52480262
#> exo1      -0.019329849 0.03760428 -0.05973000   0.02678755
#> exo2      -0.124928110 0.03031681 -0.15166958  -0.08654181
#> exo3      -0.025969352 0.03529510 -0.05722461   0.02030397
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.09565612 0.2807204  -0.4202777    0.2294198
#> lag1_var2 -0.02928427 0.2188261  -0.3004819    0.1479016
#> lag1_var3 -1.07386489 0.2987946  -1.4073829   -0.7422573
#> lag1_var4  1.06197319 0.2941937   0.7071881    1.3729241
#> const      3.92427211 3.0756414   0.3095745    7.0795243
#> exo1      -0.04090656 0.2075827  -0.2837451    0.1868541
#> exo2       1.96628280 0.2941270   1.6864947    2.3502932
#> exo3      -1.01643590 0.4982949  -1.6838040   -0.6156963
#> 
#> $ARE$A$equation3
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1   0.7974904 0.6805811  -0.08566219    1.4134047
#> lag1_var2   3.3522050 0.9199437   2.69011008    4.5993937
#> lag1_var3   5.0103412 1.0789980   4.05483409    6.4493198
#> lag1_var4  -4.1455757 0.9688413  -5.43892886   -3.2812010
#> const     -11.0796123 7.4225965 -18.51136365   -2.2169886
#> exo1       -0.7544955 0.4512710  -1.35147517   -0.3783645
#> exo2       -1.6781736 0.4670177  -2.31209594   -1.2980007
#> exo3       -0.9277800 1.0121347  -1.73709220    0.4291324
#> 
#> $ARE$A$equation4
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1   0.6861934 0.6356588  -0.02517403    1.2863821
#> lag1_var2   3.4944106 0.8376781   2.75249320    4.6039262
#> lag1_var3   4.3995085 1.0135099   3.50056479    5.7302625
#> lag1_var4  -3.5048993 0.9251772  -4.73607601   -2.7411770
#> const     -10.0510054 5.0522430 -15.35974824   -3.9744359
#> exo1       -0.7896116 0.3833184  -1.25811595   -0.3740271
#> exo2       -0.1233370 0.5706761  -0.82318274    0.4703721
#> exo3       -1.7255940 0.7157769  -2.32749690   -0.7618983
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002709717 0.0006469656 0.001935122  0.003414649
#> 
#> $ARG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05729192 0.008812552 -0.06703638   -0.0467465
#> Sigma[2,2]  2.88286064 0.624121136  2.16317613    3.5685463
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04353204 0.01108308  0.02934977   0.05477773
#> Sigma[3,2] -2.07530676 0.53845056 -2.57139424  -1.40550059
#> Sigma[3,3]  1.69239359 0.48390816  1.08680200   2.17678039
#> 
#> $ARG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.009656682 0.006444902  0.001233467   0.01582049
#> Sigma[4,2] -0.354288943 0.205684153 -0.588239538  -0.11099202
#> Sigma[4,3]  0.474911111 0.182078806  0.267833313   0.68886045
#> Sigma[4,4]  0.290035865 0.060664300  0.228809462   0.36684340
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.94993173 0.06856781  0.897246285   1.03632748
#> lag1_var2  0.06564405 0.13256324 -0.037049217   0.23245621
#> lag1_var3  0.10687554 0.22053045 -0.065482905   0.38401807
#> lag1_var4 -0.10491138 0.20081443 -0.358649355   0.05895907
#> const      1.24315922 1.40900397 -0.476247871   2.72883945
#> exo1       0.04676251 0.05105439  0.002300557   0.11129889
#> exo2      -0.13957560 0.04803974 -0.199789386  -0.09300866
#> exo3       0.03513345 0.06006080 -0.032534722   0.10686599
#> 
#> $ARG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -2.4706339  0.8305971  -3.1133268   -1.3377640
#> lag1_var2 -0.1666014  1.1053278  -1.6888743    0.5995874
#> lag1_var3 -1.4215058  1.7865209  -3.8861067   -0.1932127
#> lag1_var4  1.6017536  1.6893664   0.4140728    3.9017780
#> const     58.5166940 18.2572388  35.8644836   78.1819097
#> exo1      -1.1764850  2.2535923  -3.8849722    1.3134200
#> exo2       2.1096262  0.5493400   1.5521851    2.7893047
#> exo3      -0.1209508  0.9364644  -0.8577144    1.1218766
#> 
#> $ARG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1   1.2316810  0.9181408   0.1183394    2.1764790
#> lag1_var2   0.4931085  1.7895075  -0.7906682    2.8719550
#> lag1_var3   1.4896559  2.9731211  -0.6516462    5.4390920
#> lag1_var4  -0.8166896  2.6594550  -4.3937166    1.0712100
#> const     -15.1377715 15.1266499 -33.6914169   -0.4986045
#> exo1        0.1734487  1.1629658  -1.2151042    1.4554640
#> exo2       -5.1146411  0.5952948  -5.7754582   -4.5213713
#> exo3        2.6883372  0.8430978   1.5277118    3.2709977
#> 
#> $ARG$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.3459351 0.5310122  -0.8298054    0.3276510
#> lag1_var2  0.3578882 1.1746240  -0.5043873    1.8575696
#> lag1_var3  0.5683776 1.9947331  -0.9135614    3.1061874
#> lag1_var4  0.2920468 1.7470422  -1.9618028    1.6097298
#> const     17.7515824 9.1772021  11.0434897   30.2722896
#> exo1      -0.6296955 0.3095211  -1.0494515   -0.4606588
#> exo2      -4.3753822 0.3227945  -4.6962035   -3.9730485
#> exo3       2.9565223 0.5261452   2.4240538    3.6128990
#> 
#> 
#>