Provides posterior estimation summary for Bayesian Vector Autoregressions for dynamic panel data

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARs'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARs obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003738128 0.001247455 0.002663383  0.005402995
#> 
#> $AFG$Sigma$equation2
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.006687796 0.00422722 -0.01239331 -0.003809474
#> Sigma[2,2]  0.182400031 0.04326852  0.14847982  0.241045348
#> 
#> $AFG$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.004033088 0.008603749 -0.002686108   0.01578263
#> Sigma[3,2] -0.178053376 0.037468276 -0.227177178  -0.14233036
#> Sigma[3,3]  0.318082588 0.047785997  0.261233383   0.36854665
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.001255126 0.007469482 -0.004649646   0.01144009
#> Sigma[4,2] -0.111311938 0.022348570 -0.136779128  -0.08579348
#> Sigma[4,3]  0.272221095 0.044805695  0.220331498   0.31945917
#> Sigma[4,4]  0.254099077 0.047454831  0.203255482   0.30640240
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.97203919 0.01096784   0.9574476   0.98120469
#> lag1_var2  0.15192815 0.02090091   0.1353107   0.17978050
#> lag1_var3  0.31739202 0.04097798   0.2869174   0.37270970
#> lag1_var4 -0.27046374 0.03662657  -0.3202468  -0.24426529
#> const     -1.51973550 0.52946105  -2.1238202  -0.92411510
#> exo1      -0.04971404 0.08062822  -0.1368422   0.04907238
#> exo2      -0.04264434 0.04401971  -0.1028631  -0.01916475
#> exo3      -0.08782227 0.02282314  -0.1192911  -0.07250525
#> 
#> $AFG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.3751238 0.03791567   0.3417415    0.4244758
#> lag1_var2  0.6395589 0.22579902   0.3462036    0.8541102
#> lag1_var3 -0.8474482 0.19132844  -1.0846435   -0.6419528
#> lag1_var4  0.7483834 0.13602475   0.6144457    0.9058135
#> const     -4.1179105 5.09767539  -9.4614585    1.2710437
#> exo1       0.1455611 0.36408867  -0.2818893    0.5274434
#> exo2       0.2319418 0.36638156  -0.1095082    0.6717371
#> exo3      -0.3733563 0.53094136  -1.0325989    0.1906469
#> 
#> $AFG$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.23349016 0.1567667  -0.4449224  -0.11125444
#> lag1_var2 -0.01293163 0.2892300  -0.3036259   0.35452204
#> lag1_var3  1.69297811 0.3119526   1.4571435   2.11986880
#> lag1_var4 -0.72096021 0.2599893  -1.0574319  -0.47864325
#> const      9.24915069 4.2400933   5.4721340  14.21318968
#> exo1      -0.41943268 0.3729701  -0.6692758   0.09506649
#> exo2      -1.40777411 0.5926813  -2.1314238  -0.75608431
#> exo3       0.33173319 0.5199762  -0.1601631   0.98940403
#> 
#> $AFG$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.08113134 0.1543645 -0.28222042   0.04623792
#> lag1_var2  0.35130270 0.2165651  0.12578089   0.62059966
#> lag1_var3  1.50004106 0.2736394  1.30766931   1.87658872
#> lag1_var4 -0.41425408 0.2424639 -0.72713040  -0.17493672
#> const      0.58930360 2.8130922 -2.26280979   3.85676934
#> exo1      -0.39277149 0.4289768 -0.88225934   0.11270519
#> exo2      -1.43059773 0.6368273 -2.19841327  -0.73776693
#> exo3       0.31749620 0.3493441 -0.08503546   0.70175659
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.007675568 0.005631262 0.004217132   0.01547028
#> 
#> $AGO$Sigma$equation2
#>                  mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0106686 0.008157452 -0.02174267 -0.004434495
#> Sigma[2,2]  0.1514141 0.045030970  0.11691574  0.212403080
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.004685366 0.004862771 -0.001422081  0.009696165
#> Sigma[3,2] -0.061911992 0.017307237 -0.084198519 -0.044638593
#> Sigma[3,3]  0.041097986 0.010289015  0.029668954  0.053465803
#> 
#> $AGO$Sigma$equation4
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[4,1] -0.00426923 0.006758139 -0.01257247 0.0008070044
#> Sigma[4,2]  0.06145360 0.020239175  0.05081119 0.0890082732
#> Sigma[4,3] -0.00671671 0.012561863 -0.02309451 0.0050159758
#> Sigma[4,4]  0.04678577 0.007558807  0.03918451 0.0546327347
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  1.01864167 0.06821936  0.961436406   1.11045443
#> lag1_var2 -0.10798274 0.11798271 -0.268410289  -0.02548965
#> lag1_var3 -0.16132982 0.19487779 -0.425613323  -0.04659986
#> lag1_var4  0.16662447 0.19338018  0.047731096   0.42987384
#> const     -1.09991559 2.68107734 -4.120494740   2.03639430
#> exo1       0.02305545 0.04265528 -0.019633022   0.07726229
#> exo2      -0.06206131 0.07207094 -0.157550493  -0.00425802
#> exo3       0.10728563 0.14378364 -0.002640015   0.29543554
#> 
#> $AGO$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.01053386 0.09577632 -0.06009927   0.14174331
#> lag1_var2  0.45555329 0.14307696  0.27477205   0.56911075
#> lag1_var3 -0.59036465 0.25887460 -0.87704080  -0.32178255
#> lag1_var4  0.55603047 0.17582646  0.34667257   0.73134782
#> const      3.79751794 9.05192578 -5.81795979  14.75364583
#> exo1      -0.00628438 0.34261003 -0.47841958   0.22818137
#> exo2       0.54933610 0.19388737  0.34398762   0.77677169
#> exo3      -0.72907742 0.57896893 -1.30122426  -0.08986799
#> 
#> $AGO$A$equation3
#>                  mean         sd   5% quantile 95% quantile
#> lag1_var1 -0.18880111 0.03604543 -2.182159e-01  -0.14062011
#> lag1_var2 -0.30621744 0.30567010 -6.768109e-01  -0.03251725
#> lag1_var3  0.54515330 0.44694833  2.617245e-02   0.95730642
#> lag1_var4  0.36656484 0.37752290  4.440375e-05   0.80074271
#> const     10.65650459 5.00049164  5.366734e+00  16.55430401
#> exo1       0.08459948 0.14663457 -1.085062e-01   0.22068680
#> exo2      -1.44100581 0.07061577 -1.514352e+00  -1.35220845
#> exo3       3.42030413 0.30714436  3.158099e+00   3.82213508
#> 
#> $AGO$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.23681953 0.07991765  -0.3182859   -0.1382952
#> lag1_var2 -0.16943652 0.25741065  -0.4307072    0.1457261
#> lag1_var3 -0.16391101 0.29843185  -0.4540200    0.2074929
#> lag1_var4  1.16556586 0.25751711   0.8585933    1.4518332
#> const      6.27907974 4.29715254   2.8242545   11.2951062
#> exo1       0.09512979 0.22170709  -0.1844256    0.3001729
#> exo2      -1.24653127 0.15254280  -1.4157435   -1.0716274
#> exo3       3.39587170 0.24223728   3.1412616    3.6739097
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002622434 0.0009343928 0.001585607  0.003646814
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05911751 0.04472053  -0.1114934  -0.01912898
#> Sigma[2,2]  7.75873412 1.09088914   6.3309327   8.79891528
#> 
#> $ALB$Sigma$equation3
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.006774859 0.02250858 -0.02005866   0.02743442
#> Sigma[3,2] -5.422181385 0.82979784 -6.55449070  -4.78740109
#> Sigma[3,3]  6.006338372 1.91232388  4.70426052   8.60514406
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.04067298 0.01864251 -0.06096857  -0.01951125
#> Sigma[4,2] -0.20062361 1.33914427 -1.83098757   1.25314060
#> Sigma[4,3]  2.81141429 1.69722841  1.56757009   5.13037455
#> Sigma[4,4]  3.24163196 1.14953237  2.40146953   4.80067741
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1  1.04401667 0.008545111  1.03572433   1.05514151
#> lag1_var2 -0.01977577 0.058822155 -0.07952347   0.05548233
#> lag1_var3 -0.04269724 0.087940379 -0.12947569   0.07120996
#> lag1_var4  0.03962553 0.075968010 -0.05840069   0.11529265
#> const     -0.87172895 0.718650964 -1.76341851  -0.09994705
#> exo1       0.05944657 0.026031952  0.02441983   0.08074582
#> exo2      -0.07555567 0.037771548 -0.10994785  -0.02749658
#> exo3       0.04198375 0.053973337 -0.02503762   0.09740247
#> 
#> $ALB$A$equation2
#>                mean        sd 5% quantile 95% quantile
#> lag1_var1 -4.405844  3.017146   -8.010047  -1.05235407
#> lag1_var2  3.498091  1.655374    1.794825   5.40184201
#> lag1_var3  5.017325  2.585590    2.417744   8.04569592
#> lag1_var4 -4.163202  2.300258   -6.838749  -1.82771838
#> const     58.000573 82.559588  -39.577058 150.20647005
#> exo1      -3.124512  2.488542   -6.140899  -0.35907530
#> exo2      -2.110927  2.144170   -4.574709  -0.01144209
#> exo3      -1.011208  3.498187   -3.768205   2.96592634
#> 
#> $ALB$A$equation3
#>                   mean         sd   5% quantile 95% quantile
#> lag1_var1   2.38221723  2.5352625   -0.17747798     5.620234
#> lag1_var2  -0.86033256  2.3520870   -3.85282179     1.287789
#> lag1_var3  -0.92736072  3.5800467   -5.30056527     2.565314
#> lag1_var4   1.42370025  3.0264455   -1.58949650     5.064052
#> const     -29.06040548 81.0923219 -116.05521298    70.008241
#> exo1       -0.09406239  3.5789427   -4.57629174     3.265554
#> exo2        0.38726522  0.9388751   -0.88788146     1.119982
#> exo3        1.55537726  1.9779292    0.09742678     4.259994
#> 
#> $ALB$A$equation4
#>                mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.623131  0.8971965   -1.737570    0.3013309
#> lag1_var2  1.537390  1.9611944   -1.159810    2.7918218
#> lag1_var3  2.585569  3.0063692   -1.536629    4.7171122
#> lag1_var4 -1.357596  2.5780575   -3.157345    2.1774949
#> const      1.012487 31.4789964  -29.155515   41.8060710
#> exo1      -2.182490  2.9222359   -5.551058    0.6632940
#> exo2      -1.101423  1.9145046   -2.916654    1.2517675
#> exo3       1.252590  2.4453751   -1.547520    3.4473740
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001393402 0.0002606006 0.001084281  0.001687384
#> 
#> $ARE$Sigma$equation2
#>                    mean           sd  5% quantile 95% quantile
#> Sigma[2,1] -0.001802444 0.0005362449 -0.002431218 -0.001202438
#> Sigma[2,2]  0.054635019 0.0094063230  0.044213458  0.064348122
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.004766746 0.002615512  0.001541729   0.00746376
#> Sigma[3,2] -0.051160307 0.018028444 -0.072378253  -0.03372943
#> Sigma[3,3]  0.181792525 0.036064064  0.138706044   0.21654121
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.003525945 0.002527632  0.0003826766 0.0059435968
#> Sigma[4,2] -0.009547977 0.013080039 -0.0269773804 0.0007850609
#> Sigma[4,3]  0.147372888 0.027156749  0.1165709274 0.1778583065
#> Sigma[4,4]  0.145223017 0.026106202  0.1210811462 0.1796453966
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1  0.94653470 0.034998329  0.90896948  0.983127111
#> lag1_var2  0.35681253 0.022663089  0.32693568  0.375832297
#> lag1_var3  0.42948252 0.028672964  0.39512644  0.461697786
#> lag1_var4 -0.41428741 0.032557690 -0.45397961 -0.378344495
#> const      0.22029449 0.730772005 -0.53323306  1.070287640
#> exo1      -0.01713509 0.009944087 -0.02664135 -0.004441907
#> exo2      -0.11362433 0.020995468 -0.14220356 -0.097933515
#> exo3      -0.03145434 0.048608001 -0.08618263  0.018288416
#> 
#> $ARE$A$equation2
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1  0.17437407 0.3576130 -0.119053316    0.6547353
#> lag1_var2  0.52197132 0.1830286  0.359010712    0.7663629
#> lag1_var3 -0.27772891 0.3292726 -0.461285112    0.1736704
#> lag1_var4  0.17797462 0.3838338 -0.343565692    0.4127935
#> const      3.72035267 4.3206293 -2.064003883    7.1928317
#> exo1       0.09477502 0.1273307  0.005602491    0.2677905
#> exo2       2.04251549 0.3326418  1.713360501    2.4566256
#> exo3      -0.55809648 0.3338183 -0.879606155   -0.1658479
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.3555155 0.4651102  -0.1024565    0.9197927
#> lag1_var2 -1.7106105 0.4864538  -2.2316301   -1.2155598
#> lag1_var3 -1.6079619 0.6314928  -2.2709486   -0.9788023
#> lag1_var4  2.5280165 0.7249960   1.7535869    3.2791508
#> const     -3.7127308 5.4507782 -10.7095731    1.5186167
#> exo1      -0.1611990 0.3299361  -0.5388422    0.2102040
#> exo2      -2.6263573 0.2584307  -2.8781615   -2.3404893
#> exo3      -1.8399638 0.1978798  -2.1051640   -1.6759699
#> 
#> $ARE$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.50494315 0.3568281   0.1670411    0.9550914
#> lag1_var2 -1.33762091 0.3948810  -1.8090228   -0.9434353
#> lag1_var3 -1.85516695 0.5655423  -2.5131056   -1.2729306
#> lag1_var4  2.71660075 0.6261950   2.0108928    3.4236927
#> const     -2.63577017 5.0500659  -9.1495512    2.2363738
#> exo1      -0.09050027 0.2926531  -0.3495422    0.2735252
#> exo2      -1.04519224 0.1639483  -1.2688450   -0.9186246
#> exo3      -2.31519113 0.2487939  -2.5137192   -1.9777222
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002990021 0.0009953566 0.001888423  0.004194827
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05738659 0.01703295 -0.07578562  -0.03705389
#> Sigma[2,2]  2.64472981 0.70917197  1.93947808   3.52280844
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04369888 0.01677378  0.02473723   0.06312224
#> Sigma[3,2] -1.99305020 0.65102486 -2.83707101  -1.37789700
#> Sigma[3,3]  1.70591385 0.57380598  1.21471027   2.46197602
#> 
#> $ARG$Sigma$equation4
#>                    mean         sd  5% quantile 95% quantile
#> Sigma[4,1]  0.009834219 0.00736601  0.002469852   0.01916014
#> Sigma[4,2] -0.419870425 0.25862105 -0.759777885  -0.21009659
#> Sigma[4,3]  0.543240672 0.20763274  0.379174309   0.80895371
#> Sigma[4,4]  0.320188237 0.07160621  0.237303250   0.38412528
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.959419056 0.03167222  0.92834611   0.99821810
#> lag1_var2 -0.080948234 0.02059312 -0.10053555  -0.05752170
#> lag1_var3 -0.130994047 0.03627039 -0.16638131  -0.09206379
#> lag1_var4  0.117493151 0.03215546  0.08057458   0.14527061
#> const      1.939848768 0.37116836  1.43416676   2.22485728
#> exo1       0.001551796 0.05476735 -0.07065048   0.04749792
#> exo2      -0.082592273 0.04144305 -0.12898753  -0.04053359
#> exo3       0.058397980 0.06860086 -0.01569612   0.14329027
#> 
#> $ARG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -2.7345521  0.9995317  -3.9390280   -1.9517950
#> lag1_var2  1.4781373  0.6406779   0.8445982    2.2957419
#> lag1_var3  1.2027841  1.0525271   0.2080865    2.5525051
#> lag1_var4 -0.5913427  0.9381929  -1.7379264    0.4060588
#> const     38.9742502 30.0107141   1.6639324   64.1370233
#> exo1      -0.4376508  1.4138143  -1.8339628    1.3762781
#> exo2       2.6313654  0.7828834   1.9503025    3.6606839
#> exo3       1.5385077  1.7184440  -0.7446096    3.0819043
#> 
#> $ARG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.03821508  0.9179176   0.2199519    2.1244413
#> lag1_var2  0.03788963  0.7655419  -0.4861724    1.0909886
#> lag1_var3  0.81611829  1.2657566  -0.1480947    2.5359611
#> lag1_var4 -0.32100524  1.1284737  -1.8411914    0.6253025
#> const      1.40578366 31.1700757 -33.7127485   33.0569435
#> exo1      -0.25708784  1.1306703  -1.6583967    0.8999633
#> exo2      -5.52661531  0.6565966  -6.4334877   -5.0834428
#> exo3       1.98895492  1.2833671   0.5230643    3.5090794
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.7047267  0.3779098 -1.11521694   -0.3052658
#> lag1_var2  1.0456836  0.7443412  0.37806119    1.9392734
#> lag1_var3  1.7216362  1.2292429  0.59365399    3.2036673
#> lag1_var4 -0.7562457  1.0637425 -1.98943142    0.2039621
#> const     21.2850974 16.6330276 -0.01221764   37.8750830
#> exo1      -0.5954227  0.4975970 -1.16640320   -0.1458550
#> exo2      -4.4287681  0.4380623 -4.93604394   -3.9658440
#> exo3       3.2187528  0.4689346  2.59980208    3.6295442
#> 
#> 
#>