Provides posterior estimation summary for Bayesian Vector Autoregressions for dynamic panel data

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARs'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARs obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003548042 0.001121461 0.002528132  0.005001081
#> 
#> $AFG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01196696 0.008124426  -0.0229271 -0.005205504
#> Sigma[2,2]  0.28797245 0.073252002   0.2033936  0.368811082
#> 
#> $AFG$Sigma$equation3
#>                  mean         sd  5% quantile 95% quantile
#> Sigma[3,1]  0.0136120 0.01034505  0.004664202   0.02746512
#> Sigma[3,2] -0.2933684 0.07861228 -0.398811104  -0.23781845
#> Sigma[3,3]  0.4879639 0.14775453  0.338580597   0.66946949
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.009668727 0.008137064  0.001683358    0.0200799
#> Sigma[4,2] -0.191519449 0.074034703 -0.288611439   -0.1285218
#> Sigma[4,3]  0.414340122 0.141025978  0.268460447    0.5730097
#> Sigma[4,4]  0.382320174 0.131881428  0.243843673    0.5214448
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.99748769 0.01440290   0.9785300   1.01008841
#> lag1_var2 -0.02780279 0.07473004  -0.1241418   0.03944212
#> lag1_var3 -0.01943951 0.16455318  -0.2389067   0.11428392
#> lag1_var4  0.01311754 0.14613415  -0.1036301   0.20940817
#> const      0.57054583 0.85518104  -0.4204714   1.53094964
#> exo1      -0.03775278 0.07656182  -0.1351791   0.03932588
#> exo2      -0.05274576 0.08835071  -0.1568541   0.03185661
#> exo3      -0.17908558 0.04579087  -0.2299550  -0.12387493
#> 
#> $AFG$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.37519988 0.2630663  0.09734961    0.6675751
#> lag1_var2 -0.03483452 0.5206255 -0.53091910    0.6245517
#> lag1_var3 -2.37718102 0.9091376 -3.22814867   -1.1864182
#> lag1_var4  2.12987613 0.7285258  1.20069907    2.8490674
#> const      2.50623200 8.6259674 -7.40408825   10.1490477
#> exo1      -0.45455665 0.2565973 -0.65175186   -0.1067550
#> exo2       0.19598295 0.5456147 -0.30967734    0.8922540
#> exo3      -0.31288456 0.5768212 -0.86256629    0.4403422
#> 
#> $AFG$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.2360663 0.3814626  -0.6030640    0.2510047
#> lag1_var2  0.8383810 0.7957774   0.0218871    1.6724123
#> lag1_var3  3.6607693 1.5434602   2.0267016    5.3956639
#> lag1_var4 -2.4899312 1.3008672  -3.9844699   -1.1144561
#> const      0.5372676 8.6186738  -8.1947417   10.9658920
#> exo1       0.5467711 0.3253410   0.1834182    0.8815390
#> exo2      -1.4101238 0.8510232  -2.1874047   -0.2807258
#> exo3       0.2488466 0.4536081  -0.1319597    0.8486163
#> 
#> $AFG$A$equation4
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1 -0.08097055 0.3184725  -0.36029580    0.3402146
#> lag1_var2  0.88666112 0.6520849   0.22290358    1.6061343
#> lag1_var3  2.79614530 1.3268976   1.65072427    4.2874978
#> lag1_var4 -1.58892974 1.1525096  -2.89858331   -0.5969978
#> const     -4.53600883 5.5675000 -10.40405570    1.8971878
#> exo1       0.38420290 0.3455902  -0.04082548    0.6968043
#> exo2      -1.40605372 0.8248905  -2.07496500   -0.2980601
#> exo3       0.22312246 0.5619237  -0.27884428    0.9377740
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005877911 0.001316812 0.004657093  0.007552704
#> 
#> $AGO$Sigma$equation2
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.006718896 0.003429582 -0.01100653 -0.003164229
#> Sigma[2,2]  0.123554303 0.024888643  0.09809583  0.151486246
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.005826137 0.001487840  0.004149699   0.00760201
#> Sigma[3,2] -0.048356306 0.011764989 -0.061977502  -0.03956265
#> Sigma[3,3]  0.032707799 0.007402617  0.025397135   0.04123228
#> 
#> $AGO$Sigma$equation4
#>                    mean         sd  5% quantile 95% quantile
#> Sigma[4,1]  0.001304123 0.00232273 -0.001163405 0.0040320197
#> Sigma[4,2]  0.050625825 0.01234860  0.035811780 0.0630193008
#> Sigma[4,3] -0.004021042 0.00383798 -0.007676679 0.0004981232
#> Sigma[4,4]  0.040691433 0.01049171  0.029963642 0.0534206751
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.97385312 0.02556656   0.9541391   1.00882029
#> lag1_var2 -0.05614364 0.23412909  -0.2530156   0.25394789
#> lag1_var3 -0.05290633 0.30686872  -0.3480164   0.34044900
#> lag1_var4  0.04862806 0.27088533  -0.3046021   0.29554023
#> const      1.26527879 2.74209321  -2.3031313   3.67537038
#> exo1       0.04825525 0.09292245  -0.0636966   0.15118731
#> exo2      -0.09428621 0.10217883  -0.1930611   0.01723593
#> exo3       0.03233481 0.16044510  -0.1598145   0.19088864
#> 
#> $AGO$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.01902337 0.1765471 -0.13433825    0.2465524
#> lag1_var2  0.61653620 0.6118114  0.04206939    1.4052358
#> lag1_var3 -0.44934965 0.7902482 -1.23471199    0.5498083
#> lag1_var4  0.46471034 0.6917486 -0.40066255    1.1809539
#> const     -1.05871189 8.6198948 -9.94248380    9.7248199
#> exo1      -0.04412878 0.3380232 -0.46143118    0.3093319
#> exo2       0.38940978 0.4598554 -0.06158863    0.9842668
#> exo3      -0.68038500 0.4100056 -1.19421119   -0.2464510
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.21409787 0.03894009 -0.26236714  -0.17335002
#> lag1_var2  0.05867972 0.22878565 -0.12525215   0.36803363
#> lag1_var3  1.08203407 0.26256538  0.92153034   1.44124658
#> lag1_var4 -0.08681081 0.24435837 -0.41821648   0.06969543
#> const      5.65102544 2.94728600  2.48774620   9.11457841
#> exo1       0.09478980 0.22766409 -0.09700157   0.39941528
#> exo2      -1.39980550 0.22740622 -1.63993708  -1.14646372
#> exo3       3.33121151 0.21872491  3.08057578   3.59029172
#> 
#> $AGO$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.2982430 0.1296077  -0.4011940   -0.1477920
#> lag1_var2  0.4685621 0.5271150  -0.1104597    1.0012722
#> lag1_var3  0.6638994 0.6878577  -0.1496106    1.3661989
#> lag1_var4  0.4427701 0.6126309  -0.1700841    1.1686259
#> const     -0.2707697 6.1384786  -6.0808602    7.6342694
#> exo1       0.1021706 0.1868596  -0.1049513    0.3281034
#> exo2      -1.4389982 0.1991085  -1.5826737   -1.1665297
#> exo3       3.1532322 0.2842100   2.9417667    3.5348104
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002917973 0.0007463172 0.002096504  0.003818981
#> 
#> $ALB$Sigma$equation2
#>                  mean       sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0769348 0.016324 -0.09434116  -0.05991929
#> Sigma[2,2]  7.6804666 1.688381  5.49300996   8.99675264
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.06382846 0.03005524  0.02513909   0.09088746
#> Sigma[3,2] -5.16438140 0.99904855 -6.43505934  -4.16999390
#> Sigma[3,3]  5.46077465 0.95563109  4.46718192   6.54325798
#> 
#> $ALB$Sigma$equation4
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[4,1]  0.014437766 0.03239922 -0.02722805   0.04285403
#> Sigma[4,2] -0.004046375 0.94881604 -0.75984323   1.24667632
#> Sigma[4,3]  2.416406673 1.11432163  1.35729490   3.80863456
#> Sigma[4,4]  2.945143608 0.98782704  2.03462468   4.19226550
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.944458615 0.03680939  0.90927942   0.98707857
#> lag1_var2  0.047945671 0.07013572 -0.03186753   0.11223259
#> lag1_var3  0.071352311 0.11338904 -0.05707426   0.17380754
#> lag1_var4 -0.062464617 0.09681455 -0.15098364   0.04759029
#> const      0.699594308 0.29319178  0.31305692   0.97394375
#> exo1       0.003046244 0.08187056 -0.09252692   0.07338782
#> exo2      -0.022983634 0.04778379 -0.08697679   0.01836161
#> exo3       0.092389533 0.07203791  0.02693567   0.18681415
#> 
#> $ALB$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  -1.2271180 0.4428252   -1.672668   -0.7303572
#> lag1_var2   3.8210669 1.3924374    2.463203    5.5057252
#> lag1_var3   5.0897004 2.0839260    3.036252    7.5632365
#> lag1_var4  -4.1292266 1.8253411   -6.315201   -2.3878108
#> const     -25.5803270 8.3914121  -35.691922  -16.2974973
#> exo1       -2.0675364 5.4498689   -8.461944    4.1865150
#> exo2       -2.1199555 1.4497151   -4.067569   -1.0063712
#> exo3        0.4834923 2.9000658   -3.069768    3.5539793
#> 
#> $ALB$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.6143707 0.8488369  -0.2964888    1.5182283
#> lag1_var2 -1.4697763 1.6400225  -3.5594422   -0.1113957
#> lag1_var3 -1.6049344 2.6497003  -4.9571954    0.6103868
#> lag1_var4  1.9879617 2.2827295   0.1103180    4.8866296
#> const     21.1866977 7.7643598  15.1503703   31.7894290
#> exo1      -1.6146908 3.4834800  -6.2685883    1.3530708
#> exo2       0.8473885 1.8024277  -1.1779079    3.0195899
#> exo3      -0.3781506 2.1606601  -2.1966973    2.4973135
#> 
#> $ALB$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.31748194 0.7260797  -1.0136342    0.5514472
#> lag1_var2  0.91350160 0.9029581  -0.2193034    1.7457282
#> lag1_var3  1.60694377 1.5882303  -0.4129842    2.9751369
#> lag1_var4 -0.47555254 1.3374086  -1.6340035    1.2310129
#> const      0.08109965 7.5324748  -9.1264626    7.2742126
#> exo1      -3.16852687 1.6529170  -5.2339645   -1.5822361
#> exo2      -0.61426618 2.3844835  -2.6710047    2.5851738
#> exo3      -0.02142409 1.8229410  -2.4474447    1.4994686
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                  mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00164475 0.0003095561 0.001242957  0.001938106
#> 
#> $ARE$Sigma$equation2
#>                   mean          sd 5% quantile  95% quantile
#> Sigma[2,1] -0.00171015 0.001902446 -0.00428012 -0.0001189835
#> Sigma[2,2]  0.05756212 0.009142817  0.04708503  0.0679994705
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.003699058 0.003807065 -0.001107012  0.007084999
#> Sigma[3,2] -0.068296260 0.013350278 -0.083417710 -0.052400467
#> Sigma[3,3]  0.201150923 0.053538517  0.135477057  0.253990490
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.002282515 0.003051964 -0.001545163  0.005223652
#> Sigma[4,2] -0.024515780 0.011740658 -0.032432357 -0.008307687
#> Sigma[4,3]  0.150338632 0.044058033  0.096871263  0.195808756
#> Sigma[4,4]  0.133832412 0.036266690  0.091896171  0.172704979
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.92190319 0.03621384  0.88886596  0.963961104
#> lag1_var2 -0.18267558 0.04212399 -0.22700008 -0.131770692
#> lag1_var3 -0.25379465 0.06875916 -0.33945261 -0.192079551
#> lag1_var4  0.25885015 0.06347778  0.20704628  0.337345225
#> const      1.67889245 0.31916030  1.27426584  2.008320947
#> exo1      -0.05092109 0.06594149 -0.08852490  0.040070187
#> exo2      -0.06249871 0.06722892 -0.14305438  0.004418369
#> exo3      -0.02278120 0.02683396 -0.05707742  0.001482851
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.06316739 0.3669623 -0.53709745    0.2843178
#> lag1_var2  0.45792190 0.4094204  0.03132717    0.9541024
#> lag1_var3 -0.49931475 0.4984584 -0.95027410    0.1421624
#> lag1_var4  0.48580396 0.3975336 -0.03051188    0.8224989
#> const      3.05442002 2.6790721  0.15570740    6.3422006
#> exo1       0.11345372 0.2937712 -0.19292519    0.4627895
#> exo2       2.07836515 0.2822643  1.75848297    2.3947905
#> exo3      -0.92656715 0.3332110 -1.23909268   -0.5485336
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.2948400 0.5018814  -0.2048037    0.9404000
#> lag1_var2  1.0963338 0.5449878   0.3565946    1.4759272
#> lag1_var3  2.2821465 0.6640728   1.4142745    2.9102569
#> lag1_var4 -1.3241779 0.5467487  -1.8057645   -0.5926465
#> const     -4.5954367 5.2781800 -10.7872854    0.5942648
#> exo1      -0.5486804 0.7546861  -1.2572082    0.3848072
#> exo2      -2.0861972 0.8459041  -2.7322864   -0.9325919
#> exo3      -0.9674359 0.6322591  -1.4547634   -0.2007194
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.3147314 0.2696859 -0.01671333    0.5868105
#> lag1_var2  1.4432544 0.3232960  1.01861209    1.7441670
#> lag1_var3  1.9270452 0.3068329  1.53406839    2.2265161
#> lag1_var4 -0.9634869 0.2590254 -1.18769453   -0.6203652
#> const     -5.2037119 3.3760551 -8.98849939   -2.1279700
#> exo1      -0.4385736 0.6067609 -1.11198392    0.2608340
#> exo2      -0.4807625 0.6286198 -0.87734235    0.3881953
#> exo3      -1.6542154 0.5436066 -2.23792002   -1.0435262
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002726128 0.0004051463 0.002252764  0.003182405
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.04650625 0.01507223 -0.06334924  -0.03140807
#> Sigma[2,2]  2.46255684 0.19334573  2.26301140   2.67656331
#> 
#> $ARG$Sigma$equation3
#>                 mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.042037 0.01442591  0.02845698    0.0589418
#> Sigma[3,2] -1.720333 0.29658258 -2.10164414   -1.4764253
#> Sigma[3,3]  1.500528 0.38436854  1.18630739    2.0080749
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01557947 0.006215688  0.009444074   0.02312722
#> Sigma[4,2] -0.25014656 0.245678545 -0.573353260  -0.02237847
#> Sigma[4,3]  0.50699040 0.233331209  0.304798344   0.81752026
#> Sigma[4,4]  0.39678756 0.098120177  0.319861155   0.52753921
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.95725683 0.02478913  0.923912723   0.97744765
#> lag1_var2 -0.11544120 0.01669169 -0.135003188  -0.09749855
#> lag1_var3 -0.18841940 0.03157104 -0.223371461  -0.15198812
#> lag1_var4  0.16864582 0.02347377  0.146024300   0.19901885
#> const      2.37130623 0.19130822  2.199917903   2.60938525
#> exo1       0.01911819 0.02055638 -0.004779039   0.04255448
#> exo2      -0.09823950 0.05786666 -0.160350185  -0.03103708
#> exo3       0.09934106 0.12326754 -0.023469548   0.25820724
#> 
#> $ARG$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -1.34151224  0.582872  -2.0871773   -0.7713398
#> lag1_var2  0.04715081  1.364856  -1.2994215    1.5662746
#> lag1_var3 -1.12455393  2.264709  -3.3751943    1.3947775
#> lag1_var4  1.44677298  2.139731  -0.9496482    3.4709541
#> const     19.50263069 16.410712   1.4533122   36.3452288
#> exo1      -0.65702396  1.551582  -2.4427359    1.0906121
#> exo2       1.44066896  2.560680  -1.9817482    3.6964097
#> exo3      -1.22147480  1.572185  -3.0100520    0.5979271
#> 
#> $ARG$A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.7626747  0.4890115   0.2720730    1.3259741
#> lag1_var2 -0.3635598  0.5968691  -1.0254219    0.2045620
#> lag1_var3  0.1482941  0.9558572  -0.9445258    1.0303449
#> lag1_var4  0.3929157  0.9455638  -0.5419137    1.4626535
#> const      5.6559739 11.8110729  -7.4878827   17.7179969
#> exo1      -0.2124173  1.2883436  -1.7306028    0.9799417
#> exo2      -4.4559692  1.9817423  -6.2229437   -1.8104807
#> exo3       3.6389351  1.5339784   2.3081436    5.6392783
#> 
#> $ARG$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.08605621 0.2111103  -0.2739323   0.18692205
#> lag1_var2 -0.37470688 0.2692344  -0.6603798  -0.07314603
#> lag1_var3 -0.60506627 0.4864584  -1.1455676  -0.09926184
#> lag1_var4  1.42110770 0.4281856   0.9672372   1.91266099
#> const     13.60175775 4.7779574   8.9992776  19.35434494
#> exo1      -0.66878358 0.5728524  -1.2032905   0.04728103
#> exo2      -4.07327789 0.4994995  -4.5289190  -3.44562172
#> exo3       3.19839056 0.9165253   2.1927975   4.29260839
#> 
#> 
#>