Provides posterior estimation summary for Bayesian Vector Autoregressions for dynamic panel data

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARs'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARs obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003711893 0.0007993854 0.002935571  0.004602034
#> 
#> $AFG$Sigma$equation2
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.006749303 0.00450867 -0.01279976 -0.003150896
#> Sigma[2,2]  0.214462357 0.03612935  0.16555112  0.234468433
#> 
#> $AFG$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01131914 0.008849039  0.001829265   0.02217431
#> Sigma[3,2] -0.19940659 0.028141533 -0.230251132  -0.16609056
#> Sigma[3,3]  0.44467569 0.055475338  0.374303356   0.50077324
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.009911768 0.008166868  0.0006288678   0.01949366
#> Sigma[4,2] -0.119502653 0.038993455 -0.1575510760  -0.07240724
#> Sigma[4,3]  0.411550742 0.053610466  0.3473548360   0.47068634
#> Sigma[4,4]  0.415841031 0.050244224  0.3631983103   0.47032884
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.968418955 0.03425380  0.92735048   1.00529640
#> lag1_var2  0.076105179 0.09054294 -0.01425243   0.19104605
#> lag1_var3  0.199552982 0.19097633  0.02578344   0.44782515
#> lag1_var4 -0.179010324 0.16797183 -0.39808357  -0.02930263
#> const     -0.027542119 0.57640590 -0.59360160   0.70034476
#> exo1       0.004420125 0.02985089 -0.02614269   0.03671629
#> exo2       0.009851113 0.10705396 -0.10832610   0.13870978
#> exo3      -0.162472793 0.06351166 -0.21232784  -0.07628859
#> 
#> $AFG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.3561275  0.1191223   0.2663131    0.5176257
#> lag1_var2 -0.2403337  0.8080009  -1.1622868    0.4276320
#> lag1_var3 -2.6879647  1.5446450  -4.5244014   -1.3512553
#> lag1_var4  2.3378565  1.2634973   1.2532395    3.8750411
#> const      8.2062583 13.2736421  -2.3251271   25.0122304
#> exo1      -0.1593248  0.7789021  -0.9888396    0.7778041
#> exo2      -0.3096984  0.7801503  -1.3248271    0.3924494
#> exo3      -0.3598286  0.4805368  -1.0137487    0.0216196
#> 
#> $AFG$A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.2761654  0.2454281  -0.6051110  -0.09087224
#> lag1_var2  1.3492016  1.1950292   0.4606246   2.98104617
#> lag1_var3  4.7026280  2.2541499   3.2179490   7.81927565
#> lag1_var4 -3.4040950  1.8687797  -5.9865692  -2.22852137
#> const     -4.5766805 16.1481987 -23.5213456  10.00450210
#> exo1      -0.0386739  1.1767020  -1.4296203   1.29678607
#> exo2      -1.1591612  1.1289778  -2.0066329   0.36327187
#> exo3       0.1583652  0.5811125  -0.5558967   0.70491832
#> 
#> $AFG$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.15080470  0.2223732  -0.4507468   0.01956125
#> lag1_var2  1.31714326  1.0580537   0.5613855   2.76329719
#> lag1_var3  3.71868951  2.0174913   2.4839149   6.45470550
#> lag1_var4 -2.43329682  1.6869967  -4.7308876  -1.48270099
#> const     -6.45246262 13.4694856 -23.2127600   6.86441780
#> exo1      -0.10170472  1.1088675  -1.4761643   1.06479077
#> exo2      -1.37610246  0.9554483  -2.1412565  -0.08614711
#> exo3       0.07805981  0.7198199  -0.7291618   0.88846711
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.006027134 0.001176751 0.005072999  0.007611534
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.00771316 0.004376475 -0.01303493 -0.002944356
#> Sigma[2,2]  0.15647350 0.031495092  0.11649749  0.188206570
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.008185314 0.002279049  0.006135175   0.01102228
#> Sigma[3,2] -0.068260088 0.017693831 -0.087435395  -0.04645441
#> Sigma[3,3]  0.046316044 0.010084829  0.035215190   0.05839161
#> 
#> $AGO$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.001794097 0.002513298 -0.0009674531 0.0047889254
#> Sigma[4,2]  0.053063271 0.018303486  0.0408721891 0.0783399761
#> Sigma[4,3] -0.003907363 0.004274641 -0.0089122628 0.0009476389
#> Sigma[4,4]  0.042885003 0.013001388  0.0318619871 0.0603313011
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.01228326 0.03625267  0.98034660   1.05880814
#> lag1_var2 -0.63093214 0.04781988 -0.69253403  -0.58767989
#> lag1_var3 -0.74747937 0.08087043 -0.84837820  -0.66603823
#> lag1_var4  0.71278960 0.08791982  0.61807835   0.81328491
#> const      3.31838331 1.55162884  1.87261717   5.28924217
#> exo1       0.01169590 0.06774015 -0.05447237   0.08462955
#> exo2       0.09372575 0.10261939 -0.01315537   0.21444722
#> exo3       0.33302261 0.14196564  0.21456855   0.52428934
#> 
#> $AGO$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.2458634  0.0665084  0.17457098   0.31553611
#> lag1_var2 -0.9834561  0.8542016 -2.06672936  -0.18612209
#> lag1_var3 -2.6065946  1.2783530 -4.31072400  -1.53410453
#> lag1_var4  2.4084206  1.0375133  1.56759457   3.78854850
#> const      8.8034607 18.0882968 -8.35422779  29.50542522
#> exo1       0.0920273  0.1223404 -0.02160734   0.24560551
#> exo2       0.2980826  0.7811977 -0.75979792   0.93772416
#> exo3      -0.2192195  0.2103696 -0.48507165  -0.04240215
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.28826003 0.06428448 -0.36810038  -0.21998381
#> lag1_var2  0.17764133 0.27354899 -0.09909852   0.52250434
#> lag1_var3  1.29282837 0.43641385  0.87198467   1.83819970
#> lag1_var4 -0.29778630 0.34457414 -0.74230428   0.01109354
#> const      8.18605988 5.96294975  1.80626446  14.65188362
#> exo1       0.04547889 0.14636746 -0.07015035   0.24469612
#> exo2      -1.24894598 0.38331899 -1.60186872  -0.74730860
#> exo3       3.20620653 0.16113922  2.98624060   3.30261766
#> 
#> $AGO$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.1856317 0.1085565 -0.33117578   -0.1027358
#> lag1_var2 -0.6142929 0.4765471 -1.24186049   -0.2165707
#> lag1_var3 -0.7547751 0.6967933 -1.70754728   -0.2477482
#> lag1_var4  1.6997661 0.6170736  1.25808883    2.5454225
#> const      9.2372458 7.3825932  3.34330875   17.9185385
#> exo1       0.1950367 0.1558882  0.01140163    0.3728817
#> exo2      -1.3329526 0.2116636 -1.57186317   -1.0836971
#> exo3       3.4043130 0.2064139  3.17441072    3.5961579
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001916775 0.0001842595 0.001758049  0.002156205
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.07565959 0.03788556  -0.1237512  -0.03905575
#> Sigma[2,2]  9.18457611 4.25759778   5.2386239  14.81011519
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[3,1]  0.05172034 0.04668691   0.01131506    0.1117756
#> Sigma[3,2] -7.03522943 4.26054834 -12.60207475   -3.0050547
#> Sigma[3,3]  7.64997132 4.48150131   3.25777289   13.3985648
#> 
#> $ALB$Sigma$equation4
#>                     mean         sd 5% quantile 95% quantile
#> Sigma[4,1]  0.0008797351 0.02748207  -0.0218404   0.03613433
#> Sigma[4,2] -1.0591590754 1.82157755  -3.3333306   0.61478599
#> Sigma[4,3]  3.5422634165 2.01929960   1.4861966   6.06073977
#> Sigma[4,4]  3.4356927649 1.01642401   2.2878310   4.65375730
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.972928165 0.03686970  0.92790283   1.01358154
#> lag1_var2  0.000734322 0.02936770 -0.02302035   0.03979451
#> lag1_var3 -0.004382321 0.04671501 -0.04483549   0.05701678
#> lag1_var4  0.002189991 0.03869860 -0.04858546   0.03768774
#> const      0.740022517 1.02034199 -0.32004073   2.01895815
#> exo1       0.036846377 0.07355299 -0.05376880   0.11604714
#> exo2      -0.060055825 0.03026719 -0.09635827  -0.02643333
#> exo3       0.089148130 0.07376249 -0.01186755   0.13906143
#> 
#> $ALB$A$equation2
#>                mean        sd 5% quantile 95% quantile
#> lag1_var1 -3.641085  2.916008  -6.0764951   0.07309229
#> lag1_var2  2.194084  2.223878  -0.8121345   3.89923386
#> lag1_var3  2.903030  3.372310  -1.6499719   5.23133018
#> lag1_var4 -2.338176  2.850151  -4.4833008   1.53352082
#> const     58.004681 75.744441 -44.8647301 110.19268958
#> exo1      -2.944427  1.226578  -4.2071550  -1.55356772
#> exo2      -1.617161  2.950208  -5.0419419   1.24895342
#> exo3      -3.008852  4.991559  -8.7166210   2.89471347
#> 
#> $ALB$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1   2.0957152  2.623599  -1.3470901     4.378947
#> lag1_var2  -0.8868179  2.677831  -2.7976094     2.799571
#> lag1_var3  -0.8537428  4.207000  -3.7636361     4.934125
#> lag1_var4   1.3861892  3.424852  -3.3387202     3.748373
#> const     -23.4242577 87.453757 -80.1170848    95.136180
#> exo1        1.9550421  2.148519  -0.1168519     4.575234
#> exo2       -0.1069291  1.697701  -2.2723002     1.571749
#> exo3        2.0219235  2.510910  -0.6108724     4.810401
#> 
#> $ALB$A$equation4
#>                   mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.342784716  1.509459   -2.011974    1.1084791
#> lag1_var2  0.259513085  1.974200   -1.636919    2.5224201
#> lag1_var3  0.669760451  3.065841   -2.186572    4.2338844
#> lag1_var4  0.324248198  2.493877   -2.579580    2.5515505
#> const     10.453857185 53.095512  -41.565984   78.3841028
#> exo1       0.436696228  2.377924   -2.287767    3.1573952
#> exo2      -1.281101542  1.490070   -2.659982    0.4201186
#> exo3       0.003817261  1.991091   -2.008042    2.2460078
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001529251 0.0005766317 0.001048212  0.002289948
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd  5% quantile  95% quantile
#> Sigma[2,1] -0.002844147 0.001807263 -0.004829762 -0.0006457016
#> Sigma[2,2]  0.087145264 0.041281994  0.047924669  0.1399776457
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1] -0.000189594 0.002522696 -0.002760746  0.002870453
#> Sigma[3,2] -0.066124716 0.029798117 -0.097353997 -0.034411083
#> Sigma[3,3]  0.184835020 0.051004912  0.132483968  0.242303213
#> 
#> $ARE$Sigma$equation4
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[4,1] -0.0027161941 0.001293074 -0.004406401 -0.001572036
#> Sigma[4,2]  0.0001272967 0.016651154 -0.016957808  0.020588070
#> Sigma[4,3]  0.1373490881 0.049647011  0.097252378  0.198765520
#> Sigma[4,4]  0.1408637251 0.043134649  0.104721801  0.196062257
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1  0.89108245 0.007823748  0.88186268  0.899822367
#> lag1_var2  0.41661933 0.041695518  0.38264429  0.473247134
#> lag1_var3  0.46827555 0.046738729  0.42208660  0.528386450
#> lag1_var4 -0.44224228 0.049068840 -0.50619765 -0.395253274
#> const      0.76652812 0.303158860  0.45424177  1.152000542
#> exo1      -0.02743718 0.031037555 -0.06581502  0.001147738
#> exo2      -0.08476729 0.032608040 -0.11567715 -0.043433680
#> exo3      -0.07538064 0.046577949 -0.13535332 -0.030251984
#> 
#> $ARE$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.3550514 0.2255587  0.06182733    0.5564469
#> lag1_var2 -1.1542490 1.2091889 -2.40538737    0.3299947
#> lag1_var3 -2.3053826 1.8298381 -4.19135588   -0.0622133
#> lag1_var4  2.1373115 1.8509011 -0.12951403    4.0572904
#> const      4.4586314 3.5351008  1.51932128    9.1087850
#> exo1       0.1008311 0.3709126 -0.29807597    0.4356806
#> exo2       2.1191869 0.1687116  1.98592263    2.3319507
#> exo3      -0.2872324 0.6846379 -0.89959312    0.6117826
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.4339317 0.2529765   0.1230820  0.703321431
#> lag1_var2  0.1692015 0.9809723  -0.8951269  1.363778224
#> lag1_var3  1.0291988 1.5393226  -0.5558659  2.922982638
#> lag1_var4 -0.1345674 1.5899807  -2.0778371  1.514260152
#> const     -3.4284823 5.5841031  -7.4527926  3.881081097
#> exo1      -0.5466314 0.7045670  -1.4031789  0.191274821
#> exo2      -2.2414318 0.2030252  -2.4010537 -1.965598012
#> exo3      -1.0866574 0.8615553  -1.8891392  0.003451006
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.7508344 0.1177187   0.6006838   0.86662662
#> lag1_var2 -0.6598518 0.3953625  -1.2038678  -0.39144342
#> lag1_var3 -0.6386961 0.6398394  -1.5079464  -0.19728205
#> lag1_var4  1.4183951 0.6647750   0.9479132   2.32078730
#> const     -2.1983450 4.2256796  -6.5386311   3.11693888
#> exo1      -0.4627418 0.4873397  -1.1115455  -0.03445234
#> exo2      -0.5856735 0.1549950  -0.7375808  -0.38668542
#> exo3      -1.2881243 0.5805536  -1.9962664  -0.66850064
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002989254 0.001100511 0.002327049   0.00451016
#> 
#> $ARG$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0609648 0.02284032 -0.08949325  -0.03971301
#> Sigma[2,2]  2.9847195 1.04891752  2.15763719   4.19389256
#> 
#> $ARG$Sigma$equation3
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.0521245 0.02178935  0.03516146   0.08154962
#> Sigma[3,2] -2.2012928 0.81542935 -3.20856800  -1.60919051
#> Sigma[3,3]  1.8273110 0.65515325  1.31424810   2.67507891
#> 
#> $ARG$Sigma$equation4
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01702143 0.01110069  0.006144434   0.03143262
#> Sigma[4,2] -0.43579952 0.23562474 -0.752413984  -0.23544626
#> Sigma[4,3]  0.54755219 0.21242266  0.360887792   0.82208585
#> Sigma[4,4]  0.31633387 0.09358981  0.223694508   0.42269379
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.90274863 0.04500148  0.86115067   0.95706371
#> lag1_var2  0.09610026 0.04867799  0.03463801   0.14690641
#> lag1_var3  0.15857027 0.08044955  0.05957089   0.24616155
#> lag1_var4 -0.14864099 0.07512603 -0.22278717  -0.05178503
#> const      2.03704543 1.20019016  0.71328423   3.33068281
#> exo1       0.04960154 0.06056171 -0.03304399   0.09154127
#> exo2      -0.11963122 0.04756091 -0.16896682  -0.06835322
#> exo3       0.04699439 0.05052774 -0.01049261   0.10256351
#> 
#> $ARG$A$equation2
#>                mean         sd 5% quantile 95% quantile
#> lag1_var1 -1.681767  1.1031635 -3.09372763   -0.7159655
#> lag1_var2 -2.236497  0.9533943 -3.48439171   -1.3317343
#> lag1_var3 -4.774401  1.6406088 -6.96786497   -3.4035537
#> lag1_var4  4.806134  1.4189655  3.25662052    6.5395512
#> const     46.570150 20.5933338 23.05470618   70.5333676
#> exo1      -2.022225  1.1797819 -3.36175208   -0.7621116
#> exo2       1.272061  1.0543496 -0.02732595    2.4099694
#> exo3       1.215528  2.6213155 -2.15533695    3.5064177
#> 
#> $ARG$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.4252206 0.7170052 -0.44340133     1.222897
#> lag1_var2  1.5793335 0.7494437  0.78405914     2.516571
#> lag1_var3  3.2755426 1.2991985  1.98806503     4.945663
#> lag1_var4 -2.5160050 1.1770566 -3.85854968    -1.122717
#> const      1.3429509 7.1919731 -8.15792666     6.894493
#> exo1       0.8233363 0.8308479 -0.01061816     1.873999
#> exo2      -4.3590483 1.0750257 -5.31251808    -2.951469
#> exo3       2.3728888 1.6314211  0.86992847     4.507773
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.7015947  0.6468554 -1.51520417  -0.06043952
#> lag1_var2  0.2037215  0.2333141 -0.07561269   0.44880533
#> lag1_var3  0.3416416  0.3995582 -0.12418146   0.77557493
#> lag1_var4  0.5154428  0.3970639  0.11836650   1.00631677
#> const     27.7007858 19.7459959  5.88592485  49.30643647
#> exo1      -0.4384392  0.3084648 -0.74742476  -0.08941698
#> exo2      -4.0690014  0.6704542 -4.92148946  -3.38922148
#> exo3       3.4504260  0.2228093  3.18788508   3.69408962
#> 
#> 
#>