Provides posterior estimation summary for Bayesian Vector Autoregressions for dynamic panel data

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARs'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARs obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003996409 0.0008618734 0.003339288  0.005121461
#> 
#> $AFG$Sigma$equation2
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.009118261 0.00337599  -0.0129125 -0.005276941
#> Sigma[2,2]  0.219472813 0.04092031   0.1805238  0.271379537
#> 
#> $AFG$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01113873 0.003276636  0.008397572   0.01549838
#> Sigma[3,2] -0.20623023 0.045476057 -0.266888973  -0.16633445
#> Sigma[3,3]  0.37473680 0.089204255  0.273550066   0.48116362
#> 
#> $AFG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.00793448 0.002755485  0.005132438   0.01130909
#> Sigma[4,2] -0.12410240 0.030647015 -0.162557368  -0.09238493
#> Sigma[4,3]  0.32185312 0.099280931  0.225082158   0.44841350
#> Sigma[4,4]  0.30405168 0.110177830  0.204584531   0.44873394
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.982870090 0.01876468   0.9628343   1.00416583
#> lag1_var2 -0.109554343 0.01209174  -0.1208295  -0.09499391
#> lag1_var3 -0.207270636 0.02136116  -0.2354561  -0.18780886
#> lag1_var4  0.176247076 0.02008836   0.1560151   0.20022718
#> const      2.006759863 0.33348521   1.7158439   2.44348649
#> exo1       0.004511079 0.07259758  -0.0821008   0.08322325
#> exo2      -0.057232579 0.05061714  -0.1251683  -0.01625680
#> exo3      -0.214352536 0.09249483  -0.3038308  -0.10252508
#> 
#> $AFG$A$equation2
#>                   mean        sd  5% quantile 95% quantile
#> lag1_var1  0.242353210 0.1972688   0.02766610   0.48254031
#> lag1_var2  1.101930248 0.1461399   0.95479653   1.28055822
#> lag1_var3  0.092374121 0.2200697  -0.06336219   0.39473942
#> lag1_var4 -0.060591561 0.1815694  -0.30545904   0.06672939
#> const     -7.393739989 2.8614310 -10.97388729  -4.38056662
#> exo1      -0.284335685 0.1551965  -0.48527784  -0.14279689
#> exo2      -0.014857937 0.7370309  -0.92642965   0.72569272
#> exo3       0.003315297 0.7580720  -0.91329168   0.84108319
#> 
#> $AFG$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.09467874 0.1963903  -0.2959466   0.14717895
#> lag1_var2 -0.28068128 0.1387823  -0.4553449  -0.14602416
#> lag1_var3  1.17149019 0.2130050   0.8777474   1.31495911
#> lag1_var4 -0.29619329 0.1514166  -0.4282432  -0.09808781
#> const     10.81707412 4.6297022   5.2501095  15.35471408
#> exo1      -0.06417991 0.5047917  -0.7420143   0.37389943
#> exo2      -1.24710735 0.7973218  -2.1348593  -0.33605673
#> exo3      -0.22780066 0.9308489  -0.9820863   1.01171997
#> 
#> $AFG$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.01907044 0.15988881  -0.1362379    0.2262833
#> lag1_var2  0.28117535 0.09047142   0.1634605    0.3687753
#> lag1_var3  1.41359796 0.14258786   1.2213780    1.5318570
#> lag1_var4 -0.37555658 0.08530649  -0.4630801   -0.2689166
#> const      0.80352454 4.63071564  -4.6940195    6.0797690
#> exo1      -0.21349690 0.55561181  -0.9703207    0.2164756
#> exo2      -1.32802590 0.54675824  -1.9885993   -0.7620790
#> exo3      -0.14263749 0.70717873  -0.8278534    0.7749860
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005565439 0.001622835 0.004308801  0.007772082
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.00953617 0.003335653 -0.01296448 -0.005346839
#> Sigma[2,2]  0.11594677 0.031503087  0.09143458  0.158216075
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.006954133 0.002222825  0.00449836   0.00965445
#> Sigma[3,2] -0.049598897 0.014560430 -0.06702335  -0.03562628
#> Sigma[3,3]  0.035257031 0.009120290  0.02543258   0.04571242
#> 
#> $AGO$Sigma$equation4
#>                     mean           sd  5% quantile 95% quantile
#> Sigma[4,1]  0.0007137876 0.0006409753 -0.000119565  0.001275287
#> Sigma[4,2]  0.0440791398 0.0118938282  0.034387507  0.060092838
#> Sigma[4,3] -0.0033287512 0.0047585230 -0.008198521  0.002136447
#> Sigma[4,4]  0.0357829331 0.0064915184  0.030281410  0.044445421
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean          sd 5% quantile  95% quantile
#> lag1_var1  0.93555264 0.020504730  0.91114273  0.9588709989
#> lag1_var2  0.23814115 0.021853159  0.21133299  0.2599815375
#> lag1_var3  0.35682623 0.024602111  0.32942610  0.3822506531
#> lag1_var4 -0.32125988 0.008078789 -0.32961598 -0.3116383305
#> const     -0.51559155 1.246708738 -1.80568680  0.8982830299
#> exo1       0.06283336 0.023366064  0.04023054  0.0915372953
#> exo2      -0.09083266 0.072223679 -0.16656628 -0.0003401139
#> exo3      -0.08429708 0.073677343 -0.16889105 -0.0051757759
#> 
#> $AGO$A$equation2
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1 -0.002125636 0.05360052  -0.05183139   0.06464085
#> lag1_var2  0.855601812 0.11130613   0.74337572   0.98498863
#> lag1_var3 -0.186846072 0.13225887  -0.34865492  -0.04471359
#> lag1_var4  0.254138258 0.05047019   0.18949972   0.30371438
#> const     -5.164243144 6.64196316 -10.05350288   3.89806217
#> exo1       0.164053396 0.17361917   0.02203497   0.39458474
#> exo2       0.480062125 0.41391044  -0.02909700   0.92946412
#> exo3      -0.276137494 0.55301083  -0.90339314   0.36987234
#> 
#> $AGO$A$equation3
#>                    mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.1897484998 0.03578222 -0.21574849 -0.141125544
#> lag1_var2  0.0004120799 0.06039031 -0.08045898  0.051583548
#> lag1_var3  1.0349573699 0.07979459  0.97932694  1.144901331
#> lag1_var4 -0.0698434214 0.02919664 -0.09845168 -0.033856535
#> const      7.7305960026 4.08634328  2.24791810 11.178219611
#> exo1      -0.0364695357 0.03302954 -0.06131304  0.008194842
#> exo2      -1.5099502197 0.24913434 -1.70817761 -1.180225916
#> exo3       3.1113703871 0.21581149  2.84404437  3.340962353
#> 
#> $AGO$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.3216212 0.03318797 -0.35849279   -0.2828884
#> lag1_var2  0.6937422 0.05632541  0.63603512    0.7586092
#> lag1_var3  0.9555741 0.07938347  0.87635080    1.0420810
#> lag1_var4  0.1756099 0.04954283  0.12880485    0.2308687
#> const     -1.6117946 2.93462464 -5.32356316    1.3509584
#> exo1       0.1606533 0.15224809  0.04750193    0.3670873
#> exo2      -1.4387418 0.23876680 -1.73853776   -1.2145313
#> exo3       3.2795180 0.28327758  2.94591255    3.5954732
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002831149 0.0007674082 0.002063502  0.003681707
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.07948064 0.02732312  -0.1165271  -0.05883382
#> Sigma[2,2]  8.49749926 3.08935118   5.4530892  12.10795989
#> 
#> $ALB$Sigma$equation3
#>                  mean         sd   5% quantile 95% quantile
#> Sigma[3,1]  0.0360863 0.03186932   0.009385843   0.07687263
#> Sigma[3,2] -6.3480057 3.70182616 -11.040489991  -2.95938402
#> Sigma[3,3]  7.0744948 4.22732922   3.784519853  12.76523562
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile  95% quantile
#> Sigma[4,1] -0.02192598 0.01864748 -0.03682113 -0.0006399011
#> Sigma[4,2] -0.76144194 2.05946000 -3.54146639  0.8790757828
#> Sigma[4,3]  3.39163598 2.36574110  1.95785736  6.6421132056
#> Sigma[4,4]  3.50170047 1.49501300  1.99347298  5.4074518207
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.97634809 0.01547082  0.960005151  0.994579098
#> lag1_var2 -0.04171969 0.05311903 -0.113051054  0.001908386
#> lag1_var3 -0.07350367 0.08481306 -0.188339154 -0.005735343
#> lag1_var4  0.06170291 0.07125992  0.003589197  0.157811394
#> const      1.24590226 1.04264032  0.609746217  2.671352219
#> exo1       0.01126538 0.05384301 -0.047124237  0.063975273
#> exo2      -0.06987333 0.04649490 -0.125952828 -0.028550745
#> exo3       0.04209321 0.09331417 -0.063265495  0.125613509
#> 
#> $ALB$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -3.5066326  1.317794   -4.343581    -1.703685
#> lag1_var2  4.1483127  1.410007    2.638643     5.900953
#> lag1_var3  6.1367766  2.228321    3.701494     8.809056
#> lag1_var4 -5.0703389  1.848516   -7.372818    -3.119288
#> const     24.7113749 40.135055  -26.656538    63.485804
#> exo1      -0.7760463  2.387814   -3.840548     1.579692
#> exo2      -0.5279577  5.029101   -6.768041     4.763924
#> exo3      -1.5024587  2.226624   -3.569855     1.409871
#> 
#> $ALB$A$equation3
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1  1.91285346  1.638889  -0.09656674     3.615992
#> lag1_var2 -3.01503319  1.540677  -4.62689807    -1.098602
#> lag1_var3 -4.44084815  2.484386  -6.98860046    -1.281371
#> lag1_var4  4.35691856  1.953111   1.93330683     6.460385
#> const     17.15167924 54.872711 -40.26457352    79.287167
#> exo1      -0.61598251  1.833479  -2.21414169     1.837122
#> exo2       0.05064461  5.141201  -5.39756391     6.070862
#> exo3       0.89645352  1.272566  -0.73341348     2.175431
#> 
#> $ALB$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.51183907  1.360693   -2.099184    0.9948385
#> lag1_var2 -0.67392293  1.788759   -2.478225    1.4716741
#> lag1_var3 -0.94663389  2.890305   -4.037692    2.4564316
#> lag1_var4  1.61456834  2.299925   -1.105064    4.0971106
#> const     33.14013165 50.860382  -18.489444   93.2261185
#> exo1      -0.79674371  1.233445   -2.207985    0.3348587
#> exo2      -0.13233720  2.982477   -3.745351    2.8886469
#> exo3      -0.08732292  1.526683   -2.085038    1.2335503
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001602727 0.0001679733 0.001391785   0.00174881
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd  5% quantile  95% quantile
#> Sigma[2,1] -0.002358857 0.001638112 -0.004023372 -0.0005969633
#> Sigma[2,2]  0.062054843 0.012080174  0.045938242  0.0719550055
#> 
#> $ARE$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.00443425 0.002097124  0.002019103  0.006037641
#> Sigma[3,2] -0.06428519 0.027397557 -0.097336794 -0.035203538
#> Sigma[3,3]  0.20455711 0.075242662  0.114808097  0.281348372
#> 
#> $ARE$Sigma$equation4
#>                    mean           sd  5% quantile 95% quantile
#> Sigma[4,1]  0.002469791 0.0009399583  0.001454004  0.003430137
#> Sigma[4,2] -0.016793787 0.0257782163 -0.048143063  0.009885120
#> Sigma[4,3]  0.158440119 0.0587059487  0.088594052  0.217042081
#> Sigma[4,4]  0.149350752 0.0475732458  0.098247453  0.206559083
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.92452743 0.03980252  0.88040966   0.97246263
#> lag1_var2 -0.07649109 0.05372864 -0.13423054  -0.01230466
#> lag1_var3 -0.13753506 0.04915641 -0.18494383  -0.07641965
#> lag1_var4  0.14210396 0.05334105  0.08638602   0.20068515
#> const      1.60851781 0.62865795  0.83473697   2.24133457
#> exo1      -0.01245865 0.04081787 -0.05666493   0.03303180
#> exo2      -0.06800516 0.08025503 -0.15132195   0.03292681
#> exo3      -0.04532068 0.06246626 -0.11871100   0.02630000
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.03707595 0.2183411 -0.23944479    0.2079648
#> lag1_var2  0.49894076 0.3938692 -0.01400890    0.8288306
#> lag1_var3 -0.34023461 0.5072560 -0.95514580    0.1005195
#> lag1_var4  0.29805284 0.5160262 -0.19045791    0.8989749
#> const      2.76977824 3.2220665 -1.20751996    6.0517196
#> exo1       0.17360966 0.1840031  0.02744925    0.4181592
#> exo2       2.21657380 0.2002736  1.96086989    2.3943309
#> exo3      -0.78421730 0.6003921 -1.60013609   -0.3168247
#> 
#> $ARE$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.20677444 0.3554884  -0.1942836    0.6273487
#> lag1_var2  0.29274918 1.4431002  -1.2630410    1.6762039
#> lag1_var3  0.91729992 1.8545516  -0.9632272    2.8686219
#> lag1_var4  0.04603432 1.7724397  -1.7826532    1.8527147
#> const     -3.19198179 5.9919809  -9.8905811    3.3956242
#> exo1      -0.63810988 0.1731160  -0.8086204   -0.4166628
#> exo2      -2.59693688 0.4966228  -2.9416025   -1.9235033
#> exo3      -1.36341077 0.2177206  -1.6267074   -1.1209622
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.2119821 0.4399219  -0.3837155    0.5773680
#> lag1_var2  0.7053724 1.2514813  -0.6520418    2.0301737
#> lag1_var3  0.6839070 1.5306660  -0.9316033    2.3226483
#> lag1_var4  0.2805764 1.4392325  -1.2079179    1.8374290
#> const     -3.0401277 6.1717415  -9.9123552    1.7387419
#> exo1      -0.5211816 0.1947263  -0.7375706   -0.3200310
#> exo2      -0.9122346 0.6049606  -1.4470821   -0.1048747
#> exo3      -2.0178073 0.4738441  -2.6521328   -1.6225621
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002499635 0.0006821014 0.001931107  0.003419542
#> 
#> $ARG$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0548564 0.02260723  -0.0828448  -0.03213147
#> Sigma[2,2]  3.1584921 1.27375085   1.8806795   4.78130274
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04083417 0.01740553  0.02177257   0.06083079
#> Sigma[3,2] -2.37207434 0.99696806 -3.63477275  -1.41840087
#> Sigma[3,3]  2.07376489 0.85321951  1.27231999   3.08333180
#> 
#> $ARG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.008447119 0.005123763  0.002819425   0.01425047
#> Sigma[4,2] -0.502912725 0.270481636 -0.811037402  -0.25461096
#> Sigma[4,3]  0.703230079 0.307695487  0.411883092   1.07439115
#> Sigma[4,4]  0.444413878 0.175253889  0.292531810   0.67567766
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.90468804 0.02890167  0.874560680   0.94097516
#> lag1_var2 -0.07311903 0.04981640 -0.128771112  -0.02001299
#> lag1_var3 -0.11979558 0.08140505 -0.213008218  -0.03325715
#> lag1_var4  0.10098805 0.07564350  0.022538694   0.18802970
#> const      3.72582732 0.70955218  3.021488403   4.62627113
#> exo1       0.01989993 0.05716043 -0.043191256   0.07973582
#> exo2      -0.10097535 0.04055188 -0.147931218  -0.05338464
#> exo3       0.04036227 0.02768915  0.004596326   0.06552722
#> 
#> $ARG$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -1.4120480  1.130796  -2.6636270  -0.09448593
#> lag1_var2  1.2599972  2.796119  -2.0153981   4.08297456
#> lag1_var3  0.8932079  4.632692  -4.5054011   5.52756009
#> lag1_var4 -0.1380364  4.042646  -4.1740097   4.52595172
#> const     -5.0041667 44.483627 -56.9663734  44.49568684
#> exo1      -1.7661656  2.195733  -4.4299696   0.10807685
#> exo2       1.6987272  2.040174  -1.0301886   3.35279750
#> exo3       2.1457047  1.974991  -0.3818593   3.89260432
#> 
#> $ARG$A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.3498402  0.7683849  -0.4378103     1.213939
#> lag1_var2  1.0660912  2.1535521  -1.2316911     3.542623
#> lag1_var3  2.4508210  3.5525142  -1.2741614     6.543712
#> lag1_var4 -1.9188911  3.1384913  -5.4490186     1.395803
#> const     17.3333488 33.6909724 -18.3411230    59.196398
#> exo1       0.3002546  1.3476120  -1.0331587     1.870299
#> exo2      -4.7753630  1.6287300  -5.8526091    -2.531784
#> exo3       1.3291437  1.2308241   0.4679770     3.024651
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.6190532  0.2254034  -0.9282957   -0.4547138
#> lag1_var2  2.0295801  0.5169173   1.3877083    2.5528340
#> lag1_var3  3.3092395  0.8348356   2.3040937    4.1896309
#> lag1_var4 -2.2173149  0.7842322  -2.9960477   -1.2686661
#> const     11.1047918 10.2927137   2.5668602   24.7336708
#> exo1      -0.8694400  0.4899986  -1.2401323   -0.1992392
#> exo2      -4.2284356  0.7438854  -4.9829019   -3.3253463
#> exo3       2.8789474  0.9072381   1.8207214    3.9241950
#> 
#> 
#>