Provides posterior estimation summary for Bayesian Vector Autoregressions for dynamic panel data

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARs'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARs obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003473882 0.001053541 0.002472127  0.004848581
#> 
#> $AFG$Sigma$equation2
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.003201924 0.004054243 -0.008589996 0.0002768622
#> Sigma[2,2]  0.188782224 0.073669788  0.133738943 0.2893624041
#> 
#> $AFG$Sigma$equation3
#>                   mean          sd   5% quantile 95% quantile
#> Sigma[3,1]  0.00438767 0.003879518  0.0004598409  0.009357169
#> Sigma[3,2] -0.17397409 0.061659811 -0.2514434849 -0.110990044
#> Sigma[3,3]  0.31957549 0.063409700  0.2441387890  0.391570446
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.003299574 0.003050474 -0.0004328965  0.006307069
#> Sigma[4,2] -0.102987259 0.038187250 -0.1432952841 -0.055080576
#> Sigma[4,3]  0.274497251 0.044102590  0.2179005413  0.318369218
#> Sigma[4,4]  0.259476944 0.034552186  0.2149498657  0.289701089
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                   mean          sd 5% quantile 95% quantile
#> lag1_var1  0.968147568 0.008735078  0.95612683   0.97416283
#> lag1_var2  0.071254421 0.021591882  0.05497285   0.10024498
#> lag1_var3  0.193263015 0.011427226  0.18106792   0.20648812
#> lag1_var4 -0.175011072 0.012334083 -0.18863144  -0.16002460
#> const      0.086863855 0.581420967 -0.64631393   0.63171511
#> exo1      -0.021125192 0.064534548 -0.08817771   0.04348158
#> exo2       0.008550295 0.063305148 -0.05828756   0.08048572
#> exo3      -0.159688292 0.030943866 -0.18666568  -0.11908756
#> 
#> $AFG$A$equation2
#>                   mean         sd   5% quantile 95% quantile
#> lag1_var1  0.300160456 0.06097937   0.237678593    0.3734715
#> lag1_var2  0.995741553 0.17189710   0.825245216    1.2116441
#> lag1_var3 -0.115411336 0.29088260  -0.360590277    0.2691794
#> lag1_var4  0.147545613 0.22683220  -0.142174826    0.3203322
#> const     -8.612794931 3.24141381 -12.826563790   -5.8488165
#> exo1      -0.012086917 0.33521424  -0.456192263    0.2944826
#> exo2       0.009744388 0.36057672  -0.396834354    0.4329138
#> exo3       0.204594240 0.21671974  -0.009868828    0.4560061
#> 
#> $AFG$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.13707738 0.1564690  -0.3334838  0.002478179
#> lag1_var2 -0.50479261 0.2741756  -0.7532224 -0.145104571
#> lag1_var3  0.60468626 0.3868488   0.1384187  1.015308764
#> lag1_var4  0.20275961 0.3513311  -0.1416161  0.618183200
#> const     14.53100153 7.0307945   6.3318558 22.256953280
#> exo1       0.08808268 0.6763063  -0.7422519  0.755823051
#> exo2      -1.07036192 0.4086048  -1.5671824 -0.627435415
#> exo3      -0.66587385 0.5247952  -1.3335453 -0.161329577
#> 
#> $AFG$A$equation4
#>                    mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.0003957748 0.1616532 -0.20900651   0.14018374
#> lag1_var2 -0.0348986275 0.2438757 -0.27534125   0.28060066
#> lag1_var3  0.6485826881 0.2948631  0.31827732   0.99265151
#> lag1_var4  0.3087724478 0.2814033  0.01212158   0.62697207
#> const      4.5929054995 7.0018632 -3.17131920  12.48393282
#> exo1       0.0949023567 0.6956394 -0.72610095   0.88244773
#> exo2      -1.1216759939 0.3730369 -1.47735597  -0.63719313
#> exo3      -0.5509956588 0.4864293 -1.18076774  -0.08183616
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005858157 0.001332992 0.004473124  0.007533158
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01267869 0.005277881  -0.0198822 -0.008596907
#> Sigma[2,2]  0.14210182 0.028743728   0.1145248  0.176349116
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.009716955 0.003381383  0.006876838   0.01421430
#> Sigma[3,2] -0.064934162 0.011545248 -0.075569505  -0.04981846
#> Sigma[3,3]  0.042540404 0.009391874  0.031045516   0.05198380
#> 
#> $AGO$Sigma$equation4
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.0003243748 0.001169249 -0.001047413  0.001651798
#> Sigma[4,2]  0.0510061772 0.018678385  0.031245697  0.071593284
#> Sigma[4,3] -0.0082673710 0.009960762 -0.018662853  0.003068942
#> Sigma[4,4]  0.0366534275 0.008047398  0.025813632  0.043214444
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.00325193 0.04033141  0.95263242   1.03607361
#> lag1_var2 -0.16661964 0.07356161 -0.23466297  -0.08289119
#> lag1_var3 -0.16059213 0.08792420 -0.24388561  -0.06173941
#> lag1_var4  0.17735461 0.08403256  0.07653528   0.24527050
#> const     -0.62532071 2.05038856 -2.99656758   1.71108907
#> exo1       0.04413427 0.11052151 -0.10807002   0.11594414
#> exo2      -0.05570401 0.08394934 -0.14448499   0.04788934
#> exo3       0.14731496 0.04065453  0.09187064   0.16829438
#> 
#> $AGO$A$equation2
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.02039552 0.09738893  -0.09275591   0.12459509
#> lag1_var2  1.14831367 0.21399762   0.96025117   1.40427536
#> lag1_var3  0.12501587 0.26231633  -0.16156611   0.35936131
#> lag1_var4 -0.08512035 0.21546587  -0.32389939   0.14816661
#> const     -4.54035520 7.62231658 -12.27279907   5.21454848
#> exo1       0.12349730 0.44119816  -0.34242530   0.58210991
#> exo2       0.60911793 0.28945176   0.29073115   0.95092964
#> exo3      -0.65347297 0.62618108  -1.31409309   0.02038845
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.17818555 0.10994567 -0.30978118  -0.07607604
#> lag1_var2  0.09037314 0.13616602 -0.06285676   0.25181800
#> lag1_var3  1.17898243 0.10381735  1.06432193   1.28971793
#> lag1_var4 -0.10013819 0.12847905 -0.26533469   0.01811300
#> const     -0.92801221 7.01370598 -8.14356690   7.51548418
#> exo1      -0.07195491 0.19599217 -0.32381913   0.11233839
#> exo2      -1.44366875 0.09189703 -1.54361994  -1.33378880
#> exo3       3.53151302 0.33239814  3.12345653   3.88718338
#> 
#> $AGO$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  -0.2223938 0.08583741  -0.3377363  -0.14820375
#> lag1_var2   1.1164530 0.18861698   0.8981255   1.33597809
#> lag1_var3   1.4912265 0.31529078   1.1209253   1.81635588
#> lag1_var4  -0.1919903 0.22886217  -0.4307361   0.07920955
#> const     -17.1701488 5.18922316 -23.9364331 -12.47643471
#> exo1        0.0265431 0.24952678  -0.1722253   0.36615645
#> exo2       -1.2238269 0.18514746  -1.3461043  -0.96931713
#> exo3        3.5440812 0.31404901   3.1605576   3.89047840
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001929902 0.0004985556  0.00144431  0.002560637
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05712938 0.02887128 -0.09558956  -0.03251096
#> Sigma[2,2]  7.13742705 2.64591769  4.69571140  10.20338192
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.02784506 0.01671533  0.01643645   0.05076926
#> Sigma[3,2] -4.53564007 1.46425892 -6.31598996  -2.98263642
#> Sigma[3,3]  4.63075418 1.31341341  2.96475039   5.97903680
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.01200227 0.01463608 -0.02906177  0.004645709
#> Sigma[4,2]  0.36155104 0.93458751 -0.85726025  1.246491855
#> Sigma[4,3]  1.86119228 1.38119260  0.74253112  3.585509708
#> Sigma[4,4]  2.53812459 1.21498644  1.42330188  3.951952208
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                    mean         sd 5% quantile 95% quantile
#> lag1_var1  1.0297927812 0.03112182  0.98729296   1.05073534
#> lag1_var2  0.0005353494 0.02444823 -0.03267716   0.01875367
#> lag1_var3 -0.0104468969 0.03382655 -0.05617231   0.01563788
#> lag1_var4  0.0098220474 0.02880932 -0.01270002   0.04820798
#> const     -0.7116037329 1.07457374 -1.36216630   0.77602400
#> exo1       0.0333388002 0.02543010  0.01798394   0.06785525
#> exo2      -0.0358073196 0.01107702 -0.04450704  -0.02077839
#> exo3       0.0697410532 0.02874930  0.04189587   0.10112158
#> 
#> $ALB$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -3.1161402  1.141592   -4.293348   -1.6898529
#> lag1_var2  3.5495198  1.688316    2.198386    5.8073594
#> lag1_var3  4.8911933  2.559666    2.715637    8.2280873
#> lag1_var4 -3.9950679  2.200225   -6.840197   -2.0391082
#> const     23.8869995 46.262917  -39.961306   52.6252676
#> exo1      -3.0078535  1.915556   -5.300535   -0.8700361
#> exo2      -0.7952923  2.685212   -3.357311    2.2902262
#> exo3       0.9546585  3.172327   -2.171292    4.7612798
#> 
#> $ALB$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  1.6983561  2.291613   -1.112366    3.6931533
#> lag1_var2 -2.0773337  1.308723   -3.742312   -0.7762049
#> lag1_var3 -2.6401763  2.163129   -5.270315   -0.3253398
#> lag1_var4  2.8576677  1.871531    0.837472    5.1401493
#> const      6.2533619 49.378097  -47.233989   55.0224413
#> exo1      -0.1259217  1.813200   -2.264661    1.7655215
#> exo2       0.4476017  2.618716   -2.463726    2.9333248
#> exo3      -0.2522144  2.498300   -3.626765    1.6961666
#> 
#> $ALB$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.38598303  2.5945089  -3.5372453    1.9744899
#> lag1_var2 -0.04220783  0.7196040  -0.9415544    0.6224517
#> lag1_var3  0.17755216  1.1400367  -1.0262869    1.5328568
#> lag1_var4  0.71060753  0.8951399  -0.3704739    1.5831216
#> const     17.96613619 63.9437779 -38.0578042   91.0982554
#> exo1      -2.17725192  1.3335796  -3.0993145   -0.3455252
#> exo2       0.06759043  1.6191302  -1.5778306    2.0309284
#> exo3       0.52113671  1.5919194  -1.1211108    2.5379345
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd  5% quantile 95% quantile
#> Sigma[1,1] 0.001381857 0.0003819949 0.0008692886  0.001694782
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.001995356 0.000740160 -0.002933663  -0.00135057
#> Sigma[2,2]  0.048182910 0.008351032  0.038835273   0.05827651
#> 
#> $ARE$Sigma$equation3
#>                    mean           sd  5% quantile 95% quantile
#> Sigma[3,1]  0.003545076 0.0008603473  0.002724368  0.004588023
#> Sigma[3,2] -0.048678957 0.0132279698 -0.066166315 -0.038118059
#> Sigma[3,3]  0.162551003 0.0375204811  0.131240483  0.213303918
#> 
#> $ARE$Sigma$equation4
#>                   mean           sd  5% quantile  95% quantile
#> Sigma[4,1]  0.00194250 0.0008005484  0.001099271  0.0029124244
#> Sigma[4,2] -0.01207235 0.0097411372 -0.022494693 -0.0006830496
#> Sigma[4,3]  0.12801007 0.0298358180  0.102046352  0.1676569861
#> Sigma[4,4]  0.12185358 0.0267812789  0.098505691  0.1576532213
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.96901191 0.05289685  0.90109435   1.01473354
#> lag1_var2  0.16887346 0.01719508  0.15274282   0.19131548
#> lag1_var3  0.21708364 0.01219535  0.20379019   0.22866764
#> lag1_var4 -0.21112722 0.02258848 -0.22944747  -0.18286842
#> const      0.39280103 0.49147866 -0.03934201   0.96676143
#> exo1      -0.03420468 0.03201990 -0.07745140  -0.01011433
#> exo2      -0.06952554 0.02732169 -0.10513870  -0.04481170
#> exo3       0.02446197 0.06583109 -0.05285534   0.08976241
#> 
#> $ARE$A$equation2
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.18823893 0.27848375 -0.178367805   0.43299845
#> lag1_var2  0.81534229 0.09746810  0.691225977   0.90913516
#> lag1_var3  0.11039844 0.08808042  0.007820725   0.20230013
#> lag1_var4 -0.17758177 0.12939435 -0.317485818  -0.03438501
#> const      0.82049912 2.01966506 -0.603254799   3.56759028
#> exo1       0.05608031 0.17201328 -0.142278989   0.23690835
#> exo2       1.96159896 0.23973602  1.673484759   2.23064543
#> exo3      -0.40790792 0.12820271 -0.546133912  -0.27411672
#> 
#> $ARE$A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.2283380 0.54348438  -0.3506295   0.88062959
#> lag1_var2 -0.7490198 0.04909755  -0.8035219  -0.69475360
#> lag1_var3 -0.2974475 0.24192888  -0.5920366  -0.03770898
#> lag1_var4  1.2722857 0.33061450   0.8629337   1.62979434
#> const     -4.4367104 6.82325520  -9.4908275   4.32541403
#> exo1      -0.5743834 0.35816206  -0.9817712  -0.17220885
#> exo2      -2.1189594 0.42898908  -2.6407982  -1.65147871
#> exo3      -1.6496420 0.47642484  -2.1000707  -1.06782762
#> 
#> $ARE$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.4009327 0.48622416  -0.2350581   0.83868100
#> lag1_var2 -0.1225939 0.06674042  -0.1901881  -0.04593141
#> lag1_var3 -0.2149949 0.30322169  -0.6008378   0.06398894
#> lag1_var4  1.1567154 0.37075682   0.7810831   1.62991009
#> const     -5.9609492 6.46658325 -11.7155427   2.37488838
#> exo1      -0.5491333 0.49167028  -1.1353102   0.00110048
#> exo2      -0.5834243 0.31944136  -1.0156194  -0.34299020
#> exo3      -2.0127833 0.45034628  -2.5072048  -1.49537308
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                  mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00275667 0.0006046695 0.002146467  0.003530887
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.06388087 0.02636796 -0.09751467  -0.03900454
#> Sigma[2,2]  3.38285928 1.43700262  2.14191823   5.24543347
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04688416 0.01758365  0.03249624   0.07052325
#> Sigma[3,2] -2.36159816 1.00825786 -3.69600315  -1.47484670
#> Sigma[3,3]  1.87263636 0.71869354  1.26117273   2.83839929
#> 
#> $ARG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.009368387 0.004498496  0.004038599   0.01444873
#> Sigma[4,2] -0.331446513 0.187027035 -0.577774058  -0.17074621
#> Sigma[4,3]  0.481151281 0.133313302  0.395260020   0.66242274
#> Sigma[4,4]  0.312591819 0.035000984  0.281110196   0.35713890
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.918884330 0.05075748  0.87109261   0.98454846
#> lag1_var2  0.102648006 0.04669976  0.04507456   0.14985124
#> lag1_var3  0.165662066 0.07244724  0.07962331   0.24370554
#> lag1_var4 -0.162298995 0.07283652 -0.23814387  -0.07376056
#> const      1.983526581 1.28272342  0.37495557   3.32777874
#> exo1       0.007433035 0.06101108 -0.05602799   0.07690634
#> exo2      -0.099939884 0.02447618 -0.12118050  -0.06795347
#> exo3      -0.008847789 0.06600786 -0.06586808   0.07710285
#> 
#> $ARG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -2.5226555  0.8751988   -3.706890   -1.8526934
#> lag1_var2  0.8537136  3.0755346   -1.471024    4.8036112
#> lag1_var3  0.3499568  4.9079906   -3.313088    6.6554455
#> lag1_var4  0.1651624  4.8025564   -5.971478    3.7111988
#> const     40.0450903 28.6280410   18.456227   78.4377384
#> exo1      -1.2272293  1.8653105   -3.576947    0.5225333
#> exo2       2.0236526  0.8233850    1.077787    2.6884866
#> exo3       0.5328896  2.1779723   -2.299606    2.5790071
#> 
#> $ARG$A$equation3
#>                 mean         sd   5% quantile 95% quantile
#> lag1_var1  0.7914537  0.8178849  5.806433e-04     1.787713
#> lag1_var2  1.1864088  2.0418446 -1.496146e+00     2.868233
#> lag1_var3  2.5580964  3.2596520 -1.747545e+00     5.165415
#> lag1_var4 -1.9208419  3.1168410 -4.374356e+00     2.164416
#> const     -1.4847901 24.0272181 -3.426509e+01    14.799711
#> exo1       0.3857462  1.1455776 -6.720919e-01     1.621830
#> exo2      -5.2707701  0.5596460 -5.796852e+00    -4.582860
#> exo3       2.3445881  1.0968378  9.704408e-01     3.496600
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.8524023  0.4685211  -1.4494706   -0.5054401
#> lag1_var2  1.8936786  0.3142166   1.5174722    2.2403818
#> lag1_var3  3.0704349  0.5172620   2.4402923    3.6177899
#> lag1_var4 -2.0160024  0.4262129  -2.4902156   -1.5215544
#> const     19.4727951 11.0464295   7.1573691   32.0656991
#> exo1      -0.4314079  0.5764986  -0.9834166    0.2781645
#> exo2      -4.5493175  0.2613862  -4.7710628   -4.2643427
#> exo3       2.9228930  0.5964008   2.3132751    3.6505761
#> 
#> 
#>