Provides posterior estimation summary for Bayesian Vector Autoregressions for dynamic panel data

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARs'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARs obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003742491 0.00121851 0.002670771  0.005351018
#> 
#> $AFG$Sigma$equation2
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.005787942 0.004743631 -0.009862279 0.0005593159
#> Sigma[2,2]  0.215624362 0.078727560  0.151072485 0.3213892819
#> 
#> $AFG$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.008773777 0.003256514  0.004937753    0.0123961
#> Sigma[3,2] -0.169882344 0.060517757 -0.246668343   -0.1077140
#> Sigma[3,3]  0.236613481 0.048615290  0.186278408    0.2959175
#> 
#> $AFG$Sigma$equation4
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[4,1]  0.00729592 0.00559154  0.002830013   0.01480092
#> Sigma[4,2] -0.08678254 0.03855223 -0.126175134  -0.03796860
#> Sigma[4,3]  0.18772415 0.03611426  0.154655936   0.22950878
#> Sigma[4,4]  0.17428553 0.03474068  0.132186456   0.21216940
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.99436790 0.02401600  0.97134091   1.02369104
#> lag1_var2  0.18002299 0.03746690  0.13223291   0.21822177
#> lag1_var3  0.43232046 0.06904834  0.35542571   0.51060421
#> lag1_var4 -0.37989648 0.05570011 -0.44332937  -0.32206995
#> const     -2.07082971 1.23721167 -3.28469709  -0.56921667
#> exo1      -0.04203641 0.06745755 -0.10568145   0.04538848
#> exo2       0.03946281 0.07743139 -0.06296736   0.10900092
#> exo3      -0.13917735 0.06559676 -0.20109141  -0.06424233
#> 
#> $AFG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.2254992  0.2155910 -0.04998068    0.4169854
#> lag1_var2 -1.3796659  0.5989014 -2.09854834   -0.8088986
#> lag1_var3 -5.1130656  1.4878588 -6.84934413   -3.5761264
#> lag1_var4  4.4139428  1.3189962  3.02980263    5.9415812
#> const     27.5459122 11.5905582 16.93788384   42.2130164
#> exo1      -0.4820926  0.4028395 -1.01232981   -0.1283907
#> exo2      -0.1161694  0.4387936 -0.56880869    0.3943680
#> exo3      -0.3763987  0.5092249 -0.79138499    0.3050731
#> 
#> $AFG$A$equation3
#>                    mean        sd   5% quantile 95% quantile
#> lag1_var1   0.000191321 0.1580760  -0.172083750   0.15596513
#> lag1_var2   5.227748855 0.2690028   4.887086621   5.48321703
#> lag1_var3  13.000595973 0.6361590  12.206442467  13.62435802
#> lag1_var4 -10.527178035 0.5495236 -11.078974999  -9.85038530
#> const     -65.919887459 6.3907366 -71.227168152 -57.50290261
#> exo1        0.486547908 0.3550629  -0.003295609   0.69759926
#> exo2       -0.437061249 0.5025733  -1.061517907   0.09725732
#> exo3        0.931645080 0.3569408   0.540417819   1.33136461
#> 
#> $AFG$A$equation4
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1   0.1226008 0.09619314   0.01712764   0.21122213
#> lag1_var2   5.3258622 0.43380771   4.92006723   5.80909953
#> lag1_var3  12.3108371 1.00563596  11.28602313  13.39879336
#> lag1_var4  -9.8155200 0.88469196 -10.75102816  -8.89637916
#> const     -69.3277299 5.65160521 -75.70157806 -63.52873661
#> exo1        0.3131105 0.36331899  -0.15508900   0.65320398
#> exo2       -0.4268958 0.47215789  -1.01445725   0.07950119
#> exo3        0.9905784 0.36671464   0.63797640   1.46514684
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.005094186 0.001035906 0.004415518   0.00650218
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01158587 0.006749374 -0.01839709 -0.003517164
#> Sigma[2,2]  0.19887895 0.086861812  0.10356537  0.293705114
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.009614846 0.002218416  0.006560598   0.01116353
#> Sigma[3,2] -0.091126896 0.038563695 -0.132021650  -0.04937858
#> Sigma[3,3]  0.056931890 0.013923002  0.041399404   0.07174471
#> 
#> $AGO$Sigma$equation4
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.0006906964 0.003539534 -0.003449192  0.003984922
#> Sigma[4,2]  0.0654602418 0.026899747  0.035150127  0.093886436
#> Sigma[4,3] -0.0126177432 0.017086004 -0.027523204  0.009495391
#> Sigma[4,4]  0.0423994463 0.006522296  0.034282721  0.049247014
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.02292470 0.02277168  0.99741257   1.04354839
#> lag1_var2 -0.41202534 0.19637534 -0.57734999  -0.14746686
#> lag1_var3 -0.50156864 0.26342346 -0.73935693  -0.15733946
#> lag1_var4  0.47018781 0.22230894  0.17801359   0.67793629
#> const      2.31925410 2.94322516 -1.09085965   5.49056670
#> exo1       0.01993681 0.06923586 -0.07221942   0.07616918
#> exo2       0.04270694 0.10023231 -0.06974256   0.15469943
#> exo3       0.15959243 0.17152700 -0.03811017   0.36142734
#> 
#> $AGO$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.07032593  0.1390578  -0.1165512    0.1783558
#> lag1_var2 -2.82589315  1.0544122  -4.0000106   -1.6356114
#> lag1_var3 -4.68001859  1.4066253  -6.3730756   -3.3199607
#> lag1_var4  4.23242000  1.2148727   2.9853328    5.6664018
#> const     36.27013607 14.3224707  24.4456843   55.4145634
#> exo1      -0.04212019  0.2883586  -0.4014736    0.2472217
#> exo2       0.90565958  0.3946348   0.4009400    1.2952405
#> exo3       0.60084353  0.8377724  -0.3981577    1.4533906
#> 
#> $AGO$A$equation3
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  -0.23378082 0.04571329  -0.2917341   -0.1877977
#> lag1_var2   2.07598576 0.25774128   1.8249135    2.4003259
#> lag1_var3   3.59825785 0.41829697   3.1782999    4.1274017
#> lag1_var4  -2.28450493 0.32402053  -2.6940230   -1.9912878
#> const     -19.67564593 7.90026618 -29.2962235  -11.0675119
#> exo1        0.09924527 0.18380208  -0.1172744    0.3107584
#> exo2       -1.62983086 0.19473184  -1.8006086   -1.3784411
#> exo3        2.88717597 0.45702747   2.4097174    3.3781125
#> 
#> $AGO$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.2453316 0.09733979  -0.3766251   -0.1722356
#> lag1_var2 -0.2400916 0.54738551  -0.7945377    0.4331384
#> lag1_var3 -0.1879962 0.66012703  -0.9227189    0.5470478
#> lag1_var4  1.2684882 0.59156777   0.6306052    1.9333177
#> const      1.2580951 7.79623734  -9.5347600    6.1725881
#> exo1       0.1008041 0.26811842  -0.1160005    0.4600508
#> exo2      -1.1879718 0.23769102  -1.3684735   -0.8735530
#> exo3       3.8389473 0.23423123   3.5422406    4.0664356
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.006047828 0.008483497 0.001581379   0.01760456
#> 
#> $ALB$Sigma$equation2
#>                  mean        sd 5% quantile 95% quantile
#> Sigma[2,1] -0.1215937 0.1804202   -0.366202 -0.001287427
#> Sigma[2,2]  8.9832024 3.9739192    5.530294 14.012098759
#> 
#> $ALB$Sigma$equation3
#>                   mean        sd 5% quantile 95% quantile
#> Sigma[3,1]  0.08220909 0.1368646 -0.05227599    0.2591784
#> Sigma[3,2] -7.33278547 1.8836918 -9.78318090   -5.7610525
#> Sigma[3,3]  9.15997465 2.3339538  7.29596586   12.2925611
#> 
#> $ALB$Sigma$equation4
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[4,1]  0.004127623 0.05268662 -0.06346027   0.05590564
#> Sigma[4,2] -1.691741341 1.11432798 -3.15342638  -0.77357995
#> Sigma[4,3]  5.200855066 3.26414229  2.58659991   9.62674497
#> Sigma[4,4]  4.929984069 3.07716622  2.50660431   9.11039470
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.97407872 0.09149442  0.878070183   1.08691367
#> lag1_var2  0.03981115 0.05601474 -0.004487713   0.11567844
#> lag1_var3  0.04995818 0.08007308 -0.021334859   0.15613769
#> lag1_var4 -0.04327744 0.06850241 -0.132019015   0.01847590
#> const      0.08045087 1.91359656 -2.356994043   2.02201176
#> exo1      -0.01516358 0.06683439 -0.083277688   0.06202328
#> exo2      -0.02580545 0.04950418 -0.089311046   0.02363504
#> exo3       0.10597293 0.11893618  0.024541614   0.26878943
#> 
#> $ALB$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -4.0730561  1.805934 -5.74030890  -1.72996947
#> lag1_var2  1.5906845  1.363294  0.22160951   3.20466883
#> lag1_var3  2.2272503  2.273352 -0.03407769   5.04031103
#> lag1_var4 -1.7503773  1.930925 -4.18060670   0.14245768
#> const     76.4783570 29.444381 38.04061860 101.74970063
#> exo1      -1.4850781  1.701486 -3.44844004   0.03744467
#> exo2      -1.1815547  2.535876 -3.20506612   2.02221851
#> exo3      -0.8992389  2.880803 -3.79000434   2.55590240
#> 
#> $ALB$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.9712674  2.879672   -2.665027    3.9569519
#> lag1_var2 -0.4721729  1.376906   -1.989181    0.9445047
#> lag1_var3 -0.3412070  2.322352   -2.812572    2.1155524
#> lag1_var4  0.8095911  1.910691   -1.314480    2.9418695
#> const      4.1871456 92.392251  -65.687525  129.5891409
#> exo1      -0.4908784  1.418274   -2.058238    1.1581044
#> exo2       1.2015761  2.593478   -1.767245    3.9256149
#> exo3       1.0432301  2.914669   -2.862694    3.5943622
#> 
#> $ALB$A$equation4
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1 -2.11582552   3.4406353  -6.7082266    0.7087933
#> lag1_var2  0.28718404   0.9543711  -0.9253173    1.1797575
#> lag1_var3  0.77039973   1.5814822  -1.1762535    2.3372769
#> lag1_var4  0.07551287   1.2197148  -1.2250236    1.3579944
#> const     60.27249533 107.5835222 -18.7970376  206.9742328
#> exo1      -1.36499230   2.0516183  -3.5217221    1.0567450
#> exo2       0.62742440   1.5912796  -0.6257073    2.7829940
#> exo3       0.65137067   2.2655734  -1.8249626    3.3046129
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd  5% quantile 95% quantile
#> Sigma[1,1] 0.001535629 0.0006126066 0.0009609807  0.002290917
#> 
#> $ARE$Sigma$equation2
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.00164914 0.001415279 -0.003190562 6.067688e-05
#> Sigma[2,2]  0.04865705 0.011618639  0.037907883 6.376382e-02
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.004488842 0.002032375  0.001939437  0.006258824
#> Sigma[3,2] -0.048557797 0.017830517 -0.073008912 -0.035497540
#> Sigma[3,3]  0.151353191 0.032404336  0.125339498  0.194818861
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.003299543 0.002465652  0.0007298837  0.006291499
#> Sigma[4,2] -0.011867831 0.011867686 -0.0261319059  0.001361961
#> Sigma[4,3]  0.115455321 0.021137130  0.0956280290  0.142877686
#> Sigma[4,4]  0.107572400 0.018337665  0.0889197750  0.126569693
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd 5% quantile  95% quantile
#> lag1_var1  0.95174239 0.04247425  0.89974545  0.9976398031
#> lag1_var2 -0.12612173 0.02577959 -0.14763856 -0.0919428557
#> lag1_var3 -0.19478088 0.03332416 -0.22197533 -0.1496179428
#> lag1_var4  0.19584898 0.02771480  0.15783738  0.2140692215
#> const      1.17194773 0.66116448  0.63166843  2.0606354123
#> exo1      -0.03597119 0.04116860 -0.09060427  0.0009909524
#> exo2      -0.07621785 0.04860376 -0.11386136 -0.0128701176
#> exo3      -0.02680388 0.04676711 -0.06629541  0.0336010554
#> 
#> $ARE$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.1092054 0.1034647  0.01447276   0.24616189
#> lag1_var2  0.5325126 0.1288823  0.39664308   0.69315014
#> lag1_var3 -0.2568379 0.1110080 -0.34558759  -0.10696365
#> lag1_var4  0.1944043 0.1222031  0.03151071   0.29698732
#> const      2.5703821 2.1917583 -0.37821151   4.43867474
#> exo1      -0.1037826 0.1081967 -0.24024518  -0.01226457
#> exo2       2.0252542 0.3241924  1.58604985   2.25988556
#> exo3      -0.6743177 0.1891929 -0.90071606  -0.48235144
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.2551980 0.3329490  -0.1516681    0.5505981
#> lag1_var2 -0.6930452 0.3529580  -1.0749715   -0.2903111
#> lag1_var3 -0.3095763 0.4384580  -0.5910169    0.2963619
#> lag1_var4  1.2835348 0.4419587   0.6725934    1.5689162
#> const     -5.2629125 5.1839668 -10.6645399    1.2552438
#> exo1      -0.5278626 0.2440751  -0.7408398   -0.2335968
#> exo2      -2.5722291 0.3662172  -2.8790854   -2.1035994
#> exo3      -1.6092053 0.4530083  -2.2288494   -1.2839794
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.3501394 0.2963347   0.0311038    0.6525605
#> lag1_var2 -0.2994696 0.3679128  -0.7111204    0.1077628
#> lag1_var3 -0.5241973 0.4891054  -0.8695759    0.1392767
#> lag1_var4  1.4720769 0.4999185   0.7864554    1.8229289
#> const     -5.2513536 4.2775703 -10.0833706   -0.7707723
#> exo1      -0.6091811 0.2085198  -0.8449274   -0.3936237
#> exo2      -0.9919549 0.2126680  -1.1864722   -0.7196519
#> exo3      -2.1628139 0.5025892  -2.8202709   -1.6921072
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004183292 0.002758743 0.001827526  0.007630709
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.09334501 0.07651302  -0.1982094   -0.0384114
#> Sigma[2,2]  3.98936550 2.12988784   2.2995857    6.8884193
#> 
#> $ARG$Sigma$equation3
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.0740476 0.05817823   0.0317569     0.152649
#> Sigma[3,2] -2.8636394 1.69269083  -5.1729223    -1.558255
#> Sigma[3,3]  2.2860801 1.28430634   1.3609750     4.044588
#> 
#> $ARG$Sigma$equation4
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01969327 0.01609161  0.003360442   0.03644552
#> Sigma[4,2] -0.47515542 0.45255040 -1.094472279  -0.16613679
#> Sigma[4,3]  0.60241101 0.29469747  0.365666466   1.00068527
#> Sigma[4,4]  0.35396366 0.06244394  0.285967465   0.42715817
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.89042045 0.06852075  0.79688948   0.93789776
#> lag1_var2  0.20370249 0.13418133  0.04609934   0.33660336
#> lag1_var3  0.33179548 0.21974362  0.07063724   0.54847765
#> lag1_var4 -0.31967964 0.20718695 -0.53275907  -0.07692002
#> const      2.21366426 1.57992074  1.18898160   4.39880447
#> exo1       0.01815182 0.03848400 -0.02048234   0.06689641
#> exo2      -0.10724182 0.05331100 -0.16510115  -0.05330650
#> exo3      -0.08674779 0.07020543 -0.18024109  -0.02336657
#> 
#> $ARG$A$equation2
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1 -2.51463151  1.906721  -4.96413416   -0.5987415
#> lag1_var2  0.06392188 11.222920 -12.93172034   12.5638774
#> lag1_var3 -0.97262841 18.423474 -22.16294397   19.6800997
#> lag1_var4  1.47703483 16.709081 -17.13369120   20.7136441
#> const     40.52614844 75.039186 -17.74940791  138.0597350
#> exo1      -0.82372571  1.381480  -2.59855576    0.5215627
#> exo2       2.05619869  2.343048  -0.13338687    4.7567710
#> exo3       2.06398462  2.635642   0.01086334    5.1618533
#> 
#> $ARG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1   0.8471368  1.5138687  -0.7280225    2.7396134
#> lag1_var2   5.1165661  6.4728977  -2.1554768   12.6289715
#> lag1_var3   9.0300783 10.6301365  -3.0104598   21.2775486
#> lag1_var4  -7.8509471  9.6835122 -18.9483399    3.0677976
#> const     -35.8577490 42.7997933 -89.2032550    7.4247319
#> exo1       -0.3733564  0.9220766  -1.3273229    0.6470883
#> exo2       -4.9927743  1.9507269  -7.0257509   -2.8508781
#> exo3        0.3245324  1.9019930  -2.2533193    1.7008396
#> 
#> $ARG$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  -0.7837412  0.5329069  -1.2441821   -0.2018695
#> lag1_var2   5.5787794  1.1315256   4.7604296    7.1350067
#> lag1_var3   9.1245444  1.8840193   7.8042455   11.7216399
#> lag1_var4  -7.5046759  1.6715071  -9.7983408   -6.2881181
#> const     -16.6477890 17.4374637 -37.1209221    2.2438438
#> exo1       -0.9699304  0.5924604  -1.6594234   -0.3322217
#> exo2       -4.2515337  0.9107801  -5.0715650   -3.1856193
#> exo3        1.7244100  0.8183617   0.6959941    2.5842823
#> 
#> 
#>