Provides posterior estimation summary for Bayesian Vector Autoregressions for dynamic panel data

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARs'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARs obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004117778 0.0006509483 0.003401732  0.004788053
#> 
#> $AFG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01321128 0.002177195 -0.01618229  -0.01165448
#> Sigma[2,2]  0.20576374 0.038140763  0.17609410   0.25782031
#> 
#> $AFG$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.009220584 0.004775749  0.005052472    0.0154384
#> Sigma[3,2] -0.148859562 0.033799278 -0.191687682   -0.1144265
#> Sigma[3,3]  0.249422068 0.068022051  0.183551874    0.3360785
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.004279484 0.004517905 -0.0002410275  0.009604187
#> Sigma[4,2] -0.068197844 0.023600586 -0.0931355158 -0.038647253
#> Sigma[4,3]  0.210863634 0.072136837  0.1515586086  0.306840861
#> Sigma[4,4]  0.207485827 0.078688486  0.1537614227  0.314822287
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.99334495 0.01465086   0.9771408   1.01115753
#> lag1_var2  0.53329608 0.04703103   0.4793698   0.58073756
#> lag1_var3  1.16831817 0.08053655   1.0731414   1.25020531
#> lag1_var4 -1.01502263 0.05933251  -1.0760738  -0.94475403
#> const     -6.84710547 1.36159462  -8.3746080  -5.25490696
#> exo1      -0.05286892 0.10533213  -0.1809590   0.04817736
#> exo2       0.02491845 0.08960174  -0.0871985   0.10126791
#> exo3      -0.10118439 0.05391374  -0.1632307  -0.04762465
#> 
#> $AFG$A$equation2
#>                  mean          sd 5% quantile 95% quantile
#> lag1_var1  0.22037998  0.04779453   0.1599263    0.2681948
#> lag1_var2 -0.83707832  0.46541552  -1.3580356   -0.3367157
#> lag1_var3 -3.98421328  0.85929228  -4.9114163   -3.0762804
#> lag1_var4  3.46213194  0.67194453   2.7108443    4.1623355
#> const     19.21926552 10.73399024   7.5134981   32.0465527
#> exo1      -0.04303397  0.59204004  -0.5757285    0.7325297
#> exo2      -0.11267948  0.56008741  -0.7164530    0.5763098
#> exo3      -0.14731049  0.44682336  -0.6570226    0.3125686
#> 
#> $AFG$A$equation3
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1   0.07786844  0.1035693  -0.03599872    0.2003970
#> lag1_var2   3.71435457  0.4240617   3.20236980    4.1849061
#> lag1_var3   9.75263847  0.7540586   8.90430134   10.6551384
#> lag1_var4  -7.71197680  0.5666598  -8.40383044   -7.0870588
#> const     -47.45528745 11.1516768 -60.53968826  -34.5922560
#> exo1        0.15634385  0.4601212  -0.25307428    0.7711905
#> exo2       -0.57181371  0.4746806  -1.22455617   -0.2306796
#> exo3        0.46254440  0.3882676  -0.04356187    0.8071998
#> 
#> $AFG$A$equation4
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1   0.1984592 0.09493381   0.08829157    0.3053165
#> lag1_var2   3.8197986 0.29477517   3.45120607    4.1229830
#> lag1_var3   9.0573291 0.56940709   8.39386918    9.7089590
#> lag1_var4  -6.9814378 0.45102456  -7.49986416   -6.4820092
#> const     -51.5394549 7.92328138 -60.35251221  -42.0268162
#> exo1        0.1569428 0.34336857  -0.11223629    0.6161708
#> exo2       -0.6156252 0.42917392  -1.12313122   -0.2578630
#> exo3        0.5416813 0.33703552   0.13187009    0.8479994
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.007751354 0.002036319 0.005810244   0.01032832
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.02211054 0.009564424 -0.03499101  -0.01408496
#> Sigma[2,2]  0.24015047 0.062047563  0.18568042   0.31959843
#> 
#> $AGO$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01238321 0.005713496  0.007953421   0.02003817
#> Sigma[3,2] -0.09228334 0.024027080 -0.122194350  -0.06923499
#> Sigma[3,3]  0.05259676 0.012906891  0.039094729   0.06714901
#> 
#> $AGO$Sigma$equation4
#>                    mean          sd  5% quantile  95% quantile
#> Sigma[4,1] -0.005549468 0.005182178 -0.009936754  0.0013923405
#> Sigma[4,2]  0.110324417 0.032737923  0.091110866  0.1555393633
#> Sigma[4,3] -0.021790698 0.017640384 -0.040350655 -0.0005227573
#> Sigma[4,4]  0.076510993 0.018790489  0.058589160  0.0997272498
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.01939500 0.02248864  0.98888823   1.03111687
#> lag1_var2 -0.33417473 0.05727179 -0.37703312  -0.25583890
#> lag1_var3 -0.43003868 0.07627202 -0.50267722  -0.33142750
#> lag1_var4  0.39558891 0.04914237  0.32832015   0.43143813
#> const      2.27525488 3.34877004 -1.02285822   6.20318131
#> exo1       0.07637919 0.11173468 -0.05735388   0.19384012
#> exo2      -0.07722368 0.11344823 -0.19630264   0.06413806
#> exo3       0.09357423 0.14248396 -0.05348187   0.25401079
#> 
#> $AGO$A$equation2
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  -0.40923956  0.1793578  -0.5817465   -0.1747748
#> lag1_var2   4.29096522  0.8692314   3.4040715    5.3863370
#> lag1_var3   4.64446723  1.0482881   3.6730163    6.0264714
#> lag1_var4  -4.07891288  0.8160246  -5.0824732   -3.2327200
#> const     -28.90505095 26.7218949 -65.7194373  -11.0621605
#> exo1       -0.05878095  0.3241175  -0.2828991    0.3819591
#> exo2        0.50958720  0.7235779  -0.3997019    1.2353586
#> exo3       -1.52391835  0.4718453  -2.0830135   -0.9848024
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.06638035 0.05401786 -0.12251541 -0.008188493
#> lag1_var2 -0.75548382 0.28976001 -1.14094166 -0.496558980
#> lag1_var3 -0.02672073 0.37019239 -0.51466465  0.319549946
#> lag1_var4  0.93302189 0.26447454  0.71359486  1.289998138
#> const      8.24662276 9.78379603 -2.72134131 19.891148725
#> exo1       0.14986249 0.11870292  0.04250812  0.304209569
#> exo2      -1.50780722 0.35736726 -1.90726728 -1.083360811
#> exo3       3.60430174 0.14633973  3.46122053  3.789409513
#> 
#> $AGO$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  -0.4994687  0.1104833  -0.5833288   -0.3500888
#> lag1_var2   2.7866892  0.5525396   2.1522933    3.3886688
#> lag1_var3   3.9176941  0.6558610   3.2226498    4.6915991
#> lag1_var4  -2.4067877  0.5261356  -2.9557103   -1.8028399
#> const     -23.4100808 16.1020528 -45.5463311  -12.2613215
#> exo1        0.1015752  0.2036389  -0.0479421    0.3813821
#> exo2       -1.4059508  0.3796159  -1.7823942   -0.9716855
#> exo3        2.8000335  0.4256694   2.2385930    3.1894986
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.004119229 0.002567478 0.002129841  0.007626428
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.06104157 0.03183114 -0.09865043  -0.02616853
#> Sigma[2,2]  7.05050371 1.74977856  5.65436396   9.38384344
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.03253504 0.05359612 -0.01478315     0.103537
#> Sigma[3,2] -4.85518547 0.86681883 -5.75840575    -3.937317
#> Sigma[3,3]  5.75277747 2.25666845  4.36417050     8.811458
#> 
#> $ALB$Sigma$equation4
#>                   mean        sd 5% quantile 95% quantile
#> Sigma[4,1] -0.01083531 0.0388855 -0.03982691   0.04196076
#> Sigma[4,2] -0.10993833 1.3569528 -1.92266640   0.98768041
#> Sigma[4,3]  2.94294857 2.3647121  1.16208569   6.13660052
#> Sigma[4,4]  3.43277261 1.7656562  1.92592729   5.79333591
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.99083700 0.04924022  0.925336202   1.03176717
#> lag1_var2 -0.02483355 0.14128105 -0.215951505   0.07665332
#> lag1_var3 -0.04541044 0.21322814 -0.334335652   0.10543028
#> lag1_var4  0.03960731 0.17610550 -0.083782668   0.27857477
#> const      0.54760485 3.55250113 -2.195679075   5.31743468
#> exo1       0.01412790 0.02349022 -0.005309099   0.04494674
#> exo2      -0.04476913 0.06628223 -0.120662187   0.02416891
#> exo3       0.09462313 0.07224001  0.011723152   0.17920453
#> 
#> $ALB$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -4.5004251  2.107308   -7.210209    -2.428930
#> lag1_var2  3.2337326  1.937727    1.878863     5.897171
#> lag1_var3  4.3940655  3.098325    2.160004     8.618614
#> lag1_var4 -3.6922836  2.667543   -7.266859    -1.643424
#> const     67.5817337 57.978345   17.490921   143.020672
#> exo1      -3.8897674  2.789385   -7.351494    -1.267012
#> exo2      -0.3264611  5.149953   -6.563639     4.662470
#> exo3       0.6090748  3.647176   -2.639421     4.543040
#> 
#> $ALB$A$equation3
#>                mean         sd 5% quantile 95% quantile
#> lag1_var1  1.733640   3.322356  -2.6150593    4.8179446
#> lag1_var2 -3.208786   2.688463  -6.1541269   -0.5051048
#> lag1_var3 -4.284439   4.011143  -8.8115896   -0.1126070
#> lag1_var4  4.301691   3.252070   0.8573106    8.0803914
#> const     19.464298 117.728353 -88.0765738  174.9316455
#> exo1       1.638425   1.727862  -0.0870706    3.6305047
#> exo2      -1.583359   3.821913  -5.6753853    2.8011356
#> exo3      -1.966373   1.725346  -4.2107763   -0.3723769
#> 
#> $ALB$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -1.4743541   2.517466  -4.9345488    0.1323512
#> lag1_var2 -1.6706944   2.765288  -5.2462361    0.9962795
#> lag1_var3 -2.2180602   4.012228  -7.3172569    1.8137779
#> lag1_var4  2.7097438   3.209844  -0.5816397    6.7329709
#> const     69.3837746 108.772330  -2.6919422  218.0684758
#> exo1      -0.6771556   1.876602  -2.4318324    1.7612671
#> exo2      -2.0874636   2.201220  -4.3744776    0.6567999
#> exo3      -1.7618476   1.363858  -2.7261338    0.1148654
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd  5% quantile 95% quantile
#> Sigma[1,1] 0.001207737 0.0003567853 0.0008155805  0.001586703
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd  5% quantile  95% quantile
#> Sigma[2,1] -0.001908768 0.000882612 -0.002809555 -0.0009308326
#> Sigma[2,2]  0.046927595 0.008576915  0.036480567  0.0562076577
#> 
#> $ARE$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.002751999 0.001064422  0.001561518  0.003974165
#> Sigma[3,2] -0.044236821 0.013054671 -0.058280530 -0.031548150
#> Sigma[3,3]  0.143945469 0.041452803  0.107681632  0.197044881
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.001412089 0.001684829 -0.000627543  0.003012805
#> Sigma[4,2] -0.008639188 0.011913196 -0.021795862  0.004705947
#> Sigma[4,3]  0.113055147 0.033702257  0.081823208  0.156967838
#> Sigma[4,4]  0.109550054 0.027753431  0.087463421  0.146614089
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.96052530 0.04148000  0.90582463  0.995584605
#> lag1_var2  0.43380226 0.02778713  0.39644085  0.455716856
#> lag1_var3  0.54037017 0.04115657  0.48380550  0.566398213
#> lag1_var4 -0.52573996 0.05386134 -0.56061004 -0.451368174
#> const     -0.08893997 0.19922428 -0.31738895  0.114470576
#> exo1      -0.02429497 0.02561464 -0.05332562  0.003309091
#> exo2      -0.05649565 0.02089478 -0.08136060 -0.033622065
#> exo3      -0.01227961 0.06233527 -0.07972254  0.050826304
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.17939645 0.2508802  -0.1216955   0.46169490
#> lag1_var2  0.51681131 0.1403006   0.3712005   0.67434936
#> lag1_var3 -0.28729282 0.1910820  -0.4959253  -0.05875503
#> lag1_var4  0.20135020 0.2595104  -0.1213508   0.47623859
#> const      2.48028640 2.5205249  -0.5107976   5.28570327
#> exo1      -0.07786185 0.2908012  -0.3631780   0.29553247
#> exo2       2.08676039 0.3017816   1.7380251   2.43573934
#> exo3      -0.39582326 0.4565788  -0.7863616   0.20383931
#> 
#> $ARE$A$equation3
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1  0.23343848 0.2916976  -0.02386175    0.6104123
#> lag1_var2  0.09042576 0.4982579  -0.59512329    0.4500804
#> lag1_var3  0.84281811 0.5530995   0.08204785    1.2139250
#> lag1_var4  0.15317106 0.5811286  -0.27043672    0.9360876
#> const     -6.12787485 5.7505778 -13.80695712   -1.4962742
#> exo1      -0.52779423 0.3070194  -0.81700666   -0.1508533
#> exo2      -2.64189916 0.4094413  -3.02536718   -2.1068615
#> exo3      -1.52402726 0.3186982  -1.89555513   -1.2001681
#> 
#> $ARE$A$equation4
#>                 mean        sd  5% quantile 95% quantile
#> lag1_var1  0.3493452 0.1313521   0.19345453   0.49563949
#> lag1_var2  0.4919389 0.4582451  -0.11834553   0.87096857
#> lag1_var3  0.6195061 0.5718805  -0.13044567   1.11779311
#> lag1_var4  0.3380031 0.5687502  -0.08176193   1.11345199
#> const     -5.7730566 5.6657012 -12.98572201   0.02817017
#> exo1      -0.6070130 0.3140765  -0.94339730  -0.23856896
#> exo2      -1.0238784 0.3302970  -1.34339798  -0.59651652
#> exo3      -1.8826656 0.1889391  -2.13385067  -1.72174660
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002614053 0.0006702219 0.001771077  0.003317788
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05944027 0.01468313 -0.07581799  -0.04381163
#> Sigma[2,2]  3.46463850 0.87251766  2.32174586   4.27913398
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04678779 0.01257505  0.03531814   0.06310222
#> Sigma[3,2] -2.43730295 0.65871784 -3.02107870  -1.55816885
#> Sigma[3,3]  2.00791929 0.46814819  1.37279215   2.33891513
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01223835 0.005532831  0.007635928   0.01961863
#> Sigma[4,2] -0.36548848 0.159862215 -0.523452269  -0.16683756
#> Sigma[4,3]  0.58204910 0.150310759  0.422880979   0.76965120
#> Sigma[4,4]  0.39956392 0.181198753  0.293053866   0.64604073
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.885735185 0.04529328  0.84523581   0.94408009
#> lag1_var2  0.289213750 0.10477126  0.15469994   0.39133409
#> lag1_var3  0.474663408 0.17184218  0.25503097   0.64371062
#> lag1_var4 -0.434202930 0.15274986 -0.58713587  -0.24290614
#> const      0.639473126 2.22689255 -1.71904750   3.41883735
#> exo1      -0.002531934 0.04224565 -0.05932702   0.03290434
#> exo2      -0.157114312 0.01345968 -0.16764082  -0.13876093
#> exo3       0.014170769 0.07207051 -0.07448974   0.09193902
#> 
#> $ARG$A$equation2
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  -2.603892  2.017436  -4.2927666  -0.03625591
#> lag1_var2  -7.661171  4.048313 -12.4256245  -3.11558934
#> lag1_var3 -13.593781  6.645453 -21.3805448  -6.10144795
#> lag1_var4  12.716117  5.839664   6.1203954  19.51473781
#> const     126.947919 80.458945  48.9390363 222.99668337
#> exo1       -0.411809  1.230403  -1.3003225   1.27872906
#> exo2        3.488452  1.515746   1.7031369   5.18032406
#> exo3        1.226069  2.535909  -0.8701377   4.59482114
#> 
#> $ARG$A$equation3
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1   0.6701976  1.6817555   -1.5259250    2.2671767
#> lag1_var2   6.1230181  2.6878807    3.0865908    9.2623872
#> lag1_var3  10.6524719  4.4367412    5.6442907   15.8265165
#> lag1_var4  -9.2856981  3.8132367  -13.6874454   -4.9238202
#> const     -42.4819026 64.3374028 -118.8629481   19.3498659
#> exo1       -0.3460936  0.9673909   -1.6345407    0.4661629
#> exo2       -6.0977341  1.0970394   -7.3012452   -4.8497422
#> exo3        1.5199050  1.1630653    0.1726338    2.8667471
#> 
#> $ARG$A$equation4
#>                mean         sd 5% quantile 95% quantile
#> lag1_var1 -1.062782  0.6177181  -1.8643940   -0.4642729
#> lag1_var2  1.422132  0.6895979   0.6931038    2.2098581
#> lag1_var3  2.311100  1.1023957   1.1511737    3.5445998
#> lag1_var4 -1.420104  1.1019904  -2.7480972   -0.2548240
#> const     35.202816 24.9227152  13.8041396   67.2647093
#> exo1      -0.666429  0.5866796  -1.1643799    0.1009745
#> exo2      -4.489042  0.5384078  -4.8764204   -3.7485795
#> exo3       2.529604  1.0316090   1.2508300    3.5218826
#> 
#> 
#>