Provides posterior estimation summary for Bayesian Vector Autoregressions for dynamic panel data

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARs'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARs obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.007826464 0.001950701  0.00555913  0.009958457
#> 
#> $AFG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01140803 0.002421218 -0.01426395 -0.008717175
#> Sigma[2,2]  0.16631261 0.036356904  0.12105996  0.204172148
#> 
#> $AFG$Sigma$equation3
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.02430236 0.007451906  0.01465927   0.03155862
#> Sigma[3,2] -0.15204413 0.013550582 -0.17012403  -0.14036133
#> Sigma[3,3]  0.30225402 0.076963315  0.21147880   0.39073943
#> 
#> $AFG$Sigma$equation4
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[4,1]  0.02242399 0.008143883  0.01178318   0.02976392
#> Sigma[4,2] -0.08976369 0.016497999 -0.10630333  -0.06910438
#> Sigma[4,3]  0.26865980 0.089039870  0.16677071   0.37270753
#> Sigma[4,4]  0.26343843 0.096082804  0.15563000   0.37632261
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.967238725 0.02257450  0.94157002   0.98972661
#> lag1_var2 -0.369107358 0.08053591 -0.47207020  -0.29887062
#> lag1_var3 -0.796979583 0.18650584 -1.02906525  -0.63707006
#> lag1_var4  0.693725682 0.17057007  0.54749749   0.90468736
#> const      5.611844081 0.95055690  4.47391667   6.43015993
#> exo1      -0.003342895 0.07750909 -0.06386075   0.09172428
#> exo2      -0.111042057 0.09330013 -0.21926775  -0.01110276
#> exo3      -0.303621373 0.04743262 -0.34999982  -0.24285807
#> 
#> $AFG$A$equation2
#>                   mean        sd 5% quantile 95% quantile
#> lag1_var1  0.382911722 0.1402741   0.1964507    0.5040156
#> lag1_var2  0.769366499 0.2655598   0.5077962    1.1115036
#> lag1_var3 -0.527292590 0.6385868  -1.1707503    0.2704308
#> lag1_var4  0.462459183 0.5832196  -0.2364857    1.0570820
#> const     -5.763125253 4.1821506 -10.9191166   -1.2817200
#> exo1       0.007670718 0.3784591  -0.3763279    0.4455114
#> exo2       0.201656640 0.3344907  -0.1116624    0.6000666
#> exo3      -0.269196331 0.3633548  -0.5940573    0.1947002
#> 
#> $AFG$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.18308650 0.1791644  -0.3818228 -0.006863039
#> lag1_var2  0.16717337 0.3347492  -0.1690147  0.556111368
#> lag1_var3  2.21438760 0.6180796   1.4550043  2.852219384
#> lag1_var4 -1.23692149 0.5485142  -1.7766840 -0.526792668
#> const      8.36506002 8.9453908   3.0443055 20.756408650
#> exo1      -0.23845274 0.3521498  -0.6862472  0.081080226
#> exo2      -1.37279123 0.3690717  -1.8250564 -1.006569190
#> exo3      -0.07336904 0.3188360  -0.3754631  0.306275736
#> 
#> $AFG$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.01605104  0.1507892  -0.1896500   0.09547124
#> lag1_var2  0.60715763  0.3680061   0.1669233   1.00488822
#> lag1_var3  2.24393752  0.5579803   1.6881225   2.84773322
#> lag1_var4 -1.13927548  0.4341494  -1.6236469  -0.67115997
#> const     -1.03684421 10.3677108  -7.9552300  12.98382135
#> exo1      -0.24645207  0.3632272  -0.6691136   0.12586257
#> exo2      -1.34681470  0.3281104  -1.7569854  -1.06683192
#> exo3      -0.08685275  0.2971352  -0.3886441   0.26326169
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.006933315 0.002568847 0.004433669   0.01026444
#> 
#> $AGO$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.00238724 0.01006748 -0.01184456  0.009473294
#> Sigma[2,2]  0.17602090 0.07637130  0.08454170  0.255063317
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.004439884 0.004141618 -0.000659314  0.008823815
#> Sigma[3,2] -0.079544124 0.037720521 -0.118247500 -0.035727396
#> Sigma[3,3]  0.049979283 0.019058955  0.027108083  0.069169494
#> 
#> $AGO$Sigma$equation4
#>                    mean          sd   5% quantile  95% quantile
#> Sigma[4,1]  0.003299308 0.005205241 -0.0007235727  0.0090932935
#> Sigma[4,2]  0.069074092 0.028466042  0.0358187318  0.0968569272
#> Sigma[4,3] -0.015279278 0.013285441 -0.0280613581 -0.0004263831
#> Sigma[4,4]  0.047382908 0.013521264  0.0344045382  0.0627673802
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.95881338 0.03633792  0.91564512  0.995129261
#> lag1_var2  0.03440694 0.22484656 -0.20695905  0.272118173
#> lag1_var3  0.09151569 0.30904923 -0.24925591  0.420970932
#> lag1_var4 -0.07037937 0.27636473 -0.36930561  0.245791357
#> const      0.02425910 3.45505964 -4.22259930  3.745542668
#> exo1      -0.02760383 0.06628868 -0.07955450  0.062194591
#> exo2      -0.01755201 0.02558436 -0.05185088  0.002638484
#> exo3      -0.01317947 0.14383014 -0.18101344  0.146429873
#> 
#> $AGO$A$equation2
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  -0.2968434  0.1649613  -0.47160359  -0.09180873
#> lag1_var2   2.4931422  1.5217474   0.54774969   3.87004430
#> lag1_var3   2.2014479  2.1695918  -0.53494863   4.19697648
#> lag1_var4  -1.8378116  1.8599932  -3.56858305   0.52821847
#> const     -17.4659626 19.1668881 -37.31651088   3.32701335
#> exo1        0.3057490  0.3552902   0.08931313   0.79339736
#> exo2        0.5868638  0.2605732   0.34767452   0.87075361
#> exo3       -0.2678946  0.5990674  -1.04861450   0.30290494
#> 
#> $AGO$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.04117155 0.0883887  -0.1375914   0.06493116
#> lag1_var2 -1.26015505 0.5713726  -1.9241991  -0.68653841
#> lag1_var3 -0.79439375 0.8427902  -1.6743871   0.10313684
#> lag1_var4  1.49881488 0.7182289   0.7324711   2.26637772
#> const     21.73039172 7.5960773  14.6371440  30.38149248
#> exo1      -0.07020658 0.2503913  -0.3621541   0.21269456
#> exo2      -1.45777733 0.2073855  -1.7304425  -1.28546676
#> exo3       3.16549006 0.3027403   2.9267280   3.54568597
#> 
#> $AGO$A$equation4
#>                 mean          sd 5% quantile 95% quantile
#> lag1_var1 -0.3663042  0.07217679 -0.43242148   -0.2754455
#> lag1_var2  0.4312194  0.87798492 -0.76875341    1.1057330
#> lag1_var3  0.6429981  1.17470070 -0.94538151    1.5634283
#> lag1_var4  0.4371747  1.02885311 -0.39439564    1.8302608
#> const      3.7581301 10.19884024 -6.54039598   16.3235054
#> exo1       0.1855625  0.15698490  0.01554784    0.3608536
#> exo2      -1.2679312  0.09863127 -1.37089134   -1.1636214
#> exo3       3.5091798  0.35797463  3.13815634    3.8967680
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                  mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00178198 0.0004336648 0.001354874  0.002312035
#> 
#> $ALB$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0375389 0.01940899 -0.05890713  -0.01721455
#> Sigma[2,2]  7.3510812 3.39618300  5.36899520  11.95145294
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd   5% quantile 95% quantile
#> Sigma[3,1]  0.02017223 0.01689617  4.654888e-04   0.03982992
#> Sigma[3,2] -5.66300673 3.54381846 -1.052963e+01  -3.19133578
#> Sigma[3,3]  7.10118254 3.44211935  3.784041e+00  11.59230917
#> 
#> $ALB$Sigma$equation4
#>                    mean          sd 5% quantile 95% quantile
#> Sigma[4,1] -0.005899376 0.007505411  -0.0150494  0.001111579
#> Sigma[4,2] -0.817042534 1.727246051  -2.7880354  1.079528076
#> Sigma[4,3]  3.888090355 1.473619500   1.9521045  5.218118524
#> Sigma[4,4]  3.985503041 0.969231130   2.9569557  5.170331632
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  1.01975888 0.02788535  0.991643021  1.054950685
#> lag1_var2  0.02695867 0.02516455 -0.002376681  0.055886666
#> lag1_var3  0.03412431 0.03883741 -0.010581744  0.079680776
#> lag1_var4 -0.02844905 0.03257773 -0.068074825  0.007552394
#> const     -0.85090882 0.99075557 -2.005700881  0.295775721
#> exo1       0.01765515 0.04421055 -0.031984218  0.064096182
#> exo2      -0.04253327 0.07269631 -0.119877677  0.035638656
#> exo3       0.05481946 0.03319172  0.019469304  0.094520448
#> 
#> $ALB$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -4.60968993  2.211760   -7.602615    -3.312442
#> lag1_var2  4.34773437  1.782369    2.255269     6.348470
#> lag1_var3  6.08192232  2.539638    3.183295     8.958018
#> lag1_var4 -5.11264399  2.134089   -7.605265    -2.758203
#> const     51.99166303 77.782222    6.077545   157.627157
#> exo1      -1.34484323  2.270256   -3.732941     1.326208
#> exo2      -0.07211438  3.884343   -4.577220     4.116597
#> exo3       0.37590388  1.399446   -1.456794     1.635840
#> 
#> $ALB$A$equation3
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1   2.7145302  1.877204    1.3284876   5.24827077
#> lag1_var2  -2.1817616  1.923691   -4.3404465  -0.31153912
#> lag1_var3  -2.6851734  2.701024   -5.7385948   0.04439961
#> lag1_var4   3.0084819  2.297717    0.6861217   5.62559917
#> const     -22.1097096 66.522883 -107.2905330  36.37214810
#> exo1       -0.7587957  1.495345   -1.9319619   0.93033590
#> exo2       -2.6834319  1.175366   -3.6664611  -1.10968101
#> exo3       -1.1341142  1.978535   -2.6733334   1.40203477
#> 
#> $ALB$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.4102569  0.9953704   -1.186131   0.92645589
#> lag1_var2  0.5416440  1.6375025   -1.558935   2.02608993
#> lag1_var3  1.2003431  2.2989209   -1.775214   3.16062182
#> lag1_var4 -0.1050683  1.9450367   -1.695692   2.42823427
#> const      6.2894284 45.5733437  -50.046122  53.23249268
#> exo1      -1.5696780  1.8198835   -3.840141   0.34832861
#> exo2      -3.0564126  2.5454893   -5.792947   0.03993739
#> exo3      -0.9993667  1.9068954   -2.751217   1.19863033
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001346015 0.0002298646 0.001079176  0.001575562
#> 
#> $ARE$Sigma$equation2
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[2,1] -0.002560957 0.001526074 -0.004596905 -0.001301761
#> Sigma[2,2]  0.056867192 0.003229389  0.052484662  0.059301009
#> 
#> $ARE$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.00585016 0.002446487  0.002905004  0.007995091
#> Sigma[3,2] -0.06414392 0.015256351 -0.084514088 -0.051578564
#> Sigma[3,3]  0.17693166 0.046539507  0.141137334  0.240339036
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.004254373 0.002062601  0.002117359  0.006473894
#> Sigma[4,2] -0.021318414 0.014275001 -0.040098212 -0.008400634
#> Sigma[4,3]  0.133740797 0.036378170  0.103914056  0.182527305
#> Sigma[4,4]  0.124584124 0.027698649  0.098872836  0.160018935
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.93003685 0.04505487  0.87117154   0.96956186
#> lag1_var2  0.09560095 0.01928881  0.07461350   0.11335469
#> lag1_var3  0.09641162 0.03085445  0.06366784   0.13188754
#> lag1_var4 -0.08220785 0.03797388 -0.12022851  -0.03562652
#> const      0.74419151 0.29050583  0.35592290   0.97242116
#> exo1      -0.02180173 0.01228909 -0.03880243  -0.01448101
#> exo2      -0.13942743 0.01696771 -0.16251082  -0.12784865
#> exo3      -0.03948980 0.05543620 -0.11460981   0.00358408
#> 
#> $ARE$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.06238258 0.22339572 -0.16723141    0.3466258
#> lag1_var2  0.62761382 0.04820554  0.57364637    0.6847771
#> lag1_var3 -0.21319301 0.07187655 -0.29333934   -0.1375986
#> lag1_var4  0.18968230 0.09489537  0.08555928    0.2766726
#> const      0.62274603 2.13219676 -1.27245296    3.4383108
#> exo1       0.12900248 0.15700895 -0.03393158    0.2942044
#> exo2       1.82611339 0.23039326  1.62225646    2.1323627
#> exo3      -0.79825166 0.22349921 -1.07113222   -0.5729368
#> 
#> $ARE$A$equation3
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1  0.36087744 0.3765900   0.03733121   0.86594112
#> lag1_var2 -0.07370509 0.1245323  -0.20026284   0.06688692
#> lag1_var3  0.57934527 0.1937147   0.37116041   0.78029018
#> lag1_var4  0.35738508 0.2560596   0.09966992   0.66388250
#> const     -4.94960408 4.7985208 -10.44630097   0.50254677
#> exo1      -0.64971217 0.5563187  -1.13299255   0.05047032
#> exo2      -2.57493117 0.1600440  -2.72663476  -2.37841529
#> exo3      -1.52213231 0.8772591  -2.59496049  -0.66175698
#> 
#> $ARE$A$equation4
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1  0.3737527 0.2319194   0.1643877    0.6783040
#> lag1_var2  0.4172298 0.1102068   0.2916746    0.5322991
#> lag1_var3  0.4314877 0.2067593   0.2098858    0.6601949
#> lag1_var4  0.5253306 0.2627822   0.2804594    0.8186553
#> const     -6.4623363 3.7605454  -9.9807064   -1.7597914
#> exo1      -0.5330496 0.4834672  -0.9106109    0.1089504
#> exo2      -1.1744044 0.2181211  -1.4695083   -1.0020117
#> exo3      -2.1765489 0.9349002  -3.3408323   -1.3011282
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002894879 0.0008187749 0.002046574  0.003790796
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.07560903 0.02158382  -0.1004572  -0.05319877
#> Sigma[2,2]  3.67715047 0.75968202   2.7545984   4.48109570
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.05945822 0.01757685  0.03954859   0.07562798
#> Sigma[3,2] -2.80670833 0.65607753 -3.41739205  -2.07604930
#> Sigma[3,3]  2.44166751 0.69351012  1.73803113   3.25741203
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01561478 0.006853676  0.007450558   0.02285475
#> Sigma[4,2] -0.63573492 0.283072025 -0.970156232  -0.31401178
#> Sigma[4,3]  0.81759470 0.381951007  0.482786284   1.32804196
#> Sigma[4,4]  0.48145911 0.242748693  0.322568494   0.81671477
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.91022282 0.03566563  0.869847380  0.952447199
#> lag1_var2  0.15217723 0.02656524  0.119379706  0.180631873
#> lag1_var3  0.25518945 0.04601974  0.200260100  0.306886008
#> lag1_var4 -0.23237525 0.03462718 -0.270167459 -0.189881447
#> const      1.09070690 1.33565681  0.121215711  2.921451650
#> exo1      -0.04148092 0.03797932 -0.079826769  0.006146129
#> exo2      -0.13308682 0.05506309 -0.187446743 -0.071028228
#> exo3       0.03504423 0.03442102  0.006632868  0.074862248
#> 
#> $ARG$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -2.72924117  1.5243184  -4.6356620   -1.2559063
#> lag1_var2  1.58686206  1.5109721  -0.1727937    3.3500926
#> lag1_var3  1.38861061  2.6073348  -1.6630340    4.4068243
#> lag1_var4 -0.88440366  2.2382802  -3.3670650    1.7818438
#> const     45.25380080 50.4433975  15.5426413  115.1465774
#> exo1      -0.41810210  0.6373269  -1.2240000    0.2409605
#> exo2       3.09283350  1.4527494   1.9604854    5.0355301
#> exo3      -0.03655396  2.1558781  -1.9400931    2.8350310
#> 
#> $ARG$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  1.32450746  1.0011158   0.2557697   2.41513582
#> lag1_var2 -0.44006141  1.0022029  -1.3329807   0.82092253
#> lag1_var3  0.03980494  1.7980579  -1.5830253   2.30652408
#> lag1_var4  0.44899317  1.3057355  -1.2850889   1.52690397
#> const     -5.79321089 39.6724472 -58.7969562  28.81798944
#> exo1      -0.44258960  0.4609739  -0.9864820   0.06382091
#> exo2      -5.70225343  0.9184125  -6.6924504  -4.79556992
#> exo3       2.62472385  2.2369877  -0.4377712   3.90497321
#> 
#> $ARG$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.36245923  0.3112879  -0.6709234   0.03739205
#> lag1_var2  0.51172442  0.8944028  -0.5581096   1.51703315
#> lag1_var3  0.86295520  1.4905231  -0.9204920   2.53417579
#> lag1_var4  0.01635364  1.2400672  -1.4365375   1.37861920
#> const     17.39056668 13.6887882   5.0345144  35.15204475
#> exo1      -0.80201216  0.3201501  -1.1567413  -0.41248931
#> exo2      -4.40173054  0.4425837  -4.8855560  -4.02075516
#> exo3       2.98636945  1.2368906   1.3490678   4.07878943
#> 
#> 
#>