Provides posterior estimation summary for Bayesian Vector Autoregressions for dynamic panel data

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARs'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARs obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003245567 0.0006264247 0.002803943  0.004104295
#> 
#> $AFG$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01194226 0.005945872 -0.01801643 -0.005585819
#> Sigma[2,2]  0.26391785 0.137538701  0.19010993  0.451188817
#> 
#> $AFG$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01234807 0.007632504  0.004068347   0.02107591
#> Sigma[3,2] -0.24554476 0.108918839 -0.395135002  -0.16588411
#> Sigma[3,3]  0.37696644 0.093142840  0.298258004   0.48474271
#> 
#> $AFG$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.007802911 0.005687644  0.002075016   0.01431675
#> Sigma[4,2] -0.147002603 0.061112574 -0.222025200  -0.08796643
#> Sigma[4,3]  0.305569452 0.068847605  0.255996485   0.39850316
#> Sigma[4,4]  0.276466732 0.062539729  0.240725472   0.36201032
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.988950967 0.01165112  0.97382251   1.00031188
#> lag1_var2  0.046802626 0.12275083 -0.08353371   0.17077160
#> lag1_var3  0.124782717 0.26308092 -0.14910956   0.40119704
#> lag1_var4 -0.109199196 0.22470186 -0.34939366   0.12063623
#> const     -0.320820428 1.72907892 -1.98925858   1.75844292
#> exo1       0.006537488 0.04739341 -0.05635964   0.04874194
#> exo2       0.002687705 0.05041338 -0.05643456   0.05760468
#> exo3      -0.161516463 0.07152542 -0.23858218  -0.07298406
#> 
#> $AFG$A$equation2
#>                   mean        sd 5% quantile 95% quantile
#> lag1_var1   0.06647351 0.1581803  -0.1043050    0.2329177
#> lag1_var2   1.63087448 0.2721468   1.3351190    1.8936788
#> lag1_var3   1.24270224 0.5665684   0.5751131    1.7930632
#> lag1_var4  -1.02823093 0.5188612  -1.6037443   -0.4269201
#> const     -11.82430839 7.5849003 -20.9207406   -3.7643074
#> exo1        0.05247164 0.5060039  -0.4650636    0.6275327
#> exo2        0.17170181 0.3243589  -0.2384318    0.5016155
#> exo3        0.55045042 0.5883252  -0.1022878    1.1803710
#> 
#> $AFG$A$equation3
#>                   mean        sd  5% quantile 95% quantile
#> lag1_var1  0.151627914 0.1956168  0.004653317    0.4192648
#> lag1_var2  0.004834723 0.4810687 -0.520379146    0.5930668
#> lag1_var3  1.885076896 0.8463288  1.129268042    2.9990238
#> lag1_var4 -0.957886761 0.6896614 -1.885675571   -0.3677173
#> const      3.116018547 8.7108586 -6.673058785   10.9039705
#> exo1      -0.323020680 0.4957875 -0.960338084    0.1705326
#> exo2      -0.962610495 0.4843223 -1.509320746   -0.4489821
#> exo3      -0.571441695 0.7254642 -1.374652854    0.2732901
#> 
#> $AFG$A$equation4
#>                  mean        sd  5% quantile 95% quantile
#> lag1_var1   0.2084365 0.2106649   0.01677271   0.48369368
#> lag1_var2   0.8728891 0.4907242   0.32093481   1.44741285
#> lag1_var3   2.8021164 0.9564590   1.89853496   3.98445105
#> lag1_var4  -1.6089627 0.8011345  -2.60669671  -0.86533078
#> const     -10.3360315 8.1243727 -20.02100854  -1.39855602
#> exo1       -0.3342635 0.4257194  -0.89763624   0.03014322
#> exo2       -0.8619207 0.3841898  -1.27550097  -0.48954894
#> exo3       -0.2301808 0.5195480  -0.76924734   0.42365188
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.008916403 0.001255094  0.00734821   0.01022891
#> 
#> $AGO$Sigma$equation2
#>                  mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.0174404 0.01034429 -0.02813932 -0.006463831
#> Sigma[2,2]  0.1557171 0.05469144  0.09003925  0.210072500
#> 
#> $AGO$Sigma$equation3
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[3,1]  0.01282448 0.004409107  0.00909879   0.01776232
#> Sigma[3,2] -0.06831361 0.023000866 -0.09535209  -0.04348932
#> Sigma[3,3]  0.04302582 0.011072200  0.03215645   0.05709254
#> 
#> $AGO$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1] -0.002299896 0.005145869 -0.007104291  0.003735921
#> Sigma[4,2]  0.063951585 0.026595946  0.033069538  0.094310574
#> Sigma[4,3] -0.013762145 0.008058417 -0.022144070 -0.004412974
#> Sigma[4,4]  0.044081836 0.016536616  0.025781736  0.063324564
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.99782912 0.04885932  0.93964138   1.04903335
#> lag1_var2 -0.45465851 0.20160342 -0.69692953  -0.22756668
#> lag1_var3 -0.58887577 0.25437253 -0.88695166  -0.29371155
#> lag1_var4  0.50796633 0.22219884  0.25538673   0.77185270
#> const      6.35139732 1.99619183  4.01485059   8.42913379
#> exo1       0.08534582 0.07664895  0.02630044   0.18885390
#> exo2      -0.05686673 0.09017296 -0.16393591   0.04446462
#> exo3       0.02540271 0.09475955 -0.06610288   0.14781293
#> 
#> $AGO$A$equation2
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1 -0.17380095  0.2004375  -0.42355389    0.0352376
#> lag1_var2  1.27915933  0.8627880   0.59500078    2.4551632
#> lag1_var3  0.52036708  1.1881072  -0.29864608    2.1597877
#> lag1_var4 -0.43901988  0.9833765  -1.79164782    0.2609394
#> const     -0.06574355 11.1047135 -14.57404177   10.2308370
#> exo1       0.09354661  0.2808374  -0.20572072    0.4026820
#> exo2       0.51430132  0.4700448  -0.06761514    0.9654650
#> exo3      -0.99015239  0.2592986  -1.33493879   -0.7671317
#> 
#> $AGO$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.11800685 0.09890464 -0.23335837 -0.006690343
#> lag1_var2 -0.77974615 0.49131944 -1.37304151 -0.263788279
#> lag1_var3 -0.06308396 0.62116284 -0.83747286  0.565994952
#> lag1_var4  0.90597864 0.52025156  0.38297235  1.555537978
#> const     14.31205270 6.17193549  6.76734932 21.108859630
#> exo1       0.10537331 0.14667937 -0.04528533  0.291326329
#> exo2      -1.48877692 0.24993224 -1.82509134 -1.274224174
#> exo3       3.51881122 0.03989038  3.46816689  3.557407672
#> 
#> $AGO$A$equation4
#>                 mean         sd  5% quantile 95% quantile
#> lag1_var1 -0.3208185 0.08868205 -0.439277789   -0.2440922
#> lag1_var2  0.1203534 0.28250979 -0.146135058    0.4910299
#> lag1_var3  0.2767309 0.42794797 -0.097925578    0.8388572
#> lag1_var4  0.7796006 0.35384472  0.313278524    1.1051638
#> const      4.9497855 3.63580149  0.417857550    8.5442514
#> exo1       0.1797018 0.12845737  0.009596183    0.2886333
#> exo2      -1.3549450 0.28238283 -1.692821656   -1.0370260
#> exo3       3.3273721 0.21364347  3.041406336    3.5122849
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002397075 0.0003491041  0.00201599  0.002773722
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05810095 0.01150115  -0.0710518  -0.04584263
#> Sigma[2,2]  6.48963852 0.59118512   5.9106679   7.17968061
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd  5% quantile 95% quantile
#> Sigma[3,1]  0.03695119 0.02946456  0.003432226   0.07134224
#> Sigma[3,2] -4.43158110 0.70146314 -5.344116210  -3.79322165
#> Sigma[3,3]  5.66436554 1.17485083  4.307785801   6.97497666
#> 
#> $ALB$Sigma$equation4
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.002636795 0.02670741 -0.03425095   0.02792401
#> Sigma[4,2] -0.083866804 0.59166364 -0.83267204   0.51118021
#> Sigma[4,3]  3.168192650 0.84913998  2.07829453   3.95669843
#> Sigma[4,4]  3.687501620 0.95682145  2.58703164   4.81066227
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.99926984 0.02714566  0.96782681   1.03110739
#> lag1_var2 -0.01022651 0.03544501 -0.05474067   0.02451469
#> lag1_var3 -0.02291211 0.05254297 -0.08684122   0.03055164
#> lag1_var4  0.01863357 0.04320021 -0.02557248   0.07032956
#> const      0.25884075 1.10125422 -0.69602532   1.72387301
#> exo1      -0.01004723 0.02918540 -0.04949740   0.01000256
#> exo2      -0.07220255 0.01119213 -0.08251556  -0.05768366
#> exo3       0.07025148 0.02843455  0.03756703   0.09793737
#> 
#> $ALB$A$equation2
#>                mean        sd 5% quantile 95% quantile
#> lag1_var1 -3.061074  1.411830   -4.221470   -1.1594451
#> lag1_var2  3.165030  1.353134    1.862415    4.8585565
#> lag1_var3  4.216211  1.946019    2.243026    6.5968078
#> lag1_var4 -3.378658  1.545360   -5.274585   -1.8439714
#> const     25.636032 57.169322  -50.443324   77.2897469
#> exo1      -2.550065  1.689694   -4.146100   -0.5955118
#> exo2       2.261599  2.753630   -1.410836    4.4672195
#> exo3      -1.072726  1.413199   -2.854618    0.3626475
#> 
#> $ALB$A$equation3
#>                mean        sd 5% quantile 95% quantile
#> lag1_var1  1.343073  1.103490   0.2428579    2.5035686
#> lag1_var2 -1.742256  1.963079  -3.9404710    0.2155722
#> lag1_var3 -1.926372  2.939761  -5.3247643    1.0381871
#> lag1_var4  2.281964  2.451060  -0.1374674    5.1792569
#> const      7.531987 46.052086 -47.7202207   59.4643549
#> exo1      -0.659315  2.918275  -4.4735844    2.0717373
#> exo2      -3.896894  2.399935  -5.6402355   -0.6217000
#> exo3       1.230600  2.045668  -1.4260562    3.0262902
#> 
#> $ALB$A$equation4
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.82275701  0.6819993   -1.712310   -0.2678692
#> lag1_var2  0.06857424  1.5092292   -1.860164    1.5305243
#> lag1_var3  0.52726938  2.3056225   -2.448428    2.7062125
#> lag1_var4  0.48955792  1.9603073   -1.327217    3.0364956
#> const     21.57842372 20.0057306   -5.386932   37.9931822
#> exo1      -2.37096061  3.7041695   -6.296702    2.1437844
#> exo2      -2.60383240  1.6123751   -4.668222   -1.1071999
#> exo3       0.66608241  3.0652838   -2.562591    3.9947547
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.001148507 0.0002567569 0.000816234  0.001353116
#> 
#> $ARE$Sigma$equation2
#>                    mean           sd  5% quantile  95% quantile
#> Sigma[2,1] -0.001452457 0.0005706791 -0.001961505 -0.0007018805
#> Sigma[2,2]  0.061408219 0.0191273474  0.044507620  0.0864179453
#> 
#> $ARE$Sigma$equation3
#>                    mean           sd  5% quantile 95% quantile
#> Sigma[3,1]  0.003194954 0.0007054692  0.002249846  0.003780044
#> Sigma[3,2] -0.054082512 0.0218621727 -0.072678793 -0.025958878
#> Sigma[3,3]  0.174248547 0.0447492660  0.136686366  0.234194046
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd   5% quantile 95% quantile
#> Sigma[4,1]  0.002201205 0.001305588  0.0006805853  0.003724268
#> Sigma[4,2] -0.007461339 0.015518555 -0.0258996294  0.009969739
#> Sigma[4,3]  0.138371434 0.037359055  0.1112625970  0.188777533
#> Sigma[4,4]  0.141935490 0.033171043  0.1132039437  0.185060208
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.953599912 0.03168518  0.91827746   0.99099061
#> lag1_var2  0.033707613 0.04757068 -0.02896759   0.07517527
#> lag1_var3  0.016103839 0.06663980 -0.06978760   0.07846526
#> lag1_var4 -0.009200485 0.06347501 -0.06567977   0.07131949
#> const      0.683425092 0.35510114  0.26518053   1.04825722
#> exo1      -0.048382780 0.03594975 -0.09127876  -0.01411061
#> exo2      -0.106639856 0.02443120 -0.13746066  -0.08270904
#> exo3      -0.041569157 0.05875408 -0.08864243   0.03752639
#> 
#> $ARE$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.05202906 0.2832720 -0.28587944    0.2828788
#> lag1_var2  0.53932412 0.5741426 -0.08118554    1.1344918
#> lag1_var3 -0.31308058 0.6835899 -1.08882898    0.3596826
#> lag1_var4  0.27030100 0.6676450 -0.41449845    1.0715309
#> const      5.13594651 4.6977605 -0.67947217   10.1324626
#> exo1       0.03708008 0.1727598 -0.17333618    0.2182430
#> exo2       2.03179950 0.2146584  1.83443733    2.2755813
#> exo3      -0.48658501 0.3069460 -0.82721237   -0.1797494
#> 
#> $ARE$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1   0.7214274 0.5654403   0.1454632    1.3605630
#> lag1_var2   2.0271638 1.0374425   0.7428264    3.0752986
#> lag1_var3   3.4581485 1.2601625   1.9676061    4.6948096
#> lag1_var4  -2.5090400 1.3778795  -3.9156214   -0.9312061
#> const     -15.1803632 6.2418155 -18.9071937   -6.6857142
#> exo1       -0.5601762 0.3311390  -0.8632546   -0.2036578
#> exo2       -2.3137327 0.4262738  -2.8349271   -1.8540508
#> exo3       -1.0835525 0.5138277  -1.5578730   -0.4236564
#> 
#> $ARE$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1   0.7724111 0.3797175   0.4608530   1.28425667
#> lag1_var2   2.2130738 1.1664964   1.1890373   3.52622363
#> lag1_var3   2.9658121 1.4969050   1.7474020   4.62384322
#> lag1_var4  -2.0440170 1.5739483  -3.8319092  -0.76810469
#> const     -13.9723580 4.4689207 -17.8888148  -8.14298975
#> exo1       -0.4627172 0.3195742  -0.7947698  -0.06513175
#> exo2       -0.7457640 0.3691218  -1.2106983  -0.39205324
#> exo3       -1.3929534 0.2968879  -1.6575708  -1.00246942
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003070847 0.00129233 0.001798592  0.004673375
#> 
#> $ARG$Sigma$equation2
#>                   mean        sd 5% quantile 95% quantile
#> Sigma[2,1] -0.04782209 0.0234022 -0.08017943  -0.03322836
#> Sigma[2,2]  3.51838207 1.2516459  2.40787222   5.15927027
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04146163 0.01597739  0.02742155   0.06271154
#> Sigma[3,2] -2.43770851 0.93741067 -3.68095894  -1.69276070
#> Sigma[3,3]  1.93260954 0.66515272  1.42429987   2.81991259
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01434616 0.007295006  0.007972595   0.02406065
#> Sigma[4,2] -0.33177516 0.221914828 -0.591867699  -0.08461122
#> Sigma[4,3]  0.50377422 0.147015947  0.312978016   0.64536817
#> Sigma[4,4]  0.34185711 0.081363832  0.292908985   0.45323343
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.91432600 0.01111431  0.90628570   0.92893211
#> lag1_var2 -0.13861995 0.15867733 -0.28179140   0.03535650
#> lag1_var3 -0.22491565 0.25897223 -0.46013344   0.05851047
#> lag1_var4  0.19659217 0.23361876 -0.06162401   0.41171234
#> const      4.06641736 1.78516953  1.94121528   5.91613064
#> exo1       0.01843181 0.06600166 -0.05102052   0.10131055
#> exo2      -0.11290993 0.05439498 -0.17437589  -0.05674211
#> exo3       0.06306144 0.08735218 -0.05623543   0.12518026
#> 
#> $ARG$A$equation2
#>                mean         sd 5% quantile 95% quantile
#> lag1_var1 -2.997670  0.6433394  -3.4714409  -2.11929068
#> lag1_var2 -1.718550  1.6538449  -3.4942248  -0.08026908
#> lag1_var3 -3.833267  2.7636520  -6.8198862  -1.01788389
#> lag1_var4  4.175758  2.6069861   1.5352967   6.87242572
#> const     63.169106 24.7478648  31.1921698  87.20813089
#> exo1      -1.427143  2.3685943  -4.6357030   0.36215354
#> exo2       4.003467  1.4912068   2.3527899   5.26976180
#> exo3       2.558215  3.0418563   0.9167728   6.69831669
#> 
#> $ARG$A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  1.3843669  0.4671484   0.8239436     1.902457
#> lag1_var2  0.6016968  1.2918951  -0.7256018     2.131029
#> lag1_var3  1.5876415  2.1588198  -0.6866118     4.123215
#> lag1_var4 -1.1455575  2.1033157  -3.6483421     1.081236
#> const     -5.3603204 13.4091770 -19.4145286    11.092600
#> exo1       0.6323134  1.8049241  -0.7046984     3.058692
#> exo2      -6.0554846  1.3000802  -7.3500183    -4.521104
#> exo3       1.5964721  2.0673856  -1.2372885     3.128065
#> 
#> $ARG$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.4612725  0.2613983  -0.7357273   -0.1391761
#> lag1_var2 -0.5455427  0.5889815  -1.0655445    0.2285248
#> lag1_var3 -0.9164263  0.9800888  -1.7696309    0.3787693
#> lag1_var4  1.6439722  0.9770768   0.3981169    2.5189071
#> const     28.9653555 10.9529658  21.2561086   44.0617181
#> exo1      -0.3185467  0.6474593  -1.0753686    0.3446327
#> exo2      -4.1687598  0.5670892  -4.8312902   -3.5463395
#> exo3       3.5865301  0.7179375   2.6860581    4.3459053
#> 
#> 
#>