Provides posterior estimation summary for Bayesian Vector Autoregressions for dynamic panel data

Provides posterior mean, standard deviations, as well as 5 and 95 percentiles of the parameters for all C countries.

Usage

# S3 method for class 'PosteriorBVARs'
summary(object, ...)

Arguments

object

an object of class PosteriorBVARs obtained using the estimate() function applied to Vector Autoregressions containing draws from the posterior distribution of the parameters.

...

additional arguments affecting the summary produced.

Value

A list reporting the posterior mean, standard deviations, as well as 5 and 95 percentiles of the country-specific parameters.

See also

Author

Tomasz Woźniak wozniak.tom@pm.me

Examples

# specify the model
specification = specify_bvarPANEL$new(
      ilo_dynamic_panel[1:5], 
      exogenous = ilo_exogenous_variables[1:5])
burn_in       = estimate(specification, 5)             # run the burn-in
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
posterior     = estimate(burn_in, 5)                   # estimate the model
#> **************************************************|
#> bpvars: Forecasting with Bayesian Panel VARs      |
#> **************************************************|
#>  Progress of the MCMC simulation for 5 draws
#>     Every draw is saved via MCMC thinning
#>  Press Esc to interrupt the computations
#> **************************************************|
summary(posterior)
#> $AFG
#> $AFG$Sigma
#> $AFG$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003613499 0.001302426 0.002504676  0.005346289
#> 
#> $AFG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01268987 0.01066701 -0.02720621 -0.007313625
#> Sigma[2,2]  0.20148206 0.08849249  0.13952595  0.322500009
#> 
#> $AFG$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.006589634 0.004098258  0.001562412   0.01085633
#> Sigma[3,2] -0.130401187 0.024420905 -0.159616307  -0.10604944
#> Sigma[3,3]  0.234029789 0.041078694  0.190179444   0.28141766
#> 
#> $AFG$Sigma$equation4
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.0009870092 0.007646352 -0.008856921   0.00792858
#> Sigma[4,2] -0.0493275324 0.051242453 -0.098043940   0.01478794
#> Sigma[4,3]  0.2065962037 0.040342113  0.167220510   0.25669961
#> Sigma[4,4]  0.2171195720 0.046410415  0.174621233   0.26719279
#> 
#> 
#> $AFG$A
#> $AFG$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.95785655 0.02075619  0.92961671   0.97414347
#> lag1_var2 -0.02185725 0.17754246 -0.26313817   0.07830314
#> lag1_var3 -0.05960505 0.39937053 -0.59866156   0.21112376
#> lag1_var4  0.05578770 0.35292252 -0.19352946   0.53295523
#> const      1.15458750 2.10905006 -0.59958082   4.00010770
#> exo1       0.02434003 0.06020474 -0.03852469   0.09581915
#> exo2      -0.01669896 0.10391289 -0.14544053   0.07766581
#> exo3      -0.18526486 0.06604811 -0.25713230  -0.12153865
#> 
#> $AFG$A$equation2
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.57763262  0.1471242   0.3783390    0.6952576
#> lag1_var2 -0.10623657  1.7731158  -1.3321864    2.3335407
#> lag1_var3 -2.13900494  3.8353910  -4.8514674    3.1161909
#> lag1_var4  1.83818348  3.3382158  -2.7202074    4.2474978
#> const      1.89617286 20.4364839 -26.3499014   15.3777286
#> exo1      -0.27452334  0.5932836  -0.9993906    0.3531937
#> exo2       0.00543368  0.3997073  -0.4896259    0.4015163
#> exo3      -0.07333323  0.5931303  -0.6120872    0.7140765
#> 
#> $AFG$A$equation3
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  -0.2195981 0.1529599  -0.4184094  -0.08641492
#> lag1_var2   1.6701346 0.6504988   0.7850427   2.17177659
#> lag1_var3   5.3947523 1.5398278   3.3575301   6.78047008
#> lag1_var4  -3.9790890 1.3830574  -5.2805625  -2.17388270
#> const     -11.5123264 8.2610397 -19.0661348  -0.98006136
#> exo1        0.1640653 0.9937237  -0.7779735   1.43984011
#> exo2       -1.1997700 0.8907917  -2.3265588  -0.38826268
#> exo3        0.4389180 0.4216742  -0.1244221   0.80317654
#> 
#> $AFG$A$equation4
#>                   mean         sd  5% quantile 95% quantile
#> lag1_var1   0.03142430  0.1387911  -0.14663160    0.1537138
#> lag1_var2   1.79223279  1.2196041   0.64238030    3.3867332
#> lag1_var3   4.87858100  2.8149880   2.26872926    8.5912624
#> lag1_var4  -3.42522461  2.4769915  -6.69788183   -1.1446852
#> const     -18.11469722 15.9620345 -39.60560394   -4.5280528
#> exo1        0.09556883  0.8820755  -0.82517693    1.1800727
#> exo2       -1.27310953  0.8724240  -2.38816761   -0.4263989
#> exo3        0.58681787  0.5727946  -0.08768773    1.2089560
#> 
#> 
#> 
#> $AGO
#> $AGO$Sigma
#> $AGO$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.007989842 0.001326245 0.006458953  0.009515882
#> 
#> $AGO$Sigma$equation2
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[2,1] -0.01416905 0.008740442 -0.02552898 -0.005917264
#> Sigma[2,2]  0.13984408 0.036165513  0.09779414  0.172001988
#> 
#> $AGO$Sigma$equation3
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.009642783 0.003930074  0.005169973   0.01345355
#> Sigma[3,2] -0.056235002 0.015463248 -0.072683235  -0.03855047
#> Sigma[3,3]  0.040486138 0.010380898  0.031800766   0.05393739
#> 
#> $AGO$Sigma$equation4
#>                     mean          sd  5% quantile 95% quantile
#> Sigma[4,1] -0.0002480794 0.005288812 -0.007162558  0.004813722
#> Sigma[4,2]  0.0578387266 0.020516545  0.035493493  0.081317891
#> Sigma[4,3] -0.0024199292 0.007094974 -0.011919854  0.003656546
#> Sigma[4,4]  0.0496950752 0.016535802  0.031378096  0.069617808
#> 
#> 
#> $AGO$A
#> $AGO$A$equation1
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1  0.93164923 0.03708895  0.90608011   0.98264877
#> lag1_var2  0.23440937 0.18305627 -0.01132102   0.38840242
#> lag1_var3  0.29851510 0.25658731 -0.04452490   0.50735521
#> lag1_var4 -0.28403013 0.22792710 -0.46382678   0.02246575
#> const      0.52294757 1.34242344 -1.12869698   1.89281609
#> exo1       0.11839073 0.03041874  0.08399369   0.15103712
#> exo2      -0.16181547 0.07255837 -0.21431209  -0.06274151
#> exo3      -0.08075177 0.10357559 -0.15954415   0.04390165
#> 
#> $AGO$A$equation2
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1  0.07652021 0.1202854 -0.04203360    0.2315253
#> lag1_var2  0.41445738 1.0192554 -0.95962227    1.2808386
#> lag1_var3 -0.71266105 1.2810303 -2.45859324    0.2887895
#> lag1_var4  0.69120856 1.1932306 -0.29311441    2.3063950
#> const      0.31649176 6.0089549 -6.34003632    7.4450731
#> exo1       0.02406850 0.3281206 -0.34802085    0.3689745
#> exo2       0.40441676 0.2755042  0.04739779    0.6598923
#> exo3      -0.52954852 0.5919612 -0.96978655    0.2567840
#> 
#> $AGO$A$equation3
#>                  mean         sd   5% quantile 95% quantile
#> lag1_var1 -0.25509005 0.05794442 -0.3092919654   -0.1822200
#> lag1_var2  0.34706533 0.30358983  0.0554242893    0.7356008
#> lag1_var3  1.43462948 0.38301328  1.1291123099    1.9502362
#> lag1_var4 -0.40940831 0.34030126 -0.8604306923   -0.1103972
#> const      4.08459776 3.21060897 -0.2450853779    6.7452020
#> exo1       0.07704193 0.09590462 -0.0002582272    0.2050509
#> exo2      -1.49633075 0.18532710 -1.7122288086   -1.3137797
#> exo3       3.27342293 0.19379669  3.0640082925    3.4762416
#> 
#> $AGO$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -0.2982214 0.08645402  -0.3835264   -0.1896627
#> lag1_var2  0.5543573 0.68323394  -0.3241607    1.0871970
#> lag1_var3  0.7404541 0.86243811  -0.3737785    1.4332854
#> lag1_var4  0.3525468 0.82140535  -0.3169376    1.4162332
#> const      0.3269181 2.57671404  -3.1196810    2.5548390
#> exo1       0.1532670 0.33595927  -0.1945251    0.5259167
#> exo2      -1.4899498 0.13315650  -1.6177754   -1.3161942
#> exo3       3.2432143 0.37546999   2.9105166    3.7384105
#> 
#> 
#> 
#> $ALB
#> $ALB$Sigma
#> $ALB$Sigma$equation1
#>                   mean           sd 5% quantile 95% quantile
#> Sigma[1,1] 0.002555297 0.0006587593 0.001888477  0.003357066
#> 
#> $ALB$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.09495711 0.05609804   -0.169597  -0.04600654
#> Sigma[2,2] 10.19454913 4.29892070    7.726532  16.09145981
#> 
#> $ALB$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.03814892 0.04285348 -0.01411998   0.08405826
#> Sigma[3,2] -6.23170047 2.77094939 -9.75799868  -3.53100291
#> Sigma[3,3]  6.21993271 2.25188274  3.24055960   8.10099644
#> 
#> $ALB$Sigma$equation4
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[4,1] -0.03226509 0.02581199 -0.05645054 -0.001155852
#> Sigma[4,2]  0.75375273 1.52330860 -0.54328476  2.710848341
#> Sigma[4,3]  2.43173625 1.88259520  0.92423756  4.892241725
#> Sigma[4,4]  3.54686147 1.53406032  2.05813568  5.516198096
#> 
#> 
#> $ALB$A
#> $ALB$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.99662564 0.01628333  0.975840431  1.013007458
#> lag1_var2 -0.03648793 0.02219931 -0.056735517 -0.007654006
#> lag1_var3 -0.06220193 0.03149541 -0.089606928 -0.024443630
#> lag1_var4  0.05331281 0.02611857  0.021598141  0.076833480
#> const      0.62608383 0.53484646  0.154215249  1.337369775
#> exo1       0.05052975 0.04013268 -0.003463547  0.082894420
#> exo2      -0.08249821 0.03505664 -0.117594968 -0.044041080
#> exo3       0.08733707 0.03285243  0.046964775  0.114184943
#> 
#> $ALB$A$equation2
#>                mean        sd 5% quantile 95% quantile
#> lag1_var1 -3.411736  1.410965   -5.229140    -2.002039
#> lag1_var2  3.416645  1.268886    2.181981     4.828869
#> lag1_var3  4.816185  1.899789    3.329224     7.147599
#> lag1_var4 -3.929052  1.626422   -5.917212    -2.503565
#> const     32.451722 36.401325   -0.603242    79.918537
#> exo1      -1.566431  1.944430   -3.217699     1.040741
#> exo2      -0.643449  2.062917   -2.702666     1.434584
#> exo3      -2.287976  2.754858   -4.617683     1.391198
#> 
#> $ALB$A$equation3
#>                  mean         sd 5% quantile 95% quantile
#> lag1_var1   2.5520321  0.8334054    1.826797    3.6550804
#> lag1_var2  -2.1642739  1.6672376   -3.767478   -0.1153076
#> lag1_var3  -3.0074387  2.3886348   -5.376994   -0.2311459
#> lag1_var4   3.1540588  2.0740259    0.665530    5.1627363
#> const     -10.9646582 38.8068213  -59.195454   27.4135070
#> exo1       -0.9469667  1.1701751   -1.989496    0.3431289
#> exo2        0.1085144  1.6332367   -1.601750    1.8404099
#> exo3        0.7145907  2.2878408   -0.948488    3.8039027
#> 
#> $ALB$A$equation4
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.3383946  0.6839591  -0.1799455    1.2637376
#> lag1_var2 -0.2953917  1.2442589  -1.2979117    1.3369114
#> lag1_var3 -0.3781359  1.6924151  -1.6343877    1.8801714
#> lag1_var4  1.1615104  1.4344840  -0.7768348    2.2292936
#> const      6.5090782 32.4357278 -34.2312784   31.5332236
#> exo1      -1.9457220  1.4810406  -3.5547758   -0.1396556
#> exo2      -0.3085173  0.3085848  -0.6339078    0.0763706
#> exo3      -0.9300713  3.2335741  -4.0185371    2.6589834
#> 
#> 
#> 
#> $ARE
#> $ARE$Sigma
#> $ARE$Sigma$equation1
#>                   mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.003043984 0.001732472 0.001225604  0.005070018
#> 
#> $ARE$Sigma$equation2
#>                    mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.004314636 0.00249572 -0.00745314 -0.001706262
#> Sigma[2,2]  0.056416035 0.01027336  0.04539910  0.067811210
#> 
#> $ARE$Sigma$equation3
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[3,1]  0.01156369 0.005077202  0.006848036   0.01811483
#> Sigma[3,2] -0.05121496 0.022384326 -0.074482162  -0.02712183
#> Sigma[3,3]  0.20768549 0.048494460  0.161108890   0.26447796
#> 
#> $ARE$Sigma$equation4
#>                    mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.008237342 0.003962301  0.003417886  0.012347970
#> Sigma[4,2] -0.007341453 0.015624812 -0.024790671  0.009549773
#> Sigma[4,3]  0.170506872 0.051666805  0.121377263  0.228987387
#> Sigma[4,4]  0.168357060 0.055861963  0.112240646  0.235064496
#> 
#> 
#> $ARE$A
#> $ARE$A$equation1
#>                   mean         sd 5% quantile 95% quantile
#> lag1_var1  0.884047609 0.08150341  0.78721620   0.97703165
#> lag1_var2  0.873922551 0.19531859  0.62347281   1.04087818
#> lag1_var3  1.112369070 0.24478001  0.80017712   1.30900711
#> lag1_var4 -1.081623518 0.22786522 -1.27165697  -0.79555926
#> const      0.679764536 0.61536644  0.09468832   1.48088833
#> exo1       0.005775113 0.05730300 -0.06495369   0.06097028
#> exo2      -0.021893637 0.06406662 -0.10063511   0.03557802
#> exo3       0.053746776 0.04639968  0.01230911   0.10860142
#> 
#> $ARE$A$equation2
#>                   mean        sd 5% quantile 95% quantile
#> lag1_var1  0.110001303 0.1619035 -0.01899527    0.3286449
#> lag1_var2  0.015475042 0.8497841 -1.00878396    0.7819923
#> lag1_var3 -0.866941223 1.0535113 -2.13802420    0.0712105
#> lag1_var4  0.821407186 0.9856022 -0.07880507    1.9881045
#> const      1.357035825 3.7255789 -2.28790897    5.7504363
#> exo1      -0.001841498 0.2131367 -0.25243802    0.2170234
#> exo2       1.876757617 0.1796983  1.64504000    2.0551578
#> exo3      -0.613908760 0.2698362 -0.94220266   -0.3444607
#> 
#> $ARE$A$equation3
#>                 mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.1729816 0.1966920  -0.3651717   0.07577911
#> lag1_var2  4.8360761 3.3626762   2.0078808   8.94671131
#> lag1_var3  6.7712596 4.3963266   3.1180166  12.17547642
#> lag1_var4 -5.6947984 4.3286487 -10.9835266  -2.08385513
#> const     -2.0774077 6.9193783  -9.8060564   5.09976783
#> exo1      -0.4107719 0.5385243  -0.8951156   0.30701311
#> exo2      -1.9293461 0.5027303  -2.5072169  -1.33713755
#> exo3      -1.3853435 0.2611444  -1.5633977  -1.03186295
#> 
#> $ARE$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.08097931 0.3052561  -0.3328493    0.3291138
#> lag1_var2  5.01510764 2.8141023   2.7170827    8.5778040
#> lag1_var3  6.32803281 3.7183672   3.3290336   11.0551286
#> lag1_var4 -5.25004067 3.6954048  -9.9000193   -2.2589130
#> const     -4.01341447 4.5969671  -7.8760969    1.7975153
#> exo1      -0.42882289 0.5402717  -0.8243122    0.3048888
#> exo2      -0.50898905 0.5516985  -1.1266591    0.1362352
#> exo3      -1.89788226 0.3741697  -2.3049278   -1.4493567
#> 
#> 
#> 
#> $ARG
#> $ARG$Sigma
#> $ARG$Sigma$equation1
#>                  mean          sd 5% quantile 95% quantile
#> Sigma[1,1] 0.00314776 0.001124976  0.00189301  0.004480305
#> 
#> $ARG$Sigma$equation2
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[2,1] -0.05806179 0.01830365 -0.08029515  -0.03799197
#> Sigma[2,2]  3.02327452 0.45456190  2.62134822   3.57780025
#> 
#> $ARG$Sigma$equation3
#>                   mean         sd 5% quantile 95% quantile
#> Sigma[3,1]  0.04881865 0.01870515  0.02840882   0.07173371
#> Sigma[3,2] -2.17941157 0.51755942 -2.80275061  -1.64258311
#> Sigma[3,3]  1.89389213 0.52234391  1.35650602   2.49402328
#> 
#> $ARG$Sigma$equation4
#>                   mean          sd  5% quantile 95% quantile
#> Sigma[4,1]  0.01576863 0.008514708  0.006700148   0.02605283
#> Sigma[4,2] -0.37744697 0.277483215 -0.694331100  -0.05234357
#> Sigma[4,3]  0.63334350 0.235261774  0.379144970   0.87590875
#> Sigma[4,4]  0.45274614 0.115011402  0.298785280   0.54913486
#> 
#> 
#> $ARG$A
#> $ARG$A$equation1
#>                  mean         sd  5% quantile 95% quantile
#> lag1_var1  0.88443070 0.02099111  0.856025337   0.90179019
#> lag1_var2  0.32394233 0.07355559  0.230115710   0.39856296
#> lag1_var3  0.53309670 0.11599531  0.383976771   0.64827735
#> lag1_var4 -0.49331953 0.11735630 -0.623471080  -0.35144780
#> const      0.72038408 0.86067037 -0.370060894   1.40391995
#> exo1       0.04038662 0.03676990 -0.004623522   0.07668112
#> exo2      -0.14518997 0.03902281 -0.178469355  -0.10020785
#> exo3      -0.02325039 0.10168602 -0.109226594   0.11374493
#> 
#> $ARG$A$equation2
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1 -1.7567740  0.7101293  -2.5243824   -0.9644638
#> lag1_var2  1.5174029  3.4205659  -2.4256696    5.5191983
#> lag1_var3  1.2624326  5.5603748  -5.1156429    7.7740499
#> lag1_var4 -0.5983350  5.2648094  -6.5990348    5.6437403
#> const      9.3363133 21.6529849  -7.8445607   38.3977666
#> exo1      -2.6894878  0.9609456  -3.6541812   -1.5444475
#> exo2       1.9316174  1.1538894   0.6798355    3.1644710
#> exo3       0.3562675  2.2836052  -2.5575256    2.5289095
#> 
#> $ARG$A$equation3
#>                 mean         sd 5% quantile 95% quantile
#> lag1_var1  0.2688207  0.6356794  -0.4761174    0.9356506
#> lag1_var2  2.9749073  3.0006450  -1.0910297    5.3644813
#> lag1_var3  5.6416007  4.8886552  -0.9890747    9.5397127
#> lag1_var4 -4.7580247  4.5462606  -8.2380034    1.4001512
#> const     -1.4070617 17.7331002 -24.5918490   13.9956556
#> exo1       1.3648780  0.7702665   0.6651886    2.2876724
#> exo2      -5.2543176  1.5426411  -7.0404185   -3.5738115
#> exo3       1.4243259  1.1047181   0.3136671    2.8356994
#> 
#> $ARG$A$equation4
#>                  mean        sd 5% quantile 95% quantile
#> lag1_var1 -0.90267330 0.3431039  -1.2690101  -0.53514252
#> lag1_var2  4.19063631 1.6596886   2.1408345   5.90049741
#> lag1_var3  6.88637225 2.7352697   3.5077432   9.67284928
#> lag1_var4 -5.48383857 2.5101806  -8.0384174  -2.35097889
#> const      0.05754118 6.3740950  -6.2027082   7.84731435
#> exo1      -0.31511685 0.4291715  -0.8695794   0.08802026
#> exo2      -4.61449687 0.9774679  -5.7454277  -3.56113995
#> exo3       1.80402041 0.8837556   0.7329201   2.77235840
#> 
#> 
#>